Retrieve Asset Information from Kraken Exchange

This function fetches detailed asset information from the Kraken API, either for all available assets or a specified subset.

Usage

getAssets(assets = "All")

Arguments

assets

A character vector specifying the assets to retrieve. Use "All" to fetch data for all assets. For specific assets, provide their abbreviations (e.g., "ADA" or c("BTC", "ETH")). Default is "All".

Value

A data frame containing detailed information about the requested assets.

Examples

getAssets()
#> # A tibble: 372 × 8
#>    AssetID  aclass   altname  decimals display_decimals status  Asset   
#>    <chr>    <chr>    <chr>       <int>            <int> <chr>   <chr>   
#>  1 1INCH    currency 1INCH          10                5 enabled 1INCH   
#>  2 AAVE     currency AAVE           10                5 enabled AAVE    
#>  3 ACA      currency ACA            10                5 enabled ACA     
#>  4 ACH      currency ACH            10                5 enabled ACH     
#>  5 ADA      currency ADA             8                6 enabled ADA     
#>  6 ADA.S    currency ADA.S           8                6 enabled ADA.S   
#>  7 ADX      currency ADX            10                5 enabled ADX     
#>  8 AED      currency AED             4                2 enabled AED     
#>  9 AED.HOLD currency AED.HOLD        4                2 enabled AED.HOLD
#> 10 AEVO     currency AEVO            6                4 enabled AEVO    
#> # ℹ 362 more rows
#> # ℹ 1 more variable: collateral_value <dbl>
getAssets("ETH")
#> # A tibble: 1 × 8
#>   AssetID aclass altname decimals display_decimals collateral_value status Asset
#>   <chr>   <chr>  <chr>      <int>            <int>            <dbl> <chr>  <chr>
#> 1 XETH    curre… ETH           10                5                1 enabl… XETH 
getAssets(c("BTC", "AAVE", "ADA"))
#> # A tibble: 3 × 8
#>   AssetID aclass altname decimals display_decimals collateral_value status Asset
#>   <chr>   <chr>  <chr>      <int>            <int>            <dbl> <chr>  <chr>
#> 1 AAVE    curre… AAVE          10                5              0.8 enabl… AAVE 
#> 2 ADA     curre… ADA            8                6              0.9 enabl… ADA  
#> 3 BTC     curre… XBT           10                5              1   enabl… BTC