Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 2    0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 3    0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 4    0.960382   1643.565 2024-12-18 17:34:45        Bid        NA         NA
#> 5    0.960446   1616.468 2024-12-18 17:31:03        Bid        NA         NA
#> 6    0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 7    0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 8    0.961313  12680.200 2024-12-18 17:32:51        Bid        NA         NA
#> 9    0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 10   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 11   0.962967   1771.650 2024-12-18 17:33:14        Bid        NA         NA
#> 12   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 13   0.963306   3616.244 2024-12-18 17:33:48        Bid        NA         NA
#> 14   0.963348   3094.272 2024-12-18 17:33:16        Bid        NA         NA
#> 15   0.963454   6106.993 2024-12-18 17:34:51        Bid        NA         NA
#> 16   0.963970  12454.707 2024-12-18 17:33:25        Bid        NA         NA
#> 17   0.964109   8185.278 2024-12-18 17:34:04        Bid        NA         NA
#> 18   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 19   0.965000    145.118 2024-12-18 16:58:19        Bid        NA         NA
#> 20   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 21   0.965596  59096.553 2024-12-18 17:27:00        Bid        NA         NA
#> 22   0.965624  15171.846 2024-12-18 17:31:12        Bid        NA         NA
#> 23   0.966241  25256.400 2024-12-18 17:34:31        Bid        NA         NA
#> 24   0.966405   5053.233 2024-12-18 17:31:42        Bid        NA         NA
#> 25   0.967047   8086.404 2024-12-18 17:31:04        Bid        NA         NA
#> 26   0.967903  14794.804 2024-12-18 17:33:49        Bid        NA         NA
#> 27   0.968493  25124.632 2024-12-18 17:34:59        Bid        NA         NA
#> 28   0.968700    138.916 2024-12-18 17:34:12        Bid        NA         NA
#> 29   0.968724   3138.230 2024-12-18 17:34:52        Bid        NA         NA
#> 30   0.968911  38476.100 2024-12-18 17:34:43        Bid        NA         NA
#> 31   0.969425  15720.934 2024-12-18 17:34:00        Bid        NA         NA
#> 32   0.969468   1691.964 2024-12-18 17:34:25        Bid        NA         NA
#> 33   0.969523  19207.356 2024-12-18 17:34:43        Bid        NA         NA
#> 34   0.970070    154.165 2024-12-18 17:29:38        Bid        NA         NA
#> 35   0.970071 288995.359 2024-12-18 17:34:44        Bid        NA         NA
#> 36   0.970425   2485.489 2024-12-18 17:33:58        Bid        NA         NA
#> 37   0.970510   2289.265 2024-12-18 17:34:18        Bid        NA         NA
#> 38   0.970656   3762.044 2024-12-18 17:34:51        Bid        NA         NA
#> 39   0.970767  14751.163 2024-12-18 17:34:03        Bid        NA         NA
#> 40   0.970903  15189.107 2024-12-18 17:34:39        Bid        NA         NA
#> 41   0.971089  38248.600 2024-12-18 17:34:51        Bid        NA         NA
#> 42   0.971890   2359.956 2024-12-18 17:34:50        Bid        NA         NA
#> 43   0.972562     97.523 2024-12-18 17:34:22        Bid        NA         NA
#> 44   0.972820   6450.110 2024-12-18 17:34:51        Bid        NA         NA
#> 45   0.973922 203301.510 2024-12-18 17:34:35        Bid        NA         NA
#> 46   0.973970  10964.541 2024-12-18 17:34:51        Bid        NA         NA
#> 47   0.973971     41.743 2024-12-18 17:34:40        Bid        NA         NA
#> 48   0.973993   3255.604 2024-12-18 17:34:45        Bid        NA         NA
#> 49   0.974042   4514.300 2024-12-18 17:34:51        Bid        NA         NA
#> 50   0.974043   6877.761 2024-12-18 17:34:51        Bid        NA         NA
#> 51   0.974495  14694.722 2024-12-18 17:34:47        Bid        NA         NA
#> 52   0.974496     33.641 2024-12-18 17:34:48        Bid        NA         NA
#> 53   0.974513     49.256 2024-12-18 17:33:21        Bid        NA         NA
#> 54   0.974514     20.530 2024-12-18 17:34:48        Bid        NA         NA
#> 55   0.974542  11606.081 2024-12-18 17:34:51        Bid        NA         NA
#> 56   0.974543    111.829 2024-12-18 17:34:58        Bid        NA         NA
#> 57   0.975200   1289.509 2024-12-18 17:34:51        Bid        NA         NA
#> 58   0.975201     17.544 2024-12-18 17:34:58        Bid        NA         NA
#> 59   0.975250    644.755 2024-12-18 17:34:51        Bid        NA         NA
#> 60   0.975868     43.557 2024-12-18 17:34:39        Bid        NA         NA
#> 61   0.976092  17361.000 2024-12-18 17:34:31        Bid        NA         NA
#> 62   0.976245  14668.383 2024-12-18 17:34:43        Bid        NA         NA
#> 63   0.976608   3942.883 2024-12-18 17:34:45        Bid        NA         NA
#> 64   0.976902   5006.639 2024-12-18 17:34:45        Bid        NA         NA
#> 65   0.977142   9600.921 2024-12-18 17:34:42        Bid        NA         NA
#> 66   0.977316  14652.306 2024-12-18 17:34:54        Bid        NA         NA
#> 67   0.977476   2953.000 2024-12-18 17:34:05        Bid        NA         NA
#> 68   0.977477   5000.000 2024-12-18 17:34:43        Bid        NA         NA
#> 69   0.977478   2755.703 2024-12-18 17:34:43        Bid        NA         NA
#> 70   0.977700     49.000 2024-12-18 17:34:56        Bid        NA         NA
#> 71   0.977730  16025.100 2024-12-18 17:34:42        Bid        NA         NA
#> 72   0.978000  14640.178 2024-12-18 17:34:57        Bid        NA         NA
#> 73   0.978084     14.454 2024-12-18 17:34:47        Bid        NA         NA
#> 74   0.978160  10157.430 2024-12-18 17:34:43        Bid        NA         NA
#> 75   0.978174  10136.100 2024-12-18 17:34:54        Bid        NA         NA
#> 76   0.978208   9643.500 2024-12-18 17:34:40        Bid        NA         NA
#> 77   0.978300  10518.310 2024-12-18 17:34:43        Bid        NA         NA
#> 78   0.978397  10390.180 2024-12-18 17:34:44        Bid        NA         NA
#> 79   0.978477    504.339 2024-12-18 17:34:53        Bid        NA         NA
#> 80   0.978597   9754.161 2024-12-18 17:34:58        Bid        NA         NA
#> 81   0.978610   2449.000 2024-12-18 17:34:54        Bid        NA         NA
#> 82   0.978669   4876.000 2024-12-18 17:34:54        Bid        NA         NA
#> 83   0.978746     16.698 2024-12-18 17:34:47        Bid        NA         NA
#> 84   0.978747     88.070 2024-12-18 17:34:56        Bid        NA         NA
#> 85   0.979078   9749.375 2024-12-18 17:34:57        Bid        NA         NA
#> 86   0.979148   1043.864 2024-12-18 17:34:57        Bid        NA         NA
#> 87   0.979312    490.073 2024-12-18 17:34:45        Bid        NA         NA
#> 88   0.979366    537.800 2024-12-18 17:34:54        Bid        NA         NA
#> 89   0.979394   2100.000 2024-12-18 17:34:56        Bid        NA         NA
#> 90   0.979421   3894.759 2024-12-18 17:34:59        Bid        NA         NA
#> 91   0.979436   4872.907 2024-12-18 17:35:00        Bid        NA         NA
#> 92   0.979463    190.037 2024-12-18 17:34:59        Bid        NA         NA
#> 93   0.979467   5876.592 2024-12-18 17:35:00        Bid        NA         NA
#> 94   0.979485   1000.000 2024-12-18 17:34:59        Bid        NA         NA
#> 95   0.979488   3576.965 2024-12-18 17:35:00        Bid        NA         NA
#> 96   0.979489     67.597 2024-12-18 17:35:00        Bid        NA         NA
#> 97   0.979491  10454.293 2024-12-18 17:35:00        Bid        NA         NA
#> 98   0.979492    240.092 2024-12-18 17:35:00        Bid        NA         NA
#> 99   0.979493  14617.870 2024-12-18 17:35:00        Bid        NA         NA
#> 100  0.979494   1375.600 2024-12-18 17:35:00        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.980145    476.865
#> 102        NA         NA                <NA>        Ask  0.980372   2434.002
#> 103        NA         NA                <NA>        Ask  0.980421   1125.808
#> 104        NA         NA                <NA>        Ask  0.980422   1324.944
#> 105        NA         NA                <NA>        Ask  0.980427   4867.730
#> 106        NA         NA                <NA>        Ask  0.980540   7301.138
#> 107        NA         NA                <NA>        Ask  0.980654   9733.714
#> 108        NA         NA                <NA>        Ask  0.980826   3709.066
#> 109        NA         NA                <NA>        Ask  0.980831   9731.954
#> 110        NA         NA                <NA>        Ask  0.980944  10000.000
#> 111        NA         NA                <NA>        Ask  0.980962    537.800
#> 112        NA         NA                <NA>        Ask  0.981101   4839.316
#> 113        NA         NA                <NA>        Ask  0.981197   9643.400
#> 114        NA         NA                <NA>        Ask  0.981221   2127.859
#> 115        NA         NA                <NA>        Ask  0.981229   9441.079
#> 116        NA         NA                <NA>        Ask  0.981327   9727.035
#> 117        NA         NA                <NA>        Ask  0.981406  15299.700
#> 118        NA         NA                <NA>        Ask  0.981533   2434.361
#> 119        NA         NA                <NA>        Ask  0.981688     18.034
#> 120        NA         NA                <NA>        Ask  0.981802   9517.970
#> 121        NA         NA                <NA>        Ask  0.981809   1020.451
#> 122        NA         NA                <NA>        Ask  0.981825   1735.252
#> 123        NA         NA                <NA>        Ask  0.981968   1460.428
#> 124        NA         NA                <NA>        Ask  0.981969     24.102
#> 125        NA         NA                <NA>        Ask  0.982000     41.491
#> 126        NA         NA                <NA>        Ask  0.982073  14581.357
#> 127        NA         NA                <NA>        Ask  0.982098    138.225
#> 128        NA         NA                <NA>        Ask  0.982123   7909.540
#> 129        NA         NA                <NA>        Ask  0.982139   8356.120
#> 130        NA         NA                <NA>        Ask  0.982400   5000.000
#> 131        NA         NA                <NA>        Ask  0.982410  14574.478
#> 132        NA         NA                <NA>        Ask  0.983137    429.477
#> 133        NA         NA                <NA>        Ask  0.983138    651.408
#> 134        NA         NA                <NA>        Ask  0.983139   3257.038
#> 135        NA         NA                <NA>        Ask  0.983188  15880.171
#> 136        NA         NA                <NA>        Ask  0.983439   5862.668
#> 137        NA         NA                <NA>        Ask  0.983488  15773.184
#> 138        NA         NA                <NA>        Ask  0.983536     99.610
#> 139        NA         NA                <NA>        Ask  0.983537  14559.654
#> 140        NA         NA                <NA>        Ask  0.983583   3908.445
#> 141        NA         NA                <NA>        Ask  0.983611     43.805
#> 142        NA         NA                <NA>        Ask  0.983632  18821.704
#> 143        NA         NA                <NA>        Ask  0.984015   4559.853
#> 144        NA         NA                <NA>        Ask  0.984064  17350.300
#> 145        NA         NA                <NA>        Ask  0.984150   5211.260
#> 146        NA         NA                <NA>        Ask  0.984199  39764.400
#> 147        NA         NA                <NA>        Ask  0.985052   5862.668
#> 148        NA         NA                <NA>        Ask  0.985101  24826.063
#> 149        NA         NA                <NA>        Ask  0.985142     10.229
#> 150        NA         NA                <NA>        Ask  0.985264     48.681
#> 151        NA         NA                <NA>        Ask  0.985343   6514.075
#> 152        NA         NA                <NA>        Ask  0.985392  14532.239
#> 153        NA         NA                <NA>        Ask  0.985967  37945.700
#> 154        NA         NA                <NA>        Ask  0.986014     11.586
#> 155        NA         NA                <NA>        Ask  0.986086   4746.388
#> 156        NA         NA                <NA>        Ask  0.986415    137.620
#> 157        NA         NA                <NA>        Ask  0.986906     23.925
#> 158        NA         NA                <NA>        Ask  0.986954    144.325
#> 159        NA         NA                <NA>        Ask  0.987168  26581.100
#> 160        NA         NA                <NA>        Ask  0.987730  11687.854
#> 161        NA         NA                <NA>        Ask  0.987926   9220.683
#> 162        NA         NA                <NA>        Ask  0.987979     23.956
#> 163        NA         NA                <NA>        Ask  0.988410   8762.563
#> 164        NA         NA                <NA>        Ask  0.988411   2247.805
#> 165        NA         NA                <NA>        Ask  0.988625   2439.943
#> 166        NA         NA                <NA>        Ask  0.988869    137.097
#> 167        NA         NA                <NA>        Ask  0.988870  14481.133
#> 168        NA         NA                <NA>        Ask  0.989518  29208.541
#> 169        NA         NA                <NA>        Ask  0.990156   3687.953
#> 170        NA         NA                <NA>        Ask  0.990374 288987.161
#> 171        NA         NA                <NA>        Ask  0.990375  12732.200
#> 172        NA         NA                <NA>        Ask  0.990614  14455.636
#> 173        NA         NA                <NA>        Ask  0.990750    137.018
#> 174        NA         NA                <NA>        Ask  0.991176   3067.141
#> 175        NA         NA                <NA>        Ask  0.991454     11.529
#> 176        NA         NA                <NA>        Ask  0.992245     31.277
#> 177        NA         NA                <NA>        Ask  0.993136     55.456
#> 178        NA         NA                <NA>        Ask  0.993224     78.674
#> 179        NA         NA                <NA>        Ask  0.993634     48.271
#> 180        NA         NA                <NA>        Ask  0.994000     41.491
#> 181        NA         NA                <NA>        Ask  0.994025     23.810
#> 182        NA         NA                <NA>        Ask  0.994142     98.547
#> 183        NA         NA                <NA>        Ask  0.994601   6487.000
#> 184        NA         NA                <NA>        Ask  0.995104    136.419
#> 185        NA         NA                <NA>        Ask  0.996211     98.300
#> 186        NA         NA                <NA>        Ask  0.996922     11.466
#> 187        NA         NA                <NA>        Ask  0.997340    791.195
#> 188        NA         NA                <NA>        Ask  0.997602   2989.650
#> 189        NA         NA                <NA>        Ask  0.997608    791.118
#> 190        NA         NA                <NA>        Ask  0.998591   5444.061
#> 191        NA         NA                <NA>        Ask  0.998663  14340.634
#> 192        NA         NA                <NA>        Ask  0.998690      7.057
#> 193        NA         NA                <NA>        Ask  0.998818   9459.351
#> 194        NA         NA                <NA>        Ask  0.999478    135.822
#> 195        NA         NA                <NA>        Ask  0.999602   1709.991
#> 196        NA         NA                <NA>        Ask  1.000000   3727.390
#> 197        NA         NA                <NA>        Ask  1.000109     23.665
#> 198        NA         NA                <NA>        Ask  1.000330   6626.100
#> 199        NA         NA                <NA>        Ask  1.001000      8.018
#> 200        NA         NA                <NA>        Ask  1.001581    109.838
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:34:45
#> 102 2024-12-18 17:35:00
#> 103 2024-12-18 17:35:00
#> 104 2024-12-18 17:34:59
#> 105 2024-12-18 17:35:00
#> 106 2024-12-18 17:35:00
#> 107 2024-12-18 17:35:00
#> 108 2024-12-18 17:34:47
#> 109 2024-12-18 17:34:57
#> 110 2024-12-18 17:35:00
#> 111 2024-12-18 17:34:57
#> 112 2024-12-18 17:34:54
#> 113 2024-12-18 17:34:43
#> 114 2024-12-18 17:34:57
#> 115 2024-12-18 17:34:51
#> 116 2024-12-18 17:34:57
#> 117 2024-12-18 17:34:51
#> 118 2024-12-18 17:34:59
#> 119 2024-12-18 17:34:47
#> 120 2024-12-18 17:34:44
#> 121 2024-12-18 17:34:47
#> 122 2024-12-18 17:34:47
#> 123 2024-12-18 17:34:48
#> 124 2024-12-18 16:41:15
#> 125 2024-12-18 16:51:06
#> 126 2024-12-18 17:34:43
#> 127 2024-12-18 17:04:34
#> 128 2024-12-18 17:34:45
#> 129 2024-12-18 17:34:47
#> 130 2024-12-18 17:34:44
#> 131 2024-12-18 17:34:57
#> 132 2024-12-18 17:34:54
#> 133 2024-12-18 17:34:53
#> 134 2024-12-18 17:34:53
#> 135 2024-12-18 17:34:44
#> 136 2024-12-18 17:34:53
#> 137 2024-12-18 17:34:43
#> 138 2024-12-18 17:31:29
#> 139 2024-12-18 17:34:47
#> 140 2024-12-18 17:34:53
#> 141 2024-12-18 17:34:35
#> 142 2024-12-18 17:34:03
#> 143 2024-12-18 17:34:53
#> 144 2024-12-18 17:34:53
#> 145 2024-12-18 17:34:53
#> 146 2024-12-18 17:34:44
#> 147 2024-12-18 17:34:53
#> 148 2024-12-18 17:34:03
#> 149 2024-12-18 17:34:47
#> 150 2024-12-18 14:56:59
#> 151 2024-12-18 17:34:53
#> 152 2024-12-18 17:34:47
#> 153 2024-12-18 17:34:43
#> 154 2024-12-18 15:09:59
#> 155 2024-12-18 17:34:48
#> 156 2024-12-18 16:58:52
#> 157 2024-12-18 17:34:33
#> 158 2024-12-18 17:34:51
#> 159 2024-12-18 17:33:57
#> 160 2024-12-18 17:34:45
#> 161 2024-12-18 17:34:17
#> 162 2024-12-18 14:35:26
#> 163 2024-12-18 17:34:48
#> 164 2024-12-18 17:34:33
#> 165 2024-12-18 17:34:44
#> 166 2024-12-18 17:34:52
#> 167 2024-12-18 17:33:51
#> 168 2024-12-18 17:34:48
#> 169 2024-12-18 17:34:48
#> 170 2024-12-18 17:34:39
#> 171 2024-12-18 17:34:17
#> 172 2024-12-18 17:34:47
#> 173 2024-12-18 13:57:45
#> 174 2024-12-18 17:34:34
#> 175 2024-12-18 12:37:59
#> 176 2024-12-18 15:04:40
#> 177 2024-12-18 13:26:56
#> 178 2024-12-18 15:12:45
#> 179 2024-12-18 12:38:18
#> 180 2024-12-18 12:25:06
#> 181 2024-12-18 13:36:40
#> 182 2024-12-18 12:37:51
#> 183 2024-12-18 17:33:59
#> 184 2024-12-18 13:34:11
#> 185 2024-12-18 15:07:21
#> 186 2024-12-18 12:33:01
#> 187 2024-12-18 17:34:24
#> 188 2024-12-18 17:34:44
#> 189 2024-12-18 17:34:44
#> 190 2024-12-18 17:34:44
#> 191 2024-12-18 17:32:41
#> 192 2024-12-18 12:34:04
#> 193 2024-12-18 17:34:45
#> 194 2024-12-18 13:11:41
#> 195 2024-12-18 17:34:46
#> 196 2024-12-18 17:13:55
#> 197 2024-12-18 12:15:38
#> 198 2024-12-18 16:37:45
#> 199 2024-12-18 17:24:38
#> 200 2024-12-18 15:18:26
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 2    0.960382   1643.565 2024-12-18 17:34:45        Bid        NA         NA
#> 3    0.960446   1616.468 2024-12-18 17:31:03        Bid        NA         NA
#> 4    0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 5    0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 6    0.961313  12680.200 2024-12-18 17:32:51        Bid        NA         NA
#> 7    0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 8    0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 9    0.962967   1771.650 2024-12-18 17:33:14        Bid        NA         NA
#> 10   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 11   0.963306   3616.244 2024-12-18 17:33:48        Bid        NA         NA
#> 12   0.963348   3094.272 2024-12-18 17:33:16        Bid        NA         NA
#> 13   0.963454   6106.993 2024-12-18 17:34:51        Bid        NA         NA
#> 14   0.963970  12454.707 2024-12-18 17:33:25        Bid        NA         NA
#> 15   0.964109   8185.278 2024-12-18 17:34:04        Bid        NA         NA
#> 16   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 17   0.965000    145.118 2024-12-18 16:58:19        Bid        NA         NA
#> 18   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 19   0.965596  59096.553 2024-12-18 17:27:00        Bid        NA         NA
#> 20   0.965624  15171.846 2024-12-18 17:31:12        Bid        NA         NA
#> 21   0.966241  25256.400 2024-12-18 17:34:31        Bid        NA         NA
#> 22   0.966405   5053.233 2024-12-18 17:31:42        Bid        NA         NA
#> 23   0.967047   8086.404 2024-12-18 17:31:04        Bid        NA         NA
#> 24   0.968493  25124.632 2024-12-18 17:34:59        Bid        NA         NA
#> 25   0.968700    138.916 2024-12-18 17:34:12        Bid        NA         NA
#> 26   0.968724   3138.230 2024-12-18 17:34:52        Bid        NA         NA
#> 27   0.968911  38476.100 2024-12-18 17:34:43        Bid        NA         NA
#> 28   0.969425  15720.934 2024-12-18 17:34:00        Bid        NA         NA
#> 29   0.969468   1691.964 2024-12-18 17:34:25        Bid        NA         NA
#> 30   0.969523  19207.356 2024-12-18 17:34:43        Bid        NA         NA
#> 31   0.970070    154.165 2024-12-18 17:29:38        Bid        NA         NA
#> 32   0.970510   2289.265 2024-12-18 17:34:18        Bid        NA         NA
#> 33   0.970656   3762.044 2024-12-18 17:34:51        Bid        NA         NA
#> 34   0.970767  14751.163 2024-12-18 17:34:03        Bid        NA         NA
#> 35   0.970768 287732.288 2024-12-18 17:35:05        Bid        NA         NA
#> 36   0.970903  15189.107 2024-12-18 17:34:39        Bid        NA         NA
#> 37   0.971089  38248.600 2024-12-18 17:34:51        Bid        NA         NA
#> 38   0.971890   2359.956 2024-12-18 17:34:50        Bid        NA         NA
#> 39   0.972021   2481.837 2024-12-18 17:35:03        Bid        NA         NA
#> 40   0.972363  14724.275 2024-12-18 17:35:00        Bid        NA         NA
#> 41   0.972562     97.523 2024-12-18 17:34:22        Bid        NA         NA
#> 42   0.973222   6419.534 2024-12-18 17:35:03        Bid        NA         NA
#> 43   0.973922 203301.510 2024-12-18 17:34:35        Bid        NA         NA
#> 44   0.973970   5777.526 2024-12-18 17:35:03        Bid        NA         NA
#> 45   0.973971     41.743 2024-12-18 17:34:40        Bid        NA         NA
#> 46   0.973993   3255.604 2024-12-18 17:34:45        Bid        NA         NA
#> 47   0.974090   5135.424 2024-12-18 17:35:03        Bid        NA         NA
#> 48   0.974091   6877.422 2024-12-18 17:35:03        Bid        NA         NA
#> 49   0.974495  14694.722 2024-12-18 17:34:47        Bid        NA         NA
#> 50   0.974496     33.641 2024-12-18 17:34:48        Bid        NA         NA
#> 51   0.974513     49.256 2024-12-18 17:33:21        Bid        NA         NA
#> 52   0.974514     20.530 2024-12-18 17:34:48        Bid        NA         NA
#> 53   0.974542  11586.026 2024-12-18 17:35:03        Bid        NA         NA
#> 54   0.974543    111.829 2024-12-18 17:34:58        Bid        NA         NA
#> 55   0.975196   2567.288 2024-12-18 17:35:03        Bid        NA         NA
#> 56   0.975201     17.544 2024-12-18 17:34:58        Bid        NA         NA
#> 57   0.975319   1925.466 2024-12-18 17:35:03        Bid        NA         NA
#> 58   0.975403   1283.644 2024-12-18 17:35:03        Bid        NA         NA
#> 59   0.975868     43.557 2024-12-18 17:34:39        Bid        NA         NA
#> 60   0.976092  17361.000 2024-12-18 17:34:31        Bid        NA         NA
#> 61   0.976140    641.755 2024-12-18 17:35:03        Bid        NA         NA
#> 62   0.976902   5006.639 2024-12-18 17:34:45        Bid        NA         NA
#> 63   0.977142   9600.921 2024-12-18 17:34:42        Bid        NA         NA
#> 64   0.977172   3942.883 2024-12-18 17:35:03        Bid        NA         NA
#> 65   0.977373  14648.027 2024-12-18 17:35:03        Bid        NA         NA
#> 66   0.977730  16025.100 2024-12-18 17:34:42        Bid        NA         NA
#> 67   0.978000  14640.178 2024-12-18 17:34:57        Bid        NA         NA
#> 68   0.978084     14.454 2024-12-18 17:34:47        Bid        NA         NA
#> 69   0.978122   2755.703 2024-12-18 17:35:02        Bid        NA         NA
#> 70   0.978174  10136.100 2024-12-18 17:34:54        Bid        NA         NA
#> 71   0.978301   5000.000 2024-12-18 17:35:00        Bid        NA         NA
#> 72   0.978379     49.000 2024-12-18 17:35:01        Bid        NA         NA
#> 73   0.978477    504.339 2024-12-18 17:34:53        Bid        NA         NA
#> 74   0.978610   2449.000 2024-12-18 17:34:54        Bid        NA         NA
#> 75   0.978690  14628.320 2024-12-18 17:35:03        Bid        NA         NA
#> 76   0.978747     88.070 2024-12-18 17:34:56        Bid        NA         NA
#> 77   0.978908  10275.560 2024-12-18 17:35:02        Bid        NA         NA
#> 78   0.979094  10681.310 2024-12-18 17:35:02        Bid        NA         NA
#> 79   0.979101   9748.119 2024-12-18 17:35:04        Bid        NA         NA
#> 80   0.979119  10412.510 2024-12-18 17:35:03        Bid        NA         NA
#> 81   0.979130     16.698 2024-12-18 17:35:02        Bid        NA         NA
#> 82   0.979312    490.073 2024-12-18 17:34:45        Bid        NA         NA
#> 83   0.979393   1043.864 2024-12-18 17:35:04        Bid        NA         NA
#> 84   0.979399   9636.200 2024-12-18 17:35:02        Bid        NA         NA
#> 85   0.979408   4876.000 2024-12-18 17:35:04        Bid        NA         NA
#> 86   0.979411   2926.000 2024-12-18 17:35:04        Bid        NA         NA
#> 87   0.979421   3894.759 2024-12-18 17:34:59        Bid        NA         NA
#> 88   0.979588    537.700 2024-12-18 17:35:04        Bid        NA         NA
#> 89   0.979596   2100.000 2024-12-18 17:35:03        Bid        NA         NA
#> 90   0.979677   3576.965 2024-12-18 17:35:03        Bid        NA         NA
#> 91   0.979684    189.994 2024-12-18 17:35:03        Bid        NA         NA
#> 92   0.979686  34098.062 2024-12-18 17:35:04        Bid        NA         NA
#> 93   0.979690   1628.002 2024-12-18 17:35:05        Bid        NA         NA
#> 94   0.979691   5876.592 2024-12-18 17:35:04        Bid        NA         NA
#> 95   0.979692   3576.965 2024-12-18 17:35:05        Bid        NA         NA
#> 96   0.979698     33.799 2024-12-18 17:35:04        Bid        NA         NA
#> 97   0.979700     33.799 2024-12-18 17:35:04        Bid        NA         NA
#> 98   0.979702  12116.092 2024-12-18 17:35:05        Bid        NA         NA
#> 99   0.979704   1375.600 2024-12-18 17:35:05        Bid        NA         NA
#> 100  0.979705    236.242 2024-12-18 17:35:05        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.980653   1324.622
#> 102        NA         NA                <NA>        Ask  0.980654   2098.800
#> 103        NA         NA                <NA>        Ask  0.980696    477.265
#> 104        NA         NA                <NA>        Ask  0.980787   2432.844
#> 105        NA         NA                <NA>        Ask  0.980843   4865.410
#> 106        NA         NA                <NA>        Ask  0.980975   7297.134
#> 107        NA         NA                <NA>        Ask  0.981090   9728.365
#> 108        NA         NA                <NA>        Ask  0.981132   1628.002
#> 109        NA         NA                <NA>        Ask  0.981212  10000.000
#> 110        NA         NA                <NA>        Ask  0.981250   3709.066
#> 111        NA         NA                <NA>        Ask  0.981268   9726.604
#> 112        NA         NA                <NA>        Ask  0.981546   2127.859
#> 113        NA         NA                <NA>        Ask  0.981560    537.700
#> 114        NA         NA                <NA>        Ask  0.981969     24.102
#> 115        NA         NA                <NA>        Ask  0.982000     41.491
#> 116        NA         NA                <NA>        Ask  0.982019   9719.167
#> 117        NA         NA                <NA>        Ask  0.982072     18.034
#> 118        NA         NA                <NA>        Ask  0.982094   2098.800
#> 119        NA         NA                <NA>        Ask  0.982098    138.225
#> 120        NA         NA                <NA>        Ask  0.982146   9635.100
#> 121        NA         NA                <NA>        Ask  0.982175   4839.316
#> 122        NA         NA                <NA>        Ask  0.982398   1459.626
#> 123        NA         NA                <NA>        Ask  0.982399   1125.808
#> 124        NA         NA                <NA>        Ask  0.982400   5000.000
#> 125        NA         NA                <NA>        Ask  0.982419   2755.703
#> 126        NA         NA                <NA>        Ask  0.982480  14571.900
#> 127        NA         NA                <NA>        Ask  0.982518   9559.700
#> 128        NA         NA                <NA>        Ask  0.982532   9441.079
#> 129        NA         NA                <NA>        Ask  0.982566  15299.700
#> 130        NA         NA                <NA>        Ask  0.982852   7904.200
#> 131        NA         NA                <NA>        Ask  0.982878   8391.990
#> 132        NA         NA                <NA>        Ask  0.983138    651.408
#> 133        NA         NA                <NA>        Ask  0.983139   3257.038
#> 134        NA         NA                <NA>        Ask  0.983170  14561.685
#> 135        NA         NA                <NA>        Ask  0.983187    411.495
#> 136        NA         NA                <NA>        Ask  0.983188  15880.171
#> 137        NA         NA                <NA>        Ask  0.983439   5862.668
#> 138        NA         NA                <NA>        Ask  0.983488  15773.184
#> 139        NA         NA                <NA>        Ask  0.983536     99.610
#> 140        NA         NA                <NA>        Ask  0.983583   3908.445
#> 141        NA         NA                <NA>        Ask  0.983611     43.805
#> 142        NA         NA                <NA>        Ask  0.984014    151.962
#> 143        NA         NA                <NA>        Ask  0.984015   4559.853
#> 144        NA         NA                <NA>        Ask  0.984064  31898.756
#> 145        NA         NA                <NA>        Ask  0.984150   5211.260
#> 146        NA         NA                <NA>        Ask  0.984199  39764.400
#> 147        NA         NA                <NA>        Ask  0.985052   5862.668
#> 148        NA         NA                <NA>        Ask  0.985075  18821.704
#> 149        NA         NA                <NA>        Ask  0.985101  24826.063
#> 150        NA         NA                <NA>        Ask  0.985142     10.229
#> 151        NA         NA                <NA>        Ask  0.985264     48.681
#> 152        NA         NA                <NA>        Ask  0.985343   6514.075
#> 153        NA         NA                <NA>        Ask  0.985392  14532.239
#> 154        NA         NA                <NA>        Ask  0.985967  37945.700
#> 155        NA         NA                <NA>        Ask  0.986014     11.586
#> 156        NA         NA                <NA>        Ask  0.986401   4743.783
#> 157        NA         NA                <NA>        Ask  0.986415    137.620
#> 158        NA         NA                <NA>        Ask  0.987168  26581.100
#> 159        NA         NA                <NA>        Ask  0.987926   9220.683
#> 160        NA         NA                <NA>        Ask  0.987979     23.956
#> 161        NA         NA                <NA>        Ask  0.988186  11681.200
#> 162        NA         NA                <NA>        Ask  0.988410   8762.563
#> 163        NA         NA                <NA>        Ask  0.988869    137.097
#> 164        NA         NA                <NA>        Ask  0.988870  14481.133
#> 165        NA         NA                <NA>        Ask  0.989518  29208.541
#> 166        NA         NA                <NA>        Ask  0.989915   2244.454
#> 167        NA         NA                <NA>        Ask  0.990156   3687.953
#> 168        NA         NA                <NA>        Ask  0.990371     39.580
#> 169        NA         NA                <NA>        Ask  0.990372 288387.017
#> 170        NA         NA                <NA>        Ask  0.990375  12732.200
#> 171        NA         NA                <NA>        Ask  0.990562   2435.382
#> 172        NA         NA                <NA>        Ask  0.990614  14455.636
#> 173        NA         NA                <NA>        Ask  0.990750    137.018
#> 174        NA         NA                <NA>        Ask  0.991176   3067.141
#> 175        NA         NA                <NA>        Ask  0.991454     11.529
#> 176        NA         NA                <NA>        Ask  0.992245     31.277
#> 177        NA         NA                <NA>        Ask  0.993136     55.456
#> 178        NA         NA                <NA>        Ask  0.993224     78.674
#> 179        NA         NA                <NA>        Ask  0.993634     48.271
#> 180        NA         NA                <NA>        Ask  0.994000     41.491
#> 181        NA         NA                <NA>        Ask  0.994025     23.810
#> 182        NA         NA                <NA>        Ask  0.994142     98.547
#> 183        NA         NA                <NA>        Ask  0.994601   6487.000
#> 184        NA         NA                <NA>        Ask  0.995104    136.419
#> 185        NA         NA                <NA>        Ask  0.996211     98.300
#> 186        NA         NA                <NA>        Ask  0.996922     11.466
#> 187        NA         NA                <NA>        Ask  0.997340    791.195
#> 188        NA         NA                <NA>        Ask  0.997602   2989.650
#> 189        NA         NA                <NA>        Ask  0.997608    791.118
#> 190        NA         NA                <NA>        Ask  0.998591   5444.061
#> 191        NA         NA                <NA>        Ask  0.998690      7.057
#> 192        NA         NA                <NA>        Ask  0.998818   9459.351
#> 193        NA         NA                <NA>        Ask  0.999478    135.822
#> 194        NA         NA                <NA>        Ask  0.999602   1709.991
#> 195        NA         NA                <NA>        Ask  1.000000   3727.390
#> 196        NA         NA                <NA>        Ask  1.000109     23.665
#> 197        NA         NA                <NA>        Ask  1.000330   6626.100
#> 198        NA         NA                <NA>        Ask  1.001000      8.018
#> 199        NA         NA                <NA>        Ask  1.001270  14298.452
#> 200        NA         NA                <NA>        Ask  1.001581    109.838
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:35:04
#> 102 2024-12-18 17:35:04
#> 103 2024-12-18 17:35:03
#> 104 2024-12-18 17:35:04
#> 105 2024-12-18 17:35:04
#> 106 2024-12-18 17:35:04
#> 107 2024-12-18 17:35:04
#> 108 2024-12-18 17:35:04
#> 109 2024-12-18 17:35:04
#> 110 2024-12-18 17:35:03
#> 111 2024-12-18 17:35:03
#> 112 2024-12-18 17:35:04
#> 113 2024-12-18 17:35:03
#> 114 2024-12-18 16:41:15
#> 115 2024-12-18 16:51:06
#> 116 2024-12-18 17:35:03
#> 117 2024-12-18 17:35:02
#> 118 2024-12-18 17:35:04
#> 119 2024-12-18 17:04:34
#> 120 2024-12-18 17:35:02
#> 121 2024-12-18 17:35:03
#> 122 2024-12-18 17:35:04
#> 123 2024-12-18 17:35:03
#> 124 2024-12-18 17:34:44
#> 125 2024-12-18 17:35:03
#> 126 2024-12-18 17:35:04
#> 127 2024-12-18 17:35:03
#> 128 2024-12-18 17:35:03
#> 129 2024-12-18 17:35:03
#> 130 2024-12-18 17:35:03
#> 131 2024-12-18 17:35:03
#> 132 2024-12-18 17:34:53
#> 133 2024-12-18 17:34:53
#> 134 2024-12-18 17:35:03
#> 135 2024-12-18 17:35:04
#> 136 2024-12-18 17:34:44
#> 137 2024-12-18 17:34:53
#> 138 2024-12-18 17:34:43
#> 139 2024-12-18 17:31:29
#> 140 2024-12-18 17:34:53
#> 141 2024-12-18 17:34:35
#> 142 2024-12-18 17:35:03
#> 143 2024-12-18 17:34:53
#> 144 2024-12-18 17:35:03
#> 145 2024-12-18 17:34:53
#> 146 2024-12-18 17:34:44
#> 147 2024-12-18 17:34:53
#> 148 2024-12-18 17:35:00
#> 149 2024-12-18 17:34:03
#> 150 2024-12-18 17:34:47
#> 151 2024-12-18 14:56:59
#> 152 2024-12-18 17:34:53
#> 153 2024-12-18 17:34:47
#> 154 2024-12-18 17:34:43
#> 155 2024-12-18 15:09:59
#> 156 2024-12-18 17:35:04
#> 157 2024-12-18 16:58:52
#> 158 2024-12-18 17:33:57
#> 159 2024-12-18 17:34:17
#> 160 2024-12-18 14:35:26
#> 161 2024-12-18 17:35:00
#> 162 2024-12-18 17:34:48
#> 163 2024-12-18 17:34:52
#> 164 2024-12-18 17:33:51
#> 165 2024-12-18 17:34:48
#> 166 2024-12-18 17:35:02
#> 167 2024-12-18 17:34:48
#> 168 2024-12-18 17:35:03
#> 169 2024-12-18 17:35:01
#> 170 2024-12-18 17:34:17
#> 171 2024-12-18 17:35:04
#> 172 2024-12-18 17:34:47
#> 173 2024-12-18 13:57:45
#> 174 2024-12-18 17:34:34
#> 175 2024-12-18 12:37:59
#> 176 2024-12-18 15:04:40
#> 177 2024-12-18 13:26:56
#> 178 2024-12-18 15:12:45
#> 179 2024-12-18 12:38:18
#> 180 2024-12-18 12:25:06
#> 181 2024-12-18 13:36:40
#> 182 2024-12-18 12:37:51
#> 183 2024-12-18 17:33:59
#> 184 2024-12-18 13:34:11
#> 185 2024-12-18 15:07:21
#> 186 2024-12-18 12:33:01
#> 187 2024-12-18 17:34:24
#> 188 2024-12-18 17:34:44
#> 189 2024-12-18 17:34:44
#> 190 2024-12-18 17:34:44
#> 191 2024-12-18 12:34:04
#> 192 2024-12-18 17:34:45
#> 193 2024-12-18 13:11:41
#> 194 2024-12-18 17:34:46
#> 195 2024-12-18 17:13:55
#> 196 2024-12-18 12:15:38
#> 197 2024-12-18 16:37:45
#> 198 2024-12-18 17:24:38
#> 199 2024-12-18 17:35:03
#> 200 2024-12-18 15:18:26