Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 17:03:22 1.4004 1.4005
#> 2   2024-12-18 17:03:24 1.4004 1.4011
#> 3   2024-12-18 17:03:24 1.4005 1.4011
#> 4   2024-12-18 17:03:24 1.4006 1.4011
#> 5   2024-12-18 17:03:26 1.4004 1.4011
#> 6   2024-12-18 17:03:26 1.4007 1.4011
#> 7   2024-12-18 17:03:26 1.4005 1.4011
#> 8   2024-12-18 17:03:26 1.4006 1.4011
#> 9   2024-12-18 17:03:26 1.4006 1.4010
#> 10  2024-12-18 17:03:29 1.4006 1.4011
#> 11  2024-12-18 17:03:29 1.4007 1.4011
#> 12  2024-12-18 17:03:30 1.4008 1.4011
#> 13  2024-12-18 17:03:30 1.4003 1.4011
#> 14  2024-12-18 17:03:30 1.4004 1.4011
#> 15  2024-12-18 17:03:30 1.4004 1.4010
#> 16  2024-12-18 17:03:30 1.4005 1.4010
#> 17  2024-12-18 17:03:30 1.4006 1.4010
#> 18  2024-12-18 17:03:30 1.4006 1.4011
#> 19  2024-12-18 17:03:30 1.4007 1.4011
#> 20  2024-12-18 17:03:31 1.4008 1.4011
#> 21  2024-12-18 17:03:31 1.4008 1.4021
#> 22  2024-12-18 17:03:31 1.4009 1.4021
#> 23  2024-12-18 17:03:31 1.4010 1.4021
#> 24  2024-12-18 17:03:32 1.4006 1.4021
#> 25  2024-12-18 17:03:32 1.4007 1.4019
#> 26  2024-12-18 17:03:32 1.4004 1.4019
#> 27  2024-12-18 17:03:32 1.4005 1.4018
#> 28  2024-12-18 17:03:32 1.4006 1.4018
#> 29  2024-12-18 17:03:32 1.4007 1.4018
#> 30  2024-12-18 17:03:32 1.4007 1.4019
#> 31  2024-12-18 17:03:32 1.4007 1.4020
#> 32  2024-12-18 17:03:32 1.4008 1.4020
#> 33  2024-12-18 17:03:32 1.4009 1.4020
#> 34  2024-12-18 17:03:32 1.4006 1.4020
#> 35  2024-12-18 17:03:32 1.4008 1.4020
#> 36  2024-12-18 17:03:32 1.4009 1.4020
#> 37  2024-12-18 17:03:32 1.4010 1.4020
#> 38  2024-12-18 17:03:32 1.4011 1.4020
#> 39  2024-12-18 17:03:33 1.4011 1.4019
#> 40  2024-12-18 17:03:33 1.4007 1.4019
#> 41  2024-12-18 17:03:33 1.4008 1.4019
#> 42  2024-12-18 17:03:34 1.4009 1.4019
#> 43  2024-12-18 17:03:34 1.4010 1.4019
#> 44  2024-12-18 17:03:34 1.4011 1.4019
#> 45  2024-12-18 17:03:34 1.4012 1.4019
#> 46  2024-12-18 17:03:35 1.4013 1.4019
#> 47  2024-12-18 17:03:35 1.4014 1.4019
#> 48  2024-12-18 17:03:36 1.4015 1.4019
#> 49  2024-12-18 17:03:37 1.4016 1.4019
#> 50  2024-12-18 17:03:37 1.4017 1.4019
#> 51  2024-12-18 17:03:38 1.4018 1.4019
#> 52  2024-12-18 17:03:38 1.4018 1.4030
#> 53  2024-12-18 17:03:38 1.4019 1.4030
#> 54  2024-12-18 17:03:38 1.4020 1.4030
#> 55  2024-12-18 17:03:38 1.4021 1.4030
#> 56  2024-12-18 17:03:38 1.4022 1.4030
#> 57  2024-12-18 17:03:38 1.4022 1.4029
#> 58  2024-12-18 17:03:39 1.4023 1.4029
#> 59  2024-12-18 17:03:39 1.4024 1.4029
#> 60  2024-12-18 17:03:40 1.4024 1.4030
#> 61  2024-12-18 17:03:40 1.4025 1.4030
#> 62  2024-12-18 17:03:40 1.4026 1.4030
#> 63  2024-12-18 17:03:41 1.4027 1.4030
#> 64  2024-12-18 17:03:41 1.4028 1.4030
#> 65  2024-12-18 17:03:42 1.4029 1.4030
#> 66  2024-12-18 17:03:42 1.4028 1.4030
#> 67  2024-12-18 17:03:42 1.4029 1.4030
#> 68  2024-12-18 17:03:45 1.4029 1.4043
#> 69  2024-12-18 17:03:45 1.4030 1.4043
#> 70  2024-12-18 17:03:45 1.4030 1.4042
#> 71  2024-12-18 17:03:47 1.4031 1.4042
#> 72  2024-12-18 17:03:47 1.4032 1.4042
#> 73  2024-12-18 17:03:49 1.4033 1.4042
#> 74  2024-12-18 17:03:49 1.4034 1.4042
#> 75  2024-12-18 17:03:49 1.4035 1.4042
#> 76  2024-12-18 17:03:49 1.4036 1.4042
#> 77  2024-12-18 17:03:51 1.4032 1.4042
#> 78  2024-12-18 17:03:51 1.4033 1.4042
#> 79  2024-12-18 17:03:51 1.4037 1.4042
#> 80  2024-12-18 17:03:51 1.4036 1.4042
#> 81  2024-12-18 17:03:51 1.4032 1.4042
#> 82  2024-12-18 17:03:51 1.4033 1.4041
#> 83  2024-12-18 17:03:52 1.4034 1.4041
#> 84  2024-12-18 17:03:52 1.4034 1.4042
#> 85  2024-12-18 17:03:54 1.4035 1.4042
#> 86  2024-12-18 17:03:54 1.4036 1.4042
#> 87  2024-12-18 17:03:54 1.4037 1.4042
#> 88  2024-12-18 17:03:55 1.4038 1.4042
#> 89  2024-12-18 17:03:55 1.4039 1.4042
#> 90  2024-12-18 17:03:55 1.4038 1.4042
#> 91  2024-12-18 17:03:56 1.4038 1.4041
#> 92  2024-12-18 17:03:56 1.4038 1.4040
#> 93  2024-12-18 17:03:57 1.4035 1.4040
#> 94  2024-12-18 17:03:57 1.4036 1.4040
#> 95  2024-12-18 17:03:59 1.4037 1.4040
#> 96  2024-12-18 17:03:59 1.4036 1.4040
#> 97  2024-12-18 17:04:00 1.4036 1.4039
#> 98  2024-12-18 17:04:00 1.4034 1.4039
#> 99  2024-12-18 17:04:00 1.4034 1.4038
#> 100 2024-12-18 17:04:00 1.4029 1.4038
#> 101 2024-12-18 17:04:00 1.4033 1.4038
#> 102 2024-12-18 17:04:00 1.4026 1.4038
#> 103 2024-12-18 17:04:00 1.4024 1.4038
#> 104 2024-12-18 17:04:00 1.4024 1.4037
#> 105 2024-12-18 17:04:00 1.4025 1.4037
#> 106 2024-12-18 17:04:00 1.4026 1.4037
#> 107 2024-12-18 17:04:01 1.4026 1.4036
#> 108 2024-12-18 17:04:01 1.4026 1.4035
#> 109 2024-12-18 17:04:01 1.4026 1.4034
#> 110 2024-12-18 17:04:01 1.4026 1.4033
#> 111 2024-12-18 17:04:01 1.4027 1.4033
#> 112 2024-12-18 17:04:01 1.4028 1.4033
#> 113 2024-12-18 17:04:02 1.4028 1.4036
#> 114 2024-12-18 17:04:02 1.4029 1.4036
#> 115 2024-12-18 17:04:02 1.4030 1.4036
#> 116 2024-12-18 17:04:02 1.4030 1.4037
#> 117 2024-12-18 17:04:02 1.4027 1.4037
#> 118 2024-12-18 17:04:02 1.4029 1.4037
#> 119 2024-12-18 17:04:02 1.4030 1.4037
#> 120 2024-12-18 17:04:04 1.4031 1.4037
#> 121 2024-12-18 17:04:04 1.4032 1.4037
#> 122 2024-12-18 17:04:06 1.4033 1.4037
#> 123 2024-12-18 17:04:06 1.4034 1.4037
#> 124 2024-12-18 17:04:06 1.4035 1.4037
#> 125 2024-12-18 17:04:08 1.4036 1.4037
#> 126 2024-12-18 17:04:13 1.4034 1.4037
#> 127 2024-12-18 17:04:13 1.4031 1.4037
#> 128 2024-12-18 17:04:13 1.4032 1.4037
#> 129 2024-12-18 17:04:13 1.4034 1.4037
#> 130 2024-12-18 17:04:13 1.4032 1.4037
#> 131 2024-12-18 17:04:14 1.4033 1.4037
#> 132 2024-12-18 17:04:15 1.4034 1.4037
#> 133 2024-12-18 17:04:16 1.4035 1.4037
#> 134 2024-12-18 17:04:17 1.4033 1.4037
#> 135 2024-12-18 17:04:22 1.4034 1.4037
#> 136 2024-12-18 17:04:22 1.4035 1.4037
#> 137 2024-12-18 17:04:23 1.4034 1.4037
#> 138 2024-12-18 17:04:24 1.4035 1.4037
#> 139 2024-12-18 17:04:24 1.4034 1.4037
#> 140 2024-12-18 17:04:24 1.4031 1.4037
#> 141 2024-12-18 17:04:24 1.4026 1.4037
#> 142 2024-12-18 17:04:24 1.4027 1.4036
#> 143 2024-12-18 17:04:24 1.4029 1.4036
#> 144 2024-12-18 17:04:24 1.4030 1.4036
#> 145 2024-12-18 17:04:25 1.4029 1.4036
#> 146 2024-12-18 17:04:25 1.4027 1.4036
#> 147 2024-12-18 17:04:25 1.4028 1.4036
#> 148 2024-12-18 17:04:25 1.4028 1.4035
#> 149 2024-12-18 17:04:25 1.4028 1.4034
#> 150 2024-12-18 17:04:25 1.4028 1.4033
#> 151 2024-12-18 17:04:25 1.4027 1.4033
#> 152 2024-12-18 17:04:26 1.4028 1.4033
#> 153 2024-12-18 17:04:26 1.4027 1.4033
#> 154 2024-12-18 17:04:26 1.4027 1.4032
#> 155 2024-12-18 17:04:26 1.4027 1.4031
#> 156 2024-12-18 17:04:27 1.4025 1.4031
#> 157 2024-12-18 17:04:27 1.4018 1.4031
#> 158 2024-12-18 17:04:27 1.4023 1.4031
#> 159 2024-12-18 17:04:27 1.4021 1.4031
#> 160 2024-12-18 17:04:28 1.4019 1.4031
#> 161 2024-12-18 17:04:28 1.4020 1.4030
#> 162 2024-12-18 17:04:28 1.4021 1.4030
#> 163 2024-12-18 17:04:28 1.4021 1.4029
#> 164 2024-12-18 17:04:28 1.4022 1.4029
#> 165 2024-12-18 17:04:28 1.4013 1.4029
#> 166 2024-12-18 17:04:28 1.4013 1.4030
#> 167 2024-12-18 17:04:28 1.4014 1.4030
#> 168 2024-12-18 17:04:28 1.4015 1.4030
#> 169 2024-12-18 17:04:28 1.4015 1.4028
#> 170 2024-12-18 17:04:28 1.4015 1.4027
#> 171 2024-12-18 17:04:28 1.4015 1.4026
#> 172 2024-12-18 17:04:29 1.4015 1.4025
#> 173 2024-12-18 17:04:30 1.4015 1.4024
#> 174 2024-12-18 17:04:30 1.4016 1.4024
#> 175 2024-12-18 17:04:30 1.4017 1.4024
#> 176 2024-12-18 17:04:30 1.4017 1.4023
#> 177 2024-12-18 17:04:30 1.4017 1.4022
#> 178 2024-12-18 17:04:30 1.4017 1.4026
#> 179 2024-12-18 17:04:30 1.4018 1.4026
#> 180 2024-12-18 17:04:30 1.4019 1.4026
#> 181 2024-12-18 17:04:32 1.4020 1.4026
#> 182 2024-12-18 17:04:32 1.4021 1.4026
#> 183 2024-12-18 17:04:32 1.4022 1.4026
#> 184 2024-12-18 17:04:33 1.4022 1.4029
#> 185 2024-12-18 17:04:34 1.4023 1.4029
#> 186 2024-12-18 17:04:34 1.4024 1.4029
#> 187 2024-12-18 17:04:34 1.4025 1.4029
#> 188 2024-12-18 17:04:35 1.4023 1.4029
#> 189 2024-12-18 17:04:35 1.4025 1.4029
#> 190 2024-12-18 17:04:36 1.4026 1.4029
#> 191 2024-12-18 17:04:36 1.4027 1.4029
#> 192 2024-12-18 17:04:38 1.4028 1.4029
#> 193 2024-12-18 17:04:39 1.4026 1.4029
#> 194 2024-12-18 17:04:39 1.4020 1.4029
#> 195 2024-12-18 17:04:39 1.4026 1.4029
#> 196 2024-12-18 17:04:39 1.4021 1.4029
#> 197 2024-12-18 17:04:39 1.4020 1.4029
#> 198 2024-12-18 17:04:39 1.4020 1.4029
#> 199 2024-12-18 17:04:40 1.4022 1.4029
#> 200 2024-12-18 17:04:40 1.4023 1.4029
#> 201 2024-12-18 17:04:41 1.4024 1.4029
#> 202 2024-12-18 17:04:41 1.4025 1.4029
#> 203 2024-12-18 17:04:41 1.4026 1.4029
#> 204 2024-12-18 17:04:42 1.4025 1.4029
#> 205 2024-12-18 17:04:43 1.4021 1.4029
#> 206 2024-12-18 17:04:43 1.4022 1.4028
#> 207 2024-12-18 17:04:43 1.4023 1.4028
#> 208 2024-12-18 17:04:44 1.4020 1.4028
#> 209 2024-12-18 17:04:44 1.4020 1.4027
#> 210 2024-12-18 17:04:44 1.4020 1.4026
#> 211 2024-12-18 17:04:44 1.4018 1.4026
#> 212 2024-12-18 17:04:44 1.4018 1.4025
#> 213 2024-12-18 17:04:44 1.4018 1.4024
#> 214 2024-12-18 17:04:44 1.4013 1.4024
#> 215 2024-12-18 17:04:44 1.4015 1.4024
#> 216 2024-12-18 17:04:44 1.4016 1.4024
#> 217 2024-12-18 17:04:44 1.4016 1.4020
#> 218 2024-12-18 17:04:45 1.4013 1.4020
#> 219 2024-12-18 17:04:46 1.4015 1.4020
#> 220 2024-12-18 17:04:46 1.4016 1.4020
#> 221 2024-12-18 17:04:49 1.4017 1.4020
#> 222 2024-12-18 17:04:49 1.4013 1.4020
#> 223 2024-12-18 17:04:49 1.4017 1.4020
#> 224 2024-12-18 17:04:49 1.4014 1.4020
#> 225 2024-12-18 17:04:49 1.4015 1.4020
#> 226 2024-12-18 17:04:50 1.4016 1.4020
#> 227 2024-12-18 17:04:52 1.4017 1.4020
#> 228 2024-12-18 17:04:52 1.4018 1.4020
#> 229 2024-12-18 17:04:55 1.4015 1.4020
#> 230 2024-12-18 17:04:55 1.4016 1.4020
#> 231 2024-12-18 17:04:55 1.4017 1.4020
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 17:04:40 0.975951 0.976512
#> 2   2024-12-18 17:04:40 0.975951 0.976511
#> 3   2024-12-18 17:04:40 0.975951 0.976510
#> 4   2024-12-18 17:04:40 0.975951 0.976463
#> 5   2024-12-18 17:04:40 0.975951 0.976462
#> 6   2024-12-18 17:04:40 0.975951 0.976463
#> 7   2024-12-18 17:04:40 0.975951 0.976463
#> 8   2024-12-18 17:04:40 0.975951 0.976463
#> 9   2024-12-18 17:04:41 0.975951 0.976461
#> 10  2024-12-18 17:04:41 0.975951 0.976460
#> 11  2024-12-18 17:04:41 0.975951 0.976459
#> 12  2024-12-18 17:04:41 0.975951 0.976511
#> 13  2024-12-18 17:04:41 0.975951 0.976459
#> 14  2024-12-18 17:04:41 0.975951 0.976510
#> 15  2024-12-18 17:04:41 0.975952 0.976510
#> 16  2024-12-18 17:04:41 0.975952 0.976509
#> 17  2024-12-18 17:04:41 0.975953 0.976509
#> 18  2024-12-18 17:04:41 0.975953 0.976508
#> 19  2024-12-18 17:04:41 0.975950 0.976508
#> 20  2024-12-18 17:04:41 0.975951 0.976509
#> 21  2024-12-18 17:04:41 0.975951 0.976509
#> 22  2024-12-18 17:04:41 0.975952 0.976508
#> 23  2024-12-18 17:04:41 0.975952 0.976502
#> 24  2024-12-18 17:04:41 0.975953 0.976502
#> 25  2024-12-18 17:04:41 0.975953 0.976501
#> 26  2024-12-18 17:04:41 0.975953 0.976500
#> 27  2024-12-18 17:04:41 0.975953 0.976499
#> 28  2024-12-18 17:04:41 0.975953 0.976509
#> 29  2024-12-18 17:04:41 0.975985 0.976508
#> 30  2024-12-18 17:04:41 0.975986 0.976508
#> 31  2024-12-18 17:04:41 0.975987 0.976508
#> 32  2024-12-18 17:04:41 0.975986 0.976508
#> 33  2024-12-18 17:04:41 0.975986 0.976508
#> 34  2024-12-18 17:04:41 0.975987 0.976569
#> 35  2024-12-18 17:04:42 0.975988 0.976568
#> 36  2024-12-18 17:04:42 0.975988 0.976567
#> 37  2024-12-18 17:04:42 0.975989 0.976567
#> 38  2024-12-18 17:04:42 0.975989 0.976566
#> 39  2024-12-18 17:04:42 0.975989 0.976565
#> 40  2024-12-18 17:04:42 0.975989 0.976564
#> 41  2024-12-18 17:04:42 0.975990 0.976564
#> 42  2024-12-18 17:04:42 0.975990 0.976563
#> 43  2024-12-18 17:04:42 0.975991 0.976563
#> 44  2024-12-18 17:04:42 0.975991 0.976562
#> 45  2024-12-18 17:04:42 0.975991 0.976561
#> 46  2024-12-18 17:04:42 0.975991 0.976560
#> 47  2024-12-18 17:04:43 0.975849 0.976560
#> 48  2024-12-18 17:04:43 0.975851 0.976560
#> 49  2024-12-18 17:04:43 0.975852 0.976560
#> 50  2024-12-18 17:04:43 0.975849 0.976560
#> 51  2024-12-18 17:04:43 0.975850 0.976558
#> 52  2024-12-18 17:04:43 0.975850 0.976557
#> 53  2024-12-18 17:04:43 0.975849 0.976557
#> 54  2024-12-18 17:04:43 0.975849 0.976556
#> 55  2024-12-18 17:04:43 0.975760 0.976556
#> 56  2024-12-18 17:04:43 0.975760 0.976555
#> 57  2024-12-18 17:04:43 0.975761 0.976555
#> 58  2024-12-18 17:04:44 0.975760 0.976555
#> 59  2024-12-18 17:04:44 0.975760 0.976549
#> 60  2024-12-18 17:04:44 0.975760 0.976548
#> 61  2024-12-18 17:04:44 0.975760 0.976547
#> 62  2024-12-18 17:04:44 0.975760 0.976546
#> 63  2024-12-18 17:04:44 0.975760 0.976493
#> 64  2024-12-18 17:04:44 0.975760 0.976492
#> 65  2024-12-18 17:04:44 0.975760 0.976491
#> 66  2024-12-18 17:04:44 0.975760 0.976490
#> 67  2024-12-18 17:04:44 0.975760 0.976484
#> 68  2024-12-18 17:04:44 0.975760 0.976483
#> 69  2024-12-18 17:04:44 0.975760 0.976491
#> 70  2024-12-18 17:04:44 0.975760 0.976482
#> 71  2024-12-18 17:04:44 0.975760 0.976481
#> 72  2024-12-18 17:04:44 0.975760 0.976480
#> 73  2024-12-18 17:04:44 0.975760 0.976479
#> 74  2024-12-18 17:04:44 0.975760 0.976475
#> 75  2024-12-18 17:04:44 0.975760 0.976474
#> 76  2024-12-18 17:04:44 0.975760 0.976473
#> 77  2024-12-18 17:04:44 0.975760 0.976472
#> 78  2024-12-18 17:04:44 0.975760 0.976471
#> 79  2024-12-18 17:04:44 0.975760 0.976470
#> 80  2024-12-18 17:04:45 0.975760 0.976469
#> 81  2024-12-18 17:04:45 0.975760 0.976468
#> 82  2024-12-18 17:04:45 0.975760 0.976050
#> 83  2024-12-18 17:04:45 0.975465 0.976050
#> 84  2024-12-18 17:04:45 0.975465 0.976049
#> 85  2024-12-18 17:04:45 0.975465 0.976048
#> 86  2024-12-18 17:04:45 0.975466 0.976048
#> 87  2024-12-18 17:04:45 0.975467 0.976048
#> 88  2024-12-18 17:04:45 0.975467 0.976000
#> 89  2024-12-18 17:04:45 0.975467 0.975994
#> 90  2024-12-18 17:04:45 0.975467 0.975993
#> 91  2024-12-18 17:04:45 0.975467 0.975992
#> 92  2024-12-18 17:04:45 0.975466 0.975992
#> 93  2024-12-18 17:04:46 0.975466 0.975944
#> 94  2024-12-18 17:04:46 0.975466 0.975943
#> 95  2024-12-18 17:04:46 0.975466 0.975895
#> 96  2024-12-18 17:04:46 0.975466 0.975894
#> 97  2024-12-18 17:04:46 0.975466 0.975893
#> 98  2024-12-18 17:04:46 0.975466 0.975888
#> 99  2024-12-18 17:04:46 0.975466 0.975887
#> 100 2024-12-18 17:04:46 0.975466 0.975886
#> 101 2024-12-18 17:04:46 0.975466 0.975838
#> 102 2024-12-18 17:04:46 0.975466 0.975837
#> 103 2024-12-18 17:04:46 0.975466 0.975836
#> 104 2024-12-18 17:04:46 0.975466 0.975835
#> 105 2024-12-18 17:04:46 0.975466 0.975834
#> 106 2024-12-18 17:04:47 0.975466 0.975786
#> 107 2024-12-18 17:04:47 0.975466 0.975785
#> 108 2024-12-18 17:04:47 0.975466 0.975765
#> 109 2024-12-18 17:04:47 0.975466 0.975764
#> 110 2024-12-18 17:04:47 0.975466 0.975760
#> 111 2024-12-18 17:04:47 0.975466 0.975759
#> 112 2024-12-18 17:04:47 0.975466 0.975757
#> 113 2024-12-18 17:04:47 0.975466 0.975753
#> 114 2024-12-18 17:04:47 0.975466 0.975752
#> 115 2024-12-18 17:04:47 0.975466 0.975751
#> 116 2024-12-18 17:04:47 0.975466 0.975750
#> 117 2024-12-18 17:04:48 0.975466 0.975749
#> 118 2024-12-18 17:04:48 0.975466 0.975748
#> 119 2024-12-18 17:04:48 0.975466 0.975746
#> 120 2024-12-18 17:04:48 0.975466 0.975745
#> 121 2024-12-18 17:04:48 0.975466 0.975740
#> 122 2024-12-18 17:04:48 0.975466 0.975739
#> 123 2024-12-18 17:04:48 0.975466 0.975726
#> 124 2024-12-18 17:04:48 0.975466 0.975725
#> 125 2024-12-18 17:04:48 0.975466 0.975719
#> 126 2024-12-18 17:04:48 0.975466 0.975718
#> 127 2024-12-18 17:04:48 0.975466 0.975726
#> 128 2024-12-18 17:04:48 0.975466 0.975712
#> 129 2024-12-18 17:04:48 0.975466 0.975711
#> 130 2024-12-18 17:04:49 0.975271 0.975711
#> 131 2024-12-18 17:04:49 0.975271 0.975710
#> 132 2024-12-18 17:04:49 0.975272 0.975710
#> 133 2024-12-18 17:04:49 0.975272 0.975709
#> 134 2024-12-18 17:04:49 0.975273 0.975709
#> 135 2024-12-18 17:04:49 0.975241 0.975709
#> 136 2024-12-18 17:04:49 0.975242 0.975726
#> 137 2024-12-18 17:04:49 0.975242 0.975724
#> 138 2024-12-18 17:04:49 0.975242 0.975722
#> 139 2024-12-18 17:04:49 0.975242 0.975721
#> 140 2024-12-18 17:04:49 0.975242 0.975720
#> 141 2024-12-18 17:04:49 0.975242 0.975719
#> 142 2024-12-18 17:04:49 0.975242 0.975718
#> 143 2024-12-18 17:04:49 0.975242 0.975700
#> 144 2024-12-18 17:04:49 0.975242 0.975699
#> 145 2024-12-18 17:04:49 0.975243 0.975699
#> 146 2024-12-18 17:04:49 0.975244 0.975699
#> 147 2024-12-18 17:04:49 0.975244 0.975693
#> 148 2024-12-18 17:04:49 0.975244 0.975692
#> 149 2024-12-18 17:04:49 0.975240 0.975692
#> 150 2024-12-18 17:04:49 0.975241 0.975700
#> 151 2024-12-18 17:04:49 0.975241 0.975700
#> 152 2024-12-18 17:04:49 0.975241 0.975700
#> 153 2024-12-18 17:04:49 0.975240 0.975699
#> 154 2024-12-18 17:04:49 0.975241 0.975699
#> 155 2024-12-18 17:04:49 0.975242 0.975699
#> 156 2024-12-18 17:04:49 0.975241 0.975699
#> 157 2024-12-18 17:04:49 0.975241 0.975698
#> 158 2024-12-18 17:04:49 0.975241 0.975697
#> 159 2024-12-18 17:04:49 0.975242 0.975697
#> 160 2024-12-18 17:04:49 0.975243 0.975697
#> 161 2024-12-18 17:04:50 0.975243 0.975691
#> 162 2024-12-18 17:04:50 0.975243 0.975690
#> 163 2024-12-18 17:04:50 0.975245 0.975690
#> 164 2024-12-18 17:04:50 0.975246 0.975690
#> 165 2024-12-18 17:04:50 0.975245 0.975690
#> 166 2024-12-18 17:04:50 0.975247 0.975698
#> 167 2024-12-18 17:04:50 0.975245 0.975697
#> 168 2024-12-18 17:04:50 0.975246 0.975691
#> 169 2024-12-18 17:04:50 0.975246 0.975690
#> 170 2024-12-18 17:04:50 0.975246 0.975730
#> 171 2024-12-18 17:04:51 0.975246 0.975728
#> 172 2024-12-18 17:04:51 0.975246 0.975727
#> 173 2024-12-18 17:04:51 0.975246 0.975721
#> 174 2024-12-18 17:04:51 0.975246 0.975720
#> 175 2024-12-18 17:04:51 0.975246 0.975728
#> 176 2024-12-18 17:04:52 0.975246 0.975726
#> 177 2024-12-18 17:04:52 0.975247 0.975726
#> 178 2024-12-18 17:04:52 0.975248 0.975726
#> 179 2024-12-18 17:04:52 0.975248 0.975725
#> 180 2024-12-18 17:04:52 0.975248 0.975724
#> 181 2024-12-18 17:04:52 0.975248 0.975723
#> 182 2024-12-18 17:04:52 0.975248 0.975718
#> 183 2024-12-18 17:04:52 0.975248 0.975717
#> 184 2024-12-18 17:04:52 0.975248 0.975724
#> 185 2024-12-18 17:04:52 0.975248 0.975726
#> 186 2024-12-18 17:04:53 0.975248 0.975724
#> 187 2024-12-18 17:04:53 0.975249 0.975724
#> 188 2024-12-18 17:04:53 0.975249 0.975723
#> 189 2024-12-18 17:04:53 0.975250 0.975723
#> 190 2024-12-18 17:04:53 0.975250 0.975717
#> 191 2024-12-18 17:04:53 0.975250 0.975716
#> 192 2024-12-18 17:04:53 0.975250 0.975715
#> 193 2024-12-18 17:04:53 0.975250 0.975714
#> 194 2024-12-18 17:04:53 0.975250 0.975709
#> 195 2024-12-18 17:04:53 0.975250 0.975708
#> 196 2024-12-18 17:04:53 0.975250 0.975707
#> 197 2024-12-18 17:04:53 0.975250 0.975706
#> 198 2024-12-18 17:04:53 0.975250 0.975705
#> 199 2024-12-18 17:04:53 0.975250 0.975704
#> 200 2024-12-18 17:04:54 0.975250 0.975705
#> 201 2024-12-18 17:04:54 0.975255 0.975705
#> 202 2024-12-18 17:04:54 0.975256 0.975705
#> 203 2024-12-18 17:04:54 0.975257 0.975705
#> 204 2024-12-18 17:04:54 0.975292 0.975705
#> 205 2024-12-18 17:04:54 0.975293 0.975705
#> 206 2024-12-18 17:04:54 0.975294 0.975705
#> 207 2024-12-18 17:04:55 0.975294 0.975723
#> 208 2024-12-18 17:04:55 0.975295 0.975723
#> 209 2024-12-18 17:04:55 0.975296 0.975723
#> 210 2024-12-18 17:04:55 0.975297 0.975723
#> 211 2024-12-18 17:04:55 0.975298 0.975723
#> 212 2024-12-18 17:04:55 0.975298 0.975722
#> 213 2024-12-18 17:04:55 0.975298 0.975721
#> 214 2024-12-18 17:04:56 0.975299 0.975721
#> 215 2024-12-18 17:04:56 0.975300 0.975721
#> 
#> $last
#> [1] 1734541496
#>