Retrieve the Current System Status of the Kraken Exchange

This function fetches the current system status from the Kraken API, including both the operational status and the timestamp.

Usage

getStatus(data = "both")

Arguments

data

A character string specifying the type of data to return. Use "status" for the system status, "timestamp" for the timestamp, or "both" to retrieve both status and timestamp. The default is "both".

Value

A character string if data = "status" or data = "timestamp". A named vector if data = "both".

Examples

getStatus("both")
#>                 status              timestamp 
#>               "online" "2024-12-18T00:06:55Z" 
getStatus("status")
#> [1] "online"
getStatus("timestamp")
#> [1] "2024-12-18T00:06:55Z"