Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.971522  14733.649 2024-12-18 17:49:10        Bid        NA         NA
#> 2    0.972144     48.720 2024-12-18 17:40:34        Bid        NA         NA
#> 3    0.972958    138.308 2024-12-18 17:36:37        Bid        NA         NA
#> 4    0.973000     41.491 2024-12-18 17:46:07        Bid        NA         NA
#> 5    0.973286      4.400 2024-12-18 17:52:42        Bid        NA         NA
#> 6    0.973392   6347.900 2024-12-18 17:49:58        Bid        NA         NA
#> 7    0.973435   6118.811 2024-12-18 17:50:15        Bid        NA         NA
#> 8    0.974513     49.256 2024-12-18 17:33:21        Bid        NA         NA
#> 9    0.975000     41.491 2024-12-18 17:48:06        Bid        NA         NA
#> 10   0.975415   1595.383 2024-12-18 17:50:01        Bid        NA         NA
#> 11   0.975424   5586.617 2024-12-18 17:50:02        Bid        NA         NA
#> 12   0.975670  12933.605 2024-12-18 17:50:01        Bid        NA         NA
#> 13   0.975769     11.609 2024-12-18 17:40:27        Bid        NA         NA
#> 14   0.977000     41.491 2024-12-18 17:52:06        Bid        NA         NA
#> 15   0.977114     24.004 2024-12-18 17:40:48        Bid        NA         NA
#> 16   0.977234    137.703 2024-12-18 17:40:37        Bid        NA         NA
#> 17   0.978017  12680.200 2024-12-18 17:49:39        Bid        NA         NA
#> 18   0.978745   2709.814 2024-12-18 17:52:50        Bid        NA         NA
#> 19   0.979117   1625.108 2024-12-18 17:52:01        Bid        NA         NA
#> 20   0.980106   6106.993 2024-12-18 17:52:41        Bid        NA         NA
#> 21   0.981152     11.546 2024-12-18 17:48:23        Bid        NA         NA
#> 22   0.981529    137.100 2024-12-18 17:48:22        Bid        NA         NA
#> 23   0.981574   2613.657 2024-12-18 17:51:39        Bid        NA         NA
#> 24   0.981586  14337.234 2024-12-18 17:48:49        Bid        NA         NA
#> 25   0.982051   4773.793 2024-12-18 17:48:46        Bid        NA         NA
#> 26   0.982112  12454.707 2024-12-18 17:51:49        Bid        NA         NA
#> 27   0.982532  14569.105 2024-12-18 17:51:47        Bid        NA         NA
#> 28   0.982926  25256.400 2024-12-18 17:52:41        Bid        NA         NA
#> 29   0.982927 287987.715 2024-12-18 17:53:01        Bid        NA         NA
#> 30   0.983583  46359.549 2024-12-18 17:42:08        Bid        NA         NA
#> 31   0.984167  25124.632 2024-12-18 17:51:16        Bid        NA         NA
#> 32   0.984474   3157.475 2024-12-18 17:52:31        Bid        NA         NA
#> 33   0.984715   1886.551 2024-12-18 17:52:00        Bid        NA         NA
#> 34   0.984816  38476.100 2024-12-18 17:51:51        Bid        NA         NA
#> 35   0.984817   2056.646 2024-12-18 17:52:25        Bid        NA         NA
#> 36   0.985271   2062.241 2024-12-18 17:52:24        Bid        NA         NA
#> 37   0.985301   7500.583 2024-12-18 17:53:02        Bid        NA         NA
#> 38   0.985346  14528.398 2024-12-18 17:52:01        Bid        NA         NA
#> 39   0.986075  19207.356 2024-12-18 17:51:49        Bid        NA         NA
#> 40   0.986124   3793.879 2024-12-18 17:53:02        Bid        NA         NA
#> 41   0.986532  15720.934 2024-12-18 17:51:07        Bid        NA         NA
#> 42   0.986581   3414.444 2024-12-18 17:53:02        Bid        NA         NA
#> 43   0.987235 200560.147 2024-12-18 17:51:44        Bid        NA         NA
#> 44   0.987284   3034.745 2024-12-18 17:53:02        Bid        NA         NA
#> 45   0.987389  38248.600 2024-12-18 17:52:41        Bid        NA         NA
#> 46   0.987407   3051.887 2024-12-18 17:52:45        Bid        NA         NA
#> 47   0.987456   2655.402 2024-12-18 17:53:02        Bid        NA         NA
#> 48   0.987457  14497.193 2024-12-18 17:53:01        Bid        NA         NA
#> 49   0.987458   6784.324 2024-12-18 17:53:03        Bid        NA         NA
#> 50   0.987512    143.935 2024-12-18 17:52:28        Bid        NA         NA
#> 51   0.987514     94.863 2024-12-18 17:52:31        Bid        NA         NA
#> 52   0.987922     55.501 2024-12-18 17:52:11        Bid        NA         NA
#> 53   0.988154  15189.107 2024-12-18 17:52:53        Bid        NA         NA
#> 54   0.988203   7964.542 2024-12-18 17:53:02        Bid        NA         NA
#> 55   0.988564     78.861 2024-12-18 17:52:47        Bid        NA         NA
#> 56   0.989675     43.181 2024-12-18 17:52:46        Bid        NA         NA
#> 57   0.989688  14464.513 2024-12-18 17:53:00        Bid        NA         NA
#> 58   0.989983  17103.500 2024-12-18 17:52:59        Bid        NA         NA
#> 59   0.990026   2885.000 2024-12-18 17:51:35        Bid        NA         NA
#> 60   0.990027   5000.000 2024-12-18 17:53:00        Bid        NA         NA
#> 61   0.990028   2316.720 2024-12-18 17:53:03        Bid        NA         NA
#> 62   0.990599  14451.222 2024-12-18 17:53:00        Bid        NA         NA
#> 63   0.991057   2755.703 2024-12-18 17:52:53        Bid        NA         NA
#> 64   0.991180  14442.745 2024-12-18 17:53:02        Bid        NA         NA
#> 65   0.991190   5006.639 2024-12-18 17:53:01        Bid        NA         NA
#> 66   0.991287   9600.921 2024-12-18 17:52:45        Bid        NA         NA
#> 67   0.991319  16025.100 2024-12-18 17:52:02        Bid        NA         NA
#> 68   0.991369     14.454 2024-12-18 17:52:02        Bid        NA         NA
#> 69   0.991418  10136.100 2024-12-18 17:52:02        Bid        NA         NA
#> 70   0.991430    223.959 2024-12-18 17:51:40        Bid        NA         NA
#> 71   0.991466   3942.883 2024-12-18 17:53:01        Bid        NA         NA
#> 72   0.991760   9622.361 2024-12-18 17:53:03        Bid        NA         NA
#> 73   0.991803   9501.900 2024-12-18 17:53:01        Bid        NA         NA
#> 74   0.991822   2405.000 2024-12-18 17:53:01        Bid        NA         NA
#> 75   0.991836     16.698 2024-12-18 17:53:03        Bid        NA         NA
#> 76   0.991867  10470.130 2024-12-18 17:52:01        Bid        NA         NA
#> 77   0.991868      4.510 2024-12-18 17:53:02        Bid        NA         NA
#> 78   0.992120  10959.880 2024-12-18 17:52:01        Bid        NA         NA
#> 79   0.992121  10000.000 2024-12-18 17:53:04        Bid        NA         NA
#> 80   0.992157   4807.000 2024-12-18 17:53:02        Bid        NA         NA
#> 81   0.992188     48.000 2024-12-18 17:53:01        Bid        NA         NA
#> 82   0.992232    615.316 2024-12-18 17:53:02        Bid        NA         NA
#> 83   0.992264  11081.200 2024-12-18 17:52:01        Bid        NA         NA
#> 84   0.992407   9616.085 2024-12-18 17:53:03        Bid        NA         NA
#> 85   0.992507   5876.592 2024-12-18 17:53:03        Bid        NA         NA
#> 86   0.992571   3576.965 2024-12-18 17:53:03        Bid        NA         NA
#> 87   0.992589     33.799 2024-12-18 17:53:03        Bid        NA         NA
#> 88   0.992590   2402.175 2024-12-18 17:53:02        Bid        NA         NA
#> 89   0.992633     33.799 2024-12-18 17:53:03        Bid        NA         NA
#> 90   0.992692   2100.000 2024-12-18 17:53:03        Bid        NA         NA
#> 91   0.992750   9612.766 2024-12-18 17:53:04        Bid        NA         NA
#> 92   0.992751    390.444 2024-12-18 17:53:04        Bid        NA         NA
#> 93   0.992774   7209.404 2024-12-18 17:53:05        Bid        NA         NA
#> 94   0.992804   1628.002 2024-12-18 17:53:03        Bid        NA         NA
#> 95   0.992837   4805.962 2024-12-18 17:53:05        Bid        NA         NA
#> 96   0.992894   2402.843 2024-12-18 17:53:05        Bid        NA         NA
#> 97   0.992895  11580.100 2024-12-18 17:53:05        Bid        NA         NA
#> 98   0.992968    181.325 2024-12-18 17:53:02        Bid        NA         NA
#> 99   0.992992    603.800 2024-12-18 17:53:04        Bid        NA         NA
#> 100  0.992993   1150.831 2024-12-18 17:53:05        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.993224     25.419
#> 102        NA         NA                <NA>        Ask  0.993403     64.569
#> 103        NA         NA                <NA>        Ask  0.993404    197.940
#> 104        NA         NA                <NA>        Ask  0.993432   2401.545
#> 105        NA         NA                <NA>        Ask  0.993489   4802.812
#> 106        NA         NA                <NA>        Ask  0.993592   7203.473
#> 107        NA         NA                <NA>        Ask  0.993634     48.271
#> 108        NA         NA                <NA>        Ask  0.993716   5876.592
#> 109        NA         NA                <NA>        Ask  0.993761   9602.993
#> 110        NA         NA                <NA>        Ask  0.993891   1200.000
#> 111        NA         NA                <NA>        Ask  0.993936   9601.807
#> 112        NA         NA                <NA>        Ask  0.993984    603.800
#> 113        NA         NA                <NA>        Ask  0.994000     41.491
#> 114        NA         NA                <NA>        Ask  0.994025     23.810
#> 115        NA         NA                <NA>        Ask  0.994037  10000.000
#> 116        NA         NA                <NA>        Ask  0.994135    384.082
#> 117        NA         NA                <NA>        Ask  0.994142     98.547
#> 118        NA         NA                <NA>        Ask  0.994416   9597.174
#> 119        NA         NA                <NA>        Ask  0.994720   1063.930
#> 120        NA         NA                <NA>        Ask  0.994807   4807.000
#> 121        NA         NA                <NA>        Ask  0.994816     18.034
#> 122        NA         NA                <NA>        Ask  0.994818   4839.316
#> 123        NA         NA                <NA>        Ask  0.994876   2885.000
#> 124        NA         NA                <NA>        Ask  0.994912  14388.588
#> 125        NA         NA                <NA>        Ask  0.995081   9441.079
#> 126        NA         NA                <NA>        Ask  0.995104    136.419
#> 127        NA         NA                <NA>        Ask  0.995110   9422.200
#> 128        NA         NA                <NA>        Ask  0.995144  15299.700
#> 129        NA         NA                <NA>        Ask  0.995452    456.063
#> 130        NA         NA                <NA>        Ask  0.995453   9500.000
#> 131        NA         NA                <NA>        Ask  0.995560    969.381
#> 132        NA         NA                <NA>        Ask  0.995575   5000.000
#> 133        NA         NA                <NA>        Ask  0.995595   1786.322
#> 134        NA         NA                <NA>        Ask  0.995602  14378.615
#> 135        NA         NA                <NA>        Ask  0.995630   8173.580
#> 136        NA         NA                <NA>        Ask  0.995835   8352.360
#> 137        NA         NA                <NA>        Ask  0.995876   1439.856
#> 138        NA         NA                <NA>        Ask  0.996211     98.300
#> 139        NA         NA                <NA>        Ask  0.996499  14365.672
#> 140        NA         NA                <NA>        Ask  0.996508     48.000
#> 141        NA         NA                <NA>        Ask  0.996517   1344.119
#> 142        NA         NA                <NA>        Ask  0.996566  15880.171
#> 143        NA         NA                <NA>        Ask  0.996922     11.466
#> 144        NA         NA                <NA>        Ask  0.997040   5376.476
#> 145        NA         NA                <NA>        Ask  0.997254  15773.184
#> 146        NA         NA                <NA>        Ask  0.997462     20.000
#> 147        NA         NA                <NA>        Ask  0.997474     43.497
#> 148        NA         NA                <NA>        Ask  0.997541   2688.238
#> 149        NA         NA                <NA>        Ask  0.997590  39764.400
#> 150        NA         NA                <NA>        Ask  0.997616  17102.600
#> 151        NA         NA                <NA>        Ask  0.997792   6720.595
#> 152        NA         NA                <NA>        Ask  0.997841  18821.704
#> 153        NA         NA                <NA>        Ask  0.998524     10.229
#> 154        NA         NA                <NA>        Ask  0.998570   4032.357
#> 155        NA         NA                <NA>        Ask  0.998619  14334.421
#> 156        NA         NA                <NA>        Ask  0.998690      7.057
#> 157        NA         NA                <NA>        Ask  0.998916   4480.397
#> 158        NA         NA                <NA>        Ask  0.998965  24826.063
#> 159        NA         NA                <NA>        Ask  0.999343  37945.700
#> 160        NA         NA                <NA>        Ask  0.999478    135.822
#> 161        NA         NA                <NA>        Ask  1.000000   7459.003
#> 162        NA         NA                <NA>        Ask  1.000109     23.665
#> 163        NA         NA                <NA>        Ask  1.000210   4679.531
#> 164        NA         NA                <NA>        Ask  1.000620  26581.100
#> 165        NA         NA                <NA>        Ask  1.001000      8.018
#> 166        NA         NA                <NA>        Ask  1.001253  14297.599
#> 167        NA         NA                <NA>        Ask  1.001581    109.838
#> 168        NA         NA                <NA>        Ask  1.001920   9220.683
#> 169        NA         NA                <NA>        Ask  1.001970  11520.531
#> 170        NA         NA                <NA>        Ask  1.002077     47.865
#> 171        NA         NA                <NA>        Ask  1.002244     84.063
#> 172        NA         NA                <NA>        Ask  1.002263   1853.522
#> 173        NA         NA                <NA>        Ask  1.002364     21.780
#> 174        NA         NA                <NA>        Ask  1.002422     11.396
#> 175        NA         NA                <NA>        Ask  1.002521     18.217
#> 176        NA         NA                <NA>        Ask  1.002544   2026.711
#> 177        NA         NA                <NA>        Ask  1.002685    158.511
#> 178        NA         NA                <NA>        Ask  1.002686   3071.698
#> 179        NA         NA                <NA>        Ask  1.002870   8640.398
#> 180        NA         NA                <NA>        Ask  1.003297     98.300
#> 181        NA         NA                <NA>        Ask  1.003589     16.825
#> 182        NA         NA                <NA>        Ask  1.003590   7199.711
#> 183        NA         NA                <NA>        Ask  1.003655 289237.848
#> 184        NA         NA                <NA>        Ask  1.003799    152.565
#> 185        NA         NA                <NA>        Ask  1.003800  21594.803
#> 186        NA         NA                <NA>        Ask  1.003871    135.227
#> 187        NA         NA                <NA>        Ask  1.003906    112.589
#> 188        NA         NA                <NA>        Ask  1.004040  12732.200
#> 189        NA         NA                <NA>        Ask  1.004680  14248.822
#> 190        NA         NA                <NA>        Ask  1.004865     98.361
#> 191        NA         NA                <NA>        Ask  1.005411   2596.916
#> 192        NA         NA                <NA>        Ask  1.006230     23.521
#> 193        NA         NA                <NA>        Ask  1.006639    788.953
#> 194        NA         NA                <NA>        Ask  1.006724    682.084
#> 195        NA         NA                <NA>        Ask  1.007105    214.794
#> 196        NA         NA                <NA>        Ask  1.007951     11.334
#> 197        NA         NA                <NA>        Ask  1.008283    134.636
#> 198        NA         NA                <NA>        Ask  1.008559      7.045
#> 199        NA         NA                <NA>        Ask  1.009500   6487.000
#> 200        NA         NA                <NA>        Ask  1.009796     31.277
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:52:45
#> 102 2024-12-18 17:53:05
#> 103 2024-12-18 17:53:00
#> 104 2024-12-18 17:53:05
#> 105 2024-12-18 17:53:05
#> 106 2024-12-18 17:53:03
#> 107 2024-12-18 17:50:26
#> 108 2024-12-18 17:53:03
#> 109 2024-12-18 17:53:03
#> 110 2024-12-18 17:53:04
#> 111 2024-12-18 17:53:02
#> 112 2024-12-18 17:53:03
#> 113 2024-12-18 17:50:23
#> 114 2024-12-18 13:36:40
#> 115 2024-12-18 17:53:05
#> 116 2024-12-18 17:52:59
#> 117 2024-12-18 17:52:52
#> 118 2024-12-18 17:53:00
#> 119 2024-12-18 17:52:55
#> 120 2024-12-18 17:52:55
#> 121 2024-12-18 17:53:03
#> 122 2024-12-18 17:53:01
#> 123 2024-12-18 17:52:55
#> 124 2024-12-18 17:53:01
#> 125 2024-12-18 17:53:00
#> 126 2024-12-18 17:52:30
#> 127 2024-12-18 17:53:00
#> 128 2024-12-18 17:53:00
#> 129 2024-12-18 17:53:05
#> 130 2024-12-18 17:52:47
#> 131 2024-12-18 17:52:15
#> 132 2024-12-18 17:52:10
#> 133 2024-12-18 17:52:15
#> 134 2024-12-18 17:52:40
#> 135 2024-12-18 17:52:45
#> 136 2024-12-18 17:51:47
#> 137 2024-12-18 17:52:51
#> 138 2024-12-18 15:07:21
#> 139 2024-12-18 17:53:00
#> 140 2024-12-18 17:53:01
#> 141 2024-12-18 17:52:54
#> 142 2024-12-18 17:52:01
#> 143 2024-12-18 12:33:01
#> 144 2024-12-18 17:52:54
#> 145 2024-12-18 17:52:02
#> 146 2024-12-18 17:40:55
#> 147 2024-12-18 17:52:42
#> 148 2024-12-18 17:52:54
#> 149 2024-12-18 17:52:01
#> 150 2024-12-18 17:52:30
#> 151 2024-12-18 17:52:54
#> 152 2024-12-18 17:51:43
#> 153 2024-12-18 17:52:02
#> 154 2024-12-18 17:52:54
#> 155 2024-12-18 17:51:47
#> 156 2024-12-18 12:34:04
#> 157 2024-12-18 17:52:54
#> 158 2024-12-18 17:48:56
#> 159 2024-12-18 17:51:47
#> 160 2024-12-18 13:11:41
#> 161 2024-12-18 17:41:34
#> 162 2024-12-18 12:15:38
#> 163 2024-12-18 17:52:30
#> 164 2024-12-18 17:48:12
#> 165 2024-12-18 17:24:38
#> 166 2024-12-18 17:52:19
#> 167 2024-12-18 15:18:26
#> 168 2024-12-18 17:49:54
#> 169 2024-12-18 17:52:10
#> 170 2024-12-18 04:26:55
#> 171 2024-12-18 07:39:44
#> 172 2024-12-18 17:52:02
#> 173 2024-12-18 12:02:37
#> 174 2024-12-18 03:31:33
#> 175 2024-12-18 12:21:12
#> 176 2024-12-18 17:52:03
#> 177 2024-12-18 17:52:24
#> 178 2024-12-18 17:51:53
#> 179 2024-12-18 17:52:10
#> 180 2024-12-18 05:52:42
#> 181 2024-12-18 17:52:48
#> 182 2024-12-18 17:52:11
#> 183 2024-12-18 17:53:03
#> 184 2024-12-18 17:52:48
#> 185 2024-12-18 17:52:32
#> 186 2024-12-18 07:01:36
#> 187 2024-12-18 05:50:53
#> 188 2024-12-18 17:49:19
#> 189 2024-12-18 17:52:13
#> 190 2024-12-18 04:33:06
#> 191 2024-12-18 17:52:51
#> 192 2024-12-18 09:02:51
#> 193 2024-12-18 17:51:02
#> 194 2024-12-18 06:31:37
#> 195 2024-12-18 12:01:30
#> 196 2024-12-18 02:55:05
#> 197 2024-12-18 05:32:05
#> 198 2024-12-18 14:36:34
#> 199 2024-12-18 17:49:55
#> 200 2024-12-18 16:53:46
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.971170     24.151 2024-12-18 17:35:45        Bid        NA         NA
#> 2    0.971522  14733.649 2024-12-18 17:49:10        Bid        NA         NA
#> 3    0.972144     48.720 2024-12-18 17:40:34        Bid        NA         NA
#> 4    0.972958    138.308 2024-12-18 17:36:37        Bid        NA         NA
#> 5    0.973000     41.491 2024-12-18 17:46:07        Bid        NA         NA
#> 6    0.973286      4.400 2024-12-18 17:52:42        Bid        NA         NA
#> 7    0.973392   6347.900 2024-12-18 17:49:58        Bid        NA         NA
#> 8    0.973435   6118.811 2024-12-18 17:50:15        Bid        NA         NA
#> 9    0.974513     49.256 2024-12-18 17:33:21        Bid        NA         NA
#> 10   0.975000     41.491 2024-12-18 17:48:06        Bid        NA         NA
#> 11   0.975415   1595.383 2024-12-18 17:50:01        Bid        NA         NA
#> 12   0.975424   5586.617 2024-12-18 17:50:02        Bid        NA         NA
#> 13   0.975670  12933.605 2024-12-18 17:50:01        Bid        NA         NA
#> 14   0.975769     11.609 2024-12-18 17:40:27        Bid        NA         NA
#> 15   0.977000     41.491 2024-12-18 17:52:06        Bid        NA         NA
#> 16   0.977114     24.004 2024-12-18 17:40:48        Bid        NA         NA
#> 17   0.977234    137.703 2024-12-18 17:40:37        Bid        NA         NA
#> 18   0.978017  12680.200 2024-12-18 17:49:39        Bid        NA         NA
#> 19   0.978745   2709.814 2024-12-18 17:52:50        Bid        NA         NA
#> 20   0.979117   1625.108 2024-12-18 17:52:01        Bid        NA         NA
#> 21   0.980106   6106.993 2024-12-18 17:52:41        Bid        NA         NA
#> 22   0.981152     11.546 2024-12-18 17:48:23        Bid        NA         NA
#> 23   0.981529    137.100 2024-12-18 17:48:22        Bid        NA         NA
#> 24   0.981574   2613.657 2024-12-18 17:51:39        Bid        NA         NA
#> 25   0.981586  14337.234 2024-12-18 17:48:49        Bid        NA         NA
#> 26   0.982051   4773.793 2024-12-18 17:48:46        Bid        NA         NA
#> 27   0.982112  12454.707 2024-12-18 17:51:49        Bid        NA         NA
#> 28   0.982532  14569.105 2024-12-18 17:51:47        Bid        NA         NA
#> 29   0.982926  25256.400 2024-12-18 17:52:41        Bid        NA         NA
#> 30   0.982927 287987.715 2024-12-18 17:53:01        Bid        NA         NA
#> 31   0.983583  46359.549 2024-12-18 17:42:08        Bid        NA         NA
#> 32   0.984167  25124.632 2024-12-18 17:51:16        Bid        NA         NA
#> 33   0.984474   3157.475 2024-12-18 17:52:31        Bid        NA         NA
#> 34   0.984715   1886.551 2024-12-18 17:52:00        Bid        NA         NA
#> 35   0.984816  38476.100 2024-12-18 17:51:51        Bid        NA         NA
#> 36   0.984817   2056.646 2024-12-18 17:52:25        Bid        NA         NA
#> 37   0.985271   2062.241 2024-12-18 17:52:24        Bid        NA         NA
#> 38   0.985301   7500.583 2024-12-18 17:53:02        Bid        NA         NA
#> 39   0.985346  14528.398 2024-12-18 17:52:01        Bid        NA         NA
#> 40   0.986075  19207.356 2024-12-18 17:51:49        Bid        NA         NA
#> 41   0.986124   3793.879 2024-12-18 17:53:02        Bid        NA         NA
#> 42   0.986532  15720.934 2024-12-18 17:51:07        Bid        NA         NA
#> 43   0.986581   3414.444 2024-12-18 17:53:02        Bid        NA         NA
#> 44   0.987235 200560.147 2024-12-18 17:51:44        Bid        NA         NA
#> 45   0.987284   3034.745 2024-12-18 17:53:02        Bid        NA         NA
#> 46   0.987389  38248.600 2024-12-18 17:52:41        Bid        NA         NA
#> 47   0.987407   3051.887 2024-12-18 17:52:45        Bid        NA         NA
#> 48   0.987456   2655.402 2024-12-18 17:53:02        Bid        NA         NA
#> 49   0.987457  14497.193 2024-12-18 17:53:01        Bid        NA         NA
#> 50   0.987458   6784.324 2024-12-18 17:53:03        Bid        NA         NA
#> 51   0.987512    143.935 2024-12-18 17:52:28        Bid        NA         NA
#> 52   0.987514     94.863 2024-12-18 17:52:31        Bid        NA         NA
#> 53   0.987922     55.501 2024-12-18 17:52:11        Bid        NA         NA
#> 54   0.988154  15189.107 2024-12-18 17:52:53        Bid        NA         NA
#> 55   0.988203   7964.542 2024-12-18 17:53:02        Bid        NA         NA
#> 56   0.988564     78.861 2024-12-18 17:52:47        Bid        NA         NA
#> 57   0.989675     43.181 2024-12-18 17:52:46        Bid        NA         NA
#> 58   0.989688  14464.513 2024-12-18 17:53:00        Bid        NA         NA
#> 59   0.989983  17103.500 2024-12-18 17:52:59        Bid        NA         NA
#> 60   0.990026   2885.000 2024-12-18 17:51:35        Bid        NA         NA
#> 61   0.990027   5000.000 2024-12-18 17:53:00        Bid        NA         NA
#> 62   0.990028   2316.720 2024-12-18 17:53:03        Bid        NA         NA
#> 63   0.990599  14451.222 2024-12-18 17:53:00        Bid        NA         NA
#> 64   0.991057   2755.703 2024-12-18 17:52:53        Bid        NA         NA
#> 65   0.991180  14442.745 2024-12-18 17:53:02        Bid        NA         NA
#> 66   0.991190   5006.639 2024-12-18 17:53:01        Bid        NA         NA
#> 67   0.991287   9600.921 2024-12-18 17:52:45        Bid        NA         NA
#> 68   0.991319  16025.100 2024-12-18 17:52:02        Bid        NA         NA
#> 69   0.991369     14.454 2024-12-18 17:52:02        Bid        NA         NA
#> 70   0.991418  10136.100 2024-12-18 17:52:02        Bid        NA         NA
#> 71   0.991430    223.959 2024-12-18 17:51:40        Bid        NA         NA
#> 72   0.991466   3942.883 2024-12-18 17:53:01        Bid        NA         NA
#> 73   0.991760   9622.361 2024-12-18 17:53:03        Bid        NA         NA
#> 74   0.991803   9501.900 2024-12-18 17:53:01        Bid        NA         NA
#> 75   0.991822   2405.000 2024-12-18 17:53:01        Bid        NA         NA
#> 76   0.991836     16.698 2024-12-18 17:53:03        Bid        NA         NA
#> 77   0.991867  10470.130 2024-12-18 17:52:01        Bid        NA         NA
#> 78   0.991868      4.510 2024-12-18 17:53:02        Bid        NA         NA
#> 79   0.992120  10959.880 2024-12-18 17:52:01        Bid        NA         NA
#> 80   0.992121  10000.000 2024-12-18 17:53:04        Bid        NA         NA
#> 81   0.992157   4807.000 2024-12-18 17:53:02        Bid        NA         NA
#> 82   0.992188     48.000 2024-12-18 17:53:01        Bid        NA         NA
#> 83   0.992232    615.316 2024-12-18 17:53:02        Bid        NA         NA
#> 84   0.992264  11081.200 2024-12-18 17:52:01        Bid        NA         NA
#> 85   0.992407   9616.085 2024-12-18 17:53:03        Bid        NA         NA
#> 86   0.992507   5876.592 2024-12-18 17:53:03        Bid        NA         NA
#> 87   0.992571   3576.965 2024-12-18 17:53:03        Bid        NA         NA
#> 88   0.992589     33.799 2024-12-18 17:53:03        Bid        NA         NA
#> 89   0.992590   2402.175 2024-12-18 17:53:02        Bid        NA         NA
#> 90   0.992633     33.799 2024-12-18 17:53:03        Bid        NA         NA
#> 91   0.992692   2100.000 2024-12-18 17:53:03        Bid        NA         NA
#> 92   0.992750   9612.766 2024-12-18 17:53:04        Bid        NA         NA
#> 93   0.992751    390.444 2024-12-18 17:53:04        Bid        NA         NA
#> 94   0.992774   7209.404 2024-12-18 17:53:05        Bid        NA         NA
#> 95   0.992804   1628.002 2024-12-18 17:53:03        Bid        NA         NA
#> 96   0.992895  11580.100 2024-12-18 17:53:05        Bid        NA         NA
#> 97   0.992968    181.325 2024-12-18 17:53:02        Bid        NA         NA
#> 98   0.992992    603.800 2024-12-18 17:53:04        Bid        NA         NA
#> 99   0.992999   2100.000 2024-12-18 17:53:05        Bid        NA         NA
#> 100  0.993000   1150.822 2024-12-18 17:53:05        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.993224     25.419
#> 102        NA         NA                <NA>        Ask  0.993404    197.940
#> 103        NA         NA                <NA>        Ask  0.993591     64.580
#> 104        NA         NA                <NA>        Ask  0.993592   7203.473
#> 105        NA         NA                <NA>        Ask  0.993634     48.271
#> 106        NA         NA                <NA>        Ask  0.993716   5876.592
#> 107        NA         NA                <NA>        Ask  0.993761   9602.993
#> 108        NA         NA                <NA>        Ask  0.993891   1200.000
#> 109        NA         NA                <NA>        Ask  0.993936   9601.807
#> 110        NA         NA                <NA>        Ask  0.993984    603.800
#> 111        NA         NA                <NA>        Ask  0.994000     41.491
#> 112        NA         NA                <NA>        Ask  0.994025     23.810
#> 113        NA         NA                <NA>        Ask  0.994037  10000.000
#> 114        NA         NA                <NA>        Ask  0.994135    384.082
#> 115        NA         NA                <NA>        Ask  0.994142     98.547
#> 116        NA         NA                <NA>        Ask  0.994416   9597.174
#> 117        NA         NA                <NA>        Ask  0.994720   1063.930
#> 118        NA         NA                <NA>        Ask  0.994807   4807.000
#> 119        NA         NA                <NA>        Ask  0.994816     18.034
#> 120        NA         NA                <NA>        Ask  0.994818   4839.316
#> 121        NA         NA                <NA>        Ask  0.994876   2885.000
#> 122        NA         NA                <NA>        Ask  0.994912  14388.588
#> 123        NA         NA                <NA>        Ask  0.995081   9441.079
#> 124        NA         NA                <NA>        Ask  0.995104    136.419
#> 125        NA         NA                <NA>        Ask  0.995110   9422.200
#> 126        NA         NA                <NA>        Ask  0.995144  15299.700
#> 127        NA         NA                <NA>        Ask  0.995453   9500.000
#> 128        NA         NA                <NA>        Ask  0.995560    969.381
#> 129        NA         NA                <NA>        Ask  0.995575   5000.000
#> 130        NA         NA                <NA>        Ask  0.995595   1786.322
#> 131        NA         NA                <NA>        Ask  0.995602  14378.615
#> 132        NA         NA                <NA>        Ask  0.995630   8173.580
#> 133        NA         NA                <NA>        Ask  0.995835   8352.360
#> 134        NA         NA                <NA>        Ask  0.995876   1439.856
#> 135        NA         NA                <NA>        Ask  0.996211     98.300
#> 136        NA         NA                <NA>        Ask  0.996499  14365.672
#> 137        NA         NA                <NA>        Ask  0.996508     48.000
#> 138        NA         NA                <NA>        Ask  0.996517   1344.119
#> 139        NA         NA                <NA>        Ask  0.996566  15880.171
#> 140        NA         NA                <NA>        Ask  0.996921    460.289
#> 141        NA         NA                <NA>        Ask  0.996922     11.466
#> 142        NA         NA                <NA>        Ask  0.997040   5376.476
#> 143        NA         NA                <NA>        Ask  0.997254  15773.184
#> 144        NA         NA                <NA>        Ask  0.997462     20.000
#> 145        NA         NA                <NA>        Ask  0.997474     43.497
#> 146        NA         NA                <NA>        Ask  0.997541   2688.238
#> 147        NA         NA                <NA>        Ask  0.997590  39764.400
#> 148        NA         NA                <NA>        Ask  0.997616  17102.600
#> 149        NA         NA                <NA>        Ask  0.997792   6720.595
#> 150        NA         NA                <NA>        Ask  0.997841  18821.704
#> 151        NA         NA                <NA>        Ask  0.998524     10.229
#> 152        NA         NA                <NA>        Ask  0.998570   4032.357
#> 153        NA         NA                <NA>        Ask  0.998619  14334.421
#> 154        NA         NA                <NA>        Ask  0.998690      7.057
#> 155        NA         NA                <NA>        Ask  0.998916   4480.397
#> 156        NA         NA                <NA>        Ask  0.998965  24826.063
#> 157        NA         NA                <NA>        Ask  0.999343  37945.700
#> 158        NA         NA                <NA>        Ask  0.999478    135.822
#> 159        NA         NA                <NA>        Ask  1.000000   7459.003
#> 160        NA         NA                <NA>        Ask  1.000109     23.665
#> 161        NA         NA                <NA>        Ask  1.000210   4679.531
#> 162        NA         NA                <NA>        Ask  1.000620  26581.100
#> 163        NA         NA                <NA>        Ask  1.001000      8.018
#> 164        NA         NA                <NA>        Ask  1.001253  14297.599
#> 165        NA         NA                <NA>        Ask  1.001581    109.838
#> 166        NA         NA                <NA>        Ask  1.001920   9220.683
#> 167        NA         NA                <NA>        Ask  1.001970  11520.531
#> 168        NA         NA                <NA>        Ask  1.002077     47.865
#> 169        NA         NA                <NA>        Ask  1.002244     84.063
#> 170        NA         NA                <NA>        Ask  1.002263   1853.522
#> 171        NA         NA                <NA>        Ask  1.002364     21.780
#> 172        NA         NA                <NA>        Ask  1.002422     11.396
#> 173        NA         NA                <NA>        Ask  1.002521     18.217
#> 174        NA         NA                <NA>        Ask  1.002544   2026.711
#> 175        NA         NA                <NA>        Ask  1.002685    158.511
#> 176        NA         NA                <NA>        Ask  1.002686   3071.698
#> 177        NA         NA                <NA>        Ask  1.002870   8640.398
#> 178        NA         NA                <NA>        Ask  1.003297     98.300
#> 179        NA         NA                <NA>        Ask  1.003589     16.825
#> 180        NA         NA                <NA>        Ask  1.003590   7199.711
#> 181        NA         NA                <NA>        Ask  1.003655 289237.848
#> 182        NA         NA                <NA>        Ask  1.003799    152.565
#> 183        NA         NA                <NA>        Ask  1.003800  21594.803
#> 184        NA         NA                <NA>        Ask  1.003871    135.227
#> 185        NA         NA                <NA>        Ask  1.003906    112.589
#> 186        NA         NA                <NA>        Ask  1.004040  12732.200
#> 187        NA         NA                <NA>        Ask  1.004680  14248.822
#> 188        NA         NA                <NA>        Ask  1.004865     98.361
#> 189        NA         NA                <NA>        Ask  1.005411   2596.916
#> 190        NA         NA                <NA>        Ask  1.006230     23.521
#> 191        NA         NA                <NA>        Ask  1.006639    788.953
#> 192        NA         NA                <NA>        Ask  1.006724    682.084
#> 193        NA         NA                <NA>        Ask  1.007105    214.794
#> 194        NA         NA                <NA>        Ask  1.007951     11.334
#> 195        NA         NA                <NA>        Ask  1.008283    134.636
#> 196        NA         NA                <NA>        Ask  1.008559      7.045
#> 197        NA         NA                <NA>        Ask  1.009500   6487.000
#> 198        NA         NA                <NA>        Ask  1.009796     31.277
#> 199        NA         NA                <NA>        Ask  1.010000  41000.000
#> 200        NA         NA                <NA>        Ask  1.010433     98.300
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:52:45
#> 102 2024-12-18 17:53:00
#> 103 2024-12-18 17:53:05
#> 104 2024-12-18 17:53:03
#> 105 2024-12-18 17:50:26
#> 106 2024-12-18 17:53:03
#> 107 2024-12-18 17:53:03
#> 108 2024-12-18 17:53:04
#> 109 2024-12-18 17:53:02
#> 110 2024-12-18 17:53:03
#> 111 2024-12-18 17:50:23
#> 112 2024-12-18 13:36:40
#> 113 2024-12-18 17:53:05
#> 114 2024-12-18 17:52:59
#> 115 2024-12-18 17:52:52
#> 116 2024-12-18 17:53:00
#> 117 2024-12-18 17:52:55
#> 118 2024-12-18 17:52:55
#> 119 2024-12-18 17:53:03
#> 120 2024-12-18 17:53:01
#> 121 2024-12-18 17:52:55
#> 122 2024-12-18 17:53:01
#> 123 2024-12-18 17:53:00
#> 124 2024-12-18 17:52:30
#> 125 2024-12-18 17:53:00
#> 126 2024-12-18 17:53:00
#> 127 2024-12-18 17:52:47
#> 128 2024-12-18 17:52:15
#> 129 2024-12-18 17:52:10
#> 130 2024-12-18 17:52:15
#> 131 2024-12-18 17:52:40
#> 132 2024-12-18 17:52:45
#> 133 2024-12-18 17:51:47
#> 134 2024-12-18 17:52:51
#> 135 2024-12-18 15:07:21
#> 136 2024-12-18 17:53:00
#> 137 2024-12-18 17:53:01
#> 138 2024-12-18 17:52:54
#> 139 2024-12-18 17:52:01
#> 140 2024-12-18 17:53:05
#> 141 2024-12-18 12:33:01
#> 142 2024-12-18 17:52:54
#> 143 2024-12-18 17:52:02
#> 144 2024-12-18 17:40:55
#> 145 2024-12-18 17:52:42
#> 146 2024-12-18 17:52:54
#> 147 2024-12-18 17:52:01
#> 148 2024-12-18 17:52:30
#> 149 2024-12-18 17:52:54
#> 150 2024-12-18 17:51:43
#> 151 2024-12-18 17:52:02
#> 152 2024-12-18 17:52:54
#> 153 2024-12-18 17:51:47
#> 154 2024-12-18 12:34:04
#> 155 2024-12-18 17:52:54
#> 156 2024-12-18 17:48:56
#> 157 2024-12-18 17:51:47
#> 158 2024-12-18 13:11:41
#> 159 2024-12-18 17:41:34
#> 160 2024-12-18 12:15:38
#> 161 2024-12-18 17:52:30
#> 162 2024-12-18 17:48:12
#> 163 2024-12-18 17:24:38
#> 164 2024-12-18 17:52:19
#> 165 2024-12-18 15:18:26
#> 166 2024-12-18 17:49:54
#> 167 2024-12-18 17:52:10
#> 168 2024-12-18 04:26:55
#> 169 2024-12-18 07:39:44
#> 170 2024-12-18 17:52:02
#> 171 2024-12-18 12:02:37
#> 172 2024-12-18 03:31:33
#> 173 2024-12-18 12:21:12
#> 174 2024-12-18 17:52:03
#> 175 2024-12-18 17:52:24
#> 176 2024-12-18 17:51:53
#> 177 2024-12-18 17:52:10
#> 178 2024-12-18 05:52:42
#> 179 2024-12-18 17:52:48
#> 180 2024-12-18 17:52:11
#> 181 2024-12-18 17:53:03
#> 182 2024-12-18 17:52:48
#> 183 2024-12-18 17:52:32
#> 184 2024-12-18 07:01:36
#> 185 2024-12-18 05:50:53
#> 186 2024-12-18 17:49:19
#> 187 2024-12-18 17:52:13
#> 188 2024-12-18 04:33:06
#> 189 2024-12-18 17:52:51
#> 190 2024-12-18 09:02:51
#> 191 2024-12-18 17:51:02
#> 192 2024-12-18 06:31:37
#> 193 2024-12-18 12:01:30
#> 194 2024-12-18 02:55:05
#> 195 2024-12-18 05:32:05
#> 196 2024-12-18 14:36:34
#> 197 2024-12-18 17:49:55
#> 198 2024-12-18 16:53:46
#> 199 2024-12-18 15:53:30
#> 200 2024-12-18 14:38:21