Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 2    0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 3    0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 4    0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 5    0.960000  63095.795 2024-12-18 17:37:35        Bid        NA         NA
#> 6    0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 7    0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 8    0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 9    0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 10   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 11   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 12   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 13   0.962781   6457.100 2024-12-18 17:42:56        Bid        NA         NA
#> 14   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 15   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 16   0.965000    145.118 2024-12-18 16:58:19        Bid        NA         NA
#> 17   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 18   0.965417   1643.570 2024-12-18 17:46:02        Bid        NA         NA
#> 19   0.965419   7420.457 2024-12-18 17:46:01        Bid        NA         NA
#> 20   0.965420   1610.906 2024-12-18 17:46:02        Bid        NA         NA
#> 21   0.965430  11273.578 2024-12-18 17:46:03        Bid        NA         NA
#> 22   0.967000     41.491 2024-12-18 17:36:07        Bid        NA         NA
#> 23   0.968252     99.690 2024-12-18 17:37:33        Bid        NA         NA
#> 24   0.968500   6118.811 2024-12-18 17:41:06        Bid        NA         NA
#> 25   0.968544   6347.900 2024-12-18 17:41:13        Bid        NA         NA
#> 26   0.968700    138.916 2024-12-18 17:34:12        Bid        NA         NA
#> 27   0.968846  14776.956 2024-12-18 17:45:30        Bid        NA         NA
#> 28   0.969000     41.491 2024-12-18 17:38:06        Bid        NA         NA
#> 29   0.970070    154.165 2024-12-18 17:29:38        Bid        NA         NA
#> 30   0.971000     41.491 2024-12-18 17:41:06        Bid        NA         NA
#> 31   0.971170     24.151 2024-12-18 17:35:45        Bid        NA         NA
#> 32   0.972144     48.720 2024-12-18 17:40:34        Bid        NA         NA
#> 33   0.972958    138.308 2024-12-18 17:36:37        Bid        NA         NA
#> 34   0.973000     41.491 2024-12-18 17:46:07        Bid        NA         NA
#> 35   0.973422   6106.993 2024-12-18 17:41:06        Bid        NA         NA
#> 36   0.973808   3262.606 2024-12-18 17:45:25        Bid        NA         NA
#> 37   0.974513     49.256 2024-12-18 17:33:21        Bid        NA         NA
#> 38   0.974677  12680.200 2024-12-18 17:41:43        Bid        NA         NA
#> 39   0.975769     11.609 2024-12-18 17:40:27        Bid        NA         NA
#> 40   0.976635   4896.838 2024-12-18 17:42:09        Bid        NA         NA
#> 41   0.976764   2378.399 2024-12-18 17:46:02        Bid        NA         NA
#> 42   0.976910  12454.707 2024-12-18 17:42:12        Bid        NA         NA
#> 43   0.977114     24.004 2024-12-18 17:40:48        Bid        NA         NA
#> 44   0.977234    137.703 2024-12-18 17:40:37        Bid        NA         NA
#> 45   0.977527  14642.000 2024-12-18 17:42:09        Bid        NA         NA
#> 46   0.977626  25256.400 2024-12-18 17:43:10        Bid        NA         NA
#> 47   0.977845  14640.955 2024-12-18 17:45:39        Bid        NA         NA
#> 48   0.978840   2972.752 2024-12-18 17:45:25        Bid        NA         NA
#> 49   0.978890   7780.589 2024-12-18 17:43:46        Bid        NA         NA
#> 50   0.978891 287674.336 2024-12-18 17:46:20        Bid        NA         NA
#> 51   0.980048   1688.155 2024-12-18 17:46:32        Bid        NA         NA
#> 52   0.980344   1847.589 2024-12-18 17:46:00        Bid        NA         NA
#> 53   0.980451  25124.632 2024-12-18 17:45:03        Bid        NA         NA
#> 54   0.980452   2129.530 2024-12-18 17:45:15        Bid        NA         NA
#> 55   0.980594  14599.915 2024-12-18 17:45:41        Bid        NA         NA
#> 56   0.981035  38476.100 2024-12-18 17:45:43        Bid        NA         NA
#> 57   0.981569  19207.356 2024-12-18 17:43:37        Bid        NA         NA
#> 58   0.982012  38248.600 2024-12-18 17:42:56        Bid        NA         NA
#> 59   0.982753 201474.726 2024-12-18 17:45:21        Bid        NA         NA
#> 60   0.983157  15720.934 2024-12-18 17:45:26        Bid        NA         NA
#> 61   0.983158   6813.996 2024-12-18 17:46:24        Bid        NA         NA
#> 62   0.983583  46359.549 2024-12-18 17:42:08        Bid        NA         NA
#> 63   0.983622     70.731 2024-12-18 17:46:08        Bid        NA         NA
#> 64   0.983686  15189.107 2024-12-18 17:45:27        Bid        NA         NA
#> 65   0.983687    192.278 2024-12-18 17:46:24        Bid        NA         NA
#> 66   0.983790   3124.741 2024-12-18 17:46:35        Bid        NA         NA
#> 67   0.984156  14547.081 2024-12-18 17:46:22        Bid        NA         NA
#> 68   0.984157     42.686 2024-12-18 17:46:26        Bid        NA         NA
#> 69   0.984729     43.295 2024-12-18 17:45:43        Bid        NA         NA
#> 70   0.985598  17184.800 2024-12-18 17:45:48        Bid        NA         NA
#> 71   0.985599   2327.131 2024-12-18 17:46:36        Bid        NA         NA
#> 72   0.985828  14522.397 2024-12-18 17:46:35        Bid        NA         NA
#> 73   0.986527   2755.703 2024-12-18 17:46:36        Bid        NA         NA
#> 74   0.986654  14510.247 2024-12-18 17:46:36        Bid        NA         NA
#> 75   0.986932     14.454 2024-12-18 17:45:27        Bid        NA         NA
#> 76   0.987056  14504.331 2024-12-18 17:46:34        Bid        NA         NA
#> 77   0.987092   5000.000 2024-12-18 17:46:34        Bid        NA         NA
#> 78   0.987192   9600.921 2024-12-18 17:46:35        Bid        NA         NA
#> 79   0.987330   5006.639 2024-12-18 17:46:36        Bid        NA         NA
#> 80   0.987370   3942.883 2024-12-18 17:46:36        Bid        NA         NA
#> 81   0.987552  16025.100 2024-12-18 17:46:35        Bid        NA         NA
#> 82   0.987556     93.553 2024-12-18 17:46:17        Bid        NA         NA
#> 83   0.987572  10547.240 2024-12-18 17:46:36        Bid        NA         NA
#> 84   0.987694     16.698 2024-12-18 17:46:35        Bid        NA         NA
#> 85   0.987719   9663.069 2024-12-18 17:46:36        Bid        NA         NA
#> 86   0.987794  10136.100 2024-12-18 17:46:35        Bid        NA         NA
#> 87   0.987848  10794.570 2024-12-18 17:46:36        Bid        NA         NA
#> 88   0.987855  11072.780 2024-12-18 17:46:36        Bid        NA         NA
#> 89   0.987862   9554.300 2024-12-18 17:45:28        Bid        NA         NA
#> 90   0.988157   9658.787 2024-12-18 17:46:36        Bid        NA         NA
#> 91   0.988282    868.000 2024-12-18 17:45:58        Bid        NA         NA
#> 92   0.988424    503.790 2024-12-18 17:46:20        Bid        NA         NA
#> 93   0.988489   9655.540 2024-12-18 17:46:36        Bid        NA         NA
#> 94   0.988517   7241.452 2024-12-18 17:46:36        Bid        NA         NA
#> 95   0.988582   4827.318 2024-12-18 17:46:36        Bid        NA         NA
#> 96   0.988642    868.000 2024-12-18 17:46:24        Bid        NA         NA
#> 97   0.988661    564.800 2024-12-18 17:46:36        Bid        NA         NA
#> 98   0.988662    673.487 2024-12-18 17:46:36        Bid        NA         NA
#> 99   0.988791    589.000 2024-12-18 17:46:35        Bid        NA         NA
#> 100  0.988801  10211.544 2024-12-18 17:46:35        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.990120   1184.431
#> 102        NA         NA                <NA>        Ask  0.990121   5899.930
#> 103        NA         NA                <NA>        Ask  0.990128  19487.354
#> 104        NA         NA                <NA>        Ask  0.990129   2901.000
#> 105        NA         NA                <NA>        Ask  0.990133   1628.002
#> 106        NA         NA                <NA>        Ask  0.990136   1368.557
#> 107        NA         NA                <NA>        Ask  0.990137  33738.153
#> 108        NA         NA                <NA>        Ask  0.990143   2100.000
#> 109        NA         NA                <NA>        Ask  0.990151    224.710
#> 110        NA         NA                <NA>        Ask  0.990163    564.800
#> 111        NA         NA                <NA>        Ask  0.990165   2098.800
#> 112        NA         NA                <NA>        Ask  0.990280    489.582
#> 113        NA         NA                <NA>        Ask  0.990337     33.799
#> 114        NA         NA                <NA>        Ask  0.990401   9636.906
#> 115        NA         NA                <NA>        Ask  0.990570   4839.316
#> 116        NA         NA                <NA>        Ask  0.990571     20.984
#> 117        NA         NA                <NA>        Ask  0.990663     18.034
#> 118        NA         NA                <NA>        Ask  0.990667   1368.557
#> 119        NA         NA                <NA>        Ask  0.990750    137.018
#> 120        NA         NA                <NA>        Ask  0.990759   9441.079
#> 121        NA         NA                <NA>        Ask  0.990927    399.656
#> 122        NA         NA                <NA>        Ask  0.990928   2755.703
#> 123        NA         NA                <NA>        Ask  0.990943  14447.451
#> 124        NA         NA                <NA>        Ask  0.991260  15299.700
#> 125        NA         NA                <NA>        Ask  0.991266   9458.630
#> 126        NA         NA                <NA>        Ask  0.991410   8448.710
#> 127        NA         NA                <NA>        Ask  0.991439   9543.400
#> 128        NA         NA                <NA>        Ask  0.991454     11.529
#> 129        NA         NA                <NA>        Ask  0.991472   8136.280
#> 130        NA         NA                <NA>        Ask  0.991638  14437.325
#> 131        NA         NA                <NA>        Ask  0.991758   5000.000
#> 132        NA         NA                <NA>        Ask  0.992129   1445.677
#> 133        NA         NA                <NA>        Ask  0.992245     31.277
#> 134        NA         NA                <NA>        Ask  0.992430     43.588
#> 135        NA         NA                <NA>        Ask  0.992664  14422.403
#> 136        NA         NA                <NA>        Ask  0.993049  15880.171
#> 137        NA         NA                <NA>        Ask  0.993136     55.456
#> 138        NA         NA                <NA>        Ask  0.993224     78.674
#> 139        NA         NA                <NA>        Ask  0.993283  17178.400
#> 140        NA         NA                <NA>        Ask  0.993310     48.000
#> 141        NA         NA                <NA>        Ask  0.993316  15773.184
#> 142        NA         NA                <NA>        Ask  0.993471  39764.400
#> 143        NA         NA                <NA>        Ask  0.993634     48.271
#> 144        NA         NA                <NA>        Ask  0.994000     41.491
#> 145        NA         NA                <NA>        Ask  0.994025     23.810
#> 146        NA         NA                <NA>        Ask  0.994054     10.229
#> 147        NA         NA                <NA>        Ask  0.994142     98.547
#> 148        NA         NA                <NA>        Ask  0.994746  24826.063
#> 149        NA         NA                <NA>        Ask  0.994890  18821.704
#> 150        NA         NA                <NA>        Ask  0.995038  14388.001
#> 151        NA         NA                <NA>        Ask  0.995104    136.419
#> 152        NA         NA                <NA>        Ask  0.995443  37945.700
#> 153        NA         NA                <NA>        Ask  0.996198   4698.449
#> 154        NA         NA                <NA>        Ask  0.996211     98.300
#> 155        NA         NA                <NA>        Ask  0.996504   9220.683
#> 156        NA         NA                <NA>        Ask  0.996922     11.466
#> 157        NA         NA                <NA>        Ask  0.997461    138.454
#> 158        NA         NA                <NA>        Ask  0.997462     20.000
#> 159        NA         NA                <NA>        Ask  0.997552  14351.745
#> 160        NA         NA                <NA>        Ask  0.997656    160.789
#> 161        NA         NA                <NA>        Ask  0.997916  26581.100
#> 162        NA         NA                <NA>        Ask  0.998179  11565.411
#> 163        NA         NA                <NA>        Ask  0.998690      7.057
#> 164        NA         NA                <NA>        Ask  0.999016   1840.834
#> 165        NA         NA                <NA>        Ask  0.999134     26.368
#> 166        NA         NA                <NA>        Ask  0.999142  21697.227
#> 167        NA         NA                <NA>        Ask  0.999292 289384.387
#> 168        NA         NA                <NA>        Ask  0.999293   7231.162
#> 169        NA         NA                <NA>        Ask  0.999331   8677.395
#> 170        NA         NA                <NA>        Ask  0.999478    135.822
#> 171        NA         NA                <NA>        Ask  0.999589   2911.042
#> 172        NA         NA                <NA>        Ask  1.000000   7459.003
#> 173        NA         NA                <NA>        Ask  1.000109     23.665
#> 174        NA         NA                <NA>        Ask  1.000296   1653.982
#> 175        NA         NA                <NA>        Ask  1.000530  14309.017
#> 176        NA         NA                <NA>        Ask  1.000600  12732.200
#> 177        NA         NA                <NA>        Ask  1.001000      8.018
#> 178        NA         NA                <NA>        Ask  1.001581    109.838
#> 179        NA         NA                <NA>        Ask  1.002059   2282.684
#> 180        NA         NA                <NA>        Ask  1.002077     47.865
#> 181        NA         NA                <NA>        Ask  1.002244     84.063
#> 182        NA         NA                <NA>        Ask  1.002364     21.780
#> 183        NA         NA                <NA>        Ask  1.002422     11.396
#> 184        NA         NA                <NA>        Ask  1.002521     18.217
#> 185        NA         NA                <NA>        Ask  1.003297     98.300
#> 186        NA         NA                <NA>        Ask  1.003871    135.227
#> 187        NA         NA                <NA>        Ask  1.003906    112.589
#> 188        NA         NA                <NA>        Ask  1.004560   6487.000
#> 189        NA         NA                <NA>        Ask  1.004865     98.361
#> 190        NA         NA                <NA>        Ask  1.006230     23.521
#> 191        NA         NA                <NA>        Ask  1.006724    682.084
#> 192        NA         NA                <NA>        Ask  1.007105    214.794
#> 193        NA         NA                <NA>        Ask  1.007680   6626.100
#> 194        NA         NA                <NA>        Ask  1.007951     11.334
#> 195        NA         NA                <NA>        Ask  1.008283    134.636
#> 196        NA         NA                <NA>        Ask  1.008559      7.045
#> 197        NA         NA                <NA>        Ask  1.009236  14185.590
#> 198        NA         NA                <NA>        Ask  1.009796     31.277
#> 199        NA         NA                <NA>        Ask  1.010000  41000.000
#> 200        NA         NA                <NA>        Ask  1.010433     98.300
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:46:36
#> 102 2024-12-18 17:46:36
#> 103 2024-12-18 17:46:36
#> 104 2024-12-18 17:46:36
#> 105 2024-12-18 17:46:36
#> 106 2024-12-18 17:46:35
#> 107 2024-12-18 17:46:36
#> 108 2024-12-18 17:46:34
#> 109 2024-12-18 17:46:34
#> 110 2024-12-18 17:46:34
#> 111 2024-12-18 17:46:34
#> 112 2024-12-18 17:46:32
#> 113 2024-12-18 17:46:33
#> 114 2024-12-18 17:46:36
#> 115 2024-12-18 17:46:36
#> 116 2024-12-18 17:46:32
#> 117 2024-12-18 17:46:35
#> 118 2024-12-18 17:46:33
#> 119 2024-12-18 17:41:38
#> 120 2024-12-18 17:46:36
#> 121 2024-12-18 17:46:36
#> 122 2024-12-18 17:46:35
#> 123 2024-12-18 17:46:36
#> 124 2024-12-18 17:46:32
#> 125 2024-12-18 17:46:35
#> 126 2024-12-18 17:46:36
#> 127 2024-12-18 17:46:32
#> 128 2024-12-18 17:46:32
#> 129 2024-12-18 17:46:35
#> 130 2024-12-18 17:46:34
#> 131 2024-12-18 17:46:24
#> 132 2024-12-18 17:46:33
#> 133 2024-12-18 15:04:40
#> 134 2024-12-18 17:45:39
#> 135 2024-12-18 17:46:35
#> 136 2024-12-18 17:45:32
#> 137 2024-12-18 13:26:56
#> 138 2024-12-18 15:12:45
#> 139 2024-12-18 17:46:27
#> 140 2024-12-18 17:46:34
#> 141 2024-12-18 17:45:31
#> 142 2024-12-18 17:45:30
#> 143 2024-12-18 12:38:18
#> 144 2024-12-18 12:25:06
#> 145 2024-12-18 13:36:40
#> 146 2024-12-18 17:45:27
#> 147 2024-12-18 12:37:51
#> 148 2024-12-18 17:45:26
#> 149 2024-12-18 17:46:32
#> 150 2024-12-18 17:46:02
#> 151 2024-12-18 13:34:11
#> 152 2024-12-18 17:45:30
#> 153 2024-12-18 17:46:33
#> 154 2024-12-18 15:07:21
#> 155 2024-12-18 17:40:28
#> 156 2024-12-18 12:33:01
#> 157 2024-12-18 17:46:26
#> 158 2024-12-18 17:40:55
#> 159 2024-12-18 17:46:34
#> 160 2024-12-18 17:46:26
#> 161 2024-12-18 17:45:30
#> 162 2024-12-18 17:46:33
#> 163 2024-12-18 12:34:04
#> 164 2024-12-18 17:46:33
#> 165 2024-12-18 17:46:28
#> 166 2024-12-18 17:45:32
#> 167 2024-12-18 17:46:28
#> 168 2024-12-18 17:46:23
#> 169 2024-12-18 17:46:23
#> 170 2024-12-18 13:11:41
#> 171 2024-12-18 17:45:44
#> 172 2024-12-18 17:41:34
#> 173 2024-12-18 12:15:38
#> 174 2024-12-18 17:46:32
#> 175 2024-12-18 17:46:32
#> 176 2024-12-18 17:40:39
#> 177 2024-12-18 17:24:38
#> 178 2024-12-18 15:18:26
#> 179 2024-12-18 17:45:48
#> 180 2024-12-18 04:26:55
#> 181 2024-12-18 07:39:44
#> 182 2024-12-18 12:02:37
#> 183 2024-12-18 03:31:33
#> 184 2024-12-18 12:21:12
#> 185 2024-12-18 05:52:42
#> 186 2024-12-18 07:01:36
#> 187 2024-12-18 05:50:53
#> 188 2024-12-18 17:40:36
#> 189 2024-12-18 04:33:06
#> 190 2024-12-18 09:02:51
#> 191 2024-12-18 06:31:37
#> 192 2024-12-18 12:01:30
#> 193 2024-12-18 17:37:21
#> 194 2024-12-18 02:55:05
#> 195 2024-12-18 05:32:05
#> 196 2024-12-18 14:36:34
#> 197 2024-12-18 17:45:24
#> 198 2024-12-18 16:53:46
#> 199 2024-12-18 15:53:30
#> 200 2024-12-18 14:38:21
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 2    0.960000  63095.795 2024-12-18 17:37:35        Bid        NA         NA
#> 3    0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 4    0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 5    0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 6    0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 7    0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 8    0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 9    0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 10   0.962781   6457.100 2024-12-18 17:42:56        Bid        NA         NA
#> 11   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 12   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 13   0.965000    145.118 2024-12-18 16:58:19        Bid        NA         NA
#> 14   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 15   0.965417   1643.570 2024-12-18 17:46:02        Bid        NA         NA
#> 16   0.965419   7420.457 2024-12-18 17:46:01        Bid        NA         NA
#> 17   0.965420   1610.906 2024-12-18 17:46:02        Bid        NA         NA
#> 18   0.965430  11273.578 2024-12-18 17:46:03        Bid        NA         NA
#> 19   0.967000     41.491 2024-12-18 17:36:07        Bid        NA         NA
#> 20   0.968252     99.690 2024-12-18 17:37:33        Bid        NA         NA
#> 21   0.968500   6118.811 2024-12-18 17:41:06        Bid        NA         NA
#> 22   0.968544   6347.900 2024-12-18 17:41:13        Bid        NA         NA
#> 23   0.968700    138.916 2024-12-18 17:34:12        Bid        NA         NA
#> 24   0.968846  14776.956 2024-12-18 17:45:30        Bid        NA         NA
#> 25   0.969000     41.491 2024-12-18 17:38:06        Bid        NA         NA
#> 26   0.970070    154.165 2024-12-18 17:29:38        Bid        NA         NA
#> 27   0.971000     41.491 2024-12-18 17:41:06        Bid        NA         NA
#> 28   0.971170     24.151 2024-12-18 17:35:45        Bid        NA         NA
#> 29   0.972144     48.720 2024-12-18 17:40:34        Bid        NA         NA
#> 30   0.972958    138.308 2024-12-18 17:36:37        Bid        NA         NA
#> 31   0.973000     41.491 2024-12-18 17:46:07        Bid        NA         NA
#> 32   0.973422   6106.993 2024-12-18 17:41:06        Bid        NA         NA
#> 33   0.973808   3262.606 2024-12-18 17:45:25        Bid        NA         NA
#> 34   0.974513     49.256 2024-12-18 17:33:21        Bid        NA         NA
#> 35   0.974677  12680.200 2024-12-18 17:41:43        Bid        NA         NA
#> 36   0.975769     11.609 2024-12-18 17:40:27        Bid        NA         NA
#> 37   0.976635   4896.838 2024-12-18 17:42:09        Bid        NA         NA
#> 38   0.976764   2378.399 2024-12-18 17:46:02        Bid        NA         NA
#> 39   0.976910  12454.707 2024-12-18 17:42:12        Bid        NA         NA
#> 40   0.977114     24.004 2024-12-18 17:40:48        Bid        NA         NA
#> 41   0.977234    137.703 2024-12-18 17:40:37        Bid        NA         NA
#> 42   0.977527  14642.000 2024-12-18 17:42:09        Bid        NA         NA
#> 43   0.977626  25256.400 2024-12-18 17:43:10        Bid        NA         NA
#> 44   0.977845  14640.955 2024-12-18 17:45:39        Bid        NA         NA
#> 45   0.978840   2972.752 2024-12-18 17:45:25        Bid        NA         NA
#> 46   0.978890   7780.589 2024-12-18 17:43:46        Bid        NA         NA
#> 47   0.978891 287674.336 2024-12-18 17:46:20        Bid        NA         NA
#> 48   0.980048   1688.155 2024-12-18 17:46:32        Bid        NA         NA
#> 49   0.980344   1847.589 2024-12-18 17:46:00        Bid        NA         NA
#> 50   0.980451  25124.632 2024-12-18 17:45:03        Bid        NA         NA
#> 51   0.980452   2129.530 2024-12-18 17:45:15        Bid        NA         NA
#> 52   0.980594  14599.915 2024-12-18 17:45:41        Bid        NA         NA
#> 53   0.981035  38476.100 2024-12-18 17:45:43        Bid        NA         NA
#> 54   0.981569  19207.356 2024-12-18 17:43:37        Bid        NA         NA
#> 55   0.982012  38248.600 2024-12-18 17:42:56        Bid        NA         NA
#> 56   0.982753 201474.726 2024-12-18 17:45:21        Bid        NA         NA
#> 57   0.983157  15720.934 2024-12-18 17:45:26        Bid        NA         NA
#> 58   0.983158   6813.996 2024-12-18 17:46:24        Bid        NA         NA
#> 59   0.983583  46359.549 2024-12-18 17:42:08        Bid        NA         NA
#> 60   0.983622     70.731 2024-12-18 17:46:08        Bid        NA         NA
#> 61   0.983686  15189.107 2024-12-18 17:45:27        Bid        NA         NA
#> 62   0.983687    192.278 2024-12-18 17:46:24        Bid        NA         NA
#> 63   0.983790   3124.741 2024-12-18 17:46:35        Bid        NA         NA
#> 64   0.984156  14547.081 2024-12-18 17:46:22        Bid        NA         NA
#> 65   0.984157     42.686 2024-12-18 17:46:26        Bid        NA         NA
#> 66   0.984729     43.295 2024-12-18 17:45:43        Bid        NA         NA
#> 67   0.985598  17184.800 2024-12-18 17:45:48        Bid        NA         NA
#> 68   0.985814   2326.623 2024-12-18 17:46:24        Bid        NA         NA
#> 69   0.986385  14514.200 2024-12-18 17:46:31        Bid        NA         NA
#> 70   0.986932     14.454 2024-12-18 17:45:27        Bid        NA         NA
#> 71   0.987056  14504.331 2024-12-18 17:46:34        Bid        NA         NA
#> 72   0.987091   2755.703 2024-12-18 17:46:22        Bid        NA         NA
#> 73   0.987092   5000.000 2024-12-18 17:46:34        Bid        NA         NA
#> 74   0.987429  14498.858 2024-12-18 17:46:20        Bid        NA         NA
#> 75   0.987556     93.553 2024-12-18 17:46:17        Bid        NA         NA
#> 76   0.987624   2901.000 2024-12-18 17:46:20        Bid        NA         NA
#> 77   0.987859   9600.921 2024-12-18 17:46:32        Bid        NA         NA
#> 78   0.987862   9554.300 2024-12-18 17:45:28        Bid        NA         NA
#> 79   0.987933   3942.883 2024-12-18 17:46:22        Bid        NA         NA
#> 80   0.987979   9660.522 2024-12-18 17:46:35        Bid        NA         NA
#> 81   0.988002     16.698 2024-12-18 17:46:34        Bid        NA         NA
#> 82   0.988010   5006.639 2024-12-18 17:46:22        Bid        NA         NA
#> 83   0.988128  16025.100 2024-12-18 17:46:34        Bid        NA         NA
#> 84   0.988282    868.000 2024-12-18 17:45:58        Bid        NA         NA
#> 85   0.988290  10502.250 2024-12-18 17:46:22        Bid        NA         NA
#> 86   0.988291  10136.100 2024-12-18 17:46:32        Bid        NA         NA
#> 87   0.988424    503.790 2024-12-18 17:46:20        Bid        NA         NA
#> 88   0.988550   9654.945 2024-12-18 17:46:35        Bid        NA         NA
#> 89   0.988566  10863.120 2024-12-18 17:46:22        Bid        NA         NA
#> 90   0.988573  11109.530 2024-12-18 17:46:22        Bid        NA         NA
#> 91   0.988642    868.000 2024-12-18 17:46:24        Bid        NA         NA
#> 92   0.988779  24131.779 2024-12-18 17:46:35        Bid        NA         NA
#> 93   0.988785    194.905 2024-12-18 17:46:33        Bid        NA         NA
#> 94   0.988789    564.800 2024-12-18 17:46:34        Bid        NA         NA
#> 95   0.988791    589.000 2024-12-18 17:46:35        Bid        NA         NA
#> 96   0.988793   3576.965 2024-12-18 17:46:34        Bid        NA         NA
#> 97   0.988794   5910.390 2024-12-18 17:46:34        Bid        NA         NA
#> 98   0.988796  10211.544 2024-12-18 17:46:35        Bid        NA         NA
#> 99   0.988798     33.799 2024-12-18 17:46:35        Bid        NA         NA
#> 100  0.988799    673.341 2024-12-18 17:46:35        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.990138   1184.432
#> 102        NA         NA                <NA>        Ask  0.990139   1628.002
#> 103        NA         NA                <NA>        Ask  0.990141   1368.557
#> 104        NA         NA                <NA>        Ask  0.990143   2100.000
#> 105        NA         NA                <NA>        Ask  0.990149   3576.965
#> 106        NA         NA                <NA>        Ask  0.990151    224.710
#> 107        NA         NA                <NA>        Ask  0.990163    564.800
#> 108        NA         NA                <NA>        Ask  0.990165   2098.800
#> 109        NA         NA                <NA>        Ask  0.990227   7228.948
#> 110        NA         NA                <NA>        Ask  0.990229  19277.154
#> 111        NA         NA                <NA>        Ask  0.990280    489.582
#> 112        NA         NA                <NA>        Ask  0.990282  10000.000
#> 113        NA         NA                <NA>        Ask  0.990337     33.799
#> 114        NA         NA                <NA>        Ask  0.990571     20.984
#> 115        NA         NA                <NA>        Ask  0.990667   1368.557
#> 116        NA         NA                <NA>        Ask  0.990750    137.018
#> 117        NA         NA                <NA>        Ask  0.990884   9632.213
#> 118        NA         NA                <NA>        Ask  0.990958   4836.000
#> 119        NA         NA                <NA>        Ask  0.990972     18.034
#> 120        NA         NA                <NA>        Ask  0.990997   2901.000
#> 121        NA         NA                <NA>        Ask  0.991027   1063.930
#> 122        NA         NA                <NA>        Ask  0.991028   4839.316
#> 123        NA         NA                <NA>        Ask  0.991248    409.115
#> 124        NA         NA                <NA>        Ask  0.991256   9441.079
#> 125        NA         NA                <NA>        Ask  0.991260  15299.700
#> 126        NA         NA                <NA>        Ask  0.991293  14442.357
#> 127        NA         NA                <NA>        Ask  0.991439   9543.400
#> 128        NA         NA                <NA>        Ask  0.991454     11.529
#> 129        NA         NA                <NA>        Ask  0.991597   2755.703
#> 130        NA         NA                <NA>        Ask  0.991638  14437.325
#> 131        NA         NA                <NA>        Ask  0.991758   5000.000
#> 132        NA         NA                <NA>        Ask  0.992016   9711.320
#> 133        NA         NA                <NA>        Ask  0.992129   1445.677
#> 134        NA         NA                <NA>        Ask  0.992130   8341.900
#> 135        NA         NA                <NA>        Ask  0.992186   8005.070
#> 136        NA         NA                <NA>        Ask  0.992245     31.277
#> 137        NA         NA                <NA>        Ask  0.992430     43.588
#> 138        NA         NA                <NA>        Ask  0.993049  15880.171
#> 139        NA         NA                <NA>        Ask  0.993136     55.456
#> 140        NA         NA                <NA>        Ask  0.993224     78.674
#> 141        NA         NA                <NA>        Ask  0.993246  14413.961
#> 142        NA         NA                <NA>        Ask  0.993283  17178.400
#> 143        NA         NA                <NA>        Ask  0.993310     48.000
#> 144        NA         NA                <NA>        Ask  0.993316  15773.184
#> 145        NA         NA                <NA>        Ask  0.993471  39764.400
#> 146        NA         NA                <NA>        Ask  0.993634     48.271
#> 147        NA         NA                <NA>        Ask  0.994000     41.491
#> 148        NA         NA                <NA>        Ask  0.994025     23.810
#> 149        NA         NA                <NA>        Ask  0.994054     10.229
#> 150        NA         NA                <NA>        Ask  0.994142     98.547
#> 151        NA         NA                <NA>        Ask  0.994746  24826.063
#> 152        NA         NA                <NA>        Ask  0.994890  18821.704
#> 153        NA         NA                <NA>        Ask  0.995038  14388.001
#> 154        NA         NA                <NA>        Ask  0.995104    136.419
#> 155        NA         NA                <NA>        Ask  0.995443  37945.700
#> 156        NA         NA                <NA>        Ask  0.996198   4698.449
#> 157        NA         NA                <NA>        Ask  0.996211     98.300
#> 158        NA         NA                <NA>        Ask  0.996504   9220.683
#> 159        NA         NA                <NA>        Ask  0.996922     11.466
#> 160        NA         NA                <NA>        Ask  0.997461    138.454
#> 161        NA         NA                <NA>        Ask  0.997462     20.000
#> 162        NA         NA                <NA>        Ask  0.997552  14351.745
#> 163        NA         NA                <NA>        Ask  0.997656    160.789
#> 164        NA         NA                <NA>        Ask  0.997916  26581.100
#> 165        NA         NA                <NA>        Ask  0.998179  11565.411
#> 166        NA         NA                <NA>        Ask  0.998690      7.057
#> 167        NA         NA                <NA>        Ask  0.999016   1840.834
#> 168        NA         NA                <NA>        Ask  0.999134     26.368
#> 169        NA         NA                <NA>        Ask  0.999142  21697.227
#> 170        NA         NA                <NA>        Ask  0.999292 289384.387
#> 171        NA         NA                <NA>        Ask  0.999293   7231.162
#> 172        NA         NA                <NA>        Ask  0.999331   8677.395
#> 173        NA         NA                <NA>        Ask  0.999478    135.822
#> 174        NA         NA                <NA>        Ask  0.999589   2911.042
#> 175        NA         NA                <NA>        Ask  1.000000   7459.003
#> 176        NA         NA                <NA>        Ask  1.000109     23.665
#> 177        NA         NA                <NA>        Ask  1.000296   1653.982
#> 178        NA         NA                <NA>        Ask  1.000530  14309.017
#> 179        NA         NA                <NA>        Ask  1.000600  12732.200
#> 180        NA         NA                <NA>        Ask  1.001000      8.018
#> 181        NA         NA                <NA>        Ask  1.001581    109.838
#> 182        NA         NA                <NA>        Ask  1.002059   2282.684
#> 183        NA         NA                <NA>        Ask  1.002077     47.865
#> 184        NA         NA                <NA>        Ask  1.002244     84.063
#> 185        NA         NA                <NA>        Ask  1.002364     21.780
#> 186        NA         NA                <NA>        Ask  1.002422     11.396
#> 187        NA         NA                <NA>        Ask  1.002521     18.217
#> 188        NA         NA                <NA>        Ask  1.003297     98.300
#> 189        NA         NA                <NA>        Ask  1.003871    135.227
#> 190        NA         NA                <NA>        Ask  1.003906    112.589
#> 191        NA         NA                <NA>        Ask  1.004560   6487.000
#> 192        NA         NA                <NA>        Ask  1.004865     98.361
#> 193        NA         NA                <NA>        Ask  1.006230     23.521
#> 194        NA         NA                <NA>        Ask  1.006724    682.084
#> 195        NA         NA                <NA>        Ask  1.007105    214.794
#> 196        NA         NA                <NA>        Ask  1.007680   6626.100
#> 197        NA         NA                <NA>        Ask  1.007951     11.334
#> 198        NA         NA                <NA>        Ask  1.008283    134.636
#> 199        NA         NA                <NA>        Ask  1.008559      7.045
#> 200        NA         NA                <NA>        Ask  1.009236  14185.590
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:46:35
#> 102 2024-12-18 17:46:35
#> 103 2024-12-18 17:46:35
#> 104 2024-12-18 17:46:34
#> 105 2024-12-18 17:46:34
#> 106 2024-12-18 17:46:34
#> 107 2024-12-18 17:46:34
#> 108 2024-12-18 17:46:34
#> 109 2024-12-18 17:46:34
#> 110 2024-12-18 17:46:34
#> 111 2024-12-18 17:46:32
#> 112 2024-12-18 17:46:34
#> 113 2024-12-18 17:46:33
#> 114 2024-12-18 17:46:32
#> 115 2024-12-18 17:46:33
#> 116 2024-12-18 17:41:38
#> 117 2024-12-18 17:46:34
#> 118 2024-12-18 17:46:33
#> 119 2024-12-18 17:46:34
#> 120 2024-12-18 17:46:33
#> 121 2024-12-18 17:46:33
#> 122 2024-12-18 17:46:25
#> 123 2024-12-18 17:46:28
#> 124 2024-12-18 17:46:32
#> 125 2024-12-18 17:46:32
#> 126 2024-12-18 17:46:35
#> 127 2024-12-18 17:46:32
#> 128 2024-12-18 17:46:32
#> 129 2024-12-18 17:46:32
#> 130 2024-12-18 17:46:34
#> 131 2024-12-18 17:46:24
#> 132 2024-12-18 17:46:32
#> 133 2024-12-18 17:46:33
#> 134 2024-12-18 17:46:22
#> 135 2024-12-18 17:46:32
#> 136 2024-12-18 15:04:40
#> 137 2024-12-18 17:45:39
#> 138 2024-12-18 17:45:32
#> 139 2024-12-18 13:26:56
#> 140 2024-12-18 15:12:45
#> 141 2024-12-18 17:46:31
#> 142 2024-12-18 17:46:27
#> 143 2024-12-18 17:46:34
#> 144 2024-12-18 17:45:31
#> 145 2024-12-18 17:45:30
#> 146 2024-12-18 12:38:18
#> 147 2024-12-18 12:25:06
#> 148 2024-12-18 13:36:40
#> 149 2024-12-18 17:45:27
#> 150 2024-12-18 12:37:51
#> 151 2024-12-18 17:45:26
#> 152 2024-12-18 17:46:32
#> 153 2024-12-18 17:46:02
#> 154 2024-12-18 13:34:11
#> 155 2024-12-18 17:45:30
#> 156 2024-12-18 17:46:33
#> 157 2024-12-18 15:07:21
#> 158 2024-12-18 17:40:28
#> 159 2024-12-18 12:33:01
#> 160 2024-12-18 17:46:26
#> 161 2024-12-18 17:40:55
#> 162 2024-12-18 17:46:34
#> 163 2024-12-18 17:46:26
#> 164 2024-12-18 17:45:30
#> 165 2024-12-18 17:46:33
#> 166 2024-12-18 12:34:04
#> 167 2024-12-18 17:46:33
#> 168 2024-12-18 17:46:28
#> 169 2024-12-18 17:45:32
#> 170 2024-12-18 17:46:28
#> 171 2024-12-18 17:46:23
#> 172 2024-12-18 17:46:23
#> 173 2024-12-18 13:11:41
#> 174 2024-12-18 17:45:44
#> 175 2024-12-18 17:41:34
#> 176 2024-12-18 12:15:38
#> 177 2024-12-18 17:46:32
#> 178 2024-12-18 17:46:32
#> 179 2024-12-18 17:40:39
#> 180 2024-12-18 17:24:38
#> 181 2024-12-18 15:18:26
#> 182 2024-12-18 17:45:48
#> 183 2024-12-18 04:26:55
#> 184 2024-12-18 07:39:44
#> 185 2024-12-18 12:02:37
#> 186 2024-12-18 03:31:33
#> 187 2024-12-18 12:21:12
#> 188 2024-12-18 05:52:42
#> 189 2024-12-18 07:01:36
#> 190 2024-12-18 05:50:53
#> 191 2024-12-18 17:40:36
#> 192 2024-12-18 04:33:06
#> 193 2024-12-18 09:02:51
#> 194 2024-12-18 06:31:37
#> 195 2024-12-18 12:01:30
#> 196 2024-12-18 17:37:21
#> 197 2024-12-18 02:55:05
#> 198 2024-12-18 05:32:05
#> 199 2024-12-18 14:36:34
#> 200 2024-12-18 17:45:24