Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 2    0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 3    0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 4    0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 5    0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 6    0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 7    0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 8    0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 9    0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 10   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 11   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 12   0.958430   5655.280 2024-12-18 16:30:01        Bid        NA         NA
#> 13   0.958711   1616.828 2024-12-18 16:30:01        Bid        NA         NA
#> 14   0.958713   4875.193 2024-12-18 16:30:01        Bid        NA         NA
#> 15   0.958753   9818.348 2024-12-18 16:30:02        Bid        NA         NA
#> 16   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 17   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 18   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 19   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 20   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 21   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 22   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 23   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 24   0.960231   6106.993 2024-12-18 16:29:35        Bid        NA         NA
#> 25   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 26   0.960274  14913.350 2024-12-18 16:32:48        Bid        NA         NA
#> 27   0.960714  12680.200 2024-12-18 16:29:32        Bid        NA         NA
#> 28   0.960863   3426.766 2024-12-18 16:30:51        Bid        NA         NA
#> 29   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 30   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 31   0.961036   3139.799 2024-12-18 16:33:21        Bid        NA         NA
#> 32   0.961463  59317.270 2024-12-18 16:23:08        Bid        NA         NA
#> 33   0.961464 288469.221 2024-12-18 16:33:28        Bid        NA         NA
#> 34   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 35   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 36   0.962591  20442.902 2024-12-18 16:33:19        Bid        NA         NA
#> 37   0.962609   3743.389 2024-12-18 16:33:15        Bid        NA         NA
#> 38   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 39   0.963156   2469.598 2024-12-18 16:32:03        Bid        NA         NA
#> 40   0.963352   2274.763 2024-12-18 16:32:37        Bid        NA         NA
#> 41   0.963611  14860.915 2024-12-18 16:33:03        Bid        NA         NA
#> 42   0.963760  12454.707 2024-12-18 16:30:33        Bid        NA         NA
#> 43   0.963965  25256.400 2024-12-18 16:30:25        Bid        NA         NA
#> 44   0.964049   8171.696 2024-12-18 16:33:19        Bid        NA         NA
#> 45   0.964161   9953.163 2024-12-18 16:33:23        Bid        NA         NA
#> 46   0.964162   6908.728 2024-12-18 16:33:23        Bid        NA         NA
#> 47   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 48   0.964818   8957.002 2024-12-18 16:33:23        Bid        NA         NA
#> 49   0.965675   3317.801 2024-12-18 16:33:24        Bid        NA         NA
#> 50   0.965839  25124.632 2024-12-18 16:30:16        Bid        NA         NA
#> 51   0.965887  14942.748 2024-12-18 16:33:23        Bid        NA         NA
#> 52   0.966006  14822.368 2024-12-18 16:32:04        Bid        NA         NA
#> 53   0.966041   5970.183 2024-12-18 16:33:23        Bid        NA         NA
#> 54   0.966054   4975.123 2024-12-18 16:33:23        Bid        NA         NA
#> 55   0.966396  38476.100 2024-12-18 16:30:25        Bid        NA         NA
#> 56   0.966444   6971.661 2024-12-18 16:33:23        Bid        NA         NA
#> 57   0.966775  38248.600 2024-12-18 16:33:27        Bid        NA         NA
#> 58   0.967067 204742.770 2024-12-18 16:32:03        Bid        NA         NA
#> 59   0.967115   2990.340 2024-12-18 16:33:11        Bid        NA         NA
#> 60   0.967321     43.755 2024-12-18 16:33:15        Bid        NA         NA
#> 61   0.967472  19207.356 2024-12-18 16:32:03        Bid        NA         NA
#> 62   0.967473     28.549 2024-12-18 16:32:50        Bid        NA         NA
#> 63   0.967949  17498.900 2024-12-18 16:33:10        Bid        NA         NA
#> 64   0.967992  15720.934 2024-12-18 16:32:04        Bid        NA         NA
#> 65   0.968018    289.433 2024-12-18 16:33:21        Bid        NA         NA
#> 66   0.968019   2369.393 2024-12-18 16:33:29        Bid        NA         NA
#> 67   0.968124  15189.107 2024-12-18 16:32:04        Bid        NA         NA
#> 68   0.968125  14789.867 2024-12-18 16:33:30        Bid        NA         NA
#> 69   0.968562   5000.000 2024-12-18 16:32:48        Bid        NA         NA
#> 70   0.969088  14775.159 2024-12-18 16:33:30        Bid        NA         NA
#> 71   0.969396  20804.254 2024-12-18 16:33:26        Bid        NA         NA
#> 72   0.969487     14.454 2024-12-18 16:32:03        Bid        NA         NA
#> 73   0.969523  10430.240 2024-12-18 16:33:24        Bid        NA         NA
#> 74   0.969566   2755.703 2024-12-18 16:33:15        Bid        NA         NA
#> 75   0.969671  14766.277 2024-12-18 16:33:30        Bid        NA         NA
#> 76   0.969751  11088.760 2024-12-18 16:33:24        Bid        NA         NA
#> 77   0.970158   9839.244 2024-12-18 16:33:30        Bid        NA         NA
#> 78   0.970332   2938.000 2024-12-18 16:33:27        Bid        NA         NA
#> 79   0.970341  16025.100 2024-12-18 16:33:29        Bid        NA         NA
#> 80   0.970380     49.000 2024-12-18 16:33:29        Bid        NA         NA
#> 81   0.970458   7566.874 2024-12-18 16:33:29        Bid        NA         NA
#> 82   0.970468   9700.800 2024-12-18 16:32:03        Bid        NA         NA
#> 83   0.970481  10700.630 2024-12-18 16:33:20        Bid        NA         NA
#> 84   0.970482    592.262 2024-12-18 16:33:29        Bid        NA         NA
#> 85   0.970487   4897.000 2024-12-18 16:33:27        Bid        NA         NA
#> 86   0.970641     16.697 2024-12-18 16:33:29        Bid        NA         NA
#> 87   0.970808    640.300 2024-12-18 16:33:27        Bid        NA         NA
#> 88   0.970810   5006.639 2024-12-18 16:33:29        Bid        NA         NA
#> 89   0.970816   2446.000 2024-12-18 16:33:30        Bid        NA         NA
#> 90   0.970900    511.354 2024-12-18 16:33:24        Bid        NA         NA
#> 91   0.970901   5405.322 2024-12-18 16:33:29        Bid        NA         NA
#> 92   0.970902  11580.100 2024-12-18 16:33:29        Bid        NA         NA
#> 93   0.970961   9831.108 2024-12-18 16:33:30        Bid        NA         NA
#> 94   0.971176   7371.696 2024-12-18 16:33:30        Bid        NA         NA
#> 95   0.971297   2456.927 2024-12-18 16:33:30        Bid        NA         NA
#> 96   0.971352    640.300 2024-12-18 16:33:28        Bid        NA         NA
#> 97   0.971413   5876.592 2024-12-18 16:33:29        Bid        NA         NA
#> 98   0.971414    297.518 2024-12-18 16:33:29        Bid        NA         NA
#> 99   0.971592   3576.965 2024-12-18 16:33:29        Bid        NA         NA
#> 100  0.971593     67.597 2024-12-18 16:33:29        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.971920     87.466
#> 102        NA         NA                <NA>        Ask  0.971921   1392.377
#> 103        NA         NA                <NA>        Ask  0.971922   2455.349
#> 104        NA         NA                <NA>        Ask  0.972280    263.439
#> 105        NA         NA                <NA>        Ask  0.972659   3926.615
#> 106        NA         NA                <NA>        Ask  0.972939  10705.500
#> 107        NA         NA                <NA>        Ask  0.972940   9636.350
#> 108        NA         NA                <NA>        Ask  0.973244   4839.316
#> 109        NA         NA                <NA>        Ask  0.973245  15299.700
#> 110        NA         NA                <NA>        Ask  0.973420   1472.995
#> 111        NA         NA                <NA>        Ask  0.973421  14709.409
#> 112        NA         NA                <NA>        Ask  0.973497   8236.510
#> 113        NA         NA                <NA>        Ask  0.973535     14.769
#> 114        NA         NA                <NA>        Ask  0.973555   2938.000
#> 115        NA         NA                <NA>        Ask  0.973598   2755.703
#> 116        NA         NA                <NA>        Ask  0.973600   9719.700
#> 117        NA         NA                <NA>        Ask  0.973768   5000.000
#> 118        NA         NA                <NA>        Ask  0.973849   7840.710
#> 119        NA         NA                <NA>        Ask  0.974188  14697.826
#> 120        NA         NA                <NA>        Ask  0.974655   2036.704
#> 121        NA         NA                <NA>        Ask  0.974684   5434.091
#> 122        NA         NA                <NA>        Ask  0.974721   2717.046
#> 123        NA         NA                <NA>        Ask  0.975208    486.355
#> 124        NA         NA                <NA>        Ask  0.975209     43.973
#> 125        NA         NA                <NA>        Ask  0.975306  14264.489
#> 126        NA         NA                <NA>        Ask  0.975354  17495.400
#> 127        NA         NA                <NA>        Ask  0.975995     24.250
#> 128        NA         NA                <NA>        Ask  0.976054   6113.353
#> 129        NA         NA                <NA>        Ask  0.976484     10.229
#> 130        NA         NA                <NA>        Ask  0.976755    144.411
#> 131        NA         NA                <NA>        Ask  0.976756   6792.614
#> 132        NA         NA                <NA>        Ask  0.976803     27.659
#> 133        NA         NA                <NA>        Ask  0.976804  15880.171
#> 134        NA         NA                <NA>        Ask  0.976943  15773.184
#> 135        NA         NA                <NA>        Ask  0.977060  14654.697
#> 136        NA         NA                <NA>        Ask  0.977483    149.041
#> 137        NA         NA                <NA>        Ask  0.977484  39764.400
#> 138        NA         NA                <NA>        Ask  0.977661   4783.602
#> 139        NA         NA                <NA>        Ask  0.977800    138.833
#> 140        NA         NA                <NA>        Ask  0.978916  18821.704
#> 141        NA         NA                <NA>        Ask  0.978927   2239.631
#> 142        NA         NA                <NA>        Ask  0.979556  37945.700
#> 143        NA         NA                <NA>        Ask  0.979605  11775.019
#> 144        NA         NA                <NA>        Ask  0.979774  14614.105
#> 145        NA         NA                <NA>        Ask  0.979828   2428.796
#> 146        NA         NA                <NA>        Ask  0.980021  24826.063
#> 147        NA         NA                <NA>        Ask  0.980129   3676.475
#> 148        NA         NA                <NA>        Ask  0.981063  29437.547
#> 149        NA         NA                <NA>        Ask  0.981686   8840.595
#> 150        NA         NA                <NA>        Ask  0.981687 288277.818
#> 151        NA         NA                <NA>        Ask  0.981969     24.102
#> 152        NA         NA                <NA>        Ask  0.982098    138.225
#> 153        NA         NA                <NA>        Ask  0.982281   4214.044
#> 154        NA         NA                <NA>        Ask  0.982698  14570.618
#> 155        NA         NA                <NA>        Ask  0.983114   3066.481
#> 156        NA         NA                <NA>        Ask  0.983536     99.610
#> 157        NA         NA                <NA>        Ask  0.984148  26581.100
#> 158        NA         NA                <NA>        Ask  0.985264     48.681
#> 159        NA         NA                <NA>        Ask  0.985574    797.384
#> 160        NA         NA                <NA>        Ask  0.986014     11.586
#> 161        NA         NA                <NA>        Ask  0.986320  12732.200
#> 162        NA         NA                <NA>        Ask  0.986415    137.620
#> 163        NA         NA                <NA>        Ask  0.987979     23.956
#> 164        NA         NA                <NA>        Ask  0.988856   1567.357
#> 165        NA         NA                <NA>        Ask  0.989036   3002.621
#> 166        NA         NA                <NA>        Ask  0.989299   7933.588
#> 167        NA         NA                <NA>        Ask  0.989585   6487.000
#> 168        NA         NA                <NA>        Ask  0.989620   1721.843
#> 169        NA         NA                <NA>        Ask  0.990616   5471.535
#> 170        NA         NA                <NA>        Ask  0.990737    794.682
#> 171        NA         NA                <NA>        Ask  0.990750    137.018
#> 172        NA         NA                <NA>        Ask  0.991454     11.529
#> 173        NA         NA                <NA>        Ask  0.992245     31.277
#> 174        NA         NA                <NA>        Ask  0.992952   6626.100
#> 175        NA         NA                <NA>        Ask  0.993136     55.456
#> 176        NA         NA                <NA>        Ask  0.993224     78.674
#> 177        NA         NA                <NA>        Ask  0.993306  14414.256
#> 178        NA         NA                <NA>        Ask  0.993634     48.271
#> 179        NA         NA                <NA>        Ask  0.994000     41.491
#> 180        NA         NA                <NA>        Ask  0.994025     23.810
#> 181        NA         NA                <NA>        Ask  0.994142     98.547
#> 182        NA         NA                <NA>        Ask  0.995104    136.419
#> 183        NA         NA                <NA>        Ask  0.996211     98.300
#> 184        NA         NA                <NA>        Ask  0.996922     11.466
#> 185        NA         NA                <NA>        Ask  0.998690      7.057
#> 186        NA         NA                <NA>        Ask  0.999478    135.822
#> 187        NA         NA                <NA>        Ask  1.000000   3562.826
#> 188        NA         NA                <NA>        Ask  1.000109     23.665
#> 189        NA         NA                <NA>        Ask  1.000980  24554.063
#> 190        NA         NA                <NA>        Ask  1.001000      8.018
#> 191        NA         NA                <NA>        Ask  1.001032  47995.564
#> 192        NA         NA                <NA>        Ask  1.001042   4679.420
#> 193        NA         NA                <NA>        Ask  1.001150   6390.800
#> 194        NA         NA                <NA>        Ask  1.001581    109.838
#> 195        NA         NA                <NA>        Ask  1.002077     47.865
#> 196        NA         NA                <NA>        Ask  1.002244     84.063
#> 197        NA         NA                <NA>        Ask  1.002364     21.780
#> 198        NA         NA                <NA>        Ask  1.002422     11.396
#> 199        NA         NA                <NA>        Ask  1.002521     18.217
#> 200        NA         NA                <NA>        Ask  1.003297     98.300
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:33:30
#> 102 2024-12-18 16:33:29
#> 103 2024-12-18 16:33:30
#> 104 2024-12-18 16:33:28
#> 105 2024-12-18 16:33:21
#> 106 2024-12-18 16:33:29
#> 107 2024-12-18 16:33:24
#> 108 2024-12-18 16:33:29
#> 109 2024-12-18 16:33:24
#> 110 2024-12-18 16:33:29
#> 111 2024-12-18 16:33:27
#> 112 2024-12-18 16:32:47
#> 113 2024-12-18 16:33:29
#> 114 2024-12-18 16:32:43
#> 115 2024-12-18 16:33:15
#> 116 2024-12-18 16:33:20
#> 117 2024-12-18 16:32:45
#> 118 2024-12-18 16:33:15
#> 119 2024-12-18 16:33:25
#> 120 2024-12-18 16:32:49
#> 121 2024-12-18 16:33:25
#> 122 2024-12-18 16:33:13
#> 123 2024-12-18 16:33:24
#> 124 2024-12-18 16:33:11
#> 125 2024-12-18 16:33:13
#> 126 2024-12-18 16:33:00
#> 127 2024-12-18 16:22:03
#> 128 2024-12-18 16:33:13
#> 129 2024-12-18 16:32:03
#> 130 2024-12-18 16:32:45
#> 131 2024-12-18 16:33:13
#> 132 2024-12-18 16:33:04
#> 133 2024-12-18 16:32:00
#> 134 2024-12-18 16:33:04
#> 135 2024-12-18 16:32:04
#> 136 2024-12-18 16:33:25
#> 137 2024-12-18 16:32:56
#> 138 2024-12-18 16:33:19
#> 139 2024-12-18 16:17:28
#> 140 2024-12-18 16:31:05
#> 141 2024-12-18 16:33:09
#> 142 2024-12-18 16:32:35
#> 143 2024-12-18 16:33:19
#> 144 2024-12-18 16:32:03
#> 145 2024-12-18 16:32:49
#> 146 2024-12-18 16:31:59
#> 147 2024-12-18 16:33:11
#> 148 2024-12-18 16:33:19
#> 149 2024-12-18 16:33:29
#> 150 2024-12-18 16:33:23
#> 151 2024-12-18 16:27:13
#> 152 2024-12-18 14:56:28
#> 153 2024-12-18 16:28:15
#> 154 2024-12-18 16:32:04
#> 155 2024-12-18 16:32:04
#> 156 2024-12-18 15:57:30
#> 157 2024-12-18 16:29:01
#> 158 2024-12-18 14:56:59
#> 159 2024-12-18 16:32:48
#> 160 2024-12-18 15:09:59
#> 161 2024-12-18 16:27:13
#> 162 2024-12-18 15:24:42
#> 163 2024-12-18 14:35:26
#> 164 2024-12-18 16:32:09
#> 165 2024-12-18 16:32:05
#> 166 2024-12-18 16:32:03
#> 167 2024-12-18 16:27:02
#> 168 2024-12-18 16:32:03
#> 169 2024-12-18 16:30:06
#> 170 2024-12-18 16:30:05
#> 171 2024-12-18 13:57:45
#> 172 2024-12-18 12:37:59
#> 173 2024-12-18 15:04:40
#> 174 2024-12-18 16:22:56
#> 175 2024-12-18 13:26:56
#> 176 2024-12-18 15:12:45
#> 177 2024-12-18 16:32:04
#> 178 2024-12-18 12:38:18
#> 179 2024-12-18 12:25:06
#> 180 2024-12-18 13:36:40
#> 181 2024-12-18 12:37:51
#> 182 2024-12-18 13:34:11
#> 183 2024-12-18 15:07:21
#> 184 2024-12-18 12:33:01
#> 185 2024-12-18 12:34:04
#> 186 2024-12-18 13:11:41
#> 187 2024-12-18 14:57:48
#> 188 2024-12-18 12:15:38
#> 189 2024-12-18 16:33:19
#> 190 2024-12-18 12:22:05
#> 191 2024-12-18 16:33:20
#> 192 2024-12-18 16:33:18
#> 193 2024-12-18 16:27:06
#> 194 2024-12-18 15:18:26
#> 195 2024-12-18 04:26:55
#> 196 2024-12-18 07:39:44
#> 197 2024-12-18 12:02:37
#> 198 2024-12-18 03:31:33
#> 199 2024-12-18 12:21:12
#> 200 2024-12-18 05:52:42
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.955890     31.385 2024-12-18 13:29:07        Bid        NA         NA
#> 2    0.955970     86.060 2024-12-11 11:33:28        Bid        NA         NA
#> 3    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 4    0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 5    0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 6    0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 7    0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 8    0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 9    0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 10   0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 11   0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 12   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 13   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 14   0.958430   5655.280 2024-12-18 16:30:01        Bid        NA         NA
#> 15   0.958711   1616.828 2024-12-18 16:30:01        Bid        NA         NA
#> 16   0.958713   4875.193 2024-12-18 16:30:01        Bid        NA         NA
#> 17   0.958753   9818.348 2024-12-18 16:30:02        Bid        NA         NA
#> 18   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 19   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 20   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 21   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 22   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 23   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 24   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 25   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 26   0.960231   6106.993 2024-12-18 16:29:35        Bid        NA         NA
#> 27   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 28   0.960274  14913.350 2024-12-18 16:32:48        Bid        NA         NA
#> 29   0.960714  12680.200 2024-12-18 16:29:32        Bid        NA         NA
#> 30   0.960863   3426.766 2024-12-18 16:30:51        Bid        NA         NA
#> 31   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 32   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 33   0.961036   3139.799 2024-12-18 16:33:21        Bid        NA         NA
#> 34   0.961463  59317.270 2024-12-18 16:23:08        Bid        NA         NA
#> 35   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 36   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 37   0.962591  20442.902 2024-12-18 16:33:19        Bid        NA         NA
#> 38   0.962592 287291.369 2024-12-18 16:33:31        Bid        NA         NA
#> 39   0.962609   3743.389 2024-12-18 16:33:15        Bid        NA         NA
#> 40   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 41   0.963156   2469.598 2024-12-18 16:32:03        Bid        NA         NA
#> 42   0.963352   2274.763 2024-12-18 16:32:37        Bid        NA         NA
#> 43   0.963611  14860.915 2024-12-18 16:33:03        Bid        NA         NA
#> 44   0.963760  12454.707 2024-12-18 16:30:33        Bid        NA         NA
#> 45   0.963965  25256.400 2024-12-18 16:30:25        Bid        NA         NA
#> 46   0.964049   8171.696 2024-12-18 16:33:19        Bid        NA         NA
#> 47   0.964161   9953.163 2024-12-18 16:33:23        Bid        NA         NA
#> 48   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 49   0.964818   8957.002 2024-12-18 16:33:23        Bid        NA         NA
#> 50   0.965675   3317.801 2024-12-18 16:33:24        Bid        NA         NA
#> 51   0.965839  25124.632 2024-12-18 16:30:16        Bid        NA         NA
#> 52   0.965887  14942.748 2024-12-18 16:33:23        Bid        NA         NA
#> 53   0.965888   1974.625 2024-12-18 16:33:31        Bid        NA         NA
#> 54   0.966006  14822.368 2024-12-18 16:32:04        Bid        NA         NA
#> 55   0.966041   5970.183 2024-12-18 16:33:23        Bid        NA         NA
#> 56   0.966054   4975.123 2024-12-18 16:33:23        Bid        NA         NA
#> 57   0.966055   6895.190 2024-12-18 16:33:30        Bid        NA         NA
#> 58   0.966396  38476.100 2024-12-18 16:30:25        Bid        NA         NA
#> 59   0.966444   6971.661 2024-12-18 16:33:23        Bid        NA         NA
#> 60   0.967067 204742.770 2024-12-18 16:32:03        Bid        NA         NA
#> 61   0.967115   2990.340 2024-12-18 16:33:11        Bid        NA         NA
#> 62   0.967321     43.755 2024-12-18 16:33:15        Bid        NA         NA
#> 63   0.967472  19207.356 2024-12-18 16:32:03        Bid        NA         NA
#> 64   0.967473     28.549 2024-12-18 16:32:50        Bid        NA         NA
#> 65   0.967949  17498.900 2024-12-18 16:33:10        Bid        NA         NA
#> 66   0.967992  15720.934 2024-12-18 16:32:04        Bid        NA         NA
#> 67   0.968018    289.433 2024-12-18 16:33:21        Bid        NA         NA
#> 68   0.968019   2369.393 2024-12-18 16:33:29        Bid        NA         NA
#> 69   0.968124  15189.107 2024-12-18 16:32:04        Bid        NA         NA
#> 70   0.968494  38248.600 2024-12-18 16:33:31        Bid        NA         NA
#> 71   0.968730  14780.631 2024-12-18 16:33:30        Bid        NA         NA
#> 72   0.969487     14.454 2024-12-18 16:32:03        Bid        NA         NA
#> 73   0.969567   5000.000 2024-12-18 16:33:30        Bid        NA         NA
#> 74   0.969671  14766.277 2024-12-18 16:33:30        Bid        NA         NA
#> 75   0.970022  14760.937 2024-12-18 16:33:33        Bid        NA         NA
#> 76   0.970196   2755.703 2024-12-18 16:33:30        Bid        NA         NA
#> 77   0.970209  20804.254 2024-12-18 16:33:30        Bid        NA         NA
#> 78   0.970224  10430.000 2024-12-18 16:33:30        Bid        NA         NA
#> 79   0.970380     49.000 2024-12-18 16:33:29        Bid        NA         NA
#> 80   0.970462  11058.570 2024-12-18 16:33:30        Bid        NA         NA
#> 81   0.970468   9700.800 2024-12-18 16:32:03        Bid        NA         NA
#> 82   0.970481  10700.630 2024-12-18 16:33:20        Bid        NA         NA
#> 83   0.970482    592.262 2024-12-18 16:33:29        Bid        NA         NA
#> 84   0.970578   9834.988 2024-12-18 16:33:32        Bid        NA         NA
#> 85   0.970808    640.300 2024-12-18 16:33:27        Bid        NA         NA
#> 86   0.970998   2446.000 2024-12-18 16:33:30        Bid        NA         NA
#> 87   0.971184   2938.000 2024-12-18 16:33:32        Bid        NA         NA
#> 88   0.971220   7566.874 2024-12-18 16:33:31        Bid        NA         NA
#> 89   0.971246   9828.223 2024-12-18 16:33:33        Bid        NA         NA
#> 90   0.971364  16025.100 2024-12-18 16:33:31        Bid        NA         NA
#> 91   0.971378     16.697 2024-12-18 16:33:30        Bid        NA         NA
#> 92   0.971415   5006.639 2024-12-18 16:33:32        Bid        NA         NA
#> 93   0.971418   7369.860 2024-12-18 16:33:33        Bid        NA         NA
#> 94   0.971628   5405.322 2024-12-18 16:33:31        Bid        NA         NA
#> 95   0.971629  11580.100 2024-12-18 16:33:32        Bid        NA         NA
#> 96   0.971658   2456.014 2024-12-18 16:33:33        Bid        NA         NA
#> 97   0.971659    301.991 2024-12-18 16:33:33        Bid        NA         NA
#> 98   0.971899    497.858 2024-12-18 16:33:32        Bid        NA         NA
#> 99   0.971903    640.100 2024-12-18 16:33:32        Bid        NA         NA
#> 100  0.972175   3576.965 2024-12-18 16:33:33        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.972647   1391.530
#> 102        NA         NA                <NA>        Ask  0.972648  24535.156
#> 103        NA         NA                <NA>        Ask  0.972650   1628.002
#> 104        NA         NA                <NA>        Ask  0.972652     87.466
#> 105        NA         NA                <NA>        Ask  0.972783   2100.000
#> 106        NA         NA                <NA>        Ask  0.972790   4906.315
#> 107        NA         NA                <NA>        Ask  0.972871  10000.000
#> 108        NA         NA                <NA>        Ask  0.973049   3576.965
#> 109        NA         NA                <NA>        Ask  0.973123  10000.000
#> 110        NA         NA                <NA>        Ask  0.973158    640.300
#> 111        NA         NA                <NA>        Ask  0.973186   2453.158
#> 112        NA         NA                <NA>        Ask  0.973244    263.439
#> 113        NA         NA                <NA>        Ask  0.973304  10705.500
#> 114        NA         NA                <NA>        Ask  0.973305   9807.447
#> 115        NA         NA                <NA>        Ask  0.973600   9719.700
#> 116        NA         NA                <NA>        Ask  0.973674   2127.859
#> 117        NA         NA                <NA>        Ask  0.973722   4839.316
#> 118        NA         NA                <NA>        Ask  0.973724   9571.550
#> 119        NA         NA                <NA>        Ask  0.973843  14703.044
#> 120        NA         NA                <NA>        Ask  0.974164    464.142
#> 121        NA         NA                <NA>        Ask  0.974167   2755.703
#> 122        NA         NA                <NA>        Ask  0.974186  15299.700
#> 123        NA         NA                <NA>        Ask  0.974187   1472.385
#> 124        NA         NA                <NA>        Ask  0.974188  14697.826
#> 125        NA         NA                <NA>        Ask  0.974246   8301.680
#> 126        NA         NA                <NA>        Ask  0.974296     14.769
#> 127        NA         NA                <NA>        Ask  0.974472   7782.520
#> 128        NA         NA                <NA>        Ask  0.974654   5000.000
#> 129        NA         NA                <NA>        Ask  0.974753    679.262
#> 130        NA         NA                <NA>        Ask  0.974797   2036.704
#> 131        NA         NA                <NA>        Ask  0.974919   3396.307
#> 132        NA         NA                <NA>        Ask  0.975153   1358.523
#> 133        NA         NA                <NA>        Ask  0.975154   2717.046
#> 134        NA         NA                <NA>        Ask  0.975209     43.973
#> 135        NA         NA                <NA>        Ask  0.975306  14264.489
#> 136        NA         NA                <NA>        Ask  0.975352  14680.283
#> 137        NA         NA                <NA>        Ask  0.975354  17495.400
#> 138        NA         NA                <NA>        Ask  0.975995     24.250
#> 139        NA         NA                <NA>        Ask  0.976054   6113.353
#> 140        NA         NA                <NA>        Ask  0.976484     10.229
#> 141        NA         NA                <NA>        Ask  0.976755    144.411
#> 142        NA         NA                <NA>        Ask  0.976756   6792.614
#> 143        NA         NA                <NA>        Ask  0.976803     27.659
#> 144        NA         NA                <NA>        Ask  0.976804  15880.171
#> 145        NA         NA                <NA>        Ask  0.976943  15773.184
#> 146        NA         NA                <NA>        Ask  0.977060  14654.697
#> 147        NA         NA                <NA>        Ask  0.977483    149.041
#> 148        NA         NA                <NA>        Ask  0.977484  39764.400
#> 149        NA         NA                <NA>        Ask  0.977800    138.833
#> 150        NA         NA                <NA>        Ask  0.978216   4785.250
#> 151        NA         NA                <NA>        Ask  0.978916  18821.704
#> 152        NA         NA                <NA>        Ask  0.979556  37945.700
#> 153        NA         NA                <NA>        Ask  0.979774  14614.105
#> 154        NA         NA                <NA>        Ask  0.979828   2428.796
#> 155        NA         NA                <NA>        Ask  0.980021  24826.063
#> 156        NA         NA                <NA>        Ask  0.980161  11779.076
#> 157        NA         NA                <NA>        Ask  0.980570   8834.608
#> 158        NA         NA                <NA>        Ask  0.980571   2235.878
#> 159        NA         NA                <NA>        Ask  0.981063  22078.161
#> 160        NA         NA                <NA>        Ask  0.981620   7361.922
#> 161        NA         NA                <NA>        Ask  0.981687 288277.818
#> 162        NA         NA                <NA>        Ask  0.981969     24.102
#> 163        NA         NA                <NA>        Ask  0.982098    138.225
#> 164        NA         NA                <NA>        Ask  0.982184   3670.483
#> 165        NA         NA                <NA>        Ask  0.982281   4214.044
#> 166        NA         NA                <NA>        Ask  0.982698  14570.618
#> 167        NA         NA                <NA>        Ask  0.983114   3066.481
#> 168        NA         NA                <NA>        Ask  0.983536     99.610
#> 169        NA         NA                <NA>        Ask  0.984148  26581.100
#> 170        NA         NA                <NA>        Ask  0.985264     48.681
#> 171        NA         NA                <NA>        Ask  0.985574    797.384
#> 172        NA         NA                <NA>        Ask  0.986014     11.586
#> 173        NA         NA                <NA>        Ask  0.986320  12732.200
#> 174        NA         NA                <NA>        Ask  0.986415    137.620
#> 175        NA         NA                <NA>        Ask  0.987979     23.956
#> 176        NA         NA                <NA>        Ask  0.988856   1567.357
#> 177        NA         NA                <NA>        Ask  0.989036   3002.621
#> 178        NA         NA                <NA>        Ask  0.989299   7933.588
#> 179        NA         NA                <NA>        Ask  0.989585   6487.000
#> 180        NA         NA                <NA>        Ask  0.989620   1721.843
#> 181        NA         NA                <NA>        Ask  0.990616   5471.535
#> 182        NA         NA                <NA>        Ask  0.990737    794.682
#> 183        NA         NA                <NA>        Ask  0.990750    137.018
#> 184        NA         NA                <NA>        Ask  0.991454     11.529
#> 185        NA         NA                <NA>        Ask  0.992245     31.277
#> 186        NA         NA                <NA>        Ask  0.992952   6626.100
#> 187        NA         NA                <NA>        Ask  0.993136     55.456
#> 188        NA         NA                <NA>        Ask  0.993224     78.674
#> 189        NA         NA                <NA>        Ask  0.993306  14414.256
#> 190        NA         NA                <NA>        Ask  0.993634     48.271
#> 191        NA         NA                <NA>        Ask  0.994000     41.491
#> 192        NA         NA                <NA>        Ask  0.994025     23.810
#> 193        NA         NA                <NA>        Ask  0.994142     98.547
#> 194        NA         NA                <NA>        Ask  0.995104    136.419
#> 195        NA         NA                <NA>        Ask  0.996211     98.300
#> 196        NA         NA                <NA>        Ask  0.996922     11.466
#> 197        NA         NA                <NA>        Ask  0.998690      7.057
#> 198        NA         NA                <NA>        Ask  0.999478    135.822
#> 199        NA         NA                <NA>        Ask  1.000000   3562.826
#> 200        NA         NA                <NA>        Ask  1.000109     23.665
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:33:33
#> 102 2024-12-18 16:33:33
#> 103 2024-12-18 16:33:33
#> 104 2024-12-18 16:33:33
#> 105 2024-12-18 16:33:32
#> 106 2024-12-18 16:33:33
#> 107 2024-12-18 16:33:33
#> 108 2024-12-18 16:33:33
#> 109 2024-12-18 16:33:32
#> 110 2024-12-18 16:33:30
#> 111 2024-12-18 16:33:32
#> 112 2024-12-18 16:33:30
#> 113 2024-12-18 16:33:32
#> 114 2024-12-18 16:33:32
#> 115 2024-12-18 16:33:20
#> 116 2024-12-18 16:33:32
#> 117 2024-12-18 16:33:30
#> 118 2024-12-18 16:33:30
#> 119 2024-12-18 16:33:32
#> 120 2024-12-18 16:33:33
#> 121 2024-12-18 16:33:30
#> 122 2024-12-18 16:33:31
#> 123 2024-12-18 16:33:31
#> 124 2024-12-18 16:33:25
#> 125 2024-12-18 16:33:30
#> 126 2024-12-18 16:33:30
#> 127 2024-12-18 16:33:32
#> 128 2024-12-18 16:33:31
#> 129 2024-12-18 16:33:31
#> 130 2024-12-18 16:33:31
#> 131 2024-12-18 16:33:30
#> 132 2024-12-18 16:33:30
#> 133 2024-12-18 16:33:30
#> 134 2024-12-18 16:33:11
#> 135 2024-12-18 16:33:13
#> 136 2024-12-18 16:33:32
#> 137 2024-12-18 16:33:00
#> 138 2024-12-18 16:22:03
#> 139 2024-12-18 16:33:13
#> 140 2024-12-18 16:32:03
#> 141 2024-12-18 16:32:45
#> 142 2024-12-18 16:33:13
#> 143 2024-12-18 16:33:04
#> 144 2024-12-18 16:32:00
#> 145 2024-12-18 16:33:04
#> 146 2024-12-18 16:32:04
#> 147 2024-12-18 16:33:25
#> 148 2024-12-18 16:32:56
#> 149 2024-12-18 16:17:28
#> 150 2024-12-18 16:33:30
#> 151 2024-12-18 16:31:05
#> 152 2024-12-18 16:32:35
#> 153 2024-12-18 16:32:03
#> 154 2024-12-18 16:32:49
#> 155 2024-12-18 16:31:59
#> 156 2024-12-18 16:33:30
#> 157 2024-12-18 16:33:32
#> 158 2024-12-18 16:33:30
#> 159 2024-12-18 16:33:30
#> 160 2024-12-18 16:33:30
#> 161 2024-12-18 16:33:23
#> 162 2024-12-18 16:27:13
#> 163 2024-12-18 14:56:28
#> 164 2024-12-18 16:33:31
#> 165 2024-12-18 16:28:15
#> 166 2024-12-18 16:32:04
#> 167 2024-12-18 16:32:04
#> 168 2024-12-18 15:57:30
#> 169 2024-12-18 16:29:01
#> 170 2024-12-18 14:56:59
#> 171 2024-12-18 16:32:48
#> 172 2024-12-18 15:09:59
#> 173 2024-12-18 16:27:13
#> 174 2024-12-18 15:24:42
#> 175 2024-12-18 14:35:26
#> 176 2024-12-18 16:32:09
#> 177 2024-12-18 16:32:05
#> 178 2024-12-18 16:32:03
#> 179 2024-12-18 16:27:02
#> 180 2024-12-18 16:32:03
#> 181 2024-12-18 16:30:06
#> 182 2024-12-18 16:30:05
#> 183 2024-12-18 13:57:45
#> 184 2024-12-18 12:37:59
#> 185 2024-12-18 15:04:40
#> 186 2024-12-18 16:22:56
#> 187 2024-12-18 13:26:56
#> 188 2024-12-18 15:12:45
#> 189 2024-12-18 16:32:04
#> 190 2024-12-18 12:38:18
#> 191 2024-12-18 12:25:06
#> 192 2024-12-18 13:36:40
#> 193 2024-12-18 12:37:51
#> 194 2024-12-18 13:34:11
#> 195 2024-12-18 15:07:21
#> 196 2024-12-18 12:33:01
#> 197 2024-12-18 12:34:04
#> 198 2024-12-18 13:11:41
#> 199 2024-12-18 14:57:48
#> 200 2024-12-18 12:15:38