Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.982933     20.348 2024-12-17 10:24:24        Bid        NA         NA
#> 2    0.983000     41.482 2024-12-17 01:19:43        Bid        NA         NA
#> 3    0.983015    178.496 2024-12-17 13:28:29        Bid        NA         NA
#> 4    0.983094     23.858 2024-12-17 07:40:57        Bid        NA         NA
#> 5    0.983599    100.000 2024-12-11 12:43:55        Bid        NA         NA
#> 6    0.984000     82.976 2024-12-17 22:49:05        Bid        NA         NA
#> 7    0.984046      7.129 2024-12-17 10:13:57        Bid        NA         NA
#> 8    0.985000   7279.604 2024-12-17 23:38:13        Bid        NA         NA
#> 9    0.985052   5494.553 2024-12-17 23:38:22        Bid        NA         NA
#> 10   0.985449     33.229 2024-12-17 15:52:42        Bid        NA         NA
#> 11   0.985479   8088.849 2024-12-18 00:03:13        Bid        NA         NA
#> 12   0.985655   6137.472 2024-12-17 23:44:41        Bid        NA         NA
#> 13   0.985843    136.501 2024-12-15 07:33:31        Bid        NA         NA
#> 14   0.985867    396.920 2024-12-17 20:05:34        Bid        NA         NA
#> 15   0.986000     82.976 2024-12-17 22:51:06        Bid        NA         NA
#> 16   0.986546    101.364 2024-12-11 14:23:38        Bid        NA         NA
#> 17   0.986564     11.482 2024-12-16 13:03:56        Bid        NA         NA
#> 18   0.986570     50.546 2024-12-14 13:18:42        Bid        NA         NA
#> 19   0.987000     82.965 2024-12-16 13:28:48        Bid        NA         NA
#> 20   0.987024   2181.400 2024-12-18 00:05:04        Bid        NA         NA
#> 21   0.987538     69.750 2024-12-17 23:18:42        Bid        NA         NA
#> 22   0.987922     55.501 2024-12-16 15:40:40        Bid        NA         NA
#> 23   0.987945   1577.609 2024-12-18 00:01:05        Bid        NA         NA
#> 24   0.988000     46.041 2024-12-17 15:16:51        Bid        NA         NA
#> 25   0.988468   3040.402 2024-12-18 00:06:28        Bid        NA         NA
#> 26   0.988532    227.611 2024-12-17 23:38:15        Bid        NA         NA
#> 27   0.988599  12946.700 2024-12-17 23:47:53        Bid        NA         NA
#> 28   0.988733     47.903 2024-12-17 19:22:43        Bid        NA         NA
#> 29   0.988956     18.254 2024-12-16 13:50:23        Bid        NA         NA
#> 30   0.989000     82.965 2024-12-17 23:27:51        Bid        NA         NA
#> 31   0.989111     23.713 2024-12-17 01:56:09        Bid        NA         NA
#> 32   0.989157   6368.241 2024-12-17 23:50:00        Bid        NA         NA
#> 33   0.989250     98.440 2024-12-17 23:08:38        Bid        NA         NA
#> 34   0.989291   1773.088 2024-12-18 00:06:30        Bid        NA         NA
#> 35   0.990000   3105.961 2024-12-17 23:55:56        Bid        NA         NA
#> 36   0.990176    135.903 2024-12-17 20:27:31        Bid        NA         NA
#> 37   0.990595    100.000 2024-12-16 14:10:33        Bid        NA         NA
#> 38   0.991000   6178.944 2024-12-17 23:32:16        Bid        NA         NA
#> 39   0.991246   1627.649 2024-12-18 00:06:26        Bid        NA         NA
#> 40   0.991500  21626.929 2024-12-18 00:05:26        Bid        NA         NA
#> 41   0.992000     82.976 2024-12-17 23:25:47        Bid        NA         NA
#> 42   0.992006     11.419 2024-12-17 23:15:47        Bid        NA         NA
#> 43   0.992011   3048.925 2024-12-18 00:06:24        Bid        NA         NA
#> 44   0.992064  12394.993 2024-12-18 00:01:15        Bid        NA         NA
#> 45   0.992245     31.277 2024-12-16 15:03:24        Bid        NA         NA
#> 46   0.992714  13628.592 2024-12-18 00:06:12        Bid        NA         NA
#> 47   0.992798  26744.900 2024-12-18 00:01:36        Bid        NA         NA
#> 48   0.993000    124.453 2024-12-17 23:15:53        Bid        NA         NA
#> 49   0.993001   1422.836 2024-12-18 00:06:42        Bid        NA         NA
#> 50   0.993558   4531.150 2024-12-18 00:06:19        Bid        NA         NA
#> 51   0.993596  14386.557 2024-12-18 00:06:05        Bid        NA         NA
#> 52   0.993971   2467.598 2024-12-18 00:06:27        Bid        NA         NA
#> 53   0.993972 289961.629 2024-12-18 00:06:42        Bid        NA         NA
#> 54   0.994376   2080.911 2024-12-18 00:06:28        Bid        NA         NA
#> 55   0.994402   7264.489 2024-12-18 00:06:14        Bid        NA         NA
#> 56   0.994528    135.308 2024-12-17 23:24:24        Bid        NA         NA
#> 57   0.994855   3485.491 2024-12-18 00:06:24        Bid        NA         NA
#> 58   0.995139  24775.057 2024-12-18 00:04:28        Bid        NA         NA
#> 59   0.995300    451.672 2024-12-17 23:54:32        Bid        NA         NA
#> 60   0.995319  14361.642 2024-12-18 00:06:28        Bid        NA         NA
#> 61   0.995824  39627.900 2024-12-18 00:05:19        Bid        NA         NA
#> 62   0.996000    502.011 2024-12-17 23:22:21        Bid        NA         NA
#> 63   0.997252  37915.800 2024-12-18 00:06:12        Bid        NA         NA
#> 64   0.997643  15367.627 2024-12-18 00:04:28        Bid        NA         NA
#> 65   0.997659 198464.479 2024-12-18 00:05:27        Bid        NA         NA
#> 66   0.997692   3956.794 2024-12-18 00:06:38        Bid        NA         NA
#> 67   0.998185  18623.279 2024-12-18 00:06:41        Bid        NA         NA
#> 68   0.998404  15457.197 2024-12-18 00:06:00        Bid        NA         NA
#> 69   0.998452  11858.544 2024-12-18 00:06:38        Bid        NA         NA
#> 70   0.998453   6631.489 2024-12-18 00:06:45        Bid        NA         NA
#> 71   0.998587    161.921 2024-12-18 00:06:08        Bid        NA         NA
#> 72   0.999005  14308.661 2024-12-18 00:06:44        Bid        NA         NA
#> 73   0.999085    553.418 2024-12-18 00:04:04        Bid        NA         NA
#> 74   0.999399    161.917 2024-12-18 00:06:00        Bid        NA         NA
#> 75   0.999627   1000.000 2024-12-18 00:03:23        Bid        NA         NA
#> 76   0.999629    401.710 2024-12-17 23:57:08        Bid        NA         NA
#> 77   0.999676   5928.706 2024-12-18 00:06:38        Bid        NA         NA
#> 78   1.000000   5000.000 2024-12-18 00:04:40        Bid        NA         NA
#> 79   1.000051     43.976 2024-12-18 00:06:44        Bid        NA         NA
#> 80   1.000761   2278.778 2024-12-18 00:06:44        Bid        NA         NA
#> 81   1.000813     14.200 2024-12-18 00:06:27        Bid        NA         NA
#> 82   1.001352  14275.117 2024-12-18 00:06:44        Bid        NA         NA
#> 83   1.001353   5000.000 2024-12-18 00:06:44        Bid        NA         NA
#> 84   1.002329  14261.204 2024-12-18 00:06:42        Bid        NA         NA
#> 85   1.002730  14255.498 2024-12-18 00:06:44        Bid        NA         NA
#> 86   1.002882   2648.646 2024-12-18 00:06:44        Bid        NA         NA
#> 87   1.002999   4963.098 2024-12-18 00:06:42        Bid        NA         NA
#> 88   1.003157   1000.000 2024-12-18 00:06:34        Bid        NA         NA
#> 89   1.003350    378.913 2024-12-18 00:06:42        Bid        NA         NA
#> 90   1.003370    169.000 2024-12-18 00:06:41        Bid        NA         NA
#> 91   1.003466    219.105 2024-12-18 00:06:42        Bid        NA         NA
#> 92   1.003468   4748.342 2024-12-18 00:06:44        Bid        NA         NA
#> 93   1.003579  18051.600 2024-12-18 00:06:45        Bid        NA         NA
#> 94   1.003580    473.779 2024-12-18 00:06:45        Bid        NA         NA
#> 95   1.003825   2373.000 2024-12-18 00:06:44        Bid        NA         NA
#> 96   1.003826   4751.000 2024-12-18 00:06:44        Bid        NA         NA
#> 97   1.004061   9491.071 2024-12-18 00:06:45        Bid        NA         NA
#> 98   1.004084  23727.144 2024-12-18 00:06:44        Bid        NA         NA
#> 99   1.004086  10192.000 2024-12-18 00:06:45        Bid        NA         NA
#> 100  1.004087    708.934 2024-12-18 00:06:45        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  1.005000  10399.551
#> 102        NA         NA                <NA>        Ask  1.005906   3791.483
#> 103        NA         NA                <NA>        Ask  1.005994   2368.211
#> 104        NA         NA                <NA>        Ask  1.006058   4736.119
#> 105        NA         NA                <NA>        Ask  1.006106   6323.715
#> 106        NA         NA                <NA>        Ask  1.006107   7103.833
#> 107        NA         NA                <NA>        Ask  1.006226   9470.661
#> 108        NA         NA                <NA>        Ask  1.006230     23.521
#> 109        NA         NA                <NA>        Ask  1.006343     78.789
#> 110        NA         NA                <NA>        Ask  1.006344  15476.500
#> 111        NA         NA                <NA>        Ask  1.006582   4166.996
#> 112        NA         NA                <NA>        Ask  1.006583   9467.303
#> 113        NA         NA                <NA>        Ask  1.007139   9462.076
#> 114        NA         NA                <NA>        Ask  1.007462   1421.329
#> 115        NA         NA                <NA>        Ask  1.007463   5993.671
#> 116        NA         NA                <NA>        Ask  1.007464   4974.715
#> 117        NA         NA                <NA>        Ask  1.007790    697.823
#> 118        NA         NA                <NA>        Ask  1.007791  14183.931
#> 119        NA         NA                <NA>        Ask  1.007951     11.334
#> 120        NA         NA                <NA>        Ask  1.007985    167.764
#> 121        NA         NA                <NA>        Ask  1.008000     41.490
#> 122        NA         NA                <NA>        Ask  1.008035  19346.979
#> 123        NA         NA                <NA>        Ask  1.008072     84.973
#> 124        NA         NA                <NA>        Ask  1.008073     15.962
#> 125        NA         NA                <NA>        Ask  1.008100    162.448
#> 126        NA         NA                <NA>        Ask  1.008187     46.853
#> 127        NA         NA                <NA>        Ask  1.008209    671.054
#> 128        NA         NA                <NA>        Ask  1.008251  13675.610
#> 129        NA         NA                <NA>        Ask  1.008252  14177.444
#> 130        NA         NA                <NA>        Ask  1.008260  15587.200
#> 131        NA         NA                <NA>        Ask  1.008270  15567.917
#> 132        NA         NA                <NA>        Ask  1.008283    134.636
#> 133        NA         NA                <NA>        Ask  1.008559      7.045
#> 134        NA         NA                <NA>        Ask  1.008799   1174.343
#> 135        NA         NA                <NA>        Ask  1.008849   5000.000
#> 136        NA         NA                <NA>        Ask  1.008850  38526.800
#> 137        NA         NA                <NA>        Ask  1.008912  14168.175
#> 138        NA         NA                <NA>        Ask  1.009270  19199.892
#> 139        NA         NA                <NA>        Ask  1.009928   1677.633
#> 140        NA         NA                <NA>        Ask  1.010000     41.490
#> 141        NA         NA                <NA>        Ask  1.010590     47.461
#> 142        NA         NA                <NA>        Ask  1.010610  25570.621
#> 143        NA         NA                <NA>        Ask  1.010759  14142.281
#> 144        NA         NA                <NA>        Ask  1.011340  38227.800
#> 145        NA         NA                <NA>        Ask  1.011405   4619.499
#> 146        NA         NA                <NA>        Ask  1.012389     23.378
#> 147        NA         NA                <NA>        Ask  1.012590  39266.445
#> 148        NA         NA                <NA>        Ask  1.012715    134.046
#> 149        NA         NA                <NA>        Ask  1.013511     11.272
#> 150        NA         NA                <NA>        Ask  1.013519  11370.223
#> 151        NA         NA                <NA>        Ask  1.013630  14102.214
#> 152        NA         NA                <NA>        Ask  1.014231  21337.627
#> 153        NA         NA                <NA>        Ask  1.014360   6346.471
#> 154        NA         NA                <NA>        Ask  1.014843   8527.667
#> 155        NA         NA                <NA>        Ask  1.015027   7106.389
#> 156        NA         NA                <NA>        Ask  1.015091 288153.221
#> 157        NA         NA                <NA>        Ask  1.015092   2038.443
#> 158        NA         NA                <NA>        Ask  1.015525   3414.547
#> 159        NA         NA                <NA>        Ask  1.015702     96.455
#> 160        NA         NA                <NA>        Ask  1.016210  12623.000
#> 161        NA         NA                <NA>        Ask  1.016353    793.573
#> 162        NA         NA                <NA>        Ask  1.016431   2413.072
#> 163        NA         NA                <NA>        Ask  1.016747  14058.994
#> 164        NA         NA                <NA>        Ask  1.017167    133.460
#> 165        NA         NA                <NA>        Ask  1.018203   2951.377
#> 166        NA         NA                <NA>        Ask  1.018526      6.976
#> 167        NA         NA                <NA>        Ask  1.018585     23.236
#> 168        NA         NA                <NA>        Ask  1.018621   6859.598
#> 169        NA         NA                <NA>        Ask  1.019102     11.210
#> 170        NA         NA                <NA>        Ask  1.019176     47.062
#> 171        NA         NA                <NA>        Ask  1.019299   7823.583
#> 172        NA         NA                <NA>        Ask  1.019327   2967.454
#> 173        NA         NA                <NA>        Ask  1.020000  19900.498
#> 174        NA         NA                <NA>        Ask  1.020510   6459.700
#> 175        NA         NA                <NA>        Ask  1.020770   1475.062
#> 176        NA         NA                <NA>        Ask  1.021485    784.109
#> 177        NA         NA                <NA>        Ask  1.021637    132.876
#> 178        NA         NA                <NA>        Ask  1.022213   1715.279
#> 179        NA         NA                <NA>        Ask  1.023778    517.398
#> 180        NA         NA                <NA>        Ask  1.024614     82.228
#> 181        NA         NA                <NA>        Ask  1.024723     11.148
#> 182        NA         NA                <NA>        Ask  1.024819     23.095
#> 183        NA         NA                <NA>        Ask  1.025559     78.674
#> 184        NA         NA                <NA>        Ask  1.025960   6513.000
#> 185        NA         NA                <NA>        Ask  1.026127    132.294
#> 186        NA         NA                <NA>        Ask  1.026248  13928.832
#> 187        NA         NA                <NA>        Ask  1.026657     95.426
#> 188        NA         NA                <NA>        Ask  1.027347     31.277
#> 189        NA         NA                <NA>        Ask  1.027836     46.665
#> 190        NA         NA                <NA>        Ask  1.028592      6.852
#> 191        NA         NA                <NA>        Ask  1.029336     19.041
#> 192        NA         NA                <NA>        Ask  1.030637    131.715
#> 193        NA         NA                <NA>        Ask  1.030850   6475.300
#> 194        NA         NA                <NA>        Ask  1.031092     22.954
#> 195        NA         NA                <NA>        Ask  1.032000   4970.179
#> 196        NA         NA                <NA>        Ask  1.032133     16.392
#> 197        NA         NA                <NA>        Ask  1.032148     98.300
#> 198        NA         NA                <NA>        Ask  1.032519  23759.228
#> 199        NA         NA                <NA>        Ask  1.032861    115.107
#> 200        NA         NA                <NA>        Ask  1.034250   1114.538
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 00:05:06
#> 102 2024-12-18 00:06:45
#> 103 2024-12-18 00:06:45
#> 104 2024-12-18 00:06:45
#> 105 2024-12-18 00:06:43
#> 106 2024-12-18 00:06:45
#> 107 2024-12-18 00:06:45
#> 108 2024-12-18 00:05:15
#> 109 2024-12-18 00:06:44
#> 110 2024-12-18 00:06:42
#> 111 2024-12-18 00:06:45
#> 112 2024-12-18 00:06:44
#> 113 2024-12-18 00:06:44
#> 114 2024-12-18 00:06:44
#> 115 2024-12-18 00:06:44
#> 116 2024-12-18 00:06:44
#> 117 2024-12-18 00:06:45
#> 118 2024-12-18 00:06:42
#> 119 2024-12-17 23:33:40
#> 120 2024-12-18 00:06:35
#> 121 2024-12-17 23:38:39
#> 122 2024-12-18 00:06:26
#> 123 2024-12-18 00:06:28
#> 124 2024-12-18 00:06:24
#> 125 2024-12-18 00:06:28
#> 126 2024-12-18 00:06:43
#> 127 2024-12-18 00:06:35
#> 128 2024-12-18 00:06:45
#> 129 2024-12-18 00:06:44
#> 130 2024-12-18 00:06:24
#> 131 2024-12-18 00:06:24
#> 132 2024-12-17 23:02:03
#> 133 2024-12-17 23:25:04
#> 134 2024-12-18 00:06:44
#> 135 2024-12-18 00:06:28
#> 136 2024-12-18 00:06:24
#> 137 2024-12-18 00:06:42
#> 138 2024-12-18 00:06:24
#> 139 2024-12-18 00:06:35
#> 140 2024-12-17 23:49:43
#> 141 2024-12-17 23:02:28
#> 142 2024-12-18 00:04:35
#> 143 2024-12-18 00:06:28
#> 144 2024-12-18 00:06:24
#> 145 2024-12-18 00:06:44
#> 146 2024-12-17 23:24:22
#> 147 2024-12-18 00:00:27
#> 148 2024-12-17 22:46:02
#> 149 2024-12-17 22:46:00
#> 150 2024-12-18 00:06:25
#> 151 2024-12-18 00:06:42
#> 152 2024-12-18 00:06:24
#> 153 2024-12-17 23:38:16
#> 154 2024-12-18 00:06:25
#> 155 2024-12-18 00:06:25
#> 156 2024-12-18 00:06:45
#> 157 2024-12-18 00:06:43
#> 158 2024-12-18 00:06:44
#> 159 2024-12-17 23:02:18
#> 160 2024-12-18 00:01:19
#> 161 2024-12-18 00:00:29
#> 162 2024-12-18 00:06:44
#> 163 2024-12-18 00:06:24
#> 164 2024-12-17 22:44:21
#> 165 2024-12-18 00:03:23
#> 166 2024-12-17 22:49:16
#> 167 2024-12-17 22:44:18
#> 168 2024-12-17 23:54:08
#> 169 2024-12-17 22:41:15
#> 170 2024-12-17 22:46:29
#> 171 2024-12-17 23:47:52
#> 172 2024-12-18 00:06:44
#> 173 2024-12-17 23:52:53
#> 174 2024-12-17 23:38:18
#> 175 2024-12-17 23:38:19
#> 176 2024-12-17 23:38:27
#> 177 2024-12-17 22:40:07
#> 178 2024-12-18 00:01:04
#> 179 2024-12-17 23:26:16
#> 180 2024-12-17 23:02:16
#> 181 2024-12-17 23:11:35
#> 182 2024-12-17 22:23:52
#> 183 2024-12-17 23:18:42
#> 184 2024-12-17 23:18:09
#> 185 2024-12-17 22:13:28
#> 186 2024-12-18 00:06:24
#> 187 2024-12-17 22:25:29
#> 188 2024-12-17 22:47:29
#> 189 2024-12-17 22:23:45
#> 190 2024-12-17 22:23:54
#> 191 2024-12-17 22:48:13
#> 192 2024-12-17 21:54:10
#> 193 2024-12-17 23:03:43
#> 194 2024-12-17 20:50:20
#> 195 2024-12-17 23:21:27
#> 196 2024-12-17 23:57:55
#> 197 2024-12-17 23:05:47
#> 198 2024-12-18 00:06:05
#> 199 2024-12-17 23:02:14
#> 200 2024-12-18 00:06:35
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.982933     20.348 2024-12-17 10:24:24        Bid        NA         NA
#> 2    0.983000     41.482 2024-12-17 01:19:43        Bid        NA         NA
#> 3    0.983015    178.496 2024-12-17 13:28:29        Bid        NA         NA
#> 4    0.983094     23.858 2024-12-17 07:40:57        Bid        NA         NA
#> 5    0.983599    100.000 2024-12-11 12:43:55        Bid        NA         NA
#> 6    0.984000     82.976 2024-12-17 22:49:05        Bid        NA         NA
#> 7    0.984046      7.129 2024-12-17 10:13:57        Bid        NA         NA
#> 8    0.985000   7279.604 2024-12-17 23:38:13        Bid        NA         NA
#> 9    0.985052   5494.553 2024-12-17 23:38:22        Bid        NA         NA
#> 10   0.985449     33.229 2024-12-17 15:52:42        Bid        NA         NA
#> 11   0.985479   8088.849 2024-12-18 00:03:13        Bid        NA         NA
#> 12   0.985655   6137.472 2024-12-17 23:44:41        Bid        NA         NA
#> 13   0.985843    136.501 2024-12-15 07:33:31        Bid        NA         NA
#> 14   0.985867    396.920 2024-12-17 20:05:34        Bid        NA         NA
#> 15   0.986000     82.976 2024-12-17 22:51:06        Bid        NA         NA
#> 16   0.986546    101.364 2024-12-11 14:23:38        Bid        NA         NA
#> 17   0.986564     11.482 2024-12-16 13:03:56        Bid        NA         NA
#> 18   0.986570     50.546 2024-12-14 13:18:42        Bid        NA         NA
#> 19   0.987000     82.965 2024-12-16 13:28:48        Bid        NA         NA
#> 20   0.987024   2181.400 2024-12-18 00:05:04        Bid        NA         NA
#> 21   0.987538     69.750 2024-12-17 23:18:42        Bid        NA         NA
#> 22   0.987922     55.501 2024-12-16 15:40:40        Bid        NA         NA
#> 23   0.987945   1577.609 2024-12-18 00:01:05        Bid        NA         NA
#> 24   0.988000     46.041 2024-12-17 15:16:51        Bid        NA         NA
#> 25   0.988468   3040.402 2024-12-18 00:06:28        Bid        NA         NA
#> 26   0.988532    227.611 2024-12-17 23:38:15        Bid        NA         NA
#> 27   0.988599  12946.700 2024-12-17 23:47:53        Bid        NA         NA
#> 28   0.988733     47.903 2024-12-17 19:22:43        Bid        NA         NA
#> 29   0.988956     18.254 2024-12-16 13:50:23        Bid        NA         NA
#> 30   0.989000     82.965 2024-12-17 23:27:51        Bid        NA         NA
#> 31   0.989111     23.713 2024-12-17 01:56:09        Bid        NA         NA
#> 32   0.989157   6368.241 2024-12-17 23:50:00        Bid        NA         NA
#> 33   0.989250     98.440 2024-12-17 23:08:38        Bid        NA         NA
#> 34   0.989291   1773.088 2024-12-18 00:06:30        Bid        NA         NA
#> 35   0.990000   3105.961 2024-12-17 23:55:56        Bid        NA         NA
#> 36   0.990176    135.903 2024-12-17 20:27:31        Bid        NA         NA
#> 37   0.990595    100.000 2024-12-16 14:10:33        Bid        NA         NA
#> 38   0.991000   6178.944 2024-12-17 23:32:16        Bid        NA         NA
#> 39   0.991246   1627.649 2024-12-18 00:06:26        Bid        NA         NA
#> 40   0.991500  21626.929 2024-12-18 00:05:26        Bid        NA         NA
#> 41   0.992000     82.976 2024-12-17 23:25:47        Bid        NA         NA
#> 42   0.992006     11.419 2024-12-17 23:15:47        Bid        NA         NA
#> 43   0.992011   3048.925 2024-12-18 00:06:24        Bid        NA         NA
#> 44   0.992064  12394.993 2024-12-18 00:01:15        Bid        NA         NA
#> 45   0.992245     31.277 2024-12-16 15:03:24        Bid        NA         NA
#> 46   0.992714  13628.592 2024-12-18 00:06:12        Bid        NA         NA
#> 47   0.992798  26744.900 2024-12-18 00:01:36        Bid        NA         NA
#> 48   0.993000    124.453 2024-12-17 23:15:53        Bid        NA         NA
#> 49   0.993001   1422.836 2024-12-18 00:06:42        Bid        NA         NA
#> 50   0.993558   4531.150 2024-12-18 00:06:19        Bid        NA         NA
#> 51   0.993596  14386.557 2024-12-18 00:06:05        Bid        NA         NA
#> 52   0.993971   2467.598 2024-12-18 00:06:27        Bid        NA         NA
#> 53   0.993972 289961.629 2024-12-18 00:06:42        Bid        NA         NA
#> 54   0.994376   2080.911 2024-12-18 00:06:28        Bid        NA         NA
#> 55   0.994402   7264.489 2024-12-18 00:06:14        Bid        NA         NA
#> 56   0.994528    135.308 2024-12-17 23:24:24        Bid        NA         NA
#> 57   0.994855   3485.491 2024-12-18 00:06:24        Bid        NA         NA
#> 58   0.995139  24775.057 2024-12-18 00:04:28        Bid        NA         NA
#> 59   0.995300    451.672 2024-12-17 23:54:32        Bid        NA         NA
#> 60   0.995319  14361.642 2024-12-18 00:06:28        Bid        NA         NA
#> 61   0.995824  39627.900 2024-12-18 00:05:19        Bid        NA         NA
#> 62   0.996000    502.011 2024-12-17 23:22:21        Bid        NA         NA
#> 63   0.997252  37915.800 2024-12-18 00:06:12        Bid        NA         NA
#> 64   0.997643  15367.627 2024-12-18 00:04:28        Bid        NA         NA
#> 65   0.997659 198464.479 2024-12-18 00:05:27        Bid        NA         NA
#> 66   0.997692   3956.794 2024-12-18 00:06:38        Bid        NA         NA
#> 67   0.998185  18623.279 2024-12-18 00:06:41        Bid        NA         NA
#> 68   0.998404  15457.197 2024-12-18 00:06:00        Bid        NA         NA
#> 69   0.998452  11858.544 2024-12-18 00:06:38        Bid        NA         NA
#> 70   0.998453   6631.489 2024-12-18 00:06:45        Bid        NA         NA
#> 71   0.998587    161.921 2024-12-18 00:06:08        Bid        NA         NA
#> 72   0.999005  14308.661 2024-12-18 00:06:44        Bid        NA         NA
#> 73   0.999085    553.418 2024-12-18 00:04:04        Bid        NA         NA
#> 74   0.999399    161.917 2024-12-18 00:06:00        Bid        NA         NA
#> 75   0.999627   1000.000 2024-12-18 00:03:23        Bid        NA         NA
#> 76   0.999629    401.710 2024-12-17 23:57:08        Bid        NA         NA
#> 77   0.999676   5928.706 2024-12-18 00:06:38        Bid        NA         NA
#> 78   1.000000   5000.000 2024-12-18 00:04:40        Bid        NA         NA
#> 79   1.000051     43.976 2024-12-18 00:06:44        Bid        NA         NA
#> 80   1.000761   2278.778 2024-12-18 00:06:44        Bid        NA         NA
#> 81   1.000813     14.200 2024-12-18 00:06:27        Bid        NA         NA
#> 82   1.001352  14275.117 2024-12-18 00:06:44        Bid        NA         NA
#> 83   1.001353   5000.000 2024-12-18 00:06:44        Bid        NA         NA
#> 84   1.002329  14261.204 2024-12-18 00:06:42        Bid        NA         NA
#> 85   1.002730  14255.498 2024-12-18 00:06:44        Bid        NA         NA
#> 86   1.002882   2648.646 2024-12-18 00:06:44        Bid        NA         NA
#> 87   1.002999   4963.098 2024-12-18 00:06:42        Bid        NA         NA
#> 88   1.003157   1000.000 2024-12-18 00:06:34        Bid        NA         NA
#> 89   1.003350    378.913 2024-12-18 00:06:42        Bid        NA         NA
#> 90   1.003370    169.000 2024-12-18 00:06:41        Bid        NA         NA
#> 91   1.003466    219.105 2024-12-18 00:06:42        Bid        NA         NA
#> 92   1.003468   4748.342 2024-12-18 00:06:44        Bid        NA         NA
#> 93   1.003579  18051.600 2024-12-18 00:06:45        Bid        NA         NA
#> 94   1.003580    473.779 2024-12-18 00:06:45        Bid        NA         NA
#> 95   1.003825   2373.000 2024-12-18 00:06:44        Bid        NA         NA
#> 96   1.003826   4751.000 2024-12-18 00:06:44        Bid        NA         NA
#> 97   1.004061   9491.071 2024-12-18 00:06:45        Bid        NA         NA
#> 98   1.004084  23727.144 2024-12-18 00:06:44        Bid        NA         NA
#> 99   1.004086  10192.000 2024-12-18 00:06:45        Bid        NA         NA
#> 100  1.004087    708.934 2024-12-18 00:06:45        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  1.005000  10399.551
#> 102        NA         NA                <NA>        Ask  1.005906   3791.483
#> 103        NA         NA                <NA>        Ask  1.005994   2368.211
#> 104        NA         NA                <NA>        Ask  1.006058   4736.119
#> 105        NA         NA                <NA>        Ask  1.006106   6323.715
#> 106        NA         NA                <NA>        Ask  1.006107   7103.833
#> 107        NA         NA                <NA>        Ask  1.006226   9470.661
#> 108        NA         NA                <NA>        Ask  1.006230     23.521
#> 109        NA         NA                <NA>        Ask  1.006343     78.789
#> 110        NA         NA                <NA>        Ask  1.006344  15476.500
#> 111        NA         NA                <NA>        Ask  1.006582   4166.996
#> 112        NA         NA                <NA>        Ask  1.006583   9467.303
#> 113        NA         NA                <NA>        Ask  1.007139   9462.076
#> 114        NA         NA                <NA>        Ask  1.007462   1421.329
#> 115        NA         NA                <NA>        Ask  1.007463   5993.671
#> 116        NA         NA                <NA>        Ask  1.007464   4974.715
#> 117        NA         NA                <NA>        Ask  1.007790    697.823
#> 118        NA         NA                <NA>        Ask  1.007791  14183.931
#> 119        NA         NA                <NA>        Ask  1.007951     11.334
#> 120        NA         NA                <NA>        Ask  1.007985    167.764
#> 121        NA         NA                <NA>        Ask  1.008000     41.490
#> 122        NA         NA                <NA>        Ask  1.008035  19346.979
#> 123        NA         NA                <NA>        Ask  1.008072     84.973
#> 124        NA         NA                <NA>        Ask  1.008073     15.962
#> 125        NA         NA                <NA>        Ask  1.008100    162.448
#> 126        NA         NA                <NA>        Ask  1.008187     46.853
#> 127        NA         NA                <NA>        Ask  1.008209    671.054
#> 128        NA         NA                <NA>        Ask  1.008251  13675.610
#> 129        NA         NA                <NA>        Ask  1.008252  14177.444
#> 130        NA         NA                <NA>        Ask  1.008260  15587.200
#> 131        NA         NA                <NA>        Ask  1.008270  15567.917
#> 132        NA         NA                <NA>        Ask  1.008283    134.636
#> 133        NA         NA                <NA>        Ask  1.008559      7.045
#> 134        NA         NA                <NA>        Ask  1.008799   1174.343
#> 135        NA         NA                <NA>        Ask  1.008849   5000.000
#> 136        NA         NA                <NA>        Ask  1.008850  38526.800
#> 137        NA         NA                <NA>        Ask  1.008912  14168.175
#> 138        NA         NA                <NA>        Ask  1.009270  19199.892
#> 139        NA         NA                <NA>        Ask  1.009928   1677.633
#> 140        NA         NA                <NA>        Ask  1.010000     41.490
#> 141        NA         NA                <NA>        Ask  1.010590     47.461
#> 142        NA         NA                <NA>        Ask  1.010610  25570.621
#> 143        NA         NA                <NA>        Ask  1.010759  14142.281
#> 144        NA         NA                <NA>        Ask  1.011340  38227.800
#> 145        NA         NA                <NA>        Ask  1.011405   4619.499
#> 146        NA         NA                <NA>        Ask  1.012389     23.378
#> 147        NA         NA                <NA>        Ask  1.012590  39266.445
#> 148        NA         NA                <NA>        Ask  1.012715    134.046
#> 149        NA         NA                <NA>        Ask  1.013511     11.272
#> 150        NA         NA                <NA>        Ask  1.013519  11370.223
#> 151        NA         NA                <NA>        Ask  1.013630  14102.214
#> 152        NA         NA                <NA>        Ask  1.014231  21337.627
#> 153        NA         NA                <NA>        Ask  1.014360   6346.471
#> 154        NA         NA                <NA>        Ask  1.014843   8527.667
#> 155        NA         NA                <NA>        Ask  1.015027   7106.389
#> 156        NA         NA                <NA>        Ask  1.015091 288153.221
#> 157        NA         NA                <NA>        Ask  1.015092   2038.443
#> 158        NA         NA                <NA>        Ask  1.015525   3414.547
#> 159        NA         NA                <NA>        Ask  1.015702     96.455
#> 160        NA         NA                <NA>        Ask  1.016210  12623.000
#> 161        NA         NA                <NA>        Ask  1.016353    793.573
#> 162        NA         NA                <NA>        Ask  1.016431   2413.072
#> 163        NA         NA                <NA>        Ask  1.016747  14058.994
#> 164        NA         NA                <NA>        Ask  1.017167    133.460
#> 165        NA         NA                <NA>        Ask  1.018203   2951.377
#> 166        NA         NA                <NA>        Ask  1.018526      6.976
#> 167        NA         NA                <NA>        Ask  1.018585     23.236
#> 168        NA         NA                <NA>        Ask  1.018621   6859.598
#> 169        NA         NA                <NA>        Ask  1.019102     11.210
#> 170        NA         NA                <NA>        Ask  1.019176     47.062
#> 171        NA         NA                <NA>        Ask  1.019299   7823.583
#> 172        NA         NA                <NA>        Ask  1.019327   2967.454
#> 173        NA         NA                <NA>        Ask  1.020000  19900.498
#> 174        NA         NA                <NA>        Ask  1.020510   6459.700
#> 175        NA         NA                <NA>        Ask  1.020770   1475.062
#> 176        NA         NA                <NA>        Ask  1.021485    784.109
#> 177        NA         NA                <NA>        Ask  1.021637    132.876
#> 178        NA         NA                <NA>        Ask  1.022213   1715.279
#> 179        NA         NA                <NA>        Ask  1.023778    517.398
#> 180        NA         NA                <NA>        Ask  1.024614     82.228
#> 181        NA         NA                <NA>        Ask  1.024723     11.148
#> 182        NA         NA                <NA>        Ask  1.024819     23.095
#> 183        NA         NA                <NA>        Ask  1.025559     78.674
#> 184        NA         NA                <NA>        Ask  1.025960   6513.000
#> 185        NA         NA                <NA>        Ask  1.026127    132.294
#> 186        NA         NA                <NA>        Ask  1.026248  13928.832
#> 187        NA         NA                <NA>        Ask  1.026657     95.426
#> 188        NA         NA                <NA>        Ask  1.027347     31.277
#> 189        NA         NA                <NA>        Ask  1.027836     46.665
#> 190        NA         NA                <NA>        Ask  1.028592      6.852
#> 191        NA         NA                <NA>        Ask  1.029336     19.041
#> 192        NA         NA                <NA>        Ask  1.030637    131.715
#> 193        NA         NA                <NA>        Ask  1.030850   6475.300
#> 194        NA         NA                <NA>        Ask  1.031092     22.954
#> 195        NA         NA                <NA>        Ask  1.032000   4970.179
#> 196        NA         NA                <NA>        Ask  1.032133     16.392
#> 197        NA         NA                <NA>        Ask  1.032148     98.300
#> 198        NA         NA                <NA>        Ask  1.032519  23759.228
#> 199        NA         NA                <NA>        Ask  1.032861    115.107
#> 200        NA         NA                <NA>        Ask  1.034250   1114.538
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 00:05:06
#> 102 2024-12-18 00:06:45
#> 103 2024-12-18 00:06:45
#> 104 2024-12-18 00:06:45
#> 105 2024-12-18 00:06:43
#> 106 2024-12-18 00:06:45
#> 107 2024-12-18 00:06:45
#> 108 2024-12-18 00:05:15
#> 109 2024-12-18 00:06:44
#> 110 2024-12-18 00:06:42
#> 111 2024-12-18 00:06:45
#> 112 2024-12-18 00:06:44
#> 113 2024-12-18 00:06:44
#> 114 2024-12-18 00:06:44
#> 115 2024-12-18 00:06:44
#> 116 2024-12-18 00:06:44
#> 117 2024-12-18 00:06:45
#> 118 2024-12-18 00:06:42
#> 119 2024-12-17 23:33:40
#> 120 2024-12-18 00:06:35
#> 121 2024-12-17 23:38:39
#> 122 2024-12-18 00:06:26
#> 123 2024-12-18 00:06:28
#> 124 2024-12-18 00:06:24
#> 125 2024-12-18 00:06:28
#> 126 2024-12-18 00:06:43
#> 127 2024-12-18 00:06:35
#> 128 2024-12-18 00:06:45
#> 129 2024-12-18 00:06:44
#> 130 2024-12-18 00:06:24
#> 131 2024-12-18 00:06:24
#> 132 2024-12-17 23:02:03
#> 133 2024-12-17 23:25:04
#> 134 2024-12-18 00:06:44
#> 135 2024-12-18 00:06:28
#> 136 2024-12-18 00:06:24
#> 137 2024-12-18 00:06:42
#> 138 2024-12-18 00:06:24
#> 139 2024-12-18 00:06:35
#> 140 2024-12-17 23:49:43
#> 141 2024-12-17 23:02:28
#> 142 2024-12-18 00:04:35
#> 143 2024-12-18 00:06:28
#> 144 2024-12-18 00:06:24
#> 145 2024-12-18 00:06:44
#> 146 2024-12-17 23:24:22
#> 147 2024-12-18 00:00:27
#> 148 2024-12-17 22:46:02
#> 149 2024-12-17 22:46:00
#> 150 2024-12-18 00:06:25
#> 151 2024-12-18 00:06:42
#> 152 2024-12-18 00:06:24
#> 153 2024-12-17 23:38:16
#> 154 2024-12-18 00:06:25
#> 155 2024-12-18 00:06:25
#> 156 2024-12-18 00:06:45
#> 157 2024-12-18 00:06:43
#> 158 2024-12-18 00:06:44
#> 159 2024-12-17 23:02:18
#> 160 2024-12-18 00:01:19
#> 161 2024-12-18 00:00:29
#> 162 2024-12-18 00:06:44
#> 163 2024-12-18 00:06:24
#> 164 2024-12-17 22:44:21
#> 165 2024-12-18 00:03:23
#> 166 2024-12-17 22:49:16
#> 167 2024-12-17 22:44:18
#> 168 2024-12-17 23:54:08
#> 169 2024-12-17 22:41:15
#> 170 2024-12-17 22:46:29
#> 171 2024-12-17 23:47:52
#> 172 2024-12-18 00:06:44
#> 173 2024-12-17 23:52:53
#> 174 2024-12-17 23:38:18
#> 175 2024-12-17 23:38:19
#> 176 2024-12-17 23:38:27
#> 177 2024-12-17 22:40:07
#> 178 2024-12-18 00:01:04
#> 179 2024-12-17 23:26:16
#> 180 2024-12-17 23:02:16
#> 181 2024-12-17 23:11:35
#> 182 2024-12-17 22:23:52
#> 183 2024-12-17 23:18:42
#> 184 2024-12-17 23:18:09
#> 185 2024-12-17 22:13:28
#> 186 2024-12-18 00:06:24
#> 187 2024-12-17 22:25:29
#> 188 2024-12-17 22:47:29
#> 189 2024-12-17 22:23:45
#> 190 2024-12-17 22:23:54
#> 191 2024-12-17 22:48:13
#> 192 2024-12-17 21:54:10
#> 193 2024-12-17 23:03:43
#> 194 2024-12-17 20:50:20
#> 195 2024-12-17 23:21:27
#> 196 2024-12-17 23:57:55
#> 197 2024-12-17 23:05:47
#> 198 2024-12-18 00:06:05
#> 199 2024-12-17 23:02:14
#> 200 2024-12-18 00:06:35