Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.956188  14983.509 2024-12-18 17:01:51        Bid        NA         NA
#> 2    0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 3    0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 4    0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 5    0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 6    0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 7    0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 8    0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 9    0.958214   6118.811 2024-12-18 17:03:05        Bid        NA         NA
#> 10   0.958582   6347.900 2024-12-18 17:03:03        Bid        NA         NA
#> 11   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 12   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 13   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 14   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 15   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 16   0.960000  62991.628 2024-12-18 17:03:33        Bid        NA         NA
#> 17   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 18   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 19   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 20   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 21   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 22   0.961804   1915.541 2024-12-18 17:03:43        Bid        NA         NA
#> 23   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 24   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 25   0.962405   6482.323 2024-12-18 16:58:00        Bid        NA         NA
#> 26   0.962407   8177.747 2024-12-18 16:58:00        Bid        NA         NA
#> 27   0.962599   5642.132 2024-12-18 16:40:45        Bid        NA         NA
#> 28   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 29   0.963484   1636.437 2024-12-18 17:02:52        Bid        NA         NA
#> 30   0.963555   6106.993 2024-12-18 17:03:03        Bid        NA         NA
#> 31   0.963801   3021.197 2024-12-18 17:04:44        Bid        NA         NA
#> 32   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 33   0.964609  12680.200 2024-12-18 17:02:50        Bid        NA         NA
#> 34   0.964757  14847.878 2024-12-18 17:02:59        Bid        NA         NA
#> 35   0.965000    145.118 2024-12-18 16:58:19        Bid        NA         NA
#> 36   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 37   0.965269 288858.056 2024-12-18 17:04:39        Bid        NA         NA
#> 38   0.966552  25256.400 2024-12-18 17:02:53        Bid        NA         NA
#> 39   0.966602   5014.813 2024-12-18 17:04:02        Bid        NA         NA
#> 40   0.966664   2185.264 2024-12-18 17:04:45        Bid        NA         NA
#> 41   0.966701  12454.707 2024-12-18 17:02:52        Bid        NA         NA
#> 42   0.966738   8081.048 2024-12-18 17:03:07        Bid        NA         NA
#> 43   0.966795   3574.139 2024-12-18 17:04:16        Bid        NA         NA
#> 44   0.967133  15016.492 2024-12-18 17:04:02        Bid        NA         NA
#> 45   0.967134   6926.894 2024-12-18 17:04:44        Bid        NA         NA
#> 46   0.967660   2356.647 2024-12-18 17:03:41        Bid        NA         NA
#> 47   0.967925  14800.060 2024-12-18 17:03:39        Bid        NA         NA
#> 48   0.968700    138.916 2024-12-18 16:38:53        Bid        NA         NA
#> 49   0.969031   8087.693 2024-12-18 17:04:34        Bid        NA         NA
#> 50   0.969879   3242.344 2024-12-18 17:04:44        Bid        NA         NA
#> 51   0.970040  25124.632 2024-12-18 17:03:52        Bid        NA         NA
#> 52   0.970088  13744.696 2024-12-18 17:04:34        Bid        NA         NA
#> 53   0.970110  38476.100 2024-12-18 17:04:07        Bid        NA         NA
#> 54   0.970223   5661.137 2024-12-18 17:04:34        Bid        NA         NA
#> 55   0.970770  14756.674 2024-12-18 17:03:42        Bid        NA         NA
#> 56   0.970818   4851.971 2024-12-18 17:04:34        Bid        NA         NA
#> 57   0.971014 203910.440 2024-12-18 17:03:48        Bid        NA         NA
#> 58   0.971062   4042.459 2024-12-18 17:04:34        Bid        NA         NA
#> 59   0.971102  15720.934 2024-12-18 17:02:59        Bid        NA         NA
#> 60   0.971150   3233.819 2024-12-18 17:04:34        Bid        NA         NA
#> 61   0.971361  38248.600 2024-12-18 17:03:41        Bid        NA         NA
#> 62   0.971409   2425.364 2024-12-18 17:04:34        Bid        NA         NA
#> 63   0.971528  19207.356 2024-12-18 17:03:54        Bid        NA         NA
#> 64   0.971529     16.780 2024-12-18 17:04:07        Bid        NA         NA
#> 65   0.971576   2424.793 2024-12-18 17:04:34        Bid        NA         NA
#> 66   0.972265  58652.063 2024-12-18 17:03:05        Bid        NA         NA
#> 67   0.972266    286.291 2024-12-18 17:04:29        Bid        NA         NA
#> 68   0.972422  15189.107 2024-12-18 17:03:43        Bid        NA         NA
#> 69   0.972423     16.684 2024-12-18 17:04:38        Bid        NA         NA
#> 70   0.972675     43.591 2024-12-18 17:04:27        Bid        NA         NA
#> 71   0.972703  14725.667 2024-12-18 17:04:24        Bid        NA         NA
#> 72   0.972766     25.700 2024-12-18 16:35:22        Bid        NA         NA
#> 73   0.972932  17412.200 2024-12-18 17:04:24        Bid        NA         NA
#> 74   0.972933   2357.426 2024-12-18 17:04:26        Bid        NA         NA
#> 75   0.973623  14710.977 2024-12-18 17:04:43        Bid        NA         NA
#> 76   0.973632     14.454 2024-12-18 17:04:28        Bid        NA         NA
#> 77   0.973894   2755.703 2024-12-18 17:04:16        Bid        NA         NA
#> 78   0.974035  14704.759 2024-12-18 17:04:38        Bid        NA         NA
#> 79   0.974106  10532.180 2024-12-18 17:04:27        Bid        NA         NA
#> 80   0.974107   5000.000 2024-12-18 17:04:27        Bid        NA         NA
#> 81   0.974323  20423.127 2024-12-18 17:04:24        Bid        NA         NA
#> 82   0.974385   7566.874 2024-12-18 17:04:43        Bid        NA         NA
#> 83   0.974390   2440.000 2024-12-18 17:04:28        Bid        NA         NA
#> 84   0.974435  10630.960 2024-12-18 17:04:27        Bid        NA         NA
#> 85   0.974499   2933.000 2024-12-18 17:04:39        Bid        NA         NA
#> 86   0.974529   9798.202 2024-12-18 17:04:43        Bid        NA         NA
#> 87   0.974616  16025.100 2024-12-18 17:04:43        Bid        NA         NA
#> 88   0.974720  10090.978 2024-12-18 17:04:44        Bid        NA         NA
#> 89   0.974769     16.698 2024-12-18 17:04:44        Bid        NA         NA
#> 90   0.974903   9794.441 2024-12-18 17:04:44        Bid        NA         NA
#> 91   0.974930  10723.550 2024-12-18 17:04:34        Bid        NA         NA
#> 92   0.975058    132.085 2024-12-18 17:04:28        Bid        NA         NA
#> 93   0.975229   9791.164 2024-12-18 17:04:44        Bid        NA         NA
#> 94   0.975240   9683.500 2024-12-18 17:04:13        Bid        NA         NA
#> 95   0.975379   9789.659 2024-12-18 17:04:43        Bid        NA         NA
#> 96   0.975407   4894.693 2024-12-18 17:04:44        Bid        NA         NA
#> 97   0.975463   2447.207 2024-12-18 17:04:44        Bid        NA         NA
#> 98   0.975464   6491.225 2024-12-18 17:04:44        Bid        NA         NA
#> 99   0.975465     81.800 2024-12-18 17:04:44        Bid        NA         NA
#> 100  0.975760    564.600 2024-12-18 17:04:43        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.976470   1373.036
#> 102        NA         NA                <NA>        Ask  0.976471   5088.876
#> 103        NA         NA                <NA>        Ask  0.976473   5876.592
#> 104        NA         NA                <NA>        Ask  0.976480     33.799
#> 105        NA         NA                <NA>        Ask  0.976482  34225.133
#> 106        NA         NA                <NA>        Ask  0.976491     79.863
#> 107        NA         NA                <NA>        Ask  0.976493   1000.000
#> 108        NA         NA                <NA>        Ask  0.976547   1628.002
#> 109        NA         NA                <NA>        Ask  0.976549   2100.000
#> 110        NA         NA                <NA>        Ask  0.976558    134.363
#> 111        NA         NA                <NA>        Ask  0.976563     33.799
#> 112        NA         NA                <NA>        Ask  0.976617    564.600
#> 113        NA         NA                <NA>        Ask  0.976718   3576.965
#> 114        NA         NA                <NA>        Ask  0.976778    386.277
#> 115        NA         NA                <NA>        Ask  0.976851    848.255
#> 116        NA         NA                <NA>        Ask  0.977028   9773.149
#> 117        NA         NA                <NA>        Ask  0.977044   4839.316
#> 118        NA         NA                <NA>        Ask  0.977045  10705.500
#> 119        NA         NA                <NA>        Ask  0.977365   4584.244
#> 120        NA         NA                <NA>        Ask  0.977366   1375.600
#> 121        NA         NA                <NA>        Ask  0.977595  15299.700
#> 122        NA         NA                <NA>        Ask  0.977620   9401.560
#> 123        NA         NA                <NA>        Ask  0.977697     14.769
#> 124        NA         NA                <NA>        Ask  0.977720  14649.342
#> 125        NA         NA                <NA>        Ask  0.977776   2933.000
#> 126        NA         NA                <NA>        Ask  0.977777   7005.859
#> 127        NA         NA                <NA>        Ask  0.977821   7922.600
#> 128        NA         NA                <NA>        Ask  0.977947    437.232
#> 129        NA         NA                <NA>        Ask  0.977949   8234.030
#> 130        NA         NA                <NA>        Ask  0.977980   1466.524
#> 131        NA         NA                <NA>        Ask  0.978050  14644.412
#> 132        NA         NA                <NA>        Ask  0.978155   9677.700
#> 133        NA         NA                <NA>        Ask  0.978190   2755.703
#> 134        NA         NA                <NA>        Ask  0.979011  14630.037
#> 135        NA         NA                <NA>        Ask  0.979054   5000.000
#> 136        NA         NA                <NA>        Ask  0.979492   1310.562
#> 137        NA         NA                <NA>        Ask  0.979541   8518.653
#> 138        NA         NA                <NA>        Ask  0.979885   8518.653
#> 139        NA         NA                <NA>        Ask  0.979933  17434.500
#> 140        NA         NA                <NA>        Ask  0.980171     43.948
#> 141        NA         NA                <NA>        Ask  0.980658     10.229
#> 142        NA         NA                <NA>        Ask  0.980839    149.516
#> 143        NA         NA                <NA>        Ask  0.980840  11139.777
#> 144        NA         NA                <NA>        Ask  0.980889  15773.184
#> 145        NA         NA                <NA>        Ask  0.980958     22.839
#> 146        NA         NA                <NA>        Ask  0.981008  15880.171
#> 147        NA         NA                <NA>        Ask  0.981118  39764.400
#> 148        NA         NA                <NA>        Ask  0.981339   6552.810
#> 149        NA         NA                <NA>        Ask  0.981388  14595.365
#> 150        NA         NA                <NA>        Ask  0.981969     24.102
#> 151        NA         NA                <NA>        Ask  0.982000     41.491
#> 152        NA         NA                <NA>        Ask  0.982098    138.225
#> 153        NA         NA                <NA>        Ask  0.982713   4764.380
#> 154        NA         NA                <NA>        Ask  0.983159  18821.704
#> 155        NA         NA                <NA>        Ask  0.983206  37945.700
#> 156        NA         NA                <NA>        Ask  0.983536     99.610
#> 157        NA         NA                <NA>        Ask  0.983995  24826.063
#> 158        NA         NA                <NA>        Ask  0.984101   2140.622
#> 159        NA         NA                <NA>        Ask  0.984856  14547.372
#> 160        NA         NA                <NA>        Ask  0.985148    150.059
#> 161        NA         NA                <NA>        Ask  0.985264     48.681
#> 162        NA         NA                <NA>        Ask  0.985456   2314.089
#> 163        NA         NA                <NA>        Ask  0.985771  21997.142
#> 164        NA         NA                <NA>        Ask  0.986014     11.586
#> 165        NA         NA                <NA>        Ask  0.986130 288999.356
#> 166        NA         NA                <NA>        Ask  0.986131  11712.482
#> 167        NA         NA                <NA>        Ask  0.986190   3513.840
#> 168        NA         NA                <NA>        Ask  0.986328   7328.436
#> 169        NA         NA                <NA>        Ask  0.986415    137.620
#> 170        NA         NA                <NA>        Ask  0.986688   8794.123
#> 171        NA         NA                <NA>        Ask  0.986689  26581.100
#> 172        NA         NA                <NA>        Ask  0.986725  14518.081
#> 173        NA         NA                <NA>        Ask  0.987632   4214.044
#> 174        NA         NA                <NA>        Ask  0.987979     23.956
#> 175        NA         NA                <NA>        Ask  0.988354   2930.312
#> 176        NA         NA                <NA>        Ask  0.989835  12732.200
#> 177        NA         NA                <NA>        Ask  0.990094    798.591
#> 178        NA         NA                <NA>        Ask  0.990750    137.018
#> 179        NA         NA                <NA>        Ask  0.991454     11.529
#> 180        NA         NA                <NA>        Ask  0.991905   5473.257
#> 181        NA         NA                <NA>        Ask  0.992245     31.277
#> 182        NA         NA                <NA>        Ask  0.993136     55.456
#> 183        NA         NA                <NA>        Ask  0.993224     78.674
#> 184        NA         NA                <NA>        Ask  0.993634     48.271
#> 185        NA         NA                <NA>        Ask  0.994000     41.491
#> 186        NA         NA                <NA>        Ask  0.994025     23.810
#> 187        NA         NA                <NA>        Ask  0.994142     98.547
#> 188        NA         NA                <NA>        Ask  0.994377  14189.146
#> 189        NA         NA                <NA>        Ask  0.994575   6487.000
#> 190        NA         NA                <NA>        Ask  0.995104    136.419
#> 191        NA         NA                <NA>        Ask  0.996072  14381.984
#> 192        NA         NA                <NA>        Ask  0.996087    793.275
#> 193        NA         NA                <NA>        Ask  0.996211     98.300
#> 194        NA         NA                <NA>        Ask  0.996922     11.466
#> 195        NA         NA                <NA>        Ask  0.998690      7.057
#> 196        NA         NA                <NA>        Ask  0.999478    135.822
#> 197        NA         NA                <NA>        Ask  1.000000   3562.826
#> 198        NA         NA                <NA>        Ask  1.000109     23.665
#> 199        NA         NA                <NA>        Ask  1.000330   6626.100
#> 200        NA         NA                <NA>        Ask  1.001000      8.018
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:04:44
#> 102 2024-12-18 17:04:44
#> 103 2024-12-18 17:04:44
#> 104 2024-12-18 17:04:44
#> 105 2024-12-18 17:04:44
#> 106 2024-12-18 17:04:44
#> 107 2024-12-18 17:04:44
#> 108 2024-12-18 17:04:44
#> 109 2024-12-18 17:04:44
#> 110 2024-12-18 17:04:43
#> 111 2024-12-18 17:04:42
#> 112 2024-12-18 17:04:43
#> 113 2024-12-18 17:04:42
#> 114 2024-12-18 17:04:17
#> 115 2024-12-18 17:04:43
#> 116 2024-12-18 17:04:44
#> 117 2024-12-18 17:04:43
#> 118 2024-12-18 17:04:43
#> 119 2024-12-18 17:04:38
#> 120 2024-12-18 17:04:23
#> 121 2024-12-18 17:04:43
#> 122 2024-12-18 17:04:27
#> 123 2024-12-18 17:04:44
#> 124 2024-12-18 17:04:43
#> 125 2024-12-18 17:04:24
#> 126 2024-12-18 17:04:24
#> 127 2024-12-18 17:04:43
#> 128 2024-12-18 17:04:43
#> 129 2024-12-18 17:04:27
#> 130 2024-12-18 17:04:34
#> 131 2024-12-18 17:04:39
#> 132 2024-12-18 17:03:50
#> 133 2024-12-18 17:04:13
#> 134 2024-12-18 17:04:44
#> 135 2024-12-18 17:04:28
#> 136 2024-12-18 17:04:24
#> 137 2024-12-18 17:04:36
#> 138 2024-12-18 17:04:36
#> 139 2024-12-18 17:03:40
#> 140 2024-12-18 17:04:23
#> 141 2024-12-18 17:04:28
#> 142 2024-12-18 17:04:42
#> 143 2024-12-18 17:04:36
#> 144 2024-12-18 17:04:18
#> 145 2024-12-18 17:04:35
#> 146 2024-12-18 17:03:16
#> 147 2024-12-18 17:04:33
#> 148 2024-12-18 17:04:36
#> 149 2024-12-18 17:04:27
#> 150 2024-12-18 16:41:15
#> 151 2024-12-18 16:51:06
#> 152 2024-12-18 17:04:34
#> 153 2024-12-18 17:04:07
#> 154 2024-12-18 17:04:44
#> 155 2024-12-18 17:04:22
#> 156 2024-12-18 15:57:30
#> 157 2024-12-18 17:03:44
#> 158 2024-12-18 17:04:27
#> 159 2024-12-18 17:01:46
#> 160 2024-12-18 17:04:19
#> 161 2024-12-18 14:56:59
#> 162 2024-12-18 17:03:52
#> 163 2024-12-18 17:04:09
#> 164 2024-12-18 15:09:59
#> 165 2024-12-18 17:04:33
#> 166 2024-12-18 17:01:28
#> 167 2024-12-18 17:04:41
#> 168 2024-12-18 17:04:09
#> 169 2024-12-18 16:58:52
#> 170 2024-12-18 17:04:09
#> 171 2024-12-18 17:03:24
#> 172 2024-12-18 17:03:40
#> 173 2024-12-18 17:02:50
#> 174 2024-12-18 14:35:26
#> 175 2024-12-18 17:03:43
#> 176 2024-12-18 17:02:07
#> 177 2024-12-18 16:54:05
#> 178 2024-12-18 13:57:45
#> 179 2024-12-18 12:37:59
#> 180 2024-12-18 17:03:07
#> 181 2024-12-18 15:04:40
#> 182 2024-12-18 13:26:56
#> 183 2024-12-18 15:12:45
#> 184 2024-12-18 12:38:18
#> 185 2024-12-18 12:25:06
#> 186 2024-12-18 13:36:40
#> 187 2024-12-18 12:37:51
#> 188 2024-12-18 16:57:57
#> 189 2024-12-18 16:40:42
#> 190 2024-12-18 13:34:11
#> 191 2024-12-18 17:02:53
#> 192 2024-12-18 16:38:48
#> 193 2024-12-18 15:07:21
#> 194 2024-12-18 12:33:01
#> 195 2024-12-18 12:34:04
#> 196 2024-12-18 13:11:41
#> 197 2024-12-18 14:57:48
#> 198 2024-12-18 12:15:38
#> 199 2024-12-18 16:37:45
#> 200 2024-12-18 12:22:05
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 2    0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 3    0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 4    0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 5    0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 6    0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 7    0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 8    0.958214   6118.811 2024-12-18 17:03:05        Bid        NA         NA
#> 9    0.958582   6347.900 2024-12-18 17:03:03        Bid        NA         NA
#> 10   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 11   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 12   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 13   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 14   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 15   0.960000  62991.628 2024-12-18 17:03:33        Bid        NA         NA
#> 16   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 17   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 18   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 19   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 20   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 21   0.961804   1915.541 2024-12-18 17:03:43        Bid        NA         NA
#> 22   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 23   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 24   0.962405   6482.323 2024-12-18 16:58:00        Bid        NA         NA
#> 25   0.962407   8177.747 2024-12-18 16:58:00        Bid        NA         NA
#> 26   0.962599   5642.132 2024-12-18 16:40:45        Bid        NA         NA
#> 27   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 28   0.963484   1636.437 2024-12-18 17:02:52        Bid        NA         NA
#> 29   0.963555   6106.993 2024-12-18 17:03:03        Bid        NA         NA
#> 30   0.963801   3021.197 2024-12-18 17:04:44        Bid        NA         NA
#> 31   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 32   0.964609  12680.200 2024-12-18 17:02:50        Bid        NA         NA
#> 33   0.964757  14847.878 2024-12-18 17:02:59        Bid        NA         NA
#> 34   0.965000    145.118 2024-12-18 16:58:19        Bid        NA         NA
#> 35   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 36   0.965269 288858.056 2024-12-18 17:04:39        Bid        NA         NA
#> 37   0.966552  25256.400 2024-12-18 17:02:53        Bid        NA         NA
#> 38   0.966602   5014.813 2024-12-18 17:04:02        Bid        NA         NA
#> 39   0.966664   2185.264 2024-12-18 17:04:45        Bid        NA         NA
#> 40   0.966701  12454.707 2024-12-18 17:02:52        Bid        NA         NA
#> 41   0.966738   8081.048 2024-12-18 17:03:07        Bid        NA         NA
#> 42   0.966795   3574.139 2024-12-18 17:04:16        Bid        NA         NA
#> 43   0.967133  15016.492 2024-12-18 17:04:02        Bid        NA         NA
#> 44   0.967660   2356.647 2024-12-18 17:03:41        Bid        NA         NA
#> 45   0.967925  14800.060 2024-12-18 17:03:39        Bid        NA         NA
#> 46   0.967973   8091.303 2024-12-18 17:04:46        Bid        NA         NA
#> 47   0.967974   6920.883 2024-12-18 17:04:47        Bid        NA         NA
#> 48   0.968700    138.916 2024-12-18 16:38:53        Bid        NA         NA
#> 49   0.969073   7282.052 2024-12-18 17:04:46        Bid        NA         NA
#> 50   0.969657   3246.534 2024-12-18 17:04:46        Bid        NA         NA
#> 51   0.969927   5663.458 2024-12-18 17:04:46        Bid        NA         NA
#> 52   0.970040  25124.632 2024-12-18 17:03:52        Bid        NA         NA
#> 53   0.970088   6472.549 2024-12-18 17:04:46        Bid        NA         NA
#> 54   0.970110  38476.100 2024-12-18 17:04:07        Bid        NA         NA
#> 55   0.970158   8899.337 2024-12-18 17:04:46        Bid        NA         NA
#> 56   0.970770  14756.674 2024-12-18 17:03:42        Bid        NA         NA
#> 57   0.970818   7280.817 2024-12-18 17:04:46        Bid        NA         NA
#> 58   0.971014 203910.440 2024-12-18 17:03:48        Bid        NA         NA
#> 59   0.971062    808.907 2024-12-18 17:04:46        Bid        NA         NA
#> 60   0.971102  15720.934 2024-12-18 17:02:59        Bid        NA         NA
#> 61   0.971361  38248.600 2024-12-18 17:03:41        Bid        NA         NA
#> 62   0.971528  19207.356 2024-12-18 17:03:54        Bid        NA         NA
#> 63   0.971529    151.875 2024-12-18 17:04:49        Bid        NA         NA
#> 64   0.972155  14733.199 2024-12-18 17:04:45        Bid        NA         NA
#> 65   0.972265  58652.063 2024-12-18 17:03:05        Bid        NA         NA
#> 66   0.972266    152.558 2024-12-18 17:04:48        Bid        NA         NA
#> 67   0.972422  15189.107 2024-12-18 17:03:43        Bid        NA         NA
#> 68   0.972423   2375.346 2024-12-18 17:04:49        Bid        NA         NA
#> 69   0.972675     43.591 2024-12-18 17:04:27        Bid        NA         NA
#> 70   0.972766     25.700 2024-12-18 16:35:22        Bid        NA         NA
#> 71   0.972932  23903.425 2024-12-18 17:04:49        Bid        NA         NA
#> 72   0.973248  14716.641 2024-12-18 17:04:45        Bid        NA         NA
#> 73   0.973336   2755.703 2024-12-18 17:04:45        Bid        NA         NA
#> 74   0.973623  14710.977 2024-12-18 17:04:43        Bid        NA         NA
#> 75   0.973632     14.454 2024-12-18 17:04:28        Bid        NA         NA
#> 76   0.973650  20423.127 2024-12-18 17:04:45        Bid        NA         NA
#> 77   0.973908   7566.874 2024-12-18 17:04:49        Bid        NA         NA
#> 78   0.974004   9803.479 2024-12-18 17:04:49        Bid        NA         NA
#> 79   0.974106  10532.180 2024-12-18 17:04:27        Bid        NA         NA
#> 80   0.974107   5000.000 2024-12-18 17:04:27        Bid        NA         NA
#> 81   0.974140  16025.100 2024-12-18 17:04:49        Bid        NA         NA
#> 82   0.974216  10885.340 2024-12-18 17:04:45        Bid        NA         NA
#> 83   0.974217   9725.933 2024-12-18 17:04:49        Bid        NA         NA
#> 84   0.974218    519.557 2024-12-18 17:04:50        Bid        NA         NA
#> 85   0.974311   9677.600 2024-12-18 17:04:49        Bid        NA         NA
#> 86   0.974367   2440.000 2024-12-18 17:04:50        Bid        NA         NA
#> 87   0.974435  10630.960 2024-12-18 17:04:27        Bid        NA         NA
#> 88   0.974472     16.698 2024-12-18 17:04:49        Bid        NA         NA
#> 89   0.974484   4888.000 2024-12-18 17:04:50        Bid        NA         NA
#> 90   0.974720    365.045 2024-12-18 17:04:49        Bid        NA         NA
#> 91   0.974949   9793.980 2024-12-18 17:04:49        Bid        NA         NA
#> 92   0.974950     70.217 2024-12-18 17:04:49        Bid        NA         NA
#> 93   0.975083     33.799 2024-12-18 17:04:50        Bid        NA         NA
#> 94   0.975097   2098.800 2024-12-18 17:04:49        Bid        NA         NA
#> 95   0.975098   3576.965 2024-12-18 17:04:50        Bid        NA         NA
#> 96   0.975117   5876.592 2024-12-18 17:04:50        Bid        NA         NA
#> 97   0.975213   4895.665 2024-12-18 17:04:49        Bid        NA         NA
#> 98   0.975241    564.700 2024-12-18 17:04:49        Bid        NA         NA
#> 99   0.975242   2447.761 2024-12-18 17:04:49        Bid        NA         NA
#> 100  0.975243     81.850 2024-12-18 17:04:49        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.975690   1373.824
#> 102        NA         NA                <NA>        Ask  0.975691   2100.000
#> 103        NA         NA                <NA>        Ask  0.975698   2446.617
#> 104        NA         NA                <NA>        Ask  0.975784   1628.002
#> 105        NA         NA                <NA>        Ask  0.975805   7339.053
#> 106        NA         NA                <NA>        Ask  0.975836     79.863
#> 107        NA         NA                <NA>        Ask  0.975874     33.799
#> 108        NA         NA                <NA>        Ask  0.975880   3576.965
#> 109        NA         NA                <NA>        Ask  0.975894  10000.000
#> 110        NA         NA                <NA>        Ask  0.975993   5876.592
#> 111        NA         NA                <NA>        Ask  0.976049    363.124
#> 112        NA         NA                <NA>        Ask  0.976106   9782.386
#> 113        NA         NA                <NA>        Ask  0.976202    564.700
#> 114        NA         NA                <NA>        Ask  0.976544    848.255
#> 115        NA         NA                <NA>        Ask  0.976549   2100.000
#> 116        NA         NA                <NA>        Ask  0.976558    134.363
#> 117        NA         NA                <NA>        Ask  0.976629  10705.500
#> 118        NA         NA                <NA>        Ask  0.976776   4839.316
#> 119        NA         NA                <NA>        Ask  0.976777   9775.658
#> 120        NA         NA                <NA>        Ask  0.976958   9677.600
#> 121        NA         NA                <NA>        Ask  0.977157   5088.876
#> 122        NA         NA                <NA>        Ask  0.977158  15299.700
#> 123        NA         NA                <NA>        Ask  0.977365   4584.244
#> 124        NA         NA                <NA>        Ask  0.977366   1375.600
#> 125        NA         NA                <NA>        Ask  0.977387  14654.347
#> 126        NA         NA                <NA>        Ask  0.977399     14.769
#> 127        NA         NA                <NA>        Ask  0.977506   2755.703
#> 128        NA         NA                <NA>        Ask  0.977594    431.541
#> 129        NA         NA                <NA>        Ask  0.977620   9401.560
#> 130        NA         NA                <NA>        Ask  0.977776   2933.000
#> 131        NA         NA                <NA>        Ask  0.977777   7005.859
#> 132        NA         NA                <NA>        Ask  0.977821   7922.600
#> 133        NA         NA                <NA>        Ask  0.977949   8234.030
#> 134        NA         NA                <NA>        Ask  0.977980   1466.524
#> 135        NA         NA                <NA>        Ask  0.978050  14644.412
#> 136        NA         NA                <NA>        Ask  0.978963   3931.686
#> 137        NA         NA                <NA>        Ask  0.979006   9829.215
#> 138        NA         NA                <NA>        Ask  0.979011  14630.037
#> 139        NA         NA                <NA>        Ask  0.979054   5000.000
#> 140        NA         NA                <NA>        Ask  0.979885  15726.744
#> 141        NA         NA                <NA>        Ask  0.979933  17434.500
#> 142        NA         NA                <NA>        Ask  0.980171     43.948
#> 143        NA         NA                <NA>        Ask  0.980658     10.229
#> 144        NA         NA                <NA>        Ask  0.980839    149.516
#> 145        NA         NA                <NA>        Ask  0.980840   6552.810
#> 146        NA         NA                <NA>        Ask  0.980889  15773.184
#> 147        NA         NA                <NA>        Ask  0.980958     22.839
#> 148        NA         NA                <NA>        Ask  0.981008  15880.171
#> 149        NA         NA                <NA>        Ask  0.981118  39764.400
#> 150        NA         NA                <NA>        Ask  0.981388  14595.365
#> 151        NA         NA                <NA>        Ask  0.981969     24.102
#> 152        NA         NA                <NA>        Ask  0.982000     41.491
#> 153        NA         NA                <NA>        Ask  0.982098    138.225
#> 154        NA         NA                <NA>        Ask  0.982713   4764.380
#> 155        NA         NA                <NA>        Ask  0.983159  18821.704
#> 156        NA         NA                <NA>        Ask  0.983206  37945.700
#> 157        NA         NA                <NA>        Ask  0.983251  14566.947
#> 158        NA         NA                <NA>        Ask  0.983536     99.610
#> 159        NA         NA                <NA>        Ask  0.983701   2331.615
#> 160        NA         NA                <NA>        Ask  0.983995  24826.063
#> 161        NA         NA                <NA>        Ask  0.984101   2140.622
#> 162        NA         NA                <NA>        Ask  0.985148    150.059
#> 163        NA         NA                <NA>        Ask  0.985264     48.681
#> 164        NA         NA                <NA>        Ask  0.985771  21997.142
#> 165        NA         NA                <NA>        Ask  0.986014     11.586
#> 166        NA         NA                <NA>        Ask  0.986130 288999.356
#> 167        NA         NA                <NA>        Ask  0.986131  11712.482
#> 168        NA         NA                <NA>        Ask  0.986190   3513.840
#> 169        NA         NA                <NA>        Ask  0.986328   7328.436
#> 170        NA         NA                <NA>        Ask  0.986415    137.620
#> 171        NA         NA                <NA>        Ask  0.986688   8794.123
#> 172        NA         NA                <NA>        Ask  0.986689  26581.100
#> 173        NA         NA                <NA>        Ask  0.986725  14518.081
#> 174        NA         NA                <NA>        Ask  0.987632   4214.044
#> 175        NA         NA                <NA>        Ask  0.987979     23.956
#> 176        NA         NA                <NA>        Ask  0.988354   2930.312
#> 177        NA         NA                <NA>        Ask  0.989835  12732.200
#> 178        NA         NA                <NA>        Ask  0.990094    798.591
#> 179        NA         NA                <NA>        Ask  0.990750    137.018
#> 180        NA         NA                <NA>        Ask  0.991454     11.529
#> 181        NA         NA                <NA>        Ask  0.991905   5473.257
#> 182        NA         NA                <NA>        Ask  0.992245     31.277
#> 183        NA         NA                <NA>        Ask  0.993136     55.456
#> 184        NA         NA                <NA>        Ask  0.993224     78.674
#> 185        NA         NA                <NA>        Ask  0.993634     48.271
#> 186        NA         NA                <NA>        Ask  0.994000     41.491
#> 187        NA         NA                <NA>        Ask  0.994025     23.810
#> 188        NA         NA                <NA>        Ask  0.994142     98.547
#> 189        NA         NA                <NA>        Ask  0.994377  14189.146
#> 190        NA         NA                <NA>        Ask  0.994575   6487.000
#> 191        NA         NA                <NA>        Ask  0.995104    136.419
#> 192        NA         NA                <NA>        Ask  0.996072  14381.984
#> 193        NA         NA                <NA>        Ask  0.996087    793.275
#> 194        NA         NA                <NA>        Ask  0.996211     98.300
#> 195        NA         NA                <NA>        Ask  0.996922     11.466
#> 196        NA         NA                <NA>        Ask  0.998690      7.057
#> 197        NA         NA                <NA>        Ask  0.999478    135.822
#> 198        NA         NA                <NA>        Ask  1.000000   3562.826
#> 199        NA         NA                <NA>        Ask  1.000109     23.665
#> 200        NA         NA                <NA>        Ask  1.000330   6626.100
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 17:04:50
#> 102 2024-12-18 17:04:50
#> 103 2024-12-18 17:04:49
#> 104 2024-12-18 17:04:47
#> 105 2024-12-18 17:04:49
#> 106 2024-12-18 17:04:46
#> 107 2024-12-18 17:04:49
#> 108 2024-12-18 17:04:49
#> 109 2024-12-18 17:04:46
#> 110 2024-12-18 17:04:49
#> 111 2024-12-18 17:04:45
#> 112 2024-12-18 17:04:49
#> 113 2024-12-18 17:04:48
#> 114 2024-12-18 17:04:45
#> 115 2024-12-18 17:04:44
#> 116 2024-12-18 17:04:43
#> 117 2024-12-18 17:04:49
#> 118 2024-12-18 17:04:46
#> 119 2024-12-18 17:04:45
#> 120 2024-12-18 17:04:49
#> 121 2024-12-18 17:04:47
#> 122 2024-12-18 17:04:45
#> 123 2024-12-18 17:04:38
#> 124 2024-12-18 17:04:23
#> 125 2024-12-18 17:04:45
#> 126 2024-12-18 17:04:49
#> 127 2024-12-18 17:04:45
#> 128 2024-12-18 17:04:45
#> 129 2024-12-18 17:04:27
#> 130 2024-12-18 17:04:24
#> 131 2024-12-18 17:04:24
#> 132 2024-12-18 17:04:43
#> 133 2024-12-18 17:04:27
#> 134 2024-12-18 17:04:34
#> 135 2024-12-18 17:04:39
#> 136 2024-12-18 17:04:48
#> 137 2024-12-18 17:04:48
#> 138 2024-12-18 17:04:44
#> 139 2024-12-18 17:04:28
#> 140 2024-12-18 17:04:48
#> 141 2024-12-18 17:03:40
#> 142 2024-12-18 17:04:23
#> 143 2024-12-18 17:04:28
#> 144 2024-12-18 17:04:42
#> 145 2024-12-18 17:04:48
#> 146 2024-12-18 17:04:18
#> 147 2024-12-18 17:04:35
#> 148 2024-12-18 17:03:16
#> 149 2024-12-18 17:04:33
#> 150 2024-12-18 17:04:27
#> 151 2024-12-18 16:41:15
#> 152 2024-12-18 16:51:06
#> 153 2024-12-18 17:04:34
#> 154 2024-12-18 17:04:07
#> 155 2024-12-18 17:04:44
#> 156 2024-12-18 17:04:22
#> 157 2024-12-18 17:04:49
#> 158 2024-12-18 15:57:30
#> 159 2024-12-18 17:04:49
#> 160 2024-12-18 17:03:44
#> 161 2024-12-18 17:04:27
#> 162 2024-12-18 17:04:19
#> 163 2024-12-18 14:56:59
#> 164 2024-12-18 17:04:09
#> 165 2024-12-18 15:09:59
#> 166 2024-12-18 17:04:33
#> 167 2024-12-18 17:01:28
#> 168 2024-12-18 17:04:41
#> 169 2024-12-18 17:04:09
#> 170 2024-12-18 16:58:52
#> 171 2024-12-18 17:04:09
#> 172 2024-12-18 17:03:24
#> 173 2024-12-18 17:03:40
#> 174 2024-12-18 17:02:50
#> 175 2024-12-18 14:35:26
#> 176 2024-12-18 17:03:43
#> 177 2024-12-18 17:02:07
#> 178 2024-12-18 16:54:05
#> 179 2024-12-18 13:57:45
#> 180 2024-12-18 12:37:59
#> 181 2024-12-18 17:03:07
#> 182 2024-12-18 15:04:40
#> 183 2024-12-18 13:26:56
#> 184 2024-12-18 15:12:45
#> 185 2024-12-18 12:38:18
#> 186 2024-12-18 12:25:06
#> 187 2024-12-18 13:36:40
#> 188 2024-12-18 12:37:51
#> 189 2024-12-18 16:57:57
#> 190 2024-12-18 16:40:42
#> 191 2024-12-18 13:34:11
#> 192 2024-12-18 17:02:53
#> 193 2024-12-18 16:38:48
#> 194 2024-12-18 15:07:21
#> 195 2024-12-18 12:33:01
#> 196 2024-12-18 12:34:04
#> 197 2024-12-18 13:11:41
#> 198 2024-12-18 14:57:48
#> 199 2024-12-18 12:15:38
#> 200 2024-12-18 16:37:45