Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.954532     11.868 2024-12-11 09:46:23        Bid        NA         NA
#> 2    0.954823   5131.147 2024-12-18 16:23:08        Bid        NA         NA
#> 3    0.955000     81.491 2024-12-18 07:28:46        Bid        NA         NA
#> 4    0.955288  15375.240 2024-12-18 16:23:08        Bid        NA         NA
#> 5    0.955770   3505.028 2024-12-12 05:37:10        Bid        NA         NA
#> 6    0.955871     16.358 2024-12-18 08:50:50        Bid        NA         NA
#> 7    0.955890     31.385 2024-12-18 13:29:07        Bid        NA         NA
#> 8    0.955970     86.060 2024-12-11 11:33:28        Bid        NA         NA
#> 9    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 10   0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 11   0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 12   0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 13   0.956114  37711.107 2024-12-18 16:22:56        Bid        NA         NA
#> 14   0.956339   8197.997 2024-12-18 16:23:07        Bid        NA         NA
#> 15   0.956467   3121.336 2024-12-18 16:25:02        Bid        NA         NA
#> 16   0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 17   0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 18   0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 19   0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 20   0.957308  14953.478 2024-12-18 16:25:05        Bid        NA         NA
#> 21   0.957460   3719.440 2024-12-18 16:24:54        Bid        NA         NA
#> 22   0.957567   2458.201 2024-12-18 16:25:05        Bid        NA         NA
#> 23   0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 24   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 25   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 26   0.958431   2259.257 2024-12-18 16:24:52        Bid        NA         NA
#> 27   0.958526  63600.732 2024-12-18 16:23:04        Bid        NA         NA
#> 28   0.958845 289168.118 2024-12-18 16:24:56        Bid        NA         NA
#> 29   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 30   0.959061  14926.926 2024-12-18 16:24:36        Bid        NA         NA
#> 31   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 32   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 33   0.959739   1838.220 2024-12-18 16:24:23        Bid        NA         NA
#> 34   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 35   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 36   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 37   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 38   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 39   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 40   0.960937  15720.934 2024-12-18 16:22:56        Bid        NA         NA
#> 41   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 42   0.960985  10341.403 2024-12-18 16:25:14        Bid        NA         NA
#> 43   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 44   0.961357  19207.356 2024-12-18 16:24:28        Bid        NA         NA
#> 45   0.961463  59317.270 2024-12-18 16:23:08        Bid        NA         NA
#> 46   0.961464    151.925 2024-12-18 16:24:35        Bid        NA         NA
#> 47   0.961511   9306.922 2024-12-18 16:25:14        Bid        NA         NA
#> 48   0.961753  38248.600 2024-12-18 16:24:36        Bid        NA         NA
#> 49   0.961758   3317.415 2024-12-18 16:25:06        Bid        NA         NA
#> 50   0.961771  14884.099 2024-12-18 16:25:06        Bid        NA         NA
#> 51   0.961819   8272.271 2024-12-18 16:25:14        Bid        NA         NA
#> 52   0.961998  15189.107 2024-12-18 16:24:20        Bid        NA         NA
#> 53   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 54   0.962004    174.963 2024-12-18 16:24:42        Bid        NA         NA
#> 55   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 56   0.962248  13441.424 2024-12-18 16:25:14        Bid        NA         NA
#> 57   0.962848     49.000 2024-12-18 16:25:11        Bid        NA         NA
#> 58   0.962890 205630.889 2024-12-18 16:24:53        Bid        NA         NA
#> 59   0.962938  15507.017 2024-12-18 16:25:14        Bid        NA         NA
#> 60   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 61   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 62   0.964496     43.710 2024-12-18 16:25:11        Bid        NA         NA
#> 63   0.964608  14840.312 2024-12-18 16:25:10        Bid        NA         NA
#> 64   0.964848  17574.600 2024-12-18 16:24:50        Bid        NA         NA
#> 65   0.964849   2377.178 2024-12-18 16:25:06        Bid        NA         NA
#> 66   0.965430  14827.683 2024-12-18 16:25:12        Bid        NA         NA
#> 67   0.965801  14821.981 2024-12-18 16:25:08        Bid        NA         NA
#> 68   0.965880   2755.703 2024-12-18 16:25:08        Bid        NA         NA
#> 69   0.966079     14.454 2024-12-18 16:24:54        Bid        NA         NA
#> 70   0.966108  20804.254 2024-12-18 16:25:04        Bid        NA         NA
#> 71   0.966109    548.686 2024-12-18 16:25:06        Bid        NA         NA
#> 72   0.966234   2463.000 2024-12-18 16:25:07        Bid        NA         NA
#> 73   0.966235   2965.000 2024-12-18 16:25:06        Bid        NA         NA
#> 74   0.966339   9753.300 2024-12-18 16:25:06        Bid        NA         NA
#> 75   0.966396  10412.380 2024-12-18 16:25:11        Bid        NA         NA
#> 76   0.966435  10989.470 2024-12-18 16:25:13        Bid        NA         NA
#> 77   0.966485  10823.590 2024-12-18 16:25:13        Bid        NA         NA
#> 78   0.966666   5000.000 2024-12-18 16:24:50        Bid        NA         NA
#> 79   0.966668   9872.458 2024-12-18 16:25:11        Bid        NA         NA
#> 80   0.966846   4945.000 2024-12-18 16:25:12        Bid        NA         NA
#> 81   0.966882   5006.639 2024-12-18 16:25:14        Bid        NA         NA
#> 82   0.966911  16025.100 2024-12-18 16:25:10        Bid        NA         NA
#> 83   0.967012   9868.950 2024-12-18 16:25:12        Bid        NA         NA
#> 84   0.967067   7566.874 2024-12-18 16:25:10        Bid        NA         NA
#> 85   0.967310   5405.322 2024-12-18 16:25:11        Bid        NA         NA
#> 86   0.967327   9865.730 2024-12-18 16:25:12        Bid        NA         NA
#> 87   0.967402   1936.000 2024-12-18 16:25:08        Bid        NA         NA
#> 88   0.967403    362.363 2024-12-18 16:25:10        Bid        NA         NA
#> 89   0.967486   7398.083 2024-12-18 16:25:12        Bid        NA         NA
#> 90   0.967557   4931.694 2024-12-18 16:25:12        Bid        NA         NA
#> 91   0.967571   5876.592 2024-12-18 16:25:13        Bid        NA         NA
#> 92   0.967639   3576.965 2024-12-18 16:25:13        Bid        NA         NA
#> 93   0.967640     16.697 2024-12-18 16:25:10        Bid        NA         NA
#> 94   0.967646   2465.620 2024-12-18 16:25:13        Bid        NA         NA
#> 95   0.967676   1377.865 2024-12-18 16:25:13        Bid        NA         NA
#> 96   0.967681  11580.100 2024-12-18 16:25:12        Bid        NA         NA
#> 97   0.967690     33.799 2024-12-18 16:25:14        Bid        NA         NA
#> 98   0.967692    651.300 2024-12-18 16:25:14        Bid        NA         NA
#> 99   0.967695     33.799 2024-12-18 16:25:14        Bid        NA         NA
#> 100  0.967696    185.925 2024-12-18 16:25:14        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.968366   1377.502
#> 102        NA         NA                <NA>        Ask  0.968367   1936.000
#> 103        NA         NA                <NA>        Ask  0.968436  14781.663
#> 104        NA         NA                <NA>        Ask  0.968442   9854.388
#> 105        NA         NA                <NA>        Ask  0.968495   1000.000
#> 106        NA         NA                <NA>        Ask  0.968592   5876.592
#> 107        NA         NA                <NA>        Ask  0.968634     93.697
#> 108        NA         NA                <NA>        Ask  0.968664   3576.965
#> 109        NA         NA                <NA>        Ask  0.968711     33.799
#> 110        NA         NA                <NA>        Ask  0.968735     33.799
#> 111        NA         NA                <NA>        Ask  0.968778  10502.269
#> 112        NA         NA                <NA>        Ask  0.968793   1628.002
#> 113        NA         NA                <NA>        Ask  0.968839  10000.000
#> 114        NA         NA                <NA>        Ask  0.968848    394.158
#> 115        NA         NA                <NA>        Ask  0.969207   2100.000
#> 116        NA         NA                <NA>        Ask  0.969214   7384.899
#> 117        NA         NA                <NA>        Ask  0.969537   4839.316
#> 118        NA         NA                <NA>        Ask  0.969538  10705.500
#> 119        NA         NA                <NA>        Ask  0.969814  15299.700
#> 120        NA         NA                <NA>        Ask  0.969905  14759.280
#> 121        NA         NA                <NA>        Ask  0.969929   2127.859
#> 122        NA         NA                <NA>        Ask  0.970090    479.589
#> 123        NA         NA                <NA>        Ask  0.970092   2965.000
#> 124        NA         NA                <NA>        Ask  0.970193   9760.900
#> 125        NA         NA                <NA>        Ask  0.970348     10.595
#> 126        NA         NA                <NA>        Ask  0.970375   2755.703
#> 127        NA         NA                <NA>        Ask  0.970536  14749.687
#> 128        NA         NA                <NA>        Ask  0.970624   1449.527
#> 129        NA         NA                <NA>        Ask  0.970843   8127.700
#> 130        NA         NA                <NA>        Ask  0.970928   7866.140
#> 131        NA         NA                <NA>        Ask  0.970969   9618.290
#> 132        NA         NA                <NA>        Ask  0.971582  10213.526
#> 133        NA         NA                <NA>        Ask  0.971626  14733.133
#> 134        NA         NA                <NA>        Ask  0.972114   4085.411
#> 135        NA         NA                <NA>        Ask  0.972324     43.982
#> 136        NA         NA                <NA>        Ask  0.972365  15880.171
#> 137        NA         NA                <NA>        Ask  0.972431    158.223
#> 138        NA         NA                <NA>        Ask  0.972432   4766.312
#> 139        NA         NA                <NA>        Ask  0.972479    170.720
#> 140        NA         NA                <NA>        Ask  0.972480   5000.000
#> 141        NA         NA                <NA>        Ask  0.972481  18821.704
#> 142        NA         NA                <NA>        Ask  0.972537  15773.184
#> 143        NA         NA                <NA>        Ask  0.972601   5447.214
#> 144        NA         NA                <NA>        Ask  0.972649  17559.500
#> 145        NA         NA                <NA>        Ask  0.972887   6128.116
#> 146        NA         NA                <NA>        Ask  0.972935  39764.400
#> 147        NA         NA                <NA>        Ask  0.973051     10.229
#> 148        NA         NA                <NA>        Ask  0.973182   6809.017
#> 149        NA         NA                <NA>        Ask  0.973827  37945.700
#> 150        NA         NA                <NA>        Ask  0.973841  14699.631
#> 151        NA         NA                <NA>        Ask  0.974003  24826.063
#> 152        NA         NA                <NA>        Ask  0.974643   4710.960
#> 153        NA         NA                <NA>        Ask  0.975414  14675.927
#> 154        NA         NA                <NA>        Ask  0.975995     24.250
#> 155        NA         NA                <NA>        Ask  0.976288  11578.915
#> 156        NA         NA                <NA>        Ask  0.976559  26581.100
#> 157        NA         NA                <NA>        Ask  0.976703   4214.044
#> 158        NA         NA                <NA>        Ask  0.976735   8865.775
#> 159        NA         NA                <NA>        Ask  0.977313  21697.190
#> 160        NA         NA                <NA>        Ask  0.977800    138.833
#> 161        NA         NA                <NA>        Ask  0.977850   7245.242
#> 162        NA         NA                <NA>        Ask  0.978020   2214.007
#> 163        NA         NA                <NA>        Ask  0.978430   2405.784
#> 164        NA         NA                <NA>        Ask  0.978511 288251.286
#> 165        NA         NA                <NA>        Ask  0.978512  14630.224
#> 166        NA         NA                <NA>        Ask  0.979279   3636.565
#> 167        NA         NA                <NA>        Ask  0.979291  12732.200
#> 168        NA         NA                <NA>        Ask  0.980007   3043.122
#> 169        NA         NA                <NA>        Ask  0.981969     24.102
#> 170        NA         NA                <NA>        Ask  0.982098    138.225
#> 171        NA         NA                <NA>        Ask  0.983536     99.610
#> 172        NA         NA                <NA>        Ask  0.984755   6487.000
#> 173        NA         NA                <NA>        Ask  0.985264     48.681
#> 174        NA         NA                <NA>        Ask  0.986014     11.586
#> 175        NA         NA                <NA>        Ask  0.986415    137.620
#> 176        NA         NA                <NA>        Ask  0.987928  14490.789
#> 177        NA         NA                <NA>        Ask  0.987979     23.956
#> 178        NA         NA                <NA>        Ask  0.988173   1566.619
#> 179        NA         NA                <NA>        Ask  0.988174   2999.988
#> 180        NA         NA                <NA>        Ask  0.988552   1716.359
#> 181        NA         NA                <NA>        Ask  0.988893   7927.653
#> 182        NA         NA                <NA>        Ask  0.989589    792.648
#> 183        NA         NA                <NA>        Ask  0.989625   5463.282
#> 184        NA         NA                <NA>        Ask  0.990750    137.018
#> 185        NA         NA                <NA>        Ask  0.991454     11.529
#> 186        NA         NA                <NA>        Ask  0.992245     31.277
#> 187        NA         NA                <NA>        Ask  0.992952   6626.100
#> 188        NA         NA                <NA>        Ask  0.993116    787.990
#> 189        NA         NA                <NA>        Ask  0.993136     55.456
#> 190        NA         NA                <NA>        Ask  0.993224     78.674
#> 191        NA         NA                <NA>        Ask  0.993634     48.271
#> 192        NA         NA                <NA>        Ask  0.994000     41.491
#> 193        NA         NA                <NA>        Ask  0.994025     23.810
#> 194        NA         NA                <NA>        Ask  0.994142     98.547
#> 195        NA         NA                <NA>        Ask  0.995104    136.419
#> 196        NA         NA                <NA>        Ask  0.996211     98.300
#> 197        NA         NA                <NA>        Ask  0.996922     11.466
#> 198        NA         NA                <NA>        Ask  0.997223  52875.474
#> 199        NA         NA                <NA>        Ask  0.997597     16.698
#> 200        NA         NA                <NA>        Ask  0.997598  24630.935
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:25:14
#> 102 2024-12-18 16:25:14
#> 103 2024-12-18 16:25:14
#> 104 2024-12-18 16:25:14
#> 105 2024-12-18 16:25:13
#> 106 2024-12-18 16:25:13
#> 107 2024-12-18 16:25:11
#> 108 2024-12-18 16:25:13
#> 109 2024-12-18 16:25:13
#> 110 2024-12-18 16:25:14
#> 111 2024-12-18 16:25:14
#> 112 2024-12-18 16:25:13
#> 113 2024-12-18 16:25:13
#> 114 2024-12-18 16:25:14
#> 115 2024-12-18 16:25:11
#> 116 2024-12-18 16:25:12
#> 117 2024-12-18 16:25:14
#> 118 2024-12-18 16:25:12
#> 119 2024-12-18 16:25:12
#> 120 2024-12-18 16:25:12
#> 121 2024-12-18 16:25:11
#> 122 2024-12-18 16:25:13
#> 123 2024-12-18 16:25:11
#> 124 2024-12-18 16:25:07
#> 125 2024-12-18 16:25:12
#> 126 2024-12-18 16:25:10
#> 127 2024-12-18 16:25:12
#> 128 2024-12-18 16:25:01
#> 129 2024-12-18 16:25:13
#> 130 2024-12-18 16:25:13
#> 131 2024-12-18 16:25:11
#> 132 2024-12-18 16:25:04
#> 133 2024-12-18 16:25:10
#> 134 2024-12-18 16:25:04
#> 135 2024-12-18 16:25:07
#> 136 2024-12-18 16:25:10
#> 137 2024-12-18 16:25:07
#> 138 2024-12-18 16:25:04
#> 139 2024-12-18 16:25:07
#> 140 2024-12-18 16:24:54
#> 141 2024-12-18 16:24:49
#> 142 2024-12-18 16:24:55
#> 143 2024-12-18 16:25:04
#> 144 2024-12-18 16:24:58
#> 145 2024-12-18 16:25:04
#> 146 2024-12-18 16:24:54
#> 147 2024-12-18 16:24:54
#> 148 2024-12-18 16:24:52
#> 149 2024-12-18 16:23:04
#> 150 2024-12-18 16:25:10
#> 151 2024-12-18 16:22:56
#> 152 2024-12-18 16:25:01
#> 153 2024-12-18 16:25:07
#> 154 2024-12-18 16:22:03
#> 155 2024-12-18 16:24:54
#> 156 2024-12-18 16:22:56
#> 157 2024-12-18 16:22:56
#> 158 2024-12-18 16:24:57
#> 159 2024-12-18 16:24:57
#> 160 2024-12-18 16:17:28
#> 161 2024-12-18 16:24:57
#> 162 2024-12-18 16:24:52
#> 163 2024-12-18 16:25:08
#> 164 2024-12-18 16:25:06
#> 165 2024-12-18 16:24:51
#> 166 2024-12-18 16:24:55
#> 167 2024-12-18 16:22:56
#> 168 2024-12-18 16:24:37
#> 169 2024-12-18 16:08:58
#> 170 2024-12-18 14:56:28
#> 171 2024-12-18 15:57:30
#> 172 2024-12-18 16:23:04
#> 173 2024-12-18 14:56:59
#> 174 2024-12-18 15:09:59
#> 175 2024-12-18 15:24:42
#> 176 2024-12-18 16:24:50
#> 177 2024-12-18 14:35:26
#> 178 2024-12-18 16:18:14
#> 179 2024-12-18 16:18:12
#> 180 2024-12-18 16:16:40
#> 181 2024-12-18 16:16:28
#> 182 2024-12-18 16:16:25
#> 183 2024-12-18 16:16:22
#> 184 2024-12-18 13:57:45
#> 185 2024-12-18 12:37:59
#> 186 2024-12-18 15:04:40
#> 187 2024-12-18 16:22:56
#> 188 2024-12-18 16:24:51
#> 189 2024-12-18 13:26:56
#> 190 2024-12-18 15:12:45
#> 191 2024-12-18 12:38:18
#> 192 2024-12-18 12:25:06
#> 193 2024-12-18 13:36:40
#> 194 2024-12-18 12:37:51
#> 195 2024-12-18 13:34:11
#> 196 2024-12-18 15:07:21
#> 197 2024-12-18 12:33:01
#> 198 2024-12-18 16:25:06
#> 199 2024-12-18 16:25:04
#> 200 2024-12-18 16:24:54
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.954532     11.868 2024-12-11 09:46:23        Bid        NA         NA
#> 2    0.954823   5131.147 2024-12-18 16:23:08        Bid        NA         NA
#> 3    0.955000     81.491 2024-12-18 07:28:46        Bid        NA         NA
#> 4    0.955288  15375.240 2024-12-18 16:23:08        Bid        NA         NA
#> 5    0.955770   3505.028 2024-12-12 05:37:10        Bid        NA         NA
#> 6    0.955871     16.358 2024-12-18 08:50:50        Bid        NA         NA
#> 7    0.955890     31.385 2024-12-18 13:29:07        Bid        NA         NA
#> 8    0.955970     86.060 2024-12-11 11:33:28        Bid        NA         NA
#> 9    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 10   0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 11   0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 12   0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 13   0.956114  37711.107 2024-12-18 16:22:56        Bid        NA         NA
#> 14   0.956339   8197.997 2024-12-18 16:23:07        Bid        NA         NA
#> 15   0.956467   3121.336 2024-12-18 16:25:02        Bid        NA         NA
#> 16   0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 17   0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 18   0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 19   0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 20   0.957308  14953.478 2024-12-18 16:25:05        Bid        NA         NA
#> 21   0.957460   3719.440 2024-12-18 16:24:54        Bid        NA         NA
#> 22   0.957567   2458.201 2024-12-18 16:25:05        Bid        NA         NA
#> 23   0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 24   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 25   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 26   0.958431   2259.257 2024-12-18 16:24:52        Bid        NA         NA
#> 27   0.958526  63600.732 2024-12-18 16:23:04        Bid        NA         NA
#> 28   0.958845 289168.118 2024-12-18 16:24:56        Bid        NA         NA
#> 29   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 30   0.959061  14926.926 2024-12-18 16:24:36        Bid        NA         NA
#> 31   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 32   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 33   0.959739   1838.220 2024-12-18 16:24:23        Bid        NA         NA
#> 34   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 35   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 36   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 37   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 38   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 39   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 40   0.960937  15720.934 2024-12-18 16:22:56        Bid        NA         NA
#> 41   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 42   0.960985  10341.403 2024-12-18 16:25:14        Bid        NA         NA
#> 43   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 44   0.961357  19207.356 2024-12-18 16:24:28        Bid        NA         NA
#> 45   0.961463  59317.270 2024-12-18 16:23:08        Bid        NA         NA
#> 46   0.961464    151.925 2024-12-18 16:24:35        Bid        NA         NA
#> 47   0.961511   9306.922 2024-12-18 16:25:14        Bid        NA         NA
#> 48   0.961753  38248.600 2024-12-18 16:24:36        Bid        NA         NA
#> 49   0.961758   3317.415 2024-12-18 16:25:06        Bid        NA         NA
#> 50   0.961771  14884.099 2024-12-18 16:25:06        Bid        NA         NA
#> 51   0.961819   8272.271 2024-12-18 16:25:14        Bid        NA         NA
#> 52   0.961998  15189.107 2024-12-18 16:24:20        Bid        NA         NA
#> 53   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 54   0.962004    174.963 2024-12-18 16:24:42        Bid        NA         NA
#> 55   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 56   0.962248  13441.424 2024-12-18 16:25:14        Bid        NA         NA
#> 57   0.962848     49.000 2024-12-18 16:25:11        Bid        NA         NA
#> 58   0.962890 205630.889 2024-12-18 16:24:53        Bid        NA         NA
#> 59   0.962938  15507.017 2024-12-18 16:25:14        Bid        NA         NA
#> 60   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 61   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 62   0.964496     43.710 2024-12-18 16:25:11        Bid        NA         NA
#> 63   0.964608  14840.312 2024-12-18 16:25:10        Bid        NA         NA
#> 64   0.964848  17574.600 2024-12-18 16:24:50        Bid        NA         NA
#> 65   0.964849   2377.178 2024-12-18 16:25:06        Bid        NA         NA
#> 66   0.965430  14827.683 2024-12-18 16:25:12        Bid        NA         NA
#> 67   0.965801  14821.981 2024-12-18 16:25:08        Bid        NA         NA
#> 68   0.965880   2755.703 2024-12-18 16:25:08        Bid        NA         NA
#> 69   0.966079     14.454 2024-12-18 16:24:54        Bid        NA         NA
#> 70   0.966108  20804.254 2024-12-18 16:25:04        Bid        NA         NA
#> 71   0.966109    548.686 2024-12-18 16:25:06        Bid        NA         NA
#> 72   0.966234   2463.000 2024-12-18 16:25:07        Bid        NA         NA
#> 73   0.966235   2965.000 2024-12-18 16:25:06        Bid        NA         NA
#> 74   0.966339   9753.300 2024-12-18 16:25:06        Bid        NA         NA
#> 75   0.966396  10412.380 2024-12-18 16:25:11        Bid        NA         NA
#> 76   0.966435  10989.470 2024-12-18 16:25:13        Bid        NA         NA
#> 77   0.966485  10823.590 2024-12-18 16:25:13        Bid        NA         NA
#> 78   0.966666   5000.000 2024-12-18 16:24:50        Bid        NA         NA
#> 79   0.966668   9872.458 2024-12-18 16:25:11        Bid        NA         NA
#> 80   0.966846   4945.000 2024-12-18 16:25:12        Bid        NA         NA
#> 81   0.966882   5006.639 2024-12-18 16:25:14        Bid        NA         NA
#> 82   0.966911  16025.100 2024-12-18 16:25:10        Bid        NA         NA
#> 83   0.967012   9868.950 2024-12-18 16:25:12        Bid        NA         NA
#> 84   0.967067   7566.874 2024-12-18 16:25:10        Bid        NA         NA
#> 85   0.967310   5405.322 2024-12-18 16:25:11        Bid        NA         NA
#> 86   0.967327   9865.730 2024-12-18 16:25:12        Bid        NA         NA
#> 87   0.967402   1936.000 2024-12-18 16:25:08        Bid        NA         NA
#> 88   0.967403    362.363 2024-12-18 16:25:10        Bid        NA         NA
#> 89   0.967486   7398.083 2024-12-18 16:25:12        Bid        NA         NA
#> 90   0.967557   4931.694 2024-12-18 16:25:12        Bid        NA         NA
#> 91   0.967571   5876.592 2024-12-18 16:25:13        Bid        NA         NA
#> 92   0.967639   3576.965 2024-12-18 16:25:13        Bid        NA         NA
#> 93   0.967640     16.697 2024-12-18 16:25:10        Bid        NA         NA
#> 94   0.967646   2465.620 2024-12-18 16:25:13        Bid        NA         NA
#> 95   0.967676   1377.865 2024-12-18 16:25:13        Bid        NA         NA
#> 96   0.967681  11580.100 2024-12-18 16:25:12        Bid        NA         NA
#> 97   0.967690     33.799 2024-12-18 16:25:14        Bid        NA         NA
#> 98   0.967692    651.300 2024-12-18 16:25:14        Bid        NA         NA
#> 99   0.967695     33.799 2024-12-18 16:25:14        Bid        NA         NA
#> 100  0.967696    185.925 2024-12-18 16:25:14        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.968366   1377.502
#> 102        NA         NA                <NA>        Ask  0.968367   1936.000
#> 103        NA         NA                <NA>        Ask  0.968436  14781.663
#> 104        NA         NA                <NA>        Ask  0.968442   9854.388
#> 105        NA         NA                <NA>        Ask  0.968495   1000.000
#> 106        NA         NA                <NA>        Ask  0.968592   5876.592
#> 107        NA         NA                <NA>        Ask  0.968634     93.697
#> 108        NA         NA                <NA>        Ask  0.968664   3576.965
#> 109        NA         NA                <NA>        Ask  0.968711     33.799
#> 110        NA         NA                <NA>        Ask  0.968735     33.799
#> 111        NA         NA                <NA>        Ask  0.968778  10502.269
#> 112        NA         NA                <NA>        Ask  0.968793   1628.002
#> 113        NA         NA                <NA>        Ask  0.968839  10000.000
#> 114        NA         NA                <NA>        Ask  0.968848    394.158
#> 115        NA         NA                <NA>        Ask  0.969207   2100.000
#> 116        NA         NA                <NA>        Ask  0.969214   7384.899
#> 117        NA         NA                <NA>        Ask  0.969537   4839.316
#> 118        NA         NA                <NA>        Ask  0.969538  10705.500
#> 119        NA         NA                <NA>        Ask  0.969814  15299.700
#> 120        NA         NA                <NA>        Ask  0.969905  14759.280
#> 121        NA         NA                <NA>        Ask  0.969929   2127.859
#> 122        NA         NA                <NA>        Ask  0.970090    479.589
#> 123        NA         NA                <NA>        Ask  0.970092   2965.000
#> 124        NA         NA                <NA>        Ask  0.970193   9760.900
#> 125        NA         NA                <NA>        Ask  0.970348     10.595
#> 126        NA         NA                <NA>        Ask  0.970375   2755.703
#> 127        NA         NA                <NA>        Ask  0.970536  14749.687
#> 128        NA         NA                <NA>        Ask  0.970624   1449.527
#> 129        NA         NA                <NA>        Ask  0.970843   8127.700
#> 130        NA         NA                <NA>        Ask  0.970928   7866.140
#> 131        NA         NA                <NA>        Ask  0.970969   9618.290
#> 132        NA         NA                <NA>        Ask  0.971582  10213.526
#> 133        NA         NA                <NA>        Ask  0.971626  14733.133
#> 134        NA         NA                <NA>        Ask  0.972114   4085.411
#> 135        NA         NA                <NA>        Ask  0.972324     43.982
#> 136        NA         NA                <NA>        Ask  0.972365  15880.171
#> 137        NA         NA                <NA>        Ask  0.972431    158.223
#> 138        NA         NA                <NA>        Ask  0.972432   4766.312
#> 139        NA         NA                <NA>        Ask  0.972479    170.720
#> 140        NA         NA                <NA>        Ask  0.972480   5000.000
#> 141        NA         NA                <NA>        Ask  0.972481  18821.704
#> 142        NA         NA                <NA>        Ask  0.972537  15773.184
#> 143        NA         NA                <NA>        Ask  0.972601   5447.214
#> 144        NA         NA                <NA>        Ask  0.972649  17559.500
#> 145        NA         NA                <NA>        Ask  0.972887   6128.116
#> 146        NA         NA                <NA>        Ask  0.972935  39764.400
#> 147        NA         NA                <NA>        Ask  0.973051     10.229
#> 148        NA         NA                <NA>        Ask  0.973182   6809.017
#> 149        NA         NA                <NA>        Ask  0.973827  37945.700
#> 150        NA         NA                <NA>        Ask  0.973841  14699.631
#> 151        NA         NA                <NA>        Ask  0.974003  24826.063
#> 152        NA         NA                <NA>        Ask  0.974643   4710.960
#> 153        NA         NA                <NA>        Ask  0.975414  14675.927
#> 154        NA         NA                <NA>        Ask  0.975995     24.250
#> 155        NA         NA                <NA>        Ask  0.976288  11578.915
#> 156        NA         NA                <NA>        Ask  0.976559  26581.100
#> 157        NA         NA                <NA>        Ask  0.976703   4214.044
#> 158        NA         NA                <NA>        Ask  0.976735   8865.775
#> 159        NA         NA                <NA>        Ask  0.977313  21697.190
#> 160        NA         NA                <NA>        Ask  0.977800    138.833
#> 161        NA         NA                <NA>        Ask  0.977850   7245.242
#> 162        NA         NA                <NA>        Ask  0.978020   2214.007
#> 163        NA         NA                <NA>        Ask  0.978430   2405.784
#> 164        NA         NA                <NA>        Ask  0.978511 288251.286
#> 165        NA         NA                <NA>        Ask  0.978512  14630.224
#> 166        NA         NA                <NA>        Ask  0.979279   3636.565
#> 167        NA         NA                <NA>        Ask  0.979291  12732.200
#> 168        NA         NA                <NA>        Ask  0.980007   3043.122
#> 169        NA         NA                <NA>        Ask  0.981969     24.102
#> 170        NA         NA                <NA>        Ask  0.982098    138.225
#> 171        NA         NA                <NA>        Ask  0.983536     99.610
#> 172        NA         NA                <NA>        Ask  0.984755   6487.000
#> 173        NA         NA                <NA>        Ask  0.985264     48.681
#> 174        NA         NA                <NA>        Ask  0.986014     11.586
#> 175        NA         NA                <NA>        Ask  0.986415    137.620
#> 176        NA         NA                <NA>        Ask  0.987928  14490.789
#> 177        NA         NA                <NA>        Ask  0.987979     23.956
#> 178        NA         NA                <NA>        Ask  0.988173   1566.619
#> 179        NA         NA                <NA>        Ask  0.988174   2999.988
#> 180        NA         NA                <NA>        Ask  0.988552   1716.359
#> 181        NA         NA                <NA>        Ask  0.988893   7927.653
#> 182        NA         NA                <NA>        Ask  0.989589    792.648
#> 183        NA         NA                <NA>        Ask  0.989625   5463.282
#> 184        NA         NA                <NA>        Ask  0.990750    137.018
#> 185        NA         NA                <NA>        Ask  0.991454     11.529
#> 186        NA         NA                <NA>        Ask  0.992245     31.277
#> 187        NA         NA                <NA>        Ask  0.992952   6626.100
#> 188        NA         NA                <NA>        Ask  0.993116    787.990
#> 189        NA         NA                <NA>        Ask  0.993136     55.456
#> 190        NA         NA                <NA>        Ask  0.993224     78.674
#> 191        NA         NA                <NA>        Ask  0.993634     48.271
#> 192        NA         NA                <NA>        Ask  0.994000     41.491
#> 193        NA         NA                <NA>        Ask  0.994025     23.810
#> 194        NA         NA                <NA>        Ask  0.994142     98.547
#> 195        NA         NA                <NA>        Ask  0.995104    136.419
#> 196        NA         NA                <NA>        Ask  0.996211     98.300
#> 197        NA         NA                <NA>        Ask  0.996922     11.466
#> 198        NA         NA                <NA>        Ask  0.997223  52875.474
#> 199        NA         NA                <NA>        Ask  0.997597     16.698
#> 200        NA         NA                <NA>        Ask  0.997598  24630.935
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:25:14
#> 102 2024-12-18 16:25:14
#> 103 2024-12-18 16:25:14
#> 104 2024-12-18 16:25:14
#> 105 2024-12-18 16:25:13
#> 106 2024-12-18 16:25:13
#> 107 2024-12-18 16:25:11
#> 108 2024-12-18 16:25:13
#> 109 2024-12-18 16:25:13
#> 110 2024-12-18 16:25:14
#> 111 2024-12-18 16:25:14
#> 112 2024-12-18 16:25:13
#> 113 2024-12-18 16:25:13
#> 114 2024-12-18 16:25:14
#> 115 2024-12-18 16:25:11
#> 116 2024-12-18 16:25:12
#> 117 2024-12-18 16:25:14
#> 118 2024-12-18 16:25:12
#> 119 2024-12-18 16:25:12
#> 120 2024-12-18 16:25:12
#> 121 2024-12-18 16:25:11
#> 122 2024-12-18 16:25:13
#> 123 2024-12-18 16:25:11
#> 124 2024-12-18 16:25:07
#> 125 2024-12-18 16:25:12
#> 126 2024-12-18 16:25:10
#> 127 2024-12-18 16:25:12
#> 128 2024-12-18 16:25:01
#> 129 2024-12-18 16:25:13
#> 130 2024-12-18 16:25:13
#> 131 2024-12-18 16:25:11
#> 132 2024-12-18 16:25:04
#> 133 2024-12-18 16:25:10
#> 134 2024-12-18 16:25:04
#> 135 2024-12-18 16:25:07
#> 136 2024-12-18 16:25:10
#> 137 2024-12-18 16:25:07
#> 138 2024-12-18 16:25:04
#> 139 2024-12-18 16:25:07
#> 140 2024-12-18 16:24:54
#> 141 2024-12-18 16:24:49
#> 142 2024-12-18 16:24:55
#> 143 2024-12-18 16:25:04
#> 144 2024-12-18 16:24:58
#> 145 2024-12-18 16:25:04
#> 146 2024-12-18 16:24:54
#> 147 2024-12-18 16:24:54
#> 148 2024-12-18 16:24:52
#> 149 2024-12-18 16:23:04
#> 150 2024-12-18 16:25:10
#> 151 2024-12-18 16:22:56
#> 152 2024-12-18 16:25:01
#> 153 2024-12-18 16:25:07
#> 154 2024-12-18 16:22:03
#> 155 2024-12-18 16:24:54
#> 156 2024-12-18 16:22:56
#> 157 2024-12-18 16:22:56
#> 158 2024-12-18 16:24:57
#> 159 2024-12-18 16:24:57
#> 160 2024-12-18 16:17:28
#> 161 2024-12-18 16:24:57
#> 162 2024-12-18 16:24:52
#> 163 2024-12-18 16:25:08
#> 164 2024-12-18 16:25:06
#> 165 2024-12-18 16:24:51
#> 166 2024-12-18 16:24:55
#> 167 2024-12-18 16:22:56
#> 168 2024-12-18 16:24:37
#> 169 2024-12-18 16:08:58
#> 170 2024-12-18 14:56:28
#> 171 2024-12-18 15:57:30
#> 172 2024-12-18 16:23:04
#> 173 2024-12-18 14:56:59
#> 174 2024-12-18 15:09:59
#> 175 2024-12-18 15:24:42
#> 176 2024-12-18 16:24:50
#> 177 2024-12-18 14:35:26
#> 178 2024-12-18 16:18:14
#> 179 2024-12-18 16:18:12
#> 180 2024-12-18 16:16:40
#> 181 2024-12-18 16:16:28
#> 182 2024-12-18 16:16:25
#> 183 2024-12-18 16:16:22
#> 184 2024-12-18 13:57:45
#> 185 2024-12-18 12:37:59
#> 186 2024-12-18 15:04:40
#> 187 2024-12-18 16:22:56
#> 188 2024-12-18 16:24:51
#> 189 2024-12-18 13:26:56
#> 190 2024-12-18 15:12:45
#> 191 2024-12-18 12:38:18
#> 192 2024-12-18 12:25:06
#> 193 2024-12-18 13:36:40
#> 194 2024-12-18 12:37:51
#> 195 2024-12-18 13:34:11
#> 196 2024-12-18 15:07:21
#> 197 2024-12-18 12:33:01
#> 198 2024-12-18 16:25:06
#> 199 2024-12-18 16:25:04
#> 200 2024-12-18 16:24:54