Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 2    0.958430   5655.280 2024-12-18 16:30:01        Bid        NA         NA
#> 3    0.958711   1616.828 2024-12-18 16:30:01        Bid        NA         NA
#> 4    0.958713   4875.193 2024-12-18 16:30:01        Bid        NA         NA
#> 5    0.958753   9818.348 2024-12-18 16:30:02        Bid        NA         NA
#> 6    0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 7    0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 8    0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 9    0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 10   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 11   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 12   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 13   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 14   0.960231   6106.993 2024-12-18 16:29:35        Bid        NA         NA
#> 15   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 16   0.960714  12680.200 2024-12-18 16:29:32        Bid        NA         NA
#> 17   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 18   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 19   0.961463  59317.270 2024-12-18 16:23:08        Bid        NA         NA
#> 20   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 21   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 22   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 23   0.963101   2031.458 2024-12-18 16:37:57        Bid        NA         NA
#> 24   0.963760  12454.707 2024-12-18 16:30:33        Bid        NA         NA
#> 25   0.963965  25256.400 2024-12-18 16:30:25        Bid        NA         NA
#> 26   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 27   0.965815   3134.532 2024-12-18 16:37:46        Bid        NA         NA
#> 28   0.965839  25124.632 2024-12-18 16:30:16        Bid        NA         NA
#> 29   0.966396  38476.100 2024-12-18 16:30:25        Bid        NA         NA
#> 30   0.966397   1967.323 2024-12-18 16:36:44        Bid        NA         NA
#> 31   0.967472  19207.356 2024-12-18 16:32:03        Bid        NA         NA
#> 32   0.967487   3737.933 2024-12-18 16:37:43        Bid        NA         NA
#> 33   0.967488 289440.326 2024-12-18 16:38:00        Bid        NA         NA
#> 34   0.968222  14790.500 2024-12-18 16:37:30        Bid        NA         NA
#> 35   0.968421   8135.104 2024-12-18 16:37:38        Bid        NA         NA
#> 36   0.968488  38248.600 2024-12-18 16:35:55        Bid        NA         NA
#> 37   0.968596  20244.813 2024-12-18 16:37:57        Bid        NA         NA
#> 38   0.968962   3306.562 2024-12-18 16:37:13        Bid        NA         NA
#> 39   0.969597   2267.497 2024-12-18 16:37:54        Bid        NA         NA
#> 40   0.969599   2464.117 2024-12-18 16:37:54        Bid        NA         NA
#> 41   0.969633  15720.934 2024-12-18 16:37:51        Bid        NA         NA
#> 42   0.969787  14766.637 2024-12-18 16:37:54        Bid        NA         NA
#> 43   0.969970  15189.107 2024-12-18 16:33:45        Bid        NA         NA
#> 44   0.970002     20.637 2024-12-18 16:38:00        Bid        NA         NA
#> 45   0.970018   9699.535 2024-12-18 16:37:58        Bid        NA         NA
#> 46   0.970019     15.538 2024-12-18 16:38:00        Bid        NA         NA
#> 47   0.970294     44.150 2024-12-18 16:37:46        Bid        NA         NA
#> 48   0.970338     18.867 2024-12-18 16:37:54        Bid        NA         NA
#> 49   0.971567     43.702 2024-12-18 16:37:16        Bid        NA         NA
#> 50   0.971590 203789.585 2024-12-18 16:37:47        Bid        NA         NA
#> 51   0.971591     97.218 2024-12-18 16:37:55        Bid        NA         NA
#> 52   0.971638  16487.736 2024-12-18 16:37:58        Bid        NA         NA
#> 53   0.971640  12607.623 2024-12-18 16:37:58        Bid        NA         NA
#> 54   0.971641   6855.549 2024-12-18 16:37:58        Bid        NA         NA
#> 55   0.972357  14728.449 2024-12-18 16:37:58        Bid        NA         NA
#> 56   0.972405   8728.227 2024-12-18 16:37:58        Bid        NA         NA
#> 57   0.972668   2358.069 2024-12-18 16:37:31        Bid        NA         NA
#> 58   0.972716   5817.875 2024-12-18 16:37:58        Bid        NA         NA
#> 59   0.972766     25.700 2024-12-18 16:35:22        Bid        NA         NA
#> 60   0.973106  17418.600 2024-12-18 16:37:36        Bid        NA         NA
#> 61   0.973107     22.783 2024-12-18 16:37:55        Bid        NA         NA
#> 62   0.974046   5000.000 2024-12-18 16:37:46        Bid        NA         NA
#> 63   0.974210  14699.595 2024-12-18 16:37:30        Bid        NA         NA
#> 64   0.974716   2755.703 2024-12-18 16:37:47        Bid        NA         NA
#> 65   0.974848     14.454 2024-12-18 16:37:45        Bid        NA         NA
#> 66   0.975086   7566.874 2024-12-18 16:37:44        Bid        NA         NA
#> 67   0.975233  20804.254 2024-12-18 16:37:45        Bid        NA         NA
#> 68   0.975240  14684.063 2024-12-18 16:37:47        Bid        NA         NA
#> 69   0.975551   2444.808 2024-12-18 16:37:29        Bid        NA         NA
#> 70   0.975683   5006.639 2024-12-18 16:37:47        Bid        NA         NA
#> 71   0.975858   9673.500 2024-12-18 16:37:45        Bid        NA         NA
#> 72   0.975866  10661.250 2024-12-18 16:37:52        Bid        NA         NA
#> 73   0.975928  10784.330 2024-12-18 16:37:47        Bid        NA         NA
#> 74   0.975939  14674.383 2024-12-18 16:37:58        Bid        NA         NA
#> 75   0.975956  10624.280 2024-12-18 16:37:47        Bid        NA         NA
#> 76   0.975994   5383.207 2024-12-18 16:37:49        Bid        NA         NA
#> 77   0.975997   2944.000 2024-12-18 16:37:48        Bid        NA         NA
#> 78   0.976059   2451.000 2024-12-18 16:37:49        Bid        NA         NA
#> 79   0.976060  10361.735 2024-12-18 16:37:52        Bid        NA         NA
#> 80   0.976156  16025.100 2024-12-18 16:37:58        Bid        NA         NA
#> 81   0.976185   5405.322 2024-12-18 16:37:45        Bid        NA         NA
#> 82   0.976293   9778.814 2024-12-18 16:37:53        Bid        NA         NA
#> 83   0.976294    588.148 2024-12-18 16:37:59        Bid        NA         NA
#> 84   0.976936     16.697 2024-12-18 16:37:54        Bid        NA         NA
#> 85   0.977035   9771.950 2024-12-18 16:37:58        Bid        NA         NA
#> 86   0.977217    258.785 2024-12-18 16:37:54        Bid        NA         NA
#> 87   0.977300   2100.000 2024-12-18 16:37:55        Bid        NA         NA
#> 88   0.977334   1628.002 2024-12-18 16:38:00        Bid        NA         NA
#> 89   0.977369   9769.688 2024-12-18 16:38:00        Bid        NA         NA
#> 90   0.977380    633.600 2024-12-18 16:38:00        Bid        NA         NA
#> 91   0.977381    498.212 2024-12-18 16:38:00        Bid        NA         NA
#> 92   0.977435   4884.258 2024-12-18 16:38:00        Bid        NA         NA
#> 93   0.977492   1560.000 2024-12-18 16:37:58        Bid        NA         NA
#> 94   0.977493   3576.965 2024-12-18 16:38:00        Bid        NA         NA
#> 95   0.977494   5876.592 2024-12-18 16:38:00        Bid        NA         NA
#> 96   0.977495   3576.965 2024-12-18 16:38:00        Bid        NA         NA
#> 97   0.977497     33.799 2024-12-18 16:37:59        Bid        NA         NA
#> 98   0.977499   2442.098 2024-12-18 16:38:00        Bid        NA         NA
#> 99   0.977500   2100.000 2024-12-18 16:38:00        Bid        NA         NA
#> 100  0.977501    276.611 2024-12-18 16:38:00        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.977799     87.466
#> 102        NA         NA                <NA>        Ask  0.977800    138.833
#> 103        NA         NA                <NA>        Ask  0.978192   1391.714
#> 104        NA         NA                <NA>        Ask  0.978193   4880.731
#> 105        NA         NA                <NA>        Ask  0.978211    526.682
#> 106        NA         NA                <NA>        Ask  0.978391   1218.366
#> 107        NA         NA                <NA>        Ask  0.978407     49.000
#> 108        NA         NA                <NA>        Ask  0.978466  14741.165
#> 109        NA         NA                <NA>        Ask  0.978475   2098.800
#> 110        NA         NA                <NA>        Ask  0.978532  10000.000
#> 111        NA         NA                <NA>        Ask  0.978633     43.935
#> 112        NA         NA                <NA>        Ask  0.978683    633.700
#> 113        NA         NA                <NA>        Ask  0.978705   9755.289
#> 114        NA         NA                <NA>        Ask  0.978794   4839.316
#> 115        NA         NA                <NA>        Ask  0.978929   9671.300
#> 116        NA         NA                <NA>        Ask  0.979073  10705.500
#> 117        NA         NA                <NA>        Ask  0.979224   2755.703
#> 118        NA         NA                <NA>        Ask  0.979249   9749.872
#> 119        NA         NA                <NA>        Ask  0.979351  15299.700
#> 120        NA         NA                <NA>        Ask  0.979436   5000.000
#> 121        NA         NA                <NA>        Ask  0.979437   9604.050
#> 122        NA         NA                <NA>        Ask  0.979727   8083.660
#> 123        NA         NA                <NA>        Ask  0.979733    461.818
#> 124        NA         NA                <NA>        Ask  0.979734   1464.172
#> 125        NA         NA                <NA>        Ask  0.979735   7817.490
#> 126        NA         NA                <NA>        Ask  0.979865  14614.762
#> 127        NA         NA                <NA>        Ask  0.979871     14.769
#> 128        NA         NA                <NA>        Ask  0.980208  14609.656
#> 129        NA         NA                <NA>        Ask  0.980968   4066.113
#> 130        NA         NA                <NA>        Ask  0.981017  15880.171
#> 131        NA         NA                <NA>        Ask  0.981066   2710.742
#> 132        NA         NA                <NA>        Ask  0.981075  17408.300
#> 133        NA         NA                <NA>        Ask  0.981115  14596.155
#> 134        NA         NA                <NA>        Ask  0.981322   7454.539
#> 135        NA         NA                <NA>        Ask  0.981371  39764.400
#> 136        NA         NA                <NA>        Ask  0.981649   4743.798
#> 137        NA         NA                <NA>        Ask  0.981698  15773.184
#> 138        NA         NA                <NA>        Ask  0.981764   5421.483
#> 139        NA         NA                <NA>        Ask  0.981813  18821.704
#> 140        NA         NA                <NA>        Ask  0.981883     10.229
#> 141        NA         NA                <NA>        Ask  0.981969     24.102
#> 142        NA         NA                <NA>        Ask  0.982098    138.225
#> 143        NA         NA                <NA>        Ask  0.982539   6099.169
#> 144        NA         NA                <NA>        Ask  0.983277   6776.854
#> 145        NA         NA                <NA>        Ask  0.983326  37945.700
#> 146        NA         NA                <NA>        Ask  0.983466  14561.263
#> 147        NA         NA                <NA>        Ask  0.983536     99.610
#> 148        NA         NA                <NA>        Ask  0.983841   4758.558
#> 149        NA         NA                <NA>        Ask  0.983951  24826.063
#> 150        NA         NA                <NA>        Ask  0.984234   8915.066
#> 151        NA         NA                <NA>        Ask  0.984235  26581.100
#> 152        NA         NA                <NA>        Ask  0.984951   4214.044
#> 153        NA         NA                <NA>        Ask  0.985132  14538.431
#> 154        NA         NA                <NA>        Ask  0.985264     48.681
#> 155        NA         NA                <NA>        Ask  0.985476  11717.300
#> 156        NA         NA                <NA>        Ask  0.986013     25.671
#> 157        NA         NA                <NA>        Ask  0.986014     11.586
#> 158        NA         NA                <NA>        Ask  0.986041   2423.029
#> 159        NA         NA                <NA>        Ask  0.986319    171.688
#> 160        NA         NA                <NA>        Ask  0.986320  12732.200
#> 161        NA         NA                <NA>        Ask  0.986415    137.620
#> 162        NA         NA                <NA>        Ask  0.986715  21974.538
#> 163        NA         NA                <NA>        Ask  0.986819   2227.923
#> 164        NA         NA                <NA>        Ask  0.986943   7323.313
#> 165        NA         NA                <NA>        Ask  0.986982   3665.202
#> 166        NA         NA                <NA>        Ask  0.987101  14507.633
#> 167        NA         NA                <NA>        Ask  0.987883 289244.504
#> 168        NA         NA                <NA>        Ask  0.987884   3063.670
#> 169        NA         NA                <NA>        Ask  0.987979     23.956
#> 170        NA         NA                <NA>        Ask  0.989585   6487.000
#> 171        NA         NA                <NA>        Ask  0.990616   5471.535
#> 172        NA         NA                <NA>        Ask  0.990737    794.682
#> 173        NA         NA                <NA>        Ask  0.990750    137.018
#> 174        NA         NA                <NA>        Ask  0.991454     11.529
#> 175        NA         NA                <NA>        Ask  0.992245     31.277
#> 176        NA         NA                <NA>        Ask  0.993136     55.456
#> 177        NA         NA                <NA>        Ask  0.993224     78.674
#> 178        NA         NA                <NA>        Ask  0.993634     48.271
#> 179        NA         NA                <NA>        Ask  0.994000     41.491
#> 180        NA         NA                <NA>        Ask  0.994025     23.810
#> 181        NA         NA                <NA>        Ask  0.994142     98.547
#> 182        NA         NA                <NA>        Ask  0.994172   6271.239
#> 183        NA         NA                <NA>        Ask  0.994183   7903.606
#> 184        NA         NA                <NA>        Ask  0.994483    793.091
#> 185        NA         NA                <NA>        Ask  0.995104    136.419
#> 186        NA         NA                <NA>        Ask  0.996211     98.300
#> 187        NA         NA                <NA>        Ask  0.996922     11.466
#> 188        NA         NA                <NA>        Ask  0.998600  14340.571
#> 189        NA         NA                <NA>        Ask  0.998690      7.057
#> 190        NA         NA                <NA>        Ask  0.999478    135.822
#> 191        NA         NA                <NA>        Ask  1.000000   3562.826
#> 192        NA         NA                <NA>        Ask  1.000109     23.665
#> 193        NA         NA                <NA>        Ask  1.000330   6626.100
#> 194        NA         NA                <NA>        Ask  1.001000      8.018
#> 195        NA         NA                <NA>        Ask  1.001150   6390.800
#> 196        NA         NA                <NA>        Ask  1.001581    109.838
#> 197        NA         NA                <NA>        Ask  1.002077     47.865
#> 198        NA         NA                <NA>        Ask  1.002244     84.063
#> 199        NA         NA                <NA>        Ask  1.002364     21.780
#> 200        NA         NA                <NA>        Ask  1.002422     11.396
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:37:59
#> 102 2024-12-18 16:17:28
#> 103 2024-12-18 16:38:00
#> 104 2024-12-18 16:38:00
#> 105 2024-12-18 16:37:54
#> 106 2024-12-18 16:37:59
#> 107 2024-12-18 16:37:59
#> 108 2024-12-18 16:37:54
#> 109 2024-12-18 16:37:58
#> 110 2024-12-18 16:37:59
#> 111 2024-12-18 16:37:12
#> 112 2024-12-18 16:37:58
#> 113 2024-12-18 16:37:58
#> 114 2024-12-18 16:37:51
#> 115 2024-12-18 16:37:47
#> 116 2024-12-18 16:37:53
#> 117 2024-12-18 16:37:52
#> 118 2024-12-18 16:37:58
#> 119 2024-12-18 16:37:53
#> 120 2024-12-18 16:37:52
#> 121 2024-12-18 16:37:50
#> 122 2024-12-18 16:37:54
#> 123 2024-12-18 16:37:57
#> 124 2024-12-18 16:37:59
#> 125 2024-12-18 16:37:50
#> 126 2024-12-18 16:37:48
#> 127 2024-12-18 16:37:54
#> 128 2024-12-18 16:37:53
#> 129 2024-12-18 16:38:00
#> 130 2024-12-18 16:37:52
#> 131 2024-12-18 16:38:00
#> 132 2024-12-18 16:37:47
#> 133 2024-12-18 16:37:52
#> 134 2024-12-18 16:38:00
#> 135 2024-12-18 16:37:45
#> 136 2024-12-18 16:38:00
#> 137 2024-12-18 16:37:54
#> 138 2024-12-18 16:38:00
#> 139 2024-12-18 16:37:45
#> 140 2024-12-18 16:37:45
#> 141 2024-12-18 16:27:13
#> 142 2024-12-18 14:56:28
#> 143 2024-12-18 16:38:00
#> 144 2024-12-18 16:38:00
#> 145 2024-12-18 16:37:45
#> 146 2024-12-18 16:37:54
#> 147 2024-12-18 15:57:30
#> 148 2024-12-18 16:37:55
#> 149 2024-12-18 16:37:52
#> 150 2024-12-18 16:37:51
#> 151 2024-12-18 16:36:13
#> 152 2024-12-18 16:37:30
#> 153 2024-12-18 16:36:08
#> 154 2024-12-18 14:56:59
#> 155 2024-12-18 16:37:51
#> 156 2024-12-18 16:37:53
#> 157 2024-12-18 15:09:59
#> 158 2024-12-18 16:37:55
#> 159 2024-12-18 16:37:53
#> 160 2024-12-18 16:27:13
#> 161 2024-12-18 15:24:42
#> 162 2024-12-18 16:37:49
#> 163 2024-12-18 16:37:55
#> 164 2024-12-18 16:37:51
#> 165 2024-12-18 16:37:53
#> 166 2024-12-18 16:37:42
#> 167 2024-12-18 16:37:51
#> 168 2024-12-18 16:37:21
#> 169 2024-12-18 14:35:26
#> 170 2024-12-18 16:27:02
#> 171 2024-12-18 16:30:06
#> 172 2024-12-18 16:30:05
#> 173 2024-12-18 13:57:45
#> 174 2024-12-18 12:37:59
#> 175 2024-12-18 15:04:40
#> 176 2024-12-18 13:26:56
#> 177 2024-12-18 15:12:45
#> 178 2024-12-18 12:38:18
#> 179 2024-12-18 12:25:06
#> 180 2024-12-18 13:36:40
#> 181 2024-12-18 12:37:51
#> 182 2024-12-18 16:37:46
#> 183 2024-12-18 16:37:47
#> 184 2024-12-18 16:37:50
#> 185 2024-12-18 13:34:11
#> 186 2024-12-18 15:07:21
#> 187 2024-12-18 12:33:01
#> 188 2024-12-18 16:37:53
#> 189 2024-12-18 12:34:04
#> 190 2024-12-18 13:11:41
#> 191 2024-12-18 14:57:48
#> 192 2024-12-18 12:15:38
#> 193 2024-12-18 16:37:45
#> 194 2024-12-18 12:22:05
#> 195 2024-12-18 16:27:06
#> 196 2024-12-18 15:18:26
#> 197 2024-12-18 04:26:55
#> 198 2024-12-18 07:39:44
#> 199 2024-12-18 12:02:37
#> 200 2024-12-18 03:31:33
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.955890     31.385 2024-12-18 13:29:07        Bid        NA         NA
#> 2    0.955970     86.060 2024-12-11 11:33:28        Bid        NA         NA
#> 3    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 4    0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 5    0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 6    0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 7    0.956146  14977.304 2024-12-18 16:37:32        Bid        NA         NA
#> 8    0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 9    0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 10   0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 11   0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 12   0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 13   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 14   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 15   0.958430   5655.280 2024-12-18 16:30:01        Bid        NA         NA
#> 16   0.958711   1616.828 2024-12-18 16:30:01        Bid        NA         NA
#> 17   0.958713   4875.193 2024-12-18 16:30:01        Bid        NA         NA
#> 18   0.958753   9818.348 2024-12-18 16:30:02        Bid        NA         NA
#> 19   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 20   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 21   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 22   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 23   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 24   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 25   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 26   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 27   0.960231   6106.993 2024-12-18 16:29:35        Bid        NA         NA
#> 28   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 29   0.960714  12680.200 2024-12-18 16:29:32        Bid        NA         NA
#> 30   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 31   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 32   0.961463  59317.270 2024-12-18 16:23:08        Bid        NA         NA
#> 33   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 34   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 35   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 36   0.963101   2031.458 2024-12-18 16:37:57        Bid        NA         NA
#> 37   0.963760  12454.707 2024-12-18 16:30:33        Bid        NA         NA
#> 38   0.963965  25256.400 2024-12-18 16:30:25        Bid        NA         NA
#> 39   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 40   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 41   0.965815   3134.532 2024-12-18 16:37:46        Bid        NA         NA
#> 42   0.965839  25124.632 2024-12-18 16:30:16        Bid        NA         NA
#> 43   0.966396  38476.100 2024-12-18 16:30:25        Bid        NA         NA
#> 44   0.966397   1967.323 2024-12-18 16:36:44        Bid        NA         NA
#> 45   0.967487   3737.933 2024-12-18 16:37:43        Bid        NA         NA
#> 46   0.968222  14790.500 2024-12-18 16:37:30        Bid        NA         NA
#> 47   0.968421   8135.104 2024-12-18 16:37:38        Bid        NA         NA
#> 48   0.968596  20244.813 2024-12-18 16:37:57        Bid        NA         NA
#> 49   0.968861  19207.356 2024-12-18 16:38:02        Bid        NA         NA
#> 50   0.968962   3306.562 2024-12-18 16:37:13        Bid        NA         NA
#> 51   0.969597   2267.497 2024-12-18 16:37:54        Bid        NA         NA
#> 52   0.969599   2464.117 2024-12-18 16:37:54        Bid        NA         NA
#> 53   0.969633  15720.934 2024-12-18 16:37:51        Bid        NA         NA
#> 54   0.969787  53015.237 2024-12-18 16:38:04        Bid        NA         NA
#> 55   0.969970  15189.107 2024-12-18 16:33:45        Bid        NA         NA
#> 56   0.970018   9699.535 2024-12-18 16:37:58        Bid        NA         NA
#> 57   0.970294     44.150 2024-12-18 16:37:46        Bid        NA         NA
#> 58   0.970338     18.867 2024-12-18 16:37:54        Bid        NA         NA
#> 59   0.971567     43.702 2024-12-18 16:37:16        Bid        NA         NA
#> 60   0.971590 203789.585 2024-12-18 16:37:47        Bid        NA         NA
#> 61   0.971591     97.218 2024-12-18 16:37:55        Bid        NA         NA
#> 62   0.971638  16487.736 2024-12-18 16:37:58        Bid        NA         NA
#> 63   0.971640  12607.623 2024-12-18 16:37:58        Bid        NA         NA
#> 64   0.971641   6917.427 2024-12-18 16:38:02        Bid        NA         NA
#> 65   0.971642     78.725 2024-12-18 16:38:04        Bid        NA         NA
#> 66   0.972357  14728.449 2024-12-18 16:37:58        Bid        NA         NA
#> 67   0.972405   8728.227 2024-12-18 16:37:58        Bid        NA         NA
#> 68   0.972668   2358.069 2024-12-18 16:37:31        Bid        NA         NA
#> 69   0.972716   5817.875 2024-12-18 16:37:58        Bid        NA         NA
#> 70   0.972766     25.700 2024-12-18 16:35:22        Bid        NA         NA
#> 71   0.973107     22.783 2024-12-18 16:37:55        Bid        NA         NA
#> 72   0.973108   5000.000 2024-12-18 16:38:05        Bid        NA         NA
#> 73   0.974210  14699.595 2024-12-18 16:37:30        Bid        NA         NA
#> 74   0.974560  17390.200 2024-12-18 16:38:03        Bid        NA         NA
#> 75   0.974698   2755.703 2024-12-18 16:38:05        Bid        NA         NA
#> 76   0.974848     14.454 2024-12-18 16:37:45        Bid        NA         NA
#> 77   0.975041  14689.516 2024-12-18 16:38:04        Bid        NA         NA
#> 78   0.975233  20804.254 2024-12-18 16:37:45        Bid        NA         NA
#> 79   0.975240  14684.063 2024-12-18 16:37:47        Bid        NA         NA
#> 80   0.975551   2444.808 2024-12-18 16:37:29        Bid        NA         NA
#> 81   0.975557   7566.874 2024-12-18 16:38:01        Bid        NA         NA
#> 82   0.975858   9673.500 2024-12-18 16:37:45        Bid        NA         NA
#> 83   0.975866  10661.250 2024-12-18 16:37:52        Bid        NA         NA
#> 84   0.975928  10784.330 2024-12-18 16:37:47        Bid        NA         NA
#> 85   0.975956  10624.280 2024-12-18 16:37:47        Bid        NA         NA
#> 86   0.976046   5006.639 2024-12-18 16:38:01        Bid        NA         NA
#> 87   0.976060  10361.735 2024-12-18 16:37:52        Bid        NA         NA
#> 88   0.976102   9782.362 2024-12-18 16:38:04        Bid        NA         NA
#> 89   0.976156  16025.100 2024-12-18 16:37:58        Bid        NA         NA
#> 90   0.976356    588.645 2024-12-18 16:38:05        Bid        NA         NA
#> 91   0.976371   2944.000 2024-12-18 16:38:04        Bid        NA         NA
#> 92   0.976456   4904.000 2024-12-18 16:38:04        Bid        NA         NA
#> 93   0.976473   2451.000 2024-12-18 16:38:04        Bid        NA         NA
#> 94   0.976591    470.677 2024-12-18 16:38:04        Bid        NA         NA
#> 95   0.976593    514.472 2024-12-18 16:38:04        Bid        NA         NA
#> 96   0.976619     16.697 2024-12-18 16:38:04        Bid        NA         NA
#> 97   0.976620   5405.322 2024-12-18 16:38:04        Bid        NA         NA
#> 98   0.976622   7332.871 2024-12-18 16:38:05        Bid        NA         NA
#> 99   0.976918    633.500 2024-12-18 16:38:05        Bid        NA         NA
#> 100  0.977116    297.050 2024-12-18 16:38:05        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.977794   1391.714
#> 102        NA         NA                <NA>        Ask  0.977799   7411.509
#> 103        NA         NA                <NA>        Ask  0.977800    138.833
#> 104        NA         NA                <NA>        Ask  0.977886    958.142
#> 105        NA         NA                <NA>        Ask  0.978049   1218.366
#> 106        NA         NA                <NA>        Ask  0.978071   1628.002
#> 107        NA         NA                <NA>        Ask  0.978211    526.682
#> 108        NA         NA                <NA>        Ask  0.978229   9761.107
#> 109        NA         NA                <NA>        Ask  0.978261    633.500
#> 110        NA         NA                <NA>        Ask  0.978262   1628.002
#> 111        NA         NA                <NA>        Ask  0.978378  10000.000
#> 112        NA         NA                <NA>        Ask  0.978499  15544.816
#> 113        NA         NA                <NA>        Ask  0.978500   1391.717
#> 114        NA         NA                <NA>        Ask  0.978503  14738.443
#> 115        NA         NA                <NA>        Ask  0.978509     49.000
#> 116        NA         NA                <NA>        Ask  0.978557   2127.859
#> 117        NA         NA                <NA>        Ask  0.978633     43.935
#> 118        NA         NA                <NA>        Ask  0.978873  15299.700
#> 119        NA         NA                <NA>        Ask  0.978929   9671.300
#> 120        NA         NA                <NA>        Ask  0.979127   2944.000
#> 121        NA         NA                <NA>        Ask  0.979174   4904.000
#> 122        NA         NA                <NA>        Ask  0.979224   2755.703
#> 123        NA         NA                <NA>        Ask  0.979436   5000.000
#> 124        NA         NA                <NA>        Ask  0.979437   9604.050
#> 125        NA         NA                <NA>        Ask  0.979467  14623.154
#> 126        NA         NA                <NA>        Ask  0.979553     14.769
#> 127        NA         NA                <NA>        Ask  0.979726   1464.193
#> 128        NA         NA                <NA>        Ask  0.979727   8083.660
#> 129        NA         NA                <NA>        Ask  0.979735   7817.490
#> 130        NA         NA                <NA>        Ask  0.979865  14614.762
#> 131        NA         NA                <NA>        Ask  0.980206    469.828
#> 132        NA         NA                <NA>        Ask  0.980968   4066.113
#> 133        NA         NA                <NA>        Ask  0.981017  15880.171
#> 134        NA         NA                <NA>        Ask  0.981066   2710.742
#> 135        NA         NA                <NA>        Ask  0.981075  17408.300
#> 136        NA         NA                <NA>        Ask  0.981115  14596.155
#> 137        NA         NA                <NA>        Ask  0.981322   7454.539
#> 138        NA         NA                <NA>        Ask  0.981371  39764.400
#> 139        NA         NA                <NA>        Ask  0.981649   4743.798
#> 140        NA         NA                <NA>        Ask  0.981698  15773.184
#> 141        NA         NA                <NA>        Ask  0.981764   5421.483
#> 142        NA         NA                <NA>        Ask  0.981813  18821.704
#> 143        NA         NA                <NA>        Ask  0.981883     10.229
#> 144        NA         NA                <NA>        Ask  0.981969     24.102
#> 145        NA         NA                <NA>        Ask  0.982098    138.225
#> 146        NA         NA                <NA>        Ask  0.982539   6099.169
#> 147        NA         NA                <NA>        Ask  0.983277   6776.854
#> 148        NA         NA                <NA>        Ask  0.983326  37945.700
#> 149        NA         NA                <NA>        Ask  0.983466  14561.263
#> 150        NA         NA                <NA>        Ask  0.983536     99.610
#> 151        NA         NA                <NA>        Ask  0.983841   4758.558
#> 152        NA         NA                <NA>        Ask  0.983951  24826.063
#> 153        NA         NA                <NA>        Ask  0.984234    127.091
#> 154        NA         NA                <NA>        Ask  0.984235  26581.100
#> 155        NA         NA                <NA>        Ask  0.984950   8784.452
#> 156        NA         NA                <NA>        Ask  0.984951   4214.044
#> 157        NA         NA                <NA>        Ask  0.985132  14538.431
#> 158        NA         NA                <NA>        Ask  0.985264     48.681
#> 159        NA         NA                <NA>        Ask  0.985974  11712.603
#> 160        NA         NA                <NA>        Ask  0.986013     25.671
#> 161        NA         NA                <NA>        Ask  0.986014     11.586
#> 162        NA         NA                <NA>        Ask  0.986041   2423.029
#> 163        NA         NA                <NA>        Ask  0.986319    171.688
#> 164        NA         NA                <NA>        Ask  0.986320  12732.200
#> 165        NA         NA                <NA>        Ask  0.986415    137.620
#> 166        NA         NA                <NA>        Ask  0.986715  21974.538
#> 167        NA         NA                <NA>        Ask  0.986819   2227.923
#> 168        NA         NA                <NA>        Ask  0.986982   3665.202
#> 169        NA         NA                <NA>        Ask  0.987101  14507.633
#> 170        NA         NA                <NA>        Ask  0.987442   7320.377
#> 171        NA         NA                <NA>        Ask  0.987883 289244.504
#> 172        NA         NA                <NA>        Ask  0.987884   3063.670
#> 173        NA         NA                <NA>        Ask  0.987979     23.956
#> 174        NA         NA                <NA>        Ask  0.989585   6487.000
#> 175        NA         NA                <NA>        Ask  0.990616   5471.535
#> 176        NA         NA                <NA>        Ask  0.990737    794.682
#> 177        NA         NA                <NA>        Ask  0.990750    137.018
#> 178        NA         NA                <NA>        Ask  0.991454     11.529
#> 179        NA         NA                <NA>        Ask  0.992245     31.277
#> 180        NA         NA                <NA>        Ask  0.993136     55.456
#> 181        NA         NA                <NA>        Ask  0.993224     78.674
#> 182        NA         NA                <NA>        Ask  0.993634     48.271
#> 183        NA         NA                <NA>        Ask  0.994000     41.491
#> 184        NA         NA                <NA>        Ask  0.994025     23.810
#> 185        NA         NA                <NA>        Ask  0.994142     98.547
#> 186        NA         NA                <NA>        Ask  0.994172   6271.239
#> 187        NA         NA                <NA>        Ask  0.994183   7903.606
#> 188        NA         NA                <NA>        Ask  0.994483    793.091
#> 189        NA         NA                <NA>        Ask  0.995104    136.419
#> 190        NA         NA                <NA>        Ask  0.996211     98.300
#> 191        NA         NA                <NA>        Ask  0.996922     11.466
#> 192        NA         NA                <NA>        Ask  0.998600  14340.571
#> 193        NA         NA                <NA>        Ask  0.998690      7.057
#> 194        NA         NA                <NA>        Ask  0.999478    135.822
#> 195        NA         NA                <NA>        Ask  1.000000   3562.826
#> 196        NA         NA                <NA>        Ask  1.000109     23.665
#> 197        NA         NA                <NA>        Ask  1.000330   6626.100
#> 198        NA         NA                <NA>        Ask  1.001000      8.018
#> 199        NA         NA                <NA>        Ask  1.001150   6390.800
#> 200        NA         NA                <NA>        Ask  1.001581    109.838
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:38:05
#> 102 2024-12-18 16:38:05
#> 103 2024-12-18 16:17:28
#> 104 2024-12-18 16:38:05
#> 105 2024-12-18 16:38:04
#> 106 2024-12-18 16:38:05
#> 107 2024-12-18 16:37:54
#> 108 2024-12-18 16:38:04
#> 109 2024-12-18 16:38:04
#> 110 2024-12-18 16:38:04
#> 111 2024-12-18 16:38:04
#> 112 2024-12-18 16:38:04
#> 113 2024-12-18 16:38:04
#> 114 2024-12-18 16:38:03
#> 115 2024-12-18 16:38:04
#> 116 2024-12-18 16:38:04
#> 117 2024-12-18 16:37:12
#> 118 2024-12-18 16:38:04
#> 119 2024-12-18 16:37:47
#> 120 2024-12-18 16:38:04
#> 121 2024-12-18 16:38:04
#> 122 2024-12-18 16:37:52
#> 123 2024-12-18 16:37:52
#> 124 2024-12-18 16:37:50
#> 125 2024-12-18 16:38:04
#> 126 2024-12-18 16:38:04
#> 127 2024-12-18 16:38:00
#> 128 2024-12-18 16:37:54
#> 129 2024-12-18 16:37:50
#> 130 2024-12-18 16:37:48
#> 131 2024-12-18 16:38:04
#> 132 2024-12-18 16:38:00
#> 133 2024-12-18 16:37:52
#> 134 2024-12-18 16:38:00
#> 135 2024-12-18 16:37:47
#> 136 2024-12-18 16:37:52
#> 137 2024-12-18 16:38:00
#> 138 2024-12-18 16:37:45
#> 139 2024-12-18 16:38:00
#> 140 2024-12-18 16:37:54
#> 141 2024-12-18 16:38:00
#> 142 2024-12-18 16:37:45
#> 143 2024-12-18 16:37:45
#> 144 2024-12-18 16:27:13
#> 145 2024-12-18 14:56:28
#> 146 2024-12-18 16:38:00
#> 147 2024-12-18 16:38:00
#> 148 2024-12-18 16:37:45
#> 149 2024-12-18 16:37:54
#> 150 2024-12-18 15:57:30
#> 151 2024-12-18 16:37:55
#> 152 2024-12-18 16:37:52
#> 153 2024-12-18 16:38:03
#> 154 2024-12-18 16:36:13
#> 155 2024-12-18 16:38:03
#> 156 2024-12-18 16:37:30
#> 157 2024-12-18 16:36:08
#> 158 2024-12-18 14:56:59
#> 159 2024-12-18 16:38:03
#> 160 2024-12-18 16:37:53
#> 161 2024-12-18 15:09:59
#> 162 2024-12-18 16:37:55
#> 163 2024-12-18 16:37:53
#> 164 2024-12-18 16:27:13
#> 165 2024-12-18 15:24:42
#> 166 2024-12-18 16:37:49
#> 167 2024-12-18 16:37:55
#> 168 2024-12-18 16:37:53
#> 169 2024-12-18 16:37:42
#> 170 2024-12-18 16:38:03
#> 171 2024-12-18 16:37:51
#> 172 2024-12-18 16:37:21
#> 173 2024-12-18 14:35:26
#> 174 2024-12-18 16:27:02
#> 175 2024-12-18 16:30:06
#> 176 2024-12-18 16:30:05
#> 177 2024-12-18 13:57:45
#> 178 2024-12-18 12:37:59
#> 179 2024-12-18 15:04:40
#> 180 2024-12-18 13:26:56
#> 181 2024-12-18 15:12:45
#> 182 2024-12-18 12:38:18
#> 183 2024-12-18 12:25:06
#> 184 2024-12-18 13:36:40
#> 185 2024-12-18 12:37:51
#> 186 2024-12-18 16:37:46
#> 187 2024-12-18 16:37:47
#> 188 2024-12-18 16:37:50
#> 189 2024-12-18 13:34:11
#> 190 2024-12-18 15:07:21
#> 191 2024-12-18 12:33:01
#> 192 2024-12-18 16:37:53
#> 193 2024-12-18 12:34:04
#> 194 2024-12-18 13:11:41
#> 195 2024-12-18 14:57:48
#> 196 2024-12-18 12:15:38
#> 197 2024-12-18 16:37:45
#> 198 2024-12-18 12:22:05
#> 199 2024-12-18 16:27:06
#> 200 2024-12-18 15:18:26