Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.958245   6118.811 2024-12-18 16:38:47        Bid        NA         NA
#> 2    0.958484   6347.900 2024-12-18 16:38:27        Bid        NA         NA
#> 3    0.958673  14943.809 2024-12-18 16:41:04        Bid        NA         NA
#> 4    0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 5    0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 6    0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 7    0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 8    0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 9    0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 10   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 11   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 12   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 13   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 14   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 15   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 16   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 17   0.962599   5642.132 2024-12-18 16:40:45        Bid        NA         NA
#> 18   0.962605  12868.612 2024-12-18 16:40:28        Bid        NA         NA
#> 19   0.962607   1776.987 2024-12-18 16:40:29        Bid        NA         NA
#> 20   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 21   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 22   0.964601   3592.389 2024-12-18 16:43:27        Bid        NA         NA
#> 23   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 24   0.966830   6106.993 2024-12-18 16:43:28        Bid        NA         NA
#> 25   0.966928   1636.255 2024-12-18 16:42:01        Bid        NA         NA
#> 26   0.967894  12680.200 2024-12-18 16:45:05        Bid        NA         NA
#> 27   0.968229  14795.412 2024-12-18 16:43:53        Bid        NA         NA
#> 28   0.968700    138.916 2024-12-18 16:38:53        Bid        NA         NA
#> 29   0.968777   3131.503 2024-12-18 16:44:49        Bid        NA         NA
#> 30   0.969111  25256.400 2024-12-18 16:44:23        Bid        NA         NA
#> 31   0.969274  12454.707 2024-12-18 16:43:28        Bid        NA         NA
#> 32   0.970202   8124.487 2024-12-18 16:42:37        Bid        NA         NA
#> 33   0.970203 288615.107 2024-12-18 16:45:16        Bid        NA         NA
#> 34   0.970307   5042.692 2024-12-18 16:45:07        Bid        NA         NA
#> 35   0.970737   3734.342 2024-12-18 16:44:49        Bid        NA         NA
#> 36   0.970928  15115.990 2024-12-18 16:45:07        Bid        NA         NA
#> 37   0.971037   2465.990 2024-12-18 16:44:35        Bid        NA         NA
#> 38   0.971365   2269.505 2024-12-18 16:44:42        Bid        NA         NA
#> 39   0.971563  14744.496 2024-12-18 16:44:31        Bid        NA         NA
#> 40   0.971719  25124.632 2024-12-18 16:43:28        Bid        NA         NA
#> 41   0.972059  38476.100 2024-12-18 16:44:40        Bid        NA         NA
#> 42   0.972766     25.700 2024-12-18 16:35:22        Bid        NA         NA
#> 43   0.973185  19207.356 2024-12-18 16:43:28        Bid        NA         NA
#> 44   0.973233  17708.414 2024-12-18 16:45:21        Bid        NA         NA
#> 45   0.973234   1958.597 2024-12-18 16:45:07        Bid        NA         NA
#> 46   0.973643   3274.714 2024-12-18 16:45:11        Bid        NA         NA
#> 47   0.974058  58581.570 2024-12-18 16:39:56        Bid        NA         NA
#> 48   0.974105   7455.589 2024-12-18 16:45:21        Bid        NA         NA
#> 49   0.974106   6838.201 2024-12-18 16:45:20        Bid        NA         NA
#> 50   0.974267  15189.107 2024-12-18 16:43:01        Bid        NA         NA
#> 51   0.974314   6523.329 2024-12-18 16:45:21        Bid        NA         NA
#> 52   0.974590  15720.934 2024-12-18 16:44:18        Bid        NA         NA
#> 53   0.974638   5591.246 2024-12-18 16:45:21        Bid        NA         NA
#> 54   0.974672  14697.467 2024-12-18 16:44:34        Bid        NA         NA
#> 55   0.974719   4659.093 2024-12-18 16:45:21        Bid        NA         NA
#> 56   0.974798     49.000 2024-12-18 16:41:48        Bid        NA         NA
#> 57   0.974857  38248.600 2024-12-18 16:45:15        Bid        NA         NA
#> 58   0.974865 203105.008 2024-12-18 16:44:49        Bid        NA         NA
#> 59   0.974866    285.944 2024-12-18 16:45:18        Bid        NA         NA
#> 60   0.974912   9317.838 2024-12-18 16:45:21        Bid        NA         NA
#> 61   0.974913     32.924 2024-12-18 16:45:23        Bid        NA         NA
#> 62   0.976060     43.471 2024-12-18 16:45:19        Bid        NA         NA
#> 63   0.976274   2930.000 2024-12-18 16:43:50        Bid        NA         NA
#> 64   0.976621  17358.500 2024-12-18 16:44:42        Bid        NA         NA
#> 65   0.976623   5000.000 2024-12-18 16:45:16        Bid        NA         NA
#> 66   0.976624   2348.517 2024-12-18 16:45:23        Bid        NA         NA
#> 67   0.976955  14663.812 2024-12-18 16:45:16        Bid        NA         NA
#> 68   0.977069     14.454 2024-12-18 16:44:27        Bid        NA         NA
#> 69   0.977448  14656.418 2024-12-18 16:45:29        Bid        NA         NA
#> 70   0.977658   2755.703 2024-12-18 16:45:14        Bid        NA         NA
#> 71   0.977872  14650.132 2024-12-18 16:45:14        Bid        NA         NA
#> 72   0.977913  10451.560 2024-12-18 16:45:09        Bid        NA         NA
#> 73   0.978183   2440.000 2024-12-18 16:45:12        Bid        NA         NA
#> 74   0.978203   4881.000 2024-12-18 16:45:11        Bid        NA         NA
#> 75   0.978298  10719.530 2024-12-18 16:45:09        Bid        NA         NA
#> 76   0.978328  10994.910 2024-12-18 16:45:09        Bid        NA         NA
#> 77   0.978363    609.647 2024-12-18 16:45:12        Bid        NA         NA
#> 78   0.978686   9758.589 2024-12-18 16:45:20        Bid        NA         NA
#> 79   0.978717    617.100 2024-12-18 16:45:11        Bid        NA         NA
#> 80   0.978969     16.697 2024-12-18 16:45:20        Bid        NA         NA
#> 81   0.979008    584.902 2024-12-18 16:45:14        Bid        NA         NA
#> 82   0.979029  20804.254 2024-12-18 16:45:16        Bid        NA         NA
#> 83   0.979030   7566.874 2024-12-18 16:45:25        Bid        NA         NA
#> 84   0.979036   9755.095 2024-12-18 16:45:29        Bid        NA         NA
#> 85   0.979077   9638.900 2024-12-18 16:45:25        Bid        NA         NA
#> 86   0.979427  16025.100 2024-12-18 16:45:16        Bid        NA         NA
#> 87   0.979428   5006.639 2024-12-18 16:45:25        Bid        NA         NA
#> 88   0.979463   9750.850 2024-12-18 16:45:29        Bid        NA         NA
#> 89   0.979487  10000.000 2024-12-18 16:45:28        Bid        NA         NA
#> 90   0.979492   1628.002 2024-12-18 16:45:30        Bid        NA         NA
#> 91   0.979534    467.901 2024-12-18 16:45:15        Bid        NA         NA
#> 92   0.979580   7312.262 2024-12-18 16:45:29        Bid        NA         NA
#> 93   0.979638   3576.965 2024-12-18 16:45:30        Bid        NA         NA
#> 94   0.979643   4874.526 2024-12-18 16:45:29        Bid        NA         NA
#> 95   0.979650   5876.592 2024-12-18 16:45:29        Bid        NA         NA
#> 96   0.979651     33.799 2024-12-18 16:45:30        Bid        NA         NA
#> 97   0.979652    281.339 2024-12-18 16:45:30        Bid        NA         NA
#> 98   0.979775   5405.322 2024-12-18 16:45:22        Bid        NA         NA
#> 99   0.979806   6491.225 2024-12-18 16:45:29        Bid        NA         NA
#> 100  0.979807     33.799 2024-12-18 16:45:29        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.980000     41.491
#> 102        NA         NA                <NA>        Ask  0.980187   1382.919
#> 103        NA         NA                <NA>        Ask  0.980188   2435.912
#> 104        NA         NA                <NA>        Ask  0.980245   4871.537
#> 105        NA         NA                <NA>        Ask  0.980309   7306.831
#> 106        NA         NA                <NA>        Ask  0.980432   9741.216
#> 107        NA         NA                <NA>        Ask  0.980662   9738.933
#> 108        NA         NA                <NA>        Ask  0.980696    482.076
#> 109        NA         NA                <NA>        Ask  0.980698     87.466
#> 110        NA         NA                <NA>        Ask  0.980852   2435.201
#> 111        NA         NA                <NA>        Ask  0.980896  10000.000
#> 112        NA         NA                <NA>        Ask  0.980909    615.300
#> 113        NA         NA                <NA>        Ask  0.981088  13952.014
#> 114        NA         NA                <NA>        Ask  0.981239   5000.000
#> 115        NA         NA                <NA>        Ask  0.981385   5088.876
#> 116        NA         NA                <NA>        Ask  0.981386  12657.749
#> 117        NA         NA                <NA>        Ask  0.981441   9516.730
#> 118        NA         NA                <NA>        Ask  0.981534   9649.100
#> 119        NA         NA                <NA>        Ask  0.981579   2127.859
#> 120        NA         NA                <NA>        Ask  0.981672   4881.000
#> 121        NA         NA                <NA>        Ask  0.981769  10705.500
#> 122        NA         NA                <NA>        Ask  0.981826   8191.830
#> 123        NA         NA                <NA>        Ask  0.981856   7911.260
#> 124        NA         NA                <NA>        Ask  0.981910     14.769
#> 125        NA         NA                <NA>        Ask  0.981945   2755.703
#> 126        NA         NA                <NA>        Ask  0.981969     24.102
#> 127        NA         NA                <NA>        Ask  0.982022   4839.316
#> 128        NA         NA                <NA>        Ask  0.982023  14587.451
#> 129        NA         NA                <NA>        Ask  0.982098    138.225
#> 130        NA         NA                <NA>        Ask  0.982172   1460.887
#> 131        NA         NA                <NA>        Ask  0.982173  15299.700
#> 132        NA         NA                <NA>        Ask  0.982453  14581.766
#> 133        NA         NA                <NA>        Ask  0.982883     43.852
#> 134        NA         NA                <NA>        Ask  0.983411    471.902
#> 135        NA         NA                <NA>        Ask  0.983412     25.285
#> 136        NA         NA                <NA>        Ask  0.983439  10148.412
#> 137        NA         NA                <NA>        Ask  0.983488  14566.495
#> 138        NA         NA                <NA>        Ask  0.983536     99.610
#> 139        NA         NA                <NA>        Ask  0.983623    105.258
#> 140        NA         NA                <NA>        Ask  0.983624   4059.365
#> 141        NA         NA                <NA>        Ask  0.983672    149.434
#> 142        NA         NA                <NA>        Ask  0.983673  17355.900
#> 143        NA         NA                <NA>        Ask  0.983953  10148.412
#> 144        NA         NA                <NA>        Ask  0.983955  15880.171
#> 145        NA         NA                <NA>        Ask  0.984002  39764.400
#> 146        NA         NA                <NA>        Ask  0.984114     10.229
#> 147        NA         NA                <NA>        Ask  0.984670   6089.047
#> 148        NA         NA                <NA>        Ask  0.984719  15773.184
#> 149        NA         NA                <NA>        Ask  0.984879   6765.608
#> 150        NA         NA                <NA>        Ask  0.984928  18821.704
#> 151        NA         NA                <NA>        Ask  0.985264     48.681
#> 152        NA         NA                <NA>        Ask  0.985648  14534.498
#> 153        NA         NA                <NA>        Ask  0.985802  24826.063
#> 154        NA         NA                <NA>        Ask  0.986014     11.586
#> 155        NA         NA                <NA>        Ask  0.986409   4747.880
#> 156        NA         NA                <NA>        Ask  0.986415    137.620
#> 157        NA         NA                <NA>        Ask  0.986945   8766.941
#> 158        NA         NA                <NA>        Ask  0.987128  37945.700
#> 159        NA         NA                <NA>        Ask  0.987620   2422.941
#> 160        NA         NA                <NA>        Ask  0.987689   4214.044
#> 161        NA         NA                <NA>        Ask  0.987979     23.956
#> 162        NA         NA                <NA>        Ask  0.988113  14498.247
#> 163        NA         NA                <NA>        Ask  0.988396   2230.015
#> 164        NA         NA                <NA>        Ask  0.988410  11685.282
#> 165        NA         NA                <NA>        Ask  0.988453   3667.411
#> 166        NA         NA                <NA>        Ask  0.989581  26581.100
#> 167        NA         NA                <NA>        Ask  0.989593  29223.134
#> 168        NA         NA                <NA>        Ask  0.989681  12732.200
#> 169        NA         NA                <NA>        Ask  0.990452 289595.260
#> 170        NA         NA                <NA>        Ask  0.990453    797.969
#> 171        NA         NA                <NA>        Ask  0.990750    137.018
#> 172        NA         NA                <NA>        Ask  0.990954  14456.678
#> 173        NA         NA                <NA>        Ask  0.991209   3059.879
#> 174        NA         NA                <NA>        Ask  0.991454     11.529
#> 175        NA         NA                <NA>        Ask  0.992245     31.277
#> 176        NA         NA                <NA>        Ask  0.993136     55.456
#> 177        NA         NA                <NA>        Ask  0.993224     78.674
#> 178        NA         NA                <NA>        Ask  0.993634     48.271
#> 179        NA         NA                <NA>        Ask  0.994000     41.491
#> 180        NA         NA                <NA>        Ask  0.994025     23.810
#> 181        NA         NA                <NA>        Ask  0.994142     98.547
#> 182        NA         NA                <NA>        Ask  0.994172   6271.239
#> 183        NA         NA                <NA>        Ask  0.994183   7903.606
#> 184        NA         NA                <NA>        Ask  0.994575   6487.000
#> 185        NA         NA                <NA>        Ask  0.995104    136.419
#> 186        NA         NA                <NA>        Ask  0.996087    793.275
#> 187        NA         NA                <NA>        Ask  0.996211     98.300
#> 188        NA         NA                <NA>        Ask  0.996476   5455.014
#> 189        NA         NA                <NA>        Ask  0.996922     11.466
#> 190        NA         NA                <NA>        Ask  0.998600  14340.571
#> 191        NA         NA                <NA>        Ask  0.998690      7.057
#> 192        NA         NA                <NA>        Ask  0.999478    135.822
#> 193        NA         NA                <NA>        Ask  1.000000   3562.826
#> 194        NA         NA                <NA>        Ask  1.000109     23.665
#> 195        NA         NA                <NA>        Ask  1.000330   6626.100
#> 196        NA         NA                <NA>        Ask  1.001000      8.018
#> 197        NA         NA                <NA>        Ask  1.001581    109.838
#> 198        NA         NA                <NA>        Ask  1.002077     47.865
#> 199        NA         NA                <NA>        Ask  1.002244     84.063
#> 200        NA         NA                <NA>        Ask  1.002364     21.780
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:41:52
#> 102 2024-12-18 16:45:29
#> 103 2024-12-18 16:45:29
#> 104 2024-12-18 16:45:29
#> 105 2024-12-18 16:45:29
#> 106 2024-12-18 16:45:25
#> 107 2024-12-18 16:45:29
#> 108 2024-12-18 16:45:17
#> 109 2024-12-18 16:45:16
#> 110 2024-12-18 16:45:21
#> 111 2024-12-18 16:45:27
#> 112 2024-12-18 16:45:21
#> 113 2024-12-18 16:45:16
#> 114 2024-12-18 16:44:51
#> 115 2024-12-18 16:45:21
#> 116 2024-12-18 16:45:24
#> 117 2024-12-18 16:45:09
#> 118 2024-12-18 16:44:35
#> 119 2024-12-18 16:45:20
#> 120 2024-12-18 16:45:17
#> 121 2024-12-18 16:45:25
#> 122 2024-12-18 16:45:09
#> 123 2024-12-18 16:45:09
#> 124 2024-12-18 16:45:20
#> 125 2024-12-18 16:45:14
#> 126 2024-12-18 16:41:15
#> 127 2024-12-18 16:45:16
#> 128 2024-12-18 16:45:12
#> 129 2024-12-18 14:56:28
#> 130 2024-12-18 16:45:16
#> 131 2024-12-18 16:45:15
#> 132 2024-12-18 16:45:20
#> 133 2024-12-18 16:45:15
#> 134 2024-12-18 16:45:30
#> 135 2024-12-18 16:44:59
#> 136 2024-12-18 16:45:24
#> 137 2024-12-18 16:45:15
#> 138 2024-12-18 15:57:30
#> 139 2024-12-18 16:44:56
#> 140 2024-12-18 16:45:24
#> 141 2024-12-18 16:44:56
#> 142 2024-12-18 16:44:46
#> 143 2024-12-18 16:45:24
#> 144 2024-12-18 16:44:34
#> 145 2024-12-18 16:43:01
#> 146 2024-12-18 16:44:27
#> 147 2024-12-18 16:45:24
#> 148 2024-12-18 16:45:16
#> 149 2024-12-18 16:45:24
#> 150 2024-12-18 16:43:13
#> 151 2024-12-18 14:56:59
#> 152 2024-12-18 16:45:15
#> 153 2024-12-18 16:41:08
#> 154 2024-12-18 15:09:59
#> 155 2024-12-18 16:45:16
#> 156 2024-12-18 15:24:42
#> 157 2024-12-18 16:44:50
#> 158 2024-12-18 16:45:00
#> 159 2024-12-18 16:40:43
#> 160 2024-12-18 16:40:46
#> 161 2024-12-18 14:35:26
#> 162 2024-12-18 16:45:20
#> 163 2024-12-18 16:44:32
#> 164 2024-12-18 16:41:45
#> 165 2024-12-18 16:45:11
#> 166 2024-12-18 16:45:00
#> 167 2024-12-18 16:44:50
#> 168 2024-12-18 16:40:42
#> 169 2024-12-18 16:45:14
#> 170 2024-12-18 16:41:02
#> 171 2024-12-18 13:57:45
#> 172 2024-12-18 16:45:16
#> 173 2024-12-18 16:43:02
#> 174 2024-12-18 12:37:59
#> 175 2024-12-18 15:04:40
#> 176 2024-12-18 13:26:56
#> 177 2024-12-18 15:12:45
#> 178 2024-12-18 12:38:18
#> 179 2024-12-18 12:25:06
#> 180 2024-12-18 13:36:40
#> 181 2024-12-18 12:37:51
#> 182 2024-12-18 16:37:46
#> 183 2024-12-18 16:37:47
#> 184 2024-12-18 16:40:42
#> 185 2024-12-18 13:34:11
#> 186 2024-12-18 16:38:48
#> 187 2024-12-18 15:07:21
#> 188 2024-12-18 16:38:48
#> 189 2024-12-18 12:33:01
#> 190 2024-12-18 16:37:53
#> 191 2024-12-18 12:34:04
#> 192 2024-12-18 13:11:41
#> 193 2024-12-18 14:57:48
#> 194 2024-12-18 12:15:38
#> 195 2024-12-18 16:37:45
#> 196 2024-12-18 12:22:05
#> 197 2024-12-18 15:18:26
#> 198 2024-12-18 04:26:55
#> 199 2024-12-18 07:39:44
#> 200 2024-12-18 12:02:37
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.958245   6118.811 2024-12-18 16:38:47        Bid        NA         NA
#> 2    0.958484   6347.900 2024-12-18 16:38:27        Bid        NA         NA
#> 3    0.958673  14943.809 2024-12-18 16:41:04        Bid        NA         NA
#> 4    0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 5    0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 6    0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 7    0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 8    0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 9    0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 10   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 11   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 12   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 13   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 14   0.961000  27335.794 2024-12-18 16:18:40        Bid        NA         NA
#> 15   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 16   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 17   0.962599   5642.132 2024-12-18 16:40:45        Bid        NA         NA
#> 18   0.962605  12868.612 2024-12-18 16:40:28        Bid        NA         NA
#> 19   0.962607   1776.987 2024-12-18 16:40:29        Bid        NA         NA
#> 20   0.963000     41.491 2024-12-18 16:21:06        Bid        NA         NA
#> 21   0.964461     56.664 2024-12-18 16:24:13        Bid        NA         NA
#> 22   0.964601   3592.389 2024-12-18 16:43:27        Bid        NA         NA
#> 23   0.965262     24.299 2024-12-18 16:38:02        Bid        NA         NA
#> 24   0.966830   6106.993 2024-12-18 16:43:28        Bid        NA         NA
#> 25   0.966928   1636.255 2024-12-18 16:42:01        Bid        NA         NA
#> 26   0.967894  12680.200 2024-12-18 16:45:05        Bid        NA         NA
#> 27   0.968229  14795.412 2024-12-18 16:43:53        Bid        NA         NA
#> 28   0.968700    138.916 2024-12-18 16:38:53        Bid        NA         NA
#> 29   0.968777   3131.503 2024-12-18 16:44:49        Bid        NA         NA
#> 30   0.969111  25256.400 2024-12-18 16:44:23        Bid        NA         NA
#> 31   0.969274  12454.707 2024-12-18 16:43:28        Bid        NA         NA
#> 32   0.970202   8124.487 2024-12-18 16:42:37        Bid        NA         NA
#> 33   0.970203 288615.107 2024-12-18 16:45:16        Bid        NA         NA
#> 34   0.970307   5042.692 2024-12-18 16:45:07        Bid        NA         NA
#> 35   0.970737   3734.342 2024-12-18 16:44:49        Bid        NA         NA
#> 36   0.970928  15115.990 2024-12-18 16:45:07        Bid        NA         NA
#> 37   0.971037   2465.990 2024-12-18 16:44:35        Bid        NA         NA
#> 38   0.971365   2269.505 2024-12-18 16:44:42        Bid        NA         NA
#> 39   0.971563  14744.496 2024-12-18 16:44:31        Bid        NA         NA
#> 40   0.971719  25124.632 2024-12-18 16:43:28        Bid        NA         NA
#> 41   0.972059  38476.100 2024-12-18 16:44:40        Bid        NA         NA
#> 42   0.972766     25.700 2024-12-18 16:35:22        Bid        NA         NA
#> 43   0.973185  19207.356 2024-12-18 16:43:28        Bid        NA         NA
#> 44   0.973233  17708.414 2024-12-18 16:45:21        Bid        NA         NA
#> 45   0.973234   1958.597 2024-12-18 16:45:07        Bid        NA         NA
#> 46   0.973643   3274.714 2024-12-18 16:45:11        Bid        NA         NA
#> 47   0.974058  58581.570 2024-12-18 16:39:56        Bid        NA         NA
#> 48   0.974105   7455.589 2024-12-18 16:45:21        Bid        NA         NA
#> 49   0.974106   6838.201 2024-12-18 16:45:20        Bid        NA         NA
#> 50   0.974267  15189.107 2024-12-18 16:43:01        Bid        NA         NA
#> 51   0.974314   6523.329 2024-12-18 16:45:21        Bid        NA         NA
#> 52   0.974590  15720.934 2024-12-18 16:44:18        Bid        NA         NA
#> 53   0.974638   5591.246 2024-12-18 16:45:21        Bid        NA         NA
#> 54   0.974672  14697.467 2024-12-18 16:44:34        Bid        NA         NA
#> 55   0.974719   4659.093 2024-12-18 16:45:21        Bid        NA         NA
#> 56   0.974798     49.000 2024-12-18 16:41:48        Bid        NA         NA
#> 57   0.974857  38248.600 2024-12-18 16:45:15        Bid        NA         NA
#> 58   0.974865 203105.008 2024-12-18 16:44:49        Bid        NA         NA
#> 59   0.974866    285.944 2024-12-18 16:45:18        Bid        NA         NA
#> 60   0.974912   9317.838 2024-12-18 16:45:21        Bid        NA         NA
#> 61   0.974913     32.924 2024-12-18 16:45:23        Bid        NA         NA
#> 62   0.976060     43.471 2024-12-18 16:45:19        Bid        NA         NA
#> 63   0.976274   2930.000 2024-12-18 16:43:50        Bid        NA         NA
#> 64   0.976621  17358.500 2024-12-18 16:44:42        Bid        NA         NA
#> 65   0.976623   5000.000 2024-12-18 16:45:16        Bid        NA         NA
#> 66   0.976624   2348.517 2024-12-18 16:45:23        Bid        NA         NA
#> 67   0.976955  14663.812 2024-12-18 16:45:16        Bid        NA         NA
#> 68   0.977069     14.454 2024-12-18 16:44:27        Bid        NA         NA
#> 69   0.977448  14656.418 2024-12-18 16:45:29        Bid        NA         NA
#> 70   0.977658   2755.703 2024-12-18 16:45:14        Bid        NA         NA
#> 71   0.977872  14650.132 2024-12-18 16:45:14        Bid        NA         NA
#> 72   0.977913  10451.560 2024-12-18 16:45:09        Bid        NA         NA
#> 73   0.978183   2440.000 2024-12-18 16:45:12        Bid        NA         NA
#> 74   0.978203   4881.000 2024-12-18 16:45:11        Bid        NA         NA
#> 75   0.978298  10719.530 2024-12-18 16:45:09        Bid        NA         NA
#> 76   0.978328  10994.910 2024-12-18 16:45:09        Bid        NA         NA
#> 77   0.978363    609.647 2024-12-18 16:45:12        Bid        NA         NA
#> 78   0.978686   9758.589 2024-12-18 16:45:20        Bid        NA         NA
#> 79   0.978717    617.100 2024-12-18 16:45:11        Bid        NA         NA
#> 80   0.978969     16.697 2024-12-18 16:45:20        Bid        NA         NA
#> 81   0.979008    584.902 2024-12-18 16:45:14        Bid        NA         NA
#> 82   0.979029  20804.254 2024-12-18 16:45:16        Bid        NA         NA
#> 83   0.979030   7566.874 2024-12-18 16:45:25        Bid        NA         NA
#> 84   0.979036   9755.095 2024-12-18 16:45:29        Bid        NA         NA
#> 85   0.979077   9638.900 2024-12-18 16:45:25        Bid        NA         NA
#> 86   0.979427  16025.100 2024-12-18 16:45:16        Bid        NA         NA
#> 87   0.979428   5006.639 2024-12-18 16:45:25        Bid        NA         NA
#> 88   0.979463   9750.850 2024-12-18 16:45:29        Bid        NA         NA
#> 89   0.979487  10000.000 2024-12-18 16:45:28        Bid        NA         NA
#> 90   0.979492   1628.002 2024-12-18 16:45:30        Bid        NA         NA
#> 91   0.979534    467.901 2024-12-18 16:45:15        Bid        NA         NA
#> 92   0.979580   7312.262 2024-12-18 16:45:29        Bid        NA         NA
#> 93   0.979638   3576.965 2024-12-18 16:45:30        Bid        NA         NA
#> 94   0.979643   4874.526 2024-12-18 16:45:29        Bid        NA         NA
#> 95   0.979650   5876.592 2024-12-18 16:45:29        Bid        NA         NA
#> 96   0.979651    280.556 2024-12-18 16:45:29        Bid        NA         NA
#> 97   0.979775   5405.322 2024-12-18 16:45:22        Bid        NA         NA
#> 98   0.979796    615.300 2024-12-18 16:45:28        Bid        NA         NA
#> 99   0.979806   6491.225 2024-12-18 16:45:29        Bid        NA         NA
#> 100  0.979807     33.799 2024-12-18 16:45:29        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.980000     41.491
#> 102        NA         NA                <NA>        Ask  0.980187   1382.919
#> 103        NA         NA                <NA>        Ask  0.980188   2435.912
#> 104        NA         NA                <NA>        Ask  0.980245   4871.537
#> 105        NA         NA                <NA>        Ask  0.980309   7306.831
#> 106        NA         NA                <NA>        Ask  0.980432   9741.216
#> 107        NA         NA                <NA>        Ask  0.980662   9738.933
#> 108        NA         NA                <NA>        Ask  0.980696    482.076
#> 109        NA         NA                <NA>        Ask  0.980698     87.466
#> 110        NA         NA                <NA>        Ask  0.980852   2435.201
#> 111        NA         NA                <NA>        Ask  0.980896  10000.000
#> 112        NA         NA                <NA>        Ask  0.980909    615.300
#> 113        NA         NA                <NA>        Ask  0.981088  13952.014
#> 114        NA         NA                <NA>        Ask  0.981239   5000.000
#> 115        NA         NA                <NA>        Ask  0.981385   5088.876
#> 116        NA         NA                <NA>        Ask  0.981386  12657.749
#> 117        NA         NA                <NA>        Ask  0.981441   9516.730
#> 118        NA         NA                <NA>        Ask  0.981534   9649.100
#> 119        NA         NA                <NA>        Ask  0.981579   2127.859
#> 120        NA         NA                <NA>        Ask  0.981672   4881.000
#> 121        NA         NA                <NA>        Ask  0.981769  10705.500
#> 122        NA         NA                <NA>        Ask  0.981826   8191.830
#> 123        NA         NA                <NA>        Ask  0.981856   7911.260
#> 124        NA         NA                <NA>        Ask  0.981910     14.769
#> 125        NA         NA                <NA>        Ask  0.981945   2755.703
#> 126        NA         NA                <NA>        Ask  0.981969     24.102
#> 127        NA         NA                <NA>        Ask  0.982022   4839.316
#> 128        NA         NA                <NA>        Ask  0.982023  14587.451
#> 129        NA         NA                <NA>        Ask  0.982098    138.225
#> 130        NA         NA                <NA>        Ask  0.982172   1460.887
#> 131        NA         NA                <NA>        Ask  0.982173  15299.700
#> 132        NA         NA                <NA>        Ask  0.982453  14581.766
#> 133        NA         NA                <NA>        Ask  0.982883     43.852
#> 134        NA         NA                <NA>        Ask  0.983411    471.902
#> 135        NA         NA                <NA>        Ask  0.983412     25.285
#> 136        NA         NA                <NA>        Ask  0.983439  10148.412
#> 137        NA         NA                <NA>        Ask  0.983488  14566.495
#> 138        NA         NA                <NA>        Ask  0.983536     99.610
#> 139        NA         NA                <NA>        Ask  0.983623    105.258
#> 140        NA         NA                <NA>        Ask  0.983624   4059.365
#> 141        NA         NA                <NA>        Ask  0.983672    149.434
#> 142        NA         NA                <NA>        Ask  0.983673  17355.900
#> 143        NA         NA                <NA>        Ask  0.983953  10148.412
#> 144        NA         NA                <NA>        Ask  0.983955  15880.171
#> 145        NA         NA                <NA>        Ask  0.984002  39764.400
#> 146        NA         NA                <NA>        Ask  0.984114     10.229
#> 147        NA         NA                <NA>        Ask  0.984670   6089.047
#> 148        NA         NA                <NA>        Ask  0.984719  15773.184
#> 149        NA         NA                <NA>        Ask  0.984879   6765.608
#> 150        NA         NA                <NA>        Ask  0.984928  18821.704
#> 151        NA         NA                <NA>        Ask  0.985264     48.681
#> 152        NA         NA                <NA>        Ask  0.985648  14534.498
#> 153        NA         NA                <NA>        Ask  0.985802  24826.063
#> 154        NA         NA                <NA>        Ask  0.986014     11.586
#> 155        NA         NA                <NA>        Ask  0.986409   4747.880
#> 156        NA         NA                <NA>        Ask  0.986415    137.620
#> 157        NA         NA                <NA>        Ask  0.986945   8766.941
#> 158        NA         NA                <NA>        Ask  0.987128  37945.700
#> 159        NA         NA                <NA>        Ask  0.987620   2422.941
#> 160        NA         NA                <NA>        Ask  0.987689   4214.044
#> 161        NA         NA                <NA>        Ask  0.987979     23.956
#> 162        NA         NA                <NA>        Ask  0.988113  14498.247
#> 163        NA         NA                <NA>        Ask  0.988396   2230.015
#> 164        NA         NA                <NA>        Ask  0.988410  11685.282
#> 165        NA         NA                <NA>        Ask  0.988453   3667.411
#> 166        NA         NA                <NA>        Ask  0.989581  26581.100
#> 167        NA         NA                <NA>        Ask  0.989593  29223.134
#> 168        NA         NA                <NA>        Ask  0.989681  12732.200
#> 169        NA         NA                <NA>        Ask  0.990452 289595.260
#> 170        NA         NA                <NA>        Ask  0.990453    797.969
#> 171        NA         NA                <NA>        Ask  0.990750    137.018
#> 172        NA         NA                <NA>        Ask  0.990954  14456.678
#> 173        NA         NA                <NA>        Ask  0.991209   3059.879
#> 174        NA         NA                <NA>        Ask  0.991454     11.529
#> 175        NA         NA                <NA>        Ask  0.992245     31.277
#> 176        NA         NA                <NA>        Ask  0.993136     55.456
#> 177        NA         NA                <NA>        Ask  0.993224     78.674
#> 178        NA         NA                <NA>        Ask  0.993634     48.271
#> 179        NA         NA                <NA>        Ask  0.994000     41.491
#> 180        NA         NA                <NA>        Ask  0.994025     23.810
#> 181        NA         NA                <NA>        Ask  0.994142     98.547
#> 182        NA         NA                <NA>        Ask  0.994172   6271.239
#> 183        NA         NA                <NA>        Ask  0.994183   7903.606
#> 184        NA         NA                <NA>        Ask  0.994575   6487.000
#> 185        NA         NA                <NA>        Ask  0.995104    136.419
#> 186        NA         NA                <NA>        Ask  0.996087    793.275
#> 187        NA         NA                <NA>        Ask  0.996211     98.300
#> 188        NA         NA                <NA>        Ask  0.996476   5455.014
#> 189        NA         NA                <NA>        Ask  0.996922     11.466
#> 190        NA         NA                <NA>        Ask  0.998600  14340.571
#> 191        NA         NA                <NA>        Ask  0.998690      7.057
#> 192        NA         NA                <NA>        Ask  0.999478    135.822
#> 193        NA         NA                <NA>        Ask  1.000000   3562.826
#> 194        NA         NA                <NA>        Ask  1.000109     23.665
#> 195        NA         NA                <NA>        Ask  1.000330   6626.100
#> 196        NA         NA                <NA>        Ask  1.001000      8.018
#> 197        NA         NA                <NA>        Ask  1.001581    109.838
#> 198        NA         NA                <NA>        Ask  1.002077     47.865
#> 199        NA         NA                <NA>        Ask  1.002244     84.063
#> 200        NA         NA                <NA>        Ask  1.002364     21.780
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:41:52
#> 102 2024-12-18 16:45:29
#> 103 2024-12-18 16:45:29
#> 104 2024-12-18 16:45:29
#> 105 2024-12-18 16:45:29
#> 106 2024-12-18 16:45:25
#> 107 2024-12-18 16:45:29
#> 108 2024-12-18 16:45:17
#> 109 2024-12-18 16:45:16
#> 110 2024-12-18 16:45:21
#> 111 2024-12-18 16:45:27
#> 112 2024-12-18 16:45:21
#> 113 2024-12-18 16:45:16
#> 114 2024-12-18 16:44:51
#> 115 2024-12-18 16:45:21
#> 116 2024-12-18 16:45:24
#> 117 2024-12-18 16:45:09
#> 118 2024-12-18 16:44:35
#> 119 2024-12-18 16:45:20
#> 120 2024-12-18 16:45:17
#> 121 2024-12-18 16:45:25
#> 122 2024-12-18 16:45:09
#> 123 2024-12-18 16:45:09
#> 124 2024-12-18 16:45:20
#> 125 2024-12-18 16:45:14
#> 126 2024-12-18 16:41:15
#> 127 2024-12-18 16:45:16
#> 128 2024-12-18 16:45:12
#> 129 2024-12-18 14:56:28
#> 130 2024-12-18 16:45:16
#> 131 2024-12-18 16:45:15
#> 132 2024-12-18 16:45:20
#> 133 2024-12-18 16:45:15
#> 134 2024-12-18 16:45:30
#> 135 2024-12-18 16:44:59
#> 136 2024-12-18 16:45:24
#> 137 2024-12-18 16:45:15
#> 138 2024-12-18 15:57:30
#> 139 2024-12-18 16:44:56
#> 140 2024-12-18 16:45:24
#> 141 2024-12-18 16:44:56
#> 142 2024-12-18 16:44:46
#> 143 2024-12-18 16:45:24
#> 144 2024-12-18 16:44:34
#> 145 2024-12-18 16:43:01
#> 146 2024-12-18 16:44:27
#> 147 2024-12-18 16:45:24
#> 148 2024-12-18 16:45:16
#> 149 2024-12-18 16:45:24
#> 150 2024-12-18 16:43:13
#> 151 2024-12-18 14:56:59
#> 152 2024-12-18 16:45:15
#> 153 2024-12-18 16:41:08
#> 154 2024-12-18 15:09:59
#> 155 2024-12-18 16:45:16
#> 156 2024-12-18 15:24:42
#> 157 2024-12-18 16:44:50
#> 158 2024-12-18 16:45:00
#> 159 2024-12-18 16:40:43
#> 160 2024-12-18 16:40:46
#> 161 2024-12-18 14:35:26
#> 162 2024-12-18 16:45:20
#> 163 2024-12-18 16:44:32
#> 164 2024-12-18 16:41:45
#> 165 2024-12-18 16:45:11
#> 166 2024-12-18 16:45:00
#> 167 2024-12-18 16:44:50
#> 168 2024-12-18 16:40:42
#> 169 2024-12-18 16:45:14
#> 170 2024-12-18 16:41:02
#> 171 2024-12-18 13:57:45
#> 172 2024-12-18 16:45:16
#> 173 2024-12-18 16:43:02
#> 174 2024-12-18 12:37:59
#> 175 2024-12-18 15:04:40
#> 176 2024-12-18 13:26:56
#> 177 2024-12-18 15:12:45
#> 178 2024-12-18 12:38:18
#> 179 2024-12-18 12:25:06
#> 180 2024-12-18 13:36:40
#> 181 2024-12-18 12:37:51
#> 182 2024-12-18 16:37:46
#> 183 2024-12-18 16:37:47
#> 184 2024-12-18 16:40:42
#> 185 2024-12-18 13:34:11
#> 186 2024-12-18 16:38:48
#> 187 2024-12-18 15:07:21
#> 188 2024-12-18 16:38:48
#> 189 2024-12-18 12:33:01
#> 190 2024-12-18 16:37:53
#> 191 2024-12-18 12:34:04
#> 192 2024-12-18 13:11:41
#> 193 2024-12-18 14:57:48
#> 194 2024-12-18 12:15:38
#> 195 2024-12-18 16:37:45
#> 196 2024-12-18 12:22:05
#> 197 2024-12-18 15:18:26
#> 198 2024-12-18 04:26:55
#> 199 2024-12-18 07:39:44
#> 200 2024-12-18 12:02:37