Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.954571   2579.102 2024-12-18 16:17:07        Bid        NA         NA
#> 2    0.955000     81.491 2024-12-18 07:28:46        Bid        NA         NA
#> 3    0.955006   6118.811 2024-12-18 16:16:41        Bid        NA         NA
#> 4    0.955028   3163.339 2024-12-18 16:17:23        Bid        NA         NA
#> 5    0.955770   3505.028 2024-12-12 05:37:10        Bid        NA         NA
#> 6    0.955871     16.358 2024-12-18 08:50:50        Bid        NA         NA
#> 7    0.955890     31.385 2024-12-18 13:29:07        Bid        NA         NA
#> 8    0.955970     86.060 2024-12-11 11:33:28        Bid        NA         NA
#> 9    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 10   0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 11   0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 12   0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 13   0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 14   0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 15   0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 16   0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 17   0.957638   3767.680 2024-12-18 16:17:04        Bid        NA         NA
#> 18   0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 19   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 20   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 21   0.958116   6106.993 2024-12-18 16:17:02        Bid        NA         NA
#> 22   0.958237   2290.080 2024-12-18 16:17:22        Bid        NA         NA
#> 23   0.958925   2485.559 2024-12-18 16:17:04        Bid        NA         NA
#> 24   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 25   0.959139  12680.200 2024-12-18 16:16:48        Bid        NA         NA
#> 26   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 27   0.959243  14926.516 2024-12-18 16:17:02        Bid        NA         NA
#> 28   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 29   0.959427 288582.585 2024-12-18 16:17:22        Bid        NA         NA
#> 30   0.959758  20371.481 2024-12-18 16:17:04        Bid        NA         NA
#> 31   0.959767  25256.400 2024-12-18 16:17:04        Bid        NA         NA
#> 32   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 33   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 34   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 35   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 36   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 37   0.960160  12454.707 2024-12-18 16:17:03        Bid        NA         NA
#> 38   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 39   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 40   0.961000  27235.794 2024-12-18 15:06:11        Bid        NA         NA
#> 41   0.961013  14899.033 2024-12-18 16:16:38        Bid        NA         NA
#> 42   0.961211   8143.036 2024-12-18 16:17:04        Bid        NA         NA
#> 43   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 44   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 45   0.962201   1841.338 2024-12-18 16:17:04        Bid        NA         NA
#> 46   0.962582  25124.632 2024-12-18 16:17:03        Bid        NA         NA
#> 47   0.962630  20953.534 2024-12-18 16:17:27        Bid        NA         NA
#> 48   0.962631   6919.716 2024-12-18 16:17:32        Bid        NA         NA
#> 49   0.962934   3323.408 2024-12-18 16:17:18        Bid        NA         NA
#> 50   0.963000     41.491 2024-12-18 15:23:06        Bid        NA         NA
#> 51   0.963515  38476.100 2024-12-18 16:17:00        Bid        NA         NA
#> 52   0.963563   8822.448 2024-12-18 16:17:27        Bid        NA         NA
#> 53   0.963602 205478.860 2024-12-18 16:17:20        Bid        NA         NA
#> 54   0.963651   7719.418 2024-12-18 16:17:27        Bid        NA         NA
#> 55   0.963734  14856.967 2024-12-18 16:17:03        Bid        NA         NA
#> 56   0.964124  38248.600 2024-12-18 16:17:04        Bid        NA         NA
#> 57   0.964172  12129.558 2024-12-18 16:17:27        Bid        NA         NA
#> 58   0.964415  19207.356 2024-12-18 16:17:02        Bid        NA         NA
#> 59   0.964461    139.527 2024-12-18 15:20:06        Bid        NA         NA
#> 60   0.964463   9923.101 2024-12-18 16:17:27        Bid        NA         NA
#> 61   0.965000     41.491 2024-12-18 16:09:06        Bid        NA         NA
#> 62   0.965003  15720.934 2024-12-18 16:17:03        Bid        NA         NA
#> 63   0.965035    133.594 2024-12-18 16:17:04        Bid        NA         NA
#> 64   0.965051   1102.521 2024-12-18 16:17:27        Bid        NA         NA
#> 65   0.965107  15189.107 2024-12-18 16:17:04        Bid        NA         NA
#> 66   0.965262     24.299 2024-12-18 15:48:28        Bid        NA         NA
#> 67   0.965263    163.206 2024-12-18 16:17:04        Bid        NA         NA
#> 68   0.965342  14831.442 2024-12-18 16:17:31        Bid        NA         NA
#> 69   0.965460  47123.409 2024-12-18 16:17:03        Bid        NA         NA
#> 70   0.965461     26.331 2024-12-18 16:17:22        Bid        NA         NA
#> 71   0.965811  17515.700 2024-12-18 16:17:02        Bid        NA         NA
#> 72   0.965878   2374.646 2024-12-18 16:17:31        Bid        NA         NA
#> 73   0.965991     43.831 2024-12-18 16:17:08        Bid        NA         NA
#> 74   0.966220  14817.966 2024-12-18 16:17:32        Bid        NA         NA
#> 75   0.966270  10493.630 2024-12-18 16:17:23        Bid        NA         NA
#> 76   0.966297     28.683 2024-12-18 15:14:20        Bid        NA         NA
#> 77   0.966298     14.454 2024-12-18 16:17:03        Bid        NA         NA
#> 78   0.966499   7920.221 2024-12-18 16:17:33        Bid        NA         NA
#> 79   0.966615  14811.908 2024-12-18 16:17:30        Bid        NA         NA
#> 80   0.966740  16025.100 2024-12-18 16:17:33        Bid        NA         NA
#> 81   0.966784   5405.322 2024-12-18 16:17:33        Bid        NA         NA
#> 82   0.966820  10985.890 2024-12-18 16:17:32        Bid        NA         NA
#> 83   0.966868   5000.000 2024-12-18 16:17:30        Bid        NA         NA
#> 84   0.966999   4110.207 2024-12-18 16:17:33        Bid        NA         NA
#> 85   0.967009   5006.639 2024-12-18 16:17:31        Bid        NA         NA
#> 86   0.967028   2755.703 2024-12-18 16:17:30        Bid        NA         NA
#> 87   0.967049  10683.460 2024-12-18 16:17:23        Bid        NA         NA
#> 88   0.967052    560.412 2024-12-18 16:17:22        Bid        NA         NA
#> 89   0.967053   5437.685 2024-12-18 16:17:34        Bid        NA         NA
#> 90   0.967266   9867.961 2024-12-18 16:17:32        Bid        NA         NA
#> 91   0.967267     58.558 2024-12-18 16:17:33        Bid        NA         NA
#> 92   0.967370     16.697 2024-12-18 16:17:35        Bid        NA         NA
#> 93   0.967752   9730.900 2024-12-18 16:17:02        Bid        NA         NA
#> 94   0.967766   9862.858 2024-12-18 16:17:34        Bid        NA         NA
#> 95   0.968082    362.145 2024-12-18 16:17:31        Bid        NA         NA
#> 96   0.968161   9858.270 2024-12-18 16:17:34        Bid        NA         NA
#> 97   0.968209   7393.760 2024-12-18 16:17:34        Bid        NA         NA
#> 98   0.968299    698.000 2024-12-18 16:17:33        Bid        NA         NA
#> 99   0.968301   2464.354 2024-12-18 16:17:34        Bid        NA         NA
#> 100  0.968302     47.998 2024-12-18 16:17:34        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.969272   1363.473
#> 102        NA         NA                <NA>        Ask  0.969273     93.697
#> 103        NA         NA                <NA>        Ask  0.969275   6142.416
#> 104        NA         NA                <NA>        Ask  0.969284     33.799
#> 105        NA         NA                <NA>        Ask  0.969288   1000.000
#> 106        NA         NA                <NA>        Ask  0.969290  24618.393
#> 107        NA         NA                <NA>        Ask  0.969294   1000.000
#> 108        NA         NA                <NA>        Ask  0.969297   1628.002
#> 109        NA         NA                <NA>        Ask  0.969299   1936.000
#> 110        NA         NA                <NA>        Ask  0.969305   1363.472
#> 111        NA         NA                <NA>        Ask  0.969361     49.000
#> 112        NA         NA                <NA>        Ask  0.969364  10000.000
#> 113        NA         NA                <NA>        Ask  0.969424    344.213
#> 114        NA         NA                <NA>        Ask  0.969471   9845.527
#> 115        NA         NA                <NA>        Ask  0.969521    698.000
#> 116        NA         NA                <NA>        Ask  0.969592   2936.000
#> 117        NA         NA                <NA>        Ask  0.969679   5876.592
#> 118        NA         NA                <NA>        Ask  0.969752   2451.000
#> 119        NA         NA                <NA>        Ask  0.969780   3576.965
#> 120        NA         NA                <NA>        Ask  0.969909   9841.074
#> 121        NA         NA                <NA>        Ask  0.969923   2100.000
#> 122        NA         NA                <NA>        Ask  0.970250   3876.964
#> 123        NA         NA                <NA>        Ask  0.970278     18.836
#> 124        NA         NA                <NA>        Ask  0.970535  14752.098
#> 125        NA         NA                <NA>        Ask  0.970634  14151.540
#> 126        NA         NA                <NA>        Ask  0.970932  14746.065
#> 127        NA         NA                <NA>        Ask  0.971046    486.829
#> 128        NA         NA                <NA>        Ask  0.971047  10376.311
#> 129        NA         NA                <NA>        Ask  0.971067  12814.949
#> 130        NA         NA                <NA>        Ask  0.971096   9744.900
#> 131        NA         NA                <NA>        Ask  0.971134  11740.070
#> 132        NA         NA                <NA>        Ask  0.971206   2561.420
#> 133        NA         NA                <NA>        Ask  0.971283   4839.316
#> 134        NA         NA                <NA>        Ask  0.971572   1455.246
#> 135        NA         NA                <NA>        Ask  0.971610  10245.750
#> 136        NA         NA                <NA>        Ask  0.971852  14732.111
#> 137        NA         NA                <NA>        Ask  0.972803    673.609
#> 138        NA         NA                <NA>        Ask  0.972851   5000.000
#> 139        NA         NA                <NA>        Ask  0.972852    815.114
#> 140        NA         NA                <NA>        Ask  0.972906   1134.499
#> 141        NA         NA                <NA>        Ask  0.973278     13.638
#> 142        NA         NA                <NA>        Ask  0.973499   1453.576
#> 143        NA         NA                <NA>        Ask  0.974305   1772.654
#> 144        NA         NA                <NA>        Ask  0.974353  17487.600
#> 145        NA         NA                <NA>        Ask  0.974407  14694.253
#> 146        NA         NA                <NA>        Ask  0.974993     44.235
#> 147        NA         NA                <NA>        Ask  0.976407    156.767
#> 148        NA         NA                <NA>        Ask  0.976408   4759.957
#> 149        NA         NA                <NA>        Ask  0.976544  14661.325
#> 150        NA         NA                <NA>        Ask  0.976561     23.486
#> 151        NA         NA                <NA>        Ask  0.977476    148.610
#> 152        NA         NA                <NA>        Ask  0.977477  11675.773
#> 153        NA         NA                <NA>        Ask  0.977800    138.833
#> 154        NA         NA                <NA>        Ask  0.977880  15880.171
#> 155        NA         NA                <NA>        Ask  0.977993  39764.400
#> 156        NA         NA                <NA>        Ask  0.978012  15773.184
#> 157        NA         NA                <NA>        Ask  0.979391   2433.618
#> 158        NA         NA                <NA>        Ask  0.979489  18821.704
#> 159        NA         NA                <NA>        Ask  0.979733  21937.444
#> 160        NA         NA                <NA>        Ask  0.979745   7309.472
#> 161        NA         NA                <NA>        Ask  0.979909 287945.163
#> 162        NA         NA                <NA>        Ask  0.979910   3682.048
#> 163        NA         NA                <NA>        Ask  0.979924  14609.986
#> 164        NA         NA                <NA>        Ask  0.979959   2239.319
#> 165        NA         NA                <NA>        Ask  0.980951  24826.063
#> 166        NA         NA                <NA>        Ask  0.981328  37945.700
#> 167        NA         NA                <NA>        Ask  0.981799   3077.083
#> 168        NA         NA                <NA>        Ask  0.981969     24.102
#> 169        NA         NA                <NA>        Ask  0.982098    138.225
#> 170        NA         NA                <NA>        Ask  0.982499   9053.359
#> 171        NA         NA                <NA>        Ask  0.983067   8827.353
#> 172        NA         NA                <NA>        Ask  0.983536     99.610
#> 173        NA         NA                <NA>        Ask  0.983729  26581.100
#> 174        NA         NA                <NA>        Ask  0.985264     48.681
#> 175        NA         NA                <NA>        Ask  0.986014     11.586
#> 176        NA         NA                <NA>        Ask  0.986164  12732.200
#> 177        NA         NA                <NA>        Ask  0.986415    137.620
#> 178        NA         NA                <NA>        Ask  0.987665   3002.598
#> 179        NA         NA                <NA>        Ask  0.987979     23.956
#> 180        NA         NA                <NA>        Ask  0.988552   1716.359
#> 181        NA         NA                <NA>        Ask  0.988893   7927.653
#> 182        NA         NA                <NA>        Ask  0.989589    792.648
#> 183        NA         NA                <NA>        Ask  0.989625   5463.282
#> 184        NA         NA                <NA>        Ask  0.990604  14453.982
#> 185        NA         NA                <NA>        Ask  0.990750    137.018
#> 186        NA         NA                <NA>        Ask  0.991000   6487.000
#> 187        NA         NA                <NA>        Ask  0.991454     11.529
#> 188        NA         NA                <NA>        Ask  0.991484    789.393
#> 189        NA         NA                <NA>        Ask  0.992245     31.277
#> 190        NA         NA                <NA>        Ask  0.992917   1559.224
#> 191        NA         NA                <NA>        Ask  0.993136     55.456
#> 192        NA         NA                <NA>        Ask  0.993224     78.674
#> 193        NA         NA                <NA>        Ask  0.993634     48.271
#> 194        NA         NA                <NA>        Ask  0.994000     41.491
#> 195        NA         NA                <NA>        Ask  0.994025     23.810
#> 196        NA         NA                <NA>        Ask  0.994142     98.547
#> 197        NA         NA                <NA>        Ask  0.995104    136.419
#> 198        NA         NA                <NA>        Ask  0.996211     98.300
#> 199        NA         NA                <NA>        Ask  0.996922     11.466
#> 200        NA         NA                <NA>        Ask  0.998273   6626.100
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:17:34
#> 102 2024-12-18 16:17:34
#> 103 2024-12-18 16:17:34
#> 104 2024-12-18 16:17:34
#> 105 2024-12-18 16:17:34
#> 106 2024-12-18 16:17:34
#> 107 2024-12-18 16:17:33
#> 108 2024-12-18 16:17:33
#> 109 2024-12-18 16:17:33
#> 110 2024-12-18 16:17:34
#> 111 2024-12-18 16:17:32
#> 112 2024-12-18 16:17:32
#> 113 2024-12-18 16:17:32
#> 114 2024-12-18 16:17:34
#> 115 2024-12-18 16:17:31
#> 116 2024-12-18 16:17:34
#> 117 2024-12-18 16:17:32
#> 118 2024-12-18 16:17:34
#> 119 2024-12-18 16:17:31
#> 120 2024-12-18 16:17:33
#> 121 2024-12-18 16:17:29
#> 122 2024-12-18 16:17:26
#> 123 2024-12-18 16:17:33
#> 124 2024-12-18 16:17:32
#> 125 2024-12-18 16:17:02
#> 126 2024-12-18 16:17:30
#> 127 2024-12-18 16:17:35
#> 128 2024-12-18 16:17:27
#> 129 2024-12-18 16:17:27
#> 130 2024-12-18 16:17:26
#> 131 2024-12-18 16:17:32
#> 132 2024-12-18 16:17:30
#> 133 2024-12-18 16:17:33
#> 134 2024-12-18 16:17:28
#> 135 2024-12-18 16:17:32
#> 136 2024-12-18 16:17:33
#> 137 2024-12-18 16:17:29
#> 138 2024-12-18 16:17:28
#> 139 2024-12-18 16:17:17
#> 140 2024-12-18 16:17:29
#> 141 2024-12-18 16:17:03
#> 142 2024-12-18 16:17:29
#> 143 2024-12-18 16:17:29
#> 144 2024-12-18 16:16:42
#> 145 2024-12-18 16:17:03
#> 146 2024-12-18 16:17:04
#> 147 2024-12-18 16:17:20
#> 148 2024-12-18 16:16:53
#> 149 2024-12-18 16:17:32
#> 150 2024-12-18 16:17:25
#> 151 2024-12-18 16:17:25
#> 152 2024-12-18 16:17:02
#> 153 2024-12-18 16:17:28
#> 154 2024-12-18 16:17:09
#> 155 2024-12-18 16:17:27
#> 156 2024-12-18 16:17:22
#> 157 2024-12-18 16:17:03
#> 158 2024-12-18 16:17:03
#> 159 2024-12-18 16:16:57
#> 160 2024-12-18 16:16:59
#> 161 2024-12-18 16:17:26
#> 162 2024-12-18 16:17:04
#> 163 2024-12-18 16:17:26
#> 164 2024-12-18 16:17:26
#> 165 2024-12-18 16:17:03
#> 166 2024-12-18 16:17:02
#> 167 2024-12-18 16:17:34
#> 168 2024-12-18 16:08:58
#> 169 2024-12-18 14:56:28
#> 170 2024-12-18 16:17:33
#> 171 2024-12-18 16:16:29
#> 172 2024-12-18 15:57:30
#> 173 2024-12-18 16:17:03
#> 174 2024-12-18 14:56:59
#> 175 2024-12-18 15:09:59
#> 176 2024-12-18 16:17:03
#> 177 2024-12-18 15:24:42
#> 178 2024-12-18 16:16:51
#> 179 2024-12-18 14:35:26
#> 180 2024-12-18 16:16:40
#> 181 2024-12-18 16:16:28
#> 182 2024-12-18 16:16:25
#> 183 2024-12-18 16:16:22
#> 184 2024-12-18 16:16:56
#> 185 2024-12-18 13:57:45
#> 186 2024-12-18 16:17:03
#> 187 2024-12-18 12:37:59
#> 188 2024-12-18 16:17:20
#> 189 2024-12-18 15:04:40
#> 190 2024-12-18 16:17:29
#> 191 2024-12-18 13:26:56
#> 192 2024-12-18 15:12:45
#> 193 2024-12-18 12:38:18
#> 194 2024-12-18 12:25:06
#> 195 2024-12-18 13:36:40
#> 196 2024-12-18 12:37:51
#> 197 2024-12-18 13:34:11
#> 198 2024-12-18 15:07:21
#> 199 2024-12-18 12:33:01
#> 200 2024-12-18 16:05:38
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.955028   3163.339 2024-12-18 16:17:23        Bid        NA         NA
#> 2    0.955770   3505.028 2024-12-12 05:37:10        Bid        NA         NA
#> 3    0.955871     16.358 2024-12-18 08:50:50        Bid        NA         NA
#> 4    0.955890     31.385 2024-12-18 13:29:07        Bid        NA         NA
#> 5    0.955970     86.060 2024-12-11 11:33:28        Bid        NA         NA
#> 6    0.956000  10600.000 2024-12-18 15:54:41        Bid        NA         NA
#> 7    0.956039    140.756 2024-12-11 11:44:40        Bid        NA         NA
#> 8    0.956056     10.296 2024-12-18 13:47:21        Bid        NA         NA
#> 9    0.956106    100.000 2024-12-11 12:28:44        Bid        NA         NA
#> 10   0.956721     20.905 2024-12-18 09:36:18        Bid        NA         NA
#> 11   0.957000    145.985 2024-12-18 15:30:00        Bid        NA         NA
#> 12   0.957143     31.277 2024-12-18 10:10:17        Bid        NA         NA
#> 13   0.957244    104.467 2024-12-18 15:30:00        Bid        NA         NA
#> 14   0.957638   3767.680 2024-12-18 16:17:04        Bid        NA         NA
#> 15   0.957757   4081.633 2024-12-18 14:18:06        Bid        NA         NA
#> 16   0.957890     18.254 2024-12-15 19:22:06        Bid        NA         NA
#> 17   0.958000    501.490 2024-12-18 04:20:05        Bid        NA         NA
#> 18   0.958116   6106.993 2024-12-18 16:17:02        Bid        NA         NA
#> 19   0.958237   2290.080 2024-12-18 16:17:22        Bid        NA         NA
#> 20   0.958925   2485.559 2024-12-18 16:17:04        Bid        NA         NA
#> 21   0.959000    207.373 2024-12-18 06:46:21        Bid        NA         NA
#> 22   0.959139  12680.200 2024-12-18 16:16:48        Bid        NA         NA
#> 23   0.959183     15.639 2024-12-16 15:26:15        Bid        NA         NA
#> 24   0.959243  14926.516 2024-12-18 16:17:02        Bid        NA         NA
#> 25   0.959391     24.447 2024-12-15 04:05:27        Bid        NA         NA
#> 26   0.959427 288582.585 2024-12-18 16:17:22        Bid        NA         NA
#> 27   0.959758  20371.481 2024-12-18 16:17:04        Bid        NA         NA
#> 28   0.959767  25256.400 2024-12-18 16:17:04        Bid        NA         NA
#> 29   0.959798     11.802 2024-12-18 15:06:16        Bid        NA         NA
#> 30   0.959865     33.229 2024-12-12 18:00:40        Bid        NA         NA
#> 31   0.960000  60446.364 2024-12-18 16:09:24        Bid        NA         NA
#> 32   0.960030    147.475 2024-12-18 15:05:08        Bid        NA         NA
#> 33   0.960057    353.395 2024-12-17 18:29:35        Bid        NA         NA
#> 34   0.960160  12454.707 2024-12-18 16:17:03        Bid        NA         NA
#> 35   0.960241    140.140 2024-12-11 11:43:22        Bid        NA         NA
#> 36   0.960746   8186.272 2024-12-18 16:17:38        Bid        NA         NA
#> 37   0.960973     69.766 2024-12-18 15:12:46        Bid        NA         NA
#> 38   0.961000  27235.794 2024-12-18 15:06:11        Bid        NA         NA
#> 39   0.961013  14899.033 2024-12-18 16:16:38        Bid        NA         NA
#> 40   0.961061  11019.616 2024-12-18 16:17:39        Bid        NA         NA
#> 41   0.962000     41.490 2024-12-18 15:06:41        Bid        NA         NA
#> 42   0.962200   4890.459 2024-12-18 15:08:28        Bid        NA         NA
#> 43   0.962201   1841.338 2024-12-18 16:17:04        Bid        NA         NA
#> 44   0.962248   8815.544 2024-12-18 16:17:39        Bid        NA         NA
#> 45   0.962582  25124.632 2024-12-18 16:17:03        Bid        NA         NA
#> 46   0.962630  17631.088 2024-12-18 16:17:39        Bid        NA         NA
#> 47   0.962684   3322.693 2024-12-18 16:17:38        Bid        NA         NA
#> 48   0.962732  12121.373 2024-12-18 16:17:39        Bid        NA         NA
#> 49   0.963000     41.491 2024-12-18 15:23:06        Bid        NA         NA
#> 50   0.963515  38476.100 2024-12-18 16:17:00        Bid        NA         NA
#> 51   0.963563   9916.929 2024-12-18 16:17:39        Bid        NA         NA
#> 52   0.963602 205478.860 2024-12-18 16:17:20        Bid        NA         NA
#> 53   0.963734  14856.967 2024-12-18 16:17:03        Bid        NA         NA
#> 54   0.963782   1101.696 2024-12-18 16:17:39        Bid        NA         NA
#> 55   0.964124  38248.600 2024-12-18 16:17:04        Bid        NA         NA
#> 56   0.964415  19207.356 2024-12-18 16:17:02        Bid        NA         NA
#> 57   0.964461    139.527 2024-12-18 15:20:06        Bid        NA         NA
#> 58   0.964798  14840.651 2024-12-18 16:17:37        Bid        NA         NA
#> 59   0.965000     41.491 2024-12-18 16:09:06        Bid        NA         NA
#> 60   0.965003  15720.934 2024-12-18 16:17:03        Bid        NA         NA
#> 61   0.965035    133.594 2024-12-18 16:17:04        Bid        NA         NA
#> 62   0.965107  15189.107 2024-12-18 16:17:04        Bid        NA         NA
#> 63   0.965262     24.299 2024-12-18 15:48:28        Bid        NA         NA
#> 64   0.965263    163.206 2024-12-18 16:17:04        Bid        NA         NA
#> 65   0.965460  47123.409 2024-12-18 16:17:03        Bid        NA         NA
#> 66   0.965461   2402.002 2024-12-18 16:17:38        Bid        NA         NA
#> 67   0.965811  17515.700 2024-12-18 16:17:02        Bid        NA         NA
#> 68   0.965899  14822.953 2024-12-18 16:17:36        Bid        NA         NA
#> 69   0.965928   7920.221 2024-12-18 16:17:39        Bid        NA         NA
#> 70   0.965991     43.831 2024-12-18 16:17:08        Bid        NA         NA
#> 71   0.966119  10583.030 2024-12-18 16:17:41        Bid        NA         NA
#> 72   0.966133  10619.770 2024-12-18 16:17:36        Bid        NA         NA
#> 73   0.966163  16025.100 2024-12-18 16:17:39        Bid        NA         NA
#> 74   0.966200   5006.639 2024-12-18 16:17:39        Bid        NA         NA
#> 75   0.966208   5405.322 2024-12-18 16:17:39        Bid        NA         NA
#> 76   0.966297     28.683 2024-12-18 15:14:20        Bid        NA         NA
#> 77   0.966298     14.454 2024-12-18 16:17:03        Bid        NA         NA
#> 78   0.966437   2755.703 2024-12-18 16:17:36        Bid        NA         NA
#> 79   0.966564  14814.308 2024-12-18 16:17:40        Bid        NA         NA
#> 80   0.966808   4110.207 2024-12-18 16:17:41        Bid        NA         NA
#> 81   0.966820  10985.890 2024-12-18 16:17:32        Bid        NA         NA
#> 82   0.966868   5000.000 2024-12-18 16:17:30        Bid        NA         NA
#> 83   0.966869   5448.441 2024-12-18 16:17:41        Bid        NA         NA
#> 84   0.967014   9869.968 2024-12-18 16:17:36        Bid        NA         NA
#> 85   0.967015    560.227 2024-12-18 16:17:40        Bid        NA         NA
#> 86   0.967046     16.697 2024-12-18 16:17:36        Bid        NA         NA
#> 87   0.967229   2451.000 2024-12-18 16:17:41        Bid        NA         NA
#> 88   0.967316   4891.000 2024-12-18 16:17:41        Bid        NA         NA
#> 89   0.967628   1628.002 2024-12-18 16:17:39        Bid        NA         NA
#> 90   0.967752   9730.900 2024-12-18 16:17:02        Bid        NA         NA
#> 91   0.967753    368.199 2024-12-18 16:17:40        Bid        NA         NA
#> 92   0.967853   9863.053 2024-12-18 16:17:41        Bid        NA         NA
#> 93   0.967888    698.200 2024-12-18 16:17:37        Bid        NA         NA
#> 94   0.967977    423.000 2024-12-18 16:17:38        Bid        NA         NA
#> 95   0.967982   5876.592 2024-12-18 16:17:39        Bid        NA         NA
#> 96   0.968053   2465.256 2024-12-18 16:17:41        Bid        NA         NA
#> 97   0.968059   2100.000 2024-12-18 16:17:41        Bid        NA         NA
#> 98   0.968060   5876.592 2024-12-18 16:17:41        Bid        NA         NA
#> 99   0.968408   3576.965 2024-12-18 16:17:41        Bid        NA         NA
#> 100  0.968409     45.816 2024-12-18 16:17:41        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.968915   1363.829
#> 102        NA         NA                <NA>        Ask  0.968916   6142.416
#> 103        NA         NA                <NA>        Ask  0.968918  14778.329
#> 104        NA         NA                <NA>        Ask  0.969031     49.000
#> 105        NA         NA                <NA>        Ask  0.969089   1936.000
#> 106        NA         NA                <NA>        Ask  0.969118    698.300
#> 107        NA         NA                <NA>        Ask  0.969230   9849.052
#> 108        NA         NA                <NA>        Ask  0.969273     93.697
#> 109        NA         NA                <NA>        Ask  0.969305   1628.002
#> 110        NA         NA                <NA>        Ask  0.969352   2464.354
#> 111        NA         NA                <NA>        Ask  0.969424    344.213
#> 112        NA         NA                <NA>        Ask  0.969645   2936.000
#> 113        NA         NA                <NA>        Ask  0.969740  14066.650
#> 114        NA         NA                <NA>        Ask  0.969853   9745.400
#> 115        NA         NA                <NA>        Ask  0.969865   4891.000
#> 116        NA         NA                <NA>        Ask  0.969877   9842.486
#> 117        NA         NA                <NA>        Ask  0.969986     18.836
#> 118        NA         NA                <NA>        Ask  0.970250   3876.964
#> 119        NA         NA                <NA>        Ask  0.970329  11895.040
#> 120        NA         NA                <NA>        Ask  0.970518  10376.311
#> 121        NA         NA                <NA>        Ask  0.970532    460.309
#> 122        NA         NA                <NA>        Ask  0.970535  14752.098
#> 123        NA         NA                <NA>        Ask  0.970546   2755.703
#> 124        NA         NA                <NA>        Ask  0.970700  10108.380
#> 125        NA         NA                <NA>        Ask  0.970904  14748.103
#> 126        NA         NA                <NA>        Ask  0.971067  12814.949
#> 127        NA         NA                <NA>        Ask  0.971283   4839.316
#> 128        NA         NA                <NA>        Ask  0.971572   1455.246
#> 129        NA         NA                <NA>        Ask  0.971851   5000.000
#> 130        NA         NA                <NA>        Ask  0.971852  14732.111
#> 131        NA         NA                <NA>        Ask  0.972803    673.609
#> 132        NA         NA                <NA>        Ask  0.972852    815.114
#> 133        NA         NA                <NA>        Ask  0.972906   1134.499
#> 134        NA         NA                <NA>        Ask  0.973278     13.638
#> 135        NA         NA                <NA>        Ask  0.973499   1453.576
#> 136        NA         NA                <NA>        Ask  0.974305   1772.654
#> 137        NA         NA                <NA>        Ask  0.974353  17487.600
#> 138        NA         NA                <NA>        Ask  0.974407  14694.253
#> 139        NA         NA                <NA>        Ask  0.974993     44.235
#> 140        NA         NA                <NA>        Ask  0.976407    156.767
#> 141        NA         NA                <NA>        Ask  0.976408   4759.957
#> 142        NA         NA                <NA>        Ask  0.976544  14661.325
#> 143        NA         NA                <NA>        Ask  0.976561     23.486
#> 144        NA         NA                <NA>        Ask  0.977476    148.610
#> 145        NA         NA                <NA>        Ask  0.977477  11675.773
#> 146        NA         NA                <NA>        Ask  0.977800    138.833
#> 147        NA         NA                <NA>        Ask  0.977880  15880.171
#> 148        NA         NA                <NA>        Ask  0.977993  39764.400
#> 149        NA         NA                <NA>        Ask  0.978012  15773.184
#> 150        NA         NA                <NA>        Ask  0.979391   2433.618
#> 151        NA         NA                <NA>        Ask  0.979489  18821.704
#> 152        NA         NA                <NA>        Ask  0.979733  21937.444
#> 153        NA         NA                <NA>        Ask  0.979745   7309.472
#> 154        NA         NA                <NA>        Ask  0.979909 287945.163
#> 155        NA         NA                <NA>        Ask  0.979910   3682.048
#> 156        NA         NA                <NA>        Ask  0.979924  14609.986
#> 157        NA         NA                <NA>        Ask  0.979959   2239.319
#> 158        NA         NA                <NA>        Ask  0.980951  24826.063
#> 159        NA         NA                <NA>        Ask  0.981328  37945.700
#> 160        NA         NA                <NA>        Ask  0.981799   3077.083
#> 161        NA         NA                <NA>        Ask  0.981969     24.102
#> 162        NA         NA                <NA>        Ask  0.982098    138.225
#> 163        NA         NA                <NA>        Ask  0.982499   9053.359
#> 164        NA         NA                <NA>        Ask  0.983067   8827.353
#> 165        NA         NA                <NA>        Ask  0.983536     99.610
#> 166        NA         NA                <NA>        Ask  0.983729  26581.100
#> 167        NA         NA                <NA>        Ask  0.985264     48.681
#> 168        NA         NA                <NA>        Ask  0.986014     11.586
#> 169        NA         NA                <NA>        Ask  0.986164  12732.200
#> 170        NA         NA                <NA>        Ask  0.986415    137.620
#> 171        NA         NA                <NA>        Ask  0.987665   3002.598
#> 172        NA         NA                <NA>        Ask  0.987979     23.956
#> 173        NA         NA                <NA>        Ask  0.988552   1716.359
#> 174        NA         NA                <NA>        Ask  0.988893   7927.653
#> 175        NA         NA                <NA>        Ask  0.989589    792.648
#> 176        NA         NA                <NA>        Ask  0.989625   5463.282
#> 177        NA         NA                <NA>        Ask  0.990604  14453.982
#> 178        NA         NA                <NA>        Ask  0.990750    137.018
#> 179        NA         NA                <NA>        Ask  0.991000   6487.000
#> 180        NA         NA                <NA>        Ask  0.991454     11.529
#> 181        NA         NA                <NA>        Ask  0.991484    789.393
#> 182        NA         NA                <NA>        Ask  0.992245     31.277
#> 183        NA         NA                <NA>        Ask  0.992917   1559.224
#> 184        NA         NA                <NA>        Ask  0.993136     55.456
#> 185        NA         NA                <NA>        Ask  0.993224     78.674
#> 186        NA         NA                <NA>        Ask  0.993634     48.271
#> 187        NA         NA                <NA>        Ask  0.994000     41.491
#> 188        NA         NA                <NA>        Ask  0.994025     23.810
#> 189        NA         NA                <NA>        Ask  0.994142     98.547
#> 190        NA         NA                <NA>        Ask  0.995104    136.419
#> 191        NA         NA                <NA>        Ask  0.996211     98.300
#> 192        NA         NA                <NA>        Ask  0.996922     11.466
#> 193        NA         NA                <NA>        Ask  0.998273   6626.100
#> 194        NA         NA                <NA>        Ask  0.998445   1171.936
#> 195        NA         NA                <NA>        Ask  0.998447   3515.804
#> 196        NA         NA                <NA>        Ask  0.998690      7.057
#> 197        NA         NA                <NA>        Ask  0.998994  48093.478
#> 198        NA         NA                <NA>        Ask  0.999478    135.822
#> 199        NA         NA                <NA>        Ask  1.000000   3562.826
#> 200        NA         NA                <NA>        Ask  1.000109     23.665
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-12-18 16:17:41
#> 102 2024-12-18 16:17:41
#> 103 2024-12-18 16:17:41
#> 104 2024-12-18 16:17:38
#> 105 2024-12-18 16:17:36
#> 106 2024-12-18 16:17:40
#> 107 2024-12-18 16:17:41
#> 108 2024-12-18 16:17:34
#> 109 2024-12-18 16:17:41
#> 110 2024-12-18 16:17:35
#> 111 2024-12-18 16:17:32
#> 112 2024-12-18 16:17:41
#> 113 2024-12-18 16:17:39
#> 114 2024-12-18 16:17:36
#> 115 2024-12-18 16:17:41
#> 116 2024-12-18 16:17:40
#> 117 2024-12-18 16:17:37
#> 118 2024-12-18 16:17:26
#> 119 2024-12-18 16:17:39
#> 120 2024-12-18 16:17:41
#> 121 2024-12-18 16:17:40
#> 122 2024-12-18 16:17:32
#> 123 2024-12-18 16:17:36
#> 124 2024-12-18 16:17:39
#> 125 2024-12-18 16:17:40
#> 126 2024-12-18 16:17:27
#> 127 2024-12-18 16:17:33
#> 128 2024-12-18 16:17:28
#> 129 2024-12-18 16:17:36
#> 130 2024-12-18 16:17:33
#> 131 2024-12-18 16:17:29
#> 132 2024-12-18 16:17:17
#> 133 2024-12-18 16:17:29
#> 134 2024-12-18 16:17:03
#> 135 2024-12-18 16:17:29
#> 136 2024-12-18 16:17:29
#> 137 2024-12-18 16:16:42
#> 138 2024-12-18 16:17:03
#> 139 2024-12-18 16:17:04
#> 140 2024-12-18 16:17:20
#> 141 2024-12-18 16:16:53
#> 142 2024-12-18 16:17:32
#> 143 2024-12-18 16:17:25
#> 144 2024-12-18 16:17:25
#> 145 2024-12-18 16:17:02
#> 146 2024-12-18 16:17:28
#> 147 2024-12-18 16:17:09
#> 148 2024-12-18 16:17:27
#> 149 2024-12-18 16:17:22
#> 150 2024-12-18 16:17:03
#> 151 2024-12-18 16:17:03
#> 152 2024-12-18 16:16:57
#> 153 2024-12-18 16:16:59
#> 154 2024-12-18 16:17:26
#> 155 2024-12-18 16:17:04
#> 156 2024-12-18 16:17:26
#> 157 2024-12-18 16:17:26
#> 158 2024-12-18 16:17:03
#> 159 2024-12-18 16:17:02
#> 160 2024-12-18 16:17:34
#> 161 2024-12-18 16:08:58
#> 162 2024-12-18 14:56:28
#> 163 2024-12-18 16:17:33
#> 164 2024-12-18 16:16:29
#> 165 2024-12-18 15:57:30
#> 166 2024-12-18 16:17:03
#> 167 2024-12-18 14:56:59
#> 168 2024-12-18 15:09:59
#> 169 2024-12-18 16:17:03
#> 170 2024-12-18 15:24:42
#> 171 2024-12-18 16:16:51
#> 172 2024-12-18 14:35:26
#> 173 2024-12-18 16:16:40
#> 174 2024-12-18 16:16:28
#> 175 2024-12-18 16:16:25
#> 176 2024-12-18 16:16:22
#> 177 2024-12-18 16:16:56
#> 178 2024-12-18 13:57:45
#> 179 2024-12-18 16:17:03
#> 180 2024-12-18 12:37:59
#> 181 2024-12-18 16:17:20
#> 182 2024-12-18 15:04:40
#> 183 2024-12-18 16:17:29
#> 184 2024-12-18 13:26:56
#> 185 2024-12-18 15:12:45
#> 186 2024-12-18 12:38:18
#> 187 2024-12-18 12:25:06
#> 188 2024-12-18 13:36:40
#> 189 2024-12-18 12:37:51
#> 190 2024-12-18 13:34:11
#> 191 2024-12-18 15:07:21
#> 192 2024-12-18 12:33:01
#> 193 2024-12-18 16:05:38
#> 194 2024-12-18 16:17:33
#> 195 2024-12-18 16:17:33
#> 196 2024-12-18 12:34:04
#> 197 2024-12-18 16:17:31
#> 198 2024-12-18 13:11:41
#> 199 2024-12-18 14:57:48
#> 200 2024-12-18 12:15:38