Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.683001    260.824 2024-11-18 23:23:35        Bid        NA         NA
#> 2    0.683437   2316.357 2024-11-18 22:13:18        Bid        NA         NA
#> 3    0.683700    290.000 2024-11-18 22:40:02        Bid        NA         NA
#> 4    0.684000     71.311 2024-11-18 20:45:12        Bid        NA         NA
#> 5    0.684030     80.157 2024-11-18 19:28:02        Bid        NA         NA
#> 6    0.684051     37.237 2024-11-18 19:37:37        Bid        NA         NA
#> 7    0.684171   6656.000 2024-11-18 22:13:20        Bid        NA         NA
#> 8    0.684642    269.356 2024-11-18 20:26:25        Bid        NA         NA
#> 9    0.684845   3672.788 2024-11-18 23:23:35        Bid        NA         NA
#> 10   0.684847   6603.578 2024-11-18 23:23:56        Bid        NA         NA
#> 11   0.684849   2000.412 2024-11-18 23:23:35        Bid        NA         NA
#> 12   0.685000   5648.231 2024-11-18 23:23:48        Bid        NA         NA
#> 13   0.685263     95.000 2024-11-18 20:42:24        Bid        NA         NA
#> 14   0.686000     71.311 2024-11-18 20:45:28        Bid        NA         NA
#> 15   0.686115   6386.900 2024-11-18 23:23:34        Bid        NA         NA
#> 16   0.686210    285.461 2024-11-18 20:23:21        Bid        NA         NA
#> 17   0.686365   9538.774 2024-11-18 23:21:10        Bid        NA         NA
#> 18   0.686406   9926.968 2024-11-18 23:21:21        Bid        NA         NA
#> 19   0.686504    119.840 2024-11-18 20:27:02        Bid        NA         NA
#> 20   0.686869   4515.578 2024-11-18 23:23:35        Bid        NA         NA
#> 21   0.686877     79.825 2024-11-18 19:36:09        Bid        NA         NA
#> 22   0.687000  10207.890 2024-11-18 20:34:07        Bid        NA         NA
#> 23   0.687076   3405.758 2024-11-18 23:23:46        Bid        NA         NA
#> 24   0.687538   9637.973 2024-11-18 23:20:37        Bid        NA         NA
#> 25   0.687700    581.649 2024-11-18 19:57:19        Bid        NA         NA
#> 26   0.687982   2403.224 2024-11-18 23:20:28        Bid        NA         NA
#> 27   0.688000     71.311 2024-11-18 20:45:39        Bid        NA         NA
#> 28   0.688305     79.660 2024-11-18 19:37:16        Bid        NA         NA
#> 29   0.689000     41.311 2024-11-18 20:27:06        Bid        NA         NA
#> 30   0.689100   1547.584 2024-11-18 23:20:29        Bid        NA         NA
#> 31   0.689642   4854.521 2024-11-18 23:23:34        Bid        NA         NA
#> 32   0.689736     79.494 2024-11-18 19:37:30        Bid        NA         NA
#> 33   0.690000     51.311 2024-11-18 22:33:23        Bid        NA         NA
#> 34   0.690008  68393.509 2024-11-18 23:23:34        Bid        NA         NA
#> 35   0.690329  13028.600 2024-11-18 23:23:10        Bid        NA         NA
#> 36   0.690379   1998.620 2024-11-18 23:20:28        Bid        NA         NA
#> 37   0.690380 204830.305 2024-11-18 23:23:54        Bid        NA         NA
#> 38   0.690571     36.886 2024-11-18 20:22:52        Bid        NA         NA
#> 39   0.690879     12.363 2024-11-18 19:39:18        Bid        NA         NA
#> 40   0.690890   4432.425 2024-11-18 23:23:34        Bid        NA         NA
#> 41   0.690913   1029.235 2024-11-18 23:23:56        Bid        NA         NA
#> 42   0.691000     41.311 2024-11-18 21:14:06        Bid        NA         NA
#> 43   0.691170     79.329 2024-11-18 19:37:27        Bid        NA         NA
#> 44   0.691192   3116.304 2024-11-18 23:23:35        Bid        NA         NA
#> 45   0.691200  68283.940 2024-11-18 23:20:35        Bid        NA         NA
#> 46   0.691981  26633.100 2024-11-18 23:23:10        Bid        NA         NA
#> 47   0.692000     41.311 2024-11-18 21:18:06        Bid        NA         NA
#> 48   0.692006     51.403 2024-11-18 20:25:54        Bid        NA         NA
#> 49   0.692023   1641.990 2024-11-18 23:23:35        Bid        NA         NA
#> 50   0.692056   4818.346 2024-11-18 23:23:10        Bid        NA         NA
#> 51   0.692057   9051.415 2024-11-18 23:23:35        Bid        NA         NA
#> 52   0.692607     79.165 2024-11-18 21:52:57        Bid        NA         NA
#> 53   0.692916    266.000 2024-11-18 20:03:52        Bid        NA         NA
#> 54   0.693200   9638.595 2024-11-18 23:23:34        Bid        NA         NA
#> 55   0.693226    150.093 2024-11-18 21:42:18        Bid        NA         NA
#> 56   0.693230  40016.600 2024-11-18 23:23:34        Bid        NA         NA
#> 57   0.693399  34029.512 2024-11-18 23:23:34        Bid        NA         NA
#> 58   0.693591 259518.763 2024-11-18 23:23:35        Bid        NA         NA
#> 59   0.693904    100.000 2024-11-18 21:25:17        Bid        NA         NA
#> 60   0.694019     12.307 2024-11-18 21:00:35        Bid        NA         NA
#> 61   0.694047     79.001 2024-11-18 20:53:33        Bid        NA         NA
#> 62   0.694270     40.090 2024-11-18 23:23:44        Bid        NA         NA
#> 63   0.694484  66848.600 2024-11-18 23:23:34        Bid        NA         NA
#> 64   0.694518   2168.581 2024-11-18 23:23:52        Bid        NA         NA
#> 65   0.694774   7263.217 2024-11-18 23:23:34        Bid        NA         NA
#> 66   0.694808   1192.741 2024-11-18 23:23:40        Bid        NA         NA
#> 67   0.694822     17.698 2024-11-18 23:23:34        Bid        NA         NA
#> 68   0.694856  33958.163 2024-11-18 23:23:45        Bid        NA         NA
#> 69   0.694863   7296.600 2024-11-18 23:23:34        Bid        NA         NA
#> 70   0.694944   5902.152 2024-11-18 23:23:34        Bid        NA         NA
#> 71   0.695000   1000.000 2024-11-18 21:18:02        Bid        NA         NA
#> 72   0.695001   3490.988 2024-11-18 23:23:41        Bid        NA         NA
#> 73   0.695033    216.859 2024-11-18 23:23:40        Bid        NA         NA
#> 74   0.695358  10154.300 2024-11-18 23:23:59        Bid        NA         NA
#> 75   0.695471   5833.611 2024-11-18 23:23:34        Bid        NA         NA
#> 76   0.695490     78.837 2024-11-18 22:18:00        Bid        NA         NA
#> 77   0.695496  15915.900 2024-11-18 23:23:45        Bid        NA         NA
#> 78   0.695745  33914.776 2024-11-18 23:23:40        Bid        NA         NA
#> 79   0.695868   3028.713 2024-11-18 23:23:35        Bid        NA         NA
#> 80   0.695913   1911.536 2024-11-18 23:23:39        Bid        NA         NA
#> 81   0.696028  20340.577 2024-11-18 23:23:36        Bid        NA         NA
#> 82   0.696075   1169.000 2024-11-18 23:23:35        Bid        NA         NA
#> 83   0.696131   5000.000 2024-11-18 23:23:35        Bid        NA         NA
#> 84   0.696405   3378.000 2024-11-18 23:23:53        Bid        NA         NA
#> 85   0.696451  13552.154 2024-11-18 23:23:45        Bid        NA         NA
#> 86   0.696707 135471.770 2024-11-18 23:23:53        Bid        NA         NA
#> 87   0.696727   4057.600 2024-11-18 23:23:49        Bid        NA         NA
#> 88   0.696864     33.799 2024-11-18 23:23:59        Bid        NA         NA
#> 89   0.696872     33.799 2024-11-18 23:23:59        Bid        NA         NA
#> 90   0.696883   3384.527 2024-11-18 23:23:59        Bid        NA         NA
#> 91   0.696936     78.673 2024-11-18 21:06:35        Bid        NA         NA
#> 92   0.696937    372.570 2024-11-18 23:23:53        Bid        NA         NA
#> 93   0.696950    894.900 2024-11-18 23:23:59        Bid        NA         NA
#> 94   0.697000     14.247 2024-11-18 23:19:45        Bid        NA         NA
#> 95   0.697050  13540.495 2024-11-18 23:23:56        Bid        NA         NA
#> 96   0.697051     50.559 2024-11-18 23:23:59        Bid        NA         NA
#> 97   0.697148  10153.948 2024-11-18 23:23:59        Bid        NA         NA
#> 98   0.697152     36.537 2024-11-18 21:14:31        Bid        NA         NA
#> 99   0.697173     12.251 2024-11-18 23:23:52        Bid        NA         NA
#> 100  0.697174  10153.579 2024-11-18 23:23:59        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.697412    544.713
#> 102        NA         NA                <NA>        Ask  0.697413  33833.663
#> 103        NA         NA                <NA>        Ask  0.697521   3384.527
#> 104        NA         NA                <NA>        Ask  0.697567    300.735
#> 105        NA         NA                <NA>        Ask  0.697597  62237.546
#> 106        NA         NA                <NA>        Ask  0.697688   3576.965
#> 107        NA         NA                <NA>        Ask  0.697694   5876.592
#> 108        NA         NA                <NA>        Ask  0.697712    360.218
#> 109        NA         NA                <NA>        Ask  0.697748     33.799
#> 110        NA         NA                <NA>        Ask  0.697792    895.000
#> 111        NA         NA                <NA>        Ask  0.697914   4289.433
#> 112        NA         NA                <NA>        Ask  0.697915    953.000
#> 113        NA         NA                <NA>        Ask  0.698072  13520.707
#> 114        NA         NA                <NA>        Ask  0.698155   7414.000
#> 115        NA         NA                <NA>        Ask  0.698207    270.000
#> 116        NA         NA                <NA>        Ask  0.698444   4058.900
#> 117        NA         NA                <NA>        Ask  0.698565  16213.393
#> 118        NA         NA                <NA>        Ask  0.698584  10000.000
#> 119        NA         NA                <NA>        Ask  0.698680   1317.000
#> 120        NA         NA                <NA>        Ask  0.698857   1987.202
#> 121        NA         NA                <NA>        Ask  0.698858  38994.297
#> 122        NA         NA                <NA>        Ask  0.698920   3028.713
#> 123        NA         NA                <NA>        Ask  0.698963  21605.538
#> 124        NA         NA                <NA>        Ask  0.699112   7783.163
#> 125        NA         NA                <NA>        Ask  0.699279   1950.805
#> 126        NA         NA                <NA>        Ask  0.699354   1104.000
#> 127        NA         NA                <NA>        Ask  0.699664    763.885
#> 128        NA         NA                <NA>        Ask  0.699698  10296.000
#> 129        NA         NA                <NA>        Ask  0.699753     17.767
#> 130        NA         NA                <NA>        Ask  0.699789   1959.531
#> 131        NA         NA                <NA>        Ask  0.699823  21579.001
#> 132        NA         NA                <NA>        Ask  0.699950   1428.133
#> 133        NA         NA                <NA>        Ask  0.699984   7305.600
#> 134        NA         NA                <NA>        Ask  0.700001   1727.045
#> 135        NA         NA                <NA>        Ask  0.700036   5000.000
#> 136        NA         NA                <NA>        Ask  0.700223   2025.956
#> 137        NA         NA                <NA>        Ask  0.700258  15402.400
#> 138        NA         NA                <NA>        Ask  0.700480     47.217
#> 139        NA         NA                <NA>        Ask  0.700541     13.334
#> 140        NA         NA                <NA>        Ask  0.700646   2324.867
#> 141        NA         NA                <NA>        Ask  0.701362  21531.643
#> 142        NA         NA                <NA>        Ask  0.701876   6458.309
#> 143        NA         NA                <NA>        Ask  0.703278  15897.463
#> 144        NA         NA                <NA>        Ask  0.703565  12175.401
#> 145        NA         NA                <NA>        Ask  0.703566  34879.307
#> 146        NA         NA                <NA>        Ask  0.704166  65503.100
#> 147        NA         NA                <NA>        Ask  0.704246   9945.755
#> 148        NA         NA                <NA>        Ask  0.704309   6058.274
#> 149        NA         NA                <NA>        Ask  0.704333   6092.069
#> 150        NA         NA                <NA>        Ask  0.704701   7142.080
#> 151        NA         NA                <NA>        Ask  0.704736  40183.455
#> 152        NA         NA                <NA>        Ask  0.705142 204819.411
#> 153        NA         NA                <NA>        Ask  0.705143   4343.384
#> 154        NA         NA                <NA>        Ask  0.705487  29882.457
#> 155        NA         NA                <NA>        Ask  0.706459     77.929
#> 156        NA         NA                <NA>        Ask  0.706474  39367.900
#> 157        NA         NA                <NA>        Ask  0.707026   1563.889
#> 158        NA         NA                <NA>        Ask  0.707093   9314.215
#> 159        NA         NA                <NA>        Ask  0.707912  25490.400
#> 160        NA         NA                <NA>        Ask  0.707928     77.767
#> 161        NA         NA                <NA>        Ask  0.708121   4685.310
#> 162        NA         NA                <NA>        Ask  0.708230    752.362
#> 163        NA         NA                <NA>        Ask  0.708304   9352.330
#> 164        NA         NA                <NA>        Ask  0.709068  12946.700
#> 165        NA         NA                <NA>        Ask  0.709399     77.606
#> 166        NA         NA                <NA>        Ask  0.709725     24.292
#> 167        NA         NA                <NA>        Ask  0.710000    448.095
#> 168        NA         NA                <NA>        Ask  0.710240     12.141
#> 169        NA         NA                <NA>        Ask  0.710834    456.251
#> 170        NA         NA                <NA>        Ask  0.710846    340.206
#> 171        NA         NA                <NA>        Ask  0.710874     77.445
#> 172        NA         NA                <NA>        Ask  0.711007   4360.464
#> 173        NA         NA                <NA>        Ask  0.711042   1926.680
#> 174        NA         NA                <NA>        Ask  0.711227   6358.182
#> 175        NA         NA                <NA>        Ask  0.711318    748.927
#> 176        NA         NA                <NA>        Ask  0.711345   3535.922
#> 177        NA         NA                <NA>        Ask  0.711630   9200.119
#> 178        NA         NA                <NA>        Ask  0.711986    147.919
#> 179        NA         NA                <NA>        Ask  0.711999   9570.140
#> 180        NA         NA                <NA>        Ask  0.712114     96.912
#> 181        NA         NA                <NA>        Ask  0.712352     77.284
#> 182        NA         NA                <NA>        Ask  0.713468     12.086
#> 183        NA         NA                <NA>        Ask  0.713833     77.124
#> 184        NA         NA                <NA>        Ask  0.714023     36.149
#> 185        NA         NA                <NA>        Ask  0.714427     25.000
#> 186        NA         NA                <NA>        Ask  0.715000     50.000
#> 187        NA         NA                <NA>        Ask  0.715316     76.964
#> 188        NA         NA                <NA>        Ask  0.715392     29.098
#> 189        NA         NA                <NA>        Ask  0.715398  10908.883
#> 190        NA         NA                <NA>        Ask  0.715772    168.151
#> 191        NA         NA                <NA>        Ask  0.715851   6643.000
#> 192        NA         NA                <NA>        Ask  0.716685    749.829
#> 193        NA         NA                <NA>        Ask  0.716711     12.031
#> 194        NA         NA                <NA>        Ask  0.716800    118.800
#> 195        NA         NA                <NA>        Ask  0.716803     76.804
#> 196        NA         NA                <NA>        Ask  0.716847   1484.126
#> 197        NA         NA                <NA>        Ask  0.716902    166.053
#> 198        NA         NA                <NA>        Ask  0.717496     95.000
#> 199        NA         NA                <NA>        Ask  0.717645     96.912
#> 200        NA         NA                <NA>        Ask  0.717692    671.026
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-11-18 23:23:59
#> 102 2024-11-18 23:23:59
#> 103 2024-11-18 23:23:59
#> 104 2024-11-18 23:23:44
#> 105 2024-11-18 23:23:55
#> 106 2024-11-18 23:23:59
#> 107 2024-11-18 23:23:59
#> 108 2024-11-18 23:23:37
#> 109 2024-11-18 23:23:59
#> 110 2024-11-18 23:23:59
#> 111 2024-11-18 23:23:59
#> 112 2024-11-18 23:23:35
#> 113 2024-11-18 23:23:45
#> 114 2024-11-18 23:23:54
#> 115 2024-11-18 23:23:34
#> 116 2024-11-18 23:23:52
#> 117 2024-11-18 23:23:59
#> 118 2024-11-18 23:23:59
#> 119 2024-11-18 23:23:59
#> 120 2024-11-18 23:23:59
#> 121 2024-11-18 23:23:59
#> 122 2024-11-18 23:23:59
#> 123 2024-11-18 23:23:49
#> 124 2024-11-18 23:23:59
#> 125 2024-11-18 23:23:59
#> 126 2024-11-18 23:23:44
#> 127 2024-11-18 23:23:51
#> 128 2024-11-18 23:23:49
#> 129 2024-11-18 23:23:34
#> 130 2024-11-18 23:23:51
#> 131 2024-11-18 23:23:43
#> 132 2024-11-18 23:23:51
#> 133 2024-11-18 23:23:48
#> 134 2024-11-18 23:23:51
#> 135 2024-11-18 23:23:45
#> 136 2024-11-18 23:23:51
#> 137 2024-11-18 23:23:34
#> 138 2024-11-18 23:23:40
#> 139 2024-11-18 23:23:34
#> 140 2024-11-18 23:23:51
#> 141 2024-11-18 23:23:34
#> 142 2024-11-18 23:23:52
#> 143 2024-11-18 23:23:55
#> 144 2024-11-18 23:23:36
#> 145 2024-11-18 23:23:32
#> 146 2024-11-18 23:23:45
#> 147 2024-11-18 23:23:59
#> 148 2024-11-18 23:23:41
#> 149 2024-11-18 23:23:13
#> 150 2024-11-18 23:23:39
#> 151 2024-11-18 23:20:30
#> 152 2024-11-18 23:23:55
#> 153 2024-11-18 23:22:31
#> 154 2024-11-18 23:23:30
#> 155 2024-11-18 23:06:39
#> 156 2024-11-18 23:23:10
#> 157 2024-11-18 23:21:16
#> 158 2024-11-18 23:22:13
#> 159 2024-11-18 23:23:25
#> 160 2024-11-18 21:31:51
#> 161 2024-11-18 23:23:18
#> 162 2024-11-18 23:22:01
#> 163 2024-11-18 23:23:34
#> 164 2024-11-18 23:23:45
#> 165 2024-11-18 20:34:04
#> 166 2024-11-18 22:16:23
#> 167 2024-11-18 23:20:28
#> 168 2024-11-18 20:06:33
#> 169 2024-11-18 23:21:56
#> 170 2024-11-18 23:06:55
#> 171 2024-11-18 20:18:30
#> 172 2024-11-18 23:22:32
#> 173 2024-11-18 23:23:11
#> 174 2024-11-18 23:23:12
#> 175 2024-11-18 23:20:50
#> 176 2024-11-18 23:22:35
#> 177 2024-11-18 23:21:10
#> 178 2024-11-18 21:42:18
#> 179 2024-11-18 23:21:21
#> 180 2024-11-18 19:21:05
#> 181 2024-11-18 19:41:38
#> 182 2024-11-18 20:01:58
#> 183 2024-11-18 20:13:05
#> 184 2024-11-18 20:10:23
#> 185 2024-11-18 19:47:02
#> 186 2024-11-18 22:43:27
#> 187 2024-11-18 20:21:03
#> 188 2024-11-18 19:39:08
#> 189 2024-11-18 23:23:34
#> 190 2024-11-18 19:11:49
#> 191 2024-11-18 22:55:03
#> 192 2024-11-18 20:11:21
#> 193 2024-11-18 20:05:16
#> 194 2024-11-18 20:45:40
#> 195 2024-11-18 19:03:32
#> 196 2024-11-18 23:23:32
#> 197 2024-11-18 20:51:50
#> 198 2024-11-18 23:02:38
#> 199 2024-11-18 19:11:00
#> 200 2024-11-18 23:02:59
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.682613    287.019 2024-11-18 23:23:35        Bid        NA         NA
#> 2    0.683000    139.899 2024-11-18 21:16:37        Bid        NA         NA
#> 3    0.683001    260.824 2024-11-18 23:23:35        Bid        NA         NA
#> 4    0.683437   2316.357 2024-11-18 22:13:18        Bid        NA         NA
#> 5    0.683700    290.000 2024-11-18 22:40:02        Bid        NA         NA
#> 6    0.684000     71.311 2024-11-18 20:45:12        Bid        NA         NA
#> 7    0.684030     80.157 2024-11-18 19:28:02        Bid        NA         NA
#> 8    0.684051     37.237 2024-11-18 19:37:37        Bid        NA         NA
#> 9    0.684171   6656.000 2024-11-18 22:13:20        Bid        NA         NA
#> 10   0.684642    269.356 2024-11-18 20:26:25        Bid        NA         NA
#> 11   0.684845   3672.788 2024-11-18 23:23:35        Bid        NA         NA
#> 12   0.684847   6603.578 2024-11-18 23:23:56        Bid        NA         NA
#> 13   0.684849   2000.412 2024-11-18 23:23:35        Bid        NA         NA
#> 14   0.685000   5648.231 2024-11-18 23:23:48        Bid        NA         NA
#> 15   0.685263     95.000 2024-11-18 20:42:24        Bid        NA         NA
#> 16   0.686000     71.311 2024-11-18 20:45:28        Bid        NA         NA
#> 17   0.686115   6386.900 2024-11-18 23:23:34        Bid        NA         NA
#> 18   0.686210    285.461 2024-11-18 20:23:21        Bid        NA         NA
#> 19   0.686365   9538.774 2024-11-18 23:21:10        Bid        NA         NA
#> 20   0.686406   9926.968 2024-11-18 23:21:21        Bid        NA         NA
#> 21   0.686504    119.840 2024-11-18 20:27:02        Bid        NA         NA
#> 22   0.686869   4515.578 2024-11-18 23:23:35        Bid        NA         NA
#> 23   0.686877     79.825 2024-11-18 19:36:09        Bid        NA         NA
#> 24   0.687000  10207.890 2024-11-18 20:34:07        Bid        NA         NA
#> 25   0.687076   3405.758 2024-11-18 23:23:46        Bid        NA         NA
#> 26   0.687538   9637.973 2024-11-18 23:20:37        Bid        NA         NA
#> 27   0.687700    581.649 2024-11-18 19:57:19        Bid        NA         NA
#> 28   0.687982   2403.224 2024-11-18 23:20:28        Bid        NA         NA
#> 29   0.688000     71.311 2024-11-18 20:45:39        Bid        NA         NA
#> 30   0.688305     79.660 2024-11-18 19:37:16        Bid        NA         NA
#> 31   0.689000     41.311 2024-11-18 20:27:06        Bid        NA         NA
#> 32   0.689100   1547.584 2024-11-18 23:20:29        Bid        NA         NA
#> 33   0.689642   4854.521 2024-11-18 23:23:34        Bid        NA         NA
#> 34   0.689736     79.494 2024-11-18 19:37:30        Bid        NA         NA
#> 35   0.690000     51.311 2024-11-18 22:33:23        Bid        NA         NA
#> 36   0.690008  68393.509 2024-11-18 23:23:34        Bid        NA         NA
#> 37   0.690329  13028.600 2024-11-18 23:23:10        Bid        NA         NA
#> 38   0.690379   1998.620 2024-11-18 23:20:28        Bid        NA         NA
#> 39   0.690380 204830.305 2024-11-18 23:23:54        Bid        NA         NA
#> 40   0.690571     36.886 2024-11-18 20:22:52        Bid        NA         NA
#> 41   0.690879     12.363 2024-11-18 19:39:18        Bid        NA         NA
#> 42   0.690890   4432.425 2024-11-18 23:23:34        Bid        NA         NA
#> 43   0.690913   1029.235 2024-11-18 23:23:56        Bid        NA         NA
#> 44   0.691000     41.311 2024-11-18 21:14:06        Bid        NA         NA
#> 45   0.691170     79.329 2024-11-18 19:37:27        Bid        NA         NA
#> 46   0.691192   3116.304 2024-11-18 23:23:35        Bid        NA         NA
#> 47   0.691200  68283.940 2024-11-18 23:20:35        Bid        NA         NA
#> 48   0.691981  26633.100 2024-11-18 23:23:10        Bid        NA         NA
#> 49   0.692000     41.311 2024-11-18 21:18:06        Bid        NA         NA
#> 50   0.692006     51.403 2024-11-18 20:25:54        Bid        NA         NA
#> 51   0.692023   1641.990 2024-11-18 23:23:35        Bid        NA         NA
#> 52   0.692056   4818.346 2024-11-18 23:23:10        Bid        NA         NA
#> 53   0.692057   9051.415 2024-11-18 23:23:35        Bid        NA         NA
#> 54   0.692607     79.165 2024-11-18 21:52:57        Bid        NA         NA
#> 55   0.692916    266.000 2024-11-18 20:03:52        Bid        NA         NA
#> 56   0.693200   9638.595 2024-11-18 23:23:34        Bid        NA         NA
#> 57   0.693226    150.093 2024-11-18 21:42:18        Bid        NA         NA
#> 58   0.693230  40016.600 2024-11-18 23:23:34        Bid        NA         NA
#> 59   0.693399  34029.512 2024-11-18 23:23:34        Bid        NA         NA
#> 60   0.693591 259518.763 2024-11-18 23:23:35        Bid        NA         NA
#> 61   0.693904    100.000 2024-11-18 21:25:17        Bid        NA         NA
#> 62   0.694019     12.307 2024-11-18 21:00:35        Bid        NA         NA
#> 63   0.694047     79.001 2024-11-18 20:53:33        Bid        NA         NA
#> 64   0.694270     40.090 2024-11-18 23:23:44        Bid        NA         NA
#> 65   0.694484  66848.600 2024-11-18 23:23:34        Bid        NA         NA
#> 66   0.694518   2168.581 2024-11-18 23:23:52        Bid        NA         NA
#> 67   0.694774   7263.217 2024-11-18 23:23:34        Bid        NA         NA
#> 68   0.694808   1192.741 2024-11-18 23:23:40        Bid        NA         NA
#> 69   0.694822     17.698 2024-11-18 23:23:34        Bid        NA         NA
#> 70   0.694856  33958.163 2024-11-18 23:23:45        Bid        NA         NA
#> 71   0.694863   7296.600 2024-11-18 23:23:34        Bid        NA         NA
#> 72   0.694944   5902.152 2024-11-18 23:23:34        Bid        NA         NA
#> 73   0.695000   1000.000 2024-11-18 21:18:02        Bid        NA         NA
#> 74   0.695001   3490.988 2024-11-18 23:23:41        Bid        NA         NA
#> 75   0.695033    216.859 2024-11-18 23:23:40        Bid        NA         NA
#> 76   0.695358  10154.300 2024-11-18 23:23:59        Bid        NA         NA
#> 77   0.695471   5833.611 2024-11-18 23:23:34        Bid        NA         NA
#> 78   0.695490     78.837 2024-11-18 22:18:00        Bid        NA         NA
#> 79   0.695496  15915.900 2024-11-18 23:23:45        Bid        NA         NA
#> 80   0.695745  33914.776 2024-11-18 23:23:40        Bid        NA         NA
#> 81   0.695868   3028.713 2024-11-18 23:23:35        Bid        NA         NA
#> 82   0.695913   1911.536 2024-11-18 23:23:39        Bid        NA         NA
#> 83   0.696028  20340.577 2024-11-18 23:23:36        Bid        NA         NA
#> 84   0.696075   1169.000 2024-11-18 23:23:35        Bid        NA         NA
#> 85   0.696131   5000.000 2024-11-18 23:23:35        Bid        NA         NA
#> 86   0.696405   3378.000 2024-11-18 23:23:53        Bid        NA         NA
#> 87   0.696451  13552.154 2024-11-18 23:23:45        Bid        NA         NA
#> 88   0.696707 135471.770 2024-11-18 23:23:53        Bid        NA         NA
#> 89   0.696727   4057.600 2024-11-18 23:23:49        Bid        NA         NA
#> 90   0.696728    769.201 2024-11-18 23:24:00        Bid        NA         NA
#> 91   0.696936     78.673 2024-11-18 21:06:35        Bid        NA         NA
#> 92   0.696937    372.570 2024-11-18 23:23:53        Bid        NA         NA
#> 93   0.696950    894.900 2024-11-18 23:23:59        Bid        NA         NA
#> 94   0.697000     14.247 2024-11-18 23:19:45        Bid        NA         NA
#> 95   0.697050  13540.495 2024-11-18 23:23:56        Bid        NA         NA
#> 96   0.697058  10155.258 2024-11-18 23:24:00        Bid        NA         NA
#> 97   0.697059     50.558 2024-11-18 23:24:00        Bid        NA         NA
#> 98   0.697150   3384.643 2024-11-18 23:24:00        Bid        NA         NA
#> 99   0.697152     36.537 2024-11-18 21:14:31        Bid        NA         NA
#> 100  0.697173     12.251 2024-11-18 23:23:52        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.697404    544.713
#> 102        NA         NA                <NA>        Ask  0.697405    970.000
#> 103        NA         NA                <NA>        Ask  0.697413  33833.663
#> 104        NA         NA                <NA>        Ask  0.697486   5413.712
#> 105        NA         NA                <NA>        Ask  0.697567    300.735
#> 106        NA         NA                <NA>        Ask  0.697593   3576.965
#> 107        NA         NA                <NA>        Ask  0.697597  62237.546
#> 108        NA         NA                <NA>        Ask  0.697599   5876.592
#> 109        NA         NA                <NA>        Ask  0.697645     33.799
#> 110        NA         NA                <NA>        Ask  0.697654     33.799
#> 111        NA         NA                <NA>        Ask  0.697688   3576.965
#> 112        NA         NA                <NA>        Ask  0.697694   5876.592
#> 113        NA         NA                <NA>        Ask  0.697712    360.218
#> 114        NA         NA                <NA>        Ask  0.697740     33.799
#> 115        NA         NA                <NA>        Ask  0.697748     33.799
#> 116        NA         NA                <NA>        Ask  0.697792    895.000
#> 117        NA         NA                <NA>        Ask  0.697914   4289.433
#> 118        NA         NA                <NA>        Ask  0.697915    953.000
#> 119        NA         NA                <NA>        Ask  0.698054   6743.000
#> 120        NA         NA                <NA>        Ask  0.698072  13520.707
#> 121        NA         NA                <NA>        Ask  0.698155   7414.000
#> 122        NA         NA                <NA>        Ask  0.698207    270.000
#> 123        NA         NA                <NA>        Ask  0.698444   4058.900
#> 124        NA         NA                <NA>        Ask  0.698489  10000.000
#> 125        NA         NA                <NA>        Ask  0.698565  38999.065
#> 126        NA         NA                <NA>        Ask  0.698584  10000.000
#> 127        NA         NA                <NA>        Ask  0.698680   1317.000
#> 128        NA         NA                <NA>        Ask  0.698858  16208.625
#> 129        NA         NA                <NA>        Ask  0.698920   3028.713
#> 130        NA         NA                <NA>        Ask  0.698963  21605.538
#> 131        NA         NA                <NA>        Ask  0.699112   7783.163
#> 132        NA         NA                <NA>        Ask  0.699279   1950.805
#> 133        NA         NA                <NA>        Ask  0.699354   1104.000
#> 134        NA         NA                <NA>        Ask  0.699664    763.885
#> 135        NA         NA                <NA>        Ask  0.699698  10296.000
#> 136        NA         NA                <NA>        Ask  0.699753     17.767
#> 137        NA         NA                <NA>        Ask  0.699789   1959.531
#> 138        NA         NA                <NA>        Ask  0.699823  21579.001
#> 139        NA         NA                <NA>        Ask  0.699950   1428.133
#> 140        NA         NA                <NA>        Ask  0.699984   7305.600
#> 141        NA         NA                <NA>        Ask  0.700001   1727.045
#> 142        NA         NA                <NA>        Ask  0.700036   5000.000
#> 143        NA         NA                <NA>        Ask  0.700223   2025.956
#> 144        NA         NA                <NA>        Ask  0.700258  15402.400
#> 145        NA         NA                <NA>        Ask  0.700480     47.217
#> 146        NA         NA                <NA>        Ask  0.700541     13.334
#> 147        NA         NA                <NA>        Ask  0.700646   2324.867
#> 148        NA         NA                <NA>        Ask  0.701362  21531.643
#> 149        NA         NA                <NA>        Ask  0.701876   6458.309
#> 150        NA         NA                <NA>        Ask  0.703278  15897.463
#> 151        NA         NA                <NA>        Ask  0.703565  12175.401
#> 152        NA         NA                <NA>        Ask  0.703566  34879.307
#> 153        NA         NA                <NA>        Ask  0.704166  65503.100
#> 154        NA         NA                <NA>        Ask  0.704246   9945.755
#> 155        NA         NA                <NA>        Ask  0.704309   6058.274
#> 156        NA         NA                <NA>        Ask  0.704333   6092.069
#> 157        NA         NA                <NA>        Ask  0.704701   7142.080
#> 158        NA         NA                <NA>        Ask  0.704736  40183.455
#> 159        NA         NA                <NA>        Ask  0.705142 204819.411
#> 160        NA         NA                <NA>        Ask  0.705143   4343.384
#> 161        NA         NA                <NA>        Ask  0.705487  29882.457
#> 162        NA         NA                <NA>        Ask  0.706459     77.929
#> 163        NA         NA                <NA>        Ask  0.706474  39367.900
#> 164        NA         NA                <NA>        Ask  0.707026   1563.889
#> 165        NA         NA                <NA>        Ask  0.707093   9314.215
#> 166        NA         NA                <NA>        Ask  0.707912  25490.400
#> 167        NA         NA                <NA>        Ask  0.707928     77.767
#> 168        NA         NA                <NA>        Ask  0.708121   4685.310
#> 169        NA         NA                <NA>        Ask  0.708230    752.362
#> 170        NA         NA                <NA>        Ask  0.708304   9352.330
#> 171        NA         NA                <NA>        Ask  0.709068  12946.700
#> 172        NA         NA                <NA>        Ask  0.709399     77.606
#> 173        NA         NA                <NA>        Ask  0.709725     24.292
#> 174        NA         NA                <NA>        Ask  0.710000    448.095
#> 175        NA         NA                <NA>        Ask  0.710240     12.141
#> 176        NA         NA                <NA>        Ask  0.710834    456.251
#> 177        NA         NA                <NA>        Ask  0.710846    340.206
#> 178        NA         NA                <NA>        Ask  0.710874     77.445
#> 179        NA         NA                <NA>        Ask  0.711007   4360.464
#> 180        NA         NA                <NA>        Ask  0.711042   1926.680
#> 181        NA         NA                <NA>        Ask  0.711227   6358.182
#> 182        NA         NA                <NA>        Ask  0.711318    748.927
#> 183        NA         NA                <NA>        Ask  0.711345   3535.922
#> 184        NA         NA                <NA>        Ask  0.711630   9200.119
#> 185        NA         NA                <NA>        Ask  0.711986    147.919
#> 186        NA         NA                <NA>        Ask  0.711999   9570.140
#> 187        NA         NA                <NA>        Ask  0.712114     96.912
#> 188        NA         NA                <NA>        Ask  0.712352     77.284
#> 189        NA         NA                <NA>        Ask  0.713468     12.086
#> 190        NA         NA                <NA>        Ask  0.713833     77.124
#> 191        NA         NA                <NA>        Ask  0.714023     36.149
#> 192        NA         NA                <NA>        Ask  0.714427     25.000
#> 193        NA         NA                <NA>        Ask  0.715000     50.000
#> 194        NA         NA                <NA>        Ask  0.715316     76.964
#> 195        NA         NA                <NA>        Ask  0.715392     29.098
#> 196        NA         NA                <NA>        Ask  0.715398  10908.883
#> 197        NA         NA                <NA>        Ask  0.715772    168.151
#> 198        NA         NA                <NA>        Ask  0.715851   6643.000
#> 199        NA         NA                <NA>        Ask  0.716685    749.829
#> 200        NA         NA                <NA>        Ask  0.716711     12.031
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2024-11-18 23:24:00
#> 102 2024-11-18 23:24:00
#> 103 2024-11-18 23:23:59
#> 104 2024-11-18 23:24:00
#> 105 2024-11-18 23:23:44
#> 106 2024-11-18 23:24:00
#> 107 2024-11-18 23:23:55
#> 108 2024-11-18 23:24:00
#> 109 2024-11-18 23:24:00
#> 110 2024-11-18 23:24:00
#> 111 2024-11-18 23:23:59
#> 112 2024-11-18 23:23:59
#> 113 2024-11-18 23:23:37
#> 114 2024-11-18 23:24:00
#> 115 2024-11-18 23:23:59
#> 116 2024-11-18 23:23:59
#> 117 2024-11-18 23:23:59
#> 118 2024-11-18 23:23:35
#> 119 2024-11-18 23:24:00
#> 120 2024-11-18 23:23:45
#> 121 2024-11-18 23:23:54
#> 122 2024-11-18 23:23:34
#> 123 2024-11-18 23:23:52
#> 124 2024-11-18 23:24:00
#> 125 2024-11-18 23:24:00
#> 126 2024-11-18 23:23:59
#> 127 2024-11-18 23:24:00
#> 128 2024-11-18 23:24:00
#> 129 2024-11-18 23:23:59
#> 130 2024-11-18 23:23:49
#> 131 2024-11-18 23:23:59
#> 132 2024-11-18 23:23:59
#> 133 2024-11-18 23:23:44
#> 134 2024-11-18 23:23:51
#> 135 2024-11-18 23:23:49
#> 136 2024-11-18 23:23:34
#> 137 2024-11-18 23:23:51
#> 138 2024-11-18 23:23:43
#> 139 2024-11-18 23:23:51
#> 140 2024-11-18 23:23:48
#> 141 2024-11-18 23:23:51
#> 142 2024-11-18 23:23:45
#> 143 2024-11-18 23:23:51
#> 144 2024-11-18 23:23:34
#> 145 2024-11-18 23:23:40
#> 146 2024-11-18 23:23:34
#> 147 2024-11-18 23:23:51
#> 148 2024-11-18 23:23:34
#> 149 2024-11-18 23:23:52
#> 150 2024-11-18 23:23:55
#> 151 2024-11-18 23:23:36
#> 152 2024-11-18 23:23:32
#> 153 2024-11-18 23:23:45
#> 154 2024-11-18 23:23:59
#> 155 2024-11-18 23:23:41
#> 156 2024-11-18 23:23:13
#> 157 2024-11-18 23:23:39
#> 158 2024-11-18 23:20:30
#> 159 2024-11-18 23:23:55
#> 160 2024-11-18 23:22:31
#> 161 2024-11-18 23:23:30
#> 162 2024-11-18 23:06:39
#> 163 2024-11-18 23:23:10
#> 164 2024-11-18 23:21:16
#> 165 2024-11-18 23:22:13
#> 166 2024-11-18 23:23:25
#> 167 2024-11-18 21:31:51
#> 168 2024-11-18 23:23:18
#> 169 2024-11-18 23:22:01
#> 170 2024-11-18 23:23:34
#> 171 2024-11-18 23:23:45
#> 172 2024-11-18 20:34:04
#> 173 2024-11-18 22:16:23
#> 174 2024-11-18 23:20:28
#> 175 2024-11-18 20:06:33
#> 176 2024-11-18 23:21:56
#> 177 2024-11-18 23:06:55
#> 178 2024-11-18 20:18:30
#> 179 2024-11-18 23:22:32
#> 180 2024-11-18 23:23:11
#> 181 2024-11-18 23:23:12
#> 182 2024-11-18 23:20:50
#> 183 2024-11-18 23:22:35
#> 184 2024-11-18 23:21:10
#> 185 2024-11-18 21:42:18
#> 186 2024-11-18 23:21:21
#> 187 2024-11-18 19:21:05
#> 188 2024-11-18 19:41:38
#> 189 2024-11-18 20:01:58
#> 190 2024-11-18 20:13:05
#> 191 2024-11-18 20:10:23
#> 192 2024-11-18 19:47:02
#> 193 2024-11-18 22:43:27
#> 194 2024-11-18 20:21:03
#> 195 2024-11-18 19:39:08
#> 196 2024-11-18 23:23:34
#> 197 2024-11-18 19:11:49
#> 198 2024-11-18 22:55:03
#> 199 2024-11-18 20:11:21
#> 200 2024-11-18 20:05:16