Retrieve Order Book Data from Kraken Exchange

This function fetches order book data from the Kraken API for a specified trading pair.

Usage

getOB(pair, count = NULL)

Arguments

pair

A character string specifying the trading pair (e.g., "ADAEUR"). This is a required parameter.

count

An optional integer between 1 and 500 specifying the number of orders to retrieve. Default is NULL.

Value

A data frame containing the order book data for the requested trading pair, with bid orders appearing first (sorted by price ascending) followed by ask orders (sorted by price ascending).

Examples

getOB("ADAEUR")
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.637629   7657.576 2025-03-13 16:05:28        Bid        NA         NA
#> 2    0.638114   6342.700 2025-03-13 16:05:38        Bid        NA         NA
#> 3    0.639000    125.326 2025-03-13 08:33:06        Bid        NA         NA
#> 4    0.639010    110.934 2025-03-11 04:21:42        Bid        NA         NA
#> 5    0.639732     61.020 2025-03-13 07:20:52        Bid        NA         NA
#> 6    0.639841  62271.100 2025-03-13 16:08:15        Bid        NA         NA
#> 7    0.640000   2160.327 2025-03-13 16:05:33        Bid        NA         NA
#> 8    0.640800   1000.000 2025-03-13 12:26:22        Bid        NA         NA
#> 9    0.640960   7582.319 2025-03-13 16:05:28        Bid        NA         NA
#> 10   0.641000    167.102 2025-03-13 15:22:41        Bid        NA         NA
#> 11   0.641373  12942.800 2025-03-13 16:05:38        Bid        NA         NA
#> 12   0.641691   2814.700 2025-03-13 16:10:21        Bid        NA         NA
#> 13   0.642000    101.029 2025-03-13 09:21:42        Bid        NA         NA
#> 14   0.642661  14799.529 2025-03-13 16:05:28        Bid        NA         NA
#> 15   0.643000    167.102 2025-03-13 06:53:05        Bid        NA         NA
#> 16   0.643003  26718.900 2025-03-13 16:05:38        Bid        NA         NA
#> 17   0.643632   9782.128 2025-03-13 16:10:46        Bid        NA         NA
#> 18   0.643761   9617.331 2025-03-13 16:08:22        Bid        NA         NA
#> 19   0.643997  28801.546 2025-03-13 16:01:45        Bid        NA         NA
#> 20   0.644131 354923.237 2025-03-13 16:10:46        Bid        NA         NA
#> 21   0.644315     61.020 2025-03-13 15:48:26        Bid        NA         NA
#> 22   0.644362  22493.615 2025-03-13 16:05:28        Bid        NA         NA
#> 23   0.644413  37301.000 2025-03-13 16:06:30        Bid        NA         NA
#> 24   0.644414   5868.419 2025-03-13 16:10:20        Bid        NA         NA
#> 25   0.644470  16775.801 2025-03-13 16:10:31        Bid        NA         NA
#> 26   0.644555      4.495 2025-03-13 14:34:32        Bid        NA         NA
#> 27   0.644632  38794.600 2025-03-13 16:05:38        Bid        NA         NA
#> 28   0.644973  15358.345 2025-03-13 16:05:14        Bid        NA         NA
#> 29   0.644984  34263.300 2025-03-13 16:08:20        Bid        NA         NA
#> 30   0.645000    667.102 2025-03-13 15:48:49        Bid        NA         NA
#> 31   0.645160  18344.800 2025-03-13 16:08:22        Bid        NA         NA
#> 32   0.645411  21748.498 2025-03-13 16:05:23        Bid        NA         NA
#> 33   0.645500   9882.818 2025-03-13 16:10:57        Bid        NA         NA
#> 34   0.645692   3871.816 2025-03-13 11:55:40        Bid        NA         NA
#> 35   0.645775  19163.684 2025-03-13 16:08:14        Bid        NA         NA
#> 36   0.645804  10283.427 2025-03-13 16:10:09        Bid        NA         NA
#> 37   0.645936  39391.300 2025-03-13 16:05:34        Bid        NA         NA
#> 38   0.645937  10620.758 2025-03-13 16:10:07        Bid        NA         NA
#> 39   0.646036   6248.943 2025-03-13 16:08:22        Bid        NA         NA
#> 40   0.646068   3636.905 2025-03-13 16:10:59        Bid        NA         NA
#> 41   0.646685     29.558 2025-03-13 15:48:55        Bid        NA         NA
#> 42   0.646705     61.899 2025-03-13 15:51:37        Bid        NA         NA
#> 43   0.646715 306162.599 2025-03-13 16:11:02        Bid        NA         NA
#> 44   0.646716    264.405 2025-03-13 16:11:04        Bid        NA         NA
#> 45   0.646779   7120.266 2025-03-13 16:11:04        Bid        NA         NA
#> 46   0.646917   7118.745 2025-03-13 16:11:02        Bid        NA         NA
#> 47   0.646918     56.559 2025-03-13 16:11:04        Bid        NA         NA
#> 48   0.647019   7117.624 2025-03-13 16:10:55        Bid        NA         NA
#> 49   0.647051   2818.530 2025-03-13 16:10:59        Bid        NA         NA
#> 50   0.647234     32.773 2025-03-13 16:10:30        Bid        NA         NA
#> 51   0.647488  29186.276 2025-03-13 16:08:26        Bid        NA         NA
#> 52   0.647520   2000.247 2025-03-13 16:10:59        Bid        NA         NA
#> 53   0.647689  28004.400 2025-03-13 16:10:09        Bid        NA         NA
#> 54   0.647694   7110.299 2025-03-13 16:10:36        Bid        NA         NA
#> 55   0.647720   5000.000 2025-03-13 16:07:24        Bid        NA         NA
#> 56   0.647815   3160.675 2025-03-13 16:11:04        Bid        NA         NA
#> 57   0.647836   2586.335 2025-03-13 16:10:54        Bid        NA         NA
#> 58   0.647901   1181.597 2025-03-13 16:10:45        Bid        NA         NA
#> 59   0.648037   7106.181 2025-03-13 16:10:00        Bid        NA         NA
#> 60   0.648051   7106.288 2025-03-13 16:11:05        Bid        NA         NA
#> 61   0.648103    363.551 2025-03-13 16:10:45        Bid        NA         NA
#> 62   0.648146     15.000 2025-03-13 16:10:55        Bid        NA         NA
#> 63   0.648495   7101.154 2025-03-13 16:10:18        Bid        NA         NA
#> 64   0.648617  18669.600 2025-03-13 16:10:09        Bid        NA         NA
#> 65   0.648809   7097.983 2025-03-13 16:11:02        Bid        NA         NA
#> 66   0.648836   3255.412 2025-03-13 16:10:46        Bid        NA         NA
#> 67   0.648837   1759.715 2025-03-13 16:10:52        Bid        NA         NA
#> 68   0.649001  11173.610 2025-03-13 16:10:56        Bid        NA         NA
#> 69   0.649011  15987.448 2025-03-13 16:11:04        Bid        NA         NA
#> 70   0.649101   3026.715 2025-03-13 16:10:51        Bid        NA         NA
#> 71   0.649114   7095.008 2025-03-13 16:10:46        Bid        NA         NA
#> 72   0.649136   7094.408 2025-03-13 16:10:55        Bid        NA         NA
#> 73   0.649167   7094.431 2025-03-13 16:10:51        Bid        NA         NA
#> 74   0.649205  11583.390 2025-03-13 16:10:52        Bid        NA         NA
#> 75   0.649242  15575.800 2025-03-13 16:10:51        Bid        NA         NA
#> 76   0.649273   1324.974 2025-03-13 16:08:19        Bid        NA         NA
#> 77   0.649287  11113.430 2025-03-13 16:11:02        Bid        NA         NA
#> 78   0.649318  15340.000 2025-03-13 16:10:46        Bid        NA         NA
#> 79   0.649394   4245.000 2025-03-13 16:11:04        Bid        NA         NA
#> 80   0.649433   7076.000 2025-03-13 16:11:04        Bid        NA         NA
#> 81   0.649442     15.000 2025-03-13 16:09:41        Bid        NA         NA
#> 82   0.649478   7090.803 2025-03-13 16:09:10        Bid        NA         NA
#> 83   0.649531   3659.558 2025-03-13 16:11:04        Bid        NA         NA
#> 84   0.649568   7089.688 2025-03-13 16:10:55        Bid        NA         NA
#> 85   0.649602   8813.253 2025-03-13 16:10:54        Bid        NA         NA
#> 86   0.649604  10674.300 2025-03-13 16:10:46        Bid        NA         NA
#> 87   0.649642  12460.600 2025-03-13 16:10:51        Bid        NA         NA
#> 88   0.649651   4252.341 2025-03-13 16:11:04        Bid        NA         NA
#> 89   0.649653   5905.059 2025-03-13 16:11:01        Bid        NA         NA
#> 90   0.649689   7088.471 2025-03-13 16:10:38        Bid        NA         NA
#> 91   0.649690    169.972 2025-03-13 16:11:02        Bid        NA         NA
#> 92   0.649794  13545.465 2025-03-13 16:11:05        Bid        NA         NA
#> 93   0.649820   3160.675 2025-03-13 16:11:01        Bid        NA         NA
#> 94   0.649821    404.982 2025-03-13 16:11:02        Bid        NA         NA
#> 95   0.649859   7259.847 2025-03-13 16:10:47        Bid        NA         NA
#> 96   0.649890   3543.088 2025-03-13 16:11:05        Bid        NA         NA
#> 97   0.649896   1590.600 2025-03-13 16:11:05        Bid        NA         NA
#> 98   0.649937   8725.200 2025-03-13 16:11:01        Bid        NA         NA
#> 99   0.649940     69.061 2025-03-13 16:11:02        Bid        NA         NA
#> 100  0.649975   7085.256 2025-03-13 16:11:04        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.650688  11136.812
#> 102        NA         NA                <NA>        Ask  0.650689   4774.387
#> 103        NA         NA                <NA>        Ask  0.650691  17693.648
#> 104        NA         NA                <NA>        Ask  0.650693     26.636
#> 105        NA         NA                <NA>        Ask  0.650695   3576.965
#> 106        NA         NA                <NA>        Ask  0.650703    296.192
#> 107        NA         NA                <NA>        Ask  0.650726    974.291
#> 108        NA         NA                <NA>        Ask  0.650728   9968.500
#> 109        NA         NA                <NA>        Ask  0.650730  10312.800
#> 110        NA         NA                <NA>        Ask  0.650732   3703.211
#> 111        NA         NA                <NA>        Ask  0.650765   7076.665
#> 112        NA         NA                <NA>        Ask  0.650770  12460.600
#> 113        NA         NA                <NA>        Ask  0.650928    418.000
#> 114        NA         NA                <NA>        Ask  0.650929   7074.872
#> 115        NA         NA                <NA>        Ask  0.651016   3684.053
#> 116        NA         NA                <NA>        Ask  0.651038   7073.691
#> 117        NA         NA                <NA>        Ask  0.651129   1376.267
#> 118        NA         NA                <NA>        Ask  0.651132  18499.025
#> 119        NA         NA                <NA>        Ask  0.651229   7108.340
#> 120        NA         NA                <NA>        Ask  0.651230  15574.000
#> 121        NA         NA                <NA>        Ask  0.651285  17338.408
#> 122        NA         NA                <NA>        Ask  0.651344  13710.060
#> 123        NA         NA                <NA>        Ask  0.651346   7070.343
#> 124        NA         NA                <NA>        Ask  0.651347  12395.685
#> 125        NA         NA                <NA>        Ask  0.651359   6025.023
#> 126        NA         NA                <NA>        Ask  0.651514   4250.545
#> 127        NA         NA                <NA>        Ask  0.651538  11052.159
#> 128        NA         NA                <NA>        Ask  0.651547  12032.910
#> 129        NA         NA                <NA>        Ask  0.651610   7068.035
#> 130        NA         NA                <NA>        Ask  0.651647     15.000
#> 131        NA         NA                <NA>        Ask  0.651667   7066.964
#> 132        NA         NA                <NA>        Ask  0.651711   3025.447
#> 133        NA         NA                <NA>        Ask  0.651848   2078.122
#> 134        NA         NA                <NA>        Ask  0.651910   7064.228
#> 135        NA         NA                <NA>        Ask  0.651961   7064.041
#> 136        NA         NA                <NA>        Ask  0.652045   1774.000
#> 137        NA         NA                <NA>        Ask  0.652060   3709.380
#> 138        NA         NA                <NA>        Ask  0.652400  16078.726
#> 139        NA         NA                <NA>        Ask  0.652590   7056.869
#> 140        NA         NA                <NA>        Ask  0.652756  18691.000
#> 141        NA         NA                <NA>        Ask  0.652880     15.000
#> 142        NA         NA                <NA>        Ask  0.653000   7052.211
#> 143        NA         NA                <NA>        Ask  0.653116   7049.000
#> 144        NA         NA                <NA>        Ask  0.653644   4667.050
#> 145        NA         NA                <NA>        Ask  0.653676   7044.886
#> 146        NA         NA                <NA>        Ask  0.653867     31.295
#> 147        NA         NA                <NA>        Ask  0.653868   7043.076
#> 148        NA         NA                <NA>        Ask  0.653971   6039.711
#> 149        NA         NA                <NA>        Ask  0.654003   4229.000
#> 150        NA         NA                <NA>        Ask  0.654049   8511.299
#> 151        NA         NA                <NA>        Ask  0.654056  27974.600
#> 152        NA         NA                <NA>        Ask  0.654247    101.382
#> 153        NA         NA                <NA>        Ask  0.654317   3521.792
#> 154        NA         NA                <NA>        Ask  0.654738  10982.321
#> 155        NA         NA                <NA>        Ask  0.655375   1112.138
#> 156        NA         NA                <NA>        Ask  0.655496   9732.106
#> 157        NA         NA                <NA>        Ask  0.655785  18078.793
#> 158        NA         NA                <NA>        Ask  0.655894  10125.983
#> 159        NA         NA                <NA>        Ask  0.656110  18271.033
#> 160        NA         NA                <NA>        Ask  0.656670    236.286
#> 161        NA         NA                <NA>        Ask  0.656671   6792.244
#> 162        NA         NA                <NA>        Ask  0.656709  16463.162
#> 163        NA         NA                <NA>        Ask  0.657382    498.268
#> 164        NA         NA                <NA>        Ask  0.657383  38422.800
#> 165        NA         NA                <NA>        Ask  0.657442 356524.640
#> 166        NA         NA                <NA>        Ask  0.657443  34233.800
#> 167        NA         NA                <NA>        Ask  0.657673   9573.283
#> 168        NA         NA                <NA>        Ask  0.657902  84957.095
#> 169        NA         NA                <NA>        Ask  0.657903  22787.192
#> 170        NA         NA                <NA>        Ask  0.658212  38897.300
#> 171        NA         NA                <NA>        Ask  0.658243  21882.619
#> 172        NA         NA                <NA>        Ask  0.658373  12699.741
#> 173        NA         NA                <NA>        Ask  0.658810  16819.027
#> 174        NA         NA                <NA>        Ask  0.659471  31499.712
#> 175        NA         NA                <NA>        Ask  0.659485  25431.900
#> 176        NA         NA                <NA>        Ask  0.659585  37312.300
#> 177        NA         NA                <NA>        Ask  0.659742     52.359
#> 178        NA         NA                <NA>        Ask  0.659791  10511.892
#> 179        NA         NA                <NA>        Ask  0.659910  13185.312
#> 180        NA         NA                <NA>        Ask  0.660786  62323.200
#> 181        NA         NA                <NA>        Ask  0.663596  12815.400
#> 182        NA         NA                <NA>        Ask  0.664272     75.873
#> 183        NA         NA                <NA>        Ask  0.666736    108.753
#> 184        NA         NA                <NA>        Ask  0.666865   6441.500
#> 185        NA         NA                <NA>        Ask  0.669735     29.023
#> 186        NA         NA                <NA>        Ask  0.670215   1878.715
#> 187        NA         NA                <NA>        Ask  0.670217   5636.121
#> 188        NA         NA                <NA>        Ask  0.670232  23727.971
#> 189        NA         NA                <NA>        Ask  0.670234   5536.879
#> 190        NA         NA                <NA>        Ask  0.670774   1779.369
#> 191        NA         NA                <NA>        Ask  0.670775   3558.736
#> 192        NA         NA                <NA>        Ask  0.670933   7133.994
#> 193        NA         NA                <NA>        Ask  0.670942   7298.188
#> 194        NA         NA                <NA>        Ask  0.670945   1839.683
#> 195        NA         NA                <NA>        Ask  0.671377     59.983
#> 196        NA         NA                <NA>        Ask  0.671583  35317.542
#> 197        NA         NA                <NA>        Ask  0.671768   6500.000
#> 198        NA         NA                <NA>        Ask  0.672519   2078.294
#> 199        NA         NA                <NA>        Ask  0.673090    195.985
#> 200        NA         NA                <NA>        Ask  0.674647   9700.573
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2025-03-13 16:11:04
#> 102 2025-03-13 16:11:04
#> 103 2025-03-13 16:11:05
#> 104 2025-03-13 16:11:04
#> 105 2025-03-13 16:11:04
#> 106 2025-03-13 16:11:03
#> 107 2025-03-13 16:10:51
#> 108 2025-03-13 16:10:51
#> 109 2025-03-13 16:10:51
#> 110 2025-03-13 16:10:51
#> 111 2025-03-13 16:11:03
#> 112 2025-03-13 16:10:51
#> 113 2025-03-13 16:11:04
#> 114 2025-03-13 16:11:01
#> 115 2025-03-13 16:11:04
#> 116 2025-03-13 16:11:03
#> 117 2025-03-13 16:11:02
#> 118 2025-03-13 16:10:53
#> 119 2025-03-13 16:11:01
#> 120 2025-03-13 16:10:46
#> 121 2025-03-13 16:10:58
#> 122 2025-03-13 16:10:56
#> 123 2025-03-13 16:10:46
#> 124 2025-03-13 16:10:46
#> 125 2025-03-13 16:10:52
#> 126 2025-03-13 16:10:58
#> 127 2025-03-13 16:11:04
#> 128 2025-03-13 16:10:53
#> 129 2025-03-13 16:10:52
#> 130 2025-03-13 16:10:56
#> 131 2025-03-13 16:10:37
#> 132 2025-03-13 16:11:02
#> 133 2025-03-13 16:10:42
#> 134 2025-03-13 16:10:58
#> 135 2025-03-13 16:10:45
#> 136 2025-03-13 16:10:38
#> 137 2025-03-13 16:10:44
#> 138 2025-03-13 16:10:50
#> 139 2025-03-13 16:11:03
#> 140 2025-03-13 16:10:32
#> 141 2025-03-13 16:11:01
#> 142 2025-03-13 16:09:48
#> 143 2025-03-13 16:09:24
#> 144 2025-03-13 16:10:44
#> 145 2025-03-13 16:09:54
#> 146 2025-03-13 16:10:26
#> 147 2025-03-13 16:11:04
#> 148 2025-03-13 16:10:44
#> 149 2025-03-13 16:09:10
#> 150 2025-03-13 16:10:58
#> 151 2025-03-13 16:09:30
#> 152 2025-03-13 16:05:39
#> 153 2025-03-13 16:06:07
#> 154 2025-03-13 16:10:58
#> 155 2025-03-13 16:07:42
#> 156 2025-03-13 16:10:46
#> 157 2025-03-13 16:10:09
#> 158 2025-03-13 16:10:39
#> 159 2025-03-13 16:10:09
#> 160 2025-03-13 16:10:58
#> 161 2025-03-13 16:09:02
#> 162 2025-03-13 16:10:51
#> 163 2025-03-13 16:10:04
#> 164 2025-03-13 16:09:27
#> 165 2025-03-13 16:10:46
#> 166 2025-03-13 16:08:50
#> 167 2025-03-13 16:10:39
#> 168 2025-03-13 16:09:03
#> 169 2025-03-13 16:06:33
#> 170 2025-03-13 16:09:33
#> 171 2025-03-13 16:09:35
#> 172 2025-03-13 16:09:09
#> 173 2025-03-13 16:05:01
#> 174 2025-03-13 16:05:01
#> 175 2025-03-13 16:10:11
#> 176 2025-03-13 16:08:56
#> 177 2025-03-13 15:43:30
#> 178 2025-03-13 16:05:01
#> 179 2025-03-13 16:07:37
#> 180 2025-03-13 16:11:01
#> 181 2025-03-13 16:05:38
#> 182 2025-03-13 14:53:51
#> 183 2025-03-13 15:36:08
#> 184 2025-03-13 16:05:38
#> 185 2025-03-13 16:04:42
#> 186 2025-03-13 16:11:02
#> 187 2025-03-13 16:11:03
#> 188 2025-03-13 16:11:01
#> 189 2025-03-13 16:11:00
#> 190 2025-03-13 16:10:33
#> 191 2025-03-13 16:10:34
#> 192 2025-03-13 16:10:36
#> 193 2025-03-13 16:10:35
#> 194 2025-03-13 16:10:35
#> 195 2025-03-13 16:09:46
#> 196 2025-03-13 16:00:34
#> 197 2025-03-13 16:05:39
#> 198 2025-03-13 16:07:41
#> 199 2025-03-13 16:07:30
#> 200 2025-03-13 16:06:14
getOB("ADAEUR", count = 100)
#>     Bid_Price Bid_Volume       Bid_Timestamp Order_Type Ask_Price Ask_Volume
#> 1    0.639841  62271.100 2025-03-13 16:08:15        Bid        NA         NA
#> 2    0.640000   2160.327 2025-03-13 16:05:33        Bid        NA         NA
#> 3    0.640800   1000.000 2025-03-13 12:26:22        Bid        NA         NA
#> 4    0.640960   7582.319 2025-03-13 16:05:28        Bid        NA         NA
#> 5    0.641000    167.102 2025-03-13 15:22:41        Bid        NA         NA
#> 6    0.641373  12942.800 2025-03-13 16:05:38        Bid        NA         NA
#> 7    0.641691   2814.700 2025-03-13 16:10:21        Bid        NA         NA
#> 8    0.642000    101.029 2025-03-13 09:21:42        Bid        NA         NA
#> 9    0.642661  14799.529 2025-03-13 16:05:28        Bid        NA         NA
#> 10   0.643000    167.102 2025-03-13 06:53:05        Bid        NA         NA
#> 11   0.643003  26718.900 2025-03-13 16:05:38        Bid        NA         NA
#> 12   0.643632   9782.128 2025-03-13 16:10:46        Bid        NA         NA
#> 13   0.643761   9617.331 2025-03-13 16:08:22        Bid        NA         NA
#> 14   0.643997  28801.546 2025-03-13 16:01:45        Bid        NA         NA
#> 15   0.644000 356955.025 2025-03-13 16:11:06        Bid        NA         NA
#> 16   0.644315     61.020 2025-03-13 15:48:26        Bid        NA         NA
#> 17   0.644362  22493.615 2025-03-13 16:05:28        Bid        NA         NA
#> 18   0.644413  37301.000 2025-03-13 16:06:30        Bid        NA         NA
#> 19   0.644414   5868.419 2025-03-13 16:10:20        Bid        NA         NA
#> 20   0.644470  16775.801 2025-03-13 16:10:31        Bid        NA         NA
#> 21   0.644555      4.495 2025-03-13 14:34:32        Bid        NA         NA
#> 22   0.644632  38794.600 2025-03-13 16:05:38        Bid        NA         NA
#> 23   0.644973  15358.345 2025-03-13 16:05:14        Bid        NA         NA
#> 24   0.644984  34263.300 2025-03-13 16:08:20        Bid        NA         NA
#> 25   0.645000    667.102 2025-03-13 15:48:49        Bid        NA         NA
#> 26   0.645160  18344.800 2025-03-13 16:08:22        Bid        NA         NA
#> 27   0.645411  21748.498 2025-03-13 16:05:23        Bid        NA         NA
#> 28   0.645500   9882.818 2025-03-13 16:10:57        Bid        NA         NA
#> 29   0.645692   3871.816 2025-03-13 11:55:40        Bid        NA         NA
#> 30   0.645775  19163.684 2025-03-13 16:08:14        Bid        NA         NA
#> 31   0.645804  10283.427 2025-03-13 16:10:09        Bid        NA         NA
#> 32   0.645936  39391.300 2025-03-13 16:05:34        Bid        NA         NA
#> 33   0.645937  10620.758 2025-03-13 16:10:07        Bid        NA         NA
#> 34   0.646036   6248.943 2025-03-13 16:08:22        Bid        NA         NA
#> 35   0.646068   3636.905 2025-03-13 16:10:59        Bid        NA         NA
#> 36   0.646685     29.558 2025-03-13 15:48:55        Bid        NA         NA
#> 37   0.646705     61.899 2025-03-13 15:51:37        Bid        NA         NA
#> 38   0.646715 306162.599 2025-03-13 16:11:02        Bid        NA         NA
#> 39   0.646716    264.405 2025-03-13 16:11:04        Bid        NA         NA
#> 40   0.646779   7120.266 2025-03-13 16:11:04        Bid        NA         NA
#> 41   0.646917   7118.745 2025-03-13 16:11:02        Bid        NA         NA
#> 42   0.646918     56.559 2025-03-13 16:11:04        Bid        NA         NA
#> 43   0.647019   7117.624 2025-03-13 16:10:55        Bid        NA         NA
#> 44   0.647051   2818.530 2025-03-13 16:10:59        Bid        NA         NA
#> 45   0.647234     32.773 2025-03-13 16:10:30        Bid        NA         NA
#> 46   0.647488  29186.276 2025-03-13 16:08:26        Bid        NA         NA
#> 47   0.647520   2000.247 2025-03-13 16:10:59        Bid        NA         NA
#> 48   0.647689  28004.400 2025-03-13 16:10:09        Bid        NA         NA
#> 49   0.647694   7110.299 2025-03-13 16:10:36        Bid        NA         NA
#> 50   0.647720   5000.000 2025-03-13 16:07:24        Bid        NA         NA
#> 51   0.647815   3160.675 2025-03-13 16:11:04        Bid        NA         NA
#> 52   0.647836   2586.335 2025-03-13 16:10:54        Bid        NA         NA
#> 53   0.647901   1181.597 2025-03-13 16:10:45        Bid        NA         NA
#> 54   0.648037   7106.181 2025-03-13 16:10:00        Bid        NA         NA
#> 55   0.648051   7106.288 2025-03-13 16:11:05        Bid        NA         NA
#> 56   0.648103    363.551 2025-03-13 16:10:45        Bid        NA         NA
#> 57   0.648146     15.000 2025-03-13 16:10:55        Bid        NA         NA
#> 58   0.648495   7101.154 2025-03-13 16:10:18        Bid        NA         NA
#> 59   0.648617  18669.600 2025-03-13 16:10:09        Bid        NA         NA
#> 60   0.648648  11316.260 2025-03-13 16:11:05        Bid        NA         NA
#> 61   0.648809   7097.983 2025-03-13 16:11:02        Bid        NA         NA
#> 62   0.648836   3255.412 2025-03-13 16:10:46        Bid        NA         NA
#> 63   0.648928  10987.448 2025-03-13 16:11:07        Bid        NA         NA
#> 64   0.649011   5000.000 2025-03-13 16:11:05        Bid        NA         NA
#> 65   0.649012   1758.374 2025-03-13 16:11:05        Bid        NA         NA
#> 66   0.649014  15340.000 2025-03-13 16:11:05        Bid        NA         NA
#> 67   0.649097   8813.253 2025-03-13 16:11:07        Bid        NA         NA
#> 68   0.649101   3026.715 2025-03-13 16:10:51        Bid        NA         NA
#> 69   0.649114   7095.008 2025-03-13 16:10:46        Bid        NA         NA
#> 70   0.649136   7094.408 2025-03-13 16:10:55        Bid        NA         NA
#> 71   0.649167   7094.431 2025-03-13 16:10:51        Bid        NA         NA
#> 72   0.649205  11583.390 2025-03-13 16:10:52        Bid        NA         NA
#> 73   0.649242  15575.800 2025-03-13 16:10:51        Bid        NA         NA
#> 74   0.649250   3539.000 2025-03-13 16:11:06        Bid        NA         NA
#> 75   0.649273   1324.974 2025-03-13 16:08:19        Bid        NA         NA
#> 76   0.649287  11113.430 2025-03-13 16:11:02        Bid        NA         NA
#> 77   0.649335  10674.300 2025-03-13 16:11:05        Bid        NA         NA
#> 78   0.649422   5905.059 2025-03-13 16:11:07        Bid        NA         NA
#> 79   0.649442     15.000 2025-03-13 16:09:41        Bid        NA         NA
#> 80   0.649447   3659.558 2025-03-13 16:11:07        Bid        NA         NA
#> 81   0.649478   7090.803 2025-03-13 16:09:10        Bid        NA         NA
#> 82   0.649568   7089.688 2025-03-13 16:10:55        Bid        NA         NA
#> 83   0.649642  12460.600 2025-03-13 16:10:51        Bid        NA         NA
#> 84   0.649651   4252.341 2025-03-13 16:11:04        Bid        NA         NA
#> 85   0.649654    384.135 2025-03-13 16:11:05        Bid        NA         NA
#> 86   0.649655    169.972 2025-03-13 16:11:06        Bid        NA         NA
#> 87   0.649684  13000.000 2025-03-13 16:11:05        Bid        NA         NA
#> 88   0.649689   7088.471 2025-03-13 16:10:38        Bid        NA         NA
#> 89   0.649691   2482.247 2025-03-13 16:11:05        Bid        NA         NA
#> 90   0.649735   2116.403 2025-03-13 16:11:05        Bid        NA         NA
#> 91   0.649738   3543.445 2025-03-13 16:11:06        Bid        NA         NA
#> 92   0.649749     69.081 2025-03-13 16:11:05        Bid        NA         NA
#> 93   0.649794  13545.465 2025-03-13 16:11:05        Bid        NA         NA
#> 94   0.649807   8726.200 2025-03-13 16:11:05        Bid        NA         NA
#> 95   0.649813   2657.984 2025-03-13 16:11:05        Bid        NA         NA
#> 96   0.649820   3160.675 2025-03-13 16:11:01        Bid        NA         NA
#> 97   0.649840   3543.360 2025-03-13 16:11:05        Bid        NA         NA
#> 98   0.649859   7259.847 2025-03-13 16:10:47        Bid        NA         NA
#> 99   0.649892   1590.500 2025-03-13 16:11:06        Bid        NA         NA
#> 100  0.649893   3542.602 2025-03-13 16:11:07        Bid        NA         NA
#> 101        NA         NA                <NA>        Ask  0.650643  11137.134
#> 102        NA         NA                <NA>        Ask  0.650644   3576.965
#> 103        NA         NA                <NA>        Ask  0.650646  14590.700
#> 104        NA         NA                <NA>        Ask  0.650647   4774.387
#> 105        NA         NA                <NA>        Ask  0.650649   7108.340
#> 106        NA         NA                <NA>        Ask  0.650650  24769.360
#> 107        NA         NA                <NA>        Ask  0.650691   7077.460
#> 108        NA         NA                <NA>        Ask  0.650693     26.636
#> 109        NA         NA                <NA>        Ask  0.650703    296.192
#> 110        NA         NA                <NA>        Ask  0.650726    974.291
#> 111        NA         NA                <NA>        Ask  0.650728   9968.500
#> 112        NA         NA                <NA>        Ask  0.650730   8722.400
#> 113        NA         NA                <NA>        Ask  0.650732   3703.211
#> 114        NA         NA                <NA>        Ask  0.650770  12460.600
#> 115        NA         NA                <NA>        Ask  0.650924    639.000
#> 116        NA         NA                <NA>        Ask  0.650926   1376.267
#> 117        NA         NA                <NA>        Ask  0.650929   7074.872
#> 118        NA         NA                <NA>        Ask  0.650932   3684.053
#> 119        NA         NA                <NA>        Ask  0.650982  10267.400
#> 120        NA         NA                <NA>        Ask  0.650985  12395.685
#> 121        NA         NA                <NA>        Ask  0.651038   7073.691
#> 122        NA         NA                <NA>        Ask  0.651132  18499.025
#> 123        NA         NA                <NA>        Ask  0.651230  15574.000
#> 124        NA         NA                <NA>        Ask  0.651285   7071.008
#> 125        NA         NA                <NA>        Ask  0.651346   7070.343
#> 126        NA         NA                <NA>        Ask  0.651359   6025.023
#> 127        NA         NA                <NA>        Ask  0.651454  11052.159
#> 128        NA         NA                <NA>        Ask  0.651475  10886.410
#> 129        NA         NA                <NA>        Ask  0.651514   4250.545
#> 130        NA         NA                <NA>        Ask  0.651547  12032.910
#> 131        NA         NA                <NA>        Ask  0.651610   7068.035
#> 132        NA         NA                <NA>        Ask  0.651647     15.000
#> 133        NA         NA                <NA>        Ask  0.651667   7066.964
#> 134        NA         NA                <NA>        Ask  0.651711   3025.447
#> 135        NA         NA                <NA>        Ask  0.651848   2078.122
#> 136        NA         NA                <NA>        Ask  0.651961   7064.041
#> 137        NA         NA                <NA>        Ask  0.652045   1774.000
#> 138        NA         NA                <NA>        Ask  0.652060   3709.380
#> 139        NA         NA                <NA>        Ask  0.652400  16078.726
#> 140        NA         NA                <NA>        Ask  0.652590   7056.869
#> 141        NA         NA                <NA>        Ask  0.652756  18691.000
#> 142        NA         NA                <NA>        Ask  0.652880     15.000
#> 143        NA         NA                <NA>        Ask  0.653000   7052.211
#> 144        NA         NA                <NA>        Ask  0.653116   7049.000
#> 145        NA         NA                <NA>        Ask  0.653644   4667.050
#> 146        NA         NA                <NA>        Ask  0.653676   7044.886
#> 147        NA         NA                <NA>        Ask  0.653867     31.295
#> 148        NA         NA                <NA>        Ask  0.653868   7043.076
#> 149        NA         NA                <NA>        Ask  0.653971   6039.711
#> 150        NA         NA                <NA>        Ask  0.654003   4229.000
#> 151        NA         NA                <NA>        Ask  0.654049   8511.299
#> 152        NA         NA                <NA>        Ask  0.654056  27974.600
#> 153        NA         NA                <NA>        Ask  0.654247    101.382
#> 154        NA         NA                <NA>        Ask  0.654317   3521.792
#> 155        NA         NA                <NA>        Ask  0.654738  10982.321
#> 156        NA         NA                <NA>        Ask  0.655375   1112.138
#> 157        NA         NA                <NA>        Ask  0.655496   9732.106
#> 158        NA         NA                <NA>        Ask  0.655785  18078.793
#> 159        NA         NA                <NA>        Ask  0.655894  10125.983
#> 160        NA         NA                <NA>        Ask  0.656110  18271.033
#> 161        NA         NA                <NA>        Ask  0.656670    236.286
#> 162        NA         NA                <NA>        Ask  0.656671   6792.244
#> 163        NA         NA                <NA>        Ask  0.656709  16463.162
#> 164        NA         NA                <NA>        Ask  0.657381 353926.433
#> 165        NA         NA                <NA>        Ask  0.657382     63.810
#> 166        NA         NA                <NA>        Ask  0.657383  38422.800
#> 167        NA         NA                <NA>        Ask  0.657443  34233.800
#> 168        NA         NA                <NA>        Ask  0.657673   9573.283
#> 169        NA         NA                <NA>        Ask  0.657902  84957.095
#> 170        NA         NA                <NA>        Ask  0.657903  22787.192
#> 171        NA         NA                <NA>        Ask  0.658212  38897.300
#> 172        NA         NA                <NA>        Ask  0.658243  21882.619
#> 173        NA         NA                <NA>        Ask  0.658373  12699.741
#> 174        NA         NA                <NA>        Ask  0.658810  16819.027
#> 175        NA         NA                <NA>        Ask  0.659471  31499.712
#> 176        NA         NA                <NA>        Ask  0.659485  25431.900
#> 177        NA         NA                <NA>        Ask  0.659585  37312.300
#> 178        NA         NA                <NA>        Ask  0.659742     52.359
#> 179        NA         NA                <NA>        Ask  0.659791  10511.892
#> 180        NA         NA                <NA>        Ask  0.659910  13185.312
#> 181        NA         NA                <NA>        Ask  0.660786  62323.200
#> 182        NA         NA                <NA>        Ask  0.663596  12815.400
#> 183        NA         NA                <NA>        Ask  0.664272     75.873
#> 184        NA         NA                <NA>        Ask  0.666735    193.600
#> 185        NA         NA                <NA>        Ask  0.666736    108.753
#> 186        NA         NA                <NA>        Ask  0.666864    252.347
#> 187        NA         NA                <NA>        Ask  0.666865   6441.500
#> 188        NA         NA                <NA>        Ask  0.669735     29.023
#> 189        NA         NA                <NA>        Ask  0.670215   1878.715
#> 190        NA         NA                <NA>        Ask  0.670217   5636.121
#> 191        NA         NA                <NA>        Ask  0.670232  23727.971
#> 192        NA         NA                <NA>        Ask  0.670234   5536.879
#> 193        NA         NA                <NA>        Ask  0.670774   1779.369
#> 194        NA         NA                <NA>        Ask  0.670775   3558.736
#> 195        NA         NA                <NA>        Ask  0.670933   7133.994
#> 196        NA         NA                <NA>        Ask  0.670942   7298.188
#> 197        NA         NA                <NA>        Ask  0.670945   1839.683
#> 198        NA         NA                <NA>        Ask  0.671377     59.983
#> 199        NA         NA                <NA>        Ask  0.671583  35317.542
#> 200        NA         NA                <NA>        Ask  0.671768   6500.000
#>           Ask_Timestamp
#> 1                  <NA>
#> 2                  <NA>
#> 3                  <NA>
#> 4                  <NA>
#> 5                  <NA>
#> 6                  <NA>
#> 7                  <NA>
#> 8                  <NA>
#> 9                  <NA>
#> 10                 <NA>
#> 11                 <NA>
#> 12                 <NA>
#> 13                 <NA>
#> 14                 <NA>
#> 15                 <NA>
#> 16                 <NA>
#> 17                 <NA>
#> 18                 <NA>
#> 19                 <NA>
#> 20                 <NA>
#> 21                 <NA>
#> 22                 <NA>
#> 23                 <NA>
#> 24                 <NA>
#> 25                 <NA>
#> 26                 <NA>
#> 27                 <NA>
#> 28                 <NA>
#> 29                 <NA>
#> 30                 <NA>
#> 31                 <NA>
#> 32                 <NA>
#> 33                 <NA>
#> 34                 <NA>
#> 35                 <NA>
#> 36                 <NA>
#> 37                 <NA>
#> 38                 <NA>
#> 39                 <NA>
#> 40                 <NA>
#> 41                 <NA>
#> 42                 <NA>
#> 43                 <NA>
#> 44                 <NA>
#> 45                 <NA>
#> 46                 <NA>
#> 47                 <NA>
#> 48                 <NA>
#> 49                 <NA>
#> 50                 <NA>
#> 51                 <NA>
#> 52                 <NA>
#> 53                 <NA>
#> 54                 <NA>
#> 55                 <NA>
#> 56                 <NA>
#> 57                 <NA>
#> 58                 <NA>
#> 59                 <NA>
#> 60                 <NA>
#> 61                 <NA>
#> 62                 <NA>
#> 63                 <NA>
#> 64                 <NA>
#> 65                 <NA>
#> 66                 <NA>
#> 67                 <NA>
#> 68                 <NA>
#> 69                 <NA>
#> 70                 <NA>
#> 71                 <NA>
#> 72                 <NA>
#> 73                 <NA>
#> 74                 <NA>
#> 75                 <NA>
#> 76                 <NA>
#> 77                 <NA>
#> 78                 <NA>
#> 79                 <NA>
#> 80                 <NA>
#> 81                 <NA>
#> 82                 <NA>
#> 83                 <NA>
#> 84                 <NA>
#> 85                 <NA>
#> 86                 <NA>
#> 87                 <NA>
#> 88                 <NA>
#> 89                 <NA>
#> 90                 <NA>
#> 91                 <NA>
#> 92                 <NA>
#> 93                 <NA>
#> 94                 <NA>
#> 95                 <NA>
#> 96                 <NA>
#> 97                 <NA>
#> 98                 <NA>
#> 99                 <NA>
#> 100                <NA>
#> 101 2025-03-13 16:11:07
#> 102 2025-03-13 16:11:07
#> 103 2025-03-13 16:11:07
#> 104 2025-03-13 16:11:07
#> 105 2025-03-13 16:11:06
#> 106 2025-03-13 16:11:06
#> 107 2025-03-13 16:11:06
#> 108 2025-03-13 16:11:04
#> 109 2025-03-13 16:11:03
#> 110 2025-03-13 16:10:51
#> 111 2025-03-13 16:10:51
#> 112 2025-03-13 16:11:07
#> 113 2025-03-13 16:10:51
#> 114 2025-03-13 16:10:51
#> 115 2025-03-13 16:11:06
#> 116 2025-03-13 16:11:05
#> 117 2025-03-13 16:11:01
#> 118 2025-03-13 16:11:07
#> 119 2025-03-13 16:11:05
#> 120 2025-03-13 16:11:05
#> 121 2025-03-13 16:11:03
#> 122 2025-03-13 16:10:53
#> 123 2025-03-13 16:10:46
#> 124 2025-03-13 16:11:05
#> 125 2025-03-13 16:10:46
#> 126 2025-03-13 16:10:52
#> 127 2025-03-13 16:11:07
#> 128 2025-03-13 16:11:05
#> 129 2025-03-13 16:10:58
#> 130 2025-03-13 16:10:53
#> 131 2025-03-13 16:10:52
#> 132 2025-03-13 16:10:56
#> 133 2025-03-13 16:10:37
#> 134 2025-03-13 16:11:02
#> 135 2025-03-13 16:10:42
#> 136 2025-03-13 16:10:45
#> 137 2025-03-13 16:10:38
#> 138 2025-03-13 16:10:44
#> 139 2025-03-13 16:10:50
#> 140 2025-03-13 16:11:03
#> 141 2025-03-13 16:10:32
#> 142 2025-03-13 16:11:01
#> 143 2025-03-13 16:09:48
#> 144 2025-03-13 16:09:24
#> 145 2025-03-13 16:10:44
#> 146 2025-03-13 16:09:54
#> 147 2025-03-13 16:10:26
#> 148 2025-03-13 16:11:04
#> 149 2025-03-13 16:10:44
#> 150 2025-03-13 16:09:10
#> 151 2025-03-13 16:10:58
#> 152 2025-03-13 16:09:30
#> 153 2025-03-13 16:05:39
#> 154 2025-03-13 16:06:07
#> 155 2025-03-13 16:10:58
#> 156 2025-03-13 16:07:42
#> 157 2025-03-13 16:10:46
#> 158 2025-03-13 16:10:09
#> 159 2025-03-13 16:10:39
#> 160 2025-03-13 16:10:09
#> 161 2025-03-13 16:10:58
#> 162 2025-03-13 16:09:02
#> 163 2025-03-13 16:10:51
#> 164 2025-03-13 16:11:06
#> 165 2025-03-13 16:11:07
#> 166 2025-03-13 16:09:27
#> 167 2025-03-13 16:08:50
#> 168 2025-03-13 16:10:39
#> 169 2025-03-13 16:09:03
#> 170 2025-03-13 16:06:33
#> 171 2025-03-13 16:09:33
#> 172 2025-03-13 16:09:35
#> 173 2025-03-13 16:09:09
#> 174 2025-03-13 16:05:01
#> 175 2025-03-13 16:05:01
#> 176 2025-03-13 16:10:11
#> 177 2025-03-13 16:08:56
#> 178 2025-03-13 15:43:30
#> 179 2025-03-13 16:05:01
#> 180 2025-03-13 16:07:37
#> 181 2025-03-13 16:11:01
#> 182 2025-03-13 16:05:38
#> 183 2025-03-13 14:53:51
#> 184 2025-03-13 16:11:07
#> 185 2025-03-13 15:36:08
#> 186 2025-03-13 16:11:07
#> 187 2025-03-13 16:05:38
#> 188 2025-03-13 16:04:42
#> 189 2025-03-13 16:11:02
#> 190 2025-03-13 16:11:03
#> 191 2025-03-13 16:11:01
#> 192 2025-03-13 16:11:00
#> 193 2025-03-13 16:10:33
#> 194 2025-03-13 16:10:34
#> 195 2025-03-13 16:10:36
#> 196 2025-03-13 16:10:35
#> 197 2025-03-13 16:10:35
#> 198 2025-03-13 16:09:46
#> 199 2025-03-13 16:00:34
#> 200 2025-03-13 16:05:39