Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 16:44:24 1.4068 1.4076
#> 2   2024-12-18 16:44:24 1.4068 1.4077
#> 3   2024-12-18 16:44:24 1.4068 1.4079
#> 4   2024-12-18 16:44:25 1.4068 1.4078
#> 5   2024-12-18 16:44:25 1.4068 1.4081
#> 6   2024-12-18 16:44:25 1.4065 1.4081
#> 7   2024-12-18 16:44:25 1.4067 1.4081
#> 8   2024-12-18 16:44:25 1.4067 1.4080
#> 9   2024-12-18 16:44:25 1.4068 1.4080
#> 10  2024-12-18 16:44:26 1.4069 1.4080
#> 11  2024-12-18 16:44:26 1.4070 1.4080
#> 12  2024-12-18 16:44:27 1.4070 1.4079
#> 13  2024-12-18 16:44:28 1.4070 1.4081
#> 14  2024-12-18 16:44:28 1.4070 1.4080
#> 15  2024-12-18 16:44:28 1.4070 1.4079
#> 16  2024-12-18 16:44:28 1.4071 1.4079
#> 17  2024-12-18 16:44:28 1.4072 1.4079
#> 18  2024-12-18 16:44:28 1.4072 1.4080
#> 19  2024-12-18 16:44:28 1.4071 1.4080
#> 20  2024-12-18 16:44:31 1.4072 1.4080
#> 21  2024-12-18 16:44:31 1.4073 1.4080
#> 22  2024-12-18 16:44:31 1.4073 1.4081
#> 23  2024-12-18 16:44:31 1.4073 1.4080
#> 24  2024-12-18 16:44:32 1.4073 1.4082
#> 25  2024-12-18 16:44:32 1.4073 1.4081
#> 26  2024-12-18 16:44:33 1.4074 1.4081
#> 27  2024-12-18 16:44:33 1.4075 1.4081
#> 28  2024-12-18 16:44:34 1.4075 1.4083
#> 29  2024-12-18 16:44:34 1.4075 1.4082
#> 30  2024-12-18 16:44:35 1.4076 1.4082
#> 31  2024-12-18 16:44:35 1.4077 1.4082
#> 32  2024-12-18 16:44:35 1.4077 1.4093
#> 33  2024-12-18 16:44:35 1.4077 1.4100
#> 34  2024-12-18 16:44:35 1.4077 1.4099
#> 35  2024-12-18 16:44:35 1.4077 1.4098
#> 36  2024-12-18 16:44:35 1.4078 1.4098
#> 37  2024-12-18 16:44:35 1.4078 1.4097
#> 38  2024-12-18 16:44:35 1.4078 1.4095
#> 39  2024-12-18 16:44:37 1.4078 1.4094
#> 40  2024-12-18 16:44:37 1.4078 1.4092
#> 41  2024-12-18 16:44:37 1.4079 1.4092
#> 42  2024-12-18 16:44:37 1.4080 1.4092
#> 43  2024-12-18 16:44:37 1.4080 1.4091
#> 44  2024-12-18 16:44:37 1.4080 1.4090
#> 45  2024-12-18 16:44:37 1.4080 1.4089
#> 46  2024-12-18 16:44:37 1.4080 1.4088
#> 47  2024-12-18 16:44:37 1.4080 1.4087
#> 48  2024-12-18 16:44:37 1.4080 1.4088
#> 49  2024-12-18 16:44:38 1.4080 1.4089
#> 50  2024-12-18 16:44:38 1.4080 1.4091
#> 51  2024-12-18 16:44:39 1.4080 1.4089
#> 52  2024-12-18 16:44:39 1.4081 1.4089
#> 53  2024-12-18 16:44:39 1.4082 1.4089
#> 54  2024-12-18 16:44:39 1.4082 1.4088
#> 55  2024-12-18 16:44:39 1.4082 1.4089
#> 56  2024-12-18 16:44:39 1.4082 1.4090
#> 57  2024-12-18 16:44:41 1.4078 1.4090
#> 58  2024-12-18 16:44:41 1.4083 1.4090
#> 59  2024-12-18 16:44:41 1.4079 1.4090
#> 60  2024-12-18 16:44:41 1.4078 1.4090
#> 61  2024-12-18 16:44:41 1.4078 1.4090
#> 62  2024-12-18 16:44:41 1.4079 1.4089
#> 63  2024-12-18 16:44:41 1.4080 1.4089
#> 64  2024-12-18 16:44:41 1.4081 1.4089
#> 65  2024-12-18 16:44:44 1.4082 1.4089
#> 66  2024-12-18 16:44:44 1.4083 1.4089
#> 67  2024-12-18 16:44:45 1.4084 1.4089
#> 68  2024-12-18 16:44:45 1.4085 1.4089
#> 69  2024-12-18 16:44:45 1.4082 1.4089
#> 70  2024-12-18 16:44:46 1.4085 1.4089
#> 71  2024-12-18 16:44:47 1.4086 1.4089
#> 72  2024-12-18 16:44:48 1.4086 1.4091
#> 73  2024-12-18 16:44:48 1.4086 1.4090
#> 74  2024-12-18 16:44:48 1.4087 1.4090
#> 75  2024-12-18 16:44:48 1.4086 1.4090
#> 76  2024-12-18 16:44:48 1.4087 1.4090
#> 77  2024-12-18 16:44:48 1.4086 1.4090
#> 78  2024-12-18 16:44:48 1.4087 1.4090
#> 79  2024-12-18 16:44:49 1.4087 1.4089
#> 80  2024-12-18 16:44:50 1.4087 1.4090
#> 81  2024-12-18 16:44:50 1.4088 1.4090
#> 82  2024-12-18 16:44:50 1.4088 1.4091
#> 83  2024-12-18 16:44:50 1.4088 1.4090
#> 84  2024-12-18 16:44:50 1.4088 1.4089
#> 85  2024-12-18 16:44:51 1.4088 1.4090
#> 86  2024-12-18 16:44:52 1.4085 1.4090
#> 87  2024-12-18 16:44:52 1.4086 1.4090
#> 88  2024-12-18 16:44:52 1.4087 1.4090
#> 89  2024-12-18 16:44:55 1.4084 1.4090
#> 90  2024-12-18 16:44:55 1.4088 1.4090
#> 91  2024-12-18 16:44:55 1.4085 1.4090
#> 92  2024-12-18 16:44:55 1.4085 1.4089
#> 93  2024-12-18 16:44:55 1.4085 1.4090
#> 94  2024-12-18 16:44:56 1.4086 1.4090
#> 95  2024-12-18 16:44:56 1.4087 1.4090
#> 96  2024-12-18 16:44:56 1.4084 1.4090
#> 97  2024-12-18 16:44:56 1.4087 1.4090
#> 98  2024-12-18 16:44:56 1.4088 1.4090
#> 99  2024-12-18 16:45:00 1.4089 1.4090
#> 100 2024-12-18 16:45:00 1.4085 1.4090
#> 101 2024-12-18 16:45:00 1.4087 1.4090
#> 102 2024-12-18 16:45:00 1.4087 1.4089
#> 103 2024-12-18 16:45:01 1.4087 1.4090
#> 104 2024-12-18 16:45:01 1.4087 1.4089
#> 105 2024-12-18 16:45:02 1.4082 1.4089
#> 106 2024-12-18 16:45:02 1.4084 1.4089
#> 107 2024-12-18 16:45:02 1.4085 1.4089
#> 108 2024-12-18 16:45:02 1.4075 1.4089
#> 109 2024-12-18 16:45:02 1.4076 1.4089
#> 110 2024-12-18 16:45:02 1.4080 1.4089
#> 111 2024-12-18 16:45:02 1.4081 1.4089
#> 112 2024-12-18 16:45:02 1.4081 1.4088
#> 113 2024-12-18 16:45:02 1.4081 1.4087
#> 114 2024-12-18 16:45:02 1.4075 1.4087
#> 115 2024-12-18 16:45:02 1.4077 1.4087
#> 116 2024-12-18 16:45:02 1.4077 1.4086
#> 117 2024-12-18 16:45:02 1.4077 1.4085
#> 118 2024-12-18 16:45:02 1.4077 1.4084
#> 119 2024-12-18 16:45:02 1.4077 1.4083
#> 120 2024-12-18 16:45:03 1.4077 1.4082
#> 121 2024-12-18 16:45:03 1.4077 1.4081
#> 122 2024-12-18 16:45:03 1.4077 1.4080
#> 123 2024-12-18 16:45:04 1.4078 1.4080
#> 124 2024-12-18 16:45:04 1.4079 1.4080
#> 125 2024-12-18 16:45:06 1.4078 1.4080
#> 126 2024-12-18 16:45:06 1.4078 1.4079
#> 127 2024-12-18 16:45:08 1.4078 1.4087
#> 128 2024-12-18 16:45:08 1.4078 1.4086
#> 129 2024-12-18 16:45:08 1.4079 1.4086
#> 130 2024-12-18 16:45:08 1.4079 1.4090
#> 131 2024-12-18 16:45:08 1.4079 1.4089
#> 132 2024-12-18 16:45:08 1.4079 1.4096
#> 133 2024-12-18 16:45:08 1.4079 1.4095
#> 134 2024-12-18 16:45:08 1.4080 1.4095
#> 135 2024-12-18 16:45:08 1.4080 1.4093
#> 136 2024-12-18 16:45:08 1.4080 1.4092
#> 137 2024-12-18 16:45:08 1.4080 1.4095
#> 138 2024-12-18 16:45:08 1.4080 1.4094
#> 139 2024-12-18 16:45:09 1.4080 1.4093
#> 140 2024-12-18 16:45:09 1.4080 1.4092
#> 141 2024-12-18 16:45:09 1.4080 1.4091
#> 142 2024-12-18 16:45:09 1.4080 1.4090
#> 143 2024-12-18 16:45:09 1.4080 1.4089
#> 144 2024-12-18 16:45:09 1.4080 1.4091
#> 145 2024-12-18 16:45:09 1.4080 1.4090
#> 146 2024-12-18 16:45:09 1.4081 1.4090
#> 147 2024-12-18 16:45:09 1.4082 1.4090
#> 148 2024-12-18 16:45:10 1.4082 1.4089
#> 149 2024-12-18 16:45:10 1.4083 1.4089
#> 150 2024-12-18 16:45:10 1.4084 1.4089
#> 151 2024-12-18 16:45:12 1.4084 1.4090
#> 152 2024-12-18 16:45:13 1.4085 1.4090
#> 153 2024-12-18 16:45:13 1.4086 1.4090
#> 154 2024-12-18 16:45:13 1.4087 1.4090
#> 155 2024-12-18 16:45:14 1.4082 1.4090
#> 156 2024-12-18 16:45:14 1.4083 1.4090
#> 157 2024-12-18 16:45:14 1.4087 1.4090
#> 158 2024-12-18 16:45:14 1.4088 1.4090
#> 159 2024-12-18 16:45:14 1.4089 1.4090
#> 160 2024-12-18 16:45:14 1.4082 1.4090
#> 161 2024-12-18 16:45:14 1.4083 1.4090
#> 162 2024-12-18 16:45:14 1.4084 1.4090
#> 163 2024-12-18 16:45:14 1.4085 1.4090
#> 164 2024-12-18 16:45:14 1.4085 1.4091
#> 165 2024-12-18 16:45:14 1.4085 1.4097
#> 166 2024-12-18 16:45:14 1.4085 1.4096
#> 167 2024-12-18 16:45:14 1.4085 1.4104
#> 168 2024-12-18 16:45:14 1.4085 1.4102
#> 169 2024-12-18 16:45:14 1.4086 1.4102
#> 170 2024-12-18 16:45:14 1.4086 1.4103
#> 171 2024-12-18 16:45:14 1.4086 1.4102
#> 172 2024-12-18 16:45:14 1.4086 1.4101
#> 173 2024-12-18 16:45:14 1.4086 1.4100
#> 174 2024-12-18 16:45:14 1.4084 1.4100
#> 175 2024-12-18 16:45:15 1.4086 1.4100
#> 176 2024-12-18 16:45:15 1.4087 1.4100
#> 177 2024-12-18 16:45:16 1.4088 1.4100
#> 178 2024-12-18 16:45:16 1.4089 1.4100
#> 179 2024-12-18 16:45:16 1.4090 1.4100
#> 180 2024-12-18 16:45:16 1.4091 1.4100
#> 181 2024-12-18 16:45:16 1.4092 1.4100
#> 182 2024-12-18 16:45:16 1.4093 1.4100
#> 183 2024-12-18 16:45:16 1.4094 1.4100
#> 184 2024-12-18 16:45:18 1.4095 1.4100
#> 185 2024-12-18 16:45:18 1.4096 1.4100
#> 186 2024-12-18 16:45:20 1.4096 1.4099
#> 187 2024-12-18 16:45:20 1.4091 1.4099
#> 188 2024-12-18 16:45:20 1.4091 1.4099
#> 189 2024-12-18 16:45:20 1.4095 1.4099
#> 190 2024-12-18 16:45:20 1.4096 1.4099
#> 191 2024-12-18 16:45:20 1.4095 1.4099
#> 192 2024-12-18 16:45:20 1.4093 1.4099
#> 193 2024-12-18 16:45:20 1.4094 1.4099
#> 194 2024-12-18 16:45:23 1.4095 1.4099
#> 195 2024-12-18 16:45:23 1.4096 1.4099
#> 196 2024-12-18 16:45:24 1.4097 1.4099
#> 197 2024-12-18 16:45:25 1.4093 1.4099
#> 198 2024-12-18 16:45:25 1.4095 1.4099
#> 199 2024-12-18 16:45:25 1.4093 1.4099
#> 200 2024-12-18 16:45:25 1.4091 1.4099
#> 201 2024-12-18 16:45:25 1.4084 1.4099
#> 202 2024-12-18 16:45:25 1.4090 1.4099
#> 203 2024-12-18 16:45:25 1.4091 1.4099
#> 204 2024-12-18 16:45:25 1.4091 1.4098
#> 205 2024-12-18 16:45:25 1.4091 1.4097
#> 206 2024-12-18 16:45:25 1.4085 1.4097
#> 207 2024-12-18 16:45:25 1.4084 1.4097
#> 208 2024-12-18 16:45:25 1.4086 1.4097
#> 209 2024-12-18 16:45:27 1.4084 1.4097
#> 210 2024-12-18 16:45:27 1.4087 1.4097
#> 211 2024-12-18 16:45:27 1.4088 1.4097
#> 212 2024-12-18 16:45:27 1.4088 1.4100
#> 213 2024-12-18 16:45:27 1.4088 1.4099
#> 214 2024-12-18 16:45:27 1.4089 1.4099
#> 215 2024-12-18 16:45:28 1.4090 1.4099
#> 216 2024-12-18 16:45:28 1.4091 1.4099
#> 217 2024-12-18 16:45:28 1.4092 1.4099
#> 218 2024-12-18 16:45:29 1.4092 1.4098
#> 219 2024-12-18 16:45:29 1.4092 1.4097
#> 220 2024-12-18 16:45:29 1.4091 1.4097
#> 221 2024-12-18 16:45:29 1.4088 1.4097
#> 222 2024-12-18 16:45:29 1.4089 1.4097
#> 223 2024-12-18 16:45:29 1.4090 1.4097
#> 224 2024-12-18 16:45:29 1.4090 1.4096
#> 225 2024-12-18 16:45:29 1.4090 1.4095
#> 226 2024-12-18 16:45:29 1.4089 1.4095
#> 227 2024-12-18 16:45:29 1.4089 1.4094
#> 228 2024-12-18 16:45:29 1.4089 1.4093
#> 229 2024-12-18 16:45:29 1.4089 1.4092
#> 230 2024-12-18 16:45:30 1.4089 1.4091
#> 231 2024-12-18 16:45:30 1.4089 1.4090
#> 232 2024-12-18 16:45:31 1.4086 1.4090
#> 233 2024-12-18 16:45:31 1.4087 1.4090
#> 234 2024-12-18 16:45:31 1.4087 1.4089
#> 235 2024-12-18 16:45:31 1.4088 1.4089
#> 236 2024-12-18 16:45:31 1.4087 1.4089
#> 237 2024-12-18 16:45:32 1.4084 1.4089
#> 238 2024-12-18 16:45:32 1.4085 1.4089
#> 239 2024-12-18 16:45:32 1.4086 1.4089
#> 240 2024-12-18 16:45:33 1.4086 1.4090
#> 241 2024-12-18 16:45:33 1.4086 1.4089
#> 242 2024-12-18 16:45:35 1.4087 1.4089
#> 243 2024-12-18 16:45:36 1.4086 1.4089
#> 244 2024-12-18 16:45:36 1.4087 1.4089
#> 245 2024-12-18 16:45:36 1.4088 1.4089
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 16:45:12 0.979130 0.979753
#> 2   2024-12-18 16:45:12 0.979129 0.979753
#> 3   2024-12-18 16:45:12 0.979129 0.979753
#> 4   2024-12-18 16:45:12 0.979131 0.979753
#> 5   2024-12-18 16:45:12 0.979131 0.979752
#> 6   2024-12-18 16:45:12 0.979131 0.979751
#> 7   2024-12-18 16:45:12 0.979131 0.979752
#> 8   2024-12-18 16:45:12 0.979131 0.979751
#> 9   2024-12-18 16:45:12 0.979131 0.979793
#> 10  2024-12-18 16:45:12 0.979132 0.979792
#> 11  2024-12-18 16:45:12 0.979132 0.979915
#> 12  2024-12-18 16:45:12 0.979133 0.979914
#> 13  2024-12-18 16:45:12 0.979133 0.979915
#> 14  2024-12-18 16:45:12 0.979133 0.979915
#> 15  2024-12-18 16:45:12 0.979133 0.979913
#> 16  2024-12-18 16:45:12 0.979133 0.979912
#> 17  2024-12-18 16:45:12 0.979186 0.979912
#> 18  2024-12-18 16:45:12 0.979186 0.979878
#> 19  2024-12-18 16:45:12 0.979187 0.979878
#> 20  2024-12-18 16:45:12 0.979187 0.979877
#> 21  2024-12-18 16:45:12 0.979188 0.979877
#> 22  2024-12-18 16:45:12 0.979241 0.979877
#> 23  2024-12-18 16:45:12 0.979242 0.979877
#> 24  2024-12-18 16:45:12 0.979243 0.979877
#> 25  2024-12-18 16:45:13 0.979296 0.979877
#> 26  2024-12-18 16:45:13 0.979297 0.979877
#> 27  2024-12-18 16:45:13 0.979298 0.979877
#> 28  2024-12-18 16:45:13 0.979298 0.979913
#> 29  2024-12-18 16:45:13 0.979299 0.979912
#> 30  2024-12-18 16:45:13 0.979300 0.979912
#> 31  2024-12-18 16:45:13 0.979301 0.979912
#> 32  2024-12-18 16:45:13 0.979302 0.979912
#> 33  2024-12-18 16:45:13 0.979339 0.979912
#> 34  2024-12-18 16:45:13 0.979340 0.979912
#> 35  2024-12-18 16:45:13 0.979345 0.979912
#> 36  2024-12-18 16:45:13 0.979346 0.979912
#> 37  2024-12-18 16:45:13 0.979351 0.979912
#> 38  2024-12-18 16:45:13 0.979352 0.979912
#> 39  2024-12-18 16:45:13 0.979353 0.979912
#> 40  2024-12-18 16:45:14 0.979358 0.979912
#> 41  2024-12-18 16:45:14 0.979359 0.979912
#> 42  2024-12-18 16:45:14 0.979359 0.979918
#> 43  2024-12-18 16:45:14 0.979359 0.979916
#> 44  2024-12-18 16:45:14 0.979360 0.979916
#> 45  2024-12-18 16:45:14 0.979361 0.979916
#> 46  2024-12-18 16:45:14 0.979366 0.979916
#> 47  2024-12-18 16:45:14 0.979367 0.979916
#> 48  2024-12-18 16:45:14 0.979361 0.979916
#> 49  2024-12-18 16:45:14 0.979361 0.979916
#> 50  2024-12-18 16:45:14 0.979361 0.979916
#> 51  2024-12-18 16:45:14 0.979363 0.979916
#> 52  2024-12-18 16:45:14 0.979364 0.979916
#> 53  2024-12-18 16:45:14 0.979365 0.979916
#> 54  2024-12-18 16:45:14 0.979366 0.979916
#> 55  2024-12-18 16:45:14 0.979367 0.979916
#> 56  2024-12-18 16:45:14 0.979368 0.979916
#> 57  2024-12-18 16:45:14 0.979368 0.980000
#> 58  2024-12-18 16:45:14 0.979473 0.980000
#> 59  2024-12-18 16:45:14 0.979474 0.980000
#> 60  2024-12-18 16:45:14 0.979474 0.979923
#> 61  2024-12-18 16:45:15 0.979527 0.979923
#> 62  2024-12-18 16:45:15 0.979528 0.979923
#> 63  2024-12-18 16:45:15 0.979529 0.979923
#> 64  2024-12-18 16:45:15 0.979529 0.980000
#> 65  2024-12-18 16:45:15 0.979533 0.980000
#> 66  2024-12-18 16:45:15 0.979534 0.980000
#> 67  2024-12-18 16:45:15 0.979535 0.980000
#> 68  2024-12-18 16:45:15 0.979536 0.980000
#> 69  2024-12-18 16:45:15 0.979537 0.980000
#> 70  2024-12-18 16:45:15 0.979538 0.980000
#> 71  2024-12-18 16:45:15 0.979539 0.980000
#> 72  2024-12-18 16:45:15 0.979540 0.980000
#> 73  2024-12-18 16:45:15 0.979541 0.980000
#> 74  2024-12-18 16:45:15 0.979542 0.980000
#> 75  2024-12-18 16:45:15 0.979543 0.980000
#> 76  2024-12-18 16:45:15 0.979542 0.980000
#> 77  2024-12-18 16:45:15 0.979542 0.980000
#> 78  2024-12-18 16:45:15 0.979544 0.980000
#> 79  2024-12-18 16:45:15 0.979545 0.980000
#> 80  2024-12-18 16:45:15 0.979546 0.980000
#> 81  2024-12-18 16:45:15 0.979545 0.980000
#> 82  2024-12-18 16:45:15 0.979548 0.980000
#> 83  2024-12-18 16:45:15 0.979549 0.980000
#> 84  2024-12-18 16:45:15 0.979550 0.980000
#> 85  2024-12-18 16:45:15 0.979551 0.980000
#> 86  2024-12-18 16:45:15 0.979552 0.980000
#> 87  2024-12-18 16:45:16 0.979553 0.980000
#> 88  2024-12-18 16:45:16 0.979554 0.980000
#> 89  2024-12-18 16:45:16 0.979555 0.980000
#> 90  2024-12-18 16:45:16 0.979556 0.980000
#> 91  2024-12-18 16:45:16 0.979557 0.980000
#> 92  2024-12-18 16:45:16 0.979558 0.980000
#> 93  2024-12-18 16:45:16 0.979559 0.980000
#> 94  2024-12-18 16:45:16 0.979560 0.980000
#> 95  2024-12-18 16:45:16 0.979561 0.980000
#> 96  2024-12-18 16:45:16 0.979562 0.980000
#> 97  2024-12-18 16:45:16 0.979563 0.980000
#> 98  2024-12-18 16:45:16 0.979564 0.980000
#> 99  2024-12-18 16:45:16 0.979999 0.980000
#> 100 2024-12-18 16:45:20 0.979798 0.980000
#> 101 2024-12-18 16:45:20 0.979792 0.980000
#> 102 2024-12-18 16:45:20 0.979780 0.980000
#> 103 2024-12-18 16:45:20 0.979793 0.980000
#> 104 2024-12-18 16:45:21 0.979757 0.980000
#> 105 2024-12-18 16:45:21 0.979758 0.980000
#> 106 2024-12-18 16:45:21 0.979759 0.980000
#> 107 2024-12-18 16:45:21 0.979760 0.980000
#> 108 2024-12-18 16:45:21 0.979761 0.980000
#> 109 2024-12-18 16:45:21 0.979762 0.980000
#> 110 2024-12-18 16:45:21 0.979763 0.980000
#> 111 2024-12-18 16:45:21 0.979764 0.980000
#> 112 2024-12-18 16:45:21 0.979765 0.980000
#> 113 2024-12-18 16:45:21 0.979766 0.980000
#> 114 2024-12-18 16:45:21 0.979767 0.980000
#> 115 2024-12-18 16:45:21 0.979768 0.980000
#> 116 2024-12-18 16:45:21 0.979769 0.980000
#> 117 2024-12-18 16:45:21 0.979770 0.980000
#> 118 2024-12-18 16:45:21 0.979771 0.980000
#> 119 2024-12-18 16:45:21 0.979772 0.980000
#> 120 2024-12-18 16:45:22 0.979771 0.980000
#> 121 2024-12-18 16:45:22 0.979772 0.980000
#> 122 2024-12-18 16:45:22 0.979771 0.980000
#> 123 2024-12-18 16:45:22 0.979772 0.980000
#> 124 2024-12-18 16:45:22 0.979773 0.980000
#> 125 2024-12-18 16:45:22 0.979774 0.980000
#> 126 2024-12-18 16:45:22 0.979775 0.980000
#> 127 2024-12-18 16:45:22 0.979776 0.980000
#> 128 2024-12-18 16:45:22 0.979777 0.980000
#> 129 2024-12-18 16:45:23 0.979778 0.980000
#> 130 2024-12-18 16:45:23 0.979779 0.980000
#> 131 2024-12-18 16:45:23 0.979780 0.980000
#> 132 2024-12-18 16:45:23 0.979781 0.980000
#> 133 2024-12-18 16:45:23 0.979782 0.980000
#> 134 2024-12-18 16:45:23 0.979783 0.980000
#> 135 2024-12-18 16:45:23 0.979784 0.980000
#> 136 2024-12-18 16:45:23 0.979781 0.980000
#> 137 2024-12-18 16:45:23 0.979783 0.980000
#> 138 2024-12-18 16:45:23 0.979784 0.980000
#> 139 2024-12-18 16:45:23 0.979785 0.980000
#> 140 2024-12-18 16:45:23 0.979786 0.980000
#> 141 2024-12-18 16:45:23 0.979787 0.980000
#> 142 2024-12-18 16:45:23 0.979788 0.980000
#> 143 2024-12-18 16:45:23 0.979789 0.980000
#> 144 2024-12-18 16:45:23 0.979790 0.980000
#> 145 2024-12-18 16:45:23 0.979791 0.980000
#> 146 2024-12-18 16:45:24 0.979792 0.980000
#> 147 2024-12-18 16:45:24 0.979793 0.980000
#> 148 2024-12-18 16:45:24 0.979791 0.980000
#> 149 2024-12-18 16:45:24 0.979794 0.980000
#> 150 2024-12-18 16:45:24 0.979795 0.980000
#> 151 2024-12-18 16:45:24 0.979796 0.980000
#> 152 2024-12-18 16:45:24 0.979797 0.980000
#> 153 2024-12-18 16:45:24 0.979798 0.980000
#> 154 2024-12-18 16:45:24 0.979799 0.980000
#> 155 2024-12-18 16:45:24 0.979797 0.980000
#> 156 2024-12-18 16:45:24 0.979796 0.980000
#> 157 2024-12-18 16:45:25 0.979794 0.980000
#> 158 2024-12-18 16:45:25 0.979795 0.980000
#> 159 2024-12-18 16:45:25 0.979796 0.980000
#> 160 2024-12-18 16:45:25 0.979795 0.980000
#> 161 2024-12-18 16:45:25 0.979796 0.980000
#> 162 2024-12-18 16:45:26 0.979797 0.980000
#> 163 2024-12-18 16:45:27 0.979798 0.980000
#> 164 2024-12-18 16:45:27 0.979799 0.980000
#> 165 2024-12-18 16:45:28 0.979800 0.980000
#> 166 2024-12-18 16:45:28 0.979801 0.980000
#> 167 2024-12-18 16:45:28 0.979802 0.980000
#> 168 2024-12-18 16:45:28 0.979803 0.980000
#> 169 2024-12-18 16:45:28 0.979804 0.980000
#> 170 2024-12-18 16:45:28 0.979805 0.980000
#> 171 2024-12-18 16:45:29 0.979806 0.980000
#> 172 2024-12-18 16:45:29 0.979807 0.980000
#> 173 2024-12-18 16:45:32 0.979806 0.980000
#> 174 2024-12-18 16:45:33 0.979807 0.980000
#> 175 2024-12-18 16:45:33 0.979808 0.980000
#> 176 2024-12-18 16:45:33 0.979809 0.980000
#> 177 2024-12-18 16:45:33 0.979810 0.980000
#> 178 2024-12-18 16:45:33 0.979811 0.980000
#> 179 2024-12-18 16:45:33 0.979812 0.980000
#> 180 2024-12-18 16:45:33 0.979813 0.980000
#> 181 2024-12-18 16:45:33 0.979814 0.980000
#> 182 2024-12-18 16:45:33 0.979815 0.980000
#> 183 2024-12-18 16:45:33 0.979816 0.980000
#> 184 2024-12-18 16:45:33 0.979817 0.980000
#> 185 2024-12-18 16:45:33 0.979818 0.980000
#> 186 2024-12-18 16:45:33 0.979819 0.980000
#> 187 2024-12-18 16:45:34 0.979820 0.980000
#> 188 2024-12-18 16:45:34 0.979819 0.980000
#> 189 2024-12-18 16:45:34 0.979820 0.980000
#> 190 2024-12-18 16:45:34 0.979819 0.980000
#> 191 2024-12-18 16:45:34 0.979775 0.980000
#> 192 2024-12-18 16:45:34 0.979775 0.979999
#> 193 2024-12-18 16:45:34 0.979775 0.979998
#> 194 2024-12-18 16:45:35 0.979775 0.979999
#> 195 2024-12-18 16:45:36 0.979775 0.979998
#> 196 2024-12-18 16:45:36 0.979775 0.979997
#> 197 2024-12-18 16:45:36 0.979775 0.979996
#> 198 2024-12-18 16:45:36 0.979775 0.979995
#> 199 2024-12-18 16:45:36 0.979775 0.979992
#> 200 2024-12-18 16:45:36 0.979775 0.979991
#> 201 2024-12-18 16:45:36 0.979775 0.979984
#> 202 2024-12-18 16:45:36 0.979775 0.979983
#> 203 2024-12-18 16:45:36 0.979775 0.979973
#> 204 2024-12-18 16:45:36 0.979775 0.979972
#> 205 2024-12-18 16:45:36 0.979775 0.979969
#> 206 2024-12-18 16:45:36 0.979775 0.979968
#> 207 2024-12-18 16:45:36 0.979775 0.979966
#> 208 2024-12-18 16:45:36 0.979775 0.979965
#> 209 2024-12-18 16:45:36 0.979775 0.979954
#> 210 2024-12-18 16:45:36 0.979775 0.979953
#> 211 2024-12-18 16:45:36 0.979775 0.979951
#> 212 2024-12-18 16:45:36 0.979775 0.979950
#> 213 2024-12-18 16:45:36 0.979667 0.979950
#> 
#> $last
#> [1] 1734540336
#>