Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 16:24:30 1.3885 1.3907
#> 2   2024-12-18 16:24:30 1.3885 1.3906
#> 3   2024-12-18 16:24:30 1.3885 1.3904
#> 4   2024-12-18 16:24:30 1.3885 1.3894
#> 5   2024-12-18 16:24:30 1.3886 1.3894
#> 6   2024-12-18 16:24:30 1.3887 1.3894
#> 7   2024-12-18 16:24:31 1.3887 1.3893
#> 8   2024-12-18 16:24:31 1.3888 1.3893
#> 9   2024-12-18 16:24:31 1.3889 1.3893
#> 10  2024-12-18 16:24:32 1.3890 1.3893
#> 11  2024-12-18 16:24:33 1.3890 1.3894
#> 12  2024-12-18 16:24:33 1.3890 1.3899
#> 13  2024-12-18 16:24:33 1.3890 1.3904
#> 14  2024-12-18 16:24:33 1.3890 1.3898
#> 15  2024-12-18 16:24:33 1.3890 1.3904
#> 16  2024-12-18 16:24:33 1.3891 1.3904
#> 17  2024-12-18 16:24:33 1.3891 1.3903
#> 18  2024-12-18 16:24:33 1.3891 1.3902
#> 19  2024-12-18 16:24:33 1.3892 1.3902
#> 20  2024-12-18 16:24:33 1.3893 1.3902
#> 21  2024-12-18 16:24:33 1.3894 1.3902
#> 22  2024-12-18 16:24:33 1.3895 1.3902
#> 23  2024-12-18 16:24:33 1.3895 1.3904
#> 24  2024-12-18 16:24:33 1.3895 1.3903
#> 25  2024-12-18 16:24:34 1.3896 1.3903
#> 26  2024-12-18 16:24:34 1.3897 1.3903
#> 27  2024-12-18 16:24:36 1.3898 1.3903
#> 28  2024-12-18 16:24:37 1.3899 1.3903
#> 29  2024-12-18 16:24:37 1.3900 1.3903
#> 30  2024-12-18 16:24:39 1.3900 1.3902
#> 31  2024-12-18 16:24:39 1.3900 1.3901
#> 32  2024-12-18 16:24:41 1.3899 1.3901
#> 33  2024-12-18 16:24:41 1.3896 1.3901
#> 34  2024-12-18 16:24:41 1.3897 1.3901
#> 35  2024-12-18 16:24:41 1.3897 1.3900
#> 36  2024-12-18 16:24:41 1.3897 1.3899
#> 37  2024-12-18 16:24:41 1.3897 1.3902
#> 38  2024-12-18 16:24:41 1.3897 1.3901
#> 39  2024-12-18 16:24:42 1.3897 1.3900
#> 40  2024-12-18 16:24:42 1.3897 1.3899
#> 41  2024-12-18 16:24:42 1.3898 1.3899
#> 42  2024-12-18 16:24:45 1.3897 1.3899
#> 43  2024-12-18 16:24:45 1.3897 1.3902
#> 44  2024-12-18 16:24:45 1.3897 1.3901
#> 45  2024-12-18 16:24:46 1.3898 1.3901
#> 46  2024-12-18 16:24:47 1.3899 1.3901
#> 47  2024-12-18 16:24:49 1.3900 1.3901
#> 48  2024-12-18 16:24:49 1.3900 1.3902
#> 49  2024-12-18 16:24:49 1.3901 1.3902
#> 50  2024-12-18 16:24:49 1.3901 1.3913
#> 51  2024-12-18 16:24:49 1.3902 1.3913
#> 52  2024-12-18 16:24:49 1.3902 1.3918
#> 53  2024-12-18 16:24:49 1.3902 1.3915
#> 54  2024-12-18 16:24:49 1.3902 1.3913
#> 55  2024-12-18 16:24:49 1.3900 1.3913
#> 56  2024-12-18 16:24:49 1.3901 1.3915
#> 57  2024-12-18 16:24:49 1.3901 1.3918
#> 58  2024-12-18 16:24:49 1.3902 1.3918
#> 59  2024-12-18 16:24:49 1.3903 1.3918
#> 60  2024-12-18 16:24:49 1.3903 1.3917
#> 61  2024-12-18 16:24:49 1.3904 1.3917
#> 62  2024-12-18 16:24:50 1.3904 1.3912
#> 63  2024-12-18 16:24:50 1.3905 1.3912
#> 64  2024-12-18 16:24:50 1.3906 1.3912
#> 65  2024-12-18 16:24:50 1.3906 1.3911
#> 66  2024-12-18 16:24:51 1.3907 1.3911
#> 67  2024-12-18 16:24:51 1.3908 1.3911
#> 68  2024-12-18 16:24:51 1.3909 1.3911
#> 69  2024-12-18 16:24:52 1.3909 1.3912
#> 70  2024-12-18 16:24:52 1.3909 1.3919
#> 71  2024-12-18 16:24:52 1.3909 1.3918
#> 72  2024-12-18 16:24:52 1.3909 1.3920
#> 73  2024-12-18 16:24:52 1.3909 1.3925
#> 74  2024-12-18 16:24:52 1.3910 1.3925
#> 75  2024-12-18 16:24:52 1.3910 1.3929
#> 76  2024-12-18 16:24:52 1.3910 1.3927
#> 77  2024-12-18 16:24:52 1.3910 1.3926
#> 78  2024-12-18 16:24:52 1.3910 1.3922
#> 79  2024-12-18 16:24:52 1.3911 1.3922
#> 80  2024-12-18 16:24:52 1.3912 1.3922
#> 81  2024-12-18 16:24:52 1.3913 1.3922
#> 82  2024-12-18 16:24:52 1.3913 1.3921
#> 83  2024-12-18 16:24:53 1.3914 1.3921
#> 84  2024-12-18 16:24:53 1.3915 1.3921
#> 85  2024-12-18 16:24:54 1.3915 1.3922
#> 86  2024-12-18 16:24:55 1.3916 1.3922
#> 87  2024-12-18 16:24:55 1.3917 1.3922
#> 88  2024-12-18 16:24:55 1.3917 1.3921
#> 89  2024-12-18 16:24:56 1.3917 1.3922
#> 90  2024-12-18 16:24:56 1.3917 1.3929
#> 91  2024-12-18 16:24:56 1.3917 1.3932
#> 92  2024-12-18 16:24:56 1.3917 1.3931
#> 93  2024-12-18 16:24:56 1.3917 1.3922
#> 94  2024-12-18 16:24:56 1.3917 1.3931
#> 95  2024-12-18 16:24:56 1.3918 1.3931
#> 96  2024-12-18 16:24:56 1.3919 1.3931
#> 97  2024-12-18 16:24:56 1.3916 1.3931
#> 98  2024-12-18 16:24:57 1.3917 1.3922
#> 99  2024-12-18 16:24:57 1.3917 1.3921
#> 100 2024-12-18 16:24:57 1.3918 1.3921
#> 101 2024-12-18 16:24:57 1.3919 1.3921
#> 102 2024-12-18 16:24:57 1.3918 1.3921
#> 103 2024-12-18 16:24:58 1.3919 1.3921
#> 104 2024-12-18 16:24:59 1.3918 1.3921
#> 105 2024-12-18 16:25:00 1.3912 1.3921
#> 106 2024-12-18 16:25:00 1.3913 1.3921
#> 107 2024-12-18 16:25:00 1.3914 1.3921
#> 108 2024-12-18 16:25:01 1.3914 1.3927
#> 109 2024-12-18 16:25:01 1.3914 1.3931
#> 110 2024-12-18 16:25:01 1.3914 1.3926
#> 111 2024-12-18 16:25:01 1.3914 1.3933
#> 112 2024-12-18 16:25:01 1.3914 1.3930
#> 113 2024-12-18 16:25:01 1.3914 1.3933
#> 114 2024-12-18 16:25:01 1.3914 1.3932
#> 115 2024-12-18 16:25:01 1.3915 1.3932
#> 116 2024-12-18 16:25:01 1.3915 1.3929
#> 117 2024-12-18 16:25:01 1.3915 1.3928
#> 118 2024-12-18 16:25:01 1.3916 1.3928
#> 119 2024-12-18 16:25:01 1.3916 1.3933
#> 120 2024-12-18 16:25:01 1.3916 1.3932
#> 121 2024-12-18 16:25:02 1.3916 1.3922
#> 122 2024-12-18 16:25:02 1.3917 1.3922
#> 123 2024-12-18 16:25:02 1.3918 1.3922
#> 124 2024-12-18 16:25:03 1.3917 1.3922
#> 125 2024-12-18 16:25:04 1.3917 1.3921
#> 126 2024-12-18 16:25:04 1.3915 1.3921
#> 127 2024-12-18 16:25:04 1.3911 1.3921
#> 128 2024-12-18 16:25:04 1.3916 1.3921
#> 129 2024-12-18 16:25:04 1.3912 1.3921
#> 130 2024-12-18 16:25:04 1.3910 1.3921
#> 131 2024-12-18 16:25:04 1.3912 1.3921
#> 132 2024-12-18 16:25:04 1.3913 1.3921
#> 133 2024-12-18 16:25:05 1.3914 1.3921
#> 134 2024-12-18 16:25:05 1.3914 1.3920
#> 135 2024-12-18 16:25:05 1.3907 1.3920
#> 136 2024-12-18 16:25:05 1.3911 1.3920
#> 137 2024-12-18 16:25:05 1.3908 1.3920
#> 138 2024-12-18 16:25:05 1.3907 1.3920
#> 139 2024-12-18 16:25:05 1.3907 1.3920
#> 140 2024-12-18 16:25:05 1.3908 1.3919
#> 141 2024-12-18 16:25:05 1.3909 1.3919
#> 142 2024-12-18 16:25:05 1.3910 1.3919
#> 143 2024-12-18 16:25:05 1.3910 1.3912
#> 144 2024-12-18 16:25:05 1.3909 1.3912
#> 145 2024-12-18 16:25:05 1.3909 1.3911
#> 146 2024-12-18 16:25:06 1.3909 1.3912
#> 147 2024-12-18 16:25:06 1.3909 1.3920
#> 148 2024-12-18 16:25:06 1.3910 1.3920
#> 149 2024-12-18 16:25:06 1.3910 1.3919
#> 150 2024-12-18 16:25:06 1.3903 1.3919
#> 151 2024-12-18 16:25:06 1.3904 1.3919
#> 152 2024-12-18 16:25:06 1.3903 1.3919
#> 153 2024-12-18 16:25:06 1.3903 1.3912
#> 154 2024-12-18 16:25:06 1.3904 1.3912
#> 155 2024-12-18 16:25:06 1.3904 1.3911
#> 156 2024-12-18 16:25:06 1.3907 1.3911
#> 157 2024-12-18 16:25:06 1.3905 1.3911
#> 158 2024-12-18 16:25:07 1.3906 1.3911
#> 159 2024-12-18 16:25:07 1.3907 1.3911
#> 160 2024-12-18 16:25:07 1.3907 1.3912
#> 161 2024-12-18 16:25:07 1.3907 1.3916
#> 162 2024-12-18 16:25:07 1.3907 1.3920
#> 163 2024-12-18 16:25:07 1.3907 1.3915
#> 164 2024-12-18 16:25:07 1.3907 1.3920
#> 165 2024-12-18 16:25:07 1.3907 1.3919
#> 166 2024-12-18 16:25:07 1.3907 1.3923
#> 167 2024-12-18 16:25:07 1.3907 1.3922
#> 168 2024-12-18 16:25:07 1.3908 1.3922
#> 169 2024-12-18 16:25:08 1.3906 1.3922
#> 170 2024-12-18 16:25:08 1.3907 1.3921
#> 171 2024-12-18 16:25:08 1.3907 1.3920
#> 172 2024-12-18 16:25:08 1.3907 1.3919
#> 173 2024-12-18 16:25:08 1.3907 1.3918
#> 174 2024-12-18 16:25:08 1.3907 1.3912
#> 175 2024-12-18 16:25:09 1.3907 1.3911
#> 176 2024-12-18 16:25:09 1.3908 1.3911
#> 177 2024-12-18 16:25:09 1.3909 1.3911
#> 178 2024-12-18 16:25:10 1.3909 1.3918
#> 179 2024-12-18 16:25:10 1.3909 1.3923
#> 180 2024-12-18 16:25:10 1.3910 1.3923
#> 181 2024-12-18 16:25:10 1.3911 1.3923
#> 182 2024-12-18 16:25:10 1.3911 1.3922
#> 183 2024-12-18 16:25:10 1.3912 1.3922
#> 184 2024-12-18 16:25:10 1.3913 1.3922
#> 185 2024-12-18 16:25:10 1.3914 1.3922
#> 186 2024-12-18 16:25:11 1.3915 1.3922
#> 187 2024-12-18 16:25:11 1.3916 1.3922
#> 188 2024-12-18 16:25:12 1.3916 1.3921
#> 189 2024-12-18 16:25:12 1.3915 1.3921
#> 190 2024-12-18 16:25:12 1.3915 1.3920
#> 191 2024-12-18 16:25:12 1.3915 1.3919
#> 192 2024-12-18 16:25:14 1.3914 1.3919
#> 193 2024-12-18 16:25:14 1.3908 1.3919
#> 194 2024-12-18 16:25:14 1.3913 1.3918
#> 195 2024-12-18 16:25:14 1.3914 1.3918
#> 196 2024-12-18 16:25:14 1.3913 1.3918
#> 197 2024-12-18 16:25:14 1.3914 1.3918
#> 198 2024-12-18 16:25:15 1.3914 1.3922
#> 199 2024-12-18 16:25:15 1.3914 1.3921
#> 200 2024-12-18 16:25:15 1.3915 1.3921
#> 201 2024-12-18 16:25:15 1.3916 1.3921
#> 202 2024-12-18 16:25:17 1.3917 1.3921
#> 203 2024-12-18 16:25:17 1.3917 1.3922
#> 204 2024-12-18 16:25:17 1.3917 1.3924
#> 205 2024-12-18 16:25:17 1.3918 1.3924
#> 206 2024-12-18 16:25:17 1.3918 1.3930
#> 207 2024-12-18 16:25:17 1.3918 1.3924
#> 208 2024-12-18 16:25:17 1.3918 1.3930
#> 209 2024-12-18 16:25:17 1.3918 1.3932
#> 210 2024-12-18 16:25:17 1.3918 1.3928
#> 211 2024-12-18 16:25:17 1.3918 1.3932
#> 212 2024-12-18 16:25:17 1.3919 1.3932
#> 213 2024-12-18 16:25:17 1.3920 1.3932
#> 214 2024-12-18 16:25:18 1.3921 1.3932
#> 215 2024-12-18 16:25:18 1.3922 1.3932
#> 216 2024-12-18 16:25:18 1.3922 1.3935
#> 217 2024-12-18 16:25:19 1.3922 1.3931
#> 218 2024-12-18 16:25:19 1.3922 1.3930
#> 219 2024-12-18 16:25:19 1.3923 1.3930
#> 220 2024-12-18 16:25:19 1.3924 1.3930
#> 221 2024-12-18 16:25:19 1.3925 1.3930
#> 222 2024-12-18 16:25:20 1.3925 1.3931
#> 223 2024-12-18 16:25:22 1.3926 1.3931
#> 224 2024-12-18 16:25:22 1.3927 1.3931
#> 225 2024-12-18 16:25:22 1.3928 1.3931
#> 226 2024-12-18 16:25:23 1.3928 1.3940
#> 227 2024-12-18 16:25:23 1.3928 1.3939
#> 228 2024-12-18 16:25:23 1.3928 1.3942
#> 229 2024-12-18 16:25:23 1.3928 1.3941
#> 230 2024-12-18 16:25:23 1.3929 1.3941
#> 231 2024-12-18 16:25:23 1.3930 1.3941
#> 232 2024-12-18 16:25:23 1.3930 1.3945
#> 233 2024-12-18 16:25:23 1.3930 1.3944
#> 234 2024-12-18 16:25:23 1.3930 1.3943
#> 235 2024-12-18 16:25:23 1.3930 1.3942
#> 236 2024-12-18 16:25:23 1.3931 1.3942
#> 237 2024-12-18 16:25:23 1.3932 1.3942
#> 238 2024-12-18 16:25:23 1.3933 1.3942
#> 239 2024-12-18 16:25:23 1.3933 1.3943
#> 240 2024-12-18 16:25:23 1.3933 1.3946
#> 241 2024-12-18 16:25:23 1.3933 1.3945
#> 242 2024-12-18 16:25:23 1.3933 1.3944
#> 243 2024-12-18 16:25:23 1.3933 1.3945
#> 244 2024-12-18 16:25:23 1.3933 1.3941
#> 245 2024-12-18 16:25:23 1.3933 1.3946
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 16:25:16 0.967760 0.968365
#> 2   2024-12-18 16:25:16 0.967761 0.968360
#> 3   2024-12-18 16:25:16 0.967762 0.968360
#> 4   2024-12-18 16:25:16 0.967762 0.968365
#> 5   2024-12-18 16:25:16 0.967765 0.968361
#> 6   2024-12-18 16:25:16 0.967765 0.968365
#> 7   2024-12-18 16:25:16 0.967765 0.968365
#> 8   2024-12-18 16:25:16 0.967766 0.968360
#> 9   2024-12-18 16:25:16 0.967766 0.968365
#> 10  2024-12-18 16:25:16 0.967805 0.968364
#> 11  2024-12-18 16:25:16 0.967806 0.968364
#> 12  2024-12-18 16:25:16 0.967805 0.968364
#> 13  2024-12-18 16:25:16 0.967807 0.968478
#> 14  2024-12-18 16:25:16 0.967805 0.968478
#> 15  2024-12-18 16:25:16 0.967805 0.968478
#> 16  2024-12-18 16:25:16 0.967808 0.968363
#> 17  2024-12-18 16:25:16 0.967808 0.968478
#> 18  2024-12-18 16:25:16 0.967805 0.968477
#> 19  2024-12-18 16:25:16 0.967805 0.968477
#> 20  2024-12-18 16:25:16 0.967805 0.968477
#> 21  2024-12-18 16:25:16 0.967805 0.968477
#> 22  2024-12-18 16:25:16 0.967805 0.968477
#> 23  2024-12-18 16:25:16 0.967805 0.968477
#> 24  2024-12-18 16:25:16 0.967805 0.968477
#> 25  2024-12-18 16:25:16 0.967806 0.968634
#> 26  2024-12-18 16:25:16 0.967806 0.968476
#> 27  2024-12-18 16:25:16 0.967806 0.968634
#> 28  2024-12-18 16:25:16 0.967806 0.968738
#> 29  2024-12-18 16:25:16 0.967806 0.968737
#> 30  2024-12-18 16:25:16 0.967806 0.968736
#> 31  2024-12-18 16:25:16 0.967806 0.968634
#> 32  2024-12-18 16:25:16 0.967806 0.968736
#> 33  2024-12-18 16:25:16 0.967807 0.968736
#> 34  2024-12-18 16:25:16 0.967808 0.968736
#> 35  2024-12-18 16:25:16 0.967809 0.968736
#> 36  2024-12-18 16:25:16 0.967810 0.968736
#> 37  2024-12-18 16:25:16 0.967811 0.968736
#> 38  2024-12-18 16:25:16 0.967812 0.968736
#> 39  2024-12-18 16:25:16 0.967813 0.968736
#> 40  2024-12-18 16:25:16 0.967814 0.968736
#> 41  2024-12-18 16:25:16 0.967805 0.968736
#> 42  2024-12-18 16:25:17 0.967980 0.968736
#> 43  2024-12-18 16:25:17 0.968036 0.968736
#> 44  2024-12-18 16:25:17 0.968037 0.968736
#> 45  2024-12-18 16:25:17 0.968038 0.968736
#> 46  2024-12-18 16:25:17 0.968039 0.968736
#> 47  2024-12-18 16:25:17 0.968040 0.968736
#> 48  2024-12-18 16:25:17 0.968041 0.968736
#> 49  2024-12-18 16:25:17 0.968038 0.968736
#> 50  2024-12-18 16:25:17 0.968040 0.968736
#> 51  2024-12-18 16:25:17 0.968040 0.968735
#> 52  2024-12-18 16:25:17 0.968096 0.968735
#> 53  2024-12-18 16:25:17 0.968096 0.968736
#> 54  2024-12-18 16:25:17 0.968096 0.969179
#> 55  2024-12-18 16:25:17 0.968097 0.968734
#> 56  2024-12-18 16:25:17 0.968097 0.968874
#> 57  2024-12-18 16:25:17 0.968097 0.968733
#> 58  2024-12-18 16:25:17 0.968097 0.968874
#> 59  2024-12-18 16:25:17 0.968097 0.968874
#> 60  2024-12-18 16:25:17 0.968097 0.968874
#> 61  2024-12-18 16:25:17 0.968097 0.968874
#> 62  2024-12-18 16:25:17 0.968097 0.968874
#> 63  2024-12-18 16:25:17 0.968097 0.968874
#> 64  2024-12-18 16:25:17 0.968097 0.968874
#> 65  2024-12-18 16:25:17 0.968097 0.968874
#> 66  2024-12-18 16:25:17 0.968097 0.968874
#> 67  2024-12-18 16:25:17 0.968097 0.968874
#> 68  2024-12-18 16:25:17 0.968097 0.968874
#> 69  2024-12-18 16:25:17 0.968098 0.968873
#> 70  2024-12-18 16:25:17 0.968099 0.968873
#> 71  2024-12-18 16:25:17 0.968100 0.968873
#> 72  2024-12-18 16:25:17 0.968101 0.968873
#> 73  2024-12-18 16:25:17 0.968157 0.968873
#> 74  2024-12-18 16:25:17 0.968158 0.968873
#> 75  2024-12-18 16:25:18 0.968159 0.968873
#> 76  2024-12-18 16:25:18 0.968160 0.968873
#> 77  2024-12-18 16:25:18 0.968161 0.968873
#> 78  2024-12-18 16:25:18 0.968162 0.968873
#> 79  2024-12-18 16:25:18 0.968163 0.968873
#> 80  2024-12-18 16:25:18 0.968164 0.968873
#> 81  2024-12-18 16:25:18 0.968165 0.968873
#> 82  2024-12-18 16:25:18 0.968166 0.968873
#> 83  2024-12-18 16:25:18 0.968157 0.968873
#> 84  2024-12-18 16:25:18 0.968158 0.968872
#> 85  2024-12-18 16:25:18 0.968159 0.968872
#> 86  2024-12-18 16:25:18 0.968215 0.968872
#> 87  2024-12-18 16:25:18 0.968216 0.968872
#> 88  2024-12-18 16:25:18 0.968217 0.968872
#> 89  2024-12-18 16:25:18 0.968273 0.968872
#> 90  2024-12-18 16:25:18 0.968274 0.968872
#> 91  2024-12-18 16:25:18 0.968275 0.968872
#> 92  2024-12-18 16:25:18 0.968279 0.968872
#> 93  2024-12-18 16:25:18 0.968331 0.968872
#> 94  2024-12-18 16:25:18 0.968332 0.968872
#> 95  2024-12-18 16:25:18 0.968333 0.968872
#> 96  2024-12-18 16:25:18 0.968334 0.968872
#> 97  2024-12-18 16:25:19 0.968335 0.968872
#> 98  2024-12-18 16:25:19 0.968336 0.968872
#> 99  2024-12-18 16:25:19 0.968337 0.968872
#> 100 2024-12-18 16:25:19 0.968337 0.968911
#> 101 2024-12-18 16:25:19 0.968337 0.968871
#> 102 2024-12-18 16:25:19 0.968338 0.968871
#> 103 2024-12-18 16:25:19 0.968338 0.968911
#> 104 2024-12-18 16:25:19 0.968338 0.968911
#> 105 2024-12-18 16:25:19 0.968338 0.968911
#> 106 2024-12-18 16:25:19 0.968338 0.968911
#> 107 2024-12-18 16:25:19 0.968338 0.968911
#> 108 2024-12-18 16:25:19 0.968338 0.968911
#> 109 2024-12-18 16:25:19 0.968339 0.968910
#> 110 2024-12-18 16:25:19 0.968340 0.968910
#> 111 2024-12-18 16:25:19 0.968339 0.968910
#> 112 2024-12-18 16:25:19 0.968339 0.968910
#> 113 2024-12-18 16:25:19 0.968340 0.968916
#> 114 2024-12-18 16:25:19 0.968340 0.968915
#> 115 2024-12-18 16:25:19 0.968337 0.968915
#> 116 2024-12-18 16:25:19 0.968339 0.968915
#> 117 2024-12-18 16:25:19 0.968340 0.968915
#> 118 2024-12-18 16:25:19 0.968345 0.968915
#> 119 2024-12-18 16:25:19 0.968346 0.968915
#> 120 2024-12-18 16:25:19 0.968346 0.968914
#> 121 2024-12-18 16:25:19 0.968347 0.968914
#> 122 2024-12-18 16:25:19 0.968347 0.968913
#> 123 2024-12-18 16:25:20 0.968348 0.968913
#> 124 2024-12-18 16:25:20 0.968349 0.968913
#> 125 2024-12-18 16:25:20 0.968349 0.968998
#> 126 2024-12-18 16:25:20 0.968349 0.968998
#> 127 2024-12-18 16:25:20 0.968349 0.968998
#> 128 2024-12-18 16:25:20 0.968436 0.968997
#> 129 2024-12-18 16:25:20 0.968437 0.968997
#> 130 2024-12-18 16:25:20 0.968438 0.968997
#> 131 2024-12-18 16:25:20 0.968439 0.968997
#> 132 2024-12-18 16:25:20 0.968439 0.969391
#> 133 2024-12-18 16:25:20 0.968439 0.969391
#> 134 2024-12-18 16:25:20 0.968439 0.969344
#> 135 2024-12-18 16:25:20 0.968439 0.969343
#> 136 2024-12-18 16:25:20 0.968495 0.969343
#> 137 2024-12-18 16:25:20 0.968496 0.969343
#> 138 2024-12-18 16:25:20 0.968497 0.969343
#> 139 2024-12-18 16:25:20 0.968498 0.969343
#> 140 2024-12-18 16:25:20 0.968499 0.969343
#> 141 2024-12-18 16:25:20 0.968500 0.969343
#> 142 2024-12-18 16:25:20 0.968501 0.969343
#> 143 2024-12-18 16:25:20 0.968502 0.969343
#> 144 2024-12-18 16:25:20 0.968503 0.969343
#> 145 2024-12-18 16:25:20 0.968495 0.969343
#> 146 2024-12-18 16:25:20 0.968495 0.969343
#> 147 2024-12-18 16:25:20 0.968495 0.969343
#> 148 2024-12-18 16:25:20 0.968497 0.969343
#> 149 2024-12-18 16:25:20 0.968498 0.969343
#> 150 2024-12-18 16:25:21 0.968498 0.969339
#> 151 2024-12-18 16:25:21 0.968495 0.969339
#> 152 2024-12-18 16:25:21 0.968496 0.969338
#> 153 2024-12-18 16:25:21 0.968496 0.969299
#> 154 2024-12-18 16:25:21 0.968496 0.969298
#> 155 2024-12-18 16:25:21 0.968496 0.969269
#> 156 2024-12-18 16:25:21 0.968496 0.969268
#> 157 2024-12-18 16:25:21 0.968496 0.969250
#> 158 2024-12-18 16:25:21 0.968496 0.969249
#> 159 2024-12-18 16:25:21 0.968496 0.969246
#> 160 2024-12-18 16:25:21 0.968496 0.969245
#> 161 2024-12-18 16:25:21 0.968501 0.969245
#> 162 2024-12-18 16:25:21 0.968501 0.969344
#> 163 2024-12-18 16:25:21 0.968502 0.969343
#> 164 2024-12-18 16:25:21 0.968502 0.969344
#> 165 2024-12-18 16:25:21 0.968502 0.969344
#> 166 2024-12-18 16:25:21 0.968503 0.969342
#> 167 2024-12-18 16:25:21 0.968503 0.969344
#> 168 2024-12-18 16:25:21 0.968504 0.969343
#> 169 2024-12-18 16:25:21 0.968505 0.969343
#> 170 2024-12-18 16:25:21 0.968607 0.969343
#> 171 2024-12-18 16:25:21 0.968608 0.969343
#> 172 2024-12-18 16:25:21 0.968609 0.969343
#> 173 2024-12-18 16:25:21 0.968609 0.969342
#> 174 2024-12-18 16:25:21 0.968609 0.969341
#> 175 2024-12-18 16:25:22 0.968610 0.969341
#> 176 2024-12-18 16:25:22 0.968611 0.969341
#> 177 2024-12-18 16:25:22 0.968611 0.969598
#> 178 2024-12-18 16:25:22 0.968611 0.969374
#> 179 2024-12-18 16:25:22 0.968611 0.969373
#> 180 2024-12-18 16:25:22 0.968612 0.969373
#> 181 2024-12-18 16:25:22 0.968613 0.969373
#> 182 2024-12-18 16:25:22 0.968613 0.969598
#> 183 2024-12-18 16:25:22 0.968613 0.969598
#> 184 2024-12-18 16:25:22 0.968613 0.969598
#> 185 2024-12-18 16:25:22 0.968652 0.969597
#> 186 2024-12-18 16:25:22 0.968653 0.969597
#> 187 2024-12-18 16:25:22 0.968654 0.969597
#> 188 2024-12-18 16:25:22 0.968655 0.969597
#> 189 2024-12-18 16:25:22 0.968656 0.969597
#> 190 2024-12-18 16:25:22 0.968656 0.969596
#> 191 2024-12-18 16:25:22 0.968656 0.969595
#> 192 2024-12-18 16:25:22 0.968657 0.969595
#> 193 2024-12-18 16:25:22 0.968658 0.969595
#> 194 2024-12-18 16:25:22 0.968659 0.969595
#> 195 2024-12-18 16:25:22 0.968660 0.969595
#> 196 2024-12-18 16:25:22 0.968661 0.969595
#> 197 2024-12-18 16:25:22 0.968662 0.969595
#> 198 2024-12-18 16:25:22 0.968667 0.969595
#> 199 2024-12-18 16:25:22 0.968988 0.969595
#> 200 2024-12-18 16:25:22 0.968989 0.969595
#> 201 2024-12-18 16:25:22 0.969045 0.969595
#> 202 2024-12-18 16:25:22 0.969046 0.969595
#> 203 2024-12-18 16:25:22 0.969047 0.969595
#> 204 2024-12-18 16:25:23 0.969047 0.969871
#> 205 2024-12-18 16:25:23 0.969047 0.969658
#> 206 2024-12-18 16:25:23 0.969047 0.969657
#> 207 2024-12-18 16:25:23 0.969047 0.969871
#> 208 2024-12-18 16:25:23 0.969047 0.969871
#> 209 2024-12-18 16:25:23 0.969047 0.969871
#> 210 2024-12-18 16:25:23 0.969047 0.969871
#> 211 2024-12-18 16:25:23 0.969047 0.969871
#> 212 2024-12-18 16:25:23 0.969047 0.969716
#> 213 2024-12-18 16:25:23 0.969047 0.969869
#> 214 2024-12-18 16:25:23 0.969048 0.969869
#> 215 2024-12-18 16:25:23 0.969048 0.969868
#> 216 2024-12-18 16:25:23 0.969048 0.969869
#> 217 2024-12-18 16:25:23 0.969048 0.969869
#> 218 2024-12-18 16:25:23 0.969049 0.969868
#> 219 2024-12-18 16:25:23 0.969049 0.969869
#> 220 2024-12-18 16:25:23 0.969049 0.969869
#> 221 2024-12-18 16:25:23 0.969049 0.969869
#> 222 2024-12-18 16:25:23 0.969049 0.969869
#> 223 2024-12-18 16:25:23 0.969050 0.969868
#> 224 2024-12-18 16:25:23 0.969051 0.969868
#> 225 2024-12-18 16:25:23 0.969107 0.969868
#> 226 2024-12-18 16:25:23 0.969108 0.969868
#> 227 2024-12-18 16:25:23 0.969046 0.969868
#> 228 2024-12-18 16:25:23 0.969047 0.970149
#> 229 2024-12-18 16:25:23 0.969047 0.970147
#> 230 2024-12-18 16:25:23 0.969047 0.970146
#> 231 2024-12-18 16:25:23 0.969048 0.970146
#> 232 2024-12-18 16:25:23 0.969049 0.970146
#> 233 2024-12-18 16:25:24 0.969049 0.970145
#> 234 2024-12-18 16:25:24 0.969049 0.970144
#> 235 2024-12-18 16:25:24 0.969049 0.970143
#> 236 2024-12-18 16:25:24 0.969049 0.970142
#> 237 2024-12-18 16:25:24 0.969049 0.970141
#> 238 2024-12-18 16:25:24 0.969049 0.970140
#> 239 2024-12-18 16:25:24 0.969049 0.970139
#> 240 2024-12-18 16:25:24 0.969050 0.970139
#> 241 2024-12-18 16:25:24 0.969050 0.970138
#> 242 2024-12-18 16:25:24 0.969051 0.970138
#> 243 2024-12-18 16:25:24 0.969051 0.970137
#> 244 2024-12-18 16:25:24 0.969051 0.970136
#> 245 2024-12-18 16:25:24 0.969051 0.970135
#> 246 2024-12-18 16:25:24 0.969052 0.970135
#> 247 2024-12-18 16:25:24 0.969053 0.970135
#> 
#> $last
#> [1] 1734539124
#>