Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 16:17:17 1.3891 1.3895
#> 2   2024-12-18 16:17:17 1.3891 1.3894
#> 3   2024-12-18 16:17:17 1.3891 1.3895
#> 4   2024-12-18 16:17:17 1.3891 1.3894
#> 5   2024-12-18 16:17:18 1.3892 1.3894
#> 6   2024-12-18 16:17:18 1.3892 1.3895
#> 7   2024-12-18 16:17:18 1.3892 1.3896
#> 8   2024-12-18 16:17:19 1.3892 1.3895
#> 9   2024-12-18 16:17:20 1.3889 1.3895
#> 10  2024-12-18 16:17:20 1.3890 1.3895
#> 11  2024-12-18 16:17:20 1.3890 1.3894
#> 12  2024-12-18 16:17:21 1.3891 1.3894
#> 13  2024-12-18 16:17:22 1.3891 1.3893
#> 14  2024-12-18 16:17:22 1.3889 1.3893
#> 15  2024-12-18 16:17:22 1.3887 1.3893
#> 16  2024-12-18 16:17:22 1.3888 1.3893
#> 17  2024-12-18 16:17:22 1.3884 1.3893
#> 18  2024-12-18 16:17:22 1.3884 1.3893
#> 19  2024-12-18 16:17:22 1.3885 1.3893
#> 20  2024-12-18 16:17:22 1.3886 1.3893
#> 21  2024-12-18 16:17:22 1.3886 1.3895
#> 22  2024-12-18 16:17:22 1.3886 1.3894
#> 23  2024-12-18 16:17:23 1.3887 1.3894
#> 24  2024-12-18 16:17:23 1.3888 1.3894
#> 25  2024-12-18 16:17:23 1.3885 1.3894
#> 26  2024-12-18 16:17:23 1.3886 1.3897
#> 27  2024-12-18 16:17:23 1.3887 1.3896
#> 28  2024-12-18 16:17:23 1.3888 1.3896
#> 29  2024-12-18 16:17:23 1.3886 1.3896
#> 30  2024-12-18 16:17:23 1.3887 1.3896
#> 31  2024-12-18 16:17:24 1.3887 1.3895
#> 32  2024-12-18 16:17:24 1.3887 1.3907
#> 33  2024-12-18 16:17:24 1.3888 1.3906
#> 34  2024-12-18 16:17:24 1.3888 1.3901
#> 35  2024-12-18 16:17:24 1.3888 1.3905
#> 36  2024-12-18 16:17:24 1.3889 1.3905
#> 37  2024-12-18 16:17:24 1.3889 1.3900
#> 38  2024-12-18 16:17:24 1.3889 1.3905
#> 39  2024-12-18 16:17:24 1.3889 1.3907
#> 40  2024-12-18 16:17:24 1.3890 1.3906
#> 41  2024-12-18 16:17:24 1.3890 1.3905
#> 42  2024-12-18 16:17:24 1.3891 1.3905
#> 43  2024-12-18 16:17:24 1.3891 1.3904
#> 44  2024-12-18 16:17:24 1.3891 1.3907
#> 45  2024-12-18 16:17:24 1.3891 1.3905
#> 46  2024-12-18 16:17:24 1.3891 1.3904
#> 47  2024-12-18 16:17:24 1.3888 1.3904
#> 48  2024-12-18 16:17:24 1.3890 1.3904
#> 49  2024-12-18 16:17:24 1.3890 1.3903
#> 50  2024-12-18 16:17:24 1.3890 1.3902
#> 51  2024-12-18 16:17:24 1.3890 1.3905
#> 52  2024-12-18 16:17:24 1.3891 1.3904
#> 53  2024-12-18 16:17:24 1.3892 1.3904
#> 54  2024-12-18 16:17:24 1.3889 1.3904
#> 55  2024-12-18 16:17:25 1.3890 1.3903
#> 56  2024-12-18 16:17:25 1.3890 1.3902
#> 57  2024-12-18 16:17:25 1.3890 1.3898
#> 58  2024-12-18 16:17:25 1.3890 1.3897
#> 59  2024-12-18 16:17:25 1.3890 1.3904
#> 60  2024-12-18 16:17:25 1.3891 1.3903
#> 61  2024-12-18 16:17:25 1.3891 1.3902
#> 62  2024-12-18 16:17:25 1.3891 1.3901
#> 63  2024-12-18 16:17:25 1.3891 1.3904
#> 64  2024-12-18 16:17:25 1.3891 1.3898
#> 65  2024-12-18 16:17:25 1.3891 1.3897
#> 66  2024-12-18 16:17:25 1.3891 1.3896
#> 67  2024-12-18 16:17:26 1.3891 1.3904
#> 68  2024-12-18 16:17:26 1.3891 1.3904
#> 69  2024-12-18 16:17:26 1.3891 1.3903
#> 70  2024-12-18 16:17:26 1.3892 1.3903
#> 71  2024-12-18 16:17:26 1.3892 1.3902
#> 72  2024-12-18 16:17:26 1.3892 1.3901
#> 73  2024-12-18 16:17:26 1.3893 1.3901
#> 74  2024-12-18 16:17:26 1.3893 1.3903
#> 75  2024-12-18 16:17:27 1.3893 1.3905
#> 76  2024-12-18 16:17:27 1.3893 1.3904
#> 77  2024-12-18 16:17:27 1.3893 1.3907
#> 78  2024-12-18 16:17:27 1.3893 1.3906
#> 79  2024-12-18 16:17:27 1.3894 1.3906
#> 80  2024-12-18 16:17:27 1.3895 1.3906
#> 81  2024-12-18 16:17:27 1.3892 1.3906
#> 82  2024-12-18 16:17:28 1.3894 1.3906
#> 83  2024-12-18 16:17:28 1.3895 1.3906
#> 84  2024-12-18 16:17:28 1.3892 1.3906
#> 85  2024-12-18 16:17:28 1.3893 1.3905
#> 86  2024-12-18 16:17:28 1.3893 1.3904
#> 87  2024-12-18 16:17:28 1.3893 1.3903
#> 88  2024-12-18 16:17:28 1.3893 1.3902
#> 89  2024-12-18 16:17:28 1.3889 1.3902
#> 90  2024-12-18 16:17:28 1.3891 1.3902
#> 91  2024-12-18 16:17:28 1.3891 1.3901
#> 92  2024-12-18 16:17:28 1.3891 1.3898
#> 93  2024-12-18 16:17:28 1.3891 1.3897
#> 94  2024-12-18 16:17:28 1.3891 1.3902
#> 95  2024-12-18 16:17:29 1.3891 1.3904
#> 96  2024-12-18 16:17:29 1.3891 1.3903
#> 97  2024-12-18 16:17:29 1.3892 1.3903
#> 98  2024-12-18 16:17:29 1.3893 1.3903
#> 99  2024-12-18 16:17:29 1.3890 1.3903
#> 100 2024-12-18 16:17:30 1.3891 1.3898
#> 101 2024-12-18 16:17:30 1.3891 1.3897
#> 102 2024-12-18 16:17:30 1.3891 1.3896
#> 103 2024-12-18 16:17:30 1.3891 1.3904
#> 104 2024-12-18 16:17:30 1.3889 1.3903
#> 105 2024-12-18 16:17:30 1.3892 1.3903
#> 106 2024-12-18 16:17:30 1.3892 1.3902
#> 107 2024-12-18 16:17:30 1.3893 1.3902
#> 108 2024-12-18 16:17:30 1.3893 1.3901
#> 109 2024-12-18 16:17:30 1.3889 1.3901
#> 110 2024-12-18 16:17:31 1.3891 1.3901
#> 111 2024-12-18 16:17:31 1.3891 1.3900
#> 112 2024-12-18 16:17:31 1.3892 1.3900
#> 113 2024-12-18 16:17:31 1.3892 1.3899
#> 114 2024-12-18 16:17:31 1.3892 1.3898
#> 115 2024-12-18 16:17:31 1.3892 1.3897
#> 116 2024-12-18 16:17:31 1.3892 1.3896
#> 117 2024-12-18 16:17:33 1.3892 1.3895
#> 118 2024-12-18 16:17:33 1.3892 1.3894
#> 119 2024-12-18 16:17:33 1.3891 1.3894
#> 120 2024-12-18 16:17:33 1.3888 1.3894
#> 121 2024-12-18 16:17:33 1.3888 1.3893
#> 122 2024-12-18 16:17:33 1.3889 1.3893
#> 123 2024-12-18 16:17:35 1.3887 1.3893
#> 124 2024-12-18 16:17:35 1.3887 1.3896
#> 125 2024-12-18 16:17:35 1.3888 1.3896
#> 126 2024-12-18 16:17:35 1.3888 1.3894
#> 127 2024-12-18 16:17:35 1.3889 1.3894
#> 128 2024-12-18 16:17:35 1.3890 1.3894
#> 129 2024-12-18 16:17:35 1.3890 1.3893
#> 130 2024-12-18 16:17:35 1.3890 1.3894
#> 131 2024-12-18 16:17:35 1.3890 1.3893
#> 132 2024-12-18 16:17:35 1.3890 1.3894
#> 133 2024-12-18 16:17:35 1.3890 1.3893
#> 134 2024-12-18 16:17:35 1.3890 1.3894
#> 135 2024-12-18 16:17:35 1.3890 1.3893
#> 136 2024-12-18 16:17:35 1.3890 1.3894
#> 137 2024-12-18 16:17:35 1.3890 1.3893
#> 138 2024-12-18 16:17:36 1.3887 1.3893
#> 139 2024-12-18 16:17:36 1.3888 1.3901
#> 140 2024-12-18 16:17:36 1.3888 1.3896
#> 141 2024-12-18 16:17:36 1.3888 1.3895
#> 142 2024-12-18 16:17:36 1.3888 1.3909
#> 143 2024-12-18 16:17:36 1.3888 1.3909
#> 144 2024-12-18 16:17:36 1.3888 1.3907
#> 145 2024-12-18 16:17:36 1.3888 1.3898
#> 146 2024-12-18 16:17:36 1.3888 1.3897
#> 147 2024-12-18 16:17:36 1.3888 1.3907
#> 148 2024-12-18 16:17:36 1.3888 1.3897
#> 149 2024-12-18 16:17:36 1.3888 1.3896
#> 150 2024-12-18 16:17:36 1.3888 1.3907
#> 151 2024-12-18 16:17:36 1.3888 1.3907
#> 152 2024-12-18 16:17:36 1.3888 1.3904
#> 153 2024-12-18 16:17:36 1.3888 1.3903
#> 154 2024-12-18 16:17:37 1.3888 1.3902
#> 155 2024-12-18 16:17:37 1.3888 1.3898
#> 156 2024-12-18 16:17:37 1.3888 1.3897
#> 157 2024-12-18 16:17:37 1.3888 1.3896
#> 158 2024-12-18 16:17:37 1.3888 1.3895
#> 159 2024-12-18 16:17:37 1.3884 1.3895
#> 160 2024-12-18 16:17:37 1.3889 1.3895
#> 161 2024-12-18 16:17:37 1.3889 1.3894
#> 162 2024-12-18 16:17:37 1.3890 1.3894
#> 163 2024-12-18 16:17:37 1.3890 1.3897
#> 164 2024-12-18 16:17:38 1.3891 1.3896
#> 165 2024-12-18 16:17:38 1.3891 1.3895
#> 166 2024-12-18 16:17:40 1.3891 1.3897
#> 167 2024-12-18 16:17:40 1.3891 1.3896
#> 168 2024-12-18 16:17:40 1.3892 1.3896
#> 169 2024-12-18 16:17:40 1.3891 1.3896
#> 170 2024-12-18 16:17:40 1.3890 1.3896
#> 171 2024-12-18 16:17:40 1.3892 1.3896
#> 172 2024-12-18 16:17:40 1.3890 1.3896
#> 173 2024-12-18 16:17:40 1.3891 1.3896
#> 174 2024-12-18 16:17:40 1.3892 1.3896
#> 175 2024-12-18 16:17:40 1.3893 1.3896
#> 176 2024-12-18 16:17:41 1.3893 1.3906
#> 177 2024-12-18 16:17:41 1.3893 1.3899
#> 178 2024-12-18 16:17:41 1.3893 1.3898
#> 179 2024-12-18 16:17:41 1.3893 1.3903
#> 180 2024-12-18 16:17:41 1.3894 1.3902
#> 181 2024-12-18 16:17:41 1.3894 1.3898
#> 182 2024-12-18 16:17:41 1.3894 1.3897
#> 183 2024-12-18 16:17:41 1.3894 1.3903
#> 184 2024-12-18 16:17:41 1.3893 1.3902
#> 185 2024-12-18 16:17:41 1.3894 1.3902
#> 186 2024-12-18 16:17:42 1.3894 1.3904
#> 187 2024-12-18 16:17:42 1.3884 1.3903
#> 188 2024-12-18 16:17:42 1.3885 1.3902
#> 189 2024-12-18 16:17:42 1.3895 1.3902
#> 190 2024-12-18 16:17:42 1.3895 1.3901
#> 191 2024-12-18 16:17:42 1.3893 1.3901
#> 192 2024-12-18 16:17:42 1.3887 1.3901
#> 193 2024-12-18 16:17:42 1.3894 1.3901
#> 194 2024-12-18 16:17:42 1.3887 1.3901
#> 195 2024-12-18 16:17:42 1.3889 1.3901
#> 196 2024-12-18 16:17:42 1.3890 1.3901
#> 197 2024-12-18 16:17:42 1.3891 1.3901
#> 198 2024-12-18 16:17:42 1.3891 1.3906
#> 199 2024-12-18 16:17:42 1.3892 1.3905
#> 200 2024-12-18 16:17:42 1.3893 1.3905
#> 201 2024-12-18 16:17:42 1.3893 1.3904
#> 202 2024-12-18 16:17:42 1.3893 1.3907
#> 203 2024-12-18 16:17:42 1.3893 1.3906
#> 204 2024-12-18 16:17:42 1.3893 1.3905
#> 205 2024-12-18 16:17:43 1.3893 1.3904
#> 206 2024-12-18 16:17:43 1.3893 1.3907
#> 207 2024-12-18 16:17:43 1.3893 1.3906
#> 208 2024-12-18 16:17:44 1.3894 1.3906
#> 209 2024-12-18 16:17:44 1.3895 1.3906
#> 210 2024-12-18 16:17:44 1.3895 1.3913
#> 211 2024-12-18 16:17:44 1.3896 1.3907
#> 212 2024-12-18 16:17:44 1.3896 1.3913
#> 213 2024-12-18 16:17:44 1.3897 1.3912
#> 214 2024-12-18 16:17:44 1.3897 1.3917
#> 215 2024-12-18 16:17:45 1.3898 1.3916
#> 216 2024-12-18 16:17:45 1.3899 1.3916
#> 217 2024-12-18 16:17:45 1.3900 1.3916
#> 218 2024-12-18 16:17:45 1.3900 1.3919
#> 219 2024-12-18 16:17:45 1.3900 1.3915
#> 220 2024-12-18 16:17:45 1.3900 1.3914
#> 221 2024-12-18 16:17:45 1.3901 1.3914
#> 222 2024-12-18 16:17:45 1.3901 1.3919
#> 223 2024-12-18 16:17:45 1.3901 1.3919
#> 224 2024-12-18 16:17:45 1.3902 1.3918
#> 225 2024-12-18 16:17:45 1.3903 1.3918
#> 226 2024-12-18 16:17:45 1.3903 1.3921
#> 227 2024-12-18 16:17:45 1.3904 1.3920
#> 228 2024-12-18 16:17:45 1.3905 1.3920
#> 229 2024-12-18 16:17:45 1.3902 1.3920
#> 230 2024-12-18 16:17:46 1.3903 1.3918
#> 231 2024-12-18 16:17:46 1.3903 1.3917
#> 232 2024-12-18 16:17:46 1.3903 1.3916
#> 233 2024-12-18 16:17:46 1.3904 1.3916
#> 234 2024-12-18 16:17:46 1.3904 1.3915
#> 235 2024-12-18 16:17:46 1.3905 1.3915
#> 236 2024-12-18 16:17:47 1.3905 1.3917
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 16:17:35 0.968368 0.969207
#> 2   2024-12-18 16:17:35 0.968368 0.969206
#> 3   2024-12-18 16:17:35 0.968368 0.969152
#> 4   2024-12-18 16:17:35 0.968368 0.969151
#> 5   2024-12-18 16:17:35 0.968368 0.969097
#> 6   2024-12-18 16:17:35 0.968368 0.969261
#> 7   2024-12-18 16:17:35 0.968368 0.969262
#> 8   2024-12-18 16:17:35 0.968368 0.969262
#> 9   2024-12-18 16:17:35 0.968368 0.969207
#> 10  2024-12-18 16:17:35 0.968368 0.969206
#> 11  2024-12-18 16:17:35 0.968368 0.969152
#> 12  2024-12-18 16:17:35 0.968368 0.969151
#> 13  2024-12-18 16:17:35 0.968368 0.969097
#> 14  2024-12-18 16:17:35 0.968368 0.969096
#> 15  2024-12-18 16:17:35 0.968368 0.969042
#> 16  2024-12-18 16:17:35 0.968368 0.969261
#> 17  2024-12-18 16:17:35 0.968368 0.969262
#> 18  2024-12-18 16:17:35 0.968368 0.969262
#> 19  2024-12-18 16:17:36 0.968368 0.969207
#> 20  2024-12-18 16:17:36 0.968368 0.969201
#> 21  2024-12-18 16:17:36 0.968368 0.969147
#> 22  2024-12-18 16:17:36 0.968369 0.969147
#> 23  2024-12-18 16:17:36 0.968369 0.969146
#> 24  2024-12-18 16:17:36 0.968369 0.969092
#> 25  2024-12-18 16:17:36 0.968369 0.969091
#> 26  2024-12-18 16:17:36 0.968369 0.969037
#> 27  2024-12-18 16:17:36 0.968369 0.969199
#> 28  2024-12-18 16:17:36 0.968369 0.969145
#> 29  2024-12-18 16:17:36 0.968369 0.969036
#> 30  2024-12-18 16:17:36 0.968369 0.968982
#> 31  2024-12-18 16:17:36 0.968369 0.969198
#> 32  2024-12-18 16:17:36 0.968369 0.969144
#> 33  2024-12-18 16:17:36 0.968369 0.968981
#> 34  2024-12-18 16:17:36 0.968369 0.968927
#> 35  2024-12-18 16:17:36 0.968299 0.968927
#> 36  2024-12-18 16:17:36 0.968368 0.969143
#> 37  2024-12-18 16:17:36 0.968368 0.969089
#> 38  2024-12-18 16:17:36 0.968299 0.969089
#> 39  2024-12-18 16:17:36 0.968300 0.968980
#> 40  2024-12-18 16:17:36 0.968300 0.969089
#> 41  2024-12-18 16:17:36 0.968300 0.969089
#> 42  2024-12-18 16:17:36 0.968300 0.969089
#> 43  2024-12-18 16:17:36 0.968300 0.969087
#> 44  2024-12-18 16:17:36 0.968300 0.968978
#> 45  2024-12-18 16:17:36 0.968300 0.969087
#> 46  2024-12-18 16:17:36 0.968300 0.969086
#> 47  2024-12-18 16:17:36 0.968300 0.968979
#> 48  2024-12-18 16:17:36 0.968082 0.968979
#> 49  2024-12-18 16:17:36 0.968083 0.968978
#> 50  2024-12-18 16:17:36 0.967959 0.968978
#> 51  2024-12-18 16:17:36 0.968032 0.968978
#> 52  2024-12-18 16:17:36 0.967973 0.968978
#> 53  2024-12-18 16:17:36 0.968033 0.968978
#> 54  2024-12-18 16:17:36 0.967973 0.968978
#> 55  2024-12-18 16:17:36 0.967975 0.968978
#> 56  2024-12-18 16:17:36 0.967975 0.968977
#> 57  2024-12-18 16:17:36 0.967976 0.968977
#> 58  2024-12-18 16:17:36 0.967976 0.968976
#> 59  2024-12-18 16:17:37 0.967976 0.968975
#> 60  2024-12-18 16:17:37 0.967976 0.968974
#> 61  2024-12-18 16:17:37 0.967976 0.968973
#> 62  2024-12-18 16:17:37 0.967976 0.968972
#> 63  2024-12-18 16:17:37 0.967976 0.968966
#> 64  2024-12-18 16:17:37 0.967976 0.968965
#> 65  2024-12-18 16:17:37 0.967976 0.968959
#> 66  2024-12-18 16:17:37 0.967976 0.968958
#> 67  2024-12-18 16:17:37 0.967976 0.968957
#> 68  2024-12-18 16:17:37 0.967976 0.968956
#> 69  2024-12-18 16:17:37 0.967976 0.968950
#> 70  2024-12-18 16:17:37 0.967976 0.968949
#> 71  2024-12-18 16:17:37 0.967976 0.968948
#> 72  2024-12-18 16:17:38 0.967977 0.968948
#> 73  2024-12-18 16:17:38 0.967978 0.968948
#> 74  2024-12-18 16:17:38 0.967979 0.968948
#> 75  2024-12-18 16:17:38 0.967979 0.968947
#> 76  2024-12-18 16:17:38 0.967980 0.968947
#> 77  2024-12-18 16:17:38 0.967980 0.968946
#> 78  2024-12-18 16:17:38 0.967980 0.968945
#> 79  2024-12-18 16:17:38 0.967980 0.968944
#> 80  2024-12-18 16:17:38 0.967981 0.968944
#> 81  2024-12-18 16:17:38 0.967982 0.968944
#> 82  2024-12-18 16:17:38 0.967982 0.968943
#> 83  2024-12-18 16:17:38 0.968041 0.968943
#> 84  2024-12-18 16:17:38 0.968041 0.968942
#> 85  2024-12-18 16:17:38 0.968042 0.968942
#> 86  2024-12-18 16:17:39 0.967981 0.968942
#> 87  2024-12-18 16:17:39 0.967982 0.968936
#> 88  2024-12-18 16:17:39 0.967982 0.968935
#> 89  2024-12-18 16:17:39 0.968051 0.968935
#> 90  2024-12-18 16:17:39 0.968051 0.968934
#> 91  2024-12-18 16:17:39 0.968052 0.968934
#> 92  2024-12-18 16:17:40 0.968052 0.968933
#> 93  2024-12-18 16:17:40 0.968052 0.968928
#> 94  2024-12-18 16:17:40 0.968052 0.968927
#> 95  2024-12-18 16:17:40 0.968052 0.968921
#> 96  2024-12-18 16:17:40 0.968052 0.968920
#> 97  2024-12-18 16:17:40 0.968052 0.968919
#> 98  2024-12-18 16:17:40 0.968052 0.968918
#> 99  2024-12-18 16:17:40 0.968053 0.968918
#> 100 2024-12-18 16:17:40 0.968053 0.968917
#> 101 2024-12-18 16:17:40 0.968054 0.968917
#> 102 2024-12-18 16:17:40 0.968051 0.968917
#> 103 2024-12-18 16:17:41 0.968052 0.968916
#> 104 2024-12-18 16:17:41 0.968053 0.968916
#> 105 2024-12-18 16:17:41 0.968053 0.968915
#> 106 2024-12-18 16:17:41 0.968054 0.968915
#> 107 2024-12-18 16:17:41 0.968059 0.968915
#> 108 2024-12-18 16:17:41 0.968060 0.968915
#> 109 2024-12-18 16:17:41 0.968408 0.968915
#> 110 2024-12-18 16:17:41 0.968409 0.968915
#> 111 2024-12-18 16:17:42 0.968410 0.968915
#> 112 2024-12-18 16:17:42 0.968410 0.968918
#> 113 2024-12-18 16:17:42 0.968410 0.969031
#> 114 2024-12-18 16:17:42 0.968410 0.969031
#> 115 2024-12-18 16:17:42 0.968410 0.968992
#> 116 2024-12-18 16:17:42 0.968410 0.968991
#> 117 2024-12-18 16:17:42 0.968410 0.969031
#> 118 2024-12-18 16:17:42 0.968410 0.968990
#> 119 2024-12-18 16:17:42 0.968410 0.969031
#> 120 2024-12-18 16:17:42 0.968410 0.969030
#> 121 2024-12-18 16:17:42 0.968411 0.969030
#> 122 2024-12-18 16:17:42 0.968412 0.969030
#> 123 2024-12-18 16:17:42 0.968412 0.969029
#> 124 2024-12-18 16:17:42 0.968412 0.969031
#> 125 2024-12-18 16:17:42 0.968417 0.969031
#> 126 2024-12-18 16:17:42 0.968418 0.969031
#> 127 2024-12-18 16:17:42 0.968644 0.969031
#> 128 2024-12-18 16:17:42 0.968645 0.969031
#> 129 2024-12-18 16:17:42 0.968793 0.969031
#> 130 2024-12-18 16:17:43 0.968793 0.969245
#> 131 2024-12-18 16:17:43 0.968793 0.969244
#> 132 2024-12-18 16:17:43 0.968793 0.969243
#> 133 2024-12-18 16:17:43 0.968793 0.969424
#> 134 2024-12-18 16:17:43 0.968793 0.969424
#> 135 2024-12-18 16:17:43 0.968793 0.969738
#> 136 2024-12-18 16:17:43 0.968793 0.969738
#> 137 2024-12-18 16:17:43 0.968793 0.969422
#> 138 2024-12-18 16:17:43 0.968793 0.969422
#> 139 2024-12-18 16:17:43 0.968794 0.969240
#> 140 2024-12-18 16:17:43 0.968794 0.969422
#> 141 2024-12-18 16:17:43 0.968794 0.969641
#> 142 2024-12-18 16:17:43 0.968794 0.969641
#> 143 2024-12-18 16:17:43 0.968793 0.969420
#> 144 2024-12-18 16:17:43 0.968794 0.969420
#> 145 2024-12-18 16:17:43 0.968793 0.969420
#> 146 2024-12-18 16:17:43 0.968795 0.969420
#> 147 2024-12-18 16:17:43 0.968795 0.969641
#> 148 2024-12-18 16:17:43 0.968795 0.969504
#> 149 2024-12-18 16:17:43 0.968796 0.969419
#> 150 2024-12-18 16:17:43 0.968796 0.969504
#> 151 2024-12-18 16:17:43 0.968796 0.969504
#> 152 2024-12-18 16:17:43 0.968797 0.969418
#> 153 2024-12-18 16:17:43 0.968797 0.969504
#> 154 2024-12-18 16:17:43 0.968797 0.969504
#> 155 2024-12-18 16:17:43 0.968797 0.969504
#> 156 2024-12-18 16:17:43 0.968797 0.969504
#> 157 2024-12-18 16:17:43 0.968797 0.969504
#> 158 2024-12-18 16:17:44 0.968797 0.969563
#> 159 2024-12-18 16:17:44 0.968797 0.969640
#> 160 2024-12-18 16:17:44 0.968797 0.969640
#> 161 2024-12-18 16:17:44 0.968797 0.969737
#> 162 2024-12-18 16:17:44 0.968797 0.969737
#> 163 2024-12-18 16:17:44 0.968797 0.969560
#> 164 2024-12-18 16:17:44 0.969026 0.969560
#> 165 2024-12-18 16:17:44 0.969026 0.969559
#> 166 2024-12-18 16:17:44 0.969027 0.969559
#> 167 2024-12-18 16:17:44 0.969027 0.969560
#> 168 2024-12-18 16:17:44 0.969027 0.969558
#> 169 2024-12-18 16:17:44 0.969027 0.969560
#> 170 2024-12-18 16:17:44 0.969027 0.969559
#> 171 2024-12-18 16:17:44 0.969027 0.969558
#> 172 2024-12-18 16:17:44 0.969027 0.969557
#> 173 2024-12-18 16:17:44 0.969028 0.969557
#> 174 2024-12-18 16:17:44 0.969028 0.969556
#> 175 2024-12-18 16:17:44 0.969028 0.969555
#> 176 2024-12-18 16:17:45 0.969029 0.969555
#> 177 2024-12-18 16:17:45 0.969030 0.969555
#> 178 2024-12-18 16:17:45 0.969027 0.969555
#> 179 2024-12-18 16:17:45 0.969027 0.969555
#> 180 2024-12-18 16:17:45 0.969027 0.969555
#> 181 2024-12-18 16:17:45 0.969027 0.969555
#> 182 2024-12-18 16:17:45 0.968796 0.969555
#> 183 2024-12-18 16:17:45 0.968796 0.969555
#> 184 2024-12-18 16:17:45 0.968798 0.969555
#> 185 2024-12-18 16:17:45 0.968799 0.969555
#> 186 2024-12-18 16:17:45 0.968796 0.969555
#> 187 2024-12-18 16:17:45 0.968798 0.969555
#> 188 2024-12-18 16:17:45 0.968799 0.969555
#> 189 2024-12-18 16:17:45 0.968798 0.969555
#> 190 2024-12-18 16:17:45 0.968798 0.969555
#> 191 2024-12-18 16:17:45 0.968799 0.969501
#> 192 2024-12-18 16:17:45 0.968799 0.969500
#> 193 2024-12-18 16:17:45 0.968835 0.969500
#> 194 2024-12-18 16:17:45 0.968836 0.969500
#> 195 2024-12-18 16:17:46 0.968837 0.969500
#> 196 2024-12-18 16:17:46 0.968837 0.969558
#> 197 2024-12-18 16:17:46 0.969056 0.969557
#> 198 2024-12-18 16:17:46 0.969057 0.969557
#> 199 2024-12-18 16:17:46 0.969057 0.969556
#> 200 2024-12-18 16:17:46 0.969058 0.969556
#> 201 2024-12-18 16:17:46 0.969058 0.969555
#> 202 2024-12-18 16:17:46 0.969058 0.969560
#> 203 2024-12-18 16:17:46 0.969058 0.969559
#> 204 2024-12-18 16:17:46 0.969058 0.969560
#> 205 2024-12-18 16:17:47 0.969112 0.969560
#> 206 2024-12-18 16:17:47 0.969113 0.969560
#> 207 2024-12-18 16:17:47 0.969237 0.969560
#> 208 2024-12-18 16:17:47 0.969238 0.969560
#> 209 2024-12-18 16:17:47 0.969238 0.969686
#> 210 2024-12-18 16:17:47 0.969238 0.969737
#> 211 2024-12-18 16:17:47 0.969238 0.969737
#> 212 2024-12-18 16:17:47 0.969238 0.969738
#> 213 2024-12-18 16:17:47 0.969239 0.969735
#> 214 2024-12-18 16:17:47 0.969239 0.969738
#> 215 2024-12-18 16:17:47 0.969239 0.969738
#> 216 2024-12-18 16:17:47 0.969240 0.969734
#> 217 2024-12-18 16:17:47 0.969240 0.969738
#> 218 2024-12-18 16:17:47 0.969240 0.969738
#> 219 2024-12-18 16:17:47 0.969240 0.969738
#> 220 2024-12-18 16:17:47 0.969428 0.969682
#> 221 2024-12-18 16:17:47 0.969428 0.969738
#> 222 2024-12-18 16:17:47 0.969428 0.969682
#> 223 2024-12-18 16:17:47 0.969428 0.969738
#> 224 2024-12-18 16:17:47 0.969428 0.969738
#> 225 2024-12-18 16:17:47 0.969429 0.969738
#> 226 2024-12-18 16:17:47 0.969428 0.969738
#> 227 2024-12-18 16:17:47 0.969429 0.969738
#> 228 2024-12-18 16:17:47 0.969428 0.969738
#> 229 2024-12-18 16:17:47 0.969428 0.969738
#> 230 2024-12-18 16:17:47 0.969428 0.969738
#> 231 2024-12-18 16:17:47 0.969428 0.969738
#> 
#> $last
#> [1] 1734538667
#>