Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 00:06:06 1.4474 1.4504
#> 2   2024-12-18 00:06:06 1.4482 1.4504
#> 3   2024-12-18 00:06:06 1.4479 1.4504
#> 4   2024-12-18 00:06:06 1.4473 1.4504
#> 5   2024-12-18 00:06:06 1.4474 1.4504
#> 6   2024-12-18 00:06:10 1.4474 1.4502
#> 7   2024-12-18 00:06:10 1.4474 1.4504
#> 8   2024-12-18 00:06:10 1.4474 1.4497
#> 9   2024-12-18 00:06:10 1.4474 1.4496
#> 10  2024-12-18 00:06:11 1.4474 1.4495
#> 11  2024-12-18 00:06:12 1.4470 1.4495
#> 12  2024-12-18 00:06:12 1.4470 1.4494
#> 13  2024-12-18 00:06:12 1.4471 1.4494
#> 14  2024-12-18 00:06:12 1.4471 1.4493
#> 15  2024-12-18 00:06:14 1.4472 1.4493
#> 16  2024-12-18 00:06:14 1.4472 1.4494
#> 17  2024-12-18 00:06:14 1.4463 1.4494
#> 18  2024-12-18 00:06:14 1.4464 1.4494
#> 19  2024-12-18 00:06:14 1.4464 1.4492
#> 20  2024-12-18 00:06:14 1.4465 1.4492
#> 21  2024-12-18 00:06:14 1.4465 1.4491
#> 22  2024-12-18 00:06:14 1.4466 1.4491
#> 23  2024-12-18 00:06:14 1.4467 1.4491
#> 24  2024-12-18 00:06:14 1.4468 1.4491
#> 25  2024-12-18 00:06:15 1.4469 1.4491
#> 26  2024-12-18 00:06:15 1.4470 1.4491
#> 27  2024-12-18 00:06:15 1.4470 1.4490
#> 28  2024-12-18 00:06:15 1.4470 1.4495
#> 29  2024-12-18 00:06:15 1.4470 1.4494
#> 30  2024-12-18 00:06:17 1.4471 1.4494
#> 31  2024-12-18 00:06:19 1.4472 1.4494
#> 32  2024-12-18 00:06:19 1.4473 1.4494
#> 33  2024-12-18 00:06:21 1.4474 1.4494
#> 34  2024-12-18 00:06:21 1.4475 1.4494
#> 35  2024-12-18 00:06:21 1.4476 1.4494
#> 36  2024-12-18 00:06:21 1.4477 1.4494
#> 37  2024-12-18 00:06:21 1.4477 1.4495
#> 38  2024-12-18 00:06:21 1.4478 1.4495
#> 39  2024-12-18 00:06:21 1.4479 1.4495
#> 40  2024-12-18 00:06:21 1.4480 1.4495
#> 41  2024-12-18 00:06:21 1.4481 1.4495
#> 42  2024-12-18 00:06:21 1.4482 1.4495
#> 43  2024-12-18 00:06:21 1.4483 1.4495
#> 44  2024-12-18 00:06:23 1.4484 1.4495
#> 45  2024-12-18 00:06:23 1.4485 1.4495
#> 46  2024-12-18 00:06:24 1.4485 1.4512
#> 47  2024-12-18 00:06:24 1.4485 1.4511
#> 48  2024-12-18 00:06:24 1.4489 1.4511
#> 49  2024-12-18 00:06:24 1.4492 1.4511
#> 50  2024-12-18 00:06:24 1.4492 1.4518
#> 51  2024-12-18 00:06:24 1.4492 1.4517
#> 52  2024-12-18 00:06:24 1.4492 1.4515
#> 53  2024-12-18 00:06:24 1.4492 1.4511
#> 54  2024-12-18 00:06:24 1.4492 1.4510
#> 55  2024-12-18 00:06:24 1.4492 1.4511
#> 56  2024-12-18 00:06:24 1.4492 1.4518
#> 57  2024-12-18 00:06:24 1.4492 1.4517
#> 58  2024-12-18 00:06:24 1.4492 1.4518
#> 59  2024-12-18 00:06:24 1.4494 1.4518
#> 60  2024-12-18 00:06:25 1.4495 1.4518
#> 61  2024-12-18 00:06:25 1.4499 1.4518
#> 62  2024-12-18 00:06:25 1.4500 1.4518
#> 63  2024-12-18 00:06:26 1.4501 1.4518
#> 64  2024-12-18 00:06:27 1.4502 1.4518
#> 65  2024-12-18 00:06:28 1.4500 1.4518
#> 66  2024-12-18 00:06:28 1.4501 1.4518
#> 67  2024-12-18 00:06:28 1.4503 1.4518
#> 68  2024-12-18 00:06:28 1.4504 1.4518
#> 69  2024-12-18 00:06:28 1.4504 1.4517
#> 70  2024-12-18 00:06:28 1.4503 1.4517
#> 71  2024-12-18 00:06:28 1.4503 1.4516
#> 72  2024-12-18 00:06:28 1.4501 1.4516
#> 73  2024-12-18 00:06:29 1.4498 1.4516
#> 74  2024-12-18 00:06:29 1.4490 1.4516
#> 75  2024-12-18 00:06:29 1.4491 1.4516
#> 76  2024-12-18 00:06:29 1.4491 1.4515
#> 77  2024-12-18 00:06:29 1.4491 1.4514
#> 78  2024-12-18 00:06:29 1.4491 1.4513
#> 79  2024-12-18 00:06:29 1.4491 1.4512
#> 80  2024-12-18 00:06:30 1.4482 1.4512
#> 81  2024-12-18 00:06:30 1.4483 1.4512
#> 82  2024-12-18 00:06:30 1.4483 1.4511
#> 83  2024-12-18 00:06:30 1.4484 1.4511
#> 84  2024-12-18 00:06:30 1.4484 1.4510
#> 85  2024-12-18 00:06:31 1.4485 1.4510
#> 86  2024-12-18 00:06:31 1.4483 1.4510
#> 87  2024-12-18 00:06:31 1.4483 1.4509
#> 88  2024-12-18 00:06:31 1.4486 1.4509
#> 89  2024-12-18 00:06:31 1.4486 1.4508
#> 90  2024-12-18 00:06:32 1.4487 1.4508
#> 91  2024-12-18 00:06:32 1.4487 1.4511
#> 92  2024-12-18 00:06:32 1.4487 1.4510
#> 93  2024-12-18 00:06:32 1.4487 1.4507
#> 94  2024-12-18 00:06:32 1.4488 1.4507
#> 95  2024-12-18 00:06:32 1.4488 1.4506
#> 96  2024-12-18 00:06:32 1.4488 1.4511
#> 97  2024-12-18 00:06:32 1.4488 1.4510
#> 98  2024-12-18 00:06:32 1.4488 1.4513
#> 99  2024-12-18 00:06:32 1.4488 1.4512
#> 100 2024-12-18 00:06:32 1.4488 1.4510
#> 101 2024-12-18 00:06:32 1.4488 1.4509
#> 102 2024-12-18 00:06:33 1.4489 1.4509
#> 103 2024-12-18 00:06:33 1.4488 1.4509
#> 104 2024-12-18 00:06:33 1.4486 1.4509
#> 105 2024-12-18 00:06:33 1.4481 1.4509
#> 106 2024-12-18 00:06:33 1.4481 1.4508
#> 107 2024-12-18 00:06:33 1.4482 1.4508
#> 108 2024-12-18 00:06:33 1.4482 1.4507
#> 109 2024-12-18 00:06:34 1.4483 1.4507
#> 110 2024-12-18 00:06:36 1.4481 1.4507
#> 111 2024-12-18 00:06:36 1.4482 1.4507
#> 112 2024-12-18 00:06:36 1.4482 1.4506
#> 113 2024-12-18 00:06:36 1.4484 1.4506
#> 114 2024-12-18 00:06:36 1.4484 1.4505
#> 115 2024-12-18 00:06:36 1.4485 1.4505
#> 116 2024-12-18 00:06:37 1.4486 1.4505
#> 117 2024-12-18 00:06:37 1.4487 1.4505
#> 118 2024-12-18 00:06:37 1.4487 1.4509
#> 119 2024-12-18 00:06:37 1.4487 1.4508
#> 120 2024-12-18 00:06:37 1.4487 1.4507
#> 121 2024-12-18 00:06:38 1.4488 1.4507
#> 122 2024-12-18 00:06:38 1.4489 1.4507
#> 123 2024-12-18 00:06:38 1.4489 1.4508
#> 124 2024-12-18 00:06:38 1.4489 1.4507
#> 125 2024-12-18 00:06:38 1.4489 1.4513
#> 126 2024-12-18 00:06:38 1.4489 1.4512
#> 127 2024-12-18 00:06:38 1.4489 1.4511
#> 128 2024-12-18 00:06:38 1.4489 1.4510
#> 129 2024-12-18 00:06:38 1.4489 1.4513
#> 130 2024-12-18 00:06:38 1.4489 1.4512
#> 131 2024-12-18 00:06:38 1.4490 1.4512
#> 132 2024-12-18 00:06:38 1.4491 1.4512
#> 133 2024-12-18 00:06:38 1.4491 1.4513
#> 134 2024-12-18 00:06:39 1.4493 1.4513
#> 135 2024-12-18 00:06:39 1.4494 1.4513
#> 136 2024-12-18 00:06:39 1.4495 1.4513
#> 137 2024-12-18 00:06:39 1.4496 1.4513
#> 138 2024-12-18 00:06:39 1.4494 1.4513
#> 139 2024-12-18 00:06:39 1.4495 1.4513
#> 140 2024-12-18 00:06:39 1.4496 1.4513
#> 141 2024-12-18 00:06:40 1.4497 1.4513
#> 142 2024-12-18 00:06:40 1.4498 1.4513
#> 143 2024-12-18 00:06:40 1.4498 1.4512
#> 144 2024-12-18 00:06:40 1.4498 1.4513
#> 145 2024-12-18 00:06:41 1.4498 1.4518
#> 146 2024-12-18 00:06:41 1.4499 1.4518
#> 147 2024-12-18 00:06:41 1.4499 1.4517
#> 148 2024-12-18 00:06:41 1.4500 1.4517
#> 149 2024-12-18 00:06:41 1.4501 1.4517
#> 150 2024-12-18 00:06:41 1.4502 1.4517
#> 151 2024-12-18 00:06:41 1.4502 1.4518
#> 152 2024-12-18 00:06:41 1.4505 1.4518
#> 153 2024-12-18 00:06:41 1.4506 1.4518
#> 154 2024-12-18 00:06:41 1.4507 1.4518
#> 155 2024-12-18 00:06:41 1.4508 1.4518
#> 156 2024-12-18 00:06:41 1.4507 1.4518
#> 157 2024-12-18 00:06:42 1.4508 1.4518
#> 158 2024-12-18 00:06:42 1.4511 1.4518
#> 159 2024-12-18 00:06:42 1.4512 1.4518
#> 160 2024-12-18 00:06:42 1.4513 1.4518
#> 161 2024-12-18 00:06:42 1.4512 1.4518
#> 162 2024-12-18 00:06:42 1.4513 1.4518
#> 163 2024-12-18 00:06:42 1.4514 1.4518
#> 164 2024-12-18 00:06:42 1.4513 1.4518
#> 165 2024-12-18 00:06:42 1.4514 1.4518
#> 166 2024-12-18 00:06:43 1.4515 1.4518
#> 167 2024-12-18 00:06:43 1.4516 1.4518
#> 168 2024-12-18 00:06:43 1.4516 1.4538
#> 169 2024-12-18 00:06:43 1.4516 1.4537
#> 170 2024-12-18 00:06:43 1.4516 1.4539
#> 171 2024-12-18 00:06:43 1.4516 1.4538
#> 172 2024-12-18 00:06:43 1.4517 1.4538
#> 173 2024-12-18 00:06:43 1.4518 1.4538
#> 174 2024-12-18 00:06:43 1.4519 1.4538
#> 175 2024-12-18 00:06:43 1.4520 1.4538
#> 176 2024-12-18 00:06:44 1.4521 1.4538
#> 177 2024-12-18 00:06:44 1.4522 1.4538
#> 178 2024-12-18 00:06:44 1.4522 1.4539
#> 179 2024-12-18 00:06:44 1.4523 1.4539
#> 180 2024-12-18 00:06:44 1.4524 1.4539
#> 181 2024-12-18 00:06:44 1.4523 1.4539
#> 182 2024-12-18 00:06:44 1.4523 1.4544
#> 183 2024-12-18 00:06:44 1.4524 1.4544
#> 184 2024-12-18 00:06:44 1.4525 1.4544
#> 185 2024-12-18 00:06:44 1.4526 1.4544
#> 186 2024-12-18 00:06:44 1.4527 1.4544
#> 187 2024-12-18 00:06:44 1.4528 1.4544
#> 188 2024-12-18 00:06:44 1.4529 1.4544
#> 189 2024-12-18 00:06:45 1.4530 1.4544
#> 190 2024-12-18 00:06:46 1.4531 1.4544
#> 191 2024-12-18 00:06:46 1.4532 1.4544
#> 192 2024-12-18 00:06:48 1.4530 1.4544
#> 193 2024-12-18 00:06:48 1.4533 1.4544
#> 194 2024-12-18 00:06:48 1.4534 1.4544
#> 195 2024-12-18 00:06:48 1.4535 1.4544
#> 196 2024-12-18 00:06:48 1.4536 1.4544
#> 197 2024-12-18 00:06:48 1.4536 1.4562
#> 198 2024-12-18 00:06:48 1.4537 1.4562
#> 199 2024-12-18 00:06:48 1.4538 1.4562
#> 200 2024-12-18 00:06:49 1.4539 1.4562
#> 201 2024-12-18 00:06:49 1.4542 1.4562
#> 202 2024-12-18 00:06:49 1.4543 1.4562
#> 203 2024-12-18 00:06:49 1.4544 1.4562
#> 204 2024-12-18 00:06:49 1.4545 1.4562
#> 205 2024-12-18 00:06:49 1.4548 1.4562
#> 206 2024-12-18 00:06:49 1.4549 1.4562
#> 207 2024-12-18 00:06:50 1.4550 1.4562
#> 208 2024-12-18 00:06:50 1.4554 1.4562
#> 209 2024-12-18 00:06:50 1.4555 1.4562
#> 210 2024-12-18 00:06:50 1.4555 1.4580
#> 211 2024-12-18 00:06:50 1.4556 1.4580
#> 212 2024-12-18 00:06:50 1.4557 1.4580
#> 213 2024-12-18 00:06:50 1.4558 1.4580
#> 214 2024-12-18 00:06:50 1.4558 1.4581
#> 215 2024-12-18 00:06:50 1.4560 1.4581
#> 216 2024-12-18 00:06:51 1.4561 1.4581
#> 217 2024-12-18 00:06:51 1.4562 1.4581
#> 218 2024-12-18 00:06:51 1.4563 1.4581
#> 219 2024-12-18 00:06:51 1.4564 1.4581
#> 220 2024-12-18 00:06:51 1.4564 1.4593
#> 221 2024-12-18 00:06:51 1.4564 1.4592
#> 222 2024-12-18 00:06:51 1.4564 1.4591
#> 223 2024-12-18 00:06:51 1.4564 1.4590
#> 224 2024-12-18 00:06:51 1.4564 1.4591
#> 225 2024-12-18 00:06:51 1.4566 1.4591
#> 226 2024-12-18 00:06:51 1.4566 1.4593
#> 227 2024-12-18 00:06:51 1.4564 1.4593
#> 228 2024-12-18 00:06:51 1.4565 1.4593
#> 229 2024-12-18 00:06:51 1.4566 1.4593
#> 230 2024-12-18 00:06:51 1.4566 1.4594
#> 231 2024-12-18 00:06:52 1.4567 1.4594
#> 232 2024-12-18 00:06:52 1.4568 1.4594
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 00:06:22 1.002192 1.003529
#> 2   2024-12-18 00:06:22 1.002192 1.003530
#> 3   2024-12-18 00:06:22 1.002192 1.003529
#> 4   2024-12-18 00:06:22 1.002192 1.003452
#> 5   2024-12-18 00:06:22 1.002192 1.003686
#> 6   2024-12-18 00:06:22 1.002192 1.003609
#> 7   2024-12-18 00:06:22 1.002194 1.003609
#> 8   2024-12-18 00:06:22 1.002195 1.003609
#> 9   2024-12-18 00:06:22 1.002196 1.003609
#> 10  2024-12-18 00:06:22 1.002197 1.003609
#> 11  2024-12-18 00:06:22 1.002197 1.003762
#> 12  2024-12-18 00:06:22 1.002197 1.003685
#> 13  2024-12-18 00:06:22 1.002198 1.003685
#> 14  2024-12-18 00:06:22 1.002199 1.003685
#> 15  2024-12-18 00:06:23 1.002200 1.003685
#> 16  2024-12-18 00:06:23 1.002201 1.003685
#> 17  2024-12-18 00:06:23 1.002202 1.003685
#> 18  2024-12-18 00:06:23 1.002203 1.003685
#> 19  2024-12-18 00:06:23 1.002203 1.003684
#> 20  2024-12-18 00:06:24 1.002204 1.003684
#> 21  2024-12-18 00:06:24 1.002205 1.003684
#> 22  2024-12-18 00:06:24 1.002205 1.003692
#> 23  2024-12-18 00:06:24 1.002206 1.003692
#> 24  2024-12-18 00:06:24 1.002206 1.004494
#> 25  2024-12-18 00:06:24 1.002207 1.004494
#> 26  2024-12-18 00:06:24 1.002208 1.004494
#> 27  2024-12-18 00:06:24 1.002208 1.004493
#> 28  2024-12-18 00:06:24 1.003349 1.004493
#> 29  2024-12-18 00:06:24 1.003349 1.004494
#> 30  2024-12-18 00:06:24 1.003349 1.005000
#> 31  2024-12-18 00:06:24 1.002210 1.005000
#> 32  2024-12-18 00:06:24 1.002211 1.005000
#> 33  2024-12-18 00:06:24 1.003350 1.005000
#> 34  2024-12-18 00:06:24 1.003395 1.005000
#> 35  2024-12-18 00:06:24 1.003396 1.005000
#> 36  2024-12-18 00:06:24 1.003636 1.005000
#> 37  2024-12-18 00:06:25 1.003637 1.005000
#> 38  2024-12-18 00:06:25 1.003638 1.005000
#> 39  2024-12-18 00:06:25 1.003637 1.005000
#> 40  2024-12-18 00:06:25 1.003888 1.005000
#> 41  2024-12-18 00:06:25 1.003889 1.005000
#> 42  2024-12-18 00:06:25 1.004129 1.005000
#> 43  2024-12-18 00:06:25 1.004130 1.005000
#> 44  2024-12-18 00:06:26 1.004131 1.005000
#> 45  2024-12-18 00:06:26 1.004132 1.005000
#> 46  2024-12-18 00:06:27 1.004133 1.005000
#> 47  2024-12-18 00:06:27 1.004134 1.005000
#> 48  2024-12-18 00:06:27 1.004133 1.005000
#> 49  2024-12-18 00:06:27 1.004131 1.005000
#> 50  2024-12-18 00:06:28 1.004131 1.004966
#> 51  2024-12-18 00:06:28 1.004131 1.004965
#> 52  2024-12-18 00:06:28 1.004131 1.004964
#> 53  2024-12-18 00:06:29 1.004131 1.004963
#> 54  2024-12-18 00:06:29 1.004131 1.004886
#> 55  2024-12-18 00:06:29 1.004131 1.004885
#> 56  2024-12-18 00:06:29 1.004131 1.004884
#> 57  2024-12-18 00:06:29 1.004131 1.004807
#> 58  2024-12-18 00:06:29 1.004131 1.004806
#> 59  2024-12-18 00:06:29 1.004131 1.004805
#> 60  2024-12-18 00:06:29 1.004131 1.004728
#> 61  2024-12-18 00:06:29 1.004131 1.004727
#> 62  2024-12-18 00:06:29 1.004131 1.004726
#> 63  2024-12-18 00:06:29 1.004131 1.004649
#> 64  2024-12-18 00:06:29 1.004131 1.004648
#> 65  2024-12-18 00:06:29 1.004131 1.004571
#> 66  2024-12-18 00:06:29 1.004131 1.004231
#> 67  2024-12-18 00:06:30 1.004131 1.004565
#> 68  2024-12-18 00:06:30 1.004131 1.004564
#> 69  2024-12-18 00:06:31 1.003350 1.004564
#> 70  2024-12-18 00:06:32 1.003350 1.004563
#> 71  2024-12-18 00:06:32 1.003350 1.004486
#> 72  2024-12-18 00:06:32 1.003350 1.004485
#> 73  2024-12-18 00:06:33 1.003350 1.004484
#> 74  2024-12-18 00:06:33 1.003350 1.004407
#> 75  2024-12-18 00:06:33 1.003350 1.004406
#> 76  2024-12-18 00:06:33 1.003350 1.004405
#> 77  2024-12-18 00:06:33 1.003350 1.004328
#> 78  2024-12-18 00:06:33 1.003350 1.004327
#> 79  2024-12-18 00:06:33 1.003350 1.004326
#> 80  2024-12-18 00:06:33 1.003350 1.004249
#> 81  2024-12-18 00:06:33 1.003350 1.004248
#> 82  2024-12-18 00:06:33 1.003350 1.004171
#> 83  2024-12-18 00:06:34 1.003350 1.004170
#> 84  2024-12-18 00:06:34 1.003350 1.004126
#> 85  2024-12-18 00:06:34 1.003350 1.004170
#> 86  2024-12-18 00:06:34 1.003350 1.004167
#> 87  2024-12-18 00:06:34 1.003350 1.004170
#> 88  2024-12-18 00:06:35 1.003350 1.004248
#> 89  2024-12-18 00:06:35 1.003350 1.004880
#> 90  2024-12-18 00:06:35 1.003350 1.004879
#> 91  2024-12-18 00:06:35 1.003350 1.004802
#> 92  2024-12-18 00:06:35 1.003350 1.004801
#> 93  2024-12-18 00:06:35 1.003350 1.004391
#> 94  2024-12-18 00:06:35 1.003350 1.004333
#> 95  2024-12-18 00:06:35 1.003350 1.004391
#> 96  2024-12-18 00:06:36 1.003350 1.004800
#> 97  2024-12-18 00:06:36 1.003350 1.004723
#> 98  2024-12-18 00:06:36 1.003351 1.004723
#> 99  2024-12-18 00:06:36 1.003351 1.004800
#> 100 2024-12-18 00:06:37 1.003352 1.004800
#> 101 2024-12-18 00:06:37 1.003352 1.004801
#> 102 2024-12-18 00:06:37 1.003353 1.004801
#> 103 2024-12-18 00:06:37 1.003353 1.004799
#> 104 2024-12-18 00:06:37 1.003353 1.004801
#> 105 2024-12-18 00:06:37 1.003353 1.004965
#> 106 2024-12-18 00:06:37 1.003354 1.004965
#> 107 2024-12-18 00:06:37 1.003355 1.004965
#> 108 2024-12-18 00:06:37 1.003356 1.004965
#> 109 2024-12-18 00:06:37 1.003357 1.004965
#> 110 2024-12-18 00:06:37 1.003358 1.004965
#> 111 2024-12-18 00:06:38 1.003358 1.004963
#> 112 2024-12-18 00:06:38 1.003359 1.004963
#> 113 2024-12-18 00:06:38 1.003360 1.004963
#> 114 2024-12-18 00:06:38 1.003360 1.004962
#> 115 2024-12-18 00:06:38 1.003361 1.004962
#> 116 2024-12-18 00:06:38 1.003361 1.004961
#> 117 2024-12-18 00:06:39 1.003361 1.004883
#> 118 2024-12-18 00:06:39 1.003362 1.004883
#> 119 2024-12-18 00:06:39 1.003363 1.004883
#> 120 2024-12-18 00:06:40 1.003364 1.004883
#> 121 2024-12-18 00:06:40 1.003365 1.004883
#> 122 2024-12-18 00:06:41 1.003365 1.004961
#> 123 2024-12-18 00:06:41 1.003365 1.004965
#> 124 2024-12-18 00:06:41 1.003365 1.004964
#> 125 2024-12-18 00:06:41 1.003365 1.004918
#> 126 2024-12-18 00:06:41 1.003365 1.004917
#> 127 2024-12-18 00:06:41 1.003365 1.004912
#> 128 2024-12-18 00:06:41 1.003365 1.004911
#> 129 2024-12-18 00:06:41 1.003366 1.004911
#> 130 2024-12-18 00:06:41 1.003367 1.004911
#> 131 2024-12-18 00:06:41 1.003368 1.004911
#> 132 2024-12-18 00:06:41 1.003369 1.004911
#> 133 2024-12-18 00:06:41 1.003370 1.004911
#> 134 2024-12-18 00:06:41 1.003371 1.004911
#> 135 2024-12-18 00:06:42 1.003372 1.004911
#> 136 2024-12-18 00:06:42 1.003373 1.004911
#> 137 2024-12-18 00:06:42 1.003374 1.004911
#> 138 2024-12-18 00:06:42 1.003375 1.004911
#> 139 2024-12-18 00:06:42 1.003376 1.004911
#> 140 2024-12-18 00:06:42 1.003376 1.005000
#> 141 2024-12-18 00:06:42 1.003376 1.004999
#> 142 2024-12-18 00:06:42 1.003377 1.004999
#> 143 2024-12-18 00:06:42 1.003378 1.004999
#> 144 2024-12-18 00:06:42 1.003378 1.004998
#> 145 2024-12-18 00:06:42 1.003379 1.004998
#> 146 2024-12-18 00:06:42 1.003379 1.004997
#> 147 2024-12-18 00:06:42 1.003380 1.004997
#> 148 2024-12-18 00:06:42 1.003381 1.004997
#> 149 2024-12-18 00:06:42 1.003382 1.004997
#> 150 2024-12-18 00:06:42 1.003383 1.004997
#> 151 2024-12-18 00:06:42 1.003384 1.004997
#> 152 2024-12-18 00:06:42 1.003385 1.004997
#> 153 2024-12-18 00:06:42 1.003386 1.004997
#> 154 2024-12-18 00:06:42 1.003466 1.004997
#> 155 2024-12-18 00:06:42 1.003466 1.004998
#> 156 2024-12-18 00:06:42 1.003467 1.004998
#> 157 2024-12-18 00:06:42 1.003468 1.004998
#> 158 2024-12-18 00:06:42 1.003469 1.004998
#> 159 2024-12-18 00:06:43 1.003470 1.004998
#> 160 2024-12-18 00:06:43 1.003471 1.004998
#> 161 2024-12-18 00:06:43 1.003471 1.005000
#> 162 2024-12-18 00:06:43 1.003472 1.005000
#> 163 2024-12-18 00:06:43 1.003473 1.005000
#> 164 2024-12-18 00:06:43 1.003474 1.005000
#> 165 2024-12-18 00:06:43 1.003475 1.005000
#> 166 2024-12-18 00:06:43 1.003567 1.005000
#> 167 2024-12-18 00:06:43 1.003568 1.005000
#> 168 2024-12-18 00:06:43 1.003569 1.005000
#> 169 2024-12-18 00:06:43 1.003570 1.005000
#> 170 2024-12-18 00:06:43 1.003571 1.005000
#> 171 2024-12-18 00:06:43 1.003572 1.005000
#> 172 2024-12-18 00:06:43 1.003573 1.005000
#> 173 2024-12-18 00:06:44 1.003574 1.005000
#> 174 2024-12-18 00:06:44 1.003575 1.005000
#> 175 2024-12-18 00:06:44 1.003576 1.005000
#> 176 2024-12-18 00:06:44 1.003577 1.005000
#> 177 2024-12-18 00:06:44 1.003578 1.005000
#> 178 2024-12-18 00:06:44 1.003819 1.005000
#> 179 2024-12-18 00:06:44 1.003820 1.005000
#> 180 2024-12-18 00:06:44 1.004082 1.005000
#> 181 2024-12-18 00:06:44 1.004083 1.005000
#> 182 2024-12-18 00:06:44 1.004084 1.005000
#> 183 2024-12-18 00:06:45 1.004085 1.005000
#> 184 2024-12-18 00:06:45 1.004086 1.005000
#> 185 2024-12-18 00:06:45 1.004087 1.005000
#> 186 2024-12-18 00:06:46 1.004088 1.005000
#> 187 2024-12-18 00:06:46 1.004089 1.005000
#> 188 2024-12-18 00:06:47 1.004090 1.005000
#> 189 2024-12-18 00:06:47 1.004091 1.005000
#> 190 2024-12-18 00:06:48 1.004136 1.005000
#> 191 2024-12-18 00:06:48 1.004137 1.005000
#> 192 2024-12-18 00:06:48 1.004138 1.005000
#> 193 2024-12-18 00:06:48 1.004139 1.005000
#> 194 2024-12-18 00:06:49 1.004140 1.005000
#> 195 2024-12-18 00:06:49 1.004141 1.005000
#> 196 2024-12-18 00:06:49 1.004142 1.005000
#> 197 2024-12-18 00:06:49 1.004143 1.005000
#> 198 2024-12-18 00:06:49 1.004144 1.005000
#> 199 2024-12-18 00:06:49 1.004145 1.005000
#> 200 2024-12-18 00:06:49 1.004146 1.005000
#> 201 2024-12-18 00:06:49 1.004147 1.005000
#> 202 2024-12-18 00:06:49 1.004148 1.005000
#> 203 2024-12-18 00:06:49 1.004149 1.005000
#> 204 2024-12-18 00:06:50 1.004150 1.005000
#> 205 2024-12-18 00:06:50 1.004151 1.005000
#> 206 2024-12-18 00:06:50 1.004152 1.005000
#> 207 2024-12-18 00:06:50 1.004153 1.005000
#> 208 2024-12-18 00:06:50 1.004154 1.005000
#> 209 2024-12-18 00:06:50 1.004155 1.005000
#> 210 2024-12-18 00:06:50 1.004156 1.005000
#> 211 2024-12-18 00:06:50 1.004281 1.005000
#> 212 2024-12-18 00:06:50 1.004282 1.005000
#> 213 2024-12-18 00:06:51 1.004283 1.005000
#> 214 2024-12-18 00:06:51 1.004284 1.005000
#> 215 2024-12-18 00:06:51 1.004285 1.005000
#> 216 2024-12-18 00:06:51 1.004286 1.005000
#> 217 2024-12-18 00:06:51 1.004287 1.005000
#> 218 2024-12-18 00:06:51 1.004288 1.005000
#> 219 2024-12-18 00:06:51 1.004289 1.005000
#> 220 2024-12-18 00:06:51 1.004290 1.005000
#> 221 2024-12-18 00:06:51 1.004291 1.005000
#> 222 2024-12-18 00:06:51 1.004292 1.005000
#> 223 2024-12-18 00:06:51 1.004545 1.005000
#> 224 2024-12-18 00:06:52 1.004546 1.005000
#> 225 2024-12-18 00:06:52 1.004547 1.005000
#> 226 2024-12-18 00:06:52 1.004548 1.005000
#> 227 2024-12-18 00:06:52 1.004549 1.005000
#> 228 2024-12-18 00:06:52 1.004550 1.005000
#> 229 2024-12-18 00:06:52 1.004551 1.005000
#> 230 2024-12-18 00:06:52 1.004552 1.005000
#> 231 2024-12-18 00:06:52 1.004553 1.005000
#> 232 2024-12-18 00:06:52 1.004554 1.005000
#> 233 2024-12-18 00:06:52 1.004555 1.005000
#> 234 2024-12-18 00:06:53 1.004556 1.005000
#> 235 2024-12-18 00:06:53 1.004557 1.005000
#> 236 2024-12-18 00:06:53 1.004558 1.005000
#> 237 2024-12-18 00:06:53 1.004559 1.005000
#> 238 2024-12-18 00:06:53 1.004560 1.005000
#> 239 2024-12-18 00:06:53 1.004561 1.005000
#> 240 2024-12-18 00:06:53 1.004562 1.005000
#> 241 2024-12-18 00:06:53 1.004583 1.005000
#> 242 2024-12-18 00:06:53 1.004584 1.005000
#> 243 2024-12-18 00:06:53 1.004585 1.005000
#> 244 2024-12-18 00:06:53 1.004586 1.005000
#> 245 2024-12-18 00:06:53 1.004587 1.005000
#> 246 2024-12-18 00:06:54 1.004588 1.005000
#> 247 2024-12-18 00:06:54 1.004787 1.005000
#> 248 2024-12-18 00:06:54 1.004788 1.005000
#> 
#> $last
#> [1] 1734480414
#>