Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 17:34:18 1.4008 1.4013
#> 2   2024-12-18 17:34:19 1.4009 1.4013
#> 3   2024-12-18 17:34:22 1.4009 1.4016
#> 4   2024-12-18 17:34:22 1.4010 1.4015
#> 5   2024-12-18 17:34:22 1.4010 1.4020
#> 6   2024-12-18 17:34:22 1.4010 1.4019
#> 7   2024-12-18 17:34:22 1.4011 1.4019
#> 8   2024-12-18 17:34:22 1.4011 1.4022
#> 9   2024-12-18 17:34:22 1.4011 1.4025
#> 10  2024-12-18 17:34:23 1.4011 1.4023
#> 11  2024-12-18 17:34:23 1.4011 1.4022
#> 12  2024-12-18 17:34:23 1.4012 1.4022
#> 13  2024-12-18 17:34:24 1.4013 1.4022
#> 14  2024-12-18 17:34:25 1.4014 1.4022
#> 15  2024-12-18 17:34:25 1.4015 1.4022
#> 16  2024-12-18 17:34:30 1.4016 1.4022
#> 17  2024-12-18 17:34:30 1.4017 1.4022
#> 18  2024-12-18 17:34:32 1.4018 1.4022
#> 19  2024-12-18 17:34:35 1.4017 1.4022
#> 20  2024-12-18 17:34:35 1.4009 1.4022
#> 21  2024-12-18 17:34:35 1.4010 1.4021
#> 22  2024-12-18 17:34:35 1.4010 1.4020
#> 23  2024-12-18 17:34:35 1.4005 1.4020
#> 24  2024-12-18 17:34:35 1.4007 1.4020
#> 25  2024-12-18 17:34:35 1.4008 1.4020
#> 26  2024-12-18 17:34:35 1.4009 1.4020
#> 27  2024-12-18 17:34:35 1.4009 1.4019
#> 28  2024-12-18 17:34:35 1.4008 1.4019
#> 29  2024-12-18 17:34:35 1.4008 1.4018
#> 30  2024-12-18 17:34:36 1.4008 1.4017
#> 31  2024-12-18 17:34:36 1.4008 1.4016
#> 32  2024-12-18 17:34:36 1.4009 1.4016
#> 33  2024-12-18 17:34:36 1.4009 1.4021
#> 34  2024-12-18 17:34:36 1.4009 1.4021
#> 35  2024-12-18 17:34:36 1.4010 1.4020
#> 36  2024-12-18 17:34:36 1.4011 1.4020
#> 37  2024-12-18 17:34:36 1.4011 1.4023
#> 38  2024-12-18 17:34:36 1.4011 1.4021
#> 39  2024-12-18 17:34:36 1.4011 1.4020
#> 40  2024-12-18 17:34:38 1.4011 1.4023
#> 41  2024-12-18 17:34:38 1.4012 1.4022
#> 42  2024-12-18 17:34:38 1.4013 1.4022
#> 43  2024-12-18 17:34:39 1.4013 1.4024
#> 44  2024-12-18 17:34:39 1.4014 1.4023
#> 45  2024-12-18 17:34:39 1.4015 1.4023
#> 46  2024-12-18 17:34:39 1.4015 1.4022
#> 47  2024-12-18 17:34:39 1.4012 1.4022
#> 48  2024-12-18 17:34:40 1.4014 1.4022
#> 49  2024-12-18 17:34:40 1.4014 1.4023
#> 50  2024-12-18 17:34:40 1.4015 1.4023
#> 51  2024-12-18 17:34:40 1.4015 1.4022
#> 52  2024-12-18 17:34:42 1.4016 1.4022
#> 53  2024-12-18 17:34:42 1.4017 1.4022
#> 54  2024-12-18 17:34:42 1.4017 1.4025
#> 55  2024-12-18 17:34:43 1.4017 1.4033
#> 56  2024-12-18 17:34:43 1.4017 1.4031
#> 57  2024-12-18 17:34:43 1.4017 1.4030
#> 58  2024-12-18 17:34:43 1.4018 1.4030
#> 59  2024-12-18 17:34:43 1.4018 1.4034
#> 60  2024-12-18 17:34:43 1.4019 1.4033
#> 61  2024-12-18 17:34:43 1.4020 1.4033
#> 62  2024-12-18 17:34:43 1.4020 1.4036
#> 63  2024-12-18 17:34:43 1.4020 1.4034
#> 64  2024-12-18 17:34:43 1.4020 1.4033
#> 65  2024-12-18 17:34:43 1.4021 1.4033
#> 66  2024-12-18 17:34:43 1.4022 1.4033
#> 67  2024-12-18 17:34:43 1.4022 1.4032
#> 68  2024-12-18 17:34:43 1.4023 1.4032
#> 69  2024-12-18 17:34:43 1.4024 1.4032
#> 70  2024-12-18 17:34:44 1.4025 1.4032
#> 71  2024-12-18 17:34:44 1.4026 1.4032
#> 72  2024-12-18 17:34:44 1.4027 1.4032
#> 73  2024-12-18 17:34:45 1.4027 1.4034
#> 74  2024-12-18 17:34:45 1.4028 1.4033
#> 75  2024-12-18 17:34:45 1.4023 1.4033
#> 76  2024-12-18 17:34:45 1.4028 1.4033
#> 77  2024-12-18 17:34:45 1.4027 1.4033
#> 78  2024-12-18 17:34:46 1.4027 1.4035
#> 79  2024-12-18 17:34:46 1.4027 1.4042
#> 80  2024-12-18 17:34:46 1.4027 1.4042
#> 81  2024-12-18 17:34:46 1.4028 1.4041
#> 82  2024-12-18 17:34:47 1.4028 1.4040
#> 83  2024-12-18 17:34:47 1.4029 1.4040
#> 84  2024-12-18 17:34:47 1.4030 1.4040
#> 85  2024-12-18 17:34:47 1.4030 1.4039
#> 86  2024-12-18 17:34:47 1.4030 1.4041
#> 87  2024-12-18 17:34:47 1.4031 1.4040
#> 88  2024-12-18 17:34:47 1.4032 1.4040
#> 89  2024-12-18 17:34:47 1.4033 1.4040
#> 90  2024-12-18 17:34:47 1.4033 1.4043
#> 91  2024-12-18 17:34:47 1.4033 1.4042
#> 92  2024-12-18 17:34:47 1.4033 1.4044
#> 93  2024-12-18 17:34:47 1.4033 1.4043
#> 94  2024-12-18 17:34:47 1.4034 1.4043
#> 95  2024-12-18 17:34:47 1.4035 1.4043
#> 96  2024-12-18 17:34:47 1.4035 1.4042
#> 97  2024-12-18 17:34:47 1.4036 1.4042
#> 98  2024-12-18 17:34:49 1.4037 1.4042
#> 99  2024-12-18 17:34:49 1.4037 1.4041
#> 100 2024-12-18 17:34:50 1.4035 1.4041
#> 101 2024-12-18 17:34:50 1.4036 1.4041
#> 102 2024-12-18 17:34:50 1.4035 1.4041
#> 103 2024-12-18 17:34:50 1.4035 1.4040
#> 104 2024-12-18 17:34:50 1.4035 1.4039
#> 105 2024-12-18 17:34:50 1.4035 1.4042
#> 106 2024-12-18 17:34:50 1.4035 1.4041
#> 107 2024-12-18 17:34:50 1.4035 1.4040
#> 108 2024-12-18 17:34:50 1.4035 1.4039
#> 109 2024-12-18 17:34:50 1.4030 1.4039
#> 110 2024-12-18 17:34:50 1.4032 1.4039
#> 111 2024-12-18 17:34:51 1.4031 1.4039
#> 112 2024-12-18 17:34:51 1.4032 1.4039
#> 113 2024-12-18 17:34:51 1.4032 1.4038
#> 114 2024-12-18 17:34:51 1.4032 1.4040
#> 115 2024-12-18 17:34:51 1.4031 1.4039
#> 116 2024-12-18 17:34:51 1.4031 1.4038
#> 117 2024-12-18 17:34:51 1.4031 1.4037
#> 118 2024-12-18 17:34:51 1.4031 1.4036
#> 119 2024-12-18 17:34:51 1.4031 1.4038
#> 120 2024-12-18 17:34:51 1.4033 1.4037
#> 121 2024-12-18 17:34:51 1.4033 1.4046
#> 122 2024-12-18 17:34:51 1.4033 1.4046
#> 123 2024-12-18 17:34:51 1.4033 1.4043
#> 124 2024-12-18 17:34:51 1.4033 1.4042
#> 125 2024-12-18 17:34:51 1.4033 1.4046
#> 126 2024-12-18 17:34:52 1.4033 1.4045
#> 127 2024-12-18 17:34:52 1.4034 1.4045
#> 128 2024-12-18 17:34:52 1.4033 1.4045
#> 129 2024-12-18 17:34:52 1.4033 1.4046
#> 130 2024-12-18 17:34:52 1.4034 1.4046
#> 131 2024-12-18 17:34:52 1.4034 1.4045
#> 132 2024-12-18 17:34:52 1.4034 1.4044
#> 133 2024-12-18 17:34:52 1.4034 1.4043
#> 134 2024-12-18 17:34:52 1.4034 1.4042
#> 135 2024-12-18 17:34:53 1.4034 1.4045
#> 136 2024-12-18 17:34:53 1.4034 1.4044
#> 137 2024-12-18 17:34:53 1.4035 1.4044
#> 138 2024-12-18 17:34:53 1.4035 1.4043
#> 139 2024-12-18 17:34:53 1.4033 1.4043
#> 140 2024-12-18 17:34:53 1.4034 1.4043
#> 141 2024-12-18 17:34:53 1.4035 1.4043
#> 142 2024-12-18 17:34:53 1.4034 1.4043
#> 143 2024-12-18 17:34:53 1.4035 1.4043
#> 144 2024-12-18 17:34:54 1.4035 1.4042
#> 145 2024-12-18 17:34:54 1.4035 1.4041
#> 146 2024-12-18 17:34:54 1.4035 1.4043
#> 147 2024-12-18 17:34:54 1.4035 1.4042
#> 148 2024-12-18 17:34:54 1.4035 1.4043
#> 149 2024-12-18 17:34:55 1.4035 1.4045
#> 150 2024-12-18 17:34:55 1.4035 1.4042
#> 151 2024-12-18 17:34:55 1.4036 1.4042
#> 152 2024-12-18 17:34:55 1.4036 1.4044
#> 153 2024-12-18 17:34:55 1.4036 1.4041
#> 154 2024-12-18 17:34:55 1.4036 1.4044
#> 155 2024-12-18 17:34:56 1.4037 1.4043
#> 156 2024-12-18 17:34:56 1.4034 1.4043
#> 157 2024-12-18 17:34:56 1.4037 1.4043
#> 158 2024-12-18 17:34:57 1.4037 1.4042
#> 159 2024-12-18 17:34:57 1.4037 1.4041
#> 160 2024-12-18 17:34:57 1.4037 1.4043
#> 161 2024-12-18 17:34:57 1.4037 1.4042
#> 162 2024-12-18 17:35:00 1.4031 1.4042
#> 163 2024-12-18 17:35:00 1.4032 1.4042
#> 164 2024-12-18 17:35:00 1.4033 1.4042
#> 165 2024-12-18 17:35:01 1.4034 1.4042
#> 166 2024-12-18 17:35:01 1.4035 1.4042
#> 167 2024-12-18 17:35:01 1.4032 1.4042
#> 168 2024-12-18 17:35:01 1.4033 1.4042
#> 169 2024-12-18 17:35:02 1.4034 1.4042
#> 170 2024-12-18 17:35:02 1.4035 1.4042
#> 171 2024-12-18 17:35:02 1.4035 1.4045
#> 172 2024-12-18 17:35:02 1.4035 1.4044
#> 173 2024-12-18 17:35:02 1.4035 1.4046
#> 174 2024-12-18 17:35:02 1.4035 1.4045
#> 175 2024-12-18 17:35:02 1.4035 1.4044
#> 176 2024-12-18 17:35:03 1.4036 1.4044
#> 177 2024-12-18 17:35:03 1.4037 1.4044
#> 178 2024-12-18 17:35:03 1.4038 1.4044
#> 179 2024-12-18 17:35:03 1.4038 1.4046
#> 180 2024-12-18 17:35:03 1.4038 1.4046
#> 181 2024-12-18 17:35:04 1.4038 1.4045
#> 182 2024-12-18 17:35:04 1.4038 1.4046
#> 183 2024-12-18 17:35:05 1.4039 1.4046
#> 184 2024-12-18 17:35:05 1.4039 1.4045
#> 185 2024-12-18 17:35:05 1.4040 1.4045
#> 186 2024-12-18 17:35:05 1.4040 1.4044
#> 187 2024-12-18 17:35:05 1.4040 1.4043
#> 188 2024-12-18 17:35:05 1.4039 1.4043
#> 189 2024-12-18 17:35:05 1.4040 1.4043
#> 190 2024-12-18 17:35:05 1.4040 1.4044
#> 191 2024-12-18 17:35:05 1.4040 1.4045
#> 192 2024-12-18 17:35:05 1.4040 1.4044
#> 193 2024-12-18 17:35:06 1.4040 1.4046
#> 194 2024-12-18 17:35:07 1.4040 1.4045
#> 195 2024-12-18 17:35:07 1.4041 1.4045
#> 196 2024-12-18 17:35:08 1.4041 1.4046
#> 197 2024-12-18 17:35:08 1.4041 1.4045
#> 198 2024-12-18 17:35:08 1.4041 1.4044
#> 199 2024-12-18 17:35:10 1.4038 1.4044
#> 200 2024-12-18 17:35:10 1.4039 1.4044
#> 201 2024-12-18 17:35:10 1.4035 1.4044
#> 202 2024-12-18 17:35:10 1.4037 1.4044
#> 203 2024-12-18 17:35:10 1.4035 1.4044
#> 204 2024-12-18 17:35:10 1.4035 1.4043
#> 205 2024-12-18 17:35:10 1.4030 1.4043
#> 206 2024-12-18 17:35:10 1.4031 1.4042
#> 207 2024-12-18 17:35:10 1.4031 1.4041
#> 208 2024-12-18 17:35:10 1.4027 1.4041
#> 209 2024-12-18 17:35:10 1.4029 1.4041
#> 210 2024-12-18 17:35:10 1.4030 1.4041
#> 211 2024-12-18 17:35:10 1.4029 1.4041
#> 212 2024-12-18 17:35:10 1.4022 1.4041
#> 213 2024-12-18 17:35:10 1.4024 1.4041
#> 214 2024-12-18 17:35:10 1.4025 1.4041
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 17:35:01 0.979541 0.980418
#> 2   2024-12-18 17:35:01 0.979542 0.980418
#> 3   2024-12-18 17:35:01 0.979543 0.980418
#> 4   2024-12-18 17:35:01 0.979544 0.980418
#> 5   2024-12-18 17:35:01 0.979545 0.980418
#> 6   2024-12-18 17:35:01 0.979546 0.980418
#> 7   2024-12-18 17:35:01 0.979545 0.980418
#> 8   2024-12-18 17:35:01 0.979545 0.980418
#> 9   2024-12-18 17:35:01 0.979547 0.980418
#> 10  2024-12-18 17:35:01 0.979548 0.980418
#> 11  2024-12-18 17:35:01 0.979549 0.980418
#> 12  2024-12-18 17:35:01 0.979550 0.980418
#> 13  2024-12-18 17:35:02 0.979551 0.980418
#> 14  2024-12-18 17:35:02 0.979552 0.980418
#> 15  2024-12-18 17:35:02 0.979557 0.980418
#> 16  2024-12-18 17:35:02 0.979558 0.980418
#> 17  2024-12-18 17:35:02 0.979559 0.980418
#> 18  2024-12-18 17:35:02 0.979560 0.980418
#> 19  2024-12-18 17:35:02 0.979561 0.980418
#> 20  2024-12-18 17:35:02 0.979562 0.980418
#> 21  2024-12-18 17:35:02 0.979563 0.980418
#> 22  2024-12-18 17:35:02 0.979564 0.980418
#> 23  2024-12-18 17:35:02 0.979564 0.980419
#> 24  2024-12-18 17:35:02 0.979565 0.980419
#> 25  2024-12-18 17:35:02 0.979565 0.980418
#> 26  2024-12-18 17:35:02 0.979566 0.980418
#> 27  2024-12-18 17:35:02 0.979566 0.980419
#> 28  2024-12-18 17:35:02 0.979567 0.980419
#> 29  2024-12-18 17:35:02 0.979568 0.980419
#> 30  2024-12-18 17:35:02 0.979569 0.980419
#> 31  2024-12-18 17:35:02 0.979570 0.980419
#> 32  2024-12-18 17:35:02 0.979575 0.980419
#> 33  2024-12-18 17:35:02 0.979576 0.980419
#> 34  2024-12-18 17:35:02 0.979570 0.980419
#> 35  2024-12-18 17:35:02 0.979572 0.980419
#> 36  2024-12-18 17:35:02 0.979573 0.980419
#> 37  2024-12-18 17:35:02 0.979572 0.980419
#> 38  2024-12-18 17:35:02 0.979572 0.980419
#> 39  2024-12-18 17:35:02 0.979574 0.980419
#> 40  2024-12-18 17:35:02 0.979575 0.980419
#> 41  2024-12-18 17:35:02 0.979576 0.980419
#> 42  2024-12-18 17:35:02 0.979577 0.980419
#> 43  2024-12-18 17:35:02 0.979578 0.980419
#> 44  2024-12-18 17:35:02 0.979579 0.980419
#> 45  2024-12-18 17:35:02 0.979580 0.980419
#> 46  2024-12-18 17:35:02 0.979581 0.980419
#> 47  2024-12-18 17:35:03 0.979582 0.980419
#> 48  2024-12-18 17:35:03 0.979583 0.980419
#> 49  2024-12-18 17:35:03 0.979584 0.980419
#> 50  2024-12-18 17:35:03 0.979585 0.980419
#> 51  2024-12-18 17:35:03 0.979585 0.980791
#> 52  2024-12-18 17:35:03 0.979586 0.980791
#> 53  2024-12-18 17:35:03 0.979586 0.980792
#> 54  2024-12-18 17:35:03 0.979586 0.980672
#> 55  2024-12-18 17:35:03 0.979586 0.980696
#> 56  2024-12-18 17:35:03 0.980418 0.980696
#> 57  2024-12-18 17:35:03 0.979586 0.980696
#> 58  2024-12-18 17:35:03 0.979587 0.980696
#> 59  2024-12-18 17:35:03 0.979587 0.980695
#> 60  2024-12-18 17:35:03 0.979587 0.980696
#> 61  2024-12-18 17:35:03 0.979587 0.980696
#> 62  2024-12-18 17:35:03 0.979587 0.980696
#> 63  2024-12-18 17:35:03 0.979588 0.980696
#> 64  2024-12-18 17:35:03 0.979588 0.980695
#> 65  2024-12-18 17:35:03 0.979588 0.980696
#> 66  2024-12-18 17:35:03 0.979589 0.980696
#> 67  2024-12-18 17:35:03 0.979589 0.980694
#> 68  2024-12-18 17:35:03 0.979590 0.980694
#> 69  2024-12-18 17:35:03 0.979590 0.980696
#> 70  2024-12-18 17:35:03 0.979590 0.980696
#> 71  2024-12-18 17:35:03 0.979596 0.980696
#> 72  2024-12-18 17:35:03 0.979597 0.980696
#> 73  2024-12-18 17:35:03 0.979596 0.980696
#> 74  2024-12-18 17:35:03 0.979597 0.980696
#> 75  2024-12-18 17:35:03 0.979596 0.980696
#> 76  2024-12-18 17:35:03 0.979596 0.980696
#> 77  2024-12-18 17:35:03 0.979596 0.980696
#> 78  2024-12-18 17:35:03 0.979600 0.980696
#> 79  2024-12-18 17:35:03 0.979601 0.980696
#> 80  2024-12-18 17:35:03 0.979602 0.980696
#> 81  2024-12-18 17:35:03 0.979675 0.980696
#> 82  2024-12-18 17:35:03 0.979676 0.980696
#> 83  2024-12-18 17:35:03 0.979677 0.980696
#> 84  2024-12-18 17:35:03 0.979678 0.980696
#> 85  2024-12-18 17:35:03 0.979679 0.980696
#> 86  2024-12-18 17:35:03 0.979680 0.980696
#> 87  2024-12-18 17:35:03 0.979681 0.980696
#> 88  2024-12-18 17:35:03 0.979682 0.980696
#> 89  2024-12-18 17:35:03 0.979683 0.980696
#> 90  2024-12-18 17:35:03 0.979684 0.980696
#> 91  2024-12-18 17:35:03 0.979685 0.980696
#> 92  2024-12-18 17:35:04 0.979685 0.980695
#> 93  2024-12-18 17:35:04 0.979686 0.980695
#> 94  2024-12-18 17:35:04 0.979687 0.980695
#> 95  2024-12-18 17:35:04 0.979688 0.980695
#> 96  2024-12-18 17:35:04 0.979689 0.980695
#> 97  2024-12-18 17:35:04 0.979690 0.980695
#> 98  2024-12-18 17:35:04 0.979691 0.980695
#> 99  2024-12-18 17:35:04 0.979696 0.980695
#> 100 2024-12-18 17:35:04 0.979697 0.980695
#> 101 2024-12-18 17:35:04 0.979698 0.980695
#> 102 2024-12-18 17:35:04 0.979699 0.980695
#> 103 2024-12-18 17:35:04 0.979699 0.980657
#> 104 2024-12-18 17:35:04 0.979699 0.980656
#> 105 2024-12-18 17:35:04 0.979699 0.980654
#> 106 2024-12-18 17:35:04 0.979700 0.980654
#> 107 2024-12-18 17:35:04 0.979701 0.980654
#> 108 2024-12-18 17:35:04 0.979701 0.980653
#> 109 2024-12-18 17:35:05 0.979702 0.980653
#> 110 2024-12-18 17:35:05 0.979703 0.980653
#> 111 2024-12-18 17:35:05 0.979704 0.980653
#> 112 2024-12-18 17:35:05 0.979705 0.980653
#> 113 2024-12-18 17:35:05 0.979706 0.980653
#> 114 2024-12-18 17:35:05 0.979707 0.980653
#> 115 2024-12-18 17:35:05 0.979708 0.980653
#> 116 2024-12-18 17:35:05 0.979708 0.980654
#> 117 2024-12-18 17:35:05 0.979709 0.980654
#> 118 2024-12-18 17:35:05 0.979710 0.980654
#> 119 2024-12-18 17:35:05 0.979710 0.980653
#> 120 2024-12-18 17:35:05 0.979710 0.980696
#> 121 2024-12-18 17:35:05 0.979711 0.980696
#> 122 2024-12-18 17:35:05 0.979712 0.980696
#> 123 2024-12-18 17:35:05 0.979717 0.980696
#> 124 2024-12-18 17:35:05 0.979718 0.980696
#> 125 2024-12-18 17:35:05 0.979719 0.980696
#> 126 2024-12-18 17:35:06 0.979720 0.980696
#> 127 2024-12-18 17:35:06 0.979724 0.980696
#> 128 2024-12-18 17:35:06 0.979725 0.980696
#> 129 2024-12-18 17:35:06 0.979726 0.980696
#> 130 2024-12-18 17:35:06 0.979725 0.980696
#> 131 2024-12-18 17:35:06 0.979725 0.980696
#> 132 2024-12-18 17:35:06 0.979727 0.980696
#> 133 2024-12-18 17:35:06 0.979728 0.980696
#> 134 2024-12-18 17:35:06 0.979729 0.980696
#> 135 2024-12-18 17:35:06 0.979734 0.980696
#> 136 2024-12-18 17:35:06 0.979735 0.980696
#> 137 2024-12-18 17:35:06 0.979736 0.980696
#> 138 2024-12-18 17:35:06 0.979737 0.980696
#> 139 2024-12-18 17:35:06 0.979738 0.980696
#> 140 2024-12-18 17:35:06 0.979739 0.980696
#> 141 2024-12-18 17:35:06 0.979740 0.980696
#> 142 2024-12-18 17:35:06 0.979741 0.980696
#> 143 2024-12-18 17:35:06 0.979742 0.980696
#> 144 2024-12-18 17:35:06 0.979743 0.980696
#> 145 2024-12-18 17:35:06 0.979743 0.980695
#> 146 2024-12-18 17:35:06 0.979743 0.980696
#> 147 2024-12-18 17:35:06 0.979743 0.980694
#> 148 2024-12-18 17:35:06 0.979743 0.980696
#> 149 2024-12-18 17:35:07 0.979744 0.980695
#> 150 2024-12-18 17:35:07 0.979745 0.980695
#> 151 2024-12-18 17:35:07 0.979746 0.980695
#> 152 2024-12-18 17:35:07 0.979745 0.980695
#> 153 2024-12-18 17:35:07 0.979745 0.980695
#> 154 2024-12-18 17:35:07 0.979747 0.980695
#> 155 2024-12-18 17:35:07 0.979748 0.980695
#> 156 2024-12-18 17:35:07 0.979753 0.980695
#> 157 2024-12-18 17:35:07 0.979754 0.980695
#> 158 2024-12-18 17:35:07 0.979755 0.980695
#> 159 2024-12-18 17:35:07 0.979756 0.980695
#> 160 2024-12-18 17:35:08 0.979757 0.980695
#> 161 2024-12-18 17:35:08 0.979758 0.980695
#> 162 2024-12-18 17:35:08 0.979759 0.980695
#> 163 2024-12-18 17:35:08 0.979760 0.980695
#> 164 2024-12-18 17:35:08 0.979761 0.980695
#> 165 2024-12-18 17:35:08 0.979762 0.980695
#> 166 2024-12-18 17:35:08 0.979763 0.980695
#> 167 2024-12-18 17:35:08 0.979764 0.980695
#> 168 2024-12-18 17:35:08 0.979765 0.980695
#> 169 2024-12-18 17:35:08 0.979766 0.980695
#> 170 2024-12-18 17:35:08 0.979767 0.980695
#> 171 2024-12-18 17:35:08 0.979768 0.980695
#> 172 2024-12-18 17:35:08 0.979769 0.980695
#> 173 2024-12-18 17:35:08 0.979770 0.980695
#> 174 2024-12-18 17:35:08 0.979763 0.980695
#> 175 2024-12-18 17:35:08 0.979764 0.980694
#> 176 2024-12-18 17:35:08 0.979764 0.980693
#> 177 2024-12-18 17:35:08 0.979764 0.980691
#> 178 2024-12-18 17:35:08 0.979765 0.980691
#> 179 2024-12-18 17:35:08 0.979765 0.980690
#> 180 2024-12-18 17:35:08 0.979766 0.980690
#> 181 2024-12-18 17:35:09 0.979766 0.980676
#> 182 2024-12-18 17:35:09 0.979766 0.980675
#> 183 2024-12-18 17:35:09 0.979767 0.980675
#> 184 2024-12-18 17:35:09 0.979767 0.980669
#> 185 2024-12-18 17:35:09 0.979768 0.980669
#> 186 2024-12-18 17:35:09 0.979768 0.980668
#> 187 2024-12-18 17:35:09 0.979768 0.980460
#> 188 2024-12-18 17:35:09 0.979768 0.980459
#> 189 2024-12-18 17:35:10 0.979767 0.980459
#> 190 2024-12-18 17:35:10 0.979767 0.980458
#> 191 2024-12-18 17:35:10 0.979767 0.980457
#> 192 2024-12-18 17:35:10 0.979767 0.980456
#> 193 2024-12-18 17:35:10 0.979767 0.980455
#> 194 2024-12-18 17:35:10 0.979767 0.980450
#> 195 2024-12-18 17:35:10 0.979619 0.980450
#> 196 2024-12-18 17:35:10 0.979619 0.980449
#> 197 2024-12-18 17:35:10 0.979619 0.980448
#> 198 2024-12-18 17:35:10 0.979619 0.980447
#> 199 2024-12-18 17:35:10 0.979619 0.980446
#> 200 2024-12-18 17:35:10 0.979619 0.980440
#> 201 2024-12-18 17:35:10 0.979619 0.980439
#> 202 2024-12-18 17:35:10 0.979620 0.980439
#> 203 2024-12-18 17:35:10 0.979621 0.980439
#> 204 2024-12-18 17:35:10 0.979621 0.980448
#> 205 2024-12-18 17:35:10 0.979710 0.980447
#> 206 2024-12-18 17:35:10 0.979621 0.980447
#> 207 2024-12-18 17:35:10 0.979620 0.980447
#> 208 2024-12-18 17:35:10 0.979620 0.980447
#> 209 2024-12-18 17:35:10 0.979620 0.980447
#> 210 2024-12-18 17:35:10 0.979621 0.980446
#> 211 2024-12-18 17:35:10 0.979621 0.980445
#> 212 2024-12-18 17:35:10 0.979621 0.980444
#> 213 2024-12-18 17:35:10 0.979621 0.980443
#> 214 2024-12-18 17:35:11 0.979621 0.980390
#> 215 2024-12-18 17:35:11 0.979621 0.980389
#> 
#> $last
#> [1] 1734543311
#>