Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 17:52:48 1.4269 1.4285
#> 2   2024-12-18 17:52:48 1.4269 1.4284
#> 3   2024-12-18 17:52:48 1.4269 1.4283
#> 4   2024-12-18 17:52:48 1.4269 1.4287
#> 5   2024-12-18 17:52:49 1.4269 1.4285
#> 6   2024-12-18 17:52:49 1.4269 1.4284
#> 7   2024-12-18 17:52:49 1.4267 1.4284
#> 8   2024-12-18 17:52:49 1.4270 1.4284
#> 9   2024-12-18 17:52:49 1.4270 1.4283
#> 10  2024-12-18 17:52:49 1.4271 1.4283
#> 11  2024-12-18 17:52:50 1.4271 1.4282
#> 12  2024-12-18 17:52:50 1.4271 1.4281
#> 13  2024-12-18 17:52:50 1.4271 1.4280
#> 14  2024-12-18 17:52:50 1.4271 1.4283
#> 15  2024-12-18 17:52:50 1.4272 1.4282
#> 16  2024-12-18 17:52:50 1.4273 1.4282
#> 17  2024-12-18 17:52:50 1.4274 1.4282
#> 18  2024-12-18 17:52:50 1.4274 1.4289
#> 19  2024-12-18 17:52:50 1.4274 1.4287
#> 20  2024-12-18 17:52:50 1.4274 1.4286
#> 21  2024-12-18 17:52:51 1.4274 1.4285
#> 22  2024-12-18 17:52:51 1.4274 1.4284
#> 23  2024-12-18 17:52:51 1.4274 1.4283
#> 24  2024-12-18 17:52:51 1.4274 1.4282
#> 25  2024-12-18 17:52:51 1.4274 1.4281
#> 26  2024-12-18 17:52:51 1.4274 1.4280
#> 27  2024-12-18 17:52:51 1.4274 1.4283
#> 28  2024-12-18 17:52:51 1.4274 1.4286
#> 29  2024-12-18 17:52:51 1.4274 1.4285
#> 30  2024-12-18 17:52:52 1.4274 1.4284
#> 31  2024-12-18 17:52:52 1.4275 1.4284
#> 32  2024-12-18 17:52:52 1.4275 1.4283
#> 33  2024-12-18 17:52:52 1.4276 1.4283
#> 34  2024-12-18 17:52:52 1.4276 1.4282
#> 35  2024-12-18 17:52:52 1.4276 1.4281
#> 36  2024-12-18 17:52:52 1.4276 1.4280
#> 37  2024-12-18 17:52:52 1.4276 1.4283
#> 38  2024-12-18 17:52:53 1.4276 1.4281
#> 39  2024-12-18 17:52:53 1.4276 1.4280
#> 40  2024-12-18 17:52:53 1.4273 1.4280
#> 41  2024-12-18 17:52:53 1.4274 1.4279
#> 42  2024-12-18 17:52:53 1.4274 1.4278
#> 43  2024-12-18 17:52:53 1.4274 1.4280
#> 44  2024-12-18 17:52:53 1.4274 1.4279
#> 45  2024-12-18 17:52:53 1.4274 1.4278
#> 46  2024-12-18 17:52:53 1.4273 1.4278
#> 47  2024-12-18 17:52:53 1.4270 1.4278
#> 48  2024-12-18 17:52:54 1.4270 1.4277
#> 49  2024-12-18 17:52:54 1.4270 1.4276
#> 50  2024-12-18 17:52:54 1.4271 1.4276
#> 51  2024-12-18 17:52:54 1.4271 1.4277
#> 52  2024-12-18 17:52:54 1.4272 1.4277
#> 53  2024-12-18 17:52:54 1.4272 1.4276
#> 54  2024-12-18 17:52:54 1.4272 1.4275
#> 55  2024-12-18 17:52:54 1.4271 1.4275
#> 56  2024-12-18 17:52:54 1.4271 1.4274
#> 57  2024-12-18 17:52:54 1.4270 1.4274
#> 58  2024-12-18 17:52:54 1.4261 1.4274
#> 59  2024-12-18 17:52:54 1.4261 1.4274
#> 60  2024-12-18 17:52:54 1.4262 1.4273
#> 61  2024-12-18 17:52:54 1.4265 1.4273
#> 62  2024-12-18 17:52:54 1.4262 1.4273
#> 63  2024-12-18 17:52:54 1.4266 1.4273
#> 64  2024-12-18 17:52:54 1.4262 1.4273
#> 65  2024-12-18 17:52:54 1.4263 1.4272
#> 66  2024-12-18 17:52:54 1.4263 1.4277
#> 67  2024-12-18 17:52:54 1.4263 1.4277
#> 68  2024-12-18 17:52:54 1.4258 1.4276
#> 69  2024-12-18 17:52:54 1.4260 1.4276
#> 70  2024-12-18 17:52:54 1.4261 1.4276
#> 71  2024-12-18 17:52:55 1.4261 1.4275
#> 72  2024-12-18 17:52:55 1.4262 1.4275
#> 73  2024-12-18 17:52:55 1.4263 1.4275
#> 74  2024-12-18 17:52:55 1.4261 1.4275
#> 75  2024-12-18 17:52:55 1.4260 1.4275
#> 76  2024-12-18 17:52:55 1.4260 1.4275
#> 77  2024-12-18 17:52:55 1.4261 1.4274
#> 78  2024-12-18 17:52:55 1.4261 1.4273
#> 79  2024-12-18 17:52:55 1.4261 1.4272
#> 80  2024-12-18 17:52:55 1.4262 1.4272
#> 81  2024-12-18 17:52:55 1.4263 1.4272
#> 82  2024-12-18 17:52:55 1.4263 1.4271
#> 83  2024-12-18 17:52:55 1.4263 1.4270
#> 84  2024-12-18 17:52:55 1.4263 1.4269
#> 85  2024-12-18 17:52:55 1.4263 1.4270
#> 86  2024-12-18 17:52:55 1.4263 1.4276
#> 87  2024-12-18 17:52:55 1.4263 1.4275
#> 88  2024-12-18 17:52:56 1.4263 1.4274
#> 89  2024-12-18 17:52:56 1.4264 1.4274
#> 90  2024-12-18 17:52:56 1.4264 1.4273
#> 91  2024-12-18 17:52:56 1.4265 1.4273
#> 92  2024-12-18 17:52:56 1.4265 1.4272
#> 93  2024-12-18 17:52:56 1.4265 1.4271
#> 94  2024-12-18 17:52:56 1.4265 1.4274
#> 95  2024-12-18 17:52:56 1.4265 1.4272
#> 96  2024-12-18 17:52:57 1.4265 1.4271
#> 97  2024-12-18 17:52:57 1.4265 1.4270
#> 98  2024-12-18 17:52:57 1.4265 1.4273
#> 99  2024-12-18 17:52:58 1.4265 1.4272
#> 100 2024-12-18 17:52:58 1.4265 1.4271
#> 101 2024-12-18 17:52:58 1.4265 1.4270
#> 102 2024-12-18 17:52:58 1.4265 1.4269
#> 103 2024-12-18 17:52:58 1.4265 1.4272
#> 104 2024-12-18 17:52:58 1.4266 1.4272
#> 105 2024-12-18 17:52:58 1.4266 1.4271
#> 106 2024-12-18 17:52:58 1.4267 1.4271
#> 107 2024-12-18 17:52:58 1.4267 1.4270
#> 108 2024-12-18 17:52:58 1.4266 1.4270
#> 109 2024-12-18 17:52:58 1.4267 1.4270
#> 110 2024-12-18 17:52:58 1.4267 1.4271
#> 111 2024-12-18 17:52:59 1.4260 1.4271
#> 112 2024-12-18 17:52:59 1.4261 1.4270
#> 113 2024-12-18 17:52:59 1.4263 1.4270
#> 114 2024-12-18 17:52:59 1.4264 1.4270
#> 115 2024-12-18 17:52:59 1.4260 1.4270
#> 116 2024-12-18 17:52:59 1.4261 1.4269
#> 117 2024-12-18 17:52:59 1.4261 1.4268
#> 118 2024-12-18 17:52:59 1.4261 1.4271
#> 119 2024-12-18 17:53:00 1.4261 1.4268
#> 120 2024-12-18 17:53:00 1.4261 1.4267
#> 121 2024-12-18 17:53:00 1.4262 1.4267
#> 122 2024-12-18 17:53:00 1.4262 1.4271
#> 123 2024-12-18 17:53:00 1.4263 1.4270
#> 124 2024-12-18 17:53:00 1.4263 1.4267
#> 125 2024-12-18 17:53:00 1.4263 1.4266
#> 126 2024-12-18 17:53:00 1.4262 1.4266
#> 127 2024-12-18 17:53:00 1.4262 1.4270
#> 128 2024-12-18 17:53:00 1.4263 1.4269
#> 129 2024-12-18 17:53:00 1.4263 1.4266
#> 130 2024-12-18 17:53:00 1.4262 1.4266
#> 131 2024-12-18 17:53:00 1.4262 1.4269
#> 132 2024-12-18 17:53:00 1.4262 1.4270
#> 133 2024-12-18 17:53:00 1.4262 1.4270
#> 134 2024-12-18 17:53:00 1.4263 1.4269
#> 135 2024-12-18 17:53:00 1.4263 1.4275
#> 136 2024-12-18 17:53:00 1.4263 1.4269
#> 137 2024-12-18 17:53:00 1.4263 1.4268
#> 138 2024-12-18 17:53:01 1.4263 1.4267
#> 139 2024-12-18 17:53:01 1.4263 1.4266
#> 140 2024-12-18 17:53:01 1.4262 1.4266
#> 141 2024-12-18 17:53:01 1.4262 1.4265
#> 142 2024-12-18 17:53:01 1.4263 1.4265
#> 143 2024-12-18 17:53:01 1.4263 1.4266
#> 144 2024-12-18 17:53:02 1.4263 1.4268
#> 145 2024-12-18 17:53:02 1.4263 1.4267
#> 146 2024-12-18 17:53:02 1.4263 1.4268
#> 147 2024-12-18 17:53:02 1.4264 1.4268
#> 148 2024-12-18 17:53:02 1.4265 1.4268
#> 149 2024-12-18 17:53:02 1.4265 1.4270
#> 150 2024-12-18 17:53:02 1.4265 1.4269
#> 151 2024-12-18 17:53:03 1.4265 1.4268
#> 152 2024-12-18 17:53:03 1.4265 1.4270
#> 153 2024-12-18 17:53:03 1.4265 1.4269
#> 154 2024-12-18 17:53:03 1.4265 1.4273
#> 155 2024-12-18 17:53:03 1.4265 1.4280
#> 156 2024-12-18 17:53:03 1.4265 1.4280
#> 157 2024-12-18 17:53:03 1.4265 1.4278
#> 158 2024-12-18 17:53:03 1.4265 1.4277
#> 159 2024-12-18 17:53:03 1.4266 1.4277
#> 160 2024-12-18 17:53:03 1.4266 1.4276
#> 161 2024-12-18 17:53:03 1.4266 1.4275
#> 162 2024-12-18 17:53:03 1.4266 1.4277
#> 163 2024-12-18 17:53:03 1.4266 1.4276
#> 164 2024-12-18 17:53:04 1.4267 1.4276
#> 165 2024-12-18 17:53:04 1.4268 1.4276
#> 166 2024-12-18 17:53:04 1.4268 1.4275
#> 167 2024-12-18 17:53:04 1.4268 1.4274
#> 168 2024-12-18 17:53:05 1.4268 1.4275
#> 169 2024-12-18 17:53:05 1.4265 1.4275
#> 170 2024-12-18 17:53:05 1.4269 1.4275
#> 171 2024-12-18 17:53:05 1.4270 1.4275
#> 172 2024-12-18 17:53:05 1.4265 1.4275
#> 173 2024-12-18 17:53:06 1.4267 1.4275
#> 174 2024-12-18 17:53:06 1.4268 1.4275
#> 175 2024-12-18 17:53:06 1.4268 1.4274
#> 176 2024-12-18 17:53:06 1.4268 1.4275
#> 177 2024-12-18 17:53:08 1.4269 1.4275
#> 178 2024-12-18 17:53:08 1.4270 1.4275
#> 179 2024-12-18 17:53:08 1.4270 1.4276
#> 180 2024-12-18 17:53:08 1.4270 1.4280
#> 181 2024-12-18 17:53:08 1.4270 1.4279
#> 182 2024-12-18 17:53:08 1.4270 1.4278
#> 183 2024-12-18 17:53:08 1.4270 1.4277
#> 184 2024-12-18 17:53:08 1.4270 1.4276
#> 185 2024-12-18 17:53:08 1.4271 1.4276
#> 186 2024-12-18 17:53:08 1.4272 1.4276
#> 187 2024-12-18 17:53:09 1.4272 1.4286
#> 188 2024-12-18 17:53:09 1.4272 1.4289
#> 189 2024-12-18 17:53:09 1.4273 1.4289
#> 190 2024-12-18 17:53:09 1.4273 1.4288
#> 191 2024-12-18 17:53:09 1.4274 1.4288
#> 192 2024-12-18 17:53:09 1.4274 1.4289
#> 193 2024-12-18 17:53:09 1.4274 1.4288
#> 194 2024-12-18 17:53:09 1.4272 1.4288
#> 195 2024-12-18 17:53:09 1.4273 1.4288
#> 196 2024-12-18 17:53:09 1.4273 1.4287
#> 197 2024-12-18 17:53:09 1.4273 1.4286
#> 198 2024-12-18 17:53:09 1.4273 1.4285
#> 199 2024-12-18 17:53:09 1.4271 1.4285
#> 200 2024-12-18 17:53:09 1.4272 1.4289
#> 201 2024-12-18 17:53:09 1.4269 1.4288
#> 202 2024-12-18 17:53:09 1.4270 1.4287
#> 203 2024-12-18 17:53:09 1.4270 1.4286
#> 204 2024-12-18 17:53:09 1.4270 1.4285
#> 205 2024-12-18 17:53:09 1.4270 1.4284
#> 206 2024-12-18 17:53:09 1.4270 1.4283
#> 207 2024-12-18 17:53:09 1.4271 1.4283
#> 208 2024-12-18 17:53:09 1.4272 1.4283
#> 209 2024-12-18 17:53:09 1.4272 1.4287
#> 210 2024-12-18 17:53:09 1.4272 1.4289
#> 211 2024-12-18 17:53:09 1.4272 1.4287
#> 212 2024-12-18 17:53:09 1.4272 1.4286
#> 213 2024-12-18 17:53:10 1.4273 1.4286
#> 214 2024-12-18 17:53:10 1.4273 1.4285
#> 215 2024-12-18 17:53:10 1.4274 1.4285
#> 216 2024-12-18 17:53:10 1.4274 1.4284
#> 217 2024-12-18 17:53:10 1.4274 1.4283
#> 218 2024-12-18 17:53:10 1.4274 1.4282
#> 219 2024-12-18 17:53:10 1.4274 1.4281
#> 220 2024-12-18 17:53:10 1.4274 1.4284
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 17:52:00 0.992638 0.993136
#> 2   2024-12-18 17:52:01 0.992639 0.993136
#> 3   2024-12-18 17:52:01 0.992640 0.993136
#> 4   2024-12-18 17:52:01 0.992641 0.993136
#> 5   2024-12-18 17:52:01 0.992642 0.993136
#> 6   2024-12-18 17:52:01 0.992643 0.993136
#> 7   2024-12-18 17:52:01 0.992644 0.993136
#> 8   2024-12-18 17:52:01 0.992643 0.993136
#> 9   2024-12-18 17:52:01 0.992646 0.993136
#> 10  2024-12-18 17:52:01 0.992647 0.993136
#> 11  2024-12-18 17:52:01 0.992648 0.993136
#> 12  2024-12-18 17:52:01 0.992639 0.993136
#> 13  2024-12-18 17:52:01 0.992639 0.993136
#> 14  2024-12-18 17:52:01 0.992639 0.993136
#> 15  2024-12-18 17:52:01 0.992641 0.993136
#> 16  2024-12-18 17:52:01 0.992642 0.993136
#> 17  2024-12-18 17:52:01 0.992643 0.993136
#> 18  2024-12-18 17:52:01 0.992644 0.993136
#> 19  2024-12-18 17:52:01 0.992645 0.993136
#> 20  2024-12-18 17:52:01 0.992646 0.993136
#> 21  2024-12-18 17:52:01 0.992647 0.993136
#> 22  2024-12-18 17:52:01 0.992648 0.993136
#> 23  2024-12-18 17:52:01 0.992647 0.993136
#> 24  2024-12-18 17:52:01 0.992647 0.993136
#> 25  2024-12-18 17:52:01 0.992649 0.993136
#> 26  2024-12-18 17:52:01 0.992650 0.993136
#> 27  2024-12-18 17:52:01 0.992651 0.993136
#> 28  2024-12-18 17:52:01 0.992652 0.993136
#> 29  2024-12-18 17:52:01 0.992653 0.993136
#> 30  2024-12-18 17:52:01 0.992654 0.993136
#> 31  2024-12-18 17:52:02 0.992655 0.993136
#> 32  2024-12-18 17:52:02 0.992656 0.993136
#> 33  2024-12-18 17:52:02 0.992657 0.993136
#> 34  2024-12-18 17:52:02 0.992658 0.993136
#> 35  2024-12-18 17:52:02 0.992659 0.993136
#> 36  2024-12-18 17:52:02 0.992660 0.993136
#> 37  2024-12-18 17:52:02 0.992661 0.993136
#> 38  2024-12-18 17:52:02 0.992660 0.993136
#> 39  2024-12-18 17:52:02 0.993135 0.993224
#> 40  2024-12-18 17:52:02 0.993136 0.993224
#> 41  2024-12-18 17:52:02 0.993135 0.993224
#> 42  2024-12-18 17:52:02 0.993136 0.993224
#> 43  2024-12-18 17:52:02 0.993137 0.993224
#> 44  2024-12-18 17:52:02 0.993138 0.993224
#> 45  2024-12-18 17:52:02 0.993137 0.993224
#> 46  2024-12-18 17:52:02 0.993137 0.993224
#> 47  2024-12-18 17:52:02 0.993135 0.993224
#> 48  2024-12-18 17:52:02 0.993135 0.993224
#> 49  2024-12-18 17:52:02 0.993136 0.993224
#> 50  2024-12-18 17:52:03 0.993137 0.993224
#> 51  2024-12-18 17:52:03 0.993138 0.993224
#> 52  2024-12-18 17:52:03 0.993053 0.993224
#> 53  2024-12-18 17:52:03 0.993052 0.993224
#> 54  2024-12-18 17:52:03 0.993053 0.993224
#> 55  2024-12-18 17:52:03 0.993052 0.993224
#> 56  2024-12-18 17:52:03 0.993053 0.993224
#> 57  2024-12-18 17:52:03 0.993054 0.993224
#> 58  2024-12-18 17:52:03 0.993055 0.993224
#> 59  2024-12-18 17:52:03 0.993054 0.993224
#> 60  2024-12-18 17:52:03 0.993053 0.993224
#> 61  2024-12-18 17:52:04 0.993054 0.993224
#> 62  2024-12-18 17:52:04 0.993053 0.993224
#> 63  2024-12-18 17:52:04 0.993054 0.993224
#> 64  2024-12-18 17:52:05 0.993053 0.993224
#> 65  2024-12-18 17:52:06 0.993054 0.993224
#> 66  2024-12-18 17:52:06 0.993053 0.993224
#> 67  2024-12-18 17:52:07 0.993054 0.993224
#> 68  2024-12-18 17:52:07 0.993055 0.993224
#> 69  2024-12-18 17:52:07 0.993054 0.993224
#> 70  2024-12-18 17:52:08 0.993055 0.993224
#> 71  2024-12-18 17:52:08 0.993056 0.993224
#> 72  2024-12-18 17:52:08 0.993057 0.993224
#> 73  2024-12-18 17:52:08 0.993058 0.993224
#> 74  2024-12-18 17:52:08 0.993117 0.993224
#> 75  2024-12-18 17:52:08 0.993118 0.993224
#> 76  2024-12-18 17:52:08 0.993124 0.993224
#> 77  2024-12-18 17:52:08 0.993125 0.993224
#> 78  2024-12-18 17:52:08 0.993124 0.993224
#> 79  2024-12-18 17:52:14 0.993195 0.993224
#> 80  2024-12-18 17:52:14 0.993195 0.993224
#> 81  2024-12-18 17:52:14 0.993196 0.993224
#> 82  2024-12-18 17:52:14 0.993195 0.993224
#> 83  2024-12-18 17:52:14 0.993196 0.993224
#> 84  2024-12-18 17:52:14 0.993195 0.993224
#> 85  2024-12-18 17:52:14 0.993196 0.993224
#> 86  2024-12-18 17:52:14 0.993197 0.993224
#> 87  2024-12-18 17:52:14 0.993223 0.993224
#> 88  2024-12-18 17:52:17 0.993195 0.993224
#> 89  2024-12-18 17:52:17 0.993195 0.993224
#> 90  2024-12-18 17:52:17 0.993197 0.993224
#> 91  2024-12-18 17:52:17 0.993195 0.993224
#> 92  2024-12-18 17:52:17 0.993195 0.993224
#> 93  2024-12-18 17:52:17 0.993195 0.993224
#> 94  2024-12-18 17:52:18 0.993197 0.993224
#> 95  2024-12-18 17:52:18 0.993198 0.993224
#> 96  2024-12-18 17:52:18 0.993195 0.993224
#> 97  2024-12-18 17:52:18 0.993195 0.993224
#> 98  2024-12-18 17:52:19 0.993094 0.993224
#> 99  2024-12-18 17:52:19 0.992911 0.993224
#> 100 2024-12-18 17:52:19 0.992910 0.993224
#> 101 2024-12-18 17:52:19 0.992911 0.993224
#> 102 2024-12-18 17:52:20 0.992912 0.993224
#> 103 2024-12-18 17:52:20 0.992913 0.993224
#> 104 2024-12-18 17:52:20 0.992914 0.993224
#> 105 2024-12-18 17:52:20 0.992961 0.993224
#> 106 2024-12-18 17:52:20 0.992962 0.993224
#> 107 2024-12-18 17:52:20 0.992968 0.993224
#> 108 2024-12-18 17:52:20 0.992969 0.993224
#> 109 2024-12-18 17:52:20 0.992975 0.993224
#> 110 2024-12-18 17:52:20 0.992976 0.993224
#> 111 2024-12-18 17:52:20 0.992982 0.993224
#> 112 2024-12-18 17:52:20 0.992983 0.993224
#> 113 2024-12-18 17:52:20 0.992982 0.993224
#> 114 2024-12-18 17:52:20 0.992985 0.993224
#> 115 2024-12-18 17:52:20 0.992986 0.993224
#> 116 2024-12-18 17:52:21 0.992992 0.993224
#> 117 2024-12-18 17:52:21 0.992993 0.993224
#> 118 2024-12-18 17:52:21 0.992999 0.993224
#> 119 2024-12-18 17:52:21 0.993000 0.993224
#> 120 2024-12-18 17:52:21 0.992999 0.993224
#> 121 2024-12-18 17:52:46 0.993030 0.993224
#> 122 2024-12-18 17:52:47 0.993124 0.993224
#> 123 2024-12-18 17:52:47 0.993125 0.993224
#> 124 2024-12-18 17:52:47 0.993223 0.993224
#> 125 2024-12-18 17:52:53 0.993215 0.993224
#> 126 2024-12-18 17:52:53 0.993214 0.993224
#> 127 2024-12-18 17:52:55 0.993215 0.993224
#> 128 2024-12-18 17:52:55 0.993223 0.993224
#> 129 2024-12-18 17:52:55 0.993215 0.993224
#> 130 2024-12-18 17:53:00 0.993159 0.993224
#> 131 2024-12-18 17:53:00 0.993009 0.993224
#> 132 2024-12-18 17:53:00 0.992779 0.993224
#> 133 2024-12-18 17:53:00 0.992787 0.993224
#> 134 2024-12-18 17:53:00 0.992888 0.993224
#> 135 2024-12-18 17:53:00 0.992867 0.993224
#> 136 2024-12-18 17:53:00 0.992779 0.993224
#> 137 2024-12-18 17:53:00 0.992768 0.993224
#> 138 2024-12-18 17:53:00 0.993010 0.993224
#> 139 2024-12-18 17:53:00 0.992789 0.993224
#> 140 2024-12-18 17:53:00 0.992868 0.993224
#> 141 2024-12-18 17:53:00 0.992795 0.993224
#> 142 2024-12-18 17:53:00 0.993011 0.993224
#> 143 2024-12-18 17:53:00 0.992795 0.993224
#> 144 2024-12-18 17:53:00 0.992794 0.993224
#> 145 2024-12-18 17:53:00 0.992869 0.993224
#> 146 2024-12-18 17:53:00 0.992794 0.993224
#> 147 2024-12-18 17:53:01 0.992794 0.993223
#> 148 2024-12-18 17:53:01 0.992770 0.993223
#> 149 2024-12-18 17:53:01 0.992770 0.993224
#> 150 2024-12-18 17:53:01 0.992770 0.993223
#> 151 2024-12-18 17:53:01 0.992731 0.993223
#> 152 2024-12-18 17:53:01 0.992583 0.993223
#> 153 2024-12-18 17:53:01 0.992585 0.993223
#> 154 2024-12-18 17:53:01 0.992585 0.993222
#> 155 2024-12-18 17:53:01 0.992586 0.993222
#> 156 2024-12-18 17:53:01 0.992586 0.993223
#> 157 2024-12-18 17:53:01 0.992587 0.993221
#> 158 2024-12-18 17:53:01 0.992587 0.993223
#> 159 2024-12-18 17:53:01 0.992588 0.993220
#> 160 2024-12-18 17:53:01 0.992589 0.993220
#> 161 2024-12-18 17:53:01 0.992589 0.993219
#> 162 2024-12-18 17:53:01 0.992589 0.993220
#> 163 2024-12-18 17:53:01 0.992589 0.993220
#> 164 2024-12-18 17:53:01 0.992589 0.993220
#> 165 2024-12-18 17:53:01 0.992596 0.993219
#> 166 2024-12-18 17:53:01 0.992597 0.993219
#> 167 2024-12-18 17:53:01 0.992602 0.993219
#> 168 2024-12-18 17:53:01 0.992610 0.993219
#> 169 2024-12-18 17:53:01 0.992611 0.993219
#> 170 2024-12-18 17:53:01 0.992612 0.993219
#> 171 2024-12-18 17:53:01 0.992613 0.993219
#> 172 2024-12-18 17:53:01 0.992614 0.993219
#> 173 2024-12-18 17:53:01 0.992626 0.993219
#> 174 2024-12-18 17:53:01 0.992627 0.993219
#> 175 2024-12-18 17:53:01 0.992685 0.993219
#> 176 2024-12-18 17:53:01 0.992686 0.993219
#> 177 2024-12-18 17:53:02 0.992686 0.993224
#> 178 2024-12-18 17:53:02 0.992968 0.993224
#> 179 2024-12-18 17:53:03 0.992974 0.993224
#> 180 2024-12-18 17:53:03 0.992975 0.993224
#> 181 2024-12-18 17:53:03 0.992968 0.993224
#> 182 2024-12-18 17:53:03 0.992976 0.993224
#> 183 2024-12-18 17:53:03 0.992968 0.993224
#> 184 2024-12-18 17:53:04 0.992970 0.993224
#> 185 2024-12-18 17:53:04 0.992971 0.993224
#> 186 2024-12-18 17:53:04 0.992976 0.993224
#> 187 2024-12-18 17:53:04 0.992977 0.993224
#> 188 2024-12-18 17:53:04 0.992979 0.993224
#> 189 2024-12-18 17:53:04 0.992980 0.993224
#> 190 2024-12-18 17:53:04 0.992983 0.993224
#> 191 2024-12-18 17:53:04 0.992984 0.993224
#> 192 2024-12-18 17:53:04 0.992985 0.993224
#> 193 2024-12-18 17:53:04 0.992984 0.993224
#> 194 2024-12-18 17:53:04 0.992984 0.993224
#> 195 2024-12-18 17:53:04 0.992991 0.993224
#> 196 2024-12-18 17:53:04 0.992992 0.993224
#> 197 2024-12-18 17:53:04 0.992993 0.993224
#> 198 2024-12-18 17:53:04 0.992999 0.993224
#> 199 2024-12-18 17:53:04 0.993000 0.993224
#> 200 2024-12-18 17:53:04 0.993001 0.993224
#> 201 2024-12-18 17:53:05 0.993007 0.993224
#> 202 2024-12-18 17:53:05 0.993008 0.993224
#> 203 2024-12-18 17:53:05 0.993009 0.993224
#> 204 2024-12-18 17:53:05 0.993015 0.993224
#> 205 2024-12-18 17:53:05 0.993016 0.993224
#> 206 2024-12-18 17:53:05 0.992992 0.993224
#> 207 2024-12-18 17:53:05 0.992999 0.993224
#> 208 2024-12-18 17:53:05 0.993000 0.993224
#> 209 2024-12-18 17:53:05 0.993006 0.993224
#> 210 2024-12-18 17:53:05 0.993007 0.993224
#> 211 2024-12-18 17:53:06 0.992992 0.993224
#> 212 2024-12-18 17:53:06 0.992999 0.993224
#> 213 2024-12-18 17:53:06 0.993000 0.993224
#> 214 2024-12-18 17:53:06 0.993006 0.993224
#> 215 2024-12-18 17:53:06 0.993007 0.993224
#> 216 2024-12-18 17:53:06 0.993006 0.993224
#> 217 2024-12-18 17:53:06 0.993006 0.993224
#> 218 2024-12-18 17:53:06 0.992992 0.993224
#> 219 2024-12-18 17:53:06 0.992999 0.993224
#> 220 2024-12-18 17:53:06 0.993000 0.993224
#> 221 2024-12-18 17:53:06 0.993001 0.993224
#> 222 2024-12-18 17:53:07 0.993007 0.993224
#> 223 2024-12-18 17:53:07 0.993009 0.993224
#> 224 2024-12-18 17:53:07 0.993010 0.993224
#> 225 2024-12-18 17:53:07 0.993059 0.993224
#> 226 2024-12-18 17:53:08 0.993002 0.993224
#> 227 2024-12-18 17:53:08 0.993003 0.993224
#> 228 2024-12-18 17:53:08 0.993002 0.993224
#> 229 2024-12-18 17:53:08 0.993002 0.993224
#> 230 2024-12-18 17:53:08 0.993002 0.993224
#> 231 2024-12-18 17:53:08 0.993003 0.993224
#> 232 2024-12-18 17:53:08 0.993004 0.993224
#> 233 2024-12-18 17:53:08 0.993010 0.993224
#> 234 2024-12-18 17:53:08 0.993011 0.993224
#> 235 2024-12-18 17:53:08 0.993017 0.993224
#> 236 2024-12-18 17:53:08 0.993018 0.993224
#> 237 2024-12-18 17:53:08 0.993019 0.993224
#> 238 2024-12-18 17:53:08 0.993020 0.993224
#> 239 2024-12-18 17:53:08 0.993021 0.993224
#> 240 2024-12-18 17:53:08 0.993025 0.993224
#> 241 2024-12-18 17:53:08 0.993026 0.993224
#> 242 2024-12-18 17:53:08 0.993027 0.993224
#> 243 2024-12-18 17:53:08 0.993223 0.993224
#> 
#> $last
#> [1] 1734544388
#>