Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 16:32:45 1.3939 1.3954
#> 2   2024-12-18 16:32:45 1.3939 1.3953
#> 3   2024-12-18 16:32:45 1.3939 1.3952
#> 4   2024-12-18 16:32:45 1.3939 1.3951
#> 5   2024-12-18 16:32:45 1.3939 1.3950
#> 6   2024-12-18 16:32:46 1.3939 1.3949
#> 7   2024-12-18 16:32:46 1.3939 1.3948
#> 8   2024-12-18 16:32:46 1.3939 1.3950
#> 9   2024-12-18 16:32:46 1.3939 1.3949
#> 10  2024-12-18 16:32:46 1.3939 1.3948
#> 11  2024-12-18 16:32:46 1.3940 1.3948
#> 12  2024-12-18 16:32:46 1.3940 1.3950
#> 13  2024-12-18 16:32:46 1.3938 1.3950
#> 14  2024-12-18 16:32:46 1.3941 1.3950
#> 15  2024-12-18 16:32:46 1.3941 1.3949
#> 16  2024-12-18 16:32:46 1.3942 1.3949
#> 17  2024-12-18 16:32:47 1.3942 1.3948
#> 18  2024-12-18 16:32:47 1.3938 1.3948
#> 19  2024-12-18 16:32:47 1.3934 1.3948
#> 20  2024-12-18 16:32:47 1.3936 1.3948
#> 21  2024-12-18 16:32:47 1.3939 1.3948
#> 22  2024-12-18 16:32:47 1.3940 1.3948
#> 23  2024-12-18 16:32:47 1.3940 1.3947
#> 24  2024-12-18 16:32:47 1.3936 1.3947
#> 25  2024-12-18 16:32:48 1.3937 1.3940
#> 26  2024-12-18 16:32:48 1.3935 1.3940
#> 27  2024-12-18 16:32:48 1.3936 1.3940
#> 28  2024-12-18 16:32:48 1.3936 1.3939
#> 29  2024-12-18 16:32:49 1.3936 1.3938
#> 30  2024-12-18 16:32:49 1.3935 1.3938
#> 31  2024-12-18 16:32:49 1.3935 1.3939
#> 32  2024-12-18 16:32:49 1.3936 1.3939
#> 33  2024-12-18 16:32:51 1.3937 1.3939
#> 34  2024-12-18 16:32:52 1.3937 1.3940
#> 35  2024-12-18 16:32:52 1.3937 1.3942
#> 36  2024-12-18 16:32:52 1.3937 1.3943
#> 37  2024-12-18 16:32:52 1.3938 1.3943
#> 38  2024-12-18 16:32:52 1.3938 1.3950
#> 39  2024-12-18 16:32:52 1.3938 1.3949
#> 40  2024-12-18 16:32:52 1.3939 1.3949
#> 41  2024-12-18 16:32:52 1.3939 1.3958
#> 42  2024-12-18 16:32:52 1.3939 1.3957
#> 43  2024-12-18 16:32:52 1.3940 1.3957
#> 44  2024-12-18 16:32:52 1.3940 1.3956
#> 45  2024-12-18 16:32:52 1.3938 1.3956
#> 46  2024-12-18 16:32:52 1.3939 1.3955
#> 47  2024-12-18 16:32:52 1.3939 1.3953
#> 48  2024-12-18 16:32:52 1.3939 1.3940
#> 49  2024-12-18 16:32:52 1.3939 1.3956
#> 50  2024-12-18 16:32:52 1.3940 1.3956
#> 51  2024-12-18 16:32:52 1.3940 1.3955
#> 52  2024-12-18 16:32:52 1.3940 1.3954
#> 53  2024-12-18 16:32:52 1.3938 1.3954
#> 54  2024-12-18 16:32:52 1.3939 1.3955
#> 55  2024-12-18 16:32:52 1.3939 1.3950
#> 56  2024-12-18 16:32:52 1.3940 1.3950
#> 57  2024-12-18 16:32:53 1.3941 1.3950
#> 58  2024-12-18 16:32:53 1.3942 1.3950
#> 59  2024-12-18 16:32:53 1.3943 1.3950
#> 60  2024-12-18 16:32:53 1.3944 1.3950
#> 61  2024-12-18 16:32:56 1.3945 1.3950
#> 62  2024-12-18 16:32:56 1.3946 1.3950
#> 63  2024-12-18 16:32:56 1.3947 1.3950
#> 64  2024-12-18 16:32:56 1.3945 1.3950
#> 65  2024-12-18 16:32:58 1.3947 1.3950
#> 66  2024-12-18 16:32:59 1.3948 1.3950
#> 67  2024-12-18 16:33:00 1.3947 1.3950
#> 68  2024-12-18 16:33:01 1.3948 1.3950
#> 69  2024-12-18 16:33:01 1.3949 1.3950
#> 70  2024-12-18 16:33:01 1.3948 1.3950
#> 71  2024-12-18 16:33:02 1.3949 1.3950
#> 72  2024-12-18 16:33:03 1.3949 1.3958
#> 73  2024-12-18 16:33:03 1.3950 1.3958
#> 74  2024-12-18 16:33:03 1.3950 1.3966
#> 75  2024-12-18 16:33:03 1.3951 1.3966
#> 76  2024-12-18 16:33:03 1.3951 1.3972
#> 77  2024-12-18 16:33:03 1.3952 1.3972
#> 78  2024-12-18 16:33:03 1.3952 1.3965
#> 79  2024-12-18 16:33:03 1.3952 1.3960
#> 80  2024-12-18 16:33:03 1.3950 1.3960
#> 81  2024-12-18 16:33:03 1.3951 1.3959
#> 82  2024-12-18 16:33:03 1.3951 1.3958
#> 83  2024-12-18 16:33:03 1.3949 1.3958
#> 84  2024-12-18 16:33:03 1.3949 1.3958
#> 85  2024-12-18 16:33:03 1.3949 1.3950
#> 86  2024-12-18 16:33:04 1.3948 1.3950
#> 87  2024-12-18 16:33:04 1.3946 1.3950
#> 88  2024-12-18 16:33:04 1.3943 1.3950
#> 89  2024-12-18 16:33:04 1.3947 1.3950
#> 90  2024-12-18 16:33:04 1.3944 1.3950
#> 91  2024-12-18 16:33:04 1.3946 1.3950
#> 92  2024-12-18 16:33:05 1.3947 1.3950
#> 93  2024-12-18 16:33:05 1.3944 1.3950
#> 94  2024-12-18 16:33:05 1.3945 1.3950
#> 95  2024-12-18 16:33:05 1.3945 1.3948
#> 96  2024-12-18 16:33:05 1.3939 1.3948
#> 97  2024-12-18 16:33:05 1.3939 1.3948
#> 98  2024-12-18 16:33:05 1.3939 1.3948
#> 99  2024-12-18 16:33:05 1.3940 1.3948
#> 100 2024-12-18 16:33:05 1.3940 1.3947
#> 101 2024-12-18 16:33:05 1.3940 1.3948
#> 102 2024-12-18 16:33:05 1.3940 1.3953
#> 103 2024-12-18 16:33:05 1.3940 1.3952
#> 104 2024-12-18 16:33:05 1.3940 1.3951
#> 105 2024-12-18 16:33:05 1.3940 1.3950
#> 106 2024-12-18 16:33:06 1.3940 1.3949
#> 107 2024-12-18 16:33:06 1.3940 1.3948
#> 108 2024-12-18 16:33:06 1.3940 1.3947
#> 109 2024-12-18 16:33:06 1.3940 1.3946
#> 110 2024-12-18 16:33:06 1.3940 1.3947
#> 111 2024-12-18 16:33:06 1.3940 1.3952
#> 112 2024-12-18 16:33:06 1.3940 1.3948
#> 113 2024-12-18 16:33:06 1.3940 1.3952
#> 114 2024-12-18 16:33:06 1.3940 1.3951
#> 115 2024-12-18 16:33:06 1.3940 1.3950
#> 116 2024-12-18 16:33:06 1.3940 1.3949
#> 117 2024-12-18 16:33:06 1.3940 1.3948
#> 118 2024-12-18 16:33:06 1.3941 1.3948
#> 119 2024-12-18 16:33:06 1.3941 1.3947
#> 120 2024-12-18 16:33:07 1.3942 1.3947
#> 121 2024-12-18 16:33:07 1.3941 1.3947
#> 122 2024-12-18 16:33:07 1.3941 1.3946
#> 123 2024-12-18 16:33:07 1.3941 1.3945
#> 124 2024-12-18 16:33:08 1.3941 1.3946
#> 125 2024-12-18 16:33:08 1.3942 1.3946
#> 126 2024-12-18 16:33:08 1.3942 1.3953
#> 127 2024-12-18 16:33:08 1.3942 1.3952
#> 128 2024-12-18 16:33:08 1.3942 1.3950
#> 129 2024-12-18 16:33:09 1.3943 1.3950
#> 130 2024-12-18 16:33:09 1.3944 1.3950
#> 131 2024-12-18 16:33:11 1.3945 1.3950
#> 132 2024-12-18 16:33:11 1.3946 1.3950
#> 133 2024-12-18 16:33:11 1.3946 1.3948
#> 134 2024-12-18 16:33:11 1.3940 1.3948
#> 135 2024-12-18 16:33:11 1.3939 1.3948
#> 136 2024-12-18 16:33:11 1.3940 1.3947
#> 137 2024-12-18 16:33:11 1.3939 1.3947
#> 138 2024-12-18 16:33:11 1.3939 1.3940
#> 139 2024-12-18 16:33:11 1.3939 1.3947
#> 140 2024-12-18 16:33:11 1.3939 1.3948
#> 141 2024-12-18 16:33:11 1.3939 1.3950
#> 142 2024-12-18 16:33:11 1.3940 1.3950
#> 143 2024-12-18 16:33:11 1.3941 1.3950
#> 144 2024-12-18 16:33:11 1.3942 1.3950
#> 145 2024-12-18 16:33:12 1.3939 1.3950
#> 146 2024-12-18 16:33:12 1.3940 1.3948
#> 147 2024-12-18 16:33:12 1.3940 1.3947
#> 148 2024-12-18 16:33:12 1.3940 1.3946
#> 149 2024-12-18 16:33:12 1.3940 1.3948
#> 150 2024-12-18 16:33:12 1.3940 1.3950
#> 151 2024-12-18 16:33:12 1.3940 1.3955
#> 152 2024-12-18 16:33:12 1.3940 1.3954
#> 153 2024-12-18 16:33:12 1.3941 1.3954
#> 154 2024-12-18 16:33:12 1.3941 1.3950
#> 155 2024-12-18 16:33:12 1.3942 1.3950
#> 156 2024-12-18 16:33:13 1.3943 1.3950
#> 157 2024-12-18 16:33:13 1.3944 1.3950
#> 158 2024-12-18 16:33:13 1.3945 1.3950
#> 159 2024-12-18 16:33:13 1.3946 1.3950
#> 160 2024-12-18 16:33:15 1.3943 1.3950
#> 161 2024-12-18 16:33:15 1.3947 1.3950
#> 162 2024-12-18 16:33:15 1.3944 1.3950
#> 163 2024-12-18 16:33:15 1.3945 1.3950
#> 164 2024-12-18 16:33:15 1.3946 1.3950
#> 165 2024-12-18 16:33:15 1.3947 1.3950
#> 166 2024-12-18 16:33:18 1.3948 1.3950
#> 167 2024-12-18 16:33:18 1.3949 1.3950
#> 168 2024-12-18 16:33:18 1.3948 1.3950
#> 169 2024-12-18 16:33:19 1.3949 1.3950
#> 170 2024-12-18 16:33:22 1.3948 1.3950
#> 171 2024-12-18 16:33:22 1.3949 1.3950
#> 172 2024-12-18 16:33:22 1.3946 1.3950
#> 173 2024-12-18 16:33:22 1.3947 1.3950
#> 174 2024-12-18 16:33:22 1.3946 1.3950
#> 175 2024-12-18 16:33:22 1.3947 1.3950
#> 176 2024-12-18 16:33:24 1.3944 1.3950
#> 177 2024-12-18 16:33:24 1.3945 1.3950
#> 178 2024-12-18 16:33:24 1.3945 1.3949
#> 179 2024-12-18 16:33:24 1.3940 1.3949
#> 180 2024-12-18 16:33:24 1.3942 1.3949
#> 181 2024-12-18 16:33:24 1.3943 1.3949
#> 182 2024-12-18 16:33:26 1.3942 1.3949
#> 183 2024-12-18 16:33:26 1.3944 1.3949
#> 184 2024-12-18 16:33:26 1.3945 1.3949
#> 185 2024-12-18 16:33:27 1.3945 1.3950
#> 186 2024-12-18 16:33:27 1.3945 1.3949
#> 187 2024-12-18 16:33:27 1.3945 1.3950
#> 188 2024-12-18 16:33:28 1.3946 1.3950
#> 189 2024-12-18 16:33:29 1.3947 1.3950
#> 190 2024-12-18 16:33:29 1.3948 1.3950
#> 191 2024-12-18 16:33:29 1.3948 1.3954
#> 192 2024-12-18 16:33:29 1.3948 1.3957
#> 193 2024-12-18 16:33:29 1.3948 1.3959
#> 194 2024-12-18 16:33:29 1.3948 1.3956
#> 195 2024-12-18 16:33:29 1.3948 1.3959
#> 196 2024-12-18 16:33:29 1.3948 1.3958
#> 197 2024-12-18 16:33:29 1.3949 1.3958
#> 198 2024-12-18 16:33:29 1.3949 1.3960
#> 199 2024-12-18 16:33:29 1.3950 1.3960
#> 200 2024-12-18 16:33:30 1.3951 1.3960
#> 201 2024-12-18 16:33:30 1.3951 1.3963
#> 202 2024-12-18 16:33:30 1.3951 1.3964
#> 203 2024-12-18 16:33:30 1.3951 1.3969
#> 204 2024-12-18 16:33:30 1.3951 1.3968
#> 205 2024-12-18 16:33:30 1.3951 1.3963
#> 206 2024-12-18 16:33:30 1.3951 1.3961
#> 207 2024-12-18 16:33:30 1.3952 1.3961
#> 208 2024-12-18 16:33:30 1.3952 1.3968
#> 209 2024-12-18 16:33:30 1.3953 1.3968
#> 210 2024-12-18 16:33:31 1.3953 1.3972
#> 211 2024-12-18 16:33:31 1.3953 1.3970
#> 212 2024-12-18 16:33:31 1.3954 1.3970
#> 213 2024-12-18 16:33:31 1.3955 1.3970
#> 214 2024-12-18 16:33:32 1.3956 1.3970
#> 215 2024-12-18 16:33:32 1.3957 1.3970
#> 216 2024-12-18 16:33:32 1.3957 1.3969
#> 217 2024-12-18 16:33:32 1.3957 1.3961
#> 218 2024-12-18 16:33:35 1.3958 1.3961
#> 219 2024-12-18 16:33:35 1.3959 1.3961
#> 220 2024-12-18 16:33:35 1.3960 1.3961
#> 221 2024-12-18 16:33:37 1.3960 1.3968
#> 222 2024-12-18 16:33:37 1.3960 1.3971
#> 223 2024-12-18 16:33:37 1.3960 1.3970
#> 224 2024-12-18 16:33:37 1.3960 1.3974
#> 225 2024-12-18 16:33:37 1.3960 1.3973
#> 226 2024-12-18 16:33:37 1.3961 1.3973
#> 227 2024-12-18 16:33:37 1.3962 1.3973
#> 228 2024-12-18 16:33:37 1.3963 1.3973
#> 229 2024-12-18 16:33:37 1.3963 1.3972
#> 230 2024-12-18 16:33:37 1.3963 1.3971
#> 231 2024-12-18 16:33:38 1.3963 1.3969
#> 232 2024-12-18 16:33:38 1.3964 1.3969
#> 233 2024-12-18 16:33:38 1.3965 1.3969
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 16:33:24 0.971031 0.971839
#> 2   2024-12-18 16:33:24 0.971033 0.971840
#> 3   2024-12-18 16:33:24 0.971040 0.971832
#> 4   2024-12-18 16:33:24 0.971040 0.971840
#> 5   2024-12-18 16:33:24 0.971040 0.971840
#> 6   2024-12-18 16:33:24 0.971041 0.971831
#> 7   2024-12-18 16:33:24 0.971041 0.971840
#> 8   2024-12-18 16:33:24 0.971041 0.971840
#> 9   2024-12-18 16:33:24 0.971041 0.971840
#> 10  2024-12-18 16:33:24 0.971041 0.971840
#> 11  2024-12-18 16:33:24 0.971041 0.971840
#> 12  2024-12-18 16:33:24 0.971041 0.971840
#> 13  2024-12-18 16:33:25 0.971042 0.971839
#> 14  2024-12-18 16:33:25 0.971201 0.971839
#> 15  2024-12-18 16:33:25 0.971201 0.971833
#> 16  2024-12-18 16:33:25 0.971201 0.971832
#> 17  2024-12-18 16:33:25 0.971201 0.971778
#> 18  2024-12-18 16:33:25 0.971201 0.971777
#> 19  2024-12-18 16:33:25 0.971041 0.971777
#> 20  2024-12-18 16:33:25 0.971041 0.971776
#> 21  2024-12-18 16:33:25 0.971041 0.971775
#> 22  2024-12-18 16:33:25 0.971041 0.971774
#> 23  2024-12-18 16:33:25 0.971041 0.971773
#> 24  2024-12-18 16:33:25 0.971041 0.971719
#> 25  2024-12-18 16:33:25 0.971041 0.971718
#> 26  2024-12-18 16:33:25 0.971041 0.971717
#> 27  2024-12-18 16:33:25 0.971041 0.971712
#> 28  2024-12-18 16:33:25 0.971042 0.971712
#> 29  2024-12-18 16:33:25 0.971042 0.971711
#> 30  2024-12-18 16:33:25 0.971042 0.971705
#> 31  2024-12-18 16:33:25 0.971042 0.971657
#> 32  2024-12-18 16:33:25 0.971043 0.971657
#> 33  2024-12-18 16:33:25 0.971044 0.971657
#> 34  2024-12-18 16:33:25 0.971044 0.971656
#> 35  2024-12-18 16:33:26 0.971044 0.971655
#> 36  2024-12-18 16:33:26 0.971044 0.971654
#> 37  2024-12-18 16:33:26 0.971044 0.971630
#> 38  2024-12-18 16:33:26 0.971044 0.971629
#> 39  2024-12-18 16:33:26 0.971422 0.971629
#> 40  2024-12-18 16:33:27 0.971422 0.971628
#> 41  2024-12-18 16:33:27 0.971422 0.971627
#> 42  2024-12-18 16:33:27 0.971227 0.971627
#> 43  2024-12-18 16:33:27 0.970901 0.971627
#> 44  2024-12-18 16:33:27 0.970902 0.971627
#> 45  2024-12-18 16:33:27 0.970903 0.971627
#> 46  2024-12-18 16:33:27 0.970904 0.971627
#> 47  2024-12-18 16:33:27 0.970900 0.971627
#> 48  2024-12-18 16:33:27 0.970902 0.971627
#> 49  2024-12-18 16:33:27 0.970903 0.971627
#> 50  2024-12-18 16:33:27 0.970904 0.971627
#> 51  2024-12-18 16:33:27 0.970905 0.971627
#> 52  2024-12-18 16:33:27 0.970906 0.971627
#> 53  2024-12-18 16:33:28 0.971412 0.971627
#> 54  2024-12-18 16:33:28 0.971413 0.971627
#> 55  2024-12-18 16:33:29 0.971413 0.971924
#> 56  2024-12-18 16:33:29 0.971414 0.971923
#> 57  2024-12-18 16:33:29 0.971414 0.971924
#> 58  2024-12-18 16:33:29 0.971414 0.971922
#> 59  2024-12-18 16:33:29 0.971413 0.971922
#> 60  2024-12-18 16:33:29 0.971413 0.971625
#> 61  2024-12-18 16:33:29 0.971415 0.971625
#> 62  2024-12-18 16:33:29 0.971415 0.971922
#> 63  2024-12-18 16:33:29 0.971413 0.971921
#> 64  2024-12-18 16:33:29 0.971413 0.971921
#> 65  2024-12-18 16:33:29 0.971592 0.971921
#> 66  2024-12-18 16:33:29 0.971593 0.971921
#> 67  2024-12-18 16:33:30 0.971593 0.971920
#> 68  2024-12-18 16:33:30 0.971593 0.971921
#> 69  2024-12-18 16:33:30 0.971593 0.971922
#> 70  2024-12-18 16:33:30 0.971593 0.971922
#> 71  2024-12-18 16:33:30 0.971593 0.971922
#> 72  2024-12-18 16:33:30 0.971594 0.971922
#> 73  2024-12-18 16:33:30 0.971595 0.971922
#> 74  2024-12-18 16:33:30 0.971595 0.971974
#> 75  2024-12-18 16:33:30 0.971595 0.972794
#> 76  2024-12-18 16:33:30 0.971595 0.972013
#> 77  2024-12-18 16:33:30 0.971595 0.972793
#> 78  2024-12-18 16:33:30 0.971594 0.972793
#> 79  2024-12-18 16:33:30 0.971594 0.972794
#> 80  2024-12-18 16:33:30 0.971596 0.972012
#> 81  2024-12-18 16:33:30 0.971596 0.972794
#> 82  2024-12-18 16:33:30 0.971594 0.972792
#> 83  2024-12-18 16:33:30 0.971594 0.972794
#> 84  2024-12-18 16:33:30 0.971595 0.972793
#> 85  2024-12-18 16:33:30 0.971595 0.972792
#> 86  2024-12-18 16:33:30 0.971596 0.972792
#> 87  2024-12-18 16:33:30 0.971596 0.972791
#> 88  2024-12-18 16:33:30 0.971595 0.972791
#> 89  2024-12-18 16:33:30 0.971594 0.972791
#> 90  2024-12-18 16:33:30 0.971594 0.972791
#> 91  2024-12-18 16:33:30 0.971597 0.972791
#> 92  2024-12-18 16:33:30 0.971598 0.972791
#> 93  2024-12-18 16:33:30 0.971603 0.972791
#> 94  2024-12-18 16:33:30 0.971604 0.972791
#> 95  2024-12-18 16:33:30 0.971613 0.972791
#> 96  2024-12-18 16:33:30 0.971614 0.972791
#> 97  2024-12-18 16:33:30 0.971615 0.972791
#> 98  2024-12-18 16:33:30 0.971616 0.972791
#> 99  2024-12-18 16:33:31 0.971617 0.972791
#> 100 2024-12-18 16:33:31 0.971618 0.972791
#> 101 2024-12-18 16:33:31 0.971619 0.972791
#> 102 2024-12-18 16:33:31 0.971620 0.972791
#> 103 2024-12-18 16:33:31 0.971621 0.972791
#> 104 2024-12-18 16:33:31 0.971627 0.972791
#> 105 2024-12-18 16:33:31 0.972628 0.972791
#> 106 2024-12-18 16:33:31 0.972628 0.972792
#> 107 2024-12-18 16:33:31 0.972628 0.972792
#> 108 2024-12-18 16:33:31 0.972628 0.973154
#> 109 2024-12-18 16:33:31 0.972485 0.972629
#> 110 2024-12-18 16:33:31 0.972485 0.972790
#> 111 2024-12-18 16:33:31 0.972485 0.972789
#> 112 2024-12-18 16:33:31 0.972485 0.972790
#> 113 2024-12-18 16:33:31 0.972485 0.972790
#> 114 2024-12-18 16:33:31 0.972486 0.972789
#> 115 2024-12-18 16:33:31 0.972486 0.972783
#> 116 2024-12-18 16:33:31 0.972486 0.972782
#> 117 2024-12-18 16:33:32 0.972485 0.972782
#> 118 2024-12-18 16:33:32 0.972485 0.972790
#> 119 2024-12-18 16:33:32 0.972522 0.972789
#> 120 2024-12-18 16:33:32 0.972522 0.972790
#> 121 2024-12-18 16:33:32 0.972522 0.972790
#> 122 2024-12-18 16:33:32 0.972522 0.972790
#> 123 2024-12-18 16:33:32 0.972485 0.972523
#> 124 2024-12-18 16:33:32 0.972485 0.972783
#> 125 2024-12-18 16:33:32 0.971902 0.972486
#> 126 2024-12-18 16:33:32 0.972208 0.972486
#> 127 2024-12-18 16:33:32 0.971902 0.972486
#> 128 2024-12-18 16:33:32 0.971956 0.972486
#> 129 2024-12-18 16:33:32 0.971956 0.972783
#> 130 2024-12-18 16:33:32 0.971956 0.972728
#> 131 2024-12-18 16:33:32 0.971957 0.972728
#> 132 2024-12-18 16:33:32 0.971957 0.972485
#> 133 2024-12-18 16:33:32 0.971956 0.972485
#> 134 2024-12-18 16:33:32 0.971958 0.972728
#> 135 2024-12-18 16:33:32 0.971956 0.972727
#> 136 2024-12-18 16:33:32 0.971957 0.972717
#> 137 2024-12-18 16:33:32 0.971956 0.972717
#> 138 2024-12-18 16:33:32 0.971958 0.972717
#> 139 2024-12-18 16:33:32 0.971958 0.972716
#> 140 2024-12-18 16:33:32 0.971958 0.972693
#> 141 2024-12-18 16:33:32 0.971960 0.972693
#> 142 2024-12-18 16:33:32 0.971958 0.972693
#> 143 2024-12-18 16:33:32 0.971961 0.972692
#> 144 2024-12-18 16:33:32 0.971958 0.972692
#> 145 2024-12-18 16:33:32 0.971959 0.972662
#> 146 2024-12-18 16:33:32 0.971959 0.972661
#> 147 2024-12-18 16:33:32 0.971959 0.972654
#> 148 2024-12-18 16:33:32 0.971959 0.972653
#> 149 2024-12-18 16:33:32 0.971956 0.972653
#> 150 2024-12-18 16:33:32 0.971901 0.972653
#> 151 2024-12-18 16:33:32 0.971903 0.972653
#> 152 2024-12-18 16:33:32 0.971904 0.972653
#> 153 2024-12-18 16:33:33 0.971904 0.972652
#> 154 2024-12-18 16:33:33 0.971904 0.972651
#> 155 2024-12-18 16:33:33 0.972175 0.972651
#> 156 2024-12-18 16:33:33 0.972175 0.972650
#> 157 2024-12-18 16:33:33 0.972175 0.972649
#> 158 2024-12-18 16:33:33 0.972175 0.972648
#> 159 2024-12-18 16:33:33 0.972175 0.972647
#> 160 2024-12-18 16:33:33 0.972176 0.972647
#> 161 2024-12-18 16:33:33 0.972176 0.972646
#> 162 2024-12-18 16:33:33 0.972176 0.972641
#> 163 2024-12-18 16:33:33 0.972176 0.972640
#> 164 2024-12-18 16:33:33 0.972176 0.972648
#> 165 2024-12-18 16:33:34 0.972176 0.972644
#> 166 2024-12-18 16:33:34 0.972176 0.972643
#> 167 2024-12-18 16:33:34 0.972176 0.972637
#> 168 2024-12-18 16:33:34 0.972176 0.972636
#> 169 2024-12-18 16:33:34 0.972176 0.972637
#> 170 2024-12-18 16:33:34 0.972176 0.972637
#> 171 2024-12-18 16:33:34 0.972175 0.972636
#> 172 2024-12-18 16:33:34 0.971906 0.972636
#> 173 2024-12-18 16:33:34 0.971907 0.972636
#> 174 2024-12-18 16:33:34 0.971908 0.972636
#> 175 2024-12-18 16:33:34 0.971905 0.972636
#> 176 2024-12-18 16:33:34 0.971907 0.972636
#> 177 2024-12-18 16:33:34 0.971908 0.972636
#> 178 2024-12-18 16:33:34 0.971907 0.972636
#> 179 2024-12-18 16:33:34 0.971907 0.972636
#> 180 2024-12-18 16:33:35 0.971909 0.972636
#> 181 2024-12-18 16:33:35 0.971909 0.972630
#> 182 2024-12-18 16:33:35 0.971910 0.972630
#> 183 2024-12-18 16:33:35 0.971910 0.972629
#> 184 2024-12-18 16:33:35 0.971910 0.972648
#> 185 2024-12-18 16:33:35 0.971910 0.972648
#> 186 2024-12-18 16:33:35 0.971910 0.972648
#> 187 2024-12-18 16:33:35 0.971910 0.972648
#> 188 2024-12-18 16:33:35 0.971911 0.972647
#> 189 2024-12-18 16:33:35 0.971912 0.972647
#> 190 2024-12-18 16:33:35 0.971912 0.972652
#> 191 2024-12-18 16:33:35 0.971912 0.972646
#> 192 2024-12-18 16:33:35 0.971912 0.972652
#> 193 2024-12-18 16:33:35 0.971912 0.972768
#> 194 2024-12-18 16:33:35 0.971912 0.972769
#> 195 2024-12-18 16:33:35 0.971912 0.972646
#> 196 2024-12-18 16:33:35 0.971912 0.972644
#> 197 2024-12-18 16:33:35 0.971912 0.972646
#> 198 2024-12-18 16:33:35 0.971912 0.972646
#> 199 2024-12-18 16:33:35 0.971912 0.972646
#> 200 2024-12-18 16:33:35 0.971912 0.972645
#> 201 2024-12-18 16:33:35 0.971913 0.972645
#> 202 2024-12-18 16:33:35 0.971914 0.972645
#> 203 2024-12-18 16:33:36 0.971911 0.972645
#> 204 2024-12-18 16:33:36 0.971913 0.972645
#> 205 2024-12-18 16:33:36 0.971914 0.972645
#> 206 2024-12-18 16:33:36 0.971915 0.972645
#> 207 2024-12-18 16:33:36 0.971916 0.972645
#> 208 2024-12-18 16:33:36 0.971919 0.972645
#> 209 2024-12-18 16:33:36 0.971920 0.972645
#> 210 2024-12-18 16:33:36 0.971921 0.972645
#> 211 2024-12-18 16:33:36 0.972587 0.972645
#> 212 2024-12-18 16:33:36 0.972587 0.972646
#> 213 2024-12-18 16:33:36 0.972587 0.972766
#> 214 2024-12-18 16:33:36 0.972587 0.972767
#> 215 2024-12-18 16:33:36 0.972587 0.972766
#> 216 2024-12-18 16:33:36 0.972693 0.972766
#> 217 2024-12-18 16:33:36 0.972693 0.972767
#> 218 2024-12-18 16:33:37 0.972693 0.972766
#> 219 2024-12-18 16:33:37 0.972693 0.972765
#> 220 2024-12-18 16:33:37 0.972693 0.972764
#> 221 2024-12-18 16:33:37 0.972693 0.972765
#> 222 2024-12-18 16:33:37 0.972693 0.972765
#> 223 2024-12-18 16:33:37 0.972587 0.972764
#> 224 2024-12-18 16:33:37 0.972588 0.972764
#> 225 2024-12-18 16:33:37 0.972588 0.972765
#> 226 2024-12-18 16:33:38 0.972587 0.972765
#> 227 2024-12-18 16:33:38 0.972587 0.972764
#> 228 2024-12-18 16:33:38 0.972465 0.972764
#> 229 2024-12-18 16:33:38 0.972465 0.972759
#> 230 2024-12-18 16:33:38 0.972465 0.972466
#> 231 2024-12-18 16:33:39 0.972465 0.972759
#> 232 2024-12-18 16:33:39 0.972465 0.972759
#> 233 2024-12-18 16:33:39 0.972465 0.972752
#> 234 2024-12-18 16:33:39 0.972465 0.972751
#> 235 2024-12-18 16:33:39 0.972465 0.972750
#> 236 2024-12-18 16:33:39 0.972465 0.972749
#> 237 2024-12-18 16:33:39 0.972465 0.972752
#> 238 2024-12-18 16:33:39 0.972465 0.972743
#> 239 2024-12-18 16:33:39 0.972465 0.972742
#> 240 2024-12-18 16:33:39 0.972465 0.972736
#> 241 2024-12-18 16:33:39 0.972492 0.972736
#> 242 2024-12-18 16:33:39 0.972492 0.972493
#> 243 2024-12-18 16:33:39 0.972492 0.972765
#> 244 2024-12-18 16:33:39 0.972493 0.972764
#> 245 2024-12-18 16:33:39 0.972493 0.972763
#> 246 2024-12-18 16:33:39 0.972493 0.972762
#> 
#> $last
#> [1] 1734539619
#>