Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 16:37:17 1.4018 1.4030
#> 2   2024-12-18 16:37:17 1.4019 1.4030
#> 3   2024-12-18 16:37:17 1.4020 1.4030
#> 4   2024-12-18 16:37:17 1.4021 1.4030
#> 5   2024-12-18 16:37:18 1.4021 1.4031
#> 6   2024-12-18 16:37:18 1.4021 1.4030
#> 7   2024-12-18 16:37:18 1.4022 1.4030
#> 8   2024-12-18 16:37:18 1.4022 1.4032
#> 9   2024-12-18 16:37:18 1.4022 1.4031
#> 10  2024-12-18 16:37:19 1.4023 1.4031
#> 11  2024-12-18 16:37:19 1.4024 1.4031
#> 12  2024-12-18 16:37:19 1.4024 1.4032
#> 13  2024-12-18 16:37:19 1.4024 1.4036
#> 14  2024-12-18 16:37:19 1.4024 1.4038
#> 15  2024-12-18 16:37:19 1.4024 1.4036
#> 16  2024-12-18 16:37:20 1.4025 1.4036
#> 17  2024-12-18 16:37:20 1.4025 1.4038
#> 18  2024-12-18 16:37:20 1.4026 1.4038
#> 19  2024-12-18 16:37:20 1.4026 1.4045
#> 20  2024-12-18 16:37:20 1.4026 1.4046
#> 21  2024-12-18 16:37:20 1.4026 1.4041
#> 22  2024-12-18 16:37:20 1.4026 1.4040
#> 23  2024-12-18 16:37:20 1.4026 1.4042
#> 24  2024-12-18 16:37:20 1.4027 1.4042
#> 25  2024-12-18 16:37:20 1.4027 1.4041
#> 26  2024-12-18 16:37:21 1.4028 1.4041
#> 27  2024-12-18 16:37:21 1.4029 1.4041
#> 28  2024-12-18 16:37:21 1.4030 1.4041
#> 29  2024-12-18 16:37:21 1.4031 1.4041
#> 30  2024-12-18 16:37:22 1.4032 1.4041
#> 31  2024-12-18 16:37:22 1.4033 1.4041
#> 32  2024-12-18 16:37:23 1.4033 1.4040
#> 33  2024-12-18 16:37:23 1.4034 1.4040
#> 34  2024-12-18 16:37:24 1.4035 1.4040
#> 35  2024-12-18 16:37:24 1.4034 1.4040
#> 36  2024-12-18 16:37:24 1.4035 1.4040
#> 37  2024-12-18 16:37:24 1.4035 1.4041
#> 38  2024-12-18 16:37:25 1.4035 1.4040
#> 39  2024-12-18 16:37:25 1.4036 1.4040
#> 40  2024-12-18 16:37:27 1.4031 1.4040
#> 41  2024-12-18 16:37:27 1.4032 1.4040
#> 42  2024-12-18 16:37:27 1.4033 1.4040
#> 43  2024-12-18 16:37:29 1.4031 1.4040
#> 44  2024-12-18 16:37:29 1.4034 1.4040
#> 45  2024-12-18 16:37:29 1.4033 1.4040
#> 46  2024-12-18 16:37:30 1.4034 1.4040
#> 47  2024-12-18 16:37:31 1.4034 1.4042
#> 48  2024-12-18 16:37:31 1.4034 1.4041
#> 49  2024-12-18 16:37:31 1.4034 1.4046
#> 50  2024-12-18 16:37:31 1.4034 1.4045
#> 51  2024-12-18 16:37:31 1.4035 1.4045
#> 52  2024-12-18 16:37:31 1.4036 1.4045
#> 53  2024-12-18 16:37:31 1.4037 1.4045
#> 54  2024-12-18 16:37:31 1.4038 1.4045
#> 55  2024-12-18 16:37:33 1.4038 1.4046
#> 56  2024-12-18 16:37:33 1.4039 1.4046
#> 57  2024-12-18 16:37:33 1.4040 1.4046
#> 58  2024-12-18 16:37:33 1.4041 1.4046
#> 59  2024-12-18 16:37:33 1.4042 1.4046
#> 60  2024-12-18 16:37:34 1.4043 1.4046
#> 61  2024-12-18 16:37:34 1.4044 1.4046
#> 62  2024-12-18 16:37:36 1.4042 1.4046
#> 63  2024-12-18 16:37:36 1.4045 1.4046
#> 64  2024-12-18 16:37:38 1.4043 1.4046
#> 65  2024-12-18 16:37:38 1.4038 1.4046
#> 66  2024-12-18 16:37:38 1.4034 1.4046
#> 67  2024-12-18 16:37:38 1.4041 1.4046
#> 68  2024-12-18 16:37:38 1.4042 1.4046
#> 69  2024-12-18 16:37:38 1.4041 1.4046
#> 70  2024-12-18 16:37:38 1.4041 1.4045
#> 71  2024-12-18 16:37:39 1.4042 1.4045
#> 72  2024-12-18 16:37:39 1.4042 1.4046
#> 73  2024-12-18 16:37:40 1.4039 1.4046
#> 74  2024-12-18 16:37:40 1.4043 1.4046
#> 75  2024-12-18 16:37:40 1.4042 1.4046
#> 76  2024-12-18 16:37:40 1.4040 1.4046
#> 77  2024-12-18 16:37:40 1.4044 1.4046
#> 78  2024-12-18 16:37:40 1.4040 1.4046
#> 79  2024-12-18 16:37:42 1.4042 1.4046
#> 80  2024-12-18 16:37:42 1.4043 1.4046
#> 81  2024-12-18 16:37:43 1.4044 1.4046
#> 82  2024-12-18 16:37:43 1.4045 1.4046
#> 83  2024-12-18 16:37:46 1.4045 1.4060
#> 84  2024-12-18 16:37:46 1.4046 1.4060
#> 85  2024-12-18 16:37:46 1.4046 1.4059
#> 86  2024-12-18 16:37:46 1.4047 1.4059
#> 87  2024-12-18 16:37:46 1.4048 1.4059
#> 88  2024-12-18 16:37:46 1.4046 1.4059
#> 89  2024-12-18 16:37:46 1.4047 1.4063
#> 90  2024-12-18 16:37:46 1.4047 1.4062
#> 91  2024-12-18 16:37:46 1.4048 1.4062
#> 92  2024-12-18 16:37:46 1.4049 1.4062
#> 93  2024-12-18 16:37:46 1.4050 1.4062
#> 94  2024-12-18 16:37:46 1.4050 1.4063
#> 95  2024-12-18 16:37:46 1.4050 1.4066
#> 96  2024-12-18 16:37:46 1.4051 1.4066
#> 97  2024-12-18 16:37:46 1.4051 1.4070
#> 98  2024-12-18 16:37:46 1.4052 1.4070
#> 99  2024-12-18 16:37:46 1.4052 1.4065
#> 100 2024-12-18 16:37:46 1.4049 1.4065
#> 101 2024-12-18 16:37:46 1.4050 1.4069
#> 102 2024-12-18 16:37:46 1.4051 1.4069
#> 103 2024-12-18 16:37:46 1.4052 1.4069
#> 104 2024-12-18 16:37:48 1.4053 1.4069
#> 105 2024-12-18 16:37:48 1.4054 1.4069
#> 106 2024-12-18 16:37:48 1.4055 1.4069
#> 107 2024-12-18 16:37:48 1.4056 1.4069
#> 108 2024-12-18 16:37:48 1.4057 1.4069
#> 109 2024-12-18 16:37:48 1.4057 1.4068
#> 110 2024-12-18 16:37:50 1.4058 1.4068
#> 111 2024-12-18 16:37:50 1.4059 1.4068
#> 112 2024-12-18 16:37:50 1.4060 1.4068
#> 113 2024-12-18 16:37:50 1.4061 1.4068
#> 114 2024-12-18 16:37:50 1.4061 1.4075
#> 115 2024-12-18 16:37:50 1.4061 1.4074
#> 116 2024-12-18 16:37:50 1.4061 1.4076
#> 117 2024-12-18 16:37:50 1.4061 1.4075
#> 118 2024-12-18 16:37:50 1.4061 1.4074
#> 119 2024-12-18 16:37:50 1.4061 1.4073
#> 120 2024-12-18 16:37:50 1.4061 1.4072
#> 121 2024-12-18 16:37:51 1.4061 1.4071
#> 122 2024-12-18 16:37:52 1.4062 1.4071
#> 123 2024-12-18 16:37:52 1.4063 1.4071
#> 124 2024-12-18 16:37:52 1.4064 1.4071
#> 125 2024-12-18 16:37:53 1.4064 1.4072
#> 126 2024-12-18 16:37:53 1.4064 1.4076
#> 127 2024-12-18 16:37:53 1.4064 1.4075
#> 128 2024-12-18 16:37:53 1.4065 1.4075
#> 129 2024-12-18 16:37:53 1.4065 1.4079
#> 130 2024-12-18 16:37:53 1.4066 1.4079
#> 131 2024-12-18 16:37:53 1.4067 1.4079
#> 132 2024-12-18 16:37:54 1.4068 1.4079
#> 133 2024-12-18 16:37:54 1.4070 1.4079
#> 134 2024-12-18 16:37:55 1.4068 1.4079
#> 135 2024-12-18 16:37:55 1.4066 1.4079
#> 136 2024-12-18 16:37:55 1.4067 1.4079
#> 137 2024-12-18 16:37:55 1.4068 1.4079
#> 138 2024-12-18 16:37:55 1.4067 1.4079
#> 139 2024-12-18 16:37:55 1.4067 1.4078
#> 140 2024-12-18 16:37:55 1.4067 1.4079
#> 141 2024-12-18 16:37:55 1.4068 1.4079
#> 142 2024-12-18 16:37:56 1.4069 1.4079
#> 143 2024-12-18 16:37:56 1.4070 1.4079
#> 144 2024-12-18 16:37:57 1.4071 1.4079
#> 145 2024-12-18 16:37:57 1.4072 1.4079
#> 146 2024-12-18 16:37:58 1.4073 1.4079
#> 147 2024-12-18 16:38:00 1.4071 1.4079
#> 148 2024-12-18 16:38:00 1.4074 1.4079
#> 149 2024-12-18 16:38:00 1.4075 1.4079
#> 150 2024-12-18 16:38:00 1.4071 1.4079
#> 151 2024-12-18 16:38:00 1.4067 1.4079
#> 152 2024-12-18 16:38:00 1.4067 1.4079
#> 153 2024-12-18 16:38:01 1.4067 1.4078
#> 154 2024-12-18 16:38:01 1.4066 1.4078
#> 155 2024-12-18 16:38:01 1.4066 1.4077
#> 156 2024-12-18 16:38:01 1.4066 1.4076
#> 157 2024-12-18 16:38:01 1.4067 1.4076
#> 158 2024-12-18 16:38:01 1.4066 1.4076
#> 159 2024-12-18 16:38:01 1.4067 1.4076
#> 160 2024-12-18 16:38:01 1.4067 1.4079
#> 161 2024-12-18 16:38:01 1.4067 1.4078
#> 162 2024-12-18 16:38:03 1.4068 1.4078
#> 163 2024-12-18 16:38:03 1.4069 1.4078
#> 164 2024-12-18 16:38:03 1.4068 1.4078
#> 165 2024-12-18 16:38:03 1.4069 1.4078
#> 166 2024-12-18 16:38:04 1.4068 1.4078
#> 167 2024-12-18 16:38:04 1.4069 1.4078
#> 168 2024-12-18 16:38:04 1.4069 1.4077
#> 169 2024-12-18 16:38:04 1.4069 1.4076
#> 170 2024-12-18 16:38:04 1.4068 1.4076
#> 171 2024-12-18 16:38:04 1.4064 1.4076
#> 172 2024-12-18 16:38:04 1.4066 1.4076
#> 173 2024-12-18 16:38:04 1.4066 1.4075
#> 174 2024-12-18 16:38:04 1.4066 1.4074
#> 175 2024-12-18 16:38:04 1.4066 1.4073
#> 176 2024-12-18 16:38:04 1.4066 1.4072
#> 177 2024-12-18 16:38:04 1.4063 1.4072
#> 178 2024-12-18 16:38:04 1.4061 1.4072
#> 179 2024-12-18 16:38:04 1.4062 1.4072
#> 180 2024-12-18 16:38:04 1.4059 1.4072
#> 181 2024-12-18 16:38:04 1.4060 1.4072
#> 182 2024-12-18 16:38:04 1.4058 1.4072
#> 183 2024-12-18 16:38:04 1.4056 1.4072
#> 184 2024-12-18 16:38:04 1.4056 1.4071
#> 185 2024-12-18 16:38:04 1.4056 1.4070
#> 186 2024-12-18 16:38:04 1.4057 1.4070
#> 187 2024-12-18 16:38:05 1.4055 1.4070
#> 188 2024-12-18 16:38:05 1.4060 1.4070
#> 189 2024-12-18 16:38:05 1.4061 1.4070
#> 190 2024-12-18 16:38:05 1.4060 1.4070
#> 191 2024-12-18 16:38:05 1.4057 1.4070
#> 192 2024-12-18 16:38:05 1.4058 1.4070
#> 193 2024-12-18 16:38:05 1.4058 1.4069
#> 194 2024-12-18 16:38:05 1.4057 1.4069
#> 195 2024-12-18 16:38:05 1.4054 1.4069
#> 196 2024-12-18 16:38:05 1.4055 1.4069
#> 197 2024-12-18 16:38:05 1.4055 1.4068
#> 198 2024-12-18 16:38:05 1.4056 1.4068
#> 199 2024-12-18 16:38:05 1.4056 1.4067
#> 200 2024-12-18 16:38:05 1.4056 1.4066
#> 201 2024-12-18 16:38:06 1.4056 1.4065
#> 202 2024-12-18 16:38:06 1.4056 1.4064
#> 203 2024-12-18 16:38:06 1.4055 1.4064
#> 204 2024-12-18 16:38:06 1.4055 1.4063
#> 205 2024-12-18 16:38:06 1.4055 1.4062
#> 206 2024-12-18 16:38:06 1.4055 1.4061
#> 207 2024-12-18 16:38:07 1.4053 1.4061
#> 208 2024-12-18 16:38:07 1.4054 1.4061
#> 209 2024-12-18 16:38:09 1.4055 1.4061
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 16:37:54 0.977285 0.977800
#> 2   2024-12-18 16:37:54 0.977290 0.977800
#> 3   2024-12-18 16:37:54 0.977291 0.977800
#> 4   2024-12-18 16:37:54 0.977292 0.977800
#> 5   2024-12-18 16:37:55 0.977293 0.977800
#> 6   2024-12-18 16:37:55 0.977294 0.977800
#> 7   2024-12-18 16:37:55 0.977295 0.977800
#> 8   2024-12-18 16:37:55 0.977296 0.977800
#> 9   2024-12-18 16:37:55 0.977297 0.977800
#> 10  2024-12-18 16:37:55 0.977298 0.977800
#> 11  2024-12-18 16:37:55 0.977299 0.977800
#> 12  2024-12-18 16:37:55 0.977300 0.977800
#> 13  2024-12-18 16:37:55 0.977291 0.977800
#> 14  2024-12-18 16:37:55 0.977293 0.977800
#> 15  2024-12-18 16:37:55 0.977294 0.977800
#> 16  2024-12-18 16:37:55 0.977295 0.977800
#> 17  2024-12-18 16:37:55 0.977296 0.977800
#> 18  2024-12-18 16:37:55 0.977297 0.977800
#> 19  2024-12-18 16:37:55 0.977298 0.977800
#> 20  2024-12-18 16:37:55 0.977300 0.977800
#> 21  2024-12-18 16:37:55 0.977301 0.977800
#> 22  2024-12-18 16:37:56 0.977302 0.977800
#> 23  2024-12-18 16:37:56 0.977356 0.977800
#> 24  2024-12-18 16:37:56 0.977357 0.977800
#> 25  2024-12-18 16:37:56 0.977362 0.977800
#> 26  2024-12-18 16:37:56 0.977416 0.977800
#> 27  2024-12-18 16:37:56 0.977417 0.977800
#> 28  2024-12-18 16:37:57 0.977418 0.977800
#> 29  2024-12-18 16:37:57 0.977419 0.977800
#> 30  2024-12-18 16:37:57 0.977418 0.977800
#> 31  2024-12-18 16:37:57 0.977420 0.977800
#> 32  2024-12-18 16:37:57 0.977418 0.977800
#> 33  2024-12-18 16:37:57 0.977421 0.977800
#> 34  2024-12-18 16:37:57 0.977418 0.977800
#> 35  2024-12-18 16:37:57 0.977350 0.977800
#> 36  2024-12-18 16:37:57 0.977375 0.977800
#> 37  2024-12-18 16:37:57 0.977376 0.977800
#> 38  2024-12-18 16:37:57 0.977377 0.977800
#> 39  2024-12-18 16:37:57 0.977378 0.977800
#> 40  2024-12-18 16:37:58 0.977375 0.977800
#> 41  2024-12-18 16:37:58 0.977377 0.977800
#> 42  2024-12-18 16:37:58 0.977378 0.977800
#> 43  2024-12-18 16:37:58 0.977377 0.977800
#> 44  2024-12-18 16:37:58 0.977377 0.977800
#> 45  2024-12-18 16:37:58 0.977379 0.977800
#> 46  2024-12-18 16:37:58 0.977433 0.977800
#> 47  2024-12-18 16:37:58 0.977434 0.977800
#> 48  2024-12-18 16:37:58 0.977433 0.977800
#> 49  2024-12-18 16:37:58 0.977433 0.977800
#> 50  2024-12-18 16:37:58 0.977433 0.977800
#> 51  2024-12-18 16:37:58 0.977434 0.977800
#> 52  2024-12-18 16:37:58 0.977434 0.977800
#> 53  2024-12-18 16:37:58 0.977436 0.977800
#> 54  2024-12-18 16:37:58 0.977434 0.977800
#> 55  2024-12-18 16:37:58 0.977434 0.977800
#> 56  2024-12-18 16:37:58 0.977491 0.977800
#> 57  2024-12-18 16:37:58 0.977492 0.977800
#> 58  2024-12-18 16:37:58 0.977493 0.977800
#> 59  2024-12-18 16:37:58 0.977494 0.977800
#> 60  2024-12-18 16:37:58 0.977495 0.977800
#> 61  2024-12-18 16:37:58 0.977496 0.977800
#> 62  2024-12-18 16:37:59 0.977497 0.977800
#> 63  2024-12-18 16:37:59 0.977498 0.977800
#> 64  2024-12-18 16:37:59 0.977499 0.977800
#> 65  2024-12-18 16:37:59 0.977500 0.977800
#> 66  2024-12-18 16:37:59 0.977497 0.977800
#> 67  2024-12-18 16:37:59 0.977497 0.977800
#> 68  2024-12-18 16:37:59 0.977498 0.977800
#> 69  2024-12-18 16:37:59 0.977497 0.977800
#> 70  2024-12-18 16:37:59 0.977497 0.977799
#> 71  2024-12-18 16:38:00 0.977498 0.977799
#> 72  2024-12-18 16:38:00 0.977500 0.977799
#> 73  2024-12-18 16:38:00 0.977501 0.977799
#> 74  2024-12-18 16:38:00 0.977501 0.977800
#> 75  2024-12-18 16:38:00 0.977502 0.977800
#> 76  2024-12-18 16:38:00 0.977503 0.977800
#> 77  2024-12-18 16:38:00 0.977504 0.977800
#> 78  2024-12-18 16:38:00 0.977505 0.977800
#> 79  2024-12-18 16:38:00 0.977506 0.977800
#> 80  2024-12-18 16:38:00 0.977560 0.977800
#> 81  2024-12-18 16:38:00 0.977561 0.977800
#> 82  2024-12-18 16:38:01 0.977562 0.977800
#> 83  2024-12-18 16:38:01 0.977567 0.977800
#> 84  2024-12-18 16:38:01 0.977568 0.977800
#> 85  2024-12-18 16:38:01 0.977569 0.977800
#> 86  2024-12-18 16:38:01 0.977570 0.977800
#> 87  2024-12-18 16:38:01 0.977571 0.977800
#> 88  2024-12-18 16:38:01 0.977572 0.977800
#> 89  2024-12-18 16:38:01 0.977573 0.977800
#> 90  2024-12-18 16:38:02 0.977574 0.977800
#> 91  2024-12-18 16:38:02 0.977575 0.977800
#> 92  2024-12-18 16:38:02 0.977572 0.977800
#> 93  2024-12-18 16:38:02 0.977572 0.977800
#> 94  2024-12-18 16:38:02 0.977572 0.977800
#> 95  2024-12-18 16:38:02 0.977574 0.977800
#> 96  2024-12-18 16:38:02 0.977575 0.977800
#> 97  2024-12-18 16:38:02 0.977576 0.977800
#> 98  2024-12-18 16:38:02 0.977577 0.977800
#> 99  2024-12-18 16:38:02 0.977578 0.977800
#> 100 2024-12-18 16:38:02 0.977579 0.977800
#> 101 2024-12-18 16:38:03 0.977580 0.977800
#> 102 2024-12-18 16:38:03 0.977581 0.977800
#> 103 2024-12-18 16:38:03 0.977582 0.977800
#> 104 2024-12-18 16:38:03 0.977583 0.977800
#> 105 2024-12-18 16:38:03 0.977584 0.977800
#> 106 2024-12-18 16:38:03 0.977585 0.977800
#> 107 2024-12-18 16:38:03 0.977586 0.977800
#> 108 2024-12-18 16:38:03 0.977587 0.977800
#> 109 2024-12-18 16:38:03 0.977581 0.977800
#> 110 2024-12-18 16:38:03 0.977583 0.977800
#> 111 2024-12-18 16:38:03 0.977584 0.977800
#> 112 2024-12-18 16:38:03 0.977585 0.977800
#> 113 2024-12-18 16:38:03 0.977586 0.977800
#> 114 2024-12-18 16:38:03 0.977587 0.977800
#> 115 2024-12-18 16:38:03 0.977641 0.977800
#> 116 2024-12-18 16:38:03 0.977642 0.977800
#> 117 2024-12-18 16:38:04 0.977643 0.977800
#> 118 2024-12-18 16:38:04 0.977642 0.977800
#> 119 2024-12-18 16:38:04 0.977642 0.977799
#> 120 2024-12-18 16:38:04 0.977578 0.977799
#> 121 2024-12-18 16:38:04 0.977639 0.977799
#> 122 2024-12-18 16:38:04 0.977207 0.977799
#> 123 2024-12-18 16:38:04 0.977234 0.977799
#> 124 2024-12-18 16:38:04 0.977234 0.977798
#> 125 2024-12-18 16:38:04 0.977235 0.977798
#> 126 2024-12-18 16:38:04 0.977236 0.977798
#> 127 2024-12-18 16:38:04 0.976620 0.977798
#> 128 2024-12-18 16:38:04 0.977235 0.977799
#> 129 2024-12-18 16:38:04 0.977235 0.977797
#> 130 2024-12-18 16:38:04 0.976620 0.977797
#> 131 2024-12-18 16:38:04 0.976620 0.977797
#> 132 2024-12-18 16:38:04 0.976989 0.977797
#> 133 2024-12-18 16:38:04 0.977112 0.977797
#> 134 2024-12-18 16:38:04 0.977112 0.977799
#> 135 2024-12-18 16:38:05 0.977112 0.977796
#> 136 2024-12-18 16:38:05 0.977113 0.977796
#> 137 2024-12-18 16:38:05 0.977113 0.977799
#> 138 2024-12-18 16:38:05 0.977113 0.977799
#> 139 2024-12-18 16:38:05 0.977113 0.977799
#> 140 2024-12-18 16:38:05 0.977113 0.977799
#> 141 2024-12-18 16:38:05 0.976918 0.977794
#> 142 2024-12-18 16:38:05 0.976918 0.977799
#> 143 2024-12-18 16:38:05 0.977114 0.977796
#> 144 2024-12-18 16:38:05 0.977114 0.977799
#> 145 2024-12-18 16:38:05 0.976918 0.977793
#> 146 2024-12-18 16:38:05 0.976919 0.977799
#> 147 2024-12-18 16:38:05 0.976918 0.977795
#> 148 2024-12-18 16:38:05 0.977115 0.977799
#> 149 2024-12-18 16:38:05 0.976918 0.977792
#> 150 2024-12-18 16:38:05 0.976920 0.977799
#> 151 2024-12-18 16:38:05 0.976918 0.977794
#> 152 2024-12-18 16:38:05 0.977116 0.977799
#> 153 2024-12-18 16:38:05 0.976918 0.977791
#> 154 2024-12-18 16:38:05 0.976921 0.977799
#> 155 2024-12-18 16:38:05 0.976918 0.977793
#> 156 2024-12-18 16:38:05 0.977117 0.977799
#> 157 2024-12-18 16:38:05 0.976918 0.977790
#> 158 2024-12-18 16:38:05 0.976922 0.977799
#> 159 2024-12-18 16:38:05 0.976918 0.977792
#> 160 2024-12-18 16:38:05 0.977118 0.977799
#> 161 2024-12-18 16:38:05 0.976918 0.977789
#> 162 2024-12-18 16:38:05 0.976919 0.977799
#> 163 2024-12-18 16:38:05 0.976918 0.977798
#> 164 2024-12-18 16:38:05 0.976923 0.977799
#> 165 2024-12-18 16:38:05 0.976918 0.977788
#> 166 2024-12-18 16:38:05 0.976920 0.977799
#> 167 2024-12-18 16:38:05 0.976918 0.977798
#> 168 2024-12-18 16:38:05 0.976918 0.977798
#> 169 2024-12-18 16:38:05 0.976918 0.977798
#> 170 2024-12-18 16:38:05 0.976919 0.977658
#> 171 2024-12-18 16:38:05 0.976919 0.977657
#> 172 2024-12-18 16:38:05 0.976920 0.977657
#> 173 2024-12-18 16:38:05 0.976921 0.977657
#> 174 2024-12-18 16:38:05 0.976922 0.977657
#> 175 2024-12-18 16:38:05 0.976923 0.977657
#> 176 2024-12-18 16:38:05 0.976924 0.977657
#> 177 2024-12-18 16:38:05 0.976925 0.977657
#> 178 2024-12-18 16:38:06 0.976925 0.977656
#> 179 2024-12-18 16:38:06 0.976925 0.977655
#> 180 2024-12-18 16:38:06 0.976918 0.977655
#> 181 2024-12-18 16:38:06 0.976919 0.977654
#> 182 2024-12-18 16:38:06 0.976919 0.977653
#> 183 2024-12-18 16:38:06 0.976918 0.977653
#> 184 2024-12-18 16:38:06 0.976919 0.977653
#> 185 2024-12-18 16:38:06 0.976918 0.977653
#> 186 2024-12-18 16:38:06 0.976919 0.977653
#> 187 2024-12-18 16:38:06 0.976920 0.977653
#> 188 2024-12-18 16:38:06 0.976921 0.977653
#> 189 2024-12-18 16:38:07 0.976921 0.977652
#> 190 2024-12-18 16:38:07 0.976921 0.977651
#> 191 2024-12-18 16:38:07 0.976921 0.977650
#> 192 2024-12-18 16:38:07 0.976921 0.977649
#> 193 2024-12-18 16:38:07 0.976921 0.977648
#> 194 2024-12-18 16:38:07 0.976921 0.977647
#> 195 2024-12-18 16:38:07 0.976921 0.977593
#> 196 2024-12-18 16:38:07 0.976921 0.977592
#> 197 2024-12-18 16:38:07 0.976922 0.977592
#> 198 2024-12-18 16:38:07 0.976923 0.977592
#> 199 2024-12-18 16:38:07 0.976923 0.977591
#> 200 2024-12-18 16:38:07 0.976924 0.977591
#> 201 2024-12-18 16:38:07 0.976924 0.977590
#> 202 2024-12-18 16:38:07 0.976925 0.977590
#> 203 2024-12-18 16:38:08 0.976925 0.977536
#> 204 2024-12-18 16:38:08 0.976926 0.977536
#> 205 2024-12-18 16:38:08 0.976926 0.977535
#> 206 2024-12-18 16:38:08 0.976927 0.977535
#> 207 2024-12-18 16:38:08 0.976927 0.977534
#> 208 2024-12-18 16:38:08 0.976926 0.977534
#> 209 2024-12-18 16:38:08 0.976928 0.977533
#> 210 2024-12-18 16:38:08 0.976926 0.977533
#> 211 2024-12-18 16:38:08 0.976928 0.977533
#> 212 2024-12-18 16:38:08 0.976929 0.977533
#> 213 2024-12-18 16:38:08 0.976926 0.977533
#> 214 2024-12-18 16:38:08 0.976927 0.977532
#> 215 2024-12-18 16:38:08 0.976927 0.977531
#> 216 2024-12-18 16:38:08 0.976927 0.977530
#> 217 2024-12-18 16:38:08 0.976927 0.977529
#> 218 2024-12-18 16:38:08 0.976928 0.977529
#> 219 2024-12-18 16:38:08 0.976929 0.977529
#> 220 2024-12-18 16:38:09 0.976930 0.977529
#> 221 2024-12-18 16:38:09 0.976931 0.977529
#> 222 2024-12-18 16:38:09 0.976918 0.977529
#> 223 2024-12-18 16:38:09 0.976918 0.977529
#> 224 2024-12-18 16:38:09 0.976784 0.977529
#> 225 2024-12-18 16:38:09 0.976755 0.977529
#> 226 2024-12-18 16:38:09 0.976756 0.977529
#> 227 2024-12-18 16:38:09 0.976756 0.977528
#> 228 2024-12-18 16:38:09 0.976756 0.977527
#> 229 2024-12-18 16:38:09 0.976757 0.977527
#> 230 2024-12-18 16:38:09 0.976758 0.977527
#> 231 2024-12-18 16:38:09 0.976763 0.977527
#> 232 2024-12-18 16:38:09 0.976764 0.977527
#> 233 2024-12-18 16:38:09 0.976765 0.977527
#> 234 2024-12-18 16:38:10 0.976766 0.977527
#> 235 2024-12-18 16:38:10 0.976767 0.977527
#> 236 2024-12-18 16:38:10 0.976767 0.977526
#> 237 2024-12-18 16:38:10 0.976768 0.977526
#> 238 2024-12-18 16:38:10 0.976768 0.977525
#> 239 2024-12-18 16:38:10 0.976769 0.977525
#> 240 2024-12-18 16:38:10 0.976769 0.977524
#> 241 2024-12-18 16:38:10 0.976770 0.977524
#> 242 2024-12-18 16:38:10 0.976771 0.977524
#> 243 2024-12-18 16:38:10 0.976772 0.977524
#> 244 2024-12-18 16:38:11 0.976773 0.977524
#> 245 2024-12-18 16:38:11 0.976774 0.977524
#> 246 2024-12-18 16:38:11 0.976771 0.977524
#> 247 2024-12-18 16:38:11 0.976772 0.977526
#> 
#> $last
#> [1] 1734539891
#>