Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-12-18 17:46:03 1.4187 1.4198
#> 2   2024-12-18 17:46:03 1.4188 1.4198
#> 3   2024-12-18 17:46:03 1.4188 1.4197
#> 4   2024-12-18 17:46:03 1.4189 1.4197
#> 5   2024-12-18 17:46:03 1.4190 1.4197
#> 6   2024-12-18 17:46:03 1.4190 1.4196
#> 7   2024-12-18 17:46:04 1.4187 1.4196
#> 8   2024-12-18 17:46:05 1.4189 1.4196
#> 9   2024-12-18 17:46:05 1.4190 1.4196
#> 10  2024-12-18 17:46:06 1.4191 1.4196
#> 11  2024-12-18 17:46:07 1.4191 1.4195
#> 12  2024-12-18 17:46:07 1.4187 1.4195
#> 13  2024-12-18 17:46:07 1.4189 1.4195
#> 14  2024-12-18 17:46:07 1.4190 1.4195
#> 15  2024-12-18 17:46:07 1.4190 1.4194
#> 16  2024-12-18 17:46:07 1.4189 1.4194
#> 17  2024-12-18 17:46:07 1.4189 1.4196
#> 18  2024-12-18 17:46:07 1.4190 1.4195
#> 19  2024-12-18 17:46:08 1.4189 1.4195
#> 20  2024-12-18 17:46:08 1.4190 1.4195
#> 21  2024-12-18 17:46:08 1.4188 1.4195
#> 22  2024-12-18 17:46:08 1.4191 1.4195
#> 23  2024-12-18 17:46:09 1.4191 1.4194
#> 24  2024-12-18 17:46:09 1.4189 1.4194
#> 25  2024-12-18 17:46:09 1.4182 1.4194
#> 26  2024-12-18 17:46:09 1.4186 1.4194
#> 27  2024-12-18 17:46:09 1.4189 1.4194
#> 28  2024-12-18 17:46:09 1.4187 1.4194
#> 29  2024-12-18 17:46:09 1.4182 1.4194
#> 30  2024-12-18 17:46:10 1.4181 1.4194
#> 31  2024-12-18 17:46:10 1.4184 1.4194
#> 32  2024-12-18 17:46:10 1.4184 1.4193
#> 33  2024-12-18 17:46:10 1.4185 1.4193
#> 34  2024-12-18 17:46:10 1.4185 1.4192
#> 35  2024-12-18 17:46:10 1.4184 1.4192
#> 36  2024-12-18 17:46:10 1.4184 1.4191
#> 37  2024-12-18 17:46:10 1.4184 1.4190
#> 38  2024-12-18 17:46:10 1.4184 1.4189
#> 39  2024-12-18 17:46:10 1.4184 1.4188
#> 40  2024-12-18 17:46:10 1.4184 1.4194
#> 41  2024-12-18 17:46:10 1.4184 1.4193
#> 42  2024-12-18 17:46:10 1.4185 1.4193
#> 43  2024-12-18 17:46:10 1.4185 1.4192
#> 44  2024-12-18 17:46:10 1.4185 1.4191
#> 45  2024-12-18 17:46:10 1.4185 1.4195
#> 46  2024-12-18 17:46:11 1.4185 1.4193
#> 47  2024-12-18 17:46:11 1.4185 1.4192
#> 48  2024-12-18 17:46:11 1.4186 1.4192
#> 49  2024-12-18 17:46:11 1.4186 1.4194
#> 50  2024-12-18 17:46:11 1.4187 1.4193
#> 51  2024-12-18 17:46:12 1.4187 1.4196
#> 52  2024-12-18 17:46:12 1.4187 1.4195
#> 53  2024-12-18 17:46:12 1.4187 1.4194
#> 54  2024-12-18 17:46:12 1.4186 1.4194
#> 55  2024-12-18 17:46:12 1.4186 1.4193
#> 56  2024-12-18 17:46:12 1.4187 1.4193
#> 57  2024-12-18 17:46:12 1.4187 1.4192
#> 58  2024-12-18 17:46:12 1.4187 1.4195
#> 59  2024-12-18 17:46:12 1.4187 1.4193
#> 60  2024-12-18 17:46:12 1.4187 1.4192
#> 61  2024-12-18 17:46:12 1.4187 1.4194
#> 62  2024-12-18 17:46:12 1.4187 1.4193
#> 63  2024-12-18 17:46:12 1.4186 1.4193
#> 64  2024-12-18 17:46:13 1.4186 1.4192
#> 65  2024-12-18 17:46:13 1.4186 1.4191
#> 66  2024-12-18 17:46:13 1.4186 1.4190
#> 67  2024-12-18 17:46:14 1.4183 1.4190
#> 68  2024-12-18 17:46:14 1.4184 1.4190
#> 69  2024-12-18 17:46:14 1.4185 1.4190
#> 70  2024-12-18 17:46:14 1.4185 1.4189
#> 71  2024-12-18 17:46:14 1.4185 1.4188
#> 72  2024-12-18 17:46:14 1.4184 1.4188
#> 73  2024-12-18 17:46:14 1.4181 1.4188
#> 74  2024-12-18 17:46:14 1.4182 1.4188
#> 75  2024-12-18 17:46:14 1.4182 1.4187
#> 76  2024-12-18 17:46:14 1.4182 1.4186
#> 77  2024-12-18 17:46:14 1.4182 1.4185
#> 78  2024-12-18 17:46:15 1.4182 1.4186
#> 79  2024-12-18 17:46:15 1.4182 1.4191
#> 80  2024-12-18 17:46:15 1.4182 1.4190
#> 81  2024-12-18 17:46:15 1.4183 1.4190
#> 82  2024-12-18 17:46:15 1.4182 1.4190
#> 83  2024-12-18 17:46:15 1.4182 1.4189
#> 84  2024-12-18 17:46:15 1.4182 1.4188
#> 85  2024-12-18 17:46:15 1.4182 1.4187
#> 86  2024-12-18 17:46:15 1.4182 1.4186
#> 87  2024-12-18 17:46:15 1.4182 1.4185
#> 88  2024-12-18 17:46:16 1.4180 1.4185
#> 89  2024-12-18 17:46:17 1.4177 1.4185
#> 90  2024-12-18 17:46:17 1.4178 1.4185
#> 91  2024-12-18 17:46:17 1.4178 1.4184
#> 92  2024-12-18 17:46:17 1.4179 1.4184
#> 93  2024-12-18 17:46:18 1.4179 1.4186
#> 94  2024-12-18 17:46:20 1.4175 1.4185
#> 95  2024-12-18 17:46:20 1.4180 1.4185
#> 96  2024-12-18 17:46:20 1.4181 1.4185
#> 97  2024-12-18 17:46:20 1.4175 1.4185
#> 98  2024-12-18 17:46:20 1.4177 1.4185
#> 99  2024-12-18 17:46:20 1.4178 1.4185
#> 100 2024-12-18 17:46:20 1.4178 1.4200
#> 101 2024-12-18 17:46:20 1.4178 1.4198
#> 102 2024-12-18 17:46:20 1.4178 1.4197
#> 103 2024-12-18 17:46:20 1.4178 1.4194
#> 104 2024-12-18 17:46:20 1.4178 1.4193
#> 105 2024-12-18 17:46:20 1.4178 1.4198
#> 106 2024-12-18 17:46:20 1.4178 1.4196
#> 107 2024-12-18 17:46:21 1.4179 1.4196
#> 108 2024-12-18 17:46:21 1.4180 1.4196
#> 109 2024-12-18 17:46:21 1.4181 1.4196
#> 110 2024-12-18 17:46:21 1.4182 1.4196
#> 111 2024-12-18 17:46:21 1.4182 1.4198
#> 112 2024-12-18 17:46:21 1.4182 1.4196
#> 113 2024-12-18 17:46:21 1.4182 1.4195
#> 114 2024-12-18 17:46:22 1.4183 1.4195
#> 115 2024-12-18 17:46:22 1.4184 1.4195
#> 116 2024-12-18 17:46:22 1.4185 1.4195
#> 117 2024-12-18 17:46:23 1.4186 1.4195
#> 118 2024-12-18 17:46:23 1.4187 1.4195
#> 119 2024-12-18 17:46:23 1.4187 1.4194
#> 120 2024-12-18 17:46:24 1.4187 1.4196
#> 121 2024-12-18 17:46:24 1.4188 1.4195
#> 122 2024-12-18 17:46:24 1.4189 1.4195
#> 123 2024-12-18 17:46:25 1.4190 1.4195
#> 124 2024-12-18 17:46:25 1.4191 1.4195
#> 125 2024-12-18 17:46:27 1.4192 1.4195
#> 126 2024-12-18 17:46:27 1.4192 1.4194
#> 127 2024-12-18 17:46:27 1.4189 1.4194
#> 128 2024-12-18 17:46:27 1.4190 1.4194
#> 129 2024-12-18 17:46:27 1.4187 1.4194
#> 130 2024-12-18 17:46:27 1.4188 1.4194
#> 131 2024-12-18 17:46:28 1.4187 1.4194
#> 132 2024-12-18 17:46:28 1.4184 1.4194
#> 133 2024-12-18 17:46:28 1.4186 1.4194
#> 134 2024-12-18 17:46:28 1.4187 1.4194
#> 135 2024-12-18 17:46:28 1.4186 1.4194
#> 136 2024-12-18 17:46:28 1.4187 1.4194
#> 137 2024-12-18 17:46:29 1.4182 1.4194
#> 138 2024-12-18 17:46:29 1.4188 1.4194
#> 139 2024-12-18 17:46:29 1.4188 1.4193
#> 140 2024-12-18 17:46:29 1.4187 1.4193
#> 141 2024-12-18 17:46:29 1.4187 1.4192
#> 142 2024-12-18 17:46:29 1.4182 1.4192
#> 143 2024-12-18 17:46:29 1.4184 1.4192
#> 144 2024-12-18 17:46:30 1.4184 1.4191
#> 145 2024-12-18 17:46:30 1.4184 1.4190
#> 146 2024-12-18 17:46:30 1.4183 1.4190
#> 147 2024-12-18 17:46:30 1.4183 1.4189
#> 148 2024-12-18 17:46:30 1.4183 1.4188
#> 149 2024-12-18 17:46:30 1.4183 1.4193
#> 150 2024-12-18 17:46:30 1.4183 1.4192
#> 151 2024-12-18 17:46:30 1.4184 1.4192
#> 152 2024-12-18 17:46:31 1.4185 1.4192
#> 153 2024-12-18 17:46:31 1.4186 1.4192
#> 154 2024-12-18 17:46:32 1.4186 1.4195
#> 155 2024-12-18 17:46:32 1.4186 1.4194
#> 156 2024-12-18 17:46:32 1.4187 1.4194
#> 157 2024-12-18 17:46:33 1.4188 1.4194
#> 158 2024-12-18 17:46:33 1.4189 1.4194
#> 159 2024-12-18 17:46:34 1.4188 1.4194
#> 160 2024-12-18 17:46:34 1.4188 1.4193
#> 161 2024-12-18 17:46:34 1.4188 1.4192
#> 162 2024-12-18 17:46:34 1.4185 1.4192
#> 163 2024-12-18 17:46:34 1.4186 1.4192
#> 164 2024-12-18 17:46:34 1.4187 1.4192
#> 165 2024-12-18 17:46:34 1.4186 1.4192
#> 166 2024-12-18 17:46:34 1.4185 1.4192
#> 167 2024-12-18 17:46:34 1.4185 1.4191
#> 168 2024-12-18 17:46:34 1.4180 1.4191
#> 169 2024-12-18 17:46:34 1.4182 1.4190
#> 170 2024-12-18 17:46:34 1.4180 1.4190
#> 171 2024-12-18 17:46:34 1.4181 1.4189
#> 172 2024-12-18 17:46:34 1.4181 1.4188
#> 173 2024-12-18 17:46:34 1.4181 1.4187
#> 174 2024-12-18 17:46:34 1.4181 1.4186
#> 175 2024-12-18 17:46:35 1.4182 1.4186
#> 176 2024-12-18 17:46:35 1.4182 1.4185
#> 177 2024-12-18 17:46:35 1.4182 1.4184
#> 178 2024-12-18 17:46:35 1.4182 1.4186
#> 179 2024-12-18 17:46:35 1.4182 1.4185
#> 180 2024-12-18 17:46:35 1.4182 1.4186
#> 181 2024-12-18 17:46:35 1.4182 1.4185
#> 182 2024-12-18 17:46:35 1.4182 1.4184
#> 183 2024-12-18 17:46:35 1.4176 1.4184
#> 184 2024-12-18 17:46:35 1.4174 1.4184
#> 185 2024-12-18 17:46:35 1.4176 1.4184
#> 186 2024-12-18 17:46:35 1.4178 1.4184
#> 187 2024-12-18 17:46:35 1.4180 1.4184
#> 188 2024-12-18 17:46:35 1.4179 1.4184
#> 189 2024-12-18 17:46:35 1.4175 1.4184
#> 190 2024-12-18 17:46:36 1.4174 1.4184
#> 191 2024-12-18 17:46:36 1.4176 1.4184
#> 192 2024-12-18 17:46:36 1.4177 1.4184
#> 193 2024-12-18 17:46:36 1.4178 1.4184
#> 194 2024-12-18 17:46:36 1.4171 1.4184
#> 195 2024-12-18 17:46:36 1.4171 1.4183
#> 196 2024-12-18 17:46:36 1.4172 1.4183
#> 197 2024-12-18 17:46:36 1.4171 1.4183
#> 198 2024-12-18 17:46:36 1.4171 1.4182
#> 199 2024-12-18 17:46:36 1.4172 1.4182
#> 200 2024-12-18 17:46:37 1.4172 1.4181
#> 201 2024-12-18 17:46:37 1.4168 1.4181
#> 202 2024-12-18 17:46:37 1.4169 1.4180
#> 203 2024-12-18 17:46:37 1.4172 1.4180
#> 204 2024-12-18 17:46:37 1.4172 1.4179
#> 205 2024-12-18 17:46:37 1.4169 1.4179
#> 206 2024-12-18 17:46:37 1.4169 1.4178
#> 207 2024-12-18 17:46:38 1.4169 1.4177
#> 208 2024-12-18 17:46:38 1.4169 1.4176
#> 209 2024-12-18 17:46:39 1.4170 1.4176
#> 210 2024-12-18 17:46:39 1.4171 1.4176
#> 211 2024-12-18 17:46:39 1.4171 1.4175
#> 212 2024-12-18 17:46:39 1.4171 1.4174
#> 213 2024-12-18 17:46:39 1.4170 1.4174
#> 214 2024-12-18 17:46:39 1.4170 1.4177
#> 215 2024-12-18 17:46:39 1.4171 1.4176
#> 216 2024-12-18 17:46:39 1.4171 1.4175
#> 217 2024-12-18 17:46:39 1.4171 1.4174
#> 218 2024-12-18 17:46:39 1.4170 1.4174
#> 219 2024-12-18 17:46:39 1.4170 1.4177
#> 220 2024-12-18 17:46:39 1.4171 1.4176
#> 221 2024-12-18 17:46:40 1.4171 1.4175
#> 222 2024-12-18 17:46:40 1.4171 1.4174
#> 223 2024-12-18 17:46:40 1.4171 1.4173
#> 224 2024-12-18 17:46:40 1.4171 1.4176
#> 225 2024-12-18 17:46:40 1.4171 1.4174
#> 226 2024-12-18 17:46:40 1.4171 1.4173
#> 227 2024-12-18 17:46:40 1.4171 1.4176
#> 228 2024-12-18 17:46:41 1.4169 1.4175
#> 229 2024-12-18 17:46:41 1.4169 1.4174
#> 230 2024-12-18 17:46:41 1.4169 1.4173
#> 231 2024-12-18 17:46:41 1.4169 1.4176
#> 232 2024-12-18 17:46:41 1.4166 1.4175
#> 233 2024-12-18 17:46:41 1.4167 1.4175
#> 234 2024-12-18 17:46:41 1.4168 1.4175
#> 235 2024-12-18 17:46:42 1.4168 1.4174
#> 236 2024-12-18 17:46:42 1.4168 1.4173
#> 237 2024-12-18 17:46:42 1.4168 1.4172
#> 238 2024-12-18 17:46:42 1.4168 1.4171
#> 239 2024-12-18 17:46:42 1.4167 1.4171
#> 240 2024-12-18 17:46:42 1.4167 1.4174
#> 241 2024-12-18 17:46:42 1.4168 1.4173
#> 242 2024-12-18 17:46:42 1.4168 1.4172
#> 243 2024-12-18 17:46:42 1.4168 1.4171
#> 244 2024-12-18 17:46:42 1.4167 1.4171
#> 245 2024-12-18 17:46:42 1.4167 1.4174
#> 246 2024-12-18 17:46:42 1.4168 1.4173
#> 247 2024-12-18 17:46:44 1.4169 1.4173
#> 248 2024-12-18 17:46:45 1.4169 1.4174
#> 249 2024-12-18 17:46:45 1.4169 1.4173
#> 250 2024-12-18 17:46:45 1.4169 1.4172
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-12-18 17:46:32 0.988758 0.990228
#> 2   2024-12-18 17:46:32 0.988758 0.990231
#> 3   2024-12-18 17:46:32 0.988758 0.990231
#> 4   2024-12-18 17:46:32 0.988758 0.990231
#> 5   2024-12-18 17:46:32 0.988759 0.990227
#> 6   2024-12-18 17:46:32 0.988759 0.990231
#> 7   2024-12-18 17:46:32 0.988759 0.990231
#> 8   2024-12-18 17:46:32 0.988759 0.990231
#> 9   2024-12-18 17:46:32 0.988759 0.990231
#> 10  2024-12-18 17:46:32 0.988759 0.990231
#> 11  2024-12-18 17:46:32 0.988753 0.990181
#> 12  2024-12-18 17:46:32 0.988754 0.990231
#> 13  2024-12-18 17:46:32 0.988754 0.990231
#> 14  2024-12-18 17:46:32 0.988755 0.990230
#> 15  2024-12-18 17:46:32 0.988756 0.990230
#> 16  2024-12-18 17:46:32 0.988757 0.990230
#> 17  2024-12-18 17:46:32 0.988758 0.990230
#> 18  2024-12-18 17:46:32 0.988759 0.990230
#> 19  2024-12-18 17:46:32 0.988760 0.990230
#> 20  2024-12-18 17:46:32 0.988761 0.990230
#> 21  2024-12-18 17:46:32 0.988762 0.990230
#> 22  2024-12-18 17:46:32 0.988767 0.990230
#> 23  2024-12-18 17:46:32 0.988768 0.990230
#> 24  2024-12-18 17:46:32 0.988769 0.990230
#> 25  2024-12-18 17:46:32 0.988770 0.990230
#> 26  2024-12-18 17:46:32 0.988771 0.990230
#> 27  2024-12-18 17:46:32 0.988772 0.990230
#> 28  2024-12-18 17:46:32 0.988773 0.990230
#> 29  2024-12-18 17:46:32 0.988774 0.990230
#> 30  2024-12-18 17:46:33 0.988775 0.990230
#> 31  2024-12-18 17:46:33 0.988776 0.990230
#> 32  2024-12-18 17:46:33 0.988777 0.990230
#> 33  2024-12-18 17:46:33 0.988778 0.990230
#> 34  2024-12-18 17:46:33 0.988778 0.990229
#> 35  2024-12-18 17:46:33 0.988778 0.990228
#> 36  2024-12-18 17:46:33 0.988779 0.990228
#> 37  2024-12-18 17:46:33 0.988779 0.990231
#> 38  2024-12-18 17:46:33 0.988779 0.990227
#> 39  2024-12-18 17:46:33 0.988780 0.990227
#> 40  2024-12-18 17:46:33 0.988780 0.990226
#> 41  2024-12-18 17:46:33 0.988781 0.990226
#> 42  2024-12-18 17:46:33 0.988782 0.990226
#> 43  2024-12-18 17:46:33 0.988783 0.990226
#> 44  2024-12-18 17:46:33 0.988784 0.990226
#> 45  2024-12-18 17:46:33 0.988785 0.990226
#> 46  2024-12-18 17:46:33 0.988785 0.990225
#> 47  2024-12-18 17:46:33 0.988786 0.990225
#> 48  2024-12-18 17:46:33 0.988786 0.990224
#> 49  2024-12-18 17:46:33 0.988787 0.990224
#> 50  2024-12-18 17:46:33 0.988788 0.990224
#> 51  2024-12-18 17:46:34 0.988788 0.990165
#> 52  2024-12-18 17:46:34 0.988788 0.990164
#> 53  2024-12-18 17:46:34 0.988789 0.990164
#> 54  2024-12-18 17:46:34 0.988790 0.990164
#> 55  2024-12-18 17:46:34 0.988790 0.990163
#> 56  2024-12-18 17:46:34 0.988790 0.990151
#> 57  2024-12-18 17:46:34 0.988790 0.990150
#> 58  2024-12-18 17:46:34 0.988791 0.990150
#> 59  2024-12-18 17:46:34 0.988792 0.990150
#> 60  2024-12-18 17:46:34 0.988792 0.990149
#> 61  2024-12-18 17:46:34 0.988793 0.990149
#> 62  2024-12-18 17:46:34 0.988793 0.990148
#> 63  2024-12-18 17:46:34 0.988794 0.990148
#> 64  2024-12-18 17:46:34 0.988794 0.990143
#> 65  2024-12-18 17:46:34 0.988795 0.990143
#> 66  2024-12-18 17:46:34 0.988795 0.990142
#> 67  2024-12-18 17:46:35 0.988795 0.990141
#> 68  2024-12-18 17:46:35 0.988796 0.990141
#> 69  2024-12-18 17:46:35 0.988796 0.990140
#> 70  2024-12-18 17:46:35 0.988797 0.990140
#> 71  2024-12-18 17:46:35 0.988798 0.990140
#> 72  2024-12-18 17:46:35 0.988799 0.990140
#> 73  2024-12-18 17:46:35 0.988799 0.990134
#> 74  2024-12-18 17:46:35 0.988799 0.990133
#> 75  2024-12-18 17:46:35 0.988799 0.990139
#> 76  2024-12-18 17:46:35 0.988799 0.990139
#> 77  2024-12-18 17:46:35 0.988799 0.990139
#> 78  2024-12-18 17:46:35 0.988796 0.990138
#> 79  2024-12-18 17:46:35 0.988796 0.990138
#> 80  2024-12-18 17:46:35 0.988796 0.990138
#> 81  2024-12-18 17:46:35 0.988798 0.990138
#> 82  2024-12-18 17:46:35 0.988799 0.990138
#> 83  2024-12-18 17:46:35 0.988800 0.990138
#> 84  2024-12-18 17:46:35 0.988800 0.990137
#> 85  2024-12-18 17:46:35 0.988801 0.990137
#> 86  2024-12-18 17:46:35 0.988801 0.990136
#> 87  2024-12-18 17:46:35 0.988801 0.990135
#> 88  2024-12-18 17:46:35 0.988802 0.990135
#> 89  2024-12-18 17:46:36 0.988802 0.990134
#> 90  2024-12-18 17:46:36 0.988802 0.990133
#> 91  2024-12-18 17:46:36 0.988802 0.990128
#> 92  2024-12-18 17:46:36 0.988802 0.990127
#> 93  2024-12-18 17:46:36 0.988802 0.990133
#> 94  2024-12-18 17:46:36 0.988802 0.990131
#> 95  2024-12-18 17:46:36 0.988802 0.990130
#> 96  2024-12-18 17:46:36 0.988802 0.990129
#> 97  2024-12-18 17:46:36 0.988802 0.990123
#> 98  2024-12-18 17:46:36 0.988802 0.990122
#> 99  2024-12-18 17:46:36 0.988801 0.990122
#> 100 2024-12-18 17:46:36 0.988801 0.990121
#> 101 2024-12-18 17:46:36 0.988801 0.990120
#> 102 2024-12-18 17:46:36 0.988801 0.990119
#> 103 2024-12-18 17:46:36 0.988801 0.990118
#> 104 2024-12-18 17:46:37 0.988801 0.990059
#> 105 2024-12-18 17:46:37 0.988801 0.990058
#> 106 2024-12-18 17:46:37 0.988801 0.990059
#> 107 2024-12-18 17:46:37 0.988801 0.990059
#> 108 2024-12-18 17:46:37 0.988802 0.990058
#> 109 2024-12-18 17:46:37 0.988802 0.990057
#> 110 2024-12-18 17:46:37 0.988802 0.990056
#> 111 2024-12-18 17:46:37 0.988802 0.990050
#> 112 2024-12-18 17:46:37 0.988802 0.990049
#> 113 2024-12-18 17:46:37 0.988802 0.990055
#> 114 2024-12-18 17:46:37 0.988803 0.990054
#> 115 2024-12-18 17:46:37 0.988804 0.990054
#> 116 2024-12-18 17:46:37 0.988805 0.990054
#> 117 2024-12-18 17:46:37 0.988806 0.990054
#> 118 2024-12-18 17:46:37 0.988807 0.990054
#> 119 2024-12-18 17:46:37 0.988808 0.990054
#> 120 2024-12-18 17:46:37 0.988803 0.990054
#> 121 2024-12-18 17:46:37 0.988804 0.990053
#> 122 2024-12-18 17:46:37 0.988805 0.990053
#> 123 2024-12-18 17:46:37 0.988805 0.990052
#> 124 2024-12-18 17:46:37 0.988806 0.990052
#> 125 2024-12-18 17:46:37 0.988791 0.990052
#> 126 2024-12-18 17:46:38 0.988792 0.990051
#> 127 2024-12-18 17:46:38 0.988792 0.990050
#> 128 2024-12-18 17:46:38 0.988792 0.989800
#> 129 2024-12-18 17:46:38 0.988792 0.989799
#> 130 2024-12-18 17:46:38 0.988791 0.989799
#> 131 2024-12-18 17:46:38 0.988791 0.989740
#> 132 2024-12-18 17:46:38 0.988792 0.989740
#> 133 2024-12-18 17:46:38 0.988792 0.989739
#> 134 2024-12-18 17:46:38 0.988791 0.989739
#> 135 2024-12-18 17:46:38 0.988791 0.989680
#> 136 2024-12-18 17:46:38 0.988791 0.989679
#> 137 2024-12-18 17:46:38 0.988791 0.989620
#> 138 2024-12-18 17:46:38 0.988791 0.989738
#> 139 2024-12-18 17:46:38 0.988791 0.989739
#> 140 2024-12-18 17:46:38 0.988791 0.989739
#> 141 2024-12-18 17:46:38 0.988791 0.989679
#> 142 2024-12-18 17:46:38 0.988791 0.989678
#> 143 2024-12-18 17:46:38 0.988791 0.989619
#> 144 2024-12-18 17:46:38 0.988791 0.989618
#> 145 2024-12-18 17:46:38 0.988792 0.989618
#> 146 2024-12-18 17:46:38 0.988792 0.989559
#> 147 2024-12-18 17:46:38 0.988792 0.989558
#> 148 2024-12-18 17:46:38 0.988792 0.989499
#> 149 2024-12-18 17:46:38 0.988792 0.989618
#> 150 2024-12-18 17:46:38 0.988792 0.989619
#> 151 2024-12-18 17:46:38 0.988792 0.989619
#> 152 2024-12-18 17:46:38 0.988792 0.989739
#> 153 2024-12-18 17:46:38 0.988792 0.989679
#> 154 2024-12-18 17:46:38 0.988792 0.989678
#> 155 2024-12-18 17:46:38 0.988797 0.989678
#> 156 2024-12-18 17:46:38 0.988797 0.989619
#> 157 2024-12-18 17:46:38 0.988798 0.989619
#> 158 2024-12-18 17:46:38 0.988797 0.989619
#> 159 2024-12-18 17:46:38 0.988799 0.989618
#> 160 2024-12-18 17:46:38 0.988797 0.989618
#> 161 2024-12-18 17:46:38 0.988799 0.989618
#> 162 2024-12-18 17:46:38 0.988800 0.989618
#> 163 2024-12-18 17:46:39 0.988800 0.989617
#> 164 2024-12-18 17:46:39 0.988800 0.989616
#> 165 2024-12-18 17:46:39 0.988800 0.989615
#> 166 2024-12-18 17:46:39 0.988800 0.989614
#> 167 2024-12-18 17:46:39 0.988800 0.989617
#> 168 2024-12-18 17:46:39 0.988800 0.989615
#> 169 2024-12-18 17:46:39 0.988800 0.989614
#> 170 2024-12-18 17:46:39 0.988800 0.989613
#> 171 2024-12-18 17:46:39 0.988800 0.989614
#> 172 2024-12-18 17:46:39 0.988662 0.989612
#> 173 2024-12-18 17:46:39 0.988663 0.989614
#> 174 2024-12-18 17:46:40 0.988663 0.989596
#> 175 2024-12-18 17:46:40 0.988663 0.989595
#> 176 2024-12-18 17:46:40 0.988662 0.989595
#> 177 2024-12-18 17:46:40 0.988663 0.989595
#> 178 2024-12-18 17:46:40 0.988662 0.989595
#> 179 2024-12-18 17:46:40 0.988662 0.989589
#> 180 2024-12-18 17:46:40 0.988662 0.989588
#> 181 2024-12-18 17:46:40 0.988662 0.989587
#> 182 2024-12-18 17:46:41 0.988662 0.989465
#> 183 2024-12-18 17:46:41 0.988662 0.989464
#> 184 2024-12-18 17:46:41 0.988662 0.989463
#> 185 2024-12-18 17:46:41 0.988661 0.989463
#> 186 2024-12-18 17:46:41 0.988661 0.989462
#> 187 2024-12-18 17:46:41 0.988661 0.989456
#> 188 2024-12-18 17:46:41 0.988661 0.989455
#> 189 2024-12-18 17:46:41 0.988556 0.989455
#> 190 2024-12-18 17:46:41 0.988555 0.989455
#> 191 2024-12-18 17:46:41 0.988555 0.989461
#> 192 2024-12-18 17:46:41 0.988555 0.989461
#> 193 2024-12-18 17:46:41 0.988555 0.989461
#> 194 2024-12-18 17:46:41 0.988555 0.988867
#> 195 2024-12-18 17:46:41 0.988555 0.988556
#> 196 2024-12-18 17:46:41 0.988424 0.988556
#> 197 2024-12-18 17:46:41 0.988424 0.988867
#> 198 2024-12-18 17:46:41 0.988424 0.988556
#> 199 2024-12-18 17:46:41 0.988424 0.988867
#> 200 2024-12-18 17:46:42 0.988424 0.988808
#> 201 2024-12-18 17:46:42 0.988424 0.988593
#> 202 2024-12-18 17:46:42 0.988424 0.988807
#> 203 2024-12-18 17:46:42 0.988424 0.988748
#> 204 2024-12-18 17:46:42 0.988424 0.988592
#> 205 2024-12-18 17:46:42 0.988424 0.988748
#> 206 2024-12-18 17:46:42 0.988424 0.988748
#> 207 2024-12-18 17:46:42 0.988424 0.988808
#> 208 2024-12-18 17:46:42 0.988424 0.988808
#> 209 2024-12-18 17:46:42 0.988424 0.988808
#> 210 2024-12-18 17:46:42 0.988424 0.988808
#> 211 2024-12-18 17:46:43 0.988424 0.988808
#> 212 2024-12-18 17:46:43 0.988425 0.988808
#> 213 2024-12-18 17:46:43 0.988426 0.988808
#> 214 2024-12-18 17:46:43 0.988427 0.988808
#> 215 2024-12-18 17:46:43 0.988430 0.988808
#> 216 2024-12-18 17:46:43 0.988431 0.988808
#> 217 2024-12-18 17:46:43 0.988432 0.988808
#> 218 2024-12-18 17:46:43 0.988433 0.988808
#> 219 2024-12-18 17:46:43 0.988434 0.988808
#> 220 2024-12-18 17:46:43 0.988435 0.988808
#> 221 2024-12-18 17:46:43 0.988436 0.988808
#> 222 2024-12-18 17:46:43 0.988437 0.988808
#> 223 2024-12-18 17:46:44 0.988438 0.988808
#> 224 2024-12-18 17:46:44 0.988439 0.988808
#> 225 2024-12-18 17:46:44 0.988439 0.989298
#> 226 2024-12-18 17:46:44 0.988439 0.989468
#> 227 2024-12-18 17:46:44 0.988439 0.989371
#> 228 2024-12-18 17:46:44 0.988439 0.989297
#> 229 2024-12-18 17:46:44 0.988439 0.989433
#> 230 2024-12-18 17:46:44 0.988439 0.989433
#> 231 2024-12-18 17:46:44 0.988440 0.989432
#> 232 2024-12-18 17:46:44 0.988441 0.989432
#> 233 2024-12-18 17:46:44 0.988442 0.989432
#> 234 2024-12-18 17:46:44 0.988443 0.989432
#> 235 2024-12-18 17:46:45 0.988444 0.989432
#> 236 2024-12-18 17:46:45 0.988445 0.989432
#> 237 2024-12-18 17:46:45 0.988446 0.989432
#> 238 2024-12-18 17:46:45 0.988450 0.989432
#> 239 2024-12-18 17:46:45 0.988451 0.989432
#> 240 2024-12-18 17:46:45 0.988452 0.989432
#> 241 2024-12-18 17:46:45 0.988452 0.989431
#> 242 2024-12-18 17:46:45 0.988453 0.989431
#> 243 2024-12-18 17:46:45 0.988458 0.989431
#> 244 2024-12-18 17:46:45 0.988459 0.989431
#> 245 2024-12-18 17:46:45 0.988459 0.989430
#> 246 2024-12-18 17:46:45 0.988459 0.989429
#> 247 2024-12-18 17:46:45 0.988459 0.989428
#> 
#> $last
#> [1] 1734544005
#>