Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2024-11-18 23:23:19 1.2107 1.2144
#> 2   2024-11-18 23:23:19 1.2107 1.2143
#> 3   2024-11-18 23:23:19 1.2107 1.2142
#> 4   2024-11-18 23:23:19 1.2107 1.2141
#> 5   2024-11-18 23:23:19 1.2107 1.2140
#> 6   2024-11-18 23:23:19 1.2107 1.2139
#> 7   2024-11-18 23:23:19 1.2107 1.2138
#> 8   2024-11-18 23:23:19 1.2107 1.2137
#> 9   2024-11-18 23:23:19 1.2107 1.2136
#> 10  2024-11-18 23:23:19 1.2107 1.2135
#> 11  2024-11-18 23:23:20 1.2107 1.2133
#> 12  2024-11-18 23:23:20 1.2107 1.2132
#> 13  2024-11-18 23:23:20 1.2107 1.2131
#> 14  2024-11-18 23:23:20 1.2106 1.2131
#> 15  2024-11-18 23:23:21 1.2105 1.2131
#> 16  2024-11-18 23:23:21 1.2105 1.2130
#> 17  2024-11-18 23:23:21 1.2105 1.2129
#> 18  2024-11-18 23:23:21 1.2105 1.2128
#> 19  2024-11-18 23:23:22 1.2105 1.2127
#> 20  2024-11-18 23:23:22 1.2105 1.2126
#> 21  2024-11-18 23:23:22 1.2104 1.2126
#> 22  2024-11-18 23:23:22 1.2104 1.2124
#> 23  2024-11-18 23:23:22 1.2105 1.2124
#> 24  2024-11-18 23:23:22 1.2105 1.2123
#> 25  2024-11-18 23:23:22 1.2105 1.2122
#> 26  2024-11-18 23:23:23 1.2105 1.2121
#> 27  2024-11-18 23:23:23 1.2105 1.2120
#> 28  2024-11-18 23:23:23 1.2105 1.2119
#> 29  2024-11-18 23:23:24 1.2105 1.2121
#> 30  2024-11-18 23:23:24 1.2105 1.2125
#> 31  2024-11-18 23:23:24 1.2105 1.2123
#> 32  2024-11-18 23:23:24 1.2105 1.2122
#> 33  2024-11-18 23:23:24 1.2105 1.2128
#> 34  2024-11-18 23:23:25 1.2106 1.2128
#> 35  2024-11-18 23:23:25 1.2105 1.2128
#> 36  2024-11-18 23:23:25 1.2106 1.2128
#> 37  2024-11-18 23:23:25 1.2107 1.2128
#> 38  2024-11-18 23:23:26 1.2106 1.2128
#> 39  2024-11-18 23:23:26 1.2108 1.2128
#> 40  2024-11-18 23:23:26 1.2109 1.2128
#> 41  2024-11-18 23:23:27 1.2109 1.2127
#> 42  2024-11-18 23:23:27 1.2109 1.2125
#> 43  2024-11-18 23:23:28 1.2108 1.2125
#> 44  2024-11-18 23:23:28 1.2109 1.2125
#> 45  2024-11-18 23:23:28 1.2109 1.2124
#> 46  2024-11-18 23:23:28 1.2109 1.2122
#> 47  2024-11-18 23:23:29 1.2110 1.2122
#> 48  2024-11-18 23:23:30 1.2111 1.2122
#> 49  2024-11-18 23:23:31 1.2110 1.2122
#> 50  2024-11-18 23:23:31 1.2105 1.2122
#> 51  2024-11-18 23:23:31 1.2106 1.2122
#> 52  2024-11-18 23:23:31 1.2105 1.2122
#> 53  2024-11-18 23:23:31 1.2105 1.2128
#> 54  2024-11-18 23:23:31 1.2105 1.2126
#> 55  2024-11-18 23:23:31 1.2105 1.2125
#> 56  2024-11-18 23:23:31 1.2105 1.2123
#> 57  2024-11-18 23:23:31 1.2106 1.2123
#> 58  2024-11-18 23:23:31 1.2107 1.2123
#> 59  2024-11-18 23:23:32 1.2108 1.2123
#> 60  2024-11-18 23:23:32 1.2110 1.2123
#> 61  2024-11-18 23:23:32 1.2108 1.2123
#> 62  2024-11-18 23:23:32 1.2108 1.2131
#> 63  2024-11-18 23:23:32 1.2108 1.2129
#> 64  2024-11-18 23:23:32 1.2109 1.2129
#> 65  2024-11-18 23:23:32 1.2111 1.2129
#> 66  2024-11-18 23:23:32 1.2112 1.2129
#> 67  2024-11-18 23:23:32 1.2111 1.2129
#> 68  2024-11-18 23:23:32 1.2105 1.2129
#> 69  2024-11-18 23:23:33 1.2106 1.2129
#> 70  2024-11-18 23:23:33 1.2107 1.2129
#> 71  2024-11-18 23:23:33 1.2106 1.2129
#> 72  2024-11-18 23:23:34 1.2107 1.2129
#> 73  2024-11-18 23:23:34 1.2108 1.2129
#> 74  2024-11-18 23:23:34 1.2109 1.2129
#> 75  2024-11-18 23:23:34 1.2110 1.2129
#> 76  2024-11-18 23:23:34 1.2112 1.2129
#> 77  2024-11-18 23:23:34 1.2113 1.2129
#> 78  2024-11-18 23:23:35 1.2114 1.2129
#> 79  2024-11-18 23:23:35 1.2114 1.2134
#> 80  2024-11-18 23:23:35 1.2115 1.2134
#> 81  2024-11-18 23:23:35 1.2117 1.2134
#> 82  2024-11-18 23:23:35 1.2118 1.2134
#> 83  2024-11-18 23:23:36 1.2119 1.2134
#> 84  2024-11-18 23:23:36 1.2121 1.2134
#> 85  2024-11-18 23:23:36 1.2122 1.2134
#> 86  2024-11-18 23:23:37 1.2123 1.2134
#> 87  2024-11-18 23:23:38 1.2124 1.2134
#> 88  2024-11-18 23:23:38 1.2125 1.2134
#> 89  2024-11-18 23:23:38 1.2126 1.2134
#> 90  2024-11-18 23:23:39 1.2127 1.2134
#> 91  2024-11-18 23:23:39 1.2128 1.2134
#> 92  2024-11-18 23:23:39 1.2127 1.2134
#> 93  2024-11-18 23:23:39 1.2129 1.2134
#> 94  2024-11-18 23:23:39 1.2127 1.2134
#> 95  2024-11-18 23:23:39 1.2126 1.2134
#> 96  2024-11-18 23:23:39 1.2127 1.2134
#> 97  2024-11-18 23:23:39 1.2128 1.2134
#> 98  2024-11-18 23:23:39 1.2130 1.2134
#> 99  2024-11-18 23:23:39 1.2131 1.2134
#> 100 2024-11-18 23:23:40 1.2133 1.2134
#> 101 2024-11-18 23:23:42 1.2133 1.2144
#> 102 2024-11-18 23:23:42 1.2134 1.2144
#> 103 2024-11-18 23:23:42 1.2136 1.2144
#> 104 2024-11-18 23:23:42 1.2137 1.2144
#> 105 2024-11-18 23:23:42 1.2139 1.2144
#> 106 2024-11-18 23:23:43 1.2137 1.2144
#> 107 2024-11-18 23:23:43 1.2138 1.2144
#> 108 2024-11-18 23:23:43 1.2139 1.2144
#> 109 2024-11-18 23:23:44 1.2140 1.2144
#> 110 2024-11-18 23:23:44 1.2139 1.2144
#> 111 2024-11-18 23:23:44 1.2140 1.2144
#> 112 2024-11-18 23:23:44 1.2141 1.2144
#> 113 2024-11-18 23:23:45 1.2143 1.2144
#> 114 2024-11-18 23:23:45 1.2141 1.2144
#> 115 2024-11-18 23:23:45 1.2142 1.2144
#> 116 2024-11-18 23:23:45 1.2141 1.2144
#> 117 2024-11-18 23:23:45 1.2138 1.2144
#> 118 2024-11-18 23:23:47 1.2138 1.2149
#> 119 2024-11-18 23:23:47 1.2144 1.2149
#> 120 2024-11-18 23:23:48 1.2144 1.2156
#> 121 2024-11-18 23:23:48 1.2144 1.2154
#> 122 2024-11-18 23:23:48 1.2144 1.2153
#> 123 2024-11-18 23:23:48 1.2144 1.2151
#> 124 2024-11-18 23:23:48 1.2144 1.2150
#> 125 2024-11-18 23:23:48 1.2144 1.2148
#> 126 2024-11-18 23:23:48 1.2144 1.2147
#> 127 2024-11-18 23:23:49 1.2144 1.2146
#> 128 2024-11-18 23:23:49 1.2144 1.2145
#> 129 2024-11-18 23:23:49 1.2123 1.2145
#> 130 2024-11-18 23:23:49 1.2123 1.2144
#> 131 2024-11-18 23:23:49 1.2117 1.2144
#> 132 2024-11-18 23:23:49 1.2109 1.2144
#> 133 2024-11-18 23:23:49 1.2112 1.2144
#> 134 2024-11-18 23:23:49 1.2119 1.2144
#> 135 2024-11-18 23:23:49 1.2120 1.2144
#> 136 2024-11-18 23:23:49 1.2120 1.2143
#> 137 2024-11-18 23:23:49 1.2120 1.2142
#> 138 2024-11-18 23:23:49 1.2120 1.2141
#> 139 2024-11-18 23:23:49 1.2120 1.2140
#> 140 2024-11-18 23:23:50 1.2120 1.2139
#> 141 2024-11-18 23:23:50 1.2121 1.2139
#> 142 2024-11-18 23:23:50 1.2122 1.2139
#> 143 2024-11-18 23:23:51 1.2122 1.2141
#> 144 2024-11-18 23:23:51 1.2123 1.2140
#> 145 2024-11-18 23:23:52 1.2123 1.2150
#> 146 2024-11-18 23:23:52 1.2124 1.2150
#> 147 2024-11-18 23:23:52 1.2125 1.2150
#> 148 2024-11-18 23:23:52 1.2125 1.2148
#> 149 2024-11-18 23:23:52 1.2125 1.2146
#> 150 2024-11-18 23:23:53 1.2126 1.2146
#> 151 2024-11-18 23:23:53 1.2126 1.2145
#> 152 2024-11-18 23:23:53 1.2126 1.2143
#> 153 2024-11-18 23:23:53 1.2126 1.2142
#> 154 2024-11-18 23:23:53 1.2126 1.2140
#> 155 2024-11-18 23:23:53 1.2123 1.2140
#> 156 2024-11-18 23:23:53 1.2123 1.2139
#> 157 2024-11-18 23:23:53 1.2123 1.2137
#> 158 2024-11-18 23:23:53 1.2124 1.2137
#> 159 2024-11-18 23:23:54 1.2124 1.2136
#> 160 2024-11-18 23:23:54 1.2124 1.2135
#> 161 2024-11-18 23:23:54 1.2124 1.2134
#> 162 2024-11-18 23:23:55 1.2123 1.2134
#> 163 2024-11-18 23:23:55 1.2114 1.2134
#> 164 2024-11-18 23:23:55 1.2112 1.2134
#> 165 2024-11-18 23:23:55 1.2115 1.2134
#> 166 2024-11-18 23:23:55 1.2115 1.2133
#> 167 2024-11-18 23:23:55 1.2115 1.2131
#> 168 2024-11-18 23:23:55 1.2115 1.2130
#> 169 2024-11-18 23:23:55 1.2115 1.2128
#> 170 2024-11-18 23:23:55 1.2115 1.2127
#> 171 2024-11-18 23:23:55 1.2115 1.2125
#> 172 2024-11-18 23:23:56 1.2115 1.2124
#> 173 2024-11-18 23:23:56 1.2115 1.2123
#> 174 2024-11-18 23:23:57 1.2115 1.2133
#> 175 2024-11-18 23:23:57 1.2115 1.2131
#> 176 2024-11-18 23:23:57 1.2115 1.2140
#> 177 2024-11-18 23:23:57 1.2115 1.2138
#> 178 2024-11-18 23:23:57 1.2115 1.2132
#> 179 2024-11-18 23:23:57 1.2115 1.2130
#> 180 2024-11-18 23:23:57 1.2114 1.2130
#> 181 2024-11-18 23:23:57 1.2114 1.2129
#> 182 2024-11-18 23:23:57 1.2106 1.2129
#> 183 2024-11-18 23:23:57 1.2106 1.2127
#> 184 2024-11-18 23:23:57 1.2107 1.2127
#> 185 2024-11-18 23:23:58 1.2107 1.2126
#> 186 2024-11-18 23:23:58 1.2106 1.2126
#> 187 2024-11-18 23:23:58 1.2103 1.2126
#> 188 2024-11-18 23:23:58 1.2103 1.2125
#> 189 2024-11-18 23:23:58 1.2103 1.2124
#> 190 2024-11-18 23:23:58 1.2103 1.2123
#> 191 2024-11-18 23:23:58 1.2101 1.2123
#> 192 2024-11-18 23:23:58 1.2102 1.2123
#> 193 2024-11-18 23:23:58 1.2102 1.2122
#> 194 2024-11-18 23:23:58 1.2102 1.2121
#> 195 2024-11-18 23:23:58 1.2102 1.2119
#> 196 2024-11-18 23:23:59 1.2103 1.2119
#> 197 2024-11-18 23:23:59 1.2104 1.2119
#> 198 2024-11-18 23:23:59 1.2104 1.2118
#> 199 2024-11-18 23:23:59 1.2104 1.2119
#> 200 2024-11-18 23:23:59 1.2104 1.2116
#> 201 2024-11-18 23:23:59 1.2105 1.2116
#> 202 2024-11-18 23:23:59 1.2105 1.2117
#> 203 2024-11-18 23:23:59 1.2105 1.2127
#> 204 2024-11-18 23:23:59 1.2105 1.2125
#> 205 2024-11-18 23:24:00 1.2106 1.2125
#> 206 2024-11-18 23:24:00 1.2106 1.2124
#> 207 2024-11-18 23:24:00 1.2107 1.2124
#> 208 2024-11-18 23:24:00 1.2108 1.2124
#> 209 2024-11-18 23:24:00 1.2107 1.2124
#> 210 2024-11-18 23:24:01 1.2109 1.2124
#> 211 2024-11-18 23:24:02 1.2110 1.2124
#> 212 2024-11-18 23:24:02 1.2112 1.2124
#> 213 2024-11-18 23:24:03 1.2113 1.2124
#> 214 2024-11-18 23:24:03 1.2103 1.2124
#> 215 2024-11-18 23:24:04 1.2103 1.2122
#> 216 2024-11-18 23:24:04 1.2103 1.2128
#> 217 2024-11-18 23:24:04 1.2103 1.2126
#> 218 2024-11-18 23:24:04 1.2103 1.2124
#> 219 2024-11-18 23:24:05 1.2103 1.2123
#> 220 2024-11-18 23:24:05 1.2103 1.2122
#> 221 2024-11-18 23:24:06 1.2105 1.2122
#> 222 2024-11-18 23:24:06 1.2106 1.2122
#> 223 2024-11-18 23:24:06 1.2108 1.2122
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2024-11-18 23:23:34 0.697208 0.698208
#> 2   2024-11-18 23:23:34 0.697173 0.698208
#> 3   2024-11-18 23:23:34 0.697209 0.698208
#> 4   2024-11-18 23:23:34 0.697173 0.698208
#> 5   2024-11-18 23:23:34 0.697174 0.698208
#> 6   2024-11-18 23:23:34 0.697175 0.698208
#> 7   2024-11-18 23:23:34 0.697175 0.698207
#> 8   2024-11-18 23:23:34 0.697175 0.698206
#> 9   2024-11-18 23:23:34 0.697175 0.698205
#> 10  2024-11-18 23:23:34 0.697175 0.698204
#> 11  2024-11-18 23:23:34 0.697176 0.698204
#> 12  2024-11-18 23:23:34 0.697177 0.698204
#> 13  2024-11-18 23:23:34 0.697177 0.698203
#> 14  2024-11-18 23:23:34 0.697177 0.698202
#> 15  2024-11-18 23:23:34 0.697177 0.698201
#> 16  2024-11-18 23:23:34 0.697177 0.698200
#> 17  2024-11-18 23:23:34 0.697177 0.698199
#> 18  2024-11-18 23:23:34 0.697177 0.698198
#> 19  2024-11-18 23:23:34 0.697177 0.698193
#> 20  2024-11-18 23:23:34 0.697177 0.698192
#> 21  2024-11-18 23:23:34 0.697177 0.697840
#> 22  2024-11-18 23:23:34 0.697177 0.697839
#> 23  2024-11-18 23:23:34 0.697177 0.697838
#> 24  2024-11-18 23:23:34 0.697177 0.697837
#> 25  2024-11-18 23:23:35 0.697177 0.697805
#> 26  2024-11-18 23:23:35 0.697177 0.697804
#> 27  2024-11-18 23:23:35 0.697177 0.697638
#> 28  2024-11-18 23:23:35 0.697174 0.697638
#> 29  2024-11-18 23:23:35 0.697175 0.697637
#> 30  2024-11-18 23:23:35 0.697175 0.697636
#> 31  2024-11-18 23:23:35 0.697175 0.697604
#> 32  2024-11-18 23:23:35 0.697175 0.697603
#> 33  2024-11-18 23:23:35 0.697173 0.697603
#> 34  2024-11-18 23:23:35 0.697173 0.697604
#> 35  2024-11-18 23:23:35 0.697173 0.697604
#> 36  2024-11-18 23:23:35 0.697173 0.697604
#> 37  2024-11-18 23:23:35 0.697173 0.697604
#> 38  2024-11-18 23:23:35 0.697173 0.697604
#> 39  2024-11-18 23:23:35 0.697174 0.697603
#> 40  2024-11-18 23:23:35 0.697175 0.697603
#> 41  2024-11-18 23:23:35 0.697175 0.697602
#> 42  2024-11-18 23:23:35 0.697175 0.697601
#> 43  2024-11-18 23:23:35 0.697175 0.697569
#> 44  2024-11-18 23:23:35 0.697175 0.697568
#> 45  2024-11-18 23:23:36 0.697176 0.697568
#> 46  2024-11-18 23:23:36 0.697177 0.697568
#> 47  2024-11-18 23:23:37 0.697177 0.697567
#> 48  2024-11-18 23:23:37 0.697178 0.697567
#> 49  2024-11-18 23:23:37 0.697178 0.697566
#> 50  2024-11-18 23:23:37 0.697179 0.697566
#> 51  2024-11-18 23:23:38 0.697180 0.697566
#> 52  2024-11-18 23:23:38 0.697181 0.697566
#> 53  2024-11-18 23:23:38 0.697178 0.697566
#> 54  2024-11-18 23:23:39 0.697179 0.697565
#> 55  2024-11-18 23:23:39 0.697180 0.697565
#> 56  2024-11-18 23:23:39 0.697180 0.697564
#> 57  2024-11-18 23:23:39 0.697181 0.697564
#> 58  2024-11-18 23:23:40 0.697181 0.697563
#> 59  2024-11-18 23:23:40 0.697181 0.697562
#> 60  2024-11-18 23:23:40 0.697181 0.697565
#> 61  2024-11-18 23:23:40 0.697181 0.697561
#> 62  2024-11-18 23:23:40 0.697181 0.697560
#> 63  2024-11-18 23:23:40 0.697181 0.697565
#> 64  2024-11-18 23:23:40 0.697181 0.697559
#> 65  2024-11-18 23:23:40 0.697181 0.697558
#> 66  2024-11-18 23:23:40 0.697181 0.697565
#> 67  2024-11-18 23:23:40 0.697181 0.697557
#> 68  2024-11-18 23:23:40 0.697181 0.697556
#> 69  2024-11-18 23:23:40 0.697181 0.697565
#> 70  2024-11-18 23:23:41 0.697181 0.697563
#> 71  2024-11-18 23:23:41 0.697182 0.697563
#> 72  2024-11-18 23:23:41 0.697182 0.697542
#> 73  2024-11-18 23:23:41 0.697182 0.697541
#> 74  2024-11-18 23:23:41 0.697183 0.697541
#> 75  2024-11-18 23:23:41 0.697184 0.697541
#> 76  2024-11-18 23:23:41 0.697184 0.697540
#> 77  2024-11-18 23:23:41 0.697185 0.697540
#> 78  2024-11-18 23:23:41 0.697185 0.697529
#> 79  2024-11-18 23:23:41 0.697186 0.697529
#> 80  2024-11-18 23:23:41 0.697186 0.697528
#> 81  2024-11-18 23:23:41 0.697186 0.697540
#> 82  2024-11-18 23:23:41 0.697187 0.697539
#> 83  2024-11-18 23:23:41 0.697188 0.697539
#> 84  2024-11-18 23:23:41 0.697188 0.697540
#> 85  2024-11-18 23:23:42 0.697189 0.697540
#> 86  2024-11-18 23:23:42 0.697190 0.697540
#> 87  2024-11-18 23:23:42 0.697190 0.697539
#> 88  2024-11-18 23:23:42 0.697191 0.697539
#> 89  2024-11-18 23:23:42 0.697192 0.697539
#> 90  2024-11-18 23:23:42 0.697192 0.697540
#> 91  2024-11-18 23:23:43 0.697193 0.697540
#> 92  2024-11-18 23:23:43 0.697194 0.697540
#> 93  2024-11-18 23:23:43 0.697194 0.697586
#> 94  2024-11-18 23:23:43 0.697194 0.697568
#> 95  2024-11-18 23:23:43 0.697195 0.697568
#> 96  2024-11-18 23:23:43 0.697196 0.697568
#> 97  2024-11-18 23:23:43 0.697197 0.697568
#> 98  2024-11-18 23:23:43 0.697198 0.697568
#> 99  2024-11-18 23:23:43 0.697199 0.697568
#> 100 2024-11-18 23:23:43 0.697200 0.697568
#> 101 2024-11-18 23:23:43 0.697201 0.697568
#> 102 2024-11-18 23:23:43 0.697202 0.697568
#> 103 2024-11-18 23:23:43 0.697203 0.697568
#> 104 2024-11-18 23:23:44 0.697203 0.697567
#> 105 2024-11-18 23:23:44 0.697204 0.697567
#> 106 2024-11-18 23:23:44 0.697205 0.697567
#> 107 2024-11-18 23:23:44 0.697206 0.697567
#> 108 2024-11-18 23:23:44 0.697207 0.697567
#> 109 2024-11-18 23:23:44 0.697207 0.697566
#> 110 2024-11-18 23:23:44 0.697207 0.697535
#> 111 2024-11-18 23:23:44 0.697208 0.697535
#> 112 2024-11-18 23:23:44 0.697209 0.697535
#> 113 2024-11-18 23:23:45 0.697210 0.697535
#> 114 2024-11-18 23:23:45 0.697211 0.697535
#> 115 2024-11-18 23:23:45 0.697206 0.697535
#> 116 2024-11-18 23:23:45 0.697207 0.697534
#> 117 2024-11-18 23:23:45 0.697207 0.697535
#> 118 2024-11-18 23:23:45 0.697207 0.697535
#> 119 2024-11-18 23:23:45 0.697207 0.697535
#> 120 2024-11-18 23:23:45 0.697207 0.697535
#> 121 2024-11-18 23:23:46 0.697207 0.697533
#> 122 2024-11-18 23:23:46 0.697207 0.697501
#> 123 2024-11-18 23:23:46 0.697206 0.697501
#> 124 2024-11-18 23:23:46 0.697206 0.697500
#> 125 2024-11-18 23:23:47 0.697206 0.697499
#> 126 2024-11-18 23:23:47 0.697206 0.697498
#> 127 2024-11-18 23:23:47 0.697206 0.697497
#> 128 2024-11-18 23:23:47 0.697206 0.697496
#> 129 2024-11-18 23:23:47 0.697206 0.697501
#> 130 2024-11-18 23:23:47 0.697206 0.697495
#> 131 2024-11-18 23:23:47 0.697206 0.697494
#> 132 2024-11-18 23:23:47 0.697206 0.697501
#> 133 2024-11-18 23:23:47 0.697206 0.697493
#> 134 2024-11-18 23:23:47 0.697206 0.697492
#> 135 2024-11-18 23:23:47 0.697206 0.697501
#> 136 2024-11-18 23:23:48 0.697206 0.697499
#> 137 2024-11-18 23:23:48 0.697206 0.697500
#> 138 2024-11-18 23:23:48 0.697206 0.697500
#> 139 2024-11-18 23:23:48 0.697206 0.697467
#> 140 2024-11-18 23:23:48 0.697206 0.697466
#> 141 2024-11-18 23:23:49 0.697206 0.697465
#> 142 2024-11-18 23:23:49 0.697206 0.697444
#> 143 2024-11-18 23:23:49 0.697206 0.697443
#> 144 2024-11-18 23:23:49 0.697207 0.697443
#> 145 2024-11-18 23:23:49 0.697207 0.697465
#> 146 2024-11-18 23:23:49 0.697207 0.697465
#> 147 2024-11-18 23:23:49 0.697207 0.697465
#> 148 2024-11-18 23:23:50 0.697198 0.697464
#> 149 2024-11-18 23:23:50 0.697208 0.697464
#> 150 2024-11-18 23:23:50 0.697208 0.697512
#> 151 2024-11-18 23:23:50 0.697208 0.697463
#> 152 2024-11-18 23:23:50 0.697209 0.697463
#> 153 2024-11-18 23:23:50 0.697209 0.697462
#> 154 2024-11-18 23:23:51 0.697209 0.697459
#> 155 2024-11-18 23:23:51 0.697209 0.697458
#> 156 2024-11-18 23:23:52 0.697208 0.697458
#> 157 2024-11-18 23:23:52 0.697208 0.697457
#> 158 2024-11-18 23:23:52 0.697208 0.697456
#> 159 2024-11-18 23:23:52 0.697208 0.697454
#> 160 2024-11-18 23:23:52 0.697208 0.697453
#> 161 2024-11-18 23:23:53 0.697173 0.697453
#> 162 2024-11-18 23:23:54 0.697174 0.697453
#> 163 2024-11-18 23:23:54 0.697174 0.697452
#> 164 2024-11-18 23:23:54 0.697174 0.697420
#> 165 2024-11-18 23:23:54 0.697174 0.697419
#> 166 2024-11-18 23:23:55 0.697174 0.697418
#> 167 2024-11-18 23:23:55 0.697174 0.697417
#> 168 2024-11-18 23:23:55 0.697174 0.697420
#> 169 2024-11-18 23:23:55 0.697174 0.697416
#> 170 2024-11-18 23:23:55 0.697174 0.697415
#> 171 2024-11-18 23:23:55 0.697174 0.697420
#> 172 2024-11-18 23:23:55 0.697174 0.697414
#> 173 2024-11-18 23:23:55 0.697174 0.697413
#> 174 2024-11-18 23:23:55 0.697174 0.697420
#> 175 2024-11-18 23:23:55 0.697174 0.697412
#> 176 2024-11-18 23:23:55 0.697174 0.697411
#> 177 2024-11-18 23:23:55 0.697174 0.697420
#> 178 2024-11-18 23:23:56 0.697174 0.697418
#> 179 2024-11-18 23:23:56 0.697174 0.697415
#> 180 2024-11-18 23:23:56 0.697174 0.697414
#> 181 2024-11-18 23:23:58 0.697173 0.697414
#> 182 2024-11-18 23:23:59 0.697173 0.697413
#> 183 2024-11-18 23:23:59 0.697173 0.697381
#> 184 2024-11-18 23:23:59 0.697173 0.697380
#> 185 2024-11-18 23:23:59 0.697174 0.697380
#> 186 2024-11-18 23:23:59 0.697174 0.697413
#> 187 2024-11-18 23:24:00 0.697173 0.697412
#> 188 2024-11-18 23:24:00 0.697173 0.697405
#> 189 2024-11-18 23:24:00 0.697173 0.697404
#> 190 2024-11-18 23:24:01 0.697174 0.697404
#> 191 2024-11-18 23:24:01 0.697174 0.697403
#> 192 2024-11-18 23:24:01 0.697174 0.697371
#> 193 2024-11-18 23:24:01 0.697174 0.697370
#> 194 2024-11-18 23:24:01 0.697174 0.697403
#> 195 2024-11-18 23:24:01 0.697174 0.697395
#> 196 2024-11-18 23:24:01 0.697174 0.697394
#> 197 2024-11-18 23:24:02 0.697174 0.697393
#> 198 2024-11-18 23:24:02 0.697174 0.697392
#> 199 2024-11-18 23:24:03 0.697174 0.697391
#> 200 2024-11-18 23:24:03 0.697174 0.697390
#> 201 2024-11-18 23:24:03 0.697174 0.697393
#> 202 2024-11-18 23:24:03 0.697174 0.697392
#> 203 2024-11-18 23:24:03 0.697174 0.697382
#> 204 2024-11-18 23:24:03 0.697174 0.697381
#> 205 2024-11-18 23:24:03 0.697174 0.697393
#> 206 2024-11-18 23:24:03 0.697174 0.697393
#> 207 2024-11-18 23:24:03 0.697174 0.697393
#> 208 2024-11-18 23:24:04 0.697174 0.697391
#> 209 2024-11-18 23:24:04 0.697174 0.697390
#> 210 2024-11-18 23:24:05 0.697174 0.697457
#> 211 2024-11-18 23:24:05 0.697174 0.697455
#> 
#> $last
#> [1] 1731972245
#>