Retrieve Recent Spreads Data from Kraken Exchange

This function fetches recent spread data from the Kraken API for a specified trading pair.

Usage

getSpreads(pair, since = NULL, timestamp = FALSE)

Arguments

pair

A character string specifying the trading pair (e.g., "XTZUSD", "ADAEUR"). This is a required parameter.

since

A character string for a human-readable date-time (e.g., "2024-10-01 12:00:00") or a Unix timestamp. Default is NULL.

timestamp

A logical value. If TRUE, the function returns both the spreads data frame and the last timestamp for future polling. If FALSE (default), it returns only the data frame.

Value

A data frame containing the spread data or a list containing the data frame and the last timestamp.

Examples

getSpreads("XTZUSD")
#>                    Time    Bid    Ask
#> 1   2025-03-13 16:08:52 0.6759 0.6768
#> 2   2025-03-13 16:08:52 0.6760 0.6768
#> 3   2025-03-13 16:08:52 0.6761 0.6768
#> 4   2025-03-13 16:08:53 0.6759 0.6768
#> 5   2025-03-13 16:08:53 0.6761 0.6768
#> 6   2025-03-13 16:08:53 0.6762 0.6768
#> 7   2025-03-13 16:08:53 0.6759 0.6768
#> 8   2025-03-13 16:08:54 0.6760 0.6768
#> 9   2025-03-13 16:08:54 0.6761 0.6768
#> 10  2025-03-13 16:08:54 0.6759 0.6768
#> 11  2025-03-13 16:08:55 0.6760 0.6768
#> 12  2025-03-13 16:08:55 0.6761 0.6768
#> 13  2025-03-13 16:08:55 0.6760 0.6768
#> 14  2025-03-13 16:08:55 0.6760 0.6768
#> 15  2025-03-13 16:08:55 0.6759 0.6768
#> 16  2025-03-13 16:08:56 0.6760 0.6768
#> 17  2025-03-13 16:08:56 0.6761 0.6768
#> 18  2025-03-13 16:08:56 0.6759 0.6768
#> 19  2025-03-13 16:08:56 0.6760 0.6768
#> 20  2025-03-13 16:08:56 0.6760 0.6769
#> 21  2025-03-13 16:08:56 0.6759 0.6769
#> 22  2025-03-13 16:08:57 0.6760 0.6769
#> 23  2025-03-13 16:08:57 0.6761 0.6769
#> 24  2025-03-13 16:08:58 0.6762 0.6769
#> 25  2025-03-13 16:08:58 0.6763 0.6769
#> 26  2025-03-13 16:08:59 0.6764 0.6769
#> 27  2025-03-13 16:09:00 0.6765 0.6769
#> 28  2025-03-13 16:09:00 0.6766 0.6769
#> 29  2025-03-13 16:09:00 0.6767 0.6769
#> 30  2025-03-13 16:09:00 0.6768 0.6769
#> 31  2025-03-13 16:09:01 0.6768 0.6774
#> 32  2025-03-13 16:09:01 0.6768 0.6773
#> 33  2025-03-13 16:09:02 0.6769 0.6773
#> 34  2025-03-13 16:09:03 0.6770 0.6773
#> 35  2025-03-13 16:09:03 0.6770 0.6774
#> 36  2025-03-13 16:09:05 0.6769 0.6774
#> 37  2025-03-13 16:09:05 0.6770 0.6774
#> 38  2025-03-13 16:09:05 0.6771 0.6774
#> 39  2025-03-13 16:09:05 0.6769 0.6774
#> 40  2025-03-13 16:09:05 0.6770 0.6774
#> 41  2025-03-13 16:09:05 0.6771 0.6774
#> 42  2025-03-13 16:09:06 0.6771 0.6773
#> 43  2025-03-13 16:09:06 0.6771 0.6774
#> 44  2025-03-13 16:09:06 0.6771 0.6773
#> 45  2025-03-13 16:09:07 0.6771 0.6772
#> 46  2025-03-13 16:09:08 0.6769 0.6772
#> 47  2025-03-13 16:09:10 0.6769 0.6771
#> 48  2025-03-13 16:09:10 0.6769 0.6770
#> 49  2025-03-13 16:09:11 0.6763 0.6770
#> 50  2025-03-13 16:09:11 0.6763 0.6769
#> 51  2025-03-13 16:09:12 0.6763 0.6768
#> 52  2025-03-13 16:09:13 0.6763 0.6767
#> 53  2025-03-13 16:09:13 0.6763 0.6768
#> 54  2025-03-13 16:09:14 0.6763 0.6767
#> 55  2025-03-13 16:09:14 0.6762 0.6767
#> 56  2025-03-13 16:09:15 0.6762 0.6768
#> 57  2025-03-13 16:09:15 0.6762 0.6767
#> 58  2025-03-13 16:09:15 0.6762 0.6766
#> 59  2025-03-13 16:09:17 0.6763 0.6766
#> 60  2025-03-13 16:09:18 0.6763 0.6767
#> 61  2025-03-13 16:09:18 0.6763 0.6769
#> 62  2025-03-13 16:09:18 0.6763 0.6768
#> 63  2025-03-13 16:09:18 0.6763 0.6767
#> 64  2025-03-13 16:09:19 0.6763 0.6769
#> 65  2025-03-13 16:09:19 0.6763 0.6767
#> 66  2025-03-13 16:09:19 0.6763 0.6769
#> 67  2025-03-13 16:09:20 0.6763 0.6768
#> 68  2025-03-13 16:09:20 0.6763 0.6767
#> 69  2025-03-13 16:09:24 0.6763 0.6769
#> 70  2025-03-13 16:09:24 0.6763 0.6768
#> 71  2025-03-13 16:09:24 0.6763 0.6767
#> 72  2025-03-13 16:09:25 0.6763 0.6769
#> 73  2025-03-13 16:09:25 0.6763 0.6768
#> 74  2025-03-13 16:09:25 0.6763 0.6767
#> 75  2025-03-13 16:09:27 0.6763 0.6769
#> 76  2025-03-13 16:09:27 0.6763 0.6767
#> 77  2025-03-13 16:09:27 0.6763 0.6769
#> 78  2025-03-13 16:09:27 0.6762 0.6769
#> 79  2025-03-13 16:09:27 0.6762 0.6768
#> 80  2025-03-13 16:09:27 0.6762 0.6767
#> 81  2025-03-13 16:09:31 0.6762 0.6766
#> 82  2025-03-13 16:09:31 0.6762 0.6765
#> 83  2025-03-13 16:09:32 0.6761 0.6765
#> 84  2025-03-13 16:09:33 0.6761 0.6764
#> 85  2025-03-13 16:09:33 0.6761 0.6763
#> 86  2025-03-13 16:09:35 0.6759 0.6763
#> 87  2025-03-13 16:09:36 0.6759 0.6766
#> 88  2025-03-13 16:09:36 0.6759 0.6763
#> 89  2025-03-13 16:09:36 0.6759 0.6762
#> 90  2025-03-13 16:09:36 0.6759 0.6766
#> 91  2025-03-13 16:09:36 0.6759 0.6764
#> 92  2025-03-13 16:09:36 0.6759 0.6763
#> 93  2025-03-13 16:09:39 0.6759 0.6766
#> 94  2025-03-13 16:09:39 0.6759 0.6763
#> 95  2025-03-13 16:09:39 0.6759 0.6762
#> 96  2025-03-13 16:09:39 0.6759 0.6766
#> 97  2025-03-13 16:09:40 0.6759 0.6763
#> 98  2025-03-13 16:09:40 0.6759 0.6762
#> 99  2025-03-13 16:09:40 0.6759 0.6760
#> 100 2025-03-13 16:09:40 0.6759 0.6762
#> 101 2025-03-13 16:09:40 0.6759 0.6763
#> 102 2025-03-13 16:09:40 0.6759 0.6763
#> 103 2025-03-13 16:09:40 0.6759 0.6760
#> 104 2025-03-13 16:09:40 0.6752 0.6760
#> 105 2025-03-13 16:09:40 0.6752 0.6759
#> 106 2025-03-13 16:09:41 0.6753 0.6759
#> 107 2025-03-13 16:09:42 0.6754 0.6759
#> 108 2025-03-13 16:09:42 0.6753 0.6759
#> 109 2025-03-13 16:09:42 0.6753 0.6759
#> 110 2025-03-13 16:09:42 0.6755 0.6759
#> 111 2025-03-13 16:09:42 0.6755 0.6760
#> 112 2025-03-13 16:09:42 0.6756 0.6760
#> 113 2025-03-13 16:09:42 0.6756 0.6764
#> 114 2025-03-13 16:09:42 0.6757 0.6763
#> 115 2025-03-13 16:09:42 0.6758 0.6763
#> 116 2025-03-13 16:09:42 0.6758 0.6766
#> 117 2025-03-13 16:09:43 0.6759 0.6765
#> 118 2025-03-13 16:09:43 0.6759 0.6764
#> 119 2025-03-13 16:09:43 0.6759 0.6763
#> 120 2025-03-13 16:09:43 0.6759 0.6766
#> 121 2025-03-13 16:09:44 0.6760 0.6765
#> 122 2025-03-13 16:09:44 0.6760 0.6764
#> 123 2025-03-13 16:09:47 0.6761 0.6764
#> 124 2025-03-13 16:09:47 0.6761 0.6765
#> 125 2025-03-13 16:09:47 0.6761 0.6769
#> 126 2025-03-13 16:09:47 0.6761 0.6766
#> 127 2025-03-13 16:09:47 0.6761 0.6768
#> 128 2025-03-13 16:09:47 0.6761 0.6767
#> 129 2025-03-13 16:09:48 0.6761 0.6766
#> 130 2025-03-13 16:09:51 0.6762 0.6766
#> 131 2025-03-13 16:09:53 0.6762 0.6768
#> 132 2025-03-13 16:09:53 0.6762 0.6766
#> 133 2025-03-13 16:09:53 0.6762 0.6765
#> 134 2025-03-13 16:09:59 0.6762 0.6768
#> 135 2025-03-13 16:09:59 0.6762 0.6765
#> 136 2025-03-13 16:09:59 0.6762 0.6766
#> 137 2025-03-13 16:09:59 0.6762 0.6768
#> 138 2025-03-13 16:09:59 0.6760 0.6767
#> 139 2025-03-13 16:09:59 0.6760 0.6766
#> 140 2025-03-13 16:09:59 0.6760 0.6765
#> 141 2025-03-13 16:10:00 0.6760 0.6769
#> 142 2025-03-13 16:10:00 0.6760 0.6764
#> 143 2025-03-13 16:10:00 0.6760 0.6766
#> 144 2025-03-13 16:10:00 0.6760 0.6769
#> 145 2025-03-13 16:10:00 0.6760 0.6763
#> 146 2025-03-13 16:10:00 0.6760 0.6769
#> 147 2025-03-13 16:10:00 0.6760 0.6768
#> 148 2025-03-13 16:10:00 0.6760 0.6767
#> 149 2025-03-13 16:10:01 0.6759 0.6767
#> 150 2025-03-13 16:10:01 0.6759 0.6766
#> 151 2025-03-13 16:10:01 0.6759 0.6765
#> 152 2025-03-13 16:10:02 0.6759 0.6764
#> 153 2025-03-13 16:10:02 0.6759 0.6763
#> 154 2025-03-13 16:10:04 0.6759 0.6762
#> 155 2025-03-13 16:10:06 0.6759 0.6761
#> 156 2025-03-13 16:10:06 0.6759 0.6760
#> 157 2025-03-13 16:10:07 0.6757 0.6760
#> 158 2025-03-13 16:10:08 0.6757 0.6762
#> 159 2025-03-13 16:10:08 0.6757 0.6759
#> 160 2025-03-13 16:10:08 0.6757 0.6758
#> 161 2025-03-13 16:10:08 0.6757 0.6760
#> 162 2025-03-13 16:10:09 0.6754 0.6759
#> 163 2025-03-13 16:10:10 0.6754 0.6758
#> 164 2025-03-13 16:10:10 0.6754 0.6757
#> 165 2025-03-13 16:10:11 0.6752 0.6757
#> 166 2025-03-13 16:10:11 0.6752 0.6758
#> 167 2025-03-13 16:10:12 0.6753 0.6758
#> 168 2025-03-13 16:10:13 0.6754 0.6758
#> 169 2025-03-13 16:10:15 0.6752 0.6758
#> 170 2025-03-13 16:10:15 0.6752 0.6757
#> 171 2025-03-13 16:10:15 0.6752 0.6758
#> 172 2025-03-13 16:10:15 0.6752 0.6757
#> 173 2025-03-13 16:10:15 0.6752 0.6758
#> 174 2025-03-13 16:10:15 0.6751 0.6758
#> 175 2025-03-13 16:10:15 0.6752 0.6758
#> 176 2025-03-13 16:10:15 0.6753 0.6758
#> 177 2025-03-13 16:10:16 0.6753 0.6759
#> 178 2025-03-13 16:10:16 0.6753 0.6762
#> 179 2025-03-13 16:10:16 0.6753 0.6760
#> 180 2025-03-13 16:10:17 0.6754 0.6760
#> 181 2025-03-13 16:10:17 0.6755 0.6760
#> 182 2025-03-13 16:10:17 0.6754 0.6760
#> 183 2025-03-13 16:10:18 0.6754 0.6759
#> 184 2025-03-13 16:10:19 0.6754 0.6762
#> 185 2025-03-13 16:10:19 0.6754 0.6758
#> 186 2025-03-13 16:10:19 0.6754 0.6759
#> 187 2025-03-13 16:10:19 0.6754 0.6760
#> 188 2025-03-13 16:10:19 0.6754 0.6760
#> 189 2025-03-13 16:10:20 0.6753 0.6759
#> 190 2025-03-13 16:10:20 0.6753 0.6758
#> 191 2025-03-13 16:10:20 0.6753 0.6757
#> 192 2025-03-13 16:10:28 0.6753 0.6760
#> 193 2025-03-13 16:10:28 0.6753 0.6757
#> 194 2025-03-13 16:10:28 0.6753 0.6759
#> 195 2025-03-13 16:10:29 0.6753 0.6758
#> 196 2025-03-13 16:10:29 0.6753 0.6757
#> 197 2025-03-13 16:10:30 0.6753 0.6758
#> 198 2025-03-13 16:10:30 0.6753 0.6757
#> 199 2025-03-13 16:10:30 0.6753 0.6758
#> 200 2025-03-13 16:10:32 0.6753 0.6759
#> 201 2025-03-13 16:10:32 0.6753 0.6760
#> 202 2025-03-13 16:10:33 0.6753 0.6761
#> 203 2025-03-13 16:10:33 0.6750 0.6761
#> 204 2025-03-13 16:10:33 0.6754 0.6761
#> 205 2025-03-13 16:10:33 0.6750 0.6761
#> 206 2025-03-13 16:10:33 0.6750 0.6760
#> 207 2025-03-13 16:10:33 0.6751 0.6760
#> 208 2025-03-13 16:10:33 0.6752 0.6760
#> 209 2025-03-13 16:10:33 0.6753 0.6760
#> 210 2025-03-13 16:10:33 0.6753 0.6759
#> 211 2025-03-13 16:10:38 0.6753 0.6762
#> 212 2025-03-13 16:10:38 0.6753 0.6758
#> 213 2025-03-13 16:10:38 0.6753 0.6761
#> 214 2025-03-13 16:10:38 0.6753 0.6762
#> 215 2025-03-13 16:10:38 0.6753 0.6762
#> 216 2025-03-13 16:10:39 0.6753 0.6760
#> 217 2025-03-13 16:10:39 0.6753 0.6759
#> 218 2025-03-13 16:10:42 0.6753 0.6758
#> 219 2025-03-13 16:10:42 0.6753 0.6757
#> 220 2025-03-13 16:10:42 0.6753 0.6758
#> 221 2025-03-13 16:10:42 0.6753 0.6757
#> 222 2025-03-13 16:10:44 0.6753 0.6759
#> 223 2025-03-13 16:10:44 0.6753 0.6758
#> 224 2025-03-13 16:10:44 0.6753 0.6757
#> 225 2025-03-13 16:10:47 0.6753 0.6758
#> 226 2025-03-13 16:10:47 0.6753 0.6759
#> 227 2025-03-13 16:10:47 0.6753 0.6758
#> 228 2025-03-13 16:10:48 0.6754 0.6758
#> 229 2025-03-13 16:10:51 0.6754 0.6760
#> 230 2025-03-13 16:10:51 0.6754 0.6758
#> 231 2025-03-13 16:10:55 0.6754 0.6757
#> 232 2025-03-13 16:11:02 0.6754 0.6760
#> 233 2025-03-13 16:11:02 0.6754 0.6758
#> 234 2025-03-13 16:11:02 0.6754 0.6757
#> 235 2025-03-13 16:11:02 0.6752 0.6757
#> 236 2025-03-13 16:11:02 0.6751 0.6757
#> 237 2025-03-13 16:11:03 0.6752 0.6757
#> 238 2025-03-13 16:11:04 0.6751 0.6757
#> 239 2025-03-13 16:11:04 0.6753 0.6757
#> 240 2025-03-13 16:11:04 0.6751 0.6757
#> 241 2025-03-13 16:11:04 0.6752 0.6757
#> 242 2025-03-13 16:11:04 0.6751 0.6757
#> 243 2025-03-13 16:11:04 0.6751 0.6756
#> 244 2025-03-13 16:11:05 0.6751 0.6758
#> 245 2025-03-13 16:11:05 0.6751 0.6756
#> 246 2025-03-13 16:11:05 0.6751 0.6755
#> 247 2025-03-13 16:11:10 0.6751 0.6758
#> 248 2025-03-13 16:11:10 0.6752 0.6757
#> 249 2025-03-13 16:11:10 0.6752 0.6756
#> 250 2025-03-13 16:11:11 0.6753 0.6756
getSpreads("ADAEUR", since = "2024-10-01 00:00:00", timestamp = TRUE)
#> $spreads
#>                    Time      Bid      Ask
#> 1   2025-03-13 16:10:46 0.650340 0.651054
#> 2   2025-03-13 16:10:46 0.650340 0.651053
#> 3   2025-03-13 16:10:46 0.650340 0.651054
#> 4   2025-03-13 16:10:46 0.650340 0.651051
#> 5   2025-03-13 16:10:46 0.650340 0.651050
#> 6   2025-03-13 16:10:46 0.650340 0.651049
#> 7   2025-03-13 16:10:46 0.650340 0.651048
#> 8   2025-03-13 16:10:46 0.650340 0.651049
#> 9   2025-03-13 16:10:46 0.650340 0.651049
#> 10  2025-03-13 16:10:46 0.650340 0.651047
#> 11  2025-03-13 16:10:46 0.650340 0.651046
#> 12  2025-03-13 16:10:46 0.650340 0.651045
#> 13  2025-03-13 16:10:46 0.650340 0.651044
#> 14  2025-03-13 16:10:46 0.650340 0.651043
#> 15  2025-03-13 16:10:46 0.650340 0.651045
#> 16  2025-03-13 16:10:46 0.650340 0.651043
#> 17  2025-03-13 16:10:46 0.650340 0.651042
#> 18  2025-03-13 16:10:46 0.650340 0.651041
#> 19  2025-03-13 16:10:46 0.650340 0.651040
#> 20  2025-03-13 16:10:46 0.650340 0.651039
#> 21  2025-03-13 16:10:46 0.650340 0.651040
#> 22  2025-03-13 16:10:46 0.650340 0.651040
#> 23  2025-03-13 16:10:46 0.650340 0.651037
#> 24  2025-03-13 16:10:46 0.650340 0.651036
#> 25  2025-03-13 16:10:47 0.650341 0.651036
#> 26  2025-03-13 16:10:47 0.650342 0.651036
#> 27  2025-03-13 16:10:47 0.650340 0.651036
#> 28  2025-03-13 16:10:47 0.650341 0.651036
#> 29  2025-03-13 16:10:47 0.650341 0.651035
#> 30  2025-03-13 16:10:47 0.650342 0.651035
#> 31  2025-03-13 16:10:47 0.650342 0.651034
#> 32  2025-03-13 16:10:47 0.650342 0.651035
#> 33  2025-03-13 16:10:47 0.650342 0.651035
#> 34  2025-03-13 16:10:47 0.650342 0.651035
#> 35  2025-03-13 16:10:48 0.650341 0.651034
#> 36  2025-03-13 16:10:48 0.650340 0.651034
#> 37  2025-03-13 16:10:48 0.650340 0.651033
#> 38  2025-03-13 16:10:48 0.650340 0.651032
#> 39  2025-03-13 16:10:48 0.650340 0.651031
#> 40  2025-03-13 16:10:48 0.650340 0.651030
#> 41  2025-03-13 16:10:48 0.650340 0.651029
#> 42  2025-03-13 16:10:48 0.650340 0.651033
#> 43  2025-03-13 16:10:48 0.650341 0.651032
#> 44  2025-03-13 16:10:48 0.650341 0.651031
#> 45  2025-03-13 16:10:48 0.650341 0.651030
#> 46  2025-03-13 16:10:48 0.650342 0.651030
#> 47  2025-03-13 16:10:49 0.650342 0.651029
#> 48  2025-03-13 16:10:49 0.650342 0.651028
#> 49  2025-03-13 16:10:49 0.650340 0.651028
#> 50  2025-03-13 16:10:50 0.650340 0.651010
#> 51  2025-03-13 16:10:50 0.650340 0.651009
#> 52  2025-03-13 16:10:50 0.650340 0.651008
#> 53  2025-03-13 16:10:50 0.650340 0.651007
#> 54  2025-03-13 16:10:51 0.650340 0.651006
#> 55  2025-03-13 16:10:51 0.650340 0.651005
#> 56  2025-03-13 16:10:51 0.650340 0.651004
#> 57  2025-03-13 16:10:51 0.650340 0.651003
#> 58  2025-03-13 16:10:51 0.650340 0.651002
#> 59  2025-03-13 16:10:51 0.650340 0.651001
#> 60  2025-03-13 16:10:51 0.650340 0.651000
#> 61  2025-03-13 16:10:51 0.650306 0.651000
#> 62  2025-03-13 16:10:51 0.650306 0.650732
#> 63  2025-03-13 16:10:51 0.650306 0.650731
#> 64  2025-03-13 16:10:51 0.650306 0.650730
#> 65  2025-03-13 16:10:51 0.650306 0.650729
#> 66  2025-03-13 16:10:51 0.650306 0.650728
#> 67  2025-03-13 16:10:51 0.650306 0.650727
#> 68  2025-03-13 16:10:51 0.650227 0.650727
#> 69  2025-03-13 16:10:51 0.650244 0.650727
#> 70  2025-03-13 16:10:51 0.650244 0.650726
#> 71  2025-03-13 16:10:51 0.650244 0.650725
#> 72  2025-03-13 16:10:52 0.650244 0.650724
#> 73  2025-03-13 16:10:52 0.650244 0.650723
#> 74  2025-03-13 16:10:52 0.650244 0.650722
#> 75  2025-03-13 16:10:52 0.650244 0.650721
#> 76  2025-03-13 16:10:53 0.650244 0.650720
#> 77  2025-03-13 16:10:53 0.650244 0.650719
#> 78  2025-03-13 16:10:53 0.650244 0.650718
#> 79  2025-03-13 16:10:53 0.650245 0.650718
#> 80  2025-03-13 16:10:54 0.650246 0.650718
#> 81  2025-03-13 16:10:54 0.650251 0.650718
#> 82  2025-03-13 16:10:54 0.650252 0.650718
#> 83  2025-03-13 16:10:55 0.650253 0.650718
#> 84  2025-03-13 16:10:55 0.650253 0.650726
#> 85  2025-03-13 16:10:55 0.650253 0.650717
#> 86  2025-03-13 16:10:55 0.650253 0.650716
#> 87  2025-03-13 16:10:55 0.650253 0.650715
#> 88  2025-03-13 16:10:55 0.650253 0.650714
#> 89  2025-03-13 16:10:55 0.650253 0.650717
#> 90  2025-03-13 16:10:57 0.650253 0.650715
#> 91  2025-03-13 16:10:57 0.650253 0.650714
#> 92  2025-03-13 16:10:57 0.650253 0.650713
#> 93  2025-03-13 16:10:57 0.650253 0.650712
#> 94  2025-03-13 16:10:57 0.650253 0.650711
#> 95  2025-03-13 16:10:57 0.650253 0.650710
#> 96  2025-03-13 16:10:58 0.650253 0.650709
#> 97  2025-03-13 16:10:58 0.650253 0.650708
#> 98  2025-03-13 16:10:58 0.650253 0.650713
#> 99  2025-03-13 16:11:00 0.650253 0.650711
#> 100 2025-03-13 16:11:00 0.650253 0.650710
#> 101 2025-03-13 16:11:01 0.650253 0.650709
#> 102 2025-03-13 16:11:01 0.650253 0.650708
#> 103 2025-03-13 16:11:01 0.650253 0.650707
#> 104 2025-03-13 16:11:01 0.650253 0.650706
#> 105 2025-03-13 16:11:01 0.650253 0.650705
#> 106 2025-03-13 16:11:01 0.650253 0.650704
#> 107 2025-03-13 16:11:01 0.650253 0.650703
#> 108 2025-03-13 16:11:01 0.650235 0.650703
#> 109 2025-03-13 16:11:01 0.650235 0.650702
#> 110 2025-03-13 16:11:02 0.650230 0.650702
#> 111 2025-03-13 16:11:02 0.650230 0.650701
#> 112 2025-03-13 16:11:02 0.650230 0.650700
#> 113 2025-03-13 16:11:02 0.650230 0.650699
#> 114 2025-03-13 16:11:02 0.650230 0.650698
#> 115 2025-03-13 16:11:02 0.650214 0.650698
#> 116 2025-03-13 16:11:02 0.650076 0.650698
#> 117 2025-03-13 16:11:02 0.649963 0.650698
#> 118 2025-03-13 16:11:02 0.649940 0.650698
#> 119 2025-03-13 16:11:02 0.649964 0.650698
#> 120 2025-03-13 16:11:02 0.649964 0.650697
#> 121 2025-03-13 16:11:02 0.649964 0.650696
#> 122 2025-03-13 16:11:02 0.649965 0.650696
#> 123 2025-03-13 16:11:02 0.649965 0.650695
#> 124 2025-03-13 16:11:02 0.649965 0.650694
#> 125 2025-03-13 16:11:02 0.649971 0.650694
#> 126 2025-03-13 16:11:02 0.649972 0.650694
#> 127 2025-03-13 16:11:03 0.649973 0.650694
#> 128 2025-03-13 16:11:03 0.649973 0.650697
#> 129 2025-03-13 16:11:03 0.649973 0.650695
#> 130 2025-03-13 16:11:03 0.649973 0.650694
#> 131 2025-03-13 16:11:03 0.649973 0.650697
#> 132 2025-03-13 16:11:03 0.649972 0.650696
#> 133 2025-03-13 16:11:03 0.649973 0.650696
#> 134 2025-03-13 16:11:04 0.649974 0.650696
#> 135 2025-03-13 16:11:04 0.649974 0.650695
#> 136 2025-03-13 16:11:04 0.649974 0.650694
#> 137 2025-03-13 16:11:04 0.649974 0.650693
#> 138 2025-03-13 16:11:04 0.649974 0.650692
#> 139 2025-03-13 16:11:04 0.649974 0.650691
#> 140 2025-03-13 16:11:04 0.649975 0.650691
#> 141 2025-03-13 16:11:04 0.649975 0.650690
#> 142 2025-03-13 16:11:04 0.649976 0.650690
#> 143 2025-03-13 16:11:04 0.649976 0.650689
#> 144 2025-03-13 16:11:04 0.649975 0.650689
#> 145 2025-03-13 16:11:04 0.649975 0.650688
#> 146 2025-03-13 16:11:05 0.649975 0.650687
#> 147 2025-03-13 16:11:05 0.649975 0.650686
#> 148 2025-03-13 16:11:05 0.649975 0.650685
#> 149 2025-03-13 16:11:05 0.649975 0.650684
#> 150 2025-03-13 16:11:05 0.649975 0.650683
#> 151 2025-03-13 16:11:05 0.649975 0.650682
#> 152 2025-03-13 16:11:05 0.649975 0.650681
#> 153 2025-03-13 16:11:05 0.649975 0.650680
#> 154 2025-03-13 16:11:05 0.649975 0.650679
#> 155 2025-03-13 16:11:05 0.649975 0.650678
#> 156 2025-03-13 16:11:05 0.649975 0.650677
#> 157 2025-03-13 16:11:05 0.649975 0.650676
#> 158 2025-03-13 16:11:05 0.649975 0.650675
#> 159 2025-03-13 16:11:05 0.649975 0.650674
#> 160 2025-03-13 16:11:05 0.649975 0.650673
#> 161 2025-03-13 16:11:05 0.649975 0.650672
#> 162 2025-03-13 16:11:05 0.649975 0.650671
#> 163 2025-03-13 16:11:05 0.649975 0.650670
#> 164 2025-03-13 16:11:05 0.649975 0.650669
#> 165 2025-03-13 16:11:05 0.649975 0.650668
#> 166 2025-03-13 16:11:05 0.649975 0.650667
#> 167 2025-03-13 16:11:05 0.649975 0.650666
#> 168 2025-03-13 16:11:05 0.649975 0.650665
#> 169 2025-03-13 16:11:05 0.649975 0.650664
#> 170 2025-03-13 16:11:05 0.649975 0.650663
#> 171 2025-03-13 16:11:05 0.649975 0.650662
#> 172 2025-03-13 16:11:05 0.649975 0.650661
#> 173 2025-03-13 16:11:05 0.649975 0.650660
#> 174 2025-03-13 16:11:05 0.649975 0.650659
#> 175 2025-03-13 16:11:05 0.649975 0.650658
#> 176 2025-03-13 16:11:05 0.649975 0.650657
#> 177 2025-03-13 16:11:05 0.649975 0.650656
#> 178 2025-03-13 16:11:06 0.649975 0.650655
#> 179 2025-03-13 16:11:06 0.649975 0.650654
#> 180 2025-03-13 16:11:06 0.649975 0.650653
#> 181 2025-03-13 16:11:06 0.649975 0.650652
#> 182 2025-03-13 16:11:06 0.649975 0.650651
#> 183 2025-03-13 16:11:06 0.649975 0.650650
#> 184 2025-03-13 16:11:06 0.649892 0.650650
#> 185 2025-03-13 16:11:06 0.649892 0.650649
#> 186 2025-03-13 16:11:06 0.649893 0.650649
#> 187 2025-03-13 16:11:06 0.649893 0.650648
#> 188 2025-03-13 16:11:07 0.649893 0.650647
#> 189 2025-03-13 16:11:07 0.649893 0.650646
#> 190 2025-03-13 16:11:07 0.649893 0.650645
#> 191 2025-03-13 16:11:07 0.649893 0.650644
#> 192 2025-03-13 16:11:07 0.649893 0.650643
#> 193 2025-03-13 16:11:07 0.649893 0.650606
#> 194 2025-03-13 16:11:07 0.649893 0.650605
#> 195 2025-03-13 16:11:07 0.649893 0.650604
#> 196 2025-03-13 16:11:07 0.649893 0.650603
#> 197 2025-03-13 16:11:07 0.649893 0.650606
#> 198 2025-03-13 16:11:08 0.649893 0.650604
#> 199 2025-03-13 16:11:08 0.649894 0.650604
#> 200 2025-03-13 16:11:08 0.649894 0.650603
#> 201 2025-03-13 16:11:08 0.649894 0.650602
#> 202 2025-03-13 16:11:08 0.649894 0.650601
#> 203 2025-03-13 16:11:08 0.649894 0.650606
#> 204 2025-03-13 16:11:08 0.649894 0.650604
#> 205 2025-03-13 16:11:08 0.649894 0.650603
#> 206 2025-03-13 16:11:08 0.649895 0.650603
#> 207 2025-03-13 16:11:08 0.649895 0.650602
#> 208 2025-03-13 16:11:09 0.649896 0.650602
#> 209 2025-03-13 16:11:09 0.649895 0.650602
#> 210 2025-03-13 16:11:09 0.649895 0.650601
#> 211 2025-03-13 16:11:09 0.649895 0.650600
#> 212 2025-03-13 16:11:09 0.649895 0.650599
#> 213 2025-03-13 16:11:09 0.649895 0.650598
#> 214 2025-03-13 16:11:09 0.649895 0.650597
#> 215 2025-03-13 16:11:09 0.649895 0.650596
#> 216 2025-03-13 16:11:09 0.649895 0.650595
#> 217 2025-03-13 16:11:10 0.649895 0.650594
#> 218 2025-03-13 16:11:10 0.649895 0.650593
#> 219 2025-03-13 16:11:10 0.649895 0.650592
#> 220 2025-03-13 16:11:10 0.649895 0.650591
#> 221 2025-03-13 16:11:10 0.649895 0.650590
#> 222 2025-03-13 16:11:10 0.649895 0.650589
#> 223 2025-03-13 16:11:10 0.649895 0.650588
#> 224 2025-03-13 16:11:10 0.649895 0.650587
#> 225 2025-03-13 16:11:10 0.649895 0.650586
#> 226 2025-03-13 16:11:10 0.649895 0.650585
#> 227 2025-03-13 16:11:10 0.649895 0.650586
#> 228 2025-03-13 16:11:10 0.649895 0.650586
#> 229 2025-03-13 16:11:11 0.649895 0.650584
#> 230 2025-03-13 16:11:11 0.649895 0.650583
#> 231 2025-03-13 16:11:11 0.649859 0.650583
#> 232 2025-03-13 16:11:11 0.649860 0.650583
#> 233 2025-03-13 16:11:11 0.649860 0.650582
#> 234 2025-03-13 16:11:11 0.649860 0.650581
#> 235 2025-03-13 16:11:11 0.649860 0.650580
#> 236 2025-03-13 16:11:11 0.649860 0.650575
#> 237 2025-03-13 16:11:11 0.649860 0.650574
#> 238 2025-03-13 16:11:11 0.649860 0.650573
#> 239 2025-03-13 16:11:11 0.649860 0.650574
#> 240 2025-03-13 16:11:11 0.649860 0.650574
#> 241 2025-03-13 16:11:11 0.649860 0.650574
#> 242 2025-03-13 16:11:12 0.649860 0.650572
#> 243 2025-03-13 16:11:12 0.649860 0.650571
#> 244 2025-03-13 16:11:12 0.649860 0.650570
#> 245 2025-03-13 16:11:12 0.649860 0.650569
#> 246 2025-03-13 16:11:12 0.649860 0.650568
#> 247 2025-03-13 16:11:12 0.649860 0.650567
#> 248 2025-03-13 16:11:12 0.649860 0.650544
#> 249 2025-03-13 16:11:12 0.649860 0.650543
#> 250 2025-03-13 16:11:12 0.649860 0.650542
#> 
#> $last
#> [1] 1741882272
#>